Set/get global faux options

Global faux options are used, for example, to set the default separator for cell names.

faux_options(...)

Arguments

...
One of four: (1) nothing, then returns all options as a list; (2) a name of an option element, then returns its value; (3) a name-value pair which sets the corresponding option to the new value (and returns nothing), (4) a list with option-value pairs which sets all the corresponding arguments.

Value

a list of options, values of an option, or nothing

Examples

faux_options() # see all options

#> $connection
#> A connection with                      
#> description "stdin"   
#> class       "terminal"
#> mode        "r"       
#> text        "text"    
#> opened      "opened"  
#> can read    "yes"     
#> can write   "no"      
#> 
#> $plot
#> [1] TRUE
#> 
#> $sep
#> [1] "_"
#> 
#> $verbose
#> [1] TRUE
#> 

faux_options("sep") # see value of faux.sep

#> [1] "_"

if (FALSE) {
# changes cell separator (e.g., A1.B2)
faux_options(sep = ".")

# changes cell separator back to default (e.g., A1_B2)
faux_options(sep = "_")
}