Attractiveness rating subset

The faceratings dataset cut down for demos to the first 4 raters of each sex and sexpref and the first 4 faces of each sex and ethnicity with non-NA ages

Usage

fr4

Format

A data frame with 768 rows and 9 variables:

rater_id

rater's ID

rater_sex

rater's sex (female, male)

rater_age

rater's age (17.4-54.3 years)

rater_sexpref

rater's preferred sex for romantic relationships (either, men, women)

face_id

face's ID

face_sex

face's sex (female, male)

face_age

face's age (19-47 years)

face_eth

face's ethnic group (black, east_asian, west_asian, white)

rating

attractiveness rating on a scale from 1 (much less attractive than average) to 7 (much more attractive than average)

Source

https://figshare.com/articles/dataset/Face_Research_Lab_London_Set/5047666