Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A           B           C
#> 1   -1.859493581  1.01685947  0.87438821
#> 2   -2.125008628 -2.33515793 -1.47372512
#> 3    0.802443869  0.70818312 -0.30484662
#> 4   -1.259361484 -0.80594761 -1.71808886
#> 5   -0.207949509 -0.40992100 -0.17901531
#> 6   -1.306889299 -2.30994108 -0.16641375
#> 7    1.209985727 -1.81310028 -0.35183577
#> 8   -0.508851154 -0.60475152 -0.93175718
#> 9   -2.324129003 -0.10385738 -1.38818735
#> 10  -1.379859128 -1.16579062 -2.43462675
#> 11  -0.998509968 -0.41410026 -1.80916342
#> 12  -0.429011944 -0.21981488 -0.65444276
#> 13  -1.273918600  0.23706820 -1.10536606
#> 14   1.120998548 -0.42865878  0.19029407
#> 15  -0.135298035  1.05917191  0.28632479
#> 16  -1.211301949  1.35967907 -1.73479318
#> 17   0.331475367 -0.01511224  0.38068609
#> 18   0.265738849 -1.71448867  0.04099720
#> 19  -0.170113824  0.79258895  0.26447012
#> 20   1.368907623  0.56891181  0.31617019
#> 21   1.526054433  0.10423910  0.47291544
#> 22   0.548881045  0.93721194  2.17509954
#> 23  -0.280724650 -1.14272206 -0.12131551
#> 24   0.462717595  0.79184786  0.29234866
#> 25   1.050706642 -0.34140502 -0.42411468
#> 26  -0.942908340  0.25439834 -0.30032850
#> 27   0.378733597  1.13155883  0.73435585
#> 28  -0.879166969  0.08982315 -0.05232397
#> 29  -1.319243232 -1.25735291  0.06610847
#> 30   2.319189408  0.10415982  0.60623286
#> 31   0.028463089  2.06321757  1.82482896
#> 32  -1.677992481  1.07152855  0.94772020
#> 33  -0.618689470 -1.10009277 -1.85389702
#> 34  -0.936545879  2.09529878  0.44908264
#> 35   0.711695981  0.30877100  0.82429569
#> 36   0.711909254  1.57201042 -0.02711852
#> 37  -1.471075679  0.07992224 -1.07605492
#> 38  -0.155892701 -0.26714129 -0.88052846
#> 39  -1.538130003  0.54042681  0.40720219
#> 40  -0.953091420 -1.91451367 -0.81144772
#> 41  -0.115034208  1.53099798  1.04410838
#> 42  -0.581352593 -1.32652919 -0.57263194
#> 43  -1.442599745 -1.00508032 -0.59851704
#> 44   0.671305928 -0.37642874 -0.13969696
#> 45  -1.190715096 -1.74003368 -2.37055336
#> 46   0.017034779 -1.75651051 -0.57981911
#> 47   0.637520898  1.13521834  0.03361620
#> 48   0.007778282  1.43819684 -0.42686158
#> 49  -0.848099033 -0.22082379 -0.22260862
#> 50   0.984156545 -0.40846088  0.27564496
#> 51  -0.214497954  1.08242891 -0.60365745
#> 52   1.722621175 -0.40724572  0.42132953
#> 53  -1.871005576  0.72296243 -0.04454420
#> 54  -0.493204350 -1.12025928 -0.84738905
#> 55  -0.074800794 -0.59088756 -0.20673204
#> 56  -0.209819319 -1.12032558  0.69701084
#> 57   0.241787224 -0.69746347 -0.49928905
#> 58  -0.855611568 -0.25968707  0.48506486
#> 59   1.713327450  1.18016084 -0.34951197
#> 60  -0.231584433  0.13634255 -0.42550561
#> 61   0.879934199  0.57704281  1.53762193
#> 62   1.296154322 -0.56327038 -0.35889074
#> 63   1.114188562  2.44188686 -0.39061919
#> 64  -0.782637980 -0.11251194 -1.02962996
#> 65   0.732082289  1.31053118  1.10068622
#> 66   0.521347945  2.20542216  0.87015142
#> 67  -1.970944739 -1.14502134 -0.38680221
#> 68   1.098428596 -0.22745603 -0.05948697
#> 69   0.800274928  1.51902541 -0.57523066
#> 70   0.128447679 -1.02466320 -1.52470591
#> 71   0.789147974  0.86932944  0.30561662
#> 72  -0.350037881  1.67725562  2.35163979
#> 73  -1.184738250 -0.55935644 -0.14742494
#> 74  -1.234845060 -0.46846781 -0.56162676
#> 75  -0.767528266 -2.37847472 -0.19418677
#> 76  -0.231024752 -1.01960884 -1.34310750
#> 77   0.115368675 -0.30743567 -0.62784851
#> 78   0.444683165 -0.64421006 -0.02697597
#> 79  -0.639125979  0.42570571  0.53332124
#> 80  -0.059797256 -2.05243267 -2.60145918
#> 81   0.832126291 -0.33291577  1.20249926
#> 82   0.088213946  1.43228597  0.22703331
#> 83   0.209145319  0.23645576 -0.16380314
#> 84  -0.553281399 -1.02650933 -0.43177470
#> 85  -0.059490603  0.22516756  0.69366978
#> 86  -1.225521942  0.43675957  0.48816576
#> 87  -0.949521852  1.59623828 -0.36708488
#> 88  -1.139403781  0.45443185 -0.89026669
#> 89  -0.063455808 -0.75389184 -0.45860481
#> 90   0.636944128  1.11729583  1.62801089
#> 91  -0.256321470  0.25860570  1.12284801
#> 92   0.026445433 -0.11547109 -0.67182655
#> 93  -0.387669288  0.53309359  0.30711844
#> 94   0.288645946 -1.39023598 -0.20257120
#> 95  -1.864608444 -0.34859015 -0.75255847
#> 96  -0.072093187 -0.60485402  0.86827352
#> 97  -1.168373265 -1.05016471 -0.92012951
#> 98   0.788866939  0.67665268  0.22147695
#> 99  -0.617368809 -1.35519114 -0.16106229
#> 100 -0.410644613  1.17494999 -0.88925089

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A           B           C
#> 1   -0.63762439  0.79210412  2.06190376
#> 2    0.14306350 -0.60042261 -0.05598382
#> 3   -2.30028459 -0.88261211  0.59562995
#> 4    0.19065181  0.99137791  0.20019454
#> 5    0.37382624 -0.49424065  0.24363650
#> 6    2.10292097  0.56035940 -1.50336040
#> 7    0.15526704 -0.52275882 -1.08921659
#> 8    2.31235212  1.11086332 -1.32339557
#> 9   -1.40507536  0.70855311  1.50554170
#> 10  -0.41885066 -0.34608462 -0.06550499
#> 11   0.59074455 -0.38592202 -1.47457726
#> 12  -0.56658197  1.02187992  0.11390894
#> 13  -0.35427103  0.98569220  1.23292399
#> 14  -1.73541336 -0.58110852  0.98991485
#> 15  -0.50869770  0.09061112  0.58426527
#> 16  -0.78269861 -0.58250245 -0.65768318
#> 17   0.49414192 -0.29293429 -1.16895616
#> 18  -0.16158382 -1.70593875 -0.45280150
#> 19   0.12287031 -0.27212817 -0.01136165
#> 20  -0.30031140 -2.29131531 -1.17393844
#> 21  -1.22602390 -0.52634468 -1.01291434
#> 22  -0.31393900 -0.57341991 -1.39499270
#> 23  -0.74304471  0.48244003  1.69639474
#> 24   1.14916734  1.80843068  0.33812844
#> 25  -0.14483766 -1.44173252 -0.14490631
#> 26  -0.85100740 -1.37800246 -0.17715925
#> 27   0.74801144  1.32268194  0.80401935
#> 28   0.64610956 -1.73892554 -1.31766000
#> 29  -0.51121780 -1.52307861 -0.48617524
#> 30   0.85172550 -0.59359472 -0.20588401
#> 31   0.11428093 -0.23052762 -0.15966973
#> 32  -1.05807636  0.02066788  0.58611977
#> 33  -0.43669881 -0.83987734 -0.65722738
#> 34   0.36478239 -0.47527868 -0.18514662
#> 35   0.02613481  0.02401229  0.86272394
#> 36   1.96052516  0.86077351 -0.32531674
#> 37   0.28791883 -0.41439033 -1.29439367
#> 38  -0.11653118  0.11308489 -0.48846829
#> 39  -0.33686265 -0.62043510 -0.11111913
#> 40  -1.01867970  0.09230307  1.33053998
#> 41   0.72301740  1.00777558 -0.03374153
#> 42   0.48672536 -0.92905396 -0.52079381
#> 43   2.32460683 -1.17245427 -2.73906448
#> 44   1.53078970 -0.48670203 -1.53338498
#> 45   0.22422389 -0.34923299 -0.74486394
#> 46   0.38338361  0.39856330  0.07539027
#> 47   1.66123468  0.59905602 -1.05597509
#> 48   0.08844146 -0.86968111 -0.49685087
#> 49   1.70363909  0.35891350 -2.10591334
#> 50   0.61611336  0.08785902 -0.67694317
#> 51  -2.29005240  1.49381206  2.84135264
#> 52  -2.06334453  0.92066829  1.54442491
#> 53  -0.03074958 -1.19619065 -0.81240041
#> 54  -0.13664543  1.76294135  2.90769438
#> 55  -0.99163088  1.53525713  0.78115781
#> 56   0.59011548 -1.06001427 -0.95604485
#> 57   0.65836452 -0.01903674 -1.78407315
#> 58  -0.40828309 -0.08879434  0.67879271
#> 59  -0.86908449  1.17279715  3.15900280
#> 60  -1.24244879  1.29340802  2.04700999
#> 61   0.95376325  0.71022067 -0.22096430
#> 62  -0.58924091  0.61989365  1.45571576
#> 63  -0.70805441 -0.60062631  0.35725117
#> 64  -0.26325225 -1.39173668 -0.54902278
#> 65  -0.91809725 -0.23218123  0.16060469
#> 66   2.09802750  0.35527648 -0.20934803
#> 67   0.08081285 -1.13598668 -1.35844117
#> 68   0.16733177  0.46454509  0.83501617
#> 69   0.82516143 -0.21662191 -0.45489511
#> 70  -1.09716395 -1.69116346  0.85887026
#> 71  -2.15132096  0.28530479  0.77419229
#> 72   0.84270741 -1.45666464 -0.91045608
#> 73  -0.24315773  0.60084009  0.05254245
#> 74  -0.89706082 -0.13637461  0.74655888
#> 75  -1.37446689  0.02016654  1.36445951
#> 76   0.74134295 -0.34597540 -0.64141952
#> 77   0.72871264  0.87650447 -0.02507217
#> 78   0.26287789  1.48529644 -0.21347562
#> 79  -0.45270014  0.21507355  1.75266991
#> 80  -0.14854219 -0.39591914 -0.15617849
#> 81  -2.04156218  0.62272546  1.35673231
#> 82   0.22946715  0.80395073 -0.89809758
#> 83  -0.14983471 -0.58732394  0.98928157
#> 84  -0.44004517  1.53023861  1.16319933
#> 85  -0.02101311 -1.26718063 -1.47317907
#> 86  -0.02546552  1.44837760  0.76941501
#> 87   1.82588039  1.58136349 -0.24698419
#> 88   1.01133275 -1.45754398 -1.18896984
#> 89   0.34733401  0.31477361 -0.54699059
#> 90  -0.67134440  1.04615966  1.28698274
#> 91  -0.34248594 -0.72373986  0.01508847
#> 92   0.82067094  0.31469105  0.13931170
#> 93  -0.51036340 -1.37851846 -0.73688620
#> 94   0.27902196 -0.37286163 -0.19396872
#> 95  -0.49688915  0.13005188  1.09716108
#> 96  -0.72624599  0.05733712  0.55764312
#> 97   0.21954787 -1.27163957 -1.24672707
#> 98   0.33686469 -1.01974673 -0.42929297
#> 99  -0.40749794 -1.46761372 -1.66172537
#> 100  1.76658887 -0.29851737 -2.02478595