Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, empirical = FALSE, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id            y grand_i       sub_i     item_i          err
#> 1      S01     I01 -0.486960429       0  1.26043433 -0.1743304 -1.573064322
#> 2      S02     I01 -0.755323930       0  0.47861769 -0.1743304 -1.059611187
#> 3      S03     I01  0.091796759       0 -0.04319499 -0.1743304  0.309322188
#> 4      S04     I01 -0.971084695       0 -0.91687616 -0.1743304  0.120121904
#> 5      S05     I01  1.418896344       0  1.89365613 -0.1743304 -0.300429353
#> 6      S06     I01  0.672233818       0  0.32862856 -0.1743304  0.517935694
#> 7      S07     I01  1.199688643       0  1.13404574 -0.1743304  0.239973340
#> 8      S08     I01 -1.271952450       0  0.14624540 -0.1743304 -1.243867411
#> 9      S09     I01 -2.430910011       0 -1.05403017 -0.1743304 -1.202549405
#> 10     S10     I01 -1.111870400       0 -0.27331631 -0.1743304 -0.664223657
#> 11     S01     I02  3.367106167       0  1.26043433  1.4649950  0.641676832
#> 12     S02     I02  2.067732527       0  0.47861769  1.4649950  0.124119827
#> 13     S03     I02  0.684564265       0 -0.04319499  1.4649950 -0.737235749
#> 14     S04     I02  1.485871467       0 -0.91687616  1.4649950  0.937752623
#> 15     S05     I02  4.164461855       0  1.89365613  1.4649950  0.805810715
#> 16     S06     I02  1.624446421       0  0.32862856  1.4649950 -0.169177146
#> 17     S07     I02  1.873884316       0  1.13404574  1.4649950 -0.725156430
#> 18     S08     I02  3.133357492       0  0.14624540  1.4649950  1.522117087
#> 19     S09     I02 -2.008632772       0 -1.05403017  1.4649950 -2.419597609
#> 20     S10     I02  0.321908226       0 -0.27331631  1.4649950 -0.869770475
#> 21     S01     I03  2.454436635       0  1.26043433  1.2016820 -0.007679663
#> 22     S02     I03  1.255051711       0  0.47861769  1.2016820 -0.425247952
#> 23     S03     I03  1.586651547       0 -0.04319499  1.2016820  0.428164570
#> 24     S04     I03 -0.532459166       0 -0.91687616  1.2016820 -0.817264972
#> 25     S05     I03  3.143718380       0  1.89365613  1.2016820  0.048380278
#> 26     S06     I03  1.037604713       0  0.32862856  1.2016820 -0.492705817
#> 27     S07     I03  2.478911560       0  1.13404574  1.2016820  0.143183852
#> 28     S08     I03  0.728975242       0  0.14624540  1.2016820 -0.618952125
#> 29     S09     I03  1.563488998       0 -1.05403017  1.2016820  1.415837198
#> 30     S10     I03  0.898886916       0 -0.27331631  1.2016820 -0.029478747
#> 31     S01     I04  3.593194672       0  1.26043433  2.0610431  0.271717262
#> 32     S02     I04  3.698645389       0  0.47861769  2.0610431  1.158984614
#> 33     S03     I04  1.017380291       0 -0.04319499  2.0610431 -1.000467798
#> 34     S04     I04  1.079892005       0 -0.91687616  2.0610431 -0.064274914
#> 35     S05     I04  4.376270994       0  1.89365613  2.0610431  0.421571780
#> 36     S06     I04  3.797304502       0  0.32862856  2.0610431  1.407632860
#> 37     S07     I04  3.634385456       0  1.13404574  2.0610431  0.439296636
#> 38     S08     I04  1.193472749       0  0.14624540  2.0610431 -1.013815731
#> 39     S09     I04  0.880609165       0 -1.05403017  2.0610431 -0.126403747
#> 40     S10     I04  2.459518031       0 -0.27331631  2.0610431  0.671791256
#> 41     S01     I05  1.949723128       0  1.26043433  0.7757138 -0.086425043
#> 42     S02     I05  3.352133973       0  0.47861769  0.7757138  2.097802437
#> 43     S03     I05 -0.701248362       0 -0.04319499  0.7757138 -1.433767211
#> 44     S04     I05  0.075580142       0 -0.91687616  0.7757138  0.216742463
#> 45     S05     I05  2.632518839       0  1.89365613  0.7757138 -0.036851137
#> 46     S06     I05  1.990448800       0  0.32862856  0.7757138  0.886106398
#> 47     S07     I05  1.525681560       0  1.13404574  0.7757138 -0.384078021
#> 48     S08     I05  0.807192297       0  0.14624540  0.7757138 -0.114766943
#> 49     S09     I05 -1.355200065       0 -1.05403017  0.7757138 -1.076883737
#> 50     S10     I05 -0.763482333       0 -0.27331631  0.7757138 -1.265879869
#> 51     S01     I06  0.580270677       0  1.26043433 -0.8279818  0.147818099
#> 52     S02     I06 -2.854066882       0  0.47861769 -0.8279818 -2.504702825
#> 53     S03     I06  0.937576991       0 -0.04319499 -0.8279818  1.808753734
#> 54     S04     I06 -1.123462029       0 -0.91687616 -0.8279818  0.621395884
#> 55     S05     I06 -0.006185916       0  1.89365613 -0.8279818 -1.071860299
#> 56     S06     I06 -0.814132606       0  0.32862856 -0.8279818 -0.314779417
#> 57     S07     I06 -1.252455569       0  1.13404574 -0.8279818 -1.558519558
#> 58     S08     I06 -0.982446443       0  0.14624540 -0.8279818 -0.300710091
#> 59     S09     I06  0.386994209       0 -1.05403017 -0.8279818  2.269006130
#> 60     S10     I06 -0.816994954       0 -0.27331631 -0.8279818  0.284303103
#> 61     S01     I07  3.278022811       0  1.26043433  1.4056592  0.611929320
#> 62     S02     I07  2.403788481       0  0.47861769  1.4056592  0.519511625
#> 63     S03     I07  3.266849727       0 -0.04319499  1.4056592  1.904385557
#> 64     S04     I07  2.209267788       0 -0.91687616  1.4056592  1.720484788
#> 65     S05     I07  0.751990991       0  1.89365613  1.4056592 -2.547324305
#> 66     S06     I07  1.751184889       0  0.32862856  1.4056592  0.016897165
#> 67     S07     I07  3.355650240       0  1.13404574  1.4056592  0.815945338
#> 68     S08     I07  1.177491524       0  0.14624540  1.4056592 -0.374413037
#> 69     S09     I07 -1.445833339       0 -1.05403017  1.4056592 -1.797462332
#> 70     S10     I07  0.230943818       0 -0.27331631  1.4056592 -0.901399038
#> 71     S01     I08  3.106920158       0  1.26043433  0.8984211  0.948064710
#> 72     S02     I08 -0.266990487       0  0.47861769  0.8984211 -1.644029299
#> 73     S03     I08  0.335782357       0 -0.04319499  0.8984211 -0.519443769
#> 74     S04     I08 -0.267450680       0 -0.91687616  0.8984211 -0.248995635
#> 75     S05     I08  3.655098395       0  1.89365613  0.8984211  0.863021143
#> 76     S06     I08  1.961469117       0  0.32862856  0.8984211  0.734419438
#> 77     S07     I08  2.178575966       0  1.13404574  0.8984211  0.146109108
#> 78     S08     I08  0.550922957       0  0.14624540  0.8984211 -0.493743560
#> 79     S09     I08  0.285263322       0 -1.05403017  0.8984211  0.440872373
#> 80     S10     I08 -0.383713538       0 -0.27331631  0.8984211 -1.008818351
#> 81     S01     I09  1.921687079       0  1.26043433  0.8919633 -0.230710578
#> 82     S02     I09  1.197902060       0  0.47861769  0.8919633 -0.172678962
#> 83     S03     I09  1.661281795       0 -0.04319499  0.8919633  0.812513459
#> 84     S04     I09 -0.301954073       0 -0.91687616  0.8919633 -0.277041239
#> 85     S05     I09  1.365893116       0  1.89365613  0.8919633 -1.419726345
#> 86     S06     I09  2.717136928       0  0.32862856  0.8919633  1.496545038
#> 87     S07     I09  0.908768848       0  1.13404574  0.8919633 -1.117240219
#> 88     S08     I09  2.087480933       0  0.14624540  0.8919633  1.049272207
#> 89     S09     I09 -1.447587650       0 -1.05403017  0.8919633 -1.285520808
#> 90     S10     I09  0.811430129       0 -0.27331631  0.8919633  0.192783107
#> 91     S01     I10  2.496722887       0  1.26043433  1.0863042  0.149984342
#> 92     S02     I10  2.394754036       0  0.47861769  1.0863042  0.829832126
#> 93     S03     I10  1.607485960       0 -0.04319499  1.0863042  0.564376736
#> 94     S04     I10  0.958092391       0 -0.91687616  1.0863042  0.788664337
#> 95     S05     I10  2.649726743       0  1.89365613  1.0863042 -0.330233607
#> 96     S06     I10 -1.062855754       0  0.32862856  1.0863042 -2.477788531
#> 97     S07     I10  0.425212593       0  1.13404574  1.0863042 -1.795137363
#> 98     S08     I10  2.814869366       0  0.14624540  1.0863042  1.582319751
#> 99     S09     I10 -1.476451304       0 -1.05403017  1.0863042 -1.508725351
#> 100    S10     I10 -1.036033955       0 -0.27331631  1.0863042 -1.849021866