Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id",
  seed = NULL
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id           y  grand_i       sub_i      item_i          err
#> 1      S01     I01  2.84927495 3.018906 -0.24237243  0.87811937 -0.805377606
#> 15     S01     I02  2.27264342 3.018906 -0.24237243 -0.09537731 -0.408512456
#> 29     S01     I03  3.00039445 3.018906 -0.24237243  1.02064682 -0.796785557
#> 31     S01     I04  0.73550831 3.018906 -0.24237243 -0.76745916 -1.273565707
#> 44     S01     I05  2.37353363 3.018906 -0.24237243 -0.22482449 -0.178175056
#> 58     S01     I06  2.29694725 3.018906 -0.24237243  0.07006955 -0.549655484
#> 61     S01     I07  5.19894186 3.018906 -0.24237243  0.83210178  1.590306892
#> 73     S01     I08  3.62706720 3.018906 -0.24237243 -1.28263587  2.133169880
#> 87     S01     I09  2.24586876 3.018906 -0.24237243  0.63034248 -1.161006905
#> 91     S01     I10  1.47233738 3.018906 -0.24237243  0.19122648 -1.495422285
#> 7      S02     I01  2.92973042 3.018906 -0.21993591  0.87811937 -0.747358654
#> 17     S02     I02  2.70857313 3.018906 -0.21993591 -0.09537731  0.004980742
#> 30     S02     I03  4.76580295 3.018906 -0.21993591  1.02064682  0.946186431
#> 36     S02     I04  2.38466934 3.018906 -0.21993591 -0.76745916  0.353158801
#> 46     S02     I05  5.02686829 3.018906 -0.21993591 -0.22482449  2.452723089
#> 59     S02     I06  1.92937171 3.018906 -0.21993591  0.07006955 -0.939667536
#> 69     S02     I07  4.07261606 3.018906 -0.21993591  0.83210178  0.441544579
#> 75     S02     I08  1.56475783 3.018906 -0.21993591 -1.28263587  0.048424001
#> 88     S02     I09  3.15392662 3.018906 -0.21993591  0.63034248 -0.275385565
#> 98     S02     I10  2.02194971 3.018906 -0.21993591  0.19122648 -0.968246475
#> 4      S03     I01  6.68730964 3.018906  1.09714815  0.87811937  1.693136500
#> 18     S03     I02  4.24446670 3.018906  1.09714815 -0.09537731  0.223790244
#> 23     S03     I03  3.57545010 3.018906  1.09714815  1.02064682 -1.561250479
#> 33     S03     I04  3.38285359 3.018906  1.09714815 -0.76745916  0.034258989
#> 47     S03     I05  5.72087837 3.018906  1.09714815 -0.22482449  1.829649101
#> 60     S03     I06  4.05066014 3.018906  1.09714815  0.07006955 -0.135463167
#> 62     S03     I07  4.15806528 3.018906  1.09714815  0.83210178 -0.790090260
#> 76     S03     I08  2.88557739 3.018906  1.09714815 -1.28263587  0.052159494
#> 89     S03     I09  4.95289509 3.018906  1.09714815  0.63034248  0.206498850
#> 95     S03     I10  5.36457138 3.018906  1.09714815  0.19122648  1.057291133
#> 5      S04     I01  5.10165934 3.018906  0.99500707  0.87811937  0.209627283
#> 19     S04     I02  4.16150837 3.018906  0.99500707 -0.09537731  0.242972991
#> 21     S04     I03  3.80108897 3.018906  0.99500707  1.02064682 -1.233470533
#> 34     S04     I04  3.58302648 3.018906  0.99500707 -0.76745916  0.336572954
#> 48     S04     I05  1.13038943 3.018906  0.99500707 -0.22482449 -2.658698761
#> 51     S04     I06  4.84755732 3.018906  0.99500707  0.07006955  0.763575084
#> 63     S04     I07  5.79649524 3.018906  0.99500707  0.83210178  0.950480767
#> 77     S04     I08  2.64758863 3.018906  0.99500707 -1.28263587 -0.083688184
#> 90     S04     I09  5.78114908 3.018906  0.99500707  0.63034248  1.136893909
#> 92     S04     I10  4.12535166 3.018906  0.99500707  0.19122648 -0.079787510
#> 2      S05     I01  2.69297880 3.018906 -0.81834910  0.87811937 -0.385697089
#> 16     S05     I02  2.44711247 3.018906 -0.81834910 -0.09537731  0.341933273
#> 22     S05     I03  5.63568956 3.018906 -0.81834910  1.02064682  2.414486231
#> 35     S05     I04  3.07278905 3.018906 -0.81834910 -0.76745916  1.639691702
#> 45     S05     I05  1.33095877 3.018906 -0.81834910 -0.22482449 -0.644773251
#> 52     S05     I06  1.81596298 3.018906 -0.81834910  0.07006955 -0.454663079
#> 64     S05     I07  3.38563853 3.018906 -0.81834910  0.83210178  0.352980236
#> 74     S05     I08 -0.42939058 3.018906 -0.81834910 -1.28263587 -1.347311223
#> 81     S05     I09  3.38839076 3.018906 -0.81834910  0.63034248  0.557491768
#> 93     S05     I10  1.34217954 3.018906 -0.81834910  0.19122648 -1.049603451
#> 8      S06     I01  5.68526919 3.018906  0.06923933  0.87811937  1.719004874
#> 20     S06     I02  2.74742228 3.018906  0.06923933 -0.09537731 -0.245345353
#> 27     S06     I03  3.16941636 3.018906  0.06923933  1.02064682 -0.939375398
#> 37     S06     I04  0.74712574 3.018906  0.06923933 -0.76745916 -1.573560033
#> 49     S06     I05  4.23143396 3.018906  0.06923933 -0.22482449  1.368113516
#> 56     S06     I06  3.29527125 3.018906  0.06923933  0.07006955  0.137056763
#> 66     S06     I07  5.43114409 3.018906  0.06923933  0.83210178  1.510897364
#> 79     S06     I08  0.58739644 3.018906  0.06923933 -1.28263587 -1.218112631
#> 85     S06     I09  4.38380906 3.018906  0.06923933  0.63034248  0.665321636
#> 99     S06     I10  2.64096694 3.018906  0.06923933  0.19122648 -0.638404480
#> 9      S07     I01  5.64660299 3.018906  0.86961174  0.87811937  0.879966270
#> 11     S07     I02  1.24562424 3.018906  0.86961174 -0.09537731 -2.547515802
#> 24     S07     I03  5.56194323 3.018906  0.86961174  1.02064682  0.652779061
#> 38     S07     I04  3.09629892 3.018906  0.86961174 -0.76745916 -0.024759260
#> 50     S07     I05  3.64679115 3.018906  0.86961174 -0.22482449 -0.016901703
#> 53     S07     I06  2.53401812 3.018906  0.86961174  0.07006955 -1.424568775
#> 67     S07     I07  3.49921738 3.018906  0.86961174  0.83210178 -1.221401753
#> 80     S07     I08  3.26518876 3.018906  0.86961174 -1.28263587  0.659307280
#> 82     S07     I09  4.49229596 3.018906  0.86961174  0.63034248 -0.026563869
#> 96     S07     I10  4.61920492 3.018906  0.86961174  0.19122648  0.539461088
#> 6      S08     I01  5.64159619 3.018906  0.96873657  0.87811937  0.775834639
#> 12     S08     I02  3.76874898 3.018906  0.96873657 -0.09537731 -0.123515888
#> 25     S08     I03  3.45720175 3.018906  0.96873657  1.02064682 -1.551087249
#> 39     S08     I04  2.96620836 3.018906  0.96873657 -0.76745916 -0.253974655
#> 41     S08     I05  4.94138503 3.018906  0.96873657 -0.22482449  1.178567347
#> 54     S08     I06  3.77775513 3.018906  0.96873657  0.07006955 -0.279956593
#> 68     S08     I07  5.69891525 3.018906  0.96873657  0.83210178  0.879171284
#> 78     S08     I08 -0.07078241 3.018906  0.96873657 -1.28263587 -2.775788723
#> 83     S08     I09  5.64364908 3.018906  0.96873657  0.63034248  1.025664420
#> 97     S08     I10  2.66803813 3.018906  0.96873657  0.19122648 -1.510830529
#> 3      S09     I01  3.23292153 3.018906 -0.32018716  0.87811937 -0.343916299
#> 13     S09     I02  1.28029510 3.018906 -0.32018716 -0.09537731 -1.323046046
#> 26     S09     I03  2.67453600 3.018906 -0.32018716  1.02064682 -1.044829274
#> 32     S09     I04  1.14876120 3.018906 -0.32018716 -0.76745916 -0.782498083
#> 42     S09     I05  3.00671470 3.018906 -0.32018716 -0.22482449  0.532820746
#> 55     S09     I06  2.51035087 3.018906 -0.32018716  0.07006955 -0.258437124
#> 65     S09     I07  3.67178109 3.018906 -0.32018716  0.83210178  0.140960863
#> 71     S09     I08  0.50114976 3.018906 -0.32018716 -1.28263587 -0.914932828
#> 84     S09     I09  4.50272761 3.018906 -0.32018716  0.63034248  1.173666681
#> 94     S09     I10  2.77324598 3.018906 -0.32018716  0.19122648 -0.116698950
#> 10     S10     I01  4.50256744 3.018906  0.83909059  0.87811937 -0.233548137
#> 14     S10     I02  4.12529342 3.018906  0.83909059 -0.09537731  0.362674527
#> 28     S10     I03  7.13125822 3.018906  0.83909059  1.02064682  2.252615199
#> 40     S10     I04  1.05242420 3.018906  0.83909059 -0.76745916 -2.038112837
#> 43     S10     I05  3.96615683 3.018906  0.83909059 -0.22482449  0.332985120
#> 57     S10     I06  4.25937073 3.018906  0.83909059  0.07006955  0.331304978
#> 70     S10     I07  4.45766161 3.018906  0.83909059  0.83210178 -0.232436378
#> 72     S10     I08  2.36703117 3.018906  0.83909059 -1.28263587 -0.208329164
#> 86     S10     I09  5.73262385 3.018906  0.83909059  0.63034248  1.244285162
#> 100    S10     I10  2.64470515 3.018906  0.83909059  0.19122648 -1.404517533