Generate a sample with random intercepts for subjects and items

simdf_mixed Produces a dataframe with the same distributions of by-subject and by-item random intercepts as an existing dataframe

simdf_mixed(dat, sub_n = 100, item_n = 25, dv = 1, sub_id = 2,
  item_id = 3)

Arguments

dat
the existing dataframe

sub_n
the number of subjects to simulate

item_n
the number of items to simulate

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

tibble

Examples

simdf_mixed(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id       sub_i     item_i        dv
#> 1        1       1  0.15336871  0.1152507 3.4263347
#> 2        2       1 -0.07546968  0.1152507 2.6552210
#> 3        3       1 -0.09464565  0.1152507 2.5473771
#> 4        4       1  0.13680669  0.1152507 4.5736075
#> 5        5       1  0.46710613  0.1152507 4.6470300
#> 6        6       1  0.71713231  0.1152507 5.0470723
#> 7        7       1 -0.16377959  0.1152507 3.0286129
#> 8        8       1 -0.03648029  0.1152507 3.0203508
#> 9        9       1 -0.44490238  0.1152507 2.4007496
#> 10      10       1 -0.50294582  0.1152507 1.4482325
#> 11       1       2  0.15336871  0.9694343 4.2646276
#> 12       2       2 -0.07546968  0.9694343 3.5117228
#> 13       3       2 -0.09464565  0.9694343 1.9205807
#> 14       4       2  0.13680669  0.9694343 3.3332175
#> 15       5       2  0.46710613  0.9694343 4.1037056
#> 16       6       2  0.71713231  0.9694343 4.8340290
#> 17       7       2 -0.16377959  0.9694343 4.3931885
#> 18       8       2 -0.03648029  0.9694343 3.5224828
#> 19       9       2 -0.44490238  0.9694343 2.7403387
#> 20      10       2 -0.50294582  0.9694343 4.9209943
#> 21       1       3  0.15336871 -0.1809222 2.3955324
#> 22       2       3 -0.07546968 -0.1809222 5.0649247
#> 23       3       3 -0.09464565 -0.1809222 3.0032075
#> 24       4       3  0.13680669 -0.1809222 2.4135067
#> 25       5       3  0.46710613 -0.1809222 4.5656069
#> 26       6       3  0.71713231 -0.1809222 2.8650467
#> 27       7       3 -0.16377959 -0.1809222 3.6126806
#> 28       8       3 -0.03648029 -0.1809222 0.9722492
#> 29       9       3 -0.44490238 -0.1809222 0.6125408
#> 30      10       3 -0.50294582 -0.1809222 3.8909402
#> 31       1       4  0.15336871  1.0979794 5.9448229
#> 32       2       4 -0.07546968  1.0979794 3.9574235
#> 33       3       4 -0.09464565  1.0979794 4.8079004
#> 34       4       4  0.13680669  1.0979794 5.0826263
#> 35       5       4  0.46710613  1.0979794 3.3817108
#> 36       6       4  0.71713231  1.0979794 3.2445244
#> 37       7       4 -0.16377959  1.0979794 4.2746244
#> 38       8       4 -0.03648029  1.0979794 4.7967650
#> 39       9       4 -0.44490238  1.0979794 4.3737837
#> 40      10       4 -0.50294582  1.0979794 4.2923638
#> 41       1       5  0.15336871  0.2255086 5.9170173
#> 42       2       5 -0.07546968  0.2255086 2.7538325
#> 43       3       5 -0.09464565  0.2255086 3.1842397
#> 44       4       5  0.13680669  0.2255086 4.2260520
#> 45       5       5  0.46710613  0.2255086 1.5543152
#> 46       6       5  0.71713231  0.2255086 3.9556215
#> 47       7       5 -0.16377959  0.2255086 3.8447343
#> 48       8       5 -0.03648029  0.2255086 4.5385200
#> 49       9       5 -0.44490238  0.2255086 2.6456965
#> 50      10       5 -0.50294582  0.2255086 2.2025837
#> 51       1       6  0.15336871 -0.5688417 3.5177725
#> 52       2       6 -0.07546968 -0.5688417 2.9838297
#> 53       3       6 -0.09464565 -0.5688417 3.5923231
#> 54       4       6  0.13680669 -0.5688417 3.1022826
#> 55       5       6  0.46710613 -0.5688417 2.7535176
#> 56       6       6  0.71713231 -0.5688417 5.4940446
#> 57       7       6 -0.16377959 -0.5688417 2.5234260
#> 58       8       6 -0.03648029 -0.5688417 0.3511046
#> 59       9       6 -0.44490238 -0.5688417 1.8498874
#> 60      10       6 -0.50294582 -0.5688417 3.5379786
#> 61       1       7  0.15336871 -0.1003512 4.2409599
#> 62       2       7 -0.07546968 -0.1003512 3.5556705
#> 63       3       7 -0.09464565 -0.1003512 2.6899611
#> 64       4       7  0.13680669 -0.1003512 2.0439883
#> 65       5       7  0.46710613 -0.1003512 5.0354057
#> 66       6       7  0.71713231 -0.1003512 3.5521496
#> 67       7       7 -0.16377959 -0.1003512 2.3525066
#> 68       8       7 -0.03648029 -0.1003512 3.5415480
#> 69       9       7 -0.44490238 -0.1003512 4.1344729
#> 70      10       7 -0.50294582 -0.1003512 2.3269392
#> 71       1       8  0.15336871  1.6989795 6.4250556
#> 72       2       8 -0.07546968  1.6989795 5.7450023
#> 73       3       8 -0.09464565  1.6989795 3.8872425
#> 74       4       8  0.13680669  1.6989795 4.5973502
#> 75       5       8  0.46710613  1.6989795 5.7349937
#> 76       6       8  0.71713231  1.6989795 2.6852695
#> 77       7       8 -0.16377959  1.6989795 3.7919020
#> 78       8       8 -0.03648029  1.6989795 4.7810577
#> 79       9       8 -0.44490238  1.6989795 5.1999969
#> 80      10       8 -0.50294582  1.6989795 5.4375538
#> 81       1       9  0.15336871  1.2452224 5.3180720
#> 82       2       9 -0.07546968  1.2452224 2.7003936
#> 83       3       9 -0.09464565  1.2452224 4.3759073
#> 84       4       9  0.13680669  1.2452224 6.5338902
#> 85       5       9  0.46710613  1.2452224 3.8544295
#> 86       6       9  0.71713231  1.2452224 5.2503336
#> 87       7       9 -0.16377959  1.2452224 4.6527757
#> 88       8       9 -0.03648029  1.2452224 3.6883031
#> 89       9       9 -0.44490238  1.2452224 2.8378979
#> 90      10       9 -0.50294582  1.2452224 4.8626384
#> 91       1      10  0.15336871 -0.2064970 4.7460439
#> 92       2      10 -0.07546968 -0.2064970 1.8389187
#> 93       3      10 -0.09464565 -0.2064970 1.3021962
#> 94       4      10  0.13680669 -0.2064970 2.4226626
#> 95       5      10  0.46710613 -0.2064970 3.0711292
#> 96       6      10  0.71713231 -0.2064970 2.2198113
#> 97       7      10 -0.16377959 -0.2064970 0.9575364
#> 98       8      10 -0.03648029 -0.2064970 2.6135275
#> 99       9      10 -0.44490238 -0.2064970 1.2924662
#> 100     10      10 -0.50294582 -0.2064970 0.8947152