Check readline input

Check readline input

Usage

readline_check(
  prompt,
  type = c("numeric", "integer", "length", "grep"),
  min = -Inf,
  max = Inf,
  warning = NULL,
  default = NULL,
  ...
)

Arguments

prompt

the prompt for readline

type

what type of check to perform, one of c("numeric", "integer", "length", "grep")

min

the minimum value

max

the maximum value

warning

an optional custom warning message

default

the default option to return if the entry is blank, NULL allows no default, the default value will be displayed after the text as [default]

...

other arguments to pass to grep

Value

the validated result of readline

Examples

if(interactive()){
readline_check("Type a number: ", "numeric")
readline_check("Type two characters: ", "length", min = 2, max = 2)
readline_check("Type at least 3 characters: ", "length", min = 3)
readline_check("Type no more than 4 characters: ", "length", max = 44)
readline_check("Type a letter and a number: ", "grep", pattern = "^[a-zA-Z]\\d$")
}