Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A            B           C
#> 1    1.652690994 -0.178257413 -0.27544728
#> 2   -1.900543600  0.629665511 -1.87975357
#> 3    0.022026109  1.264683614  0.80077479
#> 4    0.253634225  3.640146732  0.41445237
#> 5    1.185744039  1.991582281  1.83051145
#> 6   -0.894602013 -0.895552197 -1.18194471
#> 7   -0.709097620  0.350814701  0.92698881
#> 8    0.103384768  0.105026551  1.40216836
#> 9   -0.835484482  0.635506367  1.28748485
#> 10   1.351304700  1.378582677  0.31158301
#> 11  -0.373176553  0.287290898 -0.03605547
#> 12  -2.559227734 -1.048081773 -2.01595679
#> 13  -1.157808602 -0.085049585 -0.67131025
#> 14   0.974275268 -1.441963122 -1.28850882
#> 15  -0.195669096 -0.409574985 -1.64964266
#> 16  -0.088370388 -1.144260156  0.02991775
#> 17   0.184652696  0.992770558  0.79921211
#> 18  -0.141850507  0.692186440 -0.49640160
#> 19   1.577074684  0.272892821  0.20098591
#> 20   0.434344741  0.827178001 -0.07881313
#> 21  -0.049116151  0.701035564 -0.39604434
#> 22  -0.439913977 -1.727787411  0.88938956
#> 23   0.080560084  1.496035155 -0.76115751
#> 24  -1.246490922  0.168913759 -1.24166637
#> 25  -0.427782269  0.302543376  0.09384256
#> 26  -0.269703106 -1.147824196 -0.81144753
#> 27   0.869984421 -0.923907827 -1.50472861
#> 28   0.545245894 -0.117875395  0.18314567
#> 29   0.335896438  3.111189128  1.14003568
#> 30   0.774407375  1.369696969  0.34991968
#> 31   1.987940583  1.837630427  1.62629824
#> 32   0.390917300  0.313624993  0.59665145
#> 33   0.291209885 -0.569971335  0.34199032
#> 34  -0.447249747 -0.598059365  0.32991794
#> 35   1.166144791 -1.528340251 -0.43806854
#> 36  -1.360912512 -1.444676602 -2.44483820
#> 37   0.586997572 -0.466615224  0.43820616
#> 38  -0.130987977  1.242162951  0.99304768
#> 39  -0.234533961 -0.995707684  0.48635094
#> 40   0.377620915  1.547457701 -1.17542953
#> 41  -0.776083959  0.110551637 -0.81964316
#> 42   0.868300539 -0.530577516  0.44493757
#> 43  -0.806905614 -0.076504596 -0.40320983
#> 44  -0.828032454  0.005171647 -0.78462415
#> 45   0.206932519 -0.425407357 -0.47344067
#> 46   1.308028615  0.584005149  1.30170185
#> 47  -0.473659195 -0.275943180 -0.49235564
#> 48   0.454162295 -1.447520542  0.19407478
#> 49   0.035471469 -0.420815189 -0.22961651
#> 50   0.371242603  0.556634532  0.10113023
#> 51   1.221293653 -0.933737095  1.99701648
#> 52  -0.771557403 -0.438217322 -0.70563020
#> 53   2.306781200 -0.123602517  0.22167733
#> 54   1.458932891  0.660295521  0.52279522
#> 55   0.237351789 -0.761025361 -0.36966289
#> 56  -0.359579359  0.483176011  1.23314986
#> 57   0.404714025  0.882952017 -0.23241287
#> 58   0.989334579 -0.921227318 -0.01860750
#> 59  -1.708788543 -0.328094939 -2.37506013
#> 60  -0.348664618 -0.446396904 -1.67782915
#> 61   2.695134980  0.654825798  1.64541599
#> 62  -0.802458768 -1.407720927  0.56881728
#> 63   0.595002761  0.297642550 -0.15218920
#> 64  -0.044404967 -0.184028167  0.01340438
#> 65   0.366986982  1.176098830 -0.15037147
#> 66  -1.557141203 -0.520450366 -0.49855595
#> 67   0.071948249 -0.121616067 -0.50615507
#> 68  -1.554764625  1.047848713 -0.68884261
#> 69  -0.002244005  0.336323051 -0.55769203
#> 70  -0.415804042  1.574672537  0.18351521
#> 71  -1.186825508  0.048593095 -0.66060575
#> 72   1.182820634  0.970818348  0.59693108
#> 73   0.676380327  0.466212132  0.62568842
#> 74   0.307385936 -0.197610924 -0.10512140
#> 75  -0.312676064  0.812172590  0.74689672
#> 76  -0.603366139 -0.764754840  0.21520328
#> 77   0.150529978  0.669195349  1.06900203
#> 78   0.377035917 -0.110681695 -1.17714423
#> 79   0.219623318 -1.606552328 -1.58962225
#> 80   1.511150262  0.749647326 -0.70887450
#> 81  -0.384079195  0.078935070 -1.52855322
#> 82  -1.066346839  0.125510506 -0.60329244
#> 83   0.386819949 -0.267366331  0.01183787
#> 84   0.021060709 -0.248528660 -1.29935437
#> 85  -0.485381272  0.522076491  0.31556282
#> 86  -0.456410388  0.644095983  1.03401501
#> 87   0.116200806 -0.160944535 -0.59020209
#> 88   1.970868584  0.204887203  0.47086458
#> 89   2.417114203  0.619595530  0.03101228
#> 90  -0.585473558 -2.483530279 -1.75215560
#> 91   0.027519401 -0.502430458  0.48854165
#> 92   2.366020372 -0.454171511  1.36663288
#> 93  -0.329932079 -0.542263067 -1.04213835
#> 94   0.809401905  2.149104611  0.90581093
#> 95  -0.612666692 -1.815514536 -0.18094205
#> 96  -0.631084253 -0.771121727  0.09206091
#> 97  -0.445288620  0.559740867  1.38242491
#> 98  -0.493876132 -0.904594550 -1.00729879
#> 99   1.264924032  1.651229942  1.60991993
#> 100 -0.832496718 -1.000231298 -1.13451476

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A            B            C
#> 1   -1.84804746  0.006947376  2.284650380
#> 2   -0.34700468 -1.070916895  0.056123348
#> 3    0.34988861  1.198080297  0.427704778
#> 4   -1.57536609 -1.670117586  0.214614782
#> 5    0.65388905  0.265733789  0.220262824
#> 6    2.46641237  0.360968447 -1.303210097
#> 7   -0.39766849  1.078217885  0.854378614
#> 8   -0.99080459 -0.138548120  0.384915136
#> 9    0.39397406  0.164869350 -0.208070167
#> 10  -0.55823565 -1.300310636 -0.840870293
#> 11  -0.05443303 -1.796274325 -0.274340635
#> 12  -0.02555695 -0.421345974 -0.323605963
#> 13   1.99713550 -0.762907591 -1.519105463
#> 14  -1.34969788  0.008509153 -0.144497331
#> 15   0.80225617  0.281191953  0.198620539
#> 16   1.07536288  1.123030789 -0.352260080
#> 17  -0.63552905  1.238326716  0.874914430
#> 18   0.60743489 -0.868623125 -0.168290378
#> 19  -0.80701176  1.626204538  1.798883384
#> 20  -1.65512251 -2.358812620 -0.694246595
#> 21  -0.11851547  1.273814510  1.595421151
#> 22   1.09460325 -0.531787782 -0.634340972
#> 23   0.18659122 -0.072611702 -1.287652804
#> 24   0.51028912 -2.767549561 -1.684556555
#> 25  -0.04178000 -1.009084452  0.100767663
#> 26  -1.38811657  0.677706479  1.478104522
#> 27  -1.65402320 -1.632016183 -0.123185350
#> 28   1.09416661 -0.987680831 -1.633369367
#> 29   0.58439301 -0.213072421 -0.116157081
#> 30  -0.59537660 -1.902664961 -1.306585074
#> 31   0.01973322  1.025514936 -0.600092425
#> 32   1.58758635  2.026950752  0.555000914
#> 33  -1.23614469  1.450625904  2.529773971
#> 34  -0.83032173 -1.194399858 -0.639107751
#> 35   0.55930500 -1.410294060 -0.944603595
#> 36  -0.85688464 -0.082590869 -0.297468119
#> 37   1.27356070 -0.280454117 -1.504556555
#> 38   0.41307894  0.918244600 -0.342888147
#> 39   0.15559705  0.752504448 -0.090875568
#> 40  -1.15158233 -1.712100370  0.225897908
#> 41   1.14519682  1.263344753  0.023613603
#> 42   0.87790335 -0.393299670 -0.133954133
#> 43   1.23120272 -0.738946329 -0.942357213
#> 44   1.14726139 -1.848264183 -2.021339524
#> 45  -0.89036827  0.474428720  1.477891252
#> 46  -0.83230284 -1.498736178 -1.054390691
#> 47  -0.35294385  1.413879069  1.147132207
#> 48   0.72674758  0.101268249 -0.932601607
#> 49   1.25034592  0.616103269  0.158068085
#> 50  -3.45417580 -0.900517154  1.504888280
#> 51   1.37895222  0.652702134 -0.669274332
#> 52   0.50350416 -0.790123732 -2.037559493
#> 53   0.30794593 -0.152654964 -0.384531157
#> 54  -0.66500760  0.932323383  1.563495293
#> 55   0.62773101  0.157286296 -0.942960328
#> 56  -0.85908202 -0.172951520  0.382999357
#> 57  -0.91884392  0.099149381  0.242096954
#> 58  -0.27883398  1.000834293  0.065912441
#> 59  -1.20649859 -0.970220033 -1.068425794
#> 60   1.56966894  0.286405971 -1.661335738
#> 61   1.10042229  0.441307285 -0.629890275
#> 62  -0.54146283 -0.721525905  1.195815129
#> 63  -1.36049923  1.623013356  1.409577466
#> 64   2.49274785  0.484261131 -0.783381794
#> 65   0.74270064 -0.233022660  0.629002830
#> 66   0.93399908  0.895893622 -0.639484552
#> 67   0.94438060 -0.657365030 -2.072843686
#> 68   0.16775162  0.764011537  0.909495997
#> 69   1.21340811 -1.054244614 -1.804406381
#> 70  -1.28216774 -0.556605185 -0.330193419
#> 71  -0.56194603 -2.548048799 -1.152351095
#> 72  -0.34367196 -0.985022376 -0.358115681
#> 73  -1.53419850 -1.041210384  0.738831846
#> 74   1.21050654  2.879012887  0.495766064
#> 75  -0.01370933  1.145212955  1.461887847
#> 76  -1.22174576  0.433557163  0.536541674
#> 77   0.12714324 -0.490343052 -0.795862279
#> 78   1.16512547  1.998840502  0.413446928
#> 79  -0.18819751  0.046822633  1.492296466
#> 80   0.20074680  1.171561970  1.452872702
#> 81  -2.16153004 -0.936216129  0.431614160
#> 82   2.06416121 -1.320782378 -3.020562521
#> 83  -0.91659881 -0.261904074  0.284146057
#> 84  -0.28754096 -1.255126216 -1.298030223
#> 85  -0.52924822  2.387529679  1.917932752
#> 86  -0.02855605 -0.882050632 -2.038492501
#> 87   1.59788495  0.350808919 -0.345252398
#> 88  -0.22175430  0.237217036  0.124798909
#> 89   0.24422292  0.388975508 -0.311715317
#> 90   0.76144508  0.678549405 -0.003141895
#> 91   0.38132129  2.188172750 -0.057631269
#> 92  -0.98495664 -0.253263297 -0.041918567
#> 93  -0.87312526 -2.522544545 -2.698158598
#> 94   0.38644153  1.218052298  0.293157863
#> 95   0.78000723  0.249008404  0.873421207
#> 96   0.61603917 -1.158333378 -0.702622778
#> 97   1.37176192  0.326794297 -0.060126760
#> 98   1.66420559 -1.614003056 -0.884122521
#> 99  -0.28605681  0.312318244  1.096483891
#> 100 -0.56332142  0.302029639  0.863366549