Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A           B            C
#> 1    0.050957878 -0.02105697  1.338287583
#> 2   -1.233134371  0.15688070  0.380917489
#> 3    2.302609979 -0.55369872 -0.255679421
#> 4    0.302452520  0.96486686  0.829498555
#> 5    0.047302721 -0.74027881 -0.124174527
#> 6   -1.868216767 -1.90039344 -1.113479818
#> 7    0.309672920 -0.40049780 -1.012160751
#> 8   -0.842286120 -0.88427175 -0.137377908
#> 9    0.558453800 -0.99861610  1.564685809
#> 10  -0.029524978  0.57131201 -0.579967645
#> 11  -0.260027476  1.39356992  2.187894914
#> 12   1.007473788 -0.31394938  2.321787663
#> 13   0.222337893  0.01163092  1.166561028
#> 14   1.088386440  0.28492484  0.945354995
#> 15  -0.591873819 -0.22857433 -1.068221918
#> 16  -0.099199404 -0.01419913  0.937314173
#> 17   1.722192783  1.81172672  2.483768387
#> 18  -1.963724104 -1.79802714 -1.804999326
#> 19   0.519561763 -0.82640122  0.563430173
#> 20   0.478734589 -0.57744873 -0.447373885
#> 21   0.382152923  0.88490299  1.245772178
#> 22   0.592937258  1.10947419  0.660097931
#> 23   1.544471703  2.21028066  2.035756548
#> 24   0.098672370 -1.05419114  0.659232238
#> 25   0.099562765  0.78328496 -1.009255199
#> 26   0.510075333  1.16144497 -1.136959575
#> 27  -0.020354172  0.34975660 -0.192887375
#> 28   0.643687350  0.59123321  1.296405053
#> 29   0.027890546  0.73208543  0.831234159
#> 30  -0.962251483 -0.96468897 -2.202025603
#> 31  -2.471469136  1.38215750  0.304180461
#> 32   0.073389531  1.42953844  1.424567507
#> 33  -1.249017190 -1.25424119 -0.987575281
#> 34   1.444405024 -2.26551758 -1.044712876
#> 35   0.657020944 -0.50030516 -0.356770615
#> 36  -0.710185790  0.93596613  0.124013657
#> 37  -0.913141611 -0.89131839 -0.167303643
#> 38   0.941694467 -0.81666321  0.159888482
#> 39   0.862099142  0.69046747  1.225707566
#> 40  -1.350391994 -1.01147327 -1.346615362
#> 41  -0.192835581 -0.46898476 -0.131304861
#> 42  -0.179471190  0.03844748 -0.605748225
#> 43   0.638998896 -0.33569937  0.410420878
#> 44  -0.780973494 -0.49371296  0.361086677
#> 45   0.244258434 -0.74039756  0.007798702
#> 46  -0.459159359  0.44921461 -0.693928848
#> 47   1.944754512  1.05304164  1.103559201
#> 48   0.138173551  1.77164466  0.488614516
#> 49  -0.301816790  0.19348307 -0.700573733
#> 50  -0.361571009 -1.81726367 -0.337258468
#> 51  -0.975159852 -0.84446467 -1.407412110
#> 52  -1.544342483  0.45576885 -0.594914765
#> 53   0.854262758  2.20905630  1.363525220
#> 54   0.178766935 -0.93499048 -0.174407232
#> 55   0.144693019 -0.42862715  0.294575544
#> 56   0.121725636 -0.64910942 -0.056846110
#> 57  -1.481235223  1.27388998  0.339945952
#> 58  -0.231025696  0.08262273  0.369641877
#> 59   0.260817758 -1.05752374 -1.858000646
#> 60  -1.220141707  0.32103481 -1.774515603
#> 61   0.604986599  0.36331096 -0.633983426
#> 62  -0.809542807  0.51913418 -0.127223720
#> 63   0.171726038  0.24257948  0.035209639
#> 64  -0.906862734 -0.42462741 -2.052401284
#> 65   0.372979025  0.19054139 -0.357740778
#> 66  -1.151048530  0.66754431 -0.277242352
#> 67  -1.069989648 -1.78319695  1.172047174
#> 68   1.002824539 -0.97416735  0.426219025
#> 69  -0.658983622 -0.53672767  0.206294988
#> 70  -0.237547747 -0.97768920  0.871267088
#> 71   0.694285517 -1.07127245  1.466698373
#> 72   1.811720388  1.46393111  0.754469540
#> 73  -0.346277997 -0.32970278 -0.693586678
#> 74   1.132894571  0.80438521  0.825610912
#> 75   1.835010000  0.81466739 -0.435184463
#> 76  -0.522053417  0.01833821  0.285964790
#> 77  -0.422086358  1.30242719 -0.628079145
#> 78   1.753655385  0.54092845  0.399591437
#> 79  -0.341539381  0.91506830  1.896986804
#> 80  -0.912174177 -0.62064036  0.174356737
#> 81  -1.175517380 -0.16559743 -0.138436949
#> 82  -0.677477172 -0.49695992  0.734853818
#> 83   1.068991956 -2.46756847 -1.675646436
#> 84  -0.407218236 -0.93186377 -0.664018957
#> 85   0.709057453 -0.54033016  0.608487006
#> 86   1.126327525  0.52706814 -1.366919552
#> 87   0.244455602 -0.55169938 -0.512223449
#> 88  -2.248484094 -0.69243024 -2.298760936
#> 89   0.579251882 -0.62642769  1.245703500
#> 90  -0.558378016 -0.10028440 -0.534242073
#> 91   1.579857293  2.29258851  1.783328001
#> 92  -1.159825374  0.34238668 -1.981578588
#> 93  -0.008707177  0.97937789  1.416547371
#> 94   0.221322343  0.04934062 -0.971016188
#> 95   0.190456740 -0.56656142  0.893225767
#> 96   1.479698675  0.48493130  0.748259460
#> 97   0.618979988 -0.53930969 -0.597271426
#> 98   0.519836254 -0.97172825 -0.295118795
#> 99  -0.094192341 -0.07857621 -1.127257368
#> 100  1.096727802  0.17476366  0.542098928

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A           B            C
#> 1    0.97585794 -0.02468948 -0.887085174
#> 2   -0.94713542 -0.44998663  1.317833204
#> 3   -0.33431691  0.25762060  0.174865236
#> 4   -0.72368455 -1.71666632 -0.436449816
#> 5   -0.59561373 -0.16760562  0.559326119
#> 6    0.68273731 -0.40344832 -0.854166604
#> 7   -0.56809051  0.10056298  0.599448406
#> 8   -1.45489827  0.72429026  0.537292677
#> 9   -2.94357173 -1.13167583  0.320175570
#> 10  -1.91613108 -0.45577650  0.194307532
#> 11   0.54030799  0.87471321 -0.850640159
#> 12  -0.17084883  1.58529601  1.360859407
#> 13   0.46886020  2.69344058  1.482583622
#> 14   1.79792541  0.49156973 -0.741322390
#> 15   0.06489370 -0.71178272  0.997413210
#> 16  -0.33534308  0.56744622  0.300595161
#> 17   2.71081231  1.06345268 -1.407034781
#> 18  -0.24040116 -0.53679551 -0.096028319
#> 19  -0.97053884 -0.15496262  0.485258961
#> 20   1.26571705 -0.61447163 -0.446873374
#> 21  -1.75751562 -1.12645945  0.174627332
#> 22   1.08080748 -0.92124948 -1.517697691
#> 23  -1.00013509  0.41600900  1.240228145
#> 24   1.17129073 -0.81832811 -1.493568925
#> 25  -1.27587043 -0.06882480  0.369652213
#> 26   0.64834095  0.76267960 -0.231571398
#> 27  -1.07264649  0.59978327  0.131789203
#> 28   1.70434799  0.95756680  0.511031779
#> 29   0.23992755 -1.66577709 -1.240450809
#> 30  -1.78673146  0.08344559  1.377021206
#> 31  -0.64716450  1.27358571  1.848659160
#> 32  -0.24579574  0.45348933 -0.002278003
#> 33   0.61878631  0.51806164  0.053680519
#> 34  -1.35071219 -2.32502230  0.034769974
#> 35  -2.43653967 -0.23286213  1.213651649
#> 36  -1.20362286 -0.72356599  1.184650677
#> 37  -0.54538121 -0.43061449  0.174734128
#> 38  -0.14853299  0.10597567 -0.853841238
#> 39  -1.32401512 -2.46652335 -2.724544479
#> 40  -1.72405229  0.61573454  1.571147704
#> 41  -1.03367757 -0.55884017 -0.221341753
#> 42  -1.60087777 -0.28377314  1.825967957
#> 43  -0.94063963  0.10695050  0.374362263
#> 44   0.85027989 -1.15505591 -0.910603614
#> 45   0.66363376  0.05121069 -0.598530934
#> 46   0.63040128  1.80688199 -0.167210469
#> 47   1.86242208  1.53342661 -0.507831601
#> 48  -0.99203550  1.98091547  1.732043621
#> 49   0.97457632 -0.64603495 -0.702926667
#> 50   0.24637234  0.69667795 -0.232751940
#> 51   0.88133783 -0.09367850 -0.273568123
#> 52  -0.92498704  0.60245264  1.122836885
#> 53  -0.35349424  0.69956288  1.527358038
#> 54  -1.38953413  0.67168426  1.310292520
#> 55   0.99948599  0.06588706 -0.646265802
#> 56  -0.10454452 -1.15031016 -0.453377328
#> 57   0.15398764 -0.53473426  0.134516438
#> 58   1.59740128  3.10969298  0.798347628
#> 59   0.98289727 -0.67286750 -1.993890856
#> 60   1.63288182 -1.12948562 -1.924948462
#> 61   0.03222185  0.44470942  0.309178947
#> 62  -0.18163003  0.43907557  0.289082493
#> 63   1.90931485  0.22822695 -1.396239576
#> 64   0.58401296  1.72189442  1.672720670
#> 65   0.83694291  1.37703302  0.955538396
#> 66   2.14970984 -0.01303079 -1.624843658
#> 67  -2.34010505  0.68679411  0.941974694
#> 68  -0.97591766 -0.10542862  1.597013470
#> 69  -0.63685666  0.97623225  0.498477392
#> 70   1.24314339  0.22097568  0.237294575
#> 71   0.59766129  0.51037302  0.659530520
#> 72   2.30372755  0.13388812 -0.874062105
#> 73  -0.84956919 -1.82015504 -0.589762837
#> 74  -0.83036674 -1.12431340 -0.481558201
#> 75   1.38649983 -0.68129248 -0.695486440
#> 76   1.15210368  0.20252960 -0.481089025
#> 77   1.99876500  0.45928316 -0.799165072
#> 78  -0.95533862  0.98690538  1.445488571
#> 79   0.54809971  0.78990531 -0.529365945
#> 80   0.82480618 -2.10590175 -2.005339993
#> 81  -0.29369552 -0.41646483 -1.068824002
#> 82   0.72273898  0.34173219 -0.772932817
#> 83  -0.24355244 -1.67774042 -0.907687279
#> 84  -2.71134644 -0.30457793  1.707834197
#> 85   0.46675611  0.25056305  0.003817216
#> 86   0.21979615  0.60369004 -0.251753063
#> 87   1.04504928 -0.64845288 -1.013209963
#> 88   0.70470275  0.76986147 -0.800072219
#> 89   0.02813112 -0.17569548  0.271117546
#> 90   1.72496336  1.93829593  0.420616003
#> 91   1.04575581  0.44155601  0.748884958
#> 92   0.65206088 -0.35056718 -1.101618574
#> 93  -0.52690495  2.05934484  2.058160906
#> 94  -0.16558071  0.05428708  0.591318903
#> 95  -0.74152143 -0.95382079 -0.241501551
#> 96   0.20797419 -0.36220565 -0.778845196
#> 97  -0.40505776  1.02771895  0.636374119
#> 98   0.19466175 -0.55123849 -0.924060111
#> 99  -0.12556101 -1.95029825 -1.913324350
#> 100  1.31547990  1.16877445 -0.061971397