Multiple Normally Distributed Vectors

rnorm_multi makes multiple normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, cors = 0, mu = 0, sd = 1,
  varnames = NULL, empirical = FALSE, as.matrix = FALSE)

Arguments

n
the number of samples required

vars
the number of variables to return

cors
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

varnames
optional names for the variables (string vector of length vars) defaults if cors is a matrix with column names

empirical
logical. If true, mu, sd and cors specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

Value

dataframe of vars vectors

Examples

rnorm_multi(100, 3, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A            B           C
#> 1    0.750792572  1.424873995  1.01619162
#> 2   -1.042243673 -0.275816637  0.52639710
#> 3    0.878491981  2.434684102  2.13943193
#> 4   -0.500509231  0.536184495 -0.07283162
#> 5   -1.099839859  1.395668514 -1.63782640
#> 6   -1.268203415 -0.516346650 -0.90254750
#> 7    1.062341096  2.085361364  1.05389880
#> 8   -0.072670487  0.713209683 -0.06669068
#> 9    0.799007438  0.832138301 -0.84904571
#> 10   0.274439429 -1.305844900  1.50703213
#> 11   0.483758950  0.961135166 -0.09054153
#> 12  -1.335560745  0.410231530 -0.60767290
#> 13  -1.365394967 -1.989742102 -1.39808947
#> 14   1.058685530  1.942674898  0.79418602
#> 15  -0.059286958 -0.896242839 -0.20963410
#> 16   2.732876265  0.225408869  1.48006414
#> 17   0.520040403 -0.045099167  0.69894364
#> 18  -1.261009283  0.795477058  0.38156161
#> 19   0.152387337 -0.001507871 -1.21894494
#> 20   1.153452339 -0.243123670  1.18059995
#> 21   1.221036402 -1.329257135 -0.72696920
#> 22   0.097150590 -0.488381489 -0.38403747
#> 23   1.847044170  0.215193119  1.03746265
#> 24  -0.213594069 -0.833126883 -0.60370775
#> 25  -0.944882567 -1.108280341 -2.13611782
#> 26  -0.203627401 -0.399418526  0.70486850
#> 27   0.971899030  0.251539268  0.93124695
#> 28  -0.067710098  0.330687769 -0.20521675
#> 29   0.008234813  0.643985743  1.10851032
#> 30   1.491364279  1.606960145  0.51407350
#> 31  -1.491314723 -0.173569149 -0.59828848
#> 32   0.788400822 -0.100235777 -0.87821239
#> 33  -0.275325098 -0.506920585  0.16149729
#> 34  -1.646030985  0.066253801 -2.08692166
#> 35  -0.763940091  0.400133152  0.01055207
#> 36   0.559861711  0.094147624 -0.24183194
#> 37   0.851096610  2.217705492  1.27999821
#> 38  -0.174916339  0.113335133  0.60261481
#> 39   0.579792609 -0.678784282  0.78461866
#> 40  -0.270889155 -0.210612089 -1.77436354
#> 41   0.045616422  0.372504963 -0.50383699
#> 42  -0.469471970  0.436848174  1.27704375
#> 43   1.116862858 -1.388382465  0.45742838
#> 44   0.435540978 -0.674785328  0.76644032
#> 45   0.733199928 -0.330524626 -1.17879303
#> 46  -1.202678340 -3.307972897 -1.77046080
#> 47  -0.829440876  0.569422843  0.04572102
#> 48   0.001784459 -0.684587716 -0.57477015
#> 49  -1.255359394  0.562293801  0.23466705
#> 50   2.329409361  0.563363357  1.59377323
#> 51  -0.074278634 -0.571594535 -1.19274011
#> 52   0.364196656  0.305387342 -0.05428873
#> 53  -0.399415335  1.216792632 -0.33993885
#> 54  -0.237192644  0.568221498 -0.35410471
#> 55   0.395099319 -1.135493621  0.63054444
#> 56   0.309387409 -0.471324983 -0.94209682
#> 57   1.467751541  1.584594817  2.09891298
#> 58  -1.921116085 -1.649349291 -2.52231252
#> 59   0.936534595  0.074720566 -0.05360963
#> 60   0.095504143 -1.141376090  0.49330431
#> 61  -2.100395036  0.162246518 -0.70544430
#> 62   1.437719944  0.530497597 -0.22212529
#> 63  -0.956070367 -0.668411888 -0.93985501
#> 64   0.399297618  0.122294316  0.08451302
#> 65  -0.286806047  2.194503989  0.68953463
#> 66   0.569197516  1.947087832 -0.16302798
#> 67  -1.949290727 -0.897080260  0.12529717
#> 68   0.394865364  0.065765899 -0.62937094
#> 69  -0.503900977 -0.175240568 -0.45048740
#> 70   1.439840970  0.910978267  1.54359768
#> 71   0.887303297  1.141197444  1.30315927
#> 72  -0.380841256  1.126287256 -1.27061154
#> 73   0.046357782 -1.651026981 -1.30496243
#> 74   0.501969183 -0.740051843 -0.21905098
#> 75  -0.327677152 -1.412160664 -1.18738515
#> 76  -0.041519929 -1.656652359  0.41972821
#> 77  -1.237010922 -1.007396285  0.60385664
#> 78  -0.999395609  0.507528583  0.54872639
#> 79  -0.508546869 -0.250979903  0.26216982
#> 80   1.706916012 -1.013138961 -0.24417856
#> 81  -2.250638781 -1.264743365 -1.06547039
#> 82  -1.465750791 -0.014902364 -1.49434831
#> 83  -1.819797525  0.131454869 -0.23351058
#> 84   0.644464430 -1.365010318 -0.43015447
#> 85   0.673776094 -0.023604894  0.61492934
#> 86   0.424058255 -2.337033827 -0.53427314
#> 87   1.838236640  2.403650450  1.78253998
#> 88   1.880351381 -1.351154782  0.05113393
#> 89   0.143202959 -1.430192202  0.26174046
#> 90   1.155769129 -0.186903327  0.04880420
#> 91   0.559813595 -0.862499677 -0.53711912
#> 92  -0.936341076 -0.503527111 -0.99627787
#> 93  -1.049096344  0.032194348 -1.34962850
#> 94  -0.497978396  0.162235981  0.31899526
#> 95   0.565120969 -1.919780662  0.25948394
#> 96  -0.664895124 -0.766542929 -2.21344800
#> 97   0.829280403  0.592403315 -0.43854944
#> 98  -0.017677656 -1.656619109 -0.45664659
#> 99  -0.131923116 -1.896881557 -0.91048949
#> 100  0.146219686 -0.896684910  1.27558673

rnorm_multi(100, 3, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A             B           C
#> 1    1.39081618  1.2987554842 -0.54274184
#> 2    0.12474394 -1.4341962212 -0.83385076
#> 3   -1.31234622 -0.1773139308  1.11548187
#> 4    0.13764915  1.4386412100  1.18934079
#> 5   -1.55271429 -0.7649622047 -0.14994138
#> 6   -0.69104513 -1.1444681510 -0.39532881
#> 7    0.28331723  0.6930376611  0.78522733
#> 8    0.62698282  2.6282635812  2.09300595
#> 9   -1.35745434  0.4066573669  2.08806190
#> 10  -0.06698887 -2.4653818490 -1.20034841
#> 11   0.02483736 -0.0410126675 -0.86813242
#> 12   0.79142685 -1.8084817841 -0.37123506
#> 13  -0.88409828 -0.1531384713  0.49194054
#> 14   0.60358179 -1.3450087188 -1.86485444
#> 15   0.74467121  1.2994217211  0.32121665
#> 16   0.26231671  0.7238420836 -0.21253993
#> 17  -0.08841355  0.8852067209  1.38717738
#> 18  -1.18692758 -1.5262910813 -0.22392502
#> 19   1.02473345  0.3596218704 -1.36911133
#> 20   0.29173164  0.0537979231  0.48062427
#> 21  -0.10151671  1.0013139089  0.41741153
#> 22  -0.38409912  0.1086193588  0.27613213
#> 23  -0.17473214  1.3226328277  1.56677087
#> 24  -0.68841302 -0.8568191519  0.22462907
#> 25  -3.40268201 -1.1503624457  2.19509962
#> 26   1.27476703 -0.1608727574 -1.60243175
#> 27  -0.76774044 -0.3136043209  0.40748355
#> 28  -2.93463796  0.0002708164  0.54330948
#> 29  -0.36309562  0.4812979048  0.08769210
#> 30   0.11235349  1.2436686101  0.40667721
#> 31  -0.42112919  1.1241955619  0.82149954
#> 32   0.65132497 -1.9692499386 -1.72175518
#> 33  -0.39001734  1.2671731043  1.01569959
#> 34   1.05444567  0.7305534127 -0.08594560
#> 35  -0.65195625 -0.5195396901  0.41514849
#> 36  -0.95833910  0.2003833523  1.41935582
#> 37  -0.09312194 -1.7133185727 -0.91550462
#> 38   0.15254009 -0.8956400361 -1.25958623
#> 39   0.07535271  0.1033439817  0.16790440
#> 40  -1.61236803 -1.1266807862  1.13697372
#> 41  -0.53493316 -1.6457716779  0.28787366
#> 42  -0.69921708  1.7978716255  1.97787573
#> 43  -0.58102874 -1.3412768754  0.28024767
#> 44   0.39080373 -0.9614140080 -1.51340506
#> 45   0.98098214 -0.2651307385 -1.15693664
#> 46  -1.41159592  0.9114650277  1.44179604
#> 47  -1.05225142 -0.9418931935 -0.20252229
#> 48  -0.40068116 -1.1950503292 -0.24799347
#> 49   1.38669506  0.9589220188  0.06210057
#> 50  -0.35464447  0.8920651923  0.55778884
#> 51   1.53904431  1.2176546707 -0.77686030
#> 52   0.16578814  0.4331779127 -0.40893746
#> 53  -0.60914155 -0.8140949898  1.01178545
#> 54  -2.16814062 -0.2863029111  1.76151931
#> 55  -0.85887053  0.7286470557  1.46281376
#> 56  -2.69493701 -0.5352409200  1.22842833
#> 57   0.25329846 -0.2477709459 -0.33847285
#> 58  -1.42870286  0.8775450724  1.65829026
#> 59   1.59072274  0.1852101502 -0.06485567
#> 60   0.19047954 -0.5331432083 -1.01246796
#> 61   1.12633124  0.4510249139 -0.16385936
#> 62  -0.86502391  0.4564717361  0.97700377
#> 63   0.41198377  2.1574189238  0.29625434
#> 64  -1.33266567  0.1334556024  2.13446398
#> 65  -1.26258869 -0.1687697156 -0.84856895
#> 66  -1.02980626  0.6029517902  0.65050954
#> 67   0.38682167 -1.3061125155 -1.84341380
#> 68   0.23417755 -0.4236220335 -0.99612575
#> 69   0.38936213  1.4383517669  1.20765999
#> 70   0.52613652  0.4669543053 -0.25615525
#> 71  -0.78625593 -0.1890983901 -0.06583813
#> 72  -0.52607091 -0.7252667231 -1.05297080
#> 73  -0.28890582 -1.0441594449 -1.15239430
#> 74  -1.25168046 -2.1088896749 -1.11497680
#> 75  -0.67847213 -1.4529646435 -0.09410188
#> 76  -0.99112777  0.2271702323  1.19094153
#> 77  -0.27328045 -0.6278109929  0.89942994
#> 78  -0.05788188 -0.2011886425 -0.70444974
#> 79   0.44426016  0.3086371578 -0.14094071
#> 80  -1.86391621 -0.2017716705  1.04344653
#> 81  -0.52027246  0.1144370542  0.17047033
#> 82   0.68790768  0.3227648343 -0.35982484
#> 83   0.06379928  1.0963218557 -0.79995582
#> 84   0.71240232  0.6765428974  0.08052656
#> 85   1.00680982  1.3004044786  0.46945908
#> 86  -0.19307208 -0.8284714031  0.42528039
#> 87   0.58798451  0.2443710099  0.22379468
#> 88   0.29882724  0.8581150980  1.22292625
#> 89   1.84074363  1.0970181676 -0.69553217
#> 90  -1.91299092 -1.2784569227  0.17807915
#> 91   0.64255687 -0.5425659342  0.03429620
#> 92  -0.83135155  0.4205869612  0.80329005
#> 93  -1.44298017  0.7499570038  1.75007779
#> 94  -0.80721771 -1.3782307639 -0.83537525
#> 95  -0.56849692 -0.2160438395 -0.20213552
#> 96  -1.43237796  1.0144647169  1.92039006
#> 97  -0.17925759 -1.5018571840 -1.44550197
#> 98  -0.35894894  0.9200034445  1.16485148
#> 99   0.03204885 -0.5520698956  0.01677568
#> 100  0.83799420  1.2748026410 -0.90440295