Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>               A           B             C
#> 1    0.24332102 -2.21785929 -1.460216e+00
#> 2   -0.30688291  0.23254981  1.416053e+00
#> 3    0.23014554 -1.02612137 -1.065311e-01
#> 4   -0.06186371 -0.98902678  1.166026e-01
#> 5    2.71353554  0.35840405  1.250155e+00
#> 6   -0.23988911  2.34541774  5.758502e-01
#> 7    1.42183378 -0.30268281 -2.562630e-02
#> 8   -0.81361239  0.39223337 -1.856173e-01
#> 9    0.68348824  0.56584852  3.322114e-01
#> 10   1.02574015  0.79176227  2.911226e-01
#> 11   0.28212188 -0.08961090 -6.572245e-01
#> 12  -0.89227955 -0.73908038 -7.690980e-01
#> 13  -1.77421174  0.60418402 -1.453983e+00
#> 14   1.47908719 -0.77034868 -1.532873e-01
#> 15   0.03827916 -0.16754770  8.311183e-01
#> 16   0.06435803  1.90661524  1.396589e-01
#> 17  -2.03295749  0.44170835 -1.751406e-01
#> 18   1.76636959 -0.76014342  2.712549e-01
#> 19  -0.68367677  0.25503888 -1.125852e+00
#> 20   0.10947480  1.09394209  6.457922e-01
#> 21  -1.34865515 -0.78134113 -1.413411e-01
#> 22  -1.41867377 -1.47068395 -2.084747e+00
#> 23   1.06149493 -0.48972689  1.960782e-01
#> 24  -0.78847054  0.20166891 -4.370971e-01
#> 25  -0.46392596 -0.38747183  2.057112e-01
#> 26   1.07824930  0.68775974 -3.587041e-02
#> 27  -0.21110714 -1.52954187  1.462995e-02
#> 28   0.59574731  0.65213856  1.445706e+00
#> 29   0.29169366 -0.41542039  2.622708e-01
#> 30   0.82036971 -0.25968254  7.625221e-02
#> 31   0.10181347  0.23364112  5.937274e-01
#> 32   0.44220357  0.58337853  5.272609e-01
#> 33   0.93412248 -0.36897288  4.986964e-01
#> 34  -0.11535908  1.48502263  1.245799e+00
#> 35  -0.92211635 -0.30468450 -9.469831e-01
#> 36  -0.07774043 -1.43339098 -1.457751e+00
#> 37  -1.29717609  0.02687543 -7.208074e-01
#> 38   0.46774763  0.94022835  5.410583e-01
#> 39   0.69328253  1.35366831 -3.251433e-01
#> 40  -1.29115927 -1.00198144 -1.062734e+00
#> 41   0.77363270  0.06011522  1.156583e-01
#> 42   0.77093873  0.50750252  6.206510e-01
#> 43   0.09020124 -0.06163578  1.050929e+00
#> 44  -0.04074801  0.05988283  6.438895e-01
#> 45  -1.98357530  2.15821051  2.274703e-02
#> 46  -0.48655083  0.51711880  2.020707e+00
#> 47   0.73743021  2.37283237  9.954530e-01
#> 48  -0.55415169 -0.27886835  2.327818e-01
#> 49   0.38853252 -0.06100341 -4.008855e-01
#> 50  -0.82907299  0.12386927  2.327774e-01
#> 51  -0.13810014  1.46094546  1.361371e+00
#> 52   1.10538362  0.70758223  1.596276e-01
#> 53  -0.04352064 -1.35665552 -1.262910e+00
#> 54  -1.10021115 -0.63323665  4.444964e-01
#> 55  -0.76977727 -0.39073484 -4.452786e-01
#> 56   2.35334476  1.42072290  1.830123e+00
#> 57   0.97382037  0.63826999  2.346967e+00
#> 58  -0.68755875 -0.95492742 -7.916237e-01
#> 59   0.63343850 -0.52091171  5.818521e-01
#> 60  -1.53405884 -0.53714169 -1.161130e+00
#> 61   0.03574930 -0.23358764 -6.009896e-01
#> 62   0.06288822 -0.91031737 -2.891737e-01
#> 63  -0.56342210 -0.27624718 -1.251667e+00
#> 64  -0.28219342 -0.35685014 -5.127877e-01
#> 65   0.39835546 -2.21649746 -1.363644e+00
#> 66  -0.95955597 -0.35108210  6.732179e-01
#> 67   0.85435970  0.04687564  2.029395e-01
#> 68   0.03450823  0.55867772  5.001089e-01
#> 69   0.06269156 -1.89343430  7.013132e-02
#> 70  -1.20304578  0.48174856 -2.165483e+00
#> 71   0.07661879 -1.41203830 -5.825901e-01
#> 72   0.46246472  1.60467962  4.696472e-01
#> 73   0.73098636  0.93944891  6.998442e-01
#> 74   0.35123133 -1.38614262  5.741135e-01
#> 75   0.37405500 -0.58398492 -1.092638e+00
#> 76   0.20592563 -0.40847926  1.019681e+00
#> 77  -1.44832167 -0.02566422 -8.132406e-01
#> 78   0.80932184 -0.92725452 -1.449054e-01
#> 79   0.16656071 -0.70369938  1.511537e-01
#> 80   0.89325441  0.36002673 -1.451562e-01
#> 81   1.04014021  0.39698144 -1.344224e+00
#> 82  -0.30199261  0.39480885  4.774403e-01
#> 83  -0.20325219 -0.15901612 -5.788620e-01
#> 84   0.31098437  1.66296957  1.426547e+00
#> 85   0.37042079  1.20032494 -5.747393e-01
#> 86   1.62119896 -0.20743044 -3.285429e-01
#> 87   0.76382934  1.19475170  1.098234e+00
#> 88  -0.73163100  0.18231169 -1.738957e-02
#> 89  -0.45446902 -2.19767004 -1.928748e+00
#> 90   1.36358511  1.18888327  1.849353e+00
#> 91  -0.80633319  0.86121048 -1.330035e+00
#> 92   0.65708864  1.98537208  1.561405e+00
#> 93  -0.59006277 -1.30976065  3.091451e-01
#> 94  -2.63945503 -1.27646021 -1.470579e+00
#> 95  -0.27085752  0.20454716  8.367554e-02
#> 96  -1.47153928  0.08486604  1.129796e+00
#> 97   1.20764341 -1.02803857 -1.485763e+00
#> 98  -1.69211463  0.66828431 -1.149995e-01
#> 99   0.14420677 -1.43025782  9.737805e-05
#> 100 -0.40966052  0.40785127  1.187277e+00

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>                A           B            C
#> 1   -0.161258018  0.29714746  1.070073690
#> 2   -1.238910236  0.46456906  0.062890586
#> 3   -0.735324281 -0.50492020 -0.244914905
#> 4   -0.903845649  0.18916838  0.984206766
#> 5    0.114601975 -0.32777738  0.477126540
#> 6   -0.765128575  1.78274361  1.709615176
#> 7    0.087198714  0.32897712 -0.054399863
#> 8   -0.181008987 -0.76889747 -0.315878513
#> 9    0.527402209  1.39487954 -0.280638807
#> 10   0.306582693 -0.34089474 -0.803487321
#> 11  -1.534521072  1.81372224  1.007463433
#> 12  -0.541254567 -1.55548940 -1.942464272
#> 13   0.070839733 -0.53645200  0.301935493
#> 14   1.189515076  1.19818151  0.718672732
#> 15   0.056660966  0.67427965  0.063907045
#> 16  -1.081136397  0.72122528  0.772568718
#> 17   0.905725536  1.31586957 -0.157924442
#> 18  -0.031644754 -0.28706021 -0.138508218
#> 19   0.642901997  1.28420809  0.561550550
#> 20  -0.296535507 -0.05298307 -0.718913829
#> 21   1.197460548  1.25434849 -0.548816378
#> 22  -0.146599803 -0.43360095 -0.477258059
#> 23  -1.111171498 -0.11224284  0.492499613
#> 24  -0.034451244  0.37425022  0.765720336
#> 25  -0.802697925  0.15640895  1.729552560
#> 26  -0.266866432 -1.69575185 -0.193488994
#> 27  -0.693111745  0.16750871  0.382316625
#> 28  -0.009651734  2.15833582  2.314599926
#> 29  -0.217447012 -0.42132436 -0.055369649
#> 30  -0.705449992 -0.97335625 -0.064628111
#> 31   2.273053911 -0.16028480 -0.738871941
#> 32  -0.876837987  0.32114311  0.853108694
#> 33   0.445734813 -0.25545636 -0.293471043
#> 34   0.182226359 -0.63273791 -0.758620411
#> 35   0.371604412 -1.60827705 -0.652177376
#> 36   0.906354483 -0.95510256 -1.248075952
#> 37   1.618785541  0.12574197 -1.325104440
#> 38  -1.098689834 -1.06782969  0.599896247
#> 39  -0.559753513  0.64244731 -0.391912975
#> 40  -0.311821199  1.11961550 -0.211438862
#> 41  -0.364309911  0.24636096  0.823617691
#> 42   0.825693013  1.90868161  0.275755855
#> 43  -0.379221183  0.31031031  1.143113164
#> 44  -0.067675281 -1.08065077 -0.344521126
#> 45   0.216121631  0.25565838  0.447385295
#> 46  -0.249252885  0.30373886  0.610036103
#> 47  -0.340999151 -0.91865252 -0.653413355
#> 48   0.856628958 -0.94351932 -1.232385802
#> 49   1.673288889  0.18837044 -1.057590473
#> 50   1.680815914 -1.57595325 -1.752395891
#> 51  -1.367035941  0.90356160  0.289004790
#> 52  -0.968874800 -0.40409021  0.016406919
#> 53  -0.073323398  0.43886870  1.095397131
#> 54   0.410051660  1.18607767  1.292079578
#> 55  -0.077131626  0.54855988 -0.361691117
#> 56  -1.348912922  1.00830587  2.086875271
#> 57   0.200893509 -0.89666335 -0.513571842
#> 58   0.105616776  0.30100525 -1.127097427
#> 59  -1.154916409 -0.69345839  0.721603322
#> 60   0.996896067 -0.12819743 -0.005536809
#> 61  -1.324755639  0.53096491  1.040821129
#> 62   1.097057220  0.44857365 -1.699663917
#> 63  -0.707046310 -0.54965411  0.122673214
#> 64   1.331233303  2.42608610  0.624154380
#> 65   1.417852784 -0.24167449 -1.736318778
#> 66  -0.747167993 -1.84875754 -0.918218771
#> 67   0.411080933 -0.36692941 -0.909644671
#> 68   1.315457837  0.85426873 -0.514266791
#> 69   0.200396573 -0.24250857  0.533498188
#> 70   0.426316227 -1.70974682 -2.050725407
#> 71   0.920852746  1.47715228 -0.456822522
#> 72  -0.932182315  0.52357785  0.743713253
#> 73  -0.238758244  0.42824369  0.870915788
#> 74   0.137372775 -0.31572916 -0.086771209
#> 75   0.645913221  0.54834892 -0.292266599
#> 76  -1.932199460 -0.54307423  1.915011696
#> 77   0.673113776 -1.30088242 -1.048233901
#> 78   1.365412863 -0.56563722 -0.533262163
#> 79   1.176373282 -0.15207947 -0.603278614
#> 80  -0.119127404  0.13726882  0.650365593
#> 81  -1.756759534 -0.34506632  0.568206971
#> 82   1.013705614 -0.77522670 -1.055892775
#> 83  -0.762797660  0.57016200  1.487015503
#> 84   0.537901280  2.10053120  2.390827255
#> 85   0.964350799  2.25537265  0.982598331
#> 86  -1.778351713 -0.86616548 -0.736596220
#> 87   0.567249530  0.79074084  0.315787123
#> 88   0.661377388  1.52921960  1.169416577
#> 89  -1.304123932  0.03847698  0.612649817
#> 90  -1.081979287  0.35254004  1.134414101
#> 91  -0.100269973 -0.47585684  0.201363856
#> 92  -0.304435033  0.41796790  0.247136793
#> 93  -0.248340239 -0.33134789  0.488381748
#> 94   0.626635719  1.23067934 -0.395121910
#> 95  -0.603960813 -1.78589996 -0.060369268
#> 96   1.988727693  0.67259588 -1.727580329
#> 97   1.153983222  0.12107375 -1.321991540
#> 98   1.273385962 -0.08277217 -0.787231086
#> 99   1.468850298  0.34157345 -0.208674882
#> 100  0.305218976  0.94291992  0.988436081