Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>               A            B            C
#> 1    0.21916000  0.789944739 -0.001787874
#> 2    1.16782726  0.607409577  0.507929725
#> 3   -0.11805890  1.082179234 -1.043897207
#> 4   -1.15572984 -0.460825678 -1.003547535
#> 5    0.18001921  0.191395498  1.013023489
#> 6    0.42933852  0.223447795  0.412427179
#> 7   -0.67145288  0.618867638  0.659900532
#> 8   -1.03631095  0.292670197 -0.236494180
#> 9    0.26461434 -1.626630369 -0.669459430
#> 10  -0.88696204 -0.267279823 -0.620374036
#> 11  -1.50088183 -1.304618686  0.658323753
#> 12  -0.85980285  0.694816016  0.776929199
#> 13  -1.50507852 -0.593117257 -0.450295977
#> 14  -0.43467814  0.404107023 -0.674018106
#> 15  -1.16872040 -0.554819878 -1.040603510
#> 16  -0.16230383  1.350508977  1.209020636
#> 17   0.43906823  0.731739160  1.226939128
#> 18  -0.63789992 -0.308560611  0.759235079
#> 19   1.70424301  1.884465293  2.842968381
#> 20   0.76009486 -1.007470129  0.648114834
#> 21  -0.97103721 -0.280137456 -0.629757326
#> 22   0.17390359 -0.625151136 -0.656623457
#> 23  -1.41184815 -0.206210879 -1.728962155
#> 24  -0.34962642 -0.722586184 -0.145871191
#> 25   0.86313767 -1.999225624 -2.127518121
#> 26  -2.40624374 -1.632587332 -1.510559709
#> 27  -1.31258993 -2.439008818 -0.419942904
#> 28   0.36374042  0.566065545  1.978864004
#> 29  -0.10296801 -1.302583258  0.492431888
#> 30  -0.03748323  1.185858014  1.241331988
#> 31   1.77374629 -0.437111709  1.261123268
#> 32  -1.33943380 -0.004480001 -0.591383073
#> 33   0.81750186  0.619614267  1.048118393
#> 34   0.07875661  1.231557000  1.163777165
#> 35   0.86331603  1.305611629  1.670914562
#> 36  -0.13198266 -0.081516825  0.323672395
#> 37   1.91753208  0.816027372  1.377898700
#> 38  -0.27836316 -1.021127085  0.459896506
#> 39  -2.33446932 -1.788095742 -1.615907507
#> 40  -0.32327790  0.382545024  0.072357313
#> 41   1.38322443  0.535757116  0.458040607
#> 42   1.76896132  0.282768991  2.610890101
#> 43   0.79577229 -0.180806275  0.251668398
#> 44   0.75077539 -0.557287273  1.253779535
#> 45   0.95457431 -0.267814726  0.491473299
#> 46   0.48224509  0.120686788  0.367527902
#> 47   0.33117838 -1.895499885 -0.532550385
#> 48   0.42805275  1.782631750  0.433740634
#> 49   1.26422902  1.308480576  0.186426033
#> 50  -0.47305849  0.060953206 -0.128019418
#> 51   0.21262468 -0.898944845 -0.883385929
#> 52  -0.10073275  0.108511027 -0.538555395
#> 53   0.87361856  1.574311391 -0.677992642
#> 54   1.60714279 -0.297464507  0.044741492
#> 55  -1.05472642 -1.175129128 -0.300412066
#> 56  -0.58311298 -0.887895024 -1.713710255
#> 57  -0.67277212 -1.006949704  0.280011058
#> 58   1.55801786 -0.552614585  1.439696808
#> 59   0.22803458  1.610530408  0.288662455
#> 60   0.45876341  0.620446823  0.423943283
#> 61  -0.59891530  0.099227735 -0.512190806
#> 62  -0.37050030 -0.694058800 -1.698198212
#> 63   0.14492864  1.880255493  2.416887952
#> 64   0.17634921 -0.789408698  0.792927809
#> 65  -0.52315809  0.219919807 -0.508286551
#> 66  -0.79937204  0.137104477 -0.909469705
#> 67  -0.57665718  0.485633083  0.859973581
#> 68  -1.91790676 -1.400084832 -1.169561432
#> 69  -0.79868640  0.021377621  0.334112313
#> 70   0.03060156 -0.656339649  0.220972772
#> 71  -0.33881132  1.903001601 -0.722351163
#> 72   1.11173254  1.925306892  0.441651212
#> 73   0.48677408 -0.941132478  1.651297729
#> 74  -0.36880853 -0.610048040  1.341836251
#> 75   0.44939949  0.202766672  0.728575159
#> 76  -0.19866640  0.393402791 -0.181141608
#> 77  -1.05430168  0.455147531  0.882452470
#> 78  -1.71291634  1.479090013 -0.875698603
#> 79   0.66515486  0.936045633  0.554843397
#> 80  -1.03183396  1.140354215 -1.125642693
#> 81   1.76302189 -0.422987470 -1.186092848
#> 82  -0.36721352  1.026184730  0.556674768
#> 83   0.22878873  0.107156249 -0.061909002
#> 84   0.60943381  1.309155485  0.781873789
#> 85  -0.75141101  0.636642373  0.307018797
#> 86   0.58240363 -0.811753502 -0.502469509
#> 87  -0.96052471 -0.193269295 -2.252252561
#> 88   1.00009731  0.653134998  1.500184887
#> 89   2.02417137  1.222788067  0.467053399
#> 90   0.45095953  0.953465379  1.616543846
#> 91   0.61116906 -0.395619607 -1.189716234
#> 92  -0.92731565  0.373177944  0.358007096
#> 93  -0.89565690  0.529788358 -1.002704692
#> 94  -0.87313188  0.993117589  0.467583467
#> 95  -1.59895736  0.832071982  0.760604584
#> 96  -0.60142902 -0.887354155 -1.133995314
#> 97   0.25231852  1.255181857 -0.769749197
#> 98  -1.70450431  0.193371552 -0.570570393
#> 99   0.07794962  0.157330802  1.882426057
#> 100 -0.05199213  0.011464137 -0.485304646

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>                A           B            C
#> 1   -1.109563411 -0.55122644  0.054694844
#> 2   -0.471929477  0.90777655  0.568437109
#> 3   -1.125916935 -1.42865661 -0.040356630
#> 4   -0.583589724  1.00449973  1.002450050
#> 5   -1.053192239 -0.39284132 -0.384091846
#> 6    0.423109914 -0.05158812 -0.887648059
#> 7    0.238292713  0.57890220  0.746748993
#> 8    0.410973625 -1.14850767 -1.278215977
#> 9    0.791350887 -1.55660902 -0.473013398
#> 10  -0.175614085 -1.26586726  0.466690997
#> 11  -1.440493428  0.26556850  0.373593495
#> 12   0.188305254  1.22427701  0.350171399
#> 13   1.089373451  2.25683775  1.170155173
#> 14   0.936146476  0.35405063 -0.772726732
#> 15  -1.636214460 -1.09204669  0.441963383
#> 16  -0.119534044  1.90748276  0.573914076
#> 17  -1.066878253 -0.28108619  0.968317967
#> 18   0.185739877  0.76860870  1.318556385
#> 19   0.846839125 -0.76447442 -1.193440570
#> 20   1.681797378  1.08474888  0.368080820
#> 21  -0.131721776  0.83416972  0.172504520
#> 22   0.416243013  0.43633569  0.813789560
#> 23  -0.059931022 -1.20704821 -0.485258210
#> 24  -1.154770730 -1.84909636 -1.972304312
#> 25   1.810713457  0.21389106 -0.172711611
#> 26  -0.410426305  1.27063875  0.562343567
#> 27  -0.594113993  0.74901228  1.465362937
#> 28   2.089362145  1.23847855 -1.301944206
#> 29   2.070427521  0.80327688 -0.376628385
#> 30  -0.054400500  1.44914265  0.999255317
#> 31  -0.536175553  2.16329621  2.656119796
#> 32  -0.024984326  1.12926716  0.464293150
#> 33  -1.117166606 -2.21690841 -1.017408410
#> 34  -1.192069758  0.73161010  0.630002695
#> 35  -0.769702570  0.16914941  1.158683677
#> 36  -1.396732389  1.37135159  0.966884882
#> 37  -0.861959972 -1.22891046  0.182980390
#> 38   0.298912647  1.53439611 -1.207044099
#> 39  -0.537544076  1.09267442  2.411679947
#> 40   0.457576838  0.65183406 -0.481460912
#> 41  -0.896754959 -0.24539600  0.696945317
#> 42  -0.285673552 -1.55594821 -0.714542936
#> 43   0.194814802  0.96100869  0.074081282
#> 44  -0.010628470 -1.65179002  0.185856731
#> 45  -1.135599849 -1.00413144  0.977384467
#> 46   0.305786085 -0.39056180 -0.617649875
#> 47  -1.112387118  0.01738082  1.422939032
#> 48   0.509263103 -0.29953569 -1.214933731
#> 49   0.114015266  0.39913603  0.585027886
#> 50   0.711374675  0.68404453  0.440852650
#> 51   0.008158753  0.82078667  0.070662411
#> 52   0.207799986 -0.07341362 -0.772450021
#> 53   0.712779631 -0.56758013 -0.617819568
#> 54  -1.136061400 -1.01785409 -0.524289472
#> 55  -1.188482073  0.02578678  0.957430736
#> 56   1.604959577  0.68106148 -1.249806727
#> 57   1.108365373  0.76395514  0.461929837
#> 58   0.067325069 -1.28115183 -0.546994088
#> 59   0.240866026 -0.51216072 -0.548159726
#> 60   0.857058107  0.19150753 -1.215803449
#> 61  -1.115169401  0.64511880  1.245982311
#> 62   0.428976077  1.50914691  1.247772303
#> 63  -1.427039424 -1.21619497  0.192501610
#> 64   1.141518675  0.76096946 -0.244585277
#> 65   1.435298269  0.03283902 -0.710766000
#> 66  -1.358814552 -1.45646451 -0.582558998
#> 67   0.461825321  0.49748476  0.527805213
#> 68   0.260997171 -0.38308995  0.007899547
#> 69   0.768054228 -0.66367391 -0.734667647
#> 70  -0.030192184 -1.15509575  0.253724566
#> 71   0.456605167 -0.09736196 -0.213930464
#> 72   1.204366807  0.02265832 -0.181342682
#> 73   0.049976824 -1.63353747  0.236613353
#> 74  -1.580187420 -1.72782391  1.071800465
#> 75   0.553830820 -1.29624331 -1.650399299
#> 76   0.369149519 -0.38975126  0.294409479
#> 77   1.150511095  0.64852362  0.543882612
#> 78  -0.480138313  0.22699237  0.361211717
#> 79   1.379220370  0.77015368 -0.336505295
#> 80   0.033118785  0.29827173 -1.130084574
#> 81  -0.923480621 -1.66947376 -0.806910136
#> 82  -0.618137068 -0.82054258 -0.392097147
#> 83   0.160383056  0.77543902 -0.053121219
#> 84  -0.919621268 -0.06139275  0.499226694
#> 85   0.600026401 -2.21145838 -1.731213045
#> 86  -1.658397633 -0.39222692  0.612343697
#> 87  -1.640949171 -2.16239598 -1.189897825
#> 88  -0.644669126 -0.84715356  0.486093057
#> 89   0.427710783  1.48116909  1.232250462
#> 90  -0.307540278  2.61906158  1.169257488
#> 91  -1.832507588  0.90522325  2.408473587
#> 92   0.385069645 -0.10327462 -1.138176979
#> 93  -2.283591458  0.86969599  2.200179571
#> 94  -1.047559905 -2.04557464 -0.112967715
#> 95  -0.349803996  1.31216383  0.332834972
#> 96  -0.111524249  0.38027440 -0.006413775
#> 97  -0.096301779 -0.43162170 -0.576423682
#> 98   0.335427664  0.59272255 -0.041856379
#> 99  -1.384735681 -0.21170165 -0.196190534
#> 100  1.031676046 -1.76449696 -0.611059415