Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(
  n,
  vars = 3,
  mu = 0,
  sd = 1,
  r = 0,
  varnames = NULL,
  empirical = FALSE,
  as.matrix = FALSE,
  cors = NULL
)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A           B            C
#> 1    0.038846720 -1.17486254  0.230449968
#> 2   -0.447962616  0.69691970  1.016284979
#> 3    1.233899700 -0.23319036  1.269919781
#> 4   -1.227380038 -0.30546853 -0.771135427
#> 5   -1.037249495 -0.74063206 -2.288971004
#> 6    0.633677134  0.73946064  0.459580247
#> 7   -1.525337331  0.01573329  0.924361074
#> 8    0.137449841 -1.80006720 -0.873064591
#> 9   -0.001661869 -0.48888468 -1.072379626
#> 10  -0.996124159 -0.75304837 -1.616197426
#> 11  -0.093439443 -0.45607076 -0.968910315
#> 12   1.485157951 -0.85251473 -0.442177403
#> 13   0.275735601  0.30725087  0.009372804
#> 14  -0.874745518  1.08651056 -0.149661228
#> 15  -1.808116117 -0.91926561 -1.656298049
#> 16  -2.180037167  0.10206868 -1.108606190
#> 17   1.159188408  1.28164903 -1.445631217
#> 18   0.358458438 -0.17543111  0.563522977
#> 19  -4.168234669  0.99897199 -0.418945642
#> 20   0.316510398 -0.33706906  0.807579130
#> 21   0.078053625 -0.06936209  0.318155948
#> 22   1.336017457  2.16092521  2.538725858
#> 23  -0.710729502  0.05065200 -0.433490065
#> 24  -0.911272875 -0.18898234  1.287862733
#> 25   1.728727764  0.65959699  1.675001338
#> 26   0.521245388  0.67190326  0.586706739
#> 27  -0.357574039  1.54399170 -0.781644790
#> 28  -0.580467536 -0.07969083 -0.141103358
#> 29   0.673532462  1.08474591 -0.027594373
#> 30  -0.644134602  0.90148702 -0.023939689
#> 31  -0.674139285  0.40427789  2.046525408
#> 32   0.263838587 -0.05632087  0.738507286
#> 33  -0.895052101  1.51749955  0.243660748
#> 34   0.508185873  0.92777477  0.646700503
#> 35   1.216124812 -0.06152501 -0.709767217
#> 36   0.807042054 -0.63740584 -1.913408788
#> 37  -1.229462856 -1.46533907  0.457143214
#> 38   1.944590523 -0.80184247 -0.599330114
#> 39   1.996955284  0.84086074  2.154145172
#> 40  -0.662032725  1.14425767  1.791284905
#> 41   1.030627409  0.23362438 -0.414050827
#> 42   0.256636591 -0.86116140 -0.641016643
#> 43   0.194042311 -1.11035454 -0.608314913
#> 44  -0.639269361 -0.17624740 -1.303142303
#> 45   1.087459228  0.64135527  0.356548189
#> 46   0.123535026 -0.28045516 -0.309856966
#> 47   0.487761259 -0.44655793 -0.820299728
#> 48  -1.749493784  0.58275372  0.036135247
#> 49  -2.554351762  0.85237357 -1.004064851
#> 50  -0.384693196 -0.47484165  0.464910928
#> 51  -0.441318352  1.73922530  0.482157054
#> 52   3.260177780  1.50696687  0.556516199
#> 53   0.118210954 -0.37019832  0.653313046
#> 54  -0.161425520  0.50942733  0.578363532
#> 55   0.092790910 -1.12353943 -0.037286806
#> 56  -0.174033668  1.13520532  1.067573838
#> 57   0.708928790  0.24308235  0.819796575
#> 58  -0.633353021  0.04202638  0.126824070
#> 59   2.163772555 -1.30360341 -0.707669245
#> 60  -1.067030911  0.30082455 -1.205678236
#> 61   0.381753494 -0.48908874  1.014595223
#> 62  -0.134309528 -0.84318735 -0.949576997
#> 63   0.470780812  1.12202517  1.511112350
#> 64   0.977178456 -0.72262347 -0.432182869
#> 65  -0.719505000 -0.82244052  0.160858745
#> 66  -0.858514811 -0.27504702 -1.208967152
#> 67  -1.686429452 -0.46747945  0.235865362
#> 68   0.600624685  1.13178359  1.055633570
#> 69   2.164708379  0.53576393  0.367505465
#> 70   0.408372297 -1.56632832  0.370718779
#> 71   0.318848117  0.63742050  1.200530162
#> 72   1.127733338  0.48489667  1.062235793
#> 73  -0.627263403 -0.49979843  0.195019146
#> 74  -0.342301254  1.25670942  0.043064124
#> 75  -0.196705556  0.43770883  0.023015507
#> 76   0.572695213 -0.29248837 -0.141213789
#> 77   1.095838103 -2.42550493 -0.443505030
#> 78  -0.958637352  1.63181083  1.406294541
#> 79   1.732981383 -0.30740401  0.685289795
#> 80  -2.833999194  0.19245927 -0.059064974
#> 81  -1.204684266  1.14260665  1.551692954
#> 82   0.857843692 -0.94119391  0.051951332
#> 83  -1.724888881 -0.99934762 -1.399041541
#> 84  -1.223164808  0.61765118 -0.744359415
#> 85  -0.039491707  0.13527623 -0.073270516
#> 86  -0.571210192 -1.66753818 -0.437237890
#> 87   0.147784657 -0.20289388 -0.977656617
#> 88   0.659294078  1.04270014  0.576653349
#> 89   0.178444422 -0.09025133  0.473166189
#> 90  -0.172848438 -0.21062244 -0.086570437
#> 91  -0.340457396  0.92465642  1.293953454
#> 92   0.638013503  0.90649996  0.164694348
#> 93  -0.301681670  0.07456913  1.113733879
#> 94  -0.493293463  0.27156687 -0.343384002
#> 95   1.543554153  0.95650927  2.883299718
#> 96  -1.394379268 -0.40704304 -1.063147729
#> 97  -0.062426589  1.33445610  0.473774834
#> 98   1.513132322 -0.42310714  0.657813333
#> 99   2.210801165  0.55665684  1.628438546
#> 100  0.736477516 -0.53235392 -0.151055874

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A            B            C
#> 1    1.58016753  1.787656858  1.233122780
#> 2    0.19231093 -1.864118109 -2.243680694
#> 3   -0.45994722 -0.811582552 -0.925892873
#> 4   -0.72573271  0.157231180  0.473356883
#> 5    0.36377719  0.029121425 -0.638729460
#> 6    1.28025722  1.499379358  0.600468874
#> 7    0.02421728 -0.474381688 -1.089011040
#> 8   -0.27803970  0.456275894  0.804097105
#> 9   -0.43841051 -0.561266056  0.116362655
#> 10   1.04257697 -0.486889233 -0.802376406
#> 11   0.34160542  1.016860271 -0.592123016
#> 12   0.15640394 -0.454650937 -0.024736701
#> 13   0.50928911  0.145969336  0.128580245
#> 14   0.36882505 -0.674015572  0.264797789
#> 15  -1.02572170  1.456859641  1.505317038
#> 16  -0.98915226  1.841336616  1.540043885
#> 17  -0.78261391 -0.091803566  0.356299301
#> 18  -0.15480879 -1.066996603 -0.365337603
#> 19  -0.98202738 -1.114986888 -0.942794706
#> 20  -1.50135644  1.398342354  1.203835360
#> 21  -0.92700064 -0.437534919  0.188437542
#> 22  -0.86360984  0.379671420  1.347755274
#> 23  -0.45618427  0.977470721  0.936984764
#> 24   0.21894785  0.163334061 -0.288696300
#> 25   0.33180261  0.679203646 -0.025690689
#> 26  -1.42482642 -1.038810773  0.784688030
#> 27   0.49962636  0.416042672 -0.236850673
#> 28   0.23234790 -0.474221662 -0.941169773
#> 29   0.30269267 -0.367433578 -0.354420001
#> 30  -1.17885630 -0.719396262 -0.766364898
#> 31  -0.23296819  0.078174591 -0.255934414
#> 32  -0.43637027 -0.618810081  0.524278635
#> 33   1.42543456 -1.283127181 -0.552526531
#> 34   0.94111885 -1.554553523 -1.960934077
#> 35   0.42762314  1.838295174  1.862603093
#> 36   1.01826512  0.199995547 -1.104774080
#> 37   0.22696487  0.036429431 -0.141092431
#> 38   1.61817081  0.030237374 -1.571643480
#> 39  -0.10345894 -0.218701712 -0.526614748
#> 40   0.08947029 -1.335091073 -1.655364699
#> 41   0.09083941  0.937607237  0.337090693
#> 42  -1.50320517 -0.111512829  0.847486680
#> 43   0.29587957 -0.669835085 -1.249067173
#> 44   0.46157186  0.559108252  0.230580739
#> 45   0.97374999  0.837853920  0.065865470
#> 46   1.08862168  1.296889567 -0.499242644
#> 47  -0.55197786 -1.209007081 -0.627180072
#> 48   1.58242434  0.009245386 -1.247151081
#> 49   1.34356322  2.275879031  1.015437425
#> 50   0.34445194  0.701906851 -0.005939782
#> 51   0.07080953 -1.721575713  0.084985781
#> 52   1.73238469  0.335835021 -1.602129234
#> 53   1.08582938  0.748248881 -0.202329239
#> 54   0.18940228  0.028371952 -0.737063204
#> 55   1.27569138 -0.138820123 -1.451097864
#> 56  -0.22693898 -0.410428971  0.075285795
#> 57  -1.47903087  0.311788882  0.926598868
#> 58  -0.38929985 -0.323417681  0.196946727
#> 59  -1.41486180  0.237762120  1.155819792
#> 60   0.54597537 -0.053705678 -1.445966656
#> 61  -0.15955161 -1.652252923  0.811645985
#> 62  -1.38326357  1.119223171  1.384325924
#> 63  -1.06819762 -0.419246700  0.459130538
#> 64   0.64308558  0.380509076  0.423997485
#> 65  -0.20167852 -0.282841187  0.486317928
#> 66   0.79228957 -0.553734149 -1.204561204
#> 67   0.30267632 -1.184227377 -0.513281290
#> 68   0.21344703 -0.028044741  0.407816062
#> 69   0.15623027  2.803484816  1.655107283
#> 70   0.68954323 -2.692834139 -1.197704437
#> 71  -1.95943270 -1.427051927  0.506250377
#> 72   0.33051769 -1.431596588 -0.179130826
#> 73   0.49423006 -1.759569655 -1.023046442
#> 74   0.97675356 -0.110801412 -0.807357609
#> 75   0.83710390  0.206585687 -1.131708135
#> 76   1.36626953  0.460119977 -0.963337747
#> 77   0.28977114  0.193676149 -0.022769931
#> 78  -0.66897336 -0.266925905  0.102311976
#> 79   0.83485474  0.567765185  0.477034298
#> 80   0.21435734 -0.614443228  0.009587954
#> 81   0.33811367  0.069692030  0.356131199
#> 82  -0.10242798  0.077620216  0.461196143
#> 83   0.08848788 -1.610513038 -1.749937015
#> 84  -0.37532567 -0.698516568 -0.286843137
#> 85   1.05945737  1.476812508  0.879702465
#> 86   0.15267268 -1.111630470 -1.054074005
#> 87  -1.62427587  1.680831180  1.520981056
#> 88  -1.57103880 -1.834688390 -0.078877460
#> 89   2.19600308  0.587557241 -1.535398272
#> 90  -0.28102174 -0.901511928 -0.052053751
#> 91   1.23763297 -0.797464324 -1.718162273
#> 92   0.01914682 -0.324716259  0.598962056
#> 93   0.83148326 -0.063085977 -0.647024241
#> 94  -1.58943674  1.605731600  1.622136013
#> 95   1.31505834 -0.107561755 -0.221363993
#> 96   1.07332455 -0.257773738 -1.372125201
#> 97  -0.34655871 -1.606688982 -0.714317336
#> 98   0.93840646  0.624012668 -0.935799742
#> 99   1.50496441 -0.761968363 -1.275548307
#> 100 -2.87147409 -0.745721075  1.874942887