Multiple correlated normal distributions

Make normally distributed vectors with specified relationships. See vignette("rnorm_multi", package = "faux") for details.

rnorm_multi(
  n,
  vars = NULL,
  mu = 0,
  sd = 1,
  r = 0,
  varnames = NULL,
  empirical = FALSE,
  as.matrix = FALSE,
  seed = NULL
)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

seed
DEPRECATED use set.seed() instead

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A           B            C
#> 1   -0.031164989 -0.48416524 -0.607375989
#> 2   -0.290046163  0.10501131  0.291631309
#> 3    1.688621841  0.27019969  1.607796739
#> 4    1.090500830  1.37888040  4.019378954
#> 5   -1.457990540 -0.41820276  0.908627365
#> 6   -0.104309212  0.09346967 -0.078208545
#> 7   -0.944448682 -1.06351198 -0.236208131
#> 8   -0.334398920 -1.23035907 -1.483456078
#> 9   -0.979907229  1.97074199  1.151000730
#> 10  -0.922596960 -1.48955328 -1.082635803
#> 11  -1.107869601 -0.10668845 -0.322194292
#> 12  -1.763131312 -0.89876684  0.067482043
#> 13   0.246249477  0.87631037  0.752337574
#> 14  -0.774791268 -2.06854192 -2.712674376
#> 15   1.307348402  0.42205970  1.101281255
#> 16   1.257052588  0.93727582  0.827890152
#> 17  -0.593351292  0.09085857  0.568780936
#> 18   0.479709203 -0.22288663 -0.554515302
#> 19   0.578246369  0.32742417 -0.319216514
#> 20  -0.098119487  0.93940731  0.709163993
#> 21  -1.509643012 -1.49482972 -1.138303920
#> 22   1.964504212 -0.47453168 -0.038503344
#> 23   0.754669986 -0.65423371  1.039116535
#> 24  -0.736410393  0.84488546  1.073572641
#> 25  -0.419507726  0.02402811  0.772682645
#> 26   1.178374740  1.99253717  0.931608377
#> 27  -1.287584073 -0.59368606  0.452689383
#> 28  -0.389492327  0.35430665  1.198061488
#> 29   0.066628791 -0.79649573 -1.264191403
#> 30   2.038286430 -0.21260018  2.269700589
#> 31  -1.890056880  0.20323338  0.156736669
#> 32   1.849110605  0.97720609 -0.815717102
#> 33   1.055066957 -1.98086538  0.022869603
#> 34   0.271274973  1.29085273  0.488171828
#> 35   0.029223334  1.38307507  1.431758008
#> 36   0.928188113  1.06742860  1.615789699
#> 37  -0.689793064  1.21571682  0.361762423
#> 38  -1.373855314 -0.57318006  0.399196782
#> 39   0.027982048 -0.42326312 -1.315185282
#> 40   0.064457328 -0.44474006 -0.208380787
#> 41   1.332980214  0.45149046  2.316398094
#> 42   0.240718670 -0.01134288  0.832111290
#> 43  -0.717237633 -1.00766467  0.993200693
#> 44  -0.697143195 -1.29869827 -1.510513981
#> 45  -1.453227015 -0.40709229 -2.379039034
#> 46   0.582251753  0.58361696 -0.003301991
#> 47   0.239105272  1.00978242  1.539270523
#> 48  -0.345538674  0.30688346 -1.477491295
#> 49  -0.850175337  1.27571054  0.631865772
#> 50   0.504244461  0.53725921 -0.668928912
#> 51  -0.145410869  0.81197201  0.646285112
#> 52   0.131616273  0.49744619  1.852932589
#> 53  -0.056844151  0.16697919 -0.873738571
#> 54   1.185836587  0.29162030  1.491115184
#> 55  -0.661113137 -1.44120384 -0.607972369
#> 56  -0.593302783 -0.50783754 -1.326089203
#> 57   0.101361288  1.22676950 -0.848277245
#> 58   1.428432925 -0.83760987 -1.116907859
#> 59   0.527143628 -1.32062811 -1.325002209
#> 60   0.598474283  0.22118537  0.208141175
#> 61   1.386991765 -1.04162091  0.117067885
#> 62  -0.313195029  0.56455327  0.134564531
#> 63   1.029685789 -0.24530678 -0.324707080
#> 64  -0.970490995 -0.83840379 -1.329739537
#> 65   0.515504184  0.03242235 -0.770153390
#> 66  -0.297840599 -0.80096049 -0.743422136
#> 67  -1.831246637 -0.30195111 -2.619441045
#> 68   2.003249451 -0.85739463 -1.665673241
#> 69  -0.925479846  1.09037559  0.101477837
#> 70   0.805358575  1.54322028  0.758879285
#> 71   2.015952040  0.59017745  0.860213834
#> 72   0.085414333  0.81054015  1.307162977
#> 73   1.551046672  0.62141563  0.485665909
#> 74  -0.145136053  0.76162058  1.665287083
#> 75   1.583717148 -0.46327209  1.977621738
#> 76   0.978848703 -1.03717810  0.557118339
#> 77  -0.834981157  1.11925343  1.314333581
#> 78   0.812532169 -1.97227593 -0.873035846
#> 79  -1.066547151  1.40907016 -0.015173660
#> 80   0.108138837 -1.15358477  0.779790020
#> 81   1.578429107 -1.80115745 -0.048301049
#> 82  -0.135206416  1.07364989 -0.776226451
#> 83  -0.600434723 -0.98031009  0.537310215
#> 84   0.003265433  0.62698036 -0.021387400
#> 85  -1.329701319 -0.30468518 -0.476311393
#> 86  -0.810324953 -0.32948629 -0.223660145
#> 87  -0.881791048  0.82640306 -0.133737217
#> 88  -1.029579113  0.06227840 -0.669127318
#> 89   0.698050806 -0.25056753 -0.250909004
#> 90   0.799377909 -1.22431084 -1.109906474
#> 91   0.369696864 -0.76983581  0.717756324
#> 92  -1.216766284  0.49551506  0.637885019
#> 93  -1.328351650 -1.68935729 -1.298885817
#> 94   0.583418144  1.07694843  1.005564293
#> 95   0.871845075  1.37276787  1.182500768
#> 96   0.151712207  1.06119334  0.007743209
#> 97   1.615540439  0.40076976  0.239514983
#> 98  -0.171941987  0.40161232  0.410496962
#> 99   0.069848823  0.15154019  1.027457851
#> 100  1.642966478  1.00439841  1.167481073

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A             B           C
#> 1    0.53812566 -1.4352349159 -1.33286140
#> 2   -0.34219863  0.6138994463  0.14068087
#> 3    0.46951083  1.4443882573 -0.18869774
#> 4    0.98544293  0.6846747030 -0.09862538
#> 5   -0.26110529  1.2863196217  0.79559577
#> 6    0.89743010  1.7683964184  0.44029477
#> 7   -0.16683418  0.4880817109  1.58197677
#> 8   -1.40665931 -1.7597989305 -0.80956569
#> 9    0.74009585 -0.8738195916 -1.08689077
#> 10   1.02688424  1.1052655705 -0.52127638
#> 11  -1.76316725 -0.7267457417  0.29805508
#> 12  -1.13146391 -2.1852370725 -1.49374753
#> 13  -0.10482611 -1.0260688652 -0.45488875
#> 14  -0.89681113 -1.2033399819  0.55151702
#> 15  -0.57908206  0.8721860980  1.27662931
#> 16   0.89915118 -0.5802641086 -1.74461509
#> 17  -0.02406963 -0.7114679730 -0.39819938
#> 18   0.68576581  1.2313729073  0.36770208
#> 19  -0.38432198 -1.5590461024 -0.53476924
#> 20  -0.64063266 -2.8763011855 -0.13635507
#> 21   0.25085666  1.1608916704  0.03213714
#> 22  -0.30950261 -0.0038173626 -0.70980578
#> 23  -0.92785249  0.3456553937  0.59531795
#> 24  -1.01798113 -0.5577979212  0.44669557
#> 25  -0.29759370  0.1878618739  1.42119594
#> 26  -0.90869144 -0.5777363939  0.29879894
#> 27  -0.40039456  0.0380472398  0.12249246
#> 28   1.80714384  1.5155170695 -0.36456977
#> 29   0.20370437  1.2895660923  1.07043394
#> 30  -1.38091101 -2.0906902611 -1.06207959
#> 31  -0.76176130  1.1270006758  1.30628261
#> 32   0.05624484 -0.4116752220 -0.53217419
#> 33  -1.03950598 -0.2028867374 -0.97576848
#> 34  -0.14350701 -1.7179061695 -0.36863909
#> 35   0.64972955  0.1345881990 -0.21976934
#> 36  -1.16736329  0.4964085784  1.04453081
#> 37  -1.03081181 -0.1397870066  1.14918707
#> 38  -0.52050419 -0.1310365678  0.33454824
#> 39  -0.92210500 -0.6306095147  0.85779677
#> 40   0.84344961 -0.3007840773 -0.35503163
#> 41   0.63196438 -0.6240909605 -0.59371878
#> 42   1.11753078  0.2799275598  0.35496008
#> 43   1.87284385 -0.0009892961 -1.73797503
#> 44   0.82521459  0.7299086882  0.33873296
#> 45   0.94714835  1.2220811888 -0.04624102
#> 46  -1.89713596 -0.9526051451  0.27517317
#> 47  -1.25273174  0.4461082835  1.00847404
#> 48   0.27191654  2.1059944801  1.26694767
#> 49  -0.41055999 -1.9376315529 -1.18633384
#> 50  -0.41990671  0.4761626269  0.14447154
#> 51  -0.88956347 -0.6956791978 -0.59245622
#> 52   0.28741270  0.0956330785  0.64275492
#> 53   2.51538748 -0.4920602938 -1.47510279
#> 54  -0.58316309  1.4811372323  1.20541136
#> 55   0.31333832 -0.8215381614 -1.49766899
#> 56  -2.15395472 -0.2299780771  1.26986651
#> 57  -0.84352470  0.3588477784  1.15183743
#> 58  -1.35938648 -1.4594844726  0.79330866
#> 59   0.09492242  1.8393596897  0.70919835
#> 60   0.27582180  0.4298595793  1.09210689
#> 61  -1.44600270  0.1710487360  0.44060938
#> 62   0.43778673  1.3119786745 -0.10549631
#> 63  -0.23104502 -0.5517152477  0.21029779
#> 64   0.44281756  0.0790335153 -0.95869886
#> 65   0.80914003 -0.1746323280 -0.87203247
#> 66  -1.02051389  0.5662628487  1.16228233
#> 67   0.92201022  0.5471308141 -0.41660559
#> 68  -0.11368021  1.9469454099  0.66700748
#> 69  -0.32483689 -0.1476592180 -0.31036973
#> 70   0.39179312  0.1218857813  0.07896971
#> 71  -0.19424885  0.6182151293  0.45175840
#> 72   0.45656038  0.6542808568  0.76260504
#> 73   0.45070012  0.5389397197 -0.66674565
#> 74  -0.42847963  0.6055420748  1.34766039
#> 75   0.44242012  0.7903807993  0.65752489
#> 76  -1.56328427  0.2372939452  0.95108093
#> 77  -0.48170147 -0.7956210266 -0.59248965
#> 78  -0.61388928 -1.1780831253 -0.84788246
#> 79  -1.12418352  0.9343957633  2.00836464
#> 80   0.96601125  0.2633500205 -0.96499023
#> 81  -0.95346220 -0.1058461256  0.47669103
#> 82   0.13280511  0.7635484856  0.03182757
#> 83  -0.05401680 -0.0841073722 -0.12787931
#> 84  -1.13732914 -0.7558390215 -0.42993286
#> 85   0.32113115  0.2616228656 -1.09086146
#> 86   1.16805976  0.6842891660 -0.53721979
#> 87  -0.72441802  0.8931293536  1.45226000
#> 88  -0.32953230 -0.7129472234  0.41061430
#> 89  -1.09891490 -0.4168544701  0.63427126
#> 90  -0.10821963 -1.3227632667 -1.21067849
#> 91  -0.31715538  1.2704386213  1.94619437
#> 92  -0.87847504 -0.3930716020 -0.08498610
#> 93   0.62436055  0.3434559001  0.05968153
#> 94   0.90739744  0.2670642559 -1.77632945
#> 95  -1.74335254  0.9024430729  2.41472289
#> 96   0.64678370  1.2197202388  0.09242800
#> 97   0.19366612 -0.3896991840  1.14235130
#> 98   0.26988908 -0.2803715700 -0.11474816
#> 99  -1.12683296 -0.4052643227  0.08977259
#> 100 -0.55951574 -0.3080705219 -0.13769456