Multiple Normally Distributed Vectors

rnorm_multi() makes multiple normally distributed vectors with specified relationships.

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>               A            B            C
#> 1    0.45171594  0.824352603  0.997911314
#> 2   -0.33308789  0.115403315  0.321737668
#> 3    0.14112243 -0.787470469 -1.619471073
#> 4   -0.66160025  1.171866730  2.099499492
#> 5   -0.38333302  0.576408619 -0.513008825
#> 6   -1.09488641  0.614413816  0.544146564
#> 7    0.75786712 -0.005910282 -0.757636128
#> 8    0.02325756  0.634282737  0.082947444
#> 9    0.23550192  0.050089641  0.841465542
#> 10  -0.19981207  0.292129350 -0.724730109
#> 11   0.38476742 -0.134181498  0.588531905
#> 12   1.20983038  1.393107166  1.327953379
#> 13   0.13600846 -0.178996693 -0.234585512
#> 14   0.71946924 -0.990576936 -0.869076424
#> 15   1.41428963 -1.168942868 -0.821931983
#> 16   0.48838855  0.971592644  1.442867551
#> 17  -0.31558895  1.659845238  1.347692801
#> 18   1.52845523 -0.105819522  0.383373483
#> 19   0.80003624 -0.992924320 -1.928591424
#> 20   0.63498003 -0.427565593  0.665061028
#> 21  -1.64607863 -1.503628576 -0.846064502
#> 22  -1.08338454 -1.883559948 -0.572869186
#> 23  -0.44450302 -2.573100755 -0.816779690
#> 24   0.07152515  0.027422434 -0.108881775
#> 25   0.70122257  0.111902351  0.958870228
#> 26   0.73598634  0.110281677 -0.297521430
#> 27   1.68584879  0.637895505  0.052534745
#> 28   0.29521674 -0.775167338 -0.163564771
#> 29   0.64105429  0.807606836  0.736892561
#> 30  -0.35207409  0.351566613 -1.040802978
#> 31   1.07027083  1.342847573  1.317562316
#> 32  -0.87573818  1.372613598  0.364911077
#> 33  -0.89301596  0.972276694  0.233508731
#> 34   0.85596335  0.583360085  0.195538068
#> 35   0.29084228  1.448849432  2.271864054
#> 36   1.29024121  0.921157539 -0.006081366
#> 37   0.72632795  1.449673935  1.771500349
#> 38  -0.44079509  1.123698998 -0.269929103
#> 39   1.40988642  1.576796620  1.530604277
#> 40  -1.65033966  1.018570160  1.456798307
#> 41   0.66414390  0.158295805 -0.179779937
#> 42   0.77592566  0.506210748  0.049217585
#> 43   1.31791328  0.750344193 -0.413203293
#> 44  -1.06014735  0.171683584  0.447260803
#> 45   1.50130029 -1.550536232  1.145113624
#> 46   0.43869168 -0.427291378  0.742040391
#> 47   1.30824267 -0.395828462 -0.683829682
#> 48   0.53258648  0.263778071  0.910338864
#> 49  -1.12599489  0.496876931 -1.050792048
#> 50  -1.34393559 -1.202707702  0.115144363
#> 51  -1.35305985  0.996296436  0.083180507
#> 52   0.34925398  2.444741123  0.609868742
#> 53  -0.87954138 -0.984816646 -0.959941941
#> 54   0.27553510  0.853474759  0.560819259
#> 55   1.16030620  0.310299525 -0.652412894
#> 56  -0.15310413  1.011934778  0.856544324
#> 57   1.52617249  0.162772746  0.113261860
#> 58   0.02240588  1.525555944 -0.097299576
#> 59   0.92851083  0.756350392 -0.408997992
#> 60  -0.23660263  1.482729840  0.545391662
#> 61  -1.49362649 -1.302779420 -1.506511209
#> 62  -0.40501976 -1.456818652 -0.853326473
#> 63   0.88916927  0.790543440  0.855405796
#> 64   1.07575962  1.386110786  1.819839889
#> 65  -0.73575267  0.418652475  1.058399719
#> 66  -0.39994134 -2.103187380 -1.093899708
#> 67  -0.23914643  0.526245759 -0.892902225
#> 68   0.22945761 -0.382944978  0.352603619
#> 69  -0.74634677 -0.327033867  0.079077969
#> 70  -1.94078789  0.082735435 -0.663210744
#> 71  -0.24697350 -1.140471730  0.641775516
#> 72   0.86699824  1.095384403  0.518890682
#> 73   0.41537484 -0.054622482  0.084058620
#> 74  -0.31850727 -0.236012589 -1.088947078
#> 75  -0.14567277 -0.044580335  0.469298090
#> 76  -0.12553292 -1.771355198 -0.845153173
#> 77  -1.36891006 -1.685842193  0.307874518
#> 78  -0.59409654 -0.200421225 -0.182253453
#> 79  -1.34495460 -1.180083767 -1.125048813
#> 80   0.27561172  1.532546592  1.514282710
#> 81   0.11611529 -1.527342737 -1.009633159
#> 82  -0.10320688  0.318702649  0.607929312
#> 83   0.26892393  0.053917982 -0.910940592
#> 84   0.64707680  0.578375306  0.722088508
#> 85   0.35304974 -1.680628074 -1.253147300
#> 86  -0.88716544 -0.125274480 -0.943046567
#> 87  -0.56431095  1.313213363  0.397001240
#> 88  -0.97541278  0.428285893  0.715796722
#> 89   0.95113578 -0.078819672  0.938016744
#> 90   0.44047279  0.410474500  0.504643248
#> 91  -1.44138826 -0.186543649 -1.008183911
#> 92   0.84605561  1.780440211  1.367225144
#> 93   0.33299543  1.299106707  1.129141485
#> 94   0.50152588  0.820185180  0.158002549
#> 95   0.60043038  0.589169863 -0.042511917
#> 96  -2.59432900 -0.541824410  0.080583009
#> 97   0.66956305 -1.090533343  0.700609154
#> 98   1.76748133  2.293373031  2.130681759
#> 99  -1.12634430 -0.657518060 -2.070366082
#> 100  1.30684260  0.582919472 -1.352495523

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>                A           B            C
#> 1   -0.723442129  1.32073881  1.762848837
#> 2    1.583312731 -0.13388255 -2.423886197
#> 3    0.029480003 -1.04169613 -0.717955381
#> 4   -0.846939267  0.29428267  0.503778846
#> 5   -0.804408458 -0.72033094 -0.095405479
#> 6    1.757149825 -0.59624970 -1.584703632
#> 7   -2.321222703 -2.21639094 -0.789675230
#> 8   -0.831761247  0.08228569  0.130018021
#> 9    0.675578589  0.96722819 -0.349377120
#> 10   0.799277000 -1.07921968 -1.717485123
#> 11   2.602068813  1.46294014 -0.398529444
#> 12   1.188552820  2.78689657  1.069767066
#> 13  -0.839150656  0.33254843  0.305568397
#> 14   0.079336337 -1.74351955 -1.324879949
#> 15  -0.644104326 -0.83477166  0.836343202
#> 16   0.890863524  1.30685427 -0.034763064
#> 17  -0.370618698 -1.17689757 -0.559703436
#> 18   1.033070014 -0.97460295 -0.911293536
#> 19   0.664502861 -0.11995574 -1.498505214
#> 20   1.702997900 -1.04302579 -1.845805557
#> 21  -0.297188980  0.84164085  0.874683267
#> 22  -0.309931810  1.35359737  0.012506540
#> 23   0.681721958  2.49536439  2.396368687
#> 24   1.102286068  0.12312885 -0.325628589
#> 25  -0.643946428 -0.88041095 -1.063220697
#> 26  -0.975665420 -1.84158921 -0.858938466
#> 27   0.185746858  0.30690607  0.906779372
#> 28   1.052037915  0.13584088 -1.041869361
#> 29   0.272228417 -0.36756814 -2.278443668
#> 30   1.846359471  0.16373275  0.183034699
#> 31  -0.819118694 -0.67551587 -0.614630848
#> 32  -1.477621367  0.20087746  1.373868794
#> 33   1.586930029  0.37240493 -0.833250731
#> 34  -0.556934570  0.93536174  1.496760123
#> 35  -1.047360986  0.38302741  0.313642394
#> 36   1.237101307 -0.27901272 -1.170692554
#> 37   1.363577381  1.28805559  0.367044429
#> 38   0.728670918  0.32705379 -1.094025979
#> 39  -0.885309351  0.94206762  1.228870694
#> 40   0.096793195 -0.22934802 -0.238411732
#> 41   1.001541816 -0.12477446 -1.021729113
#> 42   1.383589882  0.54433594 -0.039799629
#> 43   1.225748350  1.36803196 -0.567155619
#> 44  -0.060030360 -1.45024990 -1.663345103
#> 45  -1.813955155 -0.98245189  0.713899736
#> 46   1.109154407 -0.14813722 -1.482611049
#> 47   0.691201704 -0.95882795 -0.991061029
#> 48  -0.963113157  1.16098533  1.994893148
#> 49  -0.015399710 -0.67616435  0.843277457
#> 50   0.172144879  1.54048761  0.146195567
#> 51   0.969991498  0.71327727 -1.006389240
#> 52   0.034869870  0.85487306  1.174675427
#> 53  -0.656123118  1.84672822  0.132136246
#> 54   0.364301599 -0.25222330 -0.137697071
#> 55   1.882750102  0.19004179 -2.331152035
#> 56   0.946067242 -1.35148762 -0.422429529
#> 57  -0.004064029 -1.57903621 -0.995889746
#> 58   0.013231276  1.27455258  1.015980452
#> 59  -0.232313208  0.32422746 -0.118980473
#> 60  -0.984901266  0.25756197  1.146397623
#> 61  -1.616764116 -1.09367199 -0.586654406
#> 62   0.085984490 -0.21238141 -0.789380791
#> 63  -1.274278938 -0.09700955  1.215631273
#> 64  -0.812439348  0.01136173  0.847109266
#> 65   0.445249531 -1.63077969 -1.488152519
#> 66   1.372828328 -0.32997972 -0.430206428
#> 67   0.031557481  0.46075334  1.719371257
#> 68   0.384714793 -1.07360505 -1.259933232
#> 69   0.433638471 -0.17100543  0.069996373
#> 70  -1.691889698 -0.58152645  0.700402518
#> 71   1.325363801 -1.56916385 -1.440597753
#> 72   0.911560763  2.01559165  0.281133557
#> 73   0.347057669 -0.37404520  0.483423296
#> 74  -0.192178769  0.36736856  0.007051014
#> 75  -0.295491259  2.44555115  1.929079371
#> 76  -1.281157458 -1.04818755 -0.719469047
#> 77   2.073457032  0.22044530 -1.167253865
#> 78  -0.593963703  1.31390211  0.887666700
#> 79  -0.811307559  0.74521117  1.331424534
#> 80   1.472467438  0.75234212 -1.040458501
#> 81   0.694285299 -1.84011399 -2.124668599
#> 82   0.665230377 -2.18565552 -1.534480211
#> 83   0.118062685  0.61471677  0.660265694
#> 84   0.559132166  1.54281441  0.406627402
#> 85   0.805280680 -0.91841033  0.229021933
#> 86   1.195794191  1.44125099  0.610276371
#> 87  -2.287447277 -0.54300237  1.629503043
#> 88  -0.456310515 -1.36358699 -1.950402586
#> 89  -0.333400546  0.16208925  0.917035878
#> 90   0.085844316 -0.76461941 -0.398338606
#> 91   0.354074359  1.10084801 -0.384850486
#> 92  -2.281043126 -1.25963691  0.167272936
#> 93  -0.266475397  0.39624014  0.195050624
#> 94  -0.200612720 -2.06729255 -1.101855220
#> 95   0.177991618 -0.80640283 -0.860183414
#> 96   0.549693638 -1.52767244 -2.291509303
#> 97   1.292573922 -0.23760410 -0.184265492
#> 98  -1.386174053  0.59912386  0.787535324
#> 99   0.795527707 -0.04427783 -0.662682314
#> 100  0.051356656  0.31499765 -0.099054973