Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>               A           B           C
#> 1    0.67006492 -1.19835691  0.33123781
#> 2    1.04385672  0.18060675 -0.36315308
#> 3    0.81786747  2.10785320  1.69528021
#> 4   -0.93414635 -0.10425103 -0.73835773
#> 5    1.16679150 -0.58509667 -0.22861631
#> 6   -1.55329235 -1.32827762 -1.13045657
#> 7   -0.49026562  0.51495526  0.40834559
#> 8   -1.84384257 -1.38973873 -1.84895203
#> 9    0.09603309  0.56680169  0.34242181
#> 10   0.93003971  1.67177891 -1.17897566
#> 11  -0.60160723  0.82011149 -0.52357206
#> 12  -1.23498736  1.13482433 -0.63960785
#> 13  -0.04145501 -0.30273948 -0.53054436
#> 14   0.94813601  0.92795425  1.49305603
#> 15   0.47624495  0.43710749 -0.20197398
#> 16  -0.02230346  1.79857927  0.23080681
#> 17  -0.77081845  0.15452605  0.19485340
#> 18   1.44289376 -0.12326739  1.44434299
#> 19   0.53579426  0.03748765 -0.25010035
#> 20   1.67992987  1.03361684  1.19714410
#> 21  -1.18595621  1.79705775  1.63652232
#> 22  -0.95265033  1.32566390  0.73578825
#> 23   0.84414377 -0.48795612  0.11870378
#> 24  -1.67745634 -1.59479036 -1.18706176
#> 25   1.87045589  0.35384436  0.32838806
#> 26   1.17022709  0.78568752  1.33184016
#> 27  -0.68117344 -1.54287142 -0.83764079
#> 28   1.18905998  0.09133380  0.43626119
#> 29   1.22744339  0.69762284  1.10964919
#> 30   0.48877258 -1.65907679  0.73881621
#> 31   0.33652438  0.20931777 -0.29038622
#> 32  -0.03522946  0.48159220  0.96515872
#> 33   0.27958776  1.03771130  0.54873604
#> 34   0.28393548 -0.84960105  0.67057623
#> 35   1.18843225 -0.29289589 -0.72496813
#> 36  -0.83541926 -2.24095112 -1.09105820
#> 37  -0.56515031  0.92196940 -0.20660313
#> 38  -0.71917975  0.57583158  0.05415195
#> 39   0.26583855 -0.07641625  1.11711075
#> 40   0.66813607 -0.20793862  0.88178851
#> 41  -1.11140646 -0.75830295 -0.74800615
#> 42   1.14755294  0.13892352 -0.42624613
#> 43  -0.68700966 -1.15416494  0.03042862
#> 44  -0.46778580 -0.82383880 -0.28632716
#> 45  -0.37719010  0.07968623  0.38474816
#> 46  -0.44549769 -0.72306470 -0.04814899
#> 47  -1.04135504 -0.72676995 -1.52380036
#> 48   0.64551548  0.15411444  0.39166539
#> 49  -1.75383893  0.19252642 -1.52108315
#> 50  -0.47094420  0.01925446 -0.57683632
#> 51  -0.28796451 -0.01478058  1.27875525
#> 52  -0.41766933  1.36944276  0.61984577
#> 53   0.84952322  0.77131721  0.27787674
#> 54   0.49102823 -2.37901000 -0.42568583
#> 55  -1.66982975  0.69756142 -0.03709154
#> 56   0.88964836  0.34576263 -0.37260076
#> 57  -1.40279724  1.03822409 -0.66688927
#> 58   0.59076918 -0.22539239  0.39389631
#> 59   1.80079701 -0.87850662 -1.01613727
#> 60   0.04776284  0.13330512 -0.22043810
#> 61  -0.39934974 -0.24404994 -1.65984731
#> 62   0.06097273 -1.11105801  0.88648730
#> 63   1.29764090  1.13704973 -0.28082235
#> 64   0.85823930  0.09461432  1.42535401
#> 65   0.23283589  1.16575063  0.30698420
#> 66   0.24332102 -2.21785929 -1.46021594
#> 67  -0.30688291  0.23254981  1.41605269
#> 68   0.23014554 -1.02612137 -0.10653108
#> 69  -0.06186371 -0.98902678  0.11660257
#> 70   2.71353554  0.35840405  1.25015536
#> 71  -0.23988911  2.34541774  0.57585018
#> 72   1.42183378 -0.30268281 -0.02562630
#> 73  -0.81361239  0.39223337 -0.18561731
#> 74   0.68348824  0.56584852  0.33221141
#> 75   1.02574015  0.79176227  0.29112260
#> 76   0.28212188 -0.08961090 -0.65722446
#> 77  -0.89227955 -0.73908038 -0.76909800
#> 78  -1.77421174  0.60418402 -1.45398289
#> 79   1.47908719 -0.77034868 -0.15328726
#> 80   0.03827916 -0.16754770  0.83111833
#> 81   0.06435803  1.90661524  0.13965891
#> 82  -2.03295749  0.44170835 -0.17514065
#> 83   1.76636959 -0.76014342  0.27125490
#> 84  -0.68367677  0.25503888 -1.12585181
#> 85   0.10947480  1.09394209  0.64579218
#> 86  -1.34865515 -0.78134113 -0.14134107
#> 87  -1.41867377 -1.47068395 -2.08474697
#> 88   1.06149493 -0.48972689  0.19607818
#> 89  -0.78847054  0.20166891 -0.43709712
#> 90  -0.46392596 -0.38747183  0.20571117
#> 91   1.07824930  0.68775974 -0.03587041
#> 92  -0.21110714 -1.52954187  0.01462995
#> 93   0.59574731  0.65213856  1.44570554
#> 94   0.29169366 -0.41542039  0.26227076
#> 95   0.82036971 -0.25968254  0.07625221
#> 96   0.10181347  0.23364112  0.59372740
#> 97   0.44220357  0.58337853  0.52726093
#> 98   0.93412248 -0.36897288  0.49869640
#> 99  -0.11535908  1.48502263  1.24579948
#> 100 -0.92211635 -0.30468450 -0.94698311

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>                A           B           C
#> 1   -0.867596712 -1.33384501 -0.40028052
#> 2   -0.673875820 -1.39018272  0.15716979
#> 3    0.207572391  0.36214824 -0.79358489
#> 4    1.263652457 -0.94474264 -1.76285531
#> 5    0.249455086  0.36520894  0.43705346
#> 6    0.721466893  0.38203293 -0.32640313
#> 7    0.910128785 -0.04395513  0.38365855
#> 8    0.009488361  1.43273022  1.12947764
#> 9   -1.026366777  0.13647147  0.45938762
#> 10   0.731991338 -0.29776872 -0.63815688
#> 11  -0.790907466  1.66517007  2.02272081
#> 12   0.077612882 -1.00575257 -1.41658761
#> 13  -1.548908733 -0.60342516  0.73288399
#> 14  -0.275806903 -1.53738078 -0.23704679
#> 15  -2.056868197 -1.39775044  0.58111782
#> 16   0.477858352  1.07197262  0.28700376
#> 17   0.540220039 -0.20579302 -1.19002271
#> 18   0.862307282  0.43571766 -0.19217408
#> 19   0.110919411  1.51734267  1.82239785
#> 20  -0.049163734  0.06994460  0.29703683
#> 21   1.772204662  1.48174502 -0.38102311
#> 22  -1.600681036  0.50733402  1.42095323
#> 23  -0.555092669 -0.37205824  0.32157921
#> 24  -0.262785846  1.18800665  0.39673549
#> 25  -0.149104516 -0.49210741  0.08940664
#> 26   1.181147517  0.82624924 -0.62580832
#> 27  -0.972457472 -0.95892131  0.66695816
#> 28   0.439434556 -0.23751222 -1.08445231
#> 29   1.234768216 -0.04114625  0.55576177
#> 30  -1.156559208 -1.38019679  0.31707587
#> 31  -2.030687362  0.56253683  1.00959107
#> 32  -0.484896664 -0.79241254 -0.23162474
#> 33  -0.091930151 -0.64226178  0.47274416
#> 34  -0.844561039  1.08529337  1.55330558
#> 35  -0.427195689 -2.66843507 -2.36935373
#> 36   0.019293260 -0.12532507  0.89992999
#> 37   1.310335972  0.19740631 -1.07948806
#> 38   0.836071984  0.34282003 -0.32116335
#> 39  -1.151202418  0.81514114  1.52519331
#> 40   0.567686257 -0.89191259 -0.68149870
#> 41   0.090367778  2.45801321  0.87714019
#> 42  -1.045058999 -1.03603372 -0.25906579
#> 43  -0.786844627 -0.79587686  0.33205679
#> 44  -0.897057389 -0.40942074  0.83169766
#> 45   0.918858281 -0.17515943 -0.88298781
#> 46   1.752260311 -0.43100644 -2.66946788
#> 47  -1.171034566 -0.81247607  0.39312398
#> 48   1.277567800  0.84503870 -0.43389156
#> 49   1.898210286  1.18631051  0.54011492
#> 50   2.380094958 -0.74557293 -1.17834773
#> 51   0.133771275 -0.29615240 -1.47230455
#> 52   0.499500046 -0.08242275  0.10253827
#> 53   0.988463712  0.57878494  0.59611271
#> 54   0.536866533  0.78623805 -0.33409364
#> 55   0.588322593  1.26709490  0.24305679
#> 56   0.963861376 -0.99083548 -1.83590025
#> 57   0.472662861  0.16329278 -0.20369014
#> 58  -0.913176878  1.31585622  1.56874219
#> 59  -0.185764931  0.16352949  1.07258648
#> 60  -0.754102981  0.03027168 -0.37562068
#> 61  -1.927134471 -0.40742502  2.36667817
#> 62   0.136906934 -1.93794527 -1.14811495
#> 63  -0.676818376 -0.81467319  0.39304145
#> 64  -0.805420925 -0.02403713  1.40446912
#> 65   0.108009377  1.16250699  1.38819926
#> 66  -0.161258018  0.29714746  1.07007369
#> 67  -1.238910236  0.46456906  0.06289059
#> 68  -0.735324281 -0.50492020 -0.24491491
#> 69  -0.903845649  0.18916838  0.98420677
#> 70   0.114601975 -0.32777738  0.47712654
#> 71  -0.765128575  1.78274361  1.70961518
#> 72   0.087198714  0.32897712 -0.05439986
#> 73  -0.181008987 -0.76889747 -0.31587851
#> 74   0.527402209  1.39487954 -0.28063881
#> 75   0.306582693 -0.34089474 -0.80348732
#> 76  -1.534521072  1.81372224  1.00746343
#> 77  -0.541254567 -1.55548940 -1.94246427
#> 78   0.070839733 -0.53645200  0.30193549
#> 79   1.189515076  1.19818151  0.71867273
#> 80   0.056660966  0.67427965  0.06390704
#> 81  -1.081136397  0.72122528  0.77256872
#> 82   0.905725536  1.31586957 -0.15792444
#> 83  -0.031644754 -0.28706021 -0.13850822
#> 84   0.642901997  1.28420809  0.56155055
#> 85  -0.296535507 -0.05298307 -0.71891383
#> 86   1.197460548  1.25434849 -0.54881638
#> 87  -0.146599803 -0.43360095 -0.47725806
#> 88  -1.111171498 -0.11224284  0.49249961
#> 89  -0.034451244  0.37425022  0.76572034
#> 90  -0.802697925  0.15640895  1.72955256
#> 91  -0.266866432 -1.69575185 -0.19348899
#> 92  -0.693111745  0.16750871  0.38231663
#> 93  -0.009651734  2.15833582  2.31459993
#> 94  -0.217447012 -0.42132436 -0.05536965
#> 95  -0.705449992 -0.97335625 -0.06462811
#> 96   2.273053911 -0.16028480 -0.73887194
#> 97  -0.876837987  0.32114311  0.85310869
#> 98   0.445734813 -0.25545636 -0.29347104
#> 99   0.182226359 -0.63273791 -0.75862041
#> 100  0.371604412 -1.60827705 -0.65217738