Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A           B            C
#> 1   -1.571447845 -1.36611930 -2.009242777
#> 2   -0.836898940 -0.54496967 -1.133694026
#> 3   -0.906521858  1.18347670  1.094884677
#> 4    0.855478256 -1.79245192  1.318847015
#> 5    1.439932898 -0.72287787  0.265794657
#> 6   -0.201922300 -1.83347775 -0.945452632
#> 7   -1.342198476 -1.77343184 -0.036002667
#> 8   -2.503247049 -0.50880220 -2.292119875
#> 9    0.325813263 -0.19704109 -0.462879900
#> 10  -0.465070273 -1.17947625 -0.336502902
#> 11  -0.597491667  1.02700171  0.041802540
#> 12  -0.874127866 -0.33028466  0.415435905
#> 13   2.275399839 -1.50273694 -0.014783975
#> 14  -0.328093866 -1.05192169 -0.138259356
#> 15   1.671069700  0.96763249  0.571630218
#> 16   0.281736634 -0.31927989 -0.981331608
#> 17  -1.823946420 -0.10169715  0.274230588
#> 18  -1.265341933  0.17395552  0.812116982
#> 19  -1.709544766  1.00298976 -1.070648149
#> 20   0.720055566 -0.63647302  0.418471038
#> 21  -0.485089844  0.47468713  1.511566117
#> 22  -0.463149357 -1.65774709 -1.362710852
#> 23  -1.980640692  1.02422316 -0.439548854
#> 24  -0.517132969 -0.93898795  0.516325457
#> 25   0.559254743 -1.07849616 -0.129129128
#> 26  -1.269341925 -0.58423782 -2.536144199
#> 27   0.995577480 -0.14759260  1.917283805
#> 28   0.692102343  0.57688449 -0.632738139
#> 29   0.375843252  0.27264392  0.392561583
#> 30   1.062182825 -1.72875086 -1.112844742
#> 31   1.176003578  0.22448501  1.238946776
#> 32   0.180246971  0.52739421  0.720376687
#> 33   0.179291756  0.39277138 -0.193028398
#> 34   1.999468815 -0.13745158  0.877626806
#> 35   0.293567657  0.31057976  0.577498170
#> 36   0.235565490  1.11298505 -0.049791422
#> 37  -0.115216931  0.23026395 -0.170459759
#> 38  -0.291687831 -1.46436152 -0.272303194
#> 39   0.220889955  1.44356372  1.516639607
#> 40   0.635207313 -0.15592185  1.799305821
#> 41   0.484490740 -0.36915044  0.243474887
#> 42   0.454450071  0.70273522  0.746435293
#> 43   0.377182760  0.03168418 -0.125540801
#> 44   0.086754972  0.15109750  0.905983550
#> 45  -1.376446160 -1.48378859 -1.154079531
#> 46   0.175688896  1.01510674  0.302192131
#> 47   0.386391672  2.14025780  1.104799733
#> 48   0.296152126  0.09698998 -0.268941487
#> 49   1.009819489 -0.51411491  0.909939757
#> 50   1.672106687 -0.31502558 -0.322995570
#> 51   1.110158363 -1.74988941 -0.524644522
#> 52  -0.001352057 -1.42757769 -0.804103995
#> 53   0.028958334  0.43414119  0.901281983
#> 54  -0.216915757  1.23444215  0.069255223
#> 55  -1.525650530  0.19566855 -1.023535184
#> 56  -0.969235003  0.01558386 -0.397303290
#> 57   0.318251524 -1.05600263 -0.977492433
#> 58  -1.041520678 -0.70006714  0.153549500
#> 59  -0.444549879  1.36137587  0.007061149
#> 60   0.446601599 -0.57971930  0.626171337
#> 61  -0.915588290 -0.81711309 -1.288041881
#> 62  -1.029688905  0.05554335 -0.164555021
#> 63  -0.354023486  1.18384142 -0.079491794
#> 64   0.705071543  0.80055465 -0.379374651
#> 65   2.502972534  1.82163418  1.472100812
#> 66  -0.640199810 -0.11495555 -0.802232958
#> 67   0.521183202  2.65102450  1.210316220
#> 68   0.223809034 -1.56394634 -2.805247616
#> 69   0.960935885  0.60197079 -0.651675680
#> 70  -0.063739299 -0.34760509  1.360287043
#> 71  -0.139082568 -1.76962898 -1.581272077
#> 72   1.736981616  0.36293261  0.260639150
#> 73   0.258048499  0.87480951 -0.113325861
#> 74   0.699277088 -0.18036862  0.275840809
#> 75  -1.557120973  0.78635175 -0.462898787
#> 76   0.331973056  0.88617641  0.940049114
#> 77  -0.081732796 -1.96514316 -0.966533996
#> 78   0.925550492  0.27216374 -0.419066235
#> 79  -0.991760774 -0.62872592 -0.235273568
#> 80   1.064167614 -0.54930856  0.387196419
#> 81   0.014160808  1.21647693  1.189595126
#> 82  -0.614811341  0.55141310  1.612716985
#> 83   1.226516124 -0.44293223  0.555935213
#> 84   1.439275057  0.40603088  0.277442401
#> 85   1.226558625  1.84790563  0.203227865
#> 86   1.157322755 -0.02307746  1.653510716
#> 87   0.447523765 -0.33714715  0.558048187
#> 88   0.294764679  2.03929257  2.587345793
#> 89   0.451905030 -0.63817446 -0.935651405
#> 90   1.457819465  0.61363501  1.225994122
#> 91  -0.776624134  1.09506404  1.386549345
#> 92  -1.044707190  1.88864131  2.040202510
#> 93  -0.146130634 -1.21348230  0.712535253
#> 94  -2.101450462 -1.41894556 -0.308589039
#> 95   0.925150106  0.13968747 -0.297253528
#> 96   0.027635948 -1.48633437  0.120947305
#> 97  -1.122142662 -0.66878969 -0.183181381
#> 98   0.117807239  1.27957420  0.702283194
#> 99   0.538244690  1.66795644  1.966258567
#> 100 -0.353987732  0.37234975  0.077933421

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A             B            C
#> 1    1.02415885 -0.5948118518 -0.893447815
#> 2    1.26891948  0.2768817063 -1.246728776
#> 3    0.28712771  0.8755254950  0.478088893
#> 4    0.03055967  0.5020014679  0.314654674
#> 5   -0.20744568  1.2550827556  0.000620406
#> 6   -0.04319489 -0.6874885999 -0.476544758
#> 7    0.92514215 -0.1980124333 -1.160369434
#> 8    1.87401875  0.8267748630  0.205159131
#> 9    0.19591894 -0.1051203479  0.599843479
#> 10   1.71180202 -0.2197975835 -1.631200912
#> 11   0.12771518  0.1665588763 -0.121716999
#> 12   1.10041745 -0.9761623512 -1.292361732
#> 13  -1.23800442  1.0848302271  2.364145801
#> 14  -0.65856693 -0.3769007754  0.596460149
#> 15  -1.37932554 -0.0394376721  1.678694269
#> 16   0.61923919  0.1489934921  0.021523665
#> 17   0.65348450  1.1071908668  1.047696190
#> 18   0.68751052  1.0764413977  0.372612858
#> 19  -0.25551052 -0.5822168229 -1.012250437
#> 20  -1.10433749 -2.5897629312 -2.247631501
#> 21  -0.26020862 -0.8543432046 -0.201952495
#> 22  -0.76349170 -0.7693691231 -0.334406865
#> 23   1.49433984 -0.5133410682 -1.747325176
#> 24  -1.28712480  0.4571478081  0.523597643
#> 25   1.24223269 -1.0414474309 -0.868000989
#> 26   0.13326422  0.4589705533  0.589112813
#> 27   0.31248168  1.0565818517  0.321952557
#> 28   0.41045896 -1.6030609751 -1.251893641
#> 29  -1.27593878  0.8373394677  1.078990838
#> 30  -0.53734991  2.2324049196  2.025751486
#> 31  -0.91511254  0.5576640316  1.177708063
#> 32  -0.88681727 -0.5117048531  0.619315075
#> 33   0.90027066  0.2004489955 -1.030447515
#> 34   1.69309626  1.6259386597 -0.110563132
#> 35  -0.61619821  1.0287049582  0.604671957
#> 36  -0.68338925 -0.7439803743 -0.555716942
#> 37  -1.68992124  0.2741226366  0.873759163
#> 38  -0.82024756  1.5498194820  1.022459275
#> 39  -0.80017071 -0.0002467184  0.462661335
#> 40  -0.42170988  1.0563974677  1.436259806
#> 41  -0.83943964  0.4032177971  0.696062653
#> 42   0.82744333  0.6304594167 -0.743517049
#> 43   1.24663041 -0.7872642193 -1.474043510
#> 44  -1.02111508 -0.0877249723 -0.384147040
#> 45   1.59269187  0.5324304176 -0.872448001
#> 46  -0.23717807 -1.4315441894 -0.025924341
#> 47  -0.78753759 -1.1029511091  0.293503940
#> 48  -0.94036231  0.3083480701  1.263196831
#> 49  -0.05899490 -1.1331477524 -0.792778742
#> 50   1.67080988 -0.1124625210 -0.620714801
#> 51   2.19656029  0.7534723467 -0.119940653
#> 52  -0.73323871 -0.8603388879  0.399969738
#> 53   0.52086233 -0.6909977241 -0.386193473
#> 54  -0.40271085  0.3697077608  0.301964581
#> 55  -1.59380199 -0.9567253647  0.126650542
#> 56  -0.74272152  0.3848043255  1.234480992
#> 57   1.28144214 -0.1642350881 -0.544266650
#> 58   0.81406012  0.1775255400  0.276137262
#> 59   0.39131449  0.7493937575 -0.699149838
#> 60  -0.61820373 -0.7622561754  0.290148152
#> 61  -2.73417106 -1.6055875200  1.563259789
#> 62  -0.49412421 -1.2558148016 -1.007364163
#> 63   1.20478032  0.0302061256 -2.134679460
#> 64   0.09685578 -1.1780556110 -1.647675488
#> 65  -0.50269571 -0.3490978739  0.333231378
#> 66   0.57660520 -0.3156260110  0.046582548
#> 67  -0.37900745  0.3737365856  0.595824285
#> 68   0.89128176 -0.5632744394 -1.838061221
#> 69   1.21391293 -0.8507031900 -1.573593749
#> 70   0.39644286  0.8613545191  0.759825919
#> 71   0.82845509  0.1270391553 -1.078521495
#> 72   0.28412992 -0.1788162359 -1.076977231
#> 73  -0.66828827 -0.2220635046 -0.475602134
#> 74  -0.61348230 -0.6350246416 -0.185028254
#> 75   2.59791139  0.1594384511 -0.639322772
#> 76   0.35455888  0.5306076458 -0.691926130
#> 77  -2.82671880 -0.7601608155  0.596657953
#> 78   1.28518963  0.8784730932 -0.894544091
#> 79   0.45104850 -0.1541941549 -1.457130222
#> 80   0.46695698  0.0369788602 -0.353997949
#> 81  -0.16377633 -0.4576127909  0.532863052
#> 82  -1.12697647  0.5403754400  1.384919448
#> 83   1.61090007  1.3373887322 -0.121653975
#> 84  -0.07764590 -0.9454845501 -1.062355372
#> 85   1.61297985  1.3049899377  0.346871821
#> 86   1.34979235  0.4087748206 -0.116553903
#> 87   0.22130594  1.9375129659  1.065526619
#> 88   0.51525727 -0.6172891060 -0.465943713
#> 89   0.31329732  0.8912391087  0.273571403
#> 90   0.31319164 -0.5663994889 -0.827388501
#> 91   0.11934529 -0.2942145457 -0.568671295
#> 92   2.63391082  0.9118778108 -1.369701326
#> 93  -0.40267360 -1.6673173057  0.046546011
#> 94  -0.01176184  0.4528268415  0.128241084
#> 95   1.18285394 -0.0899197682 -1.740380327
#> 96  -0.13058646  0.2570966321 -0.754004607
#> 97   1.16393990  2.0339238247  0.373112912
#> 98   1.09201046 -0.5707726739 -1.332910448
#> 99  -0.17740672  0.8850352952  1.783621073
#> 100 -0.02991574 -0.2751252867  0.099543989