Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                 A            B            C
#> 1   -0.0644227201 -0.245667309  0.867276933
#> 2    0.4213094959  2.474510498  0.337500151
#> 3   -0.7246049022  0.559966700  1.018630164
#> 4    0.1815249836 -0.475915883  0.203264807
#> 5    0.8189776358 -0.478819406  0.395174913
#> 6    0.6577919911 -0.188292694 -0.941177410
#> 7   -0.4020506982  1.370680833  0.511254872
#> 8    1.4508218909 -0.453584834  1.255282338
#> 9    0.5136988739  0.461293901 -0.607311960
#> 10   1.4681726684  0.814181897  1.978942589
#> 11  -0.9936891768 -0.261777895  0.325850906
#> 12   0.8269879888 -0.684417931 -0.246700030
#> 13  -0.0329291087  0.483679003  1.121796017
#> 14   0.7648621694  0.757716151 -1.180590951
#> 15   0.9235064468  0.221018178  1.875383509
#> 16   0.4347385310 -0.755290479 -0.105261077
#> 17   0.4851972999  0.657045406  0.853570291
#> 18   0.5532687996  0.945370349 -0.526885996
#> 19   1.2871405754  0.541337500  0.615991383
#> 20   0.3356501033 -0.720320686  0.647128450
#> 21   0.5831526382  0.473148098  0.437007751
#> 22   1.5152256630  1.756082170  2.586626279
#> 23   0.3222319923  1.967380315  0.652808231
#> 24   0.2736219195 -1.392362049 -0.543835855
#> 25   0.9688477425 -2.345907396 -0.426443757
#> 26   0.2994126975  0.762022186  0.401132111
#> 27   1.0434249177  0.100661022 -1.385445111
#> 28  -0.1746839272  0.292562573  0.253555238
#> 29   0.0236333386 -0.259609460 -0.675982289
#> 30  -0.0006627117 -0.637194326 -0.370178562
#> 31   0.3714362581  0.756214139 -1.145099715
#> 32  -1.1655325477 -0.178160280 -0.283563931
#> 33  -0.8637870232 -1.085716092  0.320078536
#> 34  -0.2794585438 -0.086712716  0.537973555
#> 35   2.1578442318  0.009971792  1.392687847
#> 36  -1.6566758132 -1.432596493 -1.199351685
#> 37  -0.5527334689 -0.374445564 -1.000531052
#> 38  -0.4079509421 -0.862321274 -0.255539858
#> 39   0.5616296836 -0.477559762  0.003004546
#> 40  -0.7869504267  0.145351491  0.935860673
#> 41  -2.0673322062  0.083975843 -1.082291400
#> 42  -0.0644848410 -0.218594051 -0.868087596
#> 43   0.2811245244  0.996829229  0.630168193
#> 44  -0.3324426618 -0.470925380 -0.110455682
#> 45   1.6820921781  0.576158882  0.151373135
#> 46   0.6872546390 -1.715086434 -0.379252812
#> 47  -0.8828691538 -1.633712505 -1.450300641
#> 48   0.9169962870 -0.835972001  0.306201961
#> 49   1.0985168359  1.351639359  1.883320952
#> 50   0.5115312027 -0.183037220  0.089865404
#> 51   1.1558819043  0.461333957  0.145766982
#> 52  -0.5372402757 -1.038999636  0.068313865
#> 53  -1.1428558125 -0.867335233 -1.385492976
#> 54  -0.3312183372  0.457054354  0.419579369
#> 55  -2.0226351622  0.985395611  1.305112489
#> 56  -0.7453856387  0.374684787  0.872574035
#> 57   0.4826591970 -1.240275225 -0.414695468
#> 58  -0.6805929219 -0.211888399 -0.589667879
#> 59   2.1745786101  0.222151883  1.788864051
#> 60   2.7266794876 -1.536032470  0.739716859
#> 61  -1.6326748258 -0.650216733 -1.364124767
#> 62  -0.6874288426 -0.579319744 -1.872397134
#> 63   0.4386559277  1.853525121  1.332200572
#> 64  -0.1557323757  0.384590546 -0.131527442
#> 65  -0.0682234532  0.076334033 -0.962634834
#> 66   0.2556472196 -0.790404606 -1.051831754
#> 67  -0.2493980825  0.076033500  2.811021354
#> 68   0.5933604180  0.055022054  0.151644658
#> 69  -1.1989437651  1.316235655  0.270881522
#> 70   0.0334184906  1.369319474  1.363133536
#> 71   2.1267419306 -2.569839629 -0.565655453
#> 72   0.1038866329  0.408314371 -0.970913968
#> 73   0.7248214348 -0.532962677 -0.338367755
#> 74   0.1852280967  0.555654476 -0.953541155
#> 75  -1.2919875727 -0.518599203 -1.207852581
#> 76  -0.2866629275 -1.702810667 -1.183937430
#> 77   0.4744782926  0.051450031  0.163318967
#> 78  -1.3158902828 -1.259177829  0.015702705
#> 79  -0.2807041035  0.054117026  0.425813618
#> 80   0.3572543911  0.758505085  0.150558688
#> 81  -1.0410279539  0.194282441 -1.350397921
#> 82   0.1776602162 -2.409352088 -1.921549909
#> 83   0.4732490804  0.708889116  1.723936274
#> 84   0.5130439750  0.064721171  0.584174362
#> 85  -2.4965822575 -1.055285482 -1.237478751
#> 86  -0.3134352121 -0.272815970 -0.063249882
#> 87   2.0546351332  0.448871798  1.024127723
#> 88   0.8025793941  1.978047976  0.835016405
#> 89  -0.2770826461  0.787582823 -0.109682974
#> 90  -0.0582257377 -0.642360710  0.277543454
#> 91   0.6380038335 -0.437932788  1.077665950
#> 92  -1.3866901690 -0.088803811 -1.461348164
#> 93  -0.7852872405  0.527652978 -2.123411308
#> 94  -1.1928218988 -0.579609357 -0.672580697
#> 95  -1.6267209713  0.375783596 -0.586445656
#> 96   0.4899559950  0.851824773  1.632421039
#> 97   1.5624475860 -0.664550421  0.747078058
#> 98   0.2201577554  0.740444909  0.778928652
#> 99   1.3811181098 -1.468245508 -1.025102927
#> 100 -0.6151742674  1.831542078  0.843081781

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>                 A            B           C
#> 1   -0.1124147112 -0.757324634  0.05299078
#> 2   -1.5100683943  0.443196520  1.73336398
#> 3   -0.3583000330  1.235489911  0.21492851
#> 4   -0.3702908794 -1.098422769 -0.73033678
#> 5   -0.9728970082 -0.863786506 -0.69124962
#> 6    0.4305219165  1.490019126 -0.20766713
#> 7    1.9708888025  0.810636918 -0.65039265
#> 8   -0.9966314196 -0.041377803  0.52499204
#> 9   -1.1494293761  0.852194758  1.06196148
#> 10  -0.9420716711 -0.229590564 -0.93452494
#> 11  -0.3739151414 -1.203608181 -1.23261861
#> 12   1.0514659447  0.549970753  0.72168055
#> 13  -0.1767028542  0.977779771  0.87149872
#> 14   0.1153677507 -1.150380474 -0.48273292
#> 15   0.4508872837 -1.989813270 -2.27143826
#> 16  -1.5984260124  0.523495555  1.88951386
#> 17  -0.8101410430 -0.482901710  1.29392445
#> 18   1.2455612315  0.204325988 -0.32695561
#> 19   0.5560206246  0.913254492  0.24512098
#> 20   0.3295498106 -0.652083811 -0.40882743
#> 21   0.2207267431  0.833773317 -0.03731823
#> 22  -0.0975742505  0.401307510  0.27627459
#> 23  -0.3998565652  0.678420084  0.96752487
#> 24   0.6832974396  1.928239424  0.92451061
#> 25  -0.9739322566 -0.361867838 -0.20554481
#> 26  -1.2423494047  0.242611116  1.46303430
#> 27   0.5127340454 -0.628233815 -0.47495260
#> 28   0.5905271383 -0.683045605 -1.03173527
#> 29  -0.4430298368 -0.205803866 -0.28889067
#> 30  -0.2117095526 -0.276475386  1.02607942
#> 31  -0.3364338524 -1.550052316 -0.88590133
#> 32   0.0534337347  0.007098220 -0.25475444
#> 33  -0.7030917500 -1.386771467 -0.25978887
#> 34  -1.0965342937 -0.223672997  0.28542854
#> 35   0.4369282184 -0.011503896 -0.77173193
#> 36   0.0722512806  0.963665815  1.29137438
#> 37  -1.1788785085 -0.811153787  0.28455839
#> 38   1.5198859086  0.385934718 -1.16551680
#> 39  -0.4865439920  0.236174672 -0.88095065
#> 40  -0.0009858683 -0.563511645 -0.22343611
#> 41  -1.0243633991 -0.032555227  0.97210219
#> 42  -0.1119778015  0.755216899  0.43416179
#> 43   0.3258900023 -0.548979792 -0.44691986
#> 44   1.1454737015  1.968998736  0.64224544
#> 45   0.8829000516  1.010911695 -0.27380555
#> 46   0.4031274854  0.186786311 -0.21113819
#> 47   0.3376679125  0.815714916  0.36415258
#> 48  -1.5491199913 -1.419609153  0.52405382
#> 49  -0.7973193552 -0.277664762  1.38794853
#> 50   0.8521555003 -0.373914905 -0.73470686
#> 51  -0.3203799437 -2.492277811 -1.96655432
#> 52  -0.5076069203 -0.418470655 -1.09527831
#> 53  -0.1632686004  0.103278025 -0.04148546
#> 54  -0.8435546281  0.067310616  1.92574201
#> 55  -0.4491679807 -0.700543648  0.33190176
#> 56  -0.6325785747 -0.383812129  1.20570436
#> 57  -0.2315285779 -0.336869685 -0.13200552
#> 58   0.6630188669 -1.618177474 -1.07746093
#> 59   0.0973843244 -0.043025049  0.90061400
#> 60   0.2919381752  0.651426483  0.05948885
#> 61  -0.2069872243 -0.955693728 -1.54083049
#> 62  -0.9036800949  0.967298756  0.89764821
#> 63   1.4862535182  0.203867700 -1.43775486
#> 64   0.3712477661  0.715227630 -0.52044995
#> 65  -0.8502380617  0.697345295  1.43309862
#> 66  -1.4416272428 -0.372855437  0.89363472
#> 67  -1.3650463778 -0.008862178  0.99329792
#> 68   0.1118320173  0.599898849 -0.12116995
#> 69   0.6774426443 -0.567921046 -0.54860285
#> 70  -0.4355247251  0.556259360  1.30000172
#> 71   0.6843675344  0.854324270 -0.18024177
#> 72   1.3279606447  0.079045667 -0.98416476
#> 73   0.5128539062  2.291915014  0.63592161
#> 74  -0.4394384005 -1.087035858 -1.11753898
#> 75  -0.6014962576 -0.707945131 -0.83212399
#> 76   0.3513337482  0.201124799 -0.46920764
#> 77  -1.1577215281 -2.345068927  0.18774797
#> 78  -0.1584728661 -1.845172132 -1.04806255
#> 79  -0.0743556783 -1.151625110 -0.43302612
#> 80   1.1397656849 -0.108196644  0.77508245
#> 81  -0.5703303927 -0.259382833  0.81629194
#> 82  -1.0252172445 -0.976113177  1.79413235
#> 83  -1.4561439636  0.171094165  0.73205207
#> 84   0.8996032967  1.017105665  0.19475979
#> 85  -0.4679058793  0.982533109  1.16424215
#> 86  -0.5562110515 -2.222413608 -0.10467253
#> 87   0.1170414328  0.864432892 -0.04333673
#> 88   0.5543999686 -0.928561539 -1.98845867
#> 89  -1.0715501704  0.113560567  0.67194965
#> 90   1.9881415161  0.438104917 -1.50996241
#> 91   0.0886907549  0.127672220 -0.23737867
#> 92  -0.4788706240  0.553377993  1.68493704
#> 93  -1.2560154875  1.529162576  1.42833610
#> 94   1.2370271054  0.472590230 -0.65101958
#> 95  -0.8281841712  0.355517306  0.56035766
#> 96   0.4338328097  2.428942208  1.49903136
#> 97  -1.5344182463 -1.235670969  0.55562311
#> 98   0.8518001083  1.192673806 -0.29468224
#> 99  -0.1887176941 -0.344850170 -0.09630699
#> 100 -0.0937423698 -0.596414587 -0.14047241