Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A           B            C
#> 1   -0.229205984  0.05434050 -0.082553025
#> 2   -0.184527852 -0.93898438 -0.472243042
#> 3    1.774870285 -1.08730473 -0.468335108
#> 4    0.711697535 -1.80634706 -1.037812399
#> 5    1.851906397 -0.36271461  0.598864450
#> 6   -0.618183054 -0.76118626 -1.013083828
#> 7   -0.885712391  0.58694959  0.794675360
#> 8   -0.493426953 -0.39790341  0.721515393
#> 9   -1.539404041  0.88546995 -0.347348154
#> 10  -0.888815759 -1.09732866  0.475757067
#> 11  -0.811332904 -0.55588347 -1.156501516
#> 12   0.963177799 -0.12634264 -0.500906544
#> 13   1.429600007 -0.41240522  0.428424826
#> 14  -2.078066625 -1.75723116 -1.245219101
#> 15  -0.267036865  0.58265833 -0.266579289
#> 16   0.400814145  0.02981542 -0.204969498
#> 17   1.313266217  1.45659965  1.378008751
#> 18   0.082581352  1.41456699  1.424197755
#> 19   0.799312688 -0.71624375  0.659989766
#> 20   1.462930285  0.59972011  1.782443200
#> 21  -1.103901377  0.42338385  0.526183369
#> 22   0.080641625  1.18051884  0.594650491
#> 23   1.909827336 -0.37343729  1.524552081
#> 24  -0.926947492 -1.55700812 -1.289111670
#> 25   0.391537807 -0.72086099  0.704712075
#> 26  -0.733548533  0.59986117 -0.880396180
#> 27  -0.441864833  0.24711261 -0.981775815
#> 28   0.663563215 -1.00997229 -0.443379376
#> 29   1.628911525  2.31542505 -0.078466949
#> 30   0.416153203 -0.22480573 -0.129696432
#> 31  -0.198406283 -0.21381195  0.459902257
#> 32  -0.336901474  0.15895073  0.431111555
#> 33  -1.499069554  0.65520262  1.109300923
#> 34   1.113213030  1.32560376  1.946316756
#> 35  -1.061614031 -1.42265538 -0.134081489
#> 36  -0.370089288  0.61543353  2.379606801
#> 37  -0.395437116 -1.12929936  0.541651107
#> 38   2.573428975  1.01460134  0.722176033
#> 39  -0.123777990  0.37296704  0.537420809
#> 40   0.562207037  0.31896048 -0.490246708
#> 41   0.543004206  0.29237624 -0.515795801
#> 42  -0.716396047  0.69632666  0.898770316
#> 43  -0.312415421  0.91827329 -0.432473348
#> 44  -0.135941268  0.82894206 -0.454113331
#> 45   0.949396143  0.06304565 -1.039258267
#> 46   1.953316500  0.67060846  1.321584603
#> 47   0.682821328  1.36864548 -0.234276987
#> 48  -0.604050327  0.54954318  0.414525998
#> 49  -0.408811826 -0.07770751  1.930383862
#> 50   0.677769227  0.24811916  0.423623416
#> 51  -0.000625454  0.37706180 -0.591371360
#> 52   1.312733342  2.12755373  0.330413805
#> 53  -0.495726986  0.85092912  0.095128314
#> 54  -0.540134175  0.06218760 -0.786976763
#> 55  -0.633937719  0.29941125 -0.500801522
#> 56   1.977196777  0.31725869  0.767946450
#> 57  -0.853260816 -0.62406628 -0.004995766
#> 58  -0.695121107  1.07234684 -0.538742596
#> 59   0.191054474 -0.64259837 -1.395378355
#> 60  -0.006133827 -0.56303899 -0.227726784
#> 61   0.037274666  0.47327142  1.025043898
#> 62   1.127542090  0.68096898  1.245666138
#> 63  -0.044077640 -0.22225146 -0.784633699
#> 64   0.219821264  0.43143242 -0.026825319
#> 65  -1.369138988 -0.51806073 -0.072572833
#> 66   0.558319704 -0.87612289  0.890105660
#> 67   0.060427529 -0.06920719  0.465574078
#> 68   0.800880241  0.31835631 -0.340131122
#> 69  -0.084591679  0.56081983 -0.442297726
#> 70   0.446423639 -0.26530224 -0.569293366
#> 71   1.683655880  0.35191468  1.415494918
#> 72   0.292284875 -0.80131010  0.671678074
#> 73   0.100679277  3.01654198  0.455208611
#> 74  -1.659557331 -0.19323227 -1.003504201
#> 75   2.135951630 -1.05293581  0.508994672
#> 76   0.493837279  1.58531513  2.011668364
#> 77   0.393288560  0.38860807  0.099845063
#> 78   0.763664836  0.32342208  2.278734150
#> 79   0.022328609  1.57208052  1.288056086
#> 80  -0.131694955 -0.21660039 -0.076048470
#> 81   0.710006755 -0.16279609 -0.095180442
#> 82   0.665741831  0.43117211 -0.680340186
#> 83  -0.484278351 -0.20185679 -0.894753900
#> 84  -1.613613927 -0.85336194 -1.595792247
#> 85   0.893904648 -0.50150386 -0.509983523
#> 86   0.372569332  1.88938138  1.563377987
#> 87   0.472513893 -0.72726291 -0.527165147
#> 88  -0.332553020  1.05836928  1.024044777
#> 89   1.772049201 -0.70422424 -0.536590322
#> 90  -1.051259583 -0.80453690 -0.672675606
#> 91   1.551243058  1.74473178  2.085065241
#> 92  -1.168888820 -0.64552200 -1.522801324
#> 93   0.128160467 -0.42700006 -0.348402908
#> 94   0.970524630  0.20743860 -0.450721885
#> 95   0.691202687  1.07128913  0.983240154
#> 96  -0.313475645 -1.30283076 -0.773123229
#> 97  -0.869860138 -2.42274141 -0.809479958
#> 98  -0.511647556  0.25277315 -1.287321708
#> 99  -0.248815429  1.10042255 -0.651653190
#> 100  0.034289387 -0.05949643 -1.250273528

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A            B           C
#> 1    0.09582221  1.324436356  0.12819037
#> 2   -0.73238866 -0.977397537  0.17501732
#> 3   -1.42876400 -0.203915497  0.26580375
#> 4   -0.92105639 -1.374874344 -0.27226497
#> 5   -1.28076527 -1.884182758  0.15723537
#> 6   -0.61031809  0.872940302  1.05697983
#> 7   -0.74279796 -0.105383303 -0.82618123
#> 8   -1.70735226 -0.745824266  0.83023161
#> 9    0.88834496 -1.217086651 -0.43065938
#> 10  -0.27803295  0.265010301  0.09346848
#> 11  -0.34725869 -0.852127285 -0.25569864
#> 12  -0.56098259 -0.267873561  0.79176376
#> 13  -1.16384169 -1.965887790 -0.74138662
#> 14   0.51684322 -0.753379860 -0.26144697
#> 15   0.73020506  0.649222919 -0.05754758
#> 16  -0.22911879  0.318659956 -0.01472596
#> 17   1.20552898 -0.317186409 -0.86623311
#> 18   0.35384708 -0.319411345 -1.42169981
#> 19  -0.71831357  1.440833014  1.00968048
#> 20   0.27873475 -0.310169215 -0.55871557
#> 21   2.51360812  1.520651608 -0.25309508
#> 22   1.47248861 -0.327661912 -1.08019894
#> 23   1.37274664 -0.192763564 -0.53725689
#> 24  -0.48245689 -0.871541913 -0.75833717
#> 25   0.35364336 -0.523035451 -0.41050241
#> 26   1.18159390 -0.685622098 -0.80959348
#> 27   0.75044863  0.097551142  0.03530543
#> 28   1.59850419  0.636013185 -0.12860216
#> 29   0.30148888  0.871480432 -0.38883234
#> 30   0.15639733  1.673131787  1.23538850
#> 31   0.48302565 -0.447200971 -0.45818470
#> 32  -0.49970090  0.021896158  0.74741645
#> 33  -0.41860875 -1.210795723 -1.10977968
#> 34   0.03945736  0.918777239 -0.22126797
#> 35  -2.42506869  0.603979807  2.01545869
#> 36   1.43774595 -0.048097499 -0.58663011
#> 37   1.74285336  1.723755757  0.87922754
#> 38   1.11173035  1.324708000 -0.53942615
#> 39   0.72135545  1.905817873  0.51313516
#> 40   1.15240876 -0.486562049 -0.58319454
#> 41   0.24027021  1.114882877  2.39749656
#> 42   0.84386288 -0.803728193 -0.88302722
#> 43  -0.16538262  0.544137184 -0.08371529
#> 44   0.94680574 -0.568273275 -0.52578505
#> 45   0.66743829 -0.123859206  0.03372766
#> 46   0.80720177 -0.030878435  0.35961259
#> 47  -0.69413671  0.518223546  0.97047705
#> 48   0.87832661  1.467853640  0.30481960
#> 49   1.56164004 -0.212233659 -0.11080752
#> 50   0.21879897 -0.329292348 -0.58771725
#> 51  -0.83096661 -0.164153462  0.97145938
#> 52   0.52299490  2.264032892  0.83349716
#> 53  -0.37801612  0.417217756  0.86021790
#> 54   0.50176609  1.811551110  0.74518358
#> 55  -0.20921092  0.653426308  1.11571061
#> 56  -0.27533766  0.377114289  0.12279626
#> 57  -0.03338213  1.289800292  1.90426723
#> 58  -0.67270287 -1.215765596 -1.11085888
#> 59  -1.21403786 -1.160170154  0.34083363
#> 60   0.35873937 -0.284142414 -0.35771771
#> 61  -0.71219407  0.161266288  0.93371112
#> 62  -0.56411190 -0.437131322 -0.26588409
#> 63  -2.02054007 -1.896749548 -0.44780410
#> 64  -1.42318185  0.604708692  1.47451971
#> 65   1.21894183  0.616575657  0.17500055
#> 66  -1.08097585 -0.300973859 -0.07493680
#> 67   1.16589166  1.162022796  0.09468440
#> 68  -0.14460181  2.133207445  1.31720171
#> 69  -0.48442259 -1.275573105 -1.03895269
#> 70  -0.40613392 -0.021043826 -0.32117864
#> 71   0.36830170 -0.914272955 -0.13503940
#> 72  -0.80083022 -0.974182923  0.60553731
#> 73   1.68313899 -0.006179412 -1.24954755
#> 74   0.70722901  1.558807108  0.50643957
#> 75  -0.26133221 -0.494290050 -0.80385825
#> 76  -0.23203674 -0.897419467 -0.79274778
#> 77   1.28210764  0.052618233 -0.78621470
#> 78   0.97590539  0.319700652 -0.68570073
#> 79   0.81856268  0.374494778 -0.99209296
#> 80   0.28221446  0.928337695  0.35525471
#> 81  -0.91329259 -2.539930323 -1.54081552
#> 82  -1.21977391 -1.384454494 -0.48846240
#> 83   1.21353406  2.501383048  0.73699257
#> 84   1.87283440  1.555226955  0.12474182
#> 85  -0.16858292  0.821405354 -0.28271338
#> 86  -0.10789465 -0.559029228 -0.43435570
#> 87   1.70685619  0.087394125 -1.22643014
#> 88   0.49880322 -2.576299896 -2.19278141
#> 89  -0.42861085 -0.342466028 -0.20747546
#> 90  -0.51554734 -2.103746169 -1.67091452
#> 91  -2.78938737 -0.142169770  1.66096610
#> 92   0.52621395 -0.587570848 -1.10853434
#> 93  -0.25038633 -0.178790912  0.18511174
#> 94  -0.02270853 -0.789655673 -0.22343694
#> 95   0.34315840 -0.030783517 -1.65540207
#> 96  -0.60812633  0.315356172  1.29489002
#> 97  -1.36667328 -1.553850635 -0.74238863
#> 98   0.35917706  0.620726235  0.35426672
#> 99  -1.47754191 -0.901284431  1.20335340
#> 100  1.25296693 -0.020265189  0.12243695