Multiple Normally Distributed Vectors

rnorm_multi() makes multiple normally distributed vectors with specified relationships.

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A           B             C
#> 1    0.580063420 -0.09872452  1.0784038131
#> 2    0.090745757 -2.21205847 -1.1273322465
#> 3   -0.853981885 -1.05175966 -0.1363566233
#> 4   -1.211615762 -0.80127979 -0.9921388410
#> 5    0.192619570 -0.83783332 -0.0197866182
#> 6   -0.114386798  0.66023232  0.3335085124
#> 7   -0.062983007 -0.27383834  0.5704329171
#> 8    0.193590660  0.14185678  0.1785742297
#> 9   -0.340948097 -1.43084148 -2.0317941800
#> 10   2.133504111 -0.93541542 -0.6585662765
#> 11   0.396551231  0.52315754 -1.0162089235
#> 12  -0.384779171  0.38412907 -0.8367164291
#> 13   0.561300727 -0.28195229 -0.9540677145
#> 14   0.675929362 -0.93428998 -0.0396729591
#> 15   0.283146696 -2.36093571 -0.4917625244
#> 16   0.280137890  0.67161321 -0.8344114511
#> 17  -0.526655044  0.44061761 -0.0237941410
#> 18   0.408507595  1.26118948  0.2582611048
#> 19  -0.388734127 -0.35053585  0.5105104491
#> 20  -0.244792403 -0.71618395  1.4115771521
#> 21  -1.106792267  0.24557872 -1.8149257666
#> 22  -0.943906760  0.60176480  0.1351054811
#> 23  -0.267375172 -0.29709120  0.1000263582
#> 24   0.452204594 -1.17153794  0.5214464384
#> 25   0.536259704 -1.09615222 -0.5867834183
#> 26  -0.236333920  0.49410299  0.0880169640
#> 27   0.738830878 -0.03269877 -0.0777189754
#> 28  -0.376182790  1.04535590 -1.1190239099
#> 29  -0.405648929  0.69822409  1.5864914799
#> 30  -1.160958357 -1.35009209 -0.9999216075
#> 31   1.006528656  0.61391014  0.3346503048
#> 32  -1.067981697 -0.75249566 -1.1806269747
#> 33  -0.357735257 -0.50998132 -0.9337983791
#> 34   0.788906573 -0.46545444 -0.4438848537
#> 35  -0.398840231 -0.91834381 -0.6787544321
#> 36   1.375238390  0.31725327  2.3153530303
#> 37   1.309298537  1.06693175 -0.0002758011
#> 38  -0.530013223  0.71744745  0.7425554627
#> 39   1.474483534  2.06301787  0.8343343677
#> 40  -1.418679613  0.55912804  0.0881185330
#> 41   1.987764609 -1.66214704  0.3601230636
#> 42   2.552996783  1.74737931  1.3149925548
#> 43   0.524412448 -0.01102079 -0.4011135037
#> 44  -0.235328699 -0.60681908 -0.6471093657
#> 45  -0.552260875  0.48935906 -0.4639006189
#> 46  -0.808378301  1.56206858  0.4561830091
#> 47   0.376584113  0.17787723  0.3772892838
#> 48  -1.523135326 -1.12397467 -0.7134540706
#> 49   0.184684633 -0.30179709 -0.3685977391
#> 50   0.708403805  0.89554929  1.5859329703
#> 51   0.254379268  0.43257257  0.0637306254
#> 52   1.570630317 -1.16270365  0.7844504392
#> 53   0.744131290 -0.76748954 -1.5993367811
#> 54   1.131117657 -0.21915384 -0.5733556287
#> 55  -0.664541255  0.55462340 -1.9073133381
#> 56  -1.232069852 -0.76910659 -1.4118222935
#> 57   0.392813673  0.94396233 -0.4308654690
#> 58  -0.144178008 -0.64698290 -0.6334769218
#> 59   1.164475659  1.45358400  0.9452965010
#> 60  -0.233225524 -0.68819951 -0.4350951565
#> 61   0.062137521  0.31379011  0.7014771458
#> 62   0.637040390  0.58403290 -0.1763327235
#> 63  -0.270598875  0.15623892  1.0077853602
#> 64  -1.481668036 -1.54096019 -2.0061784034
#> 65  -0.919249237 -2.08932046 -0.4806280284
#> 66   0.825289240 -0.24405615  0.2990468946
#> 67  -1.216081227 -0.96032094 -0.0007913805
#> 68   0.206769111  0.24427677  0.0626262792
#> 69  -1.048041344 -1.13750948 -2.2996884331
#> 70   1.527935152  0.97355428  1.7036046674
#> 71   0.250693735 -0.41953515 -0.1767846545
#> 72  -1.293875018  1.76704580  0.9987815774
#> 73   0.390036316  1.64125076  0.5327712176
#> 74  -1.469682995 -1.17649262 -0.2694275683
#> 75  -0.887855165 -0.01672167 -1.2938965114
#> 76  -1.090398999 -0.88722919 -0.8025775367
#> 77  -0.736544396 -0.52134426 -0.2607690775
#> 78  -0.420887465 -0.74300358 -0.1066756230
#> 79   1.318045818  0.06358653  0.5941506545
#> 80  -0.416977414 -0.73062603  1.2852225469
#> 81   0.318201763  0.27278899 -0.4486797764
#> 82  -0.254429189  0.34508490  0.6871103199
#> 83   1.934886554 -1.00513721  0.2076976831
#> 84  -2.219295612 -0.31100630 -0.6100749433
#> 85  -0.178935745 -1.18453086 -0.4916501471
#> 86   0.800414923 -0.87700750 -0.7008907125
#> 87   0.083590913 -2.54822932 -1.3076891318
#> 88   0.810749904  1.74466727  2.3461962818
#> 89  -0.446075261  1.02050310 -0.1750381996
#> 90   0.748926181 -1.34145000 -0.7647587615
#> 91   0.996404056  0.44301890  0.4587844649
#> 92   2.093318974  0.42018888  1.1995588949
#> 93  -1.929648067  0.54120179 -1.4522228536
#> 94  -1.846602547 -1.01815754 -0.7388437472
#> 95  -0.009962785  0.96157126  1.5401153622
#> 96  -0.669352258  0.61491480 -1.2687851047
#> 97  -1.241527096 -1.03666393 -1.1485718522
#> 98  -0.595984914  0.39510322  1.0143729066
#> 99   0.042787189 -0.01241021 -0.5676048913
#> 100  1.578485546  0.94305738  0.4515601245

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>                A            B             C
#> 1    0.484658341 -0.742205822 -0.4449783380
#> 2    1.186786630 -1.000544878 -0.9497220323
#> 3   -0.381113323 -0.644041064 -0.6810782917
#> 4   -1.339846455  0.517809498  1.2797721443
#> 5   -0.181445384 -0.285510466 -0.7970645955
#> 6   -0.242856053 -0.690225584  0.2714375811
#> 7   -1.407509023  0.914543179  1.5745573922
#> 8   -0.608448700  1.085903958  0.6950162330
#> 9    1.006665710  0.230597685 -0.5005835875
#> 10   2.688694138  0.023284165 -1.0438745075
#> 11   0.620829156 -0.209781278 -0.3128392194
#> 12   0.489134064  0.002199792 -0.9675840914
#> 13  -0.484478368  0.254056963  0.6661222175
#> 14  -0.606086667  1.105266125  0.1869479128
#> 15  -0.356848057 -0.145414572  1.0999266440
#> 16  -0.207898361 -1.005676610 -0.3997856883
#> 17   0.696548011 -0.724568456 -1.1672952581
#> 18  -0.128278497 -0.133843063 -1.0397383583
#> 19  -0.668768791  0.413550026  0.6493829976
#> 20   0.639193185  0.466345909  0.2680866712
#> 21   1.198673387 -0.831190728 -1.3807877910
#> 22  -2.233381972 -1.364908356  0.6418456875
#> 23  -0.006388825  0.128843689 -0.0187768989
#> 24  -0.327426061 -0.336877882  0.4006878327
#> 25  -0.901291393  0.323884749  0.9662414449
#> 26  -1.155951524  1.086363905  1.1052114866
#> 27   1.251580671  0.919370611  0.1225804313
#> 28  -0.069191573 -0.396917498 -0.7994340803
#> 29  -0.280314246 -1.215826046 -0.2231116973
#> 30   1.750322871  0.857307988 -0.0007566503
#> 31   0.423765513 -1.626792490 -0.7358762655
#> 32   0.882765629 -0.499647999 -0.7955647900
#> 33  -0.644431348 -0.853866273 -1.0201752962
#> 34  -0.596586156 -1.011250991  0.1999263842
#> 35  -1.850151884  0.217921810  1.2229191838
#> 36  -0.258891740 -0.059479010 -0.0749215286
#> 37   1.261662599  0.765095233 -0.8528689506
#> 38   0.663405728  0.132448812 -1.1795552736
#> 39  -0.888862937 -1.257320157 -0.6709252565
#> 40  -1.429992304 -0.121120191  0.1786727354
#> 41   0.827544340 -0.360732604 -1.3953480577
#> 42  -0.778549707  1.456603829  0.1303537042
#> 43   0.101137027 -1.108538951 -0.7668516668
#> 44  -2.083907549 -1.664642951  0.6758047495
#> 45   0.784731572 -0.357857436 -1.2570525548
#> 46  -0.396700814  0.059989238 -0.2490777923
#> 47  -0.142656675  0.794573153 -0.0222581321
#> 48   0.883357520 -0.086309086  0.0991830449
#> 49  -1.275348327 -0.236850668  0.4418877639
#> 50  -0.063148286  1.335536045  0.7004060631
#> 51  -0.189248763  2.493607894  1.3344436428
#> 52  -0.379607355  1.199938987  1.8268748851
#> 53   2.136125617  0.784385418  0.8771598668
#> 54  -0.133719304  1.254436644  1.8660023148
#> 55   0.055052583 -0.233683471 -2.0902631686
#> 56   0.863391065  0.347070631  0.6429265108
#> 57   1.266501199  1.526597049  0.5495661158
#> 58   0.110814965 -0.709871332 -0.3375972250
#> 59   2.873482195  1.333414640 -1.7372521577
#> 60  -0.121849296  2.057985608  1.0459021110
#> 61  -1.969534427 -1.061842864  1.2104195769
#> 62   0.954160571  0.078773527 -0.4458596260
#> 63  -0.450664181 -0.381147771 -0.9290400544
#> 64   0.659332635  1.078532099 -0.4631530029
#> 65   1.753360709  0.406460694 -0.1594744993
#> 66   0.356176961 -0.182342767 -0.0863782273
#> 67  -0.378471770  0.786100705  0.6331283834
#> 68  -0.544572435  2.104225395  2.6224708911
#> 69  -0.197379421  0.125805280  0.2559255452
#> 70  -0.303661009 -0.706925285 -1.2438850305
#> 71  -0.159819774 -1.354128104 -0.5614660445
#> 72  -0.312157889 -0.090535005  0.6279715098
#> 73  -0.268737731 -2.137525108 -0.0414486934
#> 74   0.100258193 -0.170363547 -1.4314265220
#> 75   1.132417516 -0.083062632 -1.0480427931
#> 76   0.786918873  0.043642424 -1.0352274354
#> 77  -0.001441801  0.957098100  0.5364156134
#> 78   0.874144243 -1.898837326 -1.7762561578
#> 79   0.075263017  1.998466596  1.6697597415
#> 80  -0.522652730 -1.209063287  0.1022117196
#> 81   0.213258376  0.847526255  0.3208845313
#> 82  -0.136484929 -0.809384890 -0.7026234504
#> 83   0.059288270 -0.600432078 -0.7101281773
#> 84   0.737873426  1.018864267  0.0180777342
#> 85   1.076905586  1.299179888  1.4502742459
#> 86   1.069959862  0.644484824 -1.7847490104
#> 87   0.455682746  0.077335425  0.1528263410
#> 88  -1.552552000 -1.605602164 -0.2905896920
#> 89  -0.677481263 -0.086980016  0.7011170010
#> 90  -0.025159957 -0.069020087 -0.2286867488
#> 91   0.333055133  0.345835151  0.0177301819
#> 92  -1.188859716 -0.693241686  1.1911399347
#> 93   0.272511523  0.499988896 -0.1223121118
#> 94  -0.831075700 -0.846499224  0.5899186297
#> 95   0.702452657 -0.048435204 -1.0755290853
#> 96   0.553022585 -0.898320302 -0.6642299409
#> 97  -0.076327485 -0.698772198 -0.2438553542
#> 98  -0.831051326 -0.887911590 -0.4662089586
#> 99  -0.302220904  1.710761438  1.4132701210
#> 100  0.728187681 -1.090645135 -2.2160014589