Multiple correlated normal distributions

Make normally distributed vectors with specified relationships. See vignette("rnorm_multi", package = "faux") for details.

rnorm_multi(
  n,
  vars = NULL,
  mu = 0,
  sd = 1,
  r = 0,
  varnames = NULL,
  empirical = FALSE,
  as.matrix = FALSE
)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>               A           B           C
#> 1   -0.96556081 -1.20169854  0.85090419
#> 2    0.05670290  0.01700991  0.48775399
#> 3   -0.74022579  0.57910138 -0.58020629
#> 4   -0.28569039  0.49000059 -0.22942112
#> 5    0.47885909 -0.11581380  0.90855832
#> 6   -0.78988361  0.42095722 -0.11635434
#> 7   -2.05690176 -0.40400554 -0.56986906
#> 8   -0.78795983 -1.30905229 -0.43509126
#> 9   -0.89771052 -0.04537317 -0.61244957
#> 10  -1.46915932 -1.13631496 -1.36685862
#> 11   1.02366523 -0.97961633 -0.35159906
#> 12   0.96289991  2.25231022  1.43536237
#> 13   0.54622354  0.27222960 -0.48421685
#> 14  -1.43322923 -0.37246608 -1.36428787
#> 15   0.50087537  1.70435722  1.38467500
#> 16   0.90825931  2.42611954  1.52212563
#> 17   0.93114569  0.51407014  0.31449194
#> 18   1.77096168  0.36950022  1.06465284
#> 19   0.12325783  0.01531666 -0.49267538
#> 20  -0.74511082  0.25159046 -0.06429847
#> 21   0.31010632 -0.06621243  0.78332412
#> 22  -0.64948815 -0.45838064 -1.61113102
#> 23   1.88427278  1.56506517 -0.27593725
#> 24   3.59521028  0.75646345  1.17703663
#> 25  -1.04764008  0.42268757 -1.41382919
#> 26  -0.79918046  1.11283566  0.05145472
#> 27  -0.02060352  0.98520919 -0.08423622
#> 28   0.40338859  0.14152516  1.63094422
#> 29   1.68438492  0.06154474 -1.20819740
#> 30   0.66664387  0.69721412  0.82394684
#> 31  -0.10038373  0.36710245  0.23069612
#> 32  -1.54872763 -1.49469838 -1.85258368
#> 33  -0.08272891 -0.30546646 -1.62623308
#> 34   1.50689560  2.97370799  0.78397012
#> 35  -0.84378808 -0.69556788  0.78101403
#> 36   0.29810734  0.80005814 -0.46489722
#> 37  -1.17117725  2.11866526  0.67471508
#> 38   2.11178586  0.68260313  0.21461578
#> 39  -0.42702822 -1.04513560  0.21675210
#> 40   0.04130920  0.72790158  0.20989916
#> 41   0.68885773 -0.21911173  1.19578632
#> 42   0.03700794 -0.12019011  0.93097823
#> 43   1.64474168  0.74439393  0.11570275
#> 44   0.44697610 -0.65033885 -0.48599467
#> 45   0.69584120  0.02342810 -0.56712409
#> 46   0.28496705 -1.92385020 -0.36574603
#> 47  -1.09431260 -0.83853050 -0.87580572
#> 48  -0.36207652 -1.33913852 -0.58618737
#> 49  -1.23335209 -0.85354241 -1.14535486
#> 50   0.75732580  1.91166180  1.47710053
#> 51  -0.02392741  0.76068227  0.40931092
#> 52  -1.37385467 -1.15124202 -1.04976863
#> 53   0.94682442  0.90177436  1.55852834
#> 54  -0.82359977  0.18567568  0.39826548
#> 55  -0.13963610  1.38103353  0.97561976
#> 56   0.98385115 -0.38647794 -0.92172323
#> 57   0.20848761 -1.75420727 -1.27598584
#> 58  -0.45981832  0.66686079 -1.40727059
#> 59  -0.57135104  0.44847507  0.71356447
#> 60   0.86651204  1.67624058  0.98858534
#> 61   0.08031105 -0.41656343  0.93856985
#> 62   2.08933196  0.54654430  1.60638475
#> 63  -1.60447755 -0.36627121 -0.97833372
#> 64   0.53273459 -1.48432167 -0.03494961
#> 65  -0.45843262  1.25243960  0.92721224
#> 66  -1.38383397  0.22738387 -1.42672051
#> 67   1.18526319  0.54364019 -0.36367062
#> 68  -0.91991101  0.32596169 -0.26084435
#> 69  -0.56316267 -0.59833642  0.12652213
#> 70   0.22407768  0.60364018 -0.10616103
#> 71  -0.47149169 -0.17252971 -1.41148520
#> 72  -1.91252886 -0.11365838 -0.82986566
#> 73   0.15719045  1.33019562 -0.66740200
#> 74  -0.18963599 -0.97193293  0.32445673
#> 75  -0.56719062 -0.29084674  0.13618169
#> 76  -0.15504092 -1.57370829  0.04227237
#> 77  -0.93041010  0.60224621 -1.10814457
#> 78   0.87514960  0.15082219 -1.03876136
#> 79   0.24540964 -0.31204164 -1.53188440
#> 80  -0.03622377  0.99488667  0.89443679
#> 81   0.24409628  1.00045992  0.60971192
#> 82   0.45157480  1.36348999  0.80638270
#> 83   0.50237612 -0.08232024 -0.06621848
#> 84  -1.02213769 -0.38886701 -0.48317626
#> 85  -0.11159876  0.14996791  1.34334114
#> 86  -1.02326663 -0.23069496 -0.18434433
#> 87  -0.16564751 -0.03052795  0.33549538
#> 88   0.05911399  1.61427898  1.02173627
#> 89  -1.38313032  0.56764318 -0.19548931
#> 90  -0.87143366 -0.65925939 -1.17910446
#> 91   0.03780991 -0.25296779 -0.02863492
#> 92   1.51529188  0.37627691 -0.13085735
#> 93   0.09605254 -0.36241965  2.23075571
#> 94   0.01865759  0.30602535  1.59346258
#> 95   0.22776115 -1.11977681 -0.44398445
#> 96   0.32296175  1.49928864  0.10421657
#> 97  -0.13032356 -0.75486189 -0.50685795
#> 98  -1.64559603  0.81330489 -0.98422600
#> 99   0.41698022 -0.17079652  0.93075253
#> 100 -0.84470539 -0.01769492 -1.78048873

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A           B           C
#> 1   -0.99468568 -0.53489480  0.34932141
#> 2    0.49265411  0.49118050 -0.06547429
#> 3   -0.68067899  0.68367780  0.07148121
#> 4   -1.49042625 -0.39766274  0.36299837
#> 5    0.61865493 -0.19706894 -0.89999795
#> 6    1.13434731  1.16870276 -0.25104110
#> 7    0.26774085  0.61286654  0.71454158
#> 8   -0.11347291 -0.30570939 -0.45145322
#> 9    0.51033807  0.42260962 -0.23866851
#> 10   1.03926182 -0.30225717 -1.12194597
#> 11  -0.40491250 -1.60618950 -0.34256860
#> 12   0.17652146  1.29386242  0.99459854
#> 13   0.03862015  1.21569251  0.82380830
#> 14  -0.39750387  0.18322683  0.54107841
#> 15   1.67003404 -1.63255544 -1.84217377
#> 16  -1.20377631 -1.00619171  1.05002572
#> 17  -1.89569112 -1.18089416  0.14415519
#> 18   0.20756820 -0.99284204 -2.55000254
#> 19   0.81214672 -1.04738208 -1.41564919
#> 20   0.40574772  0.69331340  0.18568740
#> 21   0.28892329 -1.00733911 -0.86518806
#> 22   1.86256964  0.99809741 -0.90641982
#> 23  -0.76129749 -0.25571155  0.35291284
#> 24   0.25714737  1.30363180  1.15523461
#> 25  -0.67554609  0.29792978  0.06603650
#> 26   2.88218188 -1.35728425 -2.72285480
#> 27  -0.29884051 -0.69147011 -0.69977765
#> 28   1.43038307  2.47993102  1.03811605
#> 29   1.06285074  0.76389893  0.72580259
#> 30  -1.11002120 -1.52818291 -0.09249290
#> 31  -0.33230079 -1.70962135 -0.87496535
#> 32   1.93799457 -0.17571557 -1.21398954
#> 33   0.37070734  1.38261961  0.68012681
#> 34  -0.39831637 -1.20528033 -0.29071953
#> 35   0.67262376 -1.52496023 -1.81107697
#> 36  -2.11584055 -0.03446179  0.51498963
#> 37   0.31807647  0.47031073  0.63633128
#> 38  -0.09090542  0.69691103  0.03665340
#> 39  -0.56181305  0.70741377  1.46861642
#> 40   0.80613094  0.51223590 -0.38297355
#> 41  -1.44407500  0.88033210  1.32604739
#> 42  -0.53177891 -1.22650174  0.57442872
#> 43  -0.16582313  0.72432168  0.77302339
#> 44  -0.74240239 -0.16782073  0.22778908
#> 45  -0.46337286 -1.65884744 -0.02601632
#> 46  -0.34446733 -0.02328551 -1.14479607
#> 47  -1.63835993 -0.51941487  0.63074563
#> 48  -0.82682439 -1.03903983 -0.27956461
#> 49  -0.80354046 -0.50745460  0.30311059
#> 50  -3.16505627 -1.28203842 -0.22349482
#> 51   1.36635681  1.28964773 -0.94305519
#> 52  -0.78927767 -1.28498148 -0.74988793
#> 53  -0.12565804  0.42178318 -0.42309466
#> 54   0.63579170 -0.73503859  0.08422335
#> 55   0.54032300  1.26718797  0.83629982
#> 56  -1.46897298 -1.14188576  0.48330474
#> 57   1.68874993 -0.23545498 -0.94062586
#> 58   0.37259413  1.19838632  0.19368601
#> 59  -1.59515930  0.39531613  0.28887018
#> 60  -0.71601581 -0.28516723  1.47124448
#> 61   1.42911414  0.60484672 -0.31436309
#> 62   0.71292241  0.91665947 -0.20383250
#> 63  -0.30298062 -1.01829618  0.86216346
#> 64  -0.51246285  1.04543049  0.16223607
#> 65  -0.43797008  0.63931435  0.89461270
#> 66  -0.09201493 -0.11858434 -1.14810273
#> 67  -0.55040845 -0.20966953  0.67736258
#> 68  -1.47062415  0.54801895  1.54913138
#> 69  -1.24467053 -0.07666995  0.77060497
#> 70  -0.41695691  0.46752963  0.24752888
#> 71   0.71410955 -0.19151302 -0.13815022
#> 72  -0.96946473 -0.51888735  0.26666856
#> 73   0.48061044  0.22583339 -0.46782024
#> 74  -0.51990327  1.22825565  1.16188212
#> 75  -0.02506469  2.15180217  2.26449546
#> 76   0.31763612 -0.24193095  0.04822805
#> 77   0.38663541 -0.69226021 -0.13312860
#> 78   1.04891691 -2.68197583 -2.25866641
#> 79  -1.98800117 -0.23284391  0.87691404
#> 80   1.42257978  0.59182491  0.12401073
#> 81   0.09376620 -1.61845822 -1.11782182
#> 82  -0.96853546  0.94907121  0.97529898
#> 83  -0.66153910 -0.38305504  0.14536200
#> 84   0.88421921  1.97702395  0.22514258
#> 85   0.54690344 -0.07265456 -1.21633478
#> 86  -0.26925724  1.07245312  0.79835749
#> 87  -0.38164465 -2.83394301 -1.08925327
#> 88  -1.43357927 -0.39784761  1.46540694
#> 89   1.73078556  1.90465371  1.32010700
#> 90  -0.68889554 -1.34539195 -0.16138690
#> 91  -1.02170889 -2.50535032 -0.84324600
#> 92  -0.45106333 -0.87543715 -0.48489714
#> 93   1.48450044 -1.54221419 -0.92825003
#> 94  -0.84958212 -0.29979822  0.09115255
#> 95   0.41776586  0.57472723 -0.74958371
#> 96   0.39027294  0.90872750 -0.13184009
#> 97   2.17423316 -0.31708849 -1.30594158
#> 98   0.47918911  1.63391396  0.38463116
#> 99   1.34387813  0.23082824 -1.19470616
#> 100 -1.67664057 -1.17409236 -0.10538580