Multiple Normally Distributed Vectors

rnorm_multi() makes multiple normally distributed vectors with specified relationships.

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>               A            B            C
#> 1    0.07514087 -0.035286541 -0.032086585
#> 2   -0.46399807 -0.536892082 -0.860826318
#> 3    0.91508401  0.922850625  0.020364253
#> 4    0.31844073  0.826684606  0.141937473
#> 5   -0.65216868  0.345357352  0.581839023
#> 6    0.09998968 -0.398463979 -0.350646942
#> 7    2.00844557  1.552412181  3.426249256
#> 8    1.84272028 -0.516118556 -0.081835696
#> 9    2.47569397  0.817380682  0.213097574
#> 10  -0.60672754 -2.063631558  0.432363188
#> 11   0.22236039 -0.777675086 -1.310468301
#> 12  -0.49786832 -0.814308454 -0.989697591
#> 13   0.55189917 -0.923481054 -0.487342067
#> 14  -1.00387737 -1.237603323 -1.055592231
#> 15   1.56016269  1.955524303  2.240026836
#> 16  -0.30267978 -0.505387077 -0.749824964
#> 17   2.48609449  0.572362683  1.394409043
#> 18  -0.85552620  0.267911567 -0.185246320
#> 19   0.64467950  0.442134824  0.261910692
#> 20   1.38846011  1.357249999  0.388556704
#> 21   0.47571035 -0.303209888  0.023901363
#> 22   0.45395320  0.018101490  1.166996579
#> 23   0.32488117  0.046043575 -0.360702054
#> 24   2.16834362 -1.102000113  0.204928693
#> 25   2.03129119 -0.336173731  1.068813478
#> 26   0.81911130 -0.474421984 -0.428248650
#> 27   0.20801337 -1.182639377  0.457606776
#> 28  -0.78723838  0.714908358 -1.366749225
#> 29  -1.70776808 -0.680940479 -1.134900557
#> 30   0.47407521  0.889296201  0.349164719
#> 31  -1.10215731 -0.036137869 -1.486420882
#> 32  -1.12708828 -0.816409947 -0.021005585
#> 33   1.66602530  1.414883948  0.979602270
#> 34   0.21486985  0.898543072  0.342225362
#> 35  -0.19876080 -0.080814342 -0.860116623
#> 36  -1.43502494 -0.176891405 -0.859392555
#> 37   1.02569496  1.464785999 -0.799425664
#> 38  -0.55929880 -0.178777150 -0.235613749
#> 39   1.64376341 -1.250863843  0.464203515
#> 40  -1.26031246  0.672506615 -0.013713990
#> 41   0.73930254 -0.836374331 -0.707846825
#> 42   0.39805236  0.001103044 -0.729607650
#> 43   2.01024158  1.137487984  1.344061904
#> 44  -0.19642235 -0.081719607 -0.455556872
#> 45   2.22064908 -0.599461056  1.703688568
#> 46  -0.40234310  1.142592591  0.618378940
#> 47  -0.15118976 -0.726971893 -0.279200348
#> 48   0.50872741  0.070049185  0.739815447
#> 49  -0.72268655 -0.758157713 -0.207791170
#> 50  -0.87227121  0.055656128 -0.435703912
#> 51   1.51875335 -0.396965339  1.313525865
#> 52   2.12273724  0.716083485  1.646462295
#> 53   0.30298614  0.230632123 -0.300673523
#> 54  -1.39936241 -1.238799803 -2.240074747
#> 55  -0.11776391  0.443380767 -0.795834055
#> 56  -0.13087796 -0.089143409  0.004294045
#> 57   0.10018897  1.407516260  1.123523546
#> 58   2.00392620 -0.880456694  0.582326708
#> 59  -0.81790344 -1.635217807 -0.931913749
#> 60  -0.66195143  0.001038715  0.488828727
#> 61  -1.64728331 -0.094610911 -0.918201763
#> 62  -1.26349576  0.032929666 -1.438459469
#> 63   1.46966183  2.934884720  2.497603786
#> 64   0.30995461  0.029294889 -0.529107164
#> 65   1.97461348 -1.140085178  0.440414165
#> 66  -0.19972976 -0.985507020 -0.699699194
#> 67  -0.05826992  1.184227562  1.108604494
#> 68   0.94610542  1.538517430  0.626872964
#> 69   0.97512032  0.936809646  2.065983179
#> 70   2.04373440  0.557684507  1.437567337
#> 71   0.18293710 -0.949798879 -0.131923870
#> 72   0.66750564 -1.114671685 -0.086987725
#> 73   2.49020869  1.612001708  2.790390100
#> 74   0.13132154  0.301366989 -0.241164223
#> 75   0.90915962 -1.826357455 -0.308642348
#> 76   0.87902669  1.584603391  1.264497060
#> 77   1.28395431  2.033056494  2.240856725
#> 78  -0.29626919 -0.610751312 -0.628261027
#> 79  -1.55576207 -1.765250798 -3.632532748
#> 80   1.36276770  0.956383701  1.806516151
#> 81   1.77991746 -1.086319258  0.076207309
#> 82   0.50628871  0.008350617 -0.248283676
#> 83   1.59769769  0.668880975 -0.017613470
#> 84  -1.58398966  1.273983633 -0.474675693
#> 85   0.01011433 -0.857756242 -0.422797820
#> 86   0.59559340 -0.108571108  0.988220924
#> 87  -0.23296645 -0.614547320 -0.493069422
#> 88  -0.58111745  0.311056436 -1.527007046
#> 89  -0.37701661  0.019409776 -0.540325095
#> 90  -0.10722181  1.251862106  1.139873714
#> 91  -1.85601829 -1.116285726 -1.236027232
#> 92  -0.14983542  0.870054222 -0.267740426
#> 93  -1.01342443 -1.133405743 -1.219440549
#> 94  -0.76910153 -0.232675899  0.432405616
#> 95   0.29591236  0.261842159 -0.147513031
#> 96  -1.80708372 -0.605613606 -1.486084506
#> 97  -0.22075256 -0.572577573 -0.254603628
#> 98   1.33852807 -0.086668832 -0.717141720
#> 99  -1.61973968 -0.718346788 -0.957242721
#> 100 -0.29403662 -0.295246939 -0.501509491

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A            B            C
#> 1   -0.95109776  0.713670849  1.155923383
#> 2   -1.53587493  1.638444914  1.545981605
#> 3    0.73697347  1.186166449  0.417581180
#> 4   -1.50611868 -0.329481410  0.564340387
#> 5    0.58301741  0.478278993 -0.030980232
#> 6   -0.54831243 -0.846571728 -1.011010809
#> 7   -0.64310048 -0.687636283  0.267889742
#> 8    0.38019188 -0.143251041 -0.021186473
#> 9    1.62078329  0.068437904 -1.466831466
#> 10  -0.73572836  0.815335459  1.997067394
#> 11  -0.80933672  0.586087542  0.705094736
#> 12   0.29463294  0.142645983 -0.363291619
#> 13  -1.18121801  1.043423640  0.744279045
#> 14  -1.00989912  0.136709143  0.890039391
#> 15  -1.24860529 -0.149407345  0.383779004
#> 16  -0.13140296  1.251019560  1.295601144
#> 17  -1.60176011  1.605169070  1.693906431
#> 18   0.56861958  1.268824182  0.936472604
#> 19  -0.81611688 -2.603953332 -2.314577182
#> 20   0.27361987  0.146505340 -0.186114274
#> 21  -1.18633430 -0.393494108 -0.769228862
#> 22   0.43102049  0.689526529  0.785710629
#> 23  -0.39158048  0.585175438  0.494631075
#> 24  -0.11993690  0.781563854  0.759081490
#> 25   0.38232210 -0.964181069 -0.972442343
#> 26  -1.65793193 -0.096706027  0.725406167
#> 27  -0.06476477 -0.991636851  0.208022052
#> 28  -0.47911004 -0.299392272  0.238321300
#> 29  -1.04452444 -0.058877732 -0.448660491
#> 30  -1.03043820  1.745694959  2.374384995
#> 31   0.35344144  1.451899589  2.029462921
#> 32  -0.42606216 -0.120592533  0.671743145
#> 33  -0.57977843 -1.186201379  0.996670650
#> 34  -0.76908683  0.383740346  1.191864833
#> 35  -1.96682924 -1.803139352  0.375196809
#> 36   0.95939822  0.223634147  0.629398648
#> 37  -0.83315531  0.317671093  0.980082433
#> 38   0.45293398 -0.723876122 -0.238841185
#> 39  -0.42975765  0.849945658  1.123815933
#> 40  -0.23735936 -0.561176394 -0.687523088
#> 41   0.87292906  0.508337105  0.788697894
#> 42  -0.94342422  0.152628418  0.835030559
#> 43   0.10648280  0.917077305 -0.063230018
#> 44   0.52609824  2.312251834  1.095213428
#> 45  -0.76090829  0.682154553 -0.006426733
#> 46   0.53334191  1.300699999  0.330957470
#> 47   0.34262331 -0.481491531 -0.980378386
#> 48  -0.23019832 -0.377468748 -0.375394667
#> 49  -0.72697056  0.362435518  0.731204539
#> 50   0.36866830  0.933196912  0.894417343
#> 51  -0.01643198  0.087299058  0.530448163
#> 52  -1.36073463 -0.143728545  1.327502809
#> 53  -0.30714076  1.401164254  1.310600247
#> 54  -0.82394709 -0.831350872 -0.745505187
#> 55  -1.07806930 -0.001696335  1.548381091
#> 56  -0.17048187  0.438980273  1.025472674
#> 57   0.62899496 -0.404269373 -1.471829623
#> 58  -0.32355153 -0.745867940  0.271371209
#> 59  -0.41553842 -1.615944355 -0.734152861
#> 60  -0.27977938 -0.194875244  0.942502153
#> 61   0.36012717  0.125517921 -0.835166733
#> 62  -1.34616643 -1.597066708 -0.130722367
#> 63  -0.19236172 -0.337065599 -0.115319042
#> 64  -0.55249730 -0.941445669 -0.154270104
#> 65   0.04357475 -1.301586991 -1.344634895
#> 66   0.99310089 -0.812866170 -0.795054953
#> 67   1.93649093 -0.839247224 -2.295103483
#> 68  -0.76068999 -1.497423550  0.332520642
#> 69  -0.41246846  0.204787460  0.163523208
#> 70   0.20065992  0.041457936 -0.815410757
#> 71   1.42039186  0.352547368  0.212689786
#> 72  -0.16717481  0.173483976  0.083427506
#> 73   0.90860405 -1.358581127 -1.423551414
#> 74  -1.30334267  0.018473275  1.470684207
#> 75  -1.69217551  0.173790150  1.818497119
#> 76   1.26695159  2.159056623  1.356697990
#> 77   0.63081567  0.219896767 -0.855195230
#> 78   0.95805841  1.536581534  0.453429477
#> 79  -1.15721816 -0.208876006  0.099411140
#> 80   1.55725223  0.442065029 -1.086694113
#> 81   1.20839383 -1.591176001 -1.255080048
#> 82   2.15782459  0.145360669 -0.993279142
#> 83  -0.89362287  1.066839868  1.440368894
#> 84  -1.61894912 -1.834159217  0.931600503
#> 85  -0.53840123  1.050360102 -0.223329822
#> 86  -1.37166865 -1.554413899 -0.881314810
#> 87  -0.44000952 -0.818198878 -0.188741609
#> 88  -0.09073200  0.096000206  0.758600195
#> 89  -2.05243497 -1.891650060  0.008325293
#> 90   0.24119187 -1.869482888 -1.577510050
#> 91   1.18883372  0.812430533 -0.085497298
#> 92  -1.31341629 -1.910364786  1.283948228
#> 93   0.56036705 -0.329735627 -1.752966954
#> 94  -0.54409668 -0.278088368 -0.108665567
#> 95  -0.27655658  0.590116017  0.928550105
#> 96   0.75786886 -0.744121137 -0.482494605
#> 97   1.40071015 -0.871575233 -0.815650128
#> 98  -0.21254833  0.316479188 -0.104149357
#> 99   0.01646824  1.086590411  0.832302016
#> 100 -0.49297534 -0.463270871  0.014095154