Multiple Normally Distributed Vectors

rnorm_multi makes multiple normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, cors = 0, mu = 0, sd = 1,
  varnames = NULL, empirical = FALSE, as.matrix = FALSE)

Arguments

n
the number of samples required

vars
the number of variables to return

cors
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

varnames
optional names for the variables (string vector of length vars) defaults if cors is a matrix with column names

empirical
logical. If true, mu, sd and cors specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

Value

dataframe of vars vectors

Examples

rnorm_multi(100, 3, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A            B           C
#> 1   -1.711924939 -1.011599824 -1.32018668
#> 2    0.887134792  0.043794457  0.99462051
#> 3    0.748427556  0.410991509  1.01598343
#> 4   -0.988547303 -1.612324193  0.10827445
#> 5    0.250415810  1.967522573  1.12127447
#> 6    1.012473136  1.329047164  0.46137529
#> 7    0.290391321 -0.492076253 -1.07257786
#> 8   -0.999779485 -2.143044059 -1.61909580
#> 9    1.296369128  0.223297040  0.05172819
#> 10   1.844675802  0.525323939  1.43792983
#> 11  -0.594332509  1.134701125 -0.51513681
#> 12   2.892055704 -0.024653246 -0.87388294
#> 13   0.674912545  1.145791522 -0.16442269
#> 14   0.114929596  0.751160451  0.22516804
#> 15  -0.798672240 -0.534844472 -1.59558698
#> 16  -1.031496700  0.049687627 -0.54144674
#> 17  -0.004310488 -1.375989084  0.16455645
#> 18   1.415566785  1.598106413  1.06560667
#> 19  -1.642731803 -0.241490494 -0.33724621
#> 20   0.910502722 -0.191788473 -0.98662517
#> 21  -1.175150978 -0.479527388  0.14625695
#> 22   0.305901881  0.761645013 -0.54205693
#> 23  -0.367942472 -1.256019905 -0.12063644
#> 24   0.892308000  0.193255995  1.16876289
#> 25   2.538452134  2.240282666  2.04947504
#> 26  -0.695068465  0.222697327 -1.11980217
#> 27   0.512021718  0.508770162  0.57718656
#> 28  -0.932274424  2.643562030  2.24879422
#> 29  -0.585604074  0.622505249 -0.24917585
#> 30  -1.355941412 -0.590585765 -0.21519908
#> 31  -0.733846362  0.082328508 -0.42830930
#> 32   1.264872249  0.715577491  0.12590367
#> 33  -0.518786456  0.201122493 -0.92638734
#> 34  -0.637506919 -1.122500780 -0.86513473
#> 35   0.632266473 -0.010632207  1.05787745
#> 36   1.104769095  0.406299810 -0.21379747
#> 37   1.472185167  0.971669943  0.38215346
#> 38   0.208642678  1.232615573 -0.26677942
#> 39   0.194660642 -0.021702220 -0.36527459
#> 40   1.781764439  0.408108369  1.86009067
#> 41   2.042761937 -0.215878860  1.47377857
#> 42  -0.490815530 -0.425984391 -0.68877135
#> 43   1.655781000 -0.033659668  1.27356903
#> 44  -0.025382344  0.905039301 -0.02548762
#> 45  -0.665268785 -0.563017316  0.06761990
#> 46  -0.276302646  0.608159422  0.33566043
#> 47   0.484250987  1.126346311  1.25406657
#> 48   1.282907646  0.720836758  0.47160057
#> 49  -0.277030402 -1.021860386 -1.96611414
#> 50  -0.819145743 -0.037232249 -0.04063296
#> 51  -1.564663610 -0.636464908 -1.85042764
#> 52  -0.898058218  0.284764624 -0.34268128
#> 53   1.805225599 -0.408926161  0.83297444
#> 54   1.365198223  1.113118376  1.19018362
#> 55   0.216231182 -0.115908105  0.10565315
#> 56   0.641593299  1.738105080  2.21714184
#> 57   0.213674643  0.168441446 -0.31953324
#> 58   0.253013272  0.862077906 -0.22789514
#> 59   0.287070174 -1.951729555 -0.83911657
#> 60  -0.481794003  0.239976721  0.61024139
#> 61   0.025132801 -0.628056375 -1.53689844
#> 62  -0.167673231 -0.101515541  0.75379782
#> 63  -2.425362627 -0.498203692 -1.06272155
#> 64   1.350643220 -0.486047718  0.95719724
#> 65  -1.242626738  2.175710297  0.92752367
#> 66  -0.338836433  1.113640640 -0.15518463
#> 67  -0.026841819  0.966138768  0.37921477
#> 68  -0.453608406  0.413598279  0.23549706
#> 69  -0.493723033 -1.334332953 -0.84630857
#> 70  -1.221646818 -1.238780109  0.72514498
#> 71   0.321048652  1.743009432 -0.49620029
#> 72  -0.527149031  1.184322665  1.00890307
#> 73   0.070016525  1.224225492  0.62085618
#> 74   1.134493406  2.219461421  1.58638671
#> 75   1.504749960  0.614156332  1.09643083
#> 76   0.446530671  0.052052719  0.61180283
#> 77   1.818863849 -0.302817368  0.04494236
#> 78  -0.791160513  0.179908256  0.79171348
#> 79  -0.300914823 -0.253138072 -0.53330713
#> 80   0.926355041  1.873512137  0.35236855
#> 81   0.144573957  1.043953709 -0.49422615
#> 82  -0.027810676  0.325938318 -1.58919459
#> 83  -2.100930284  1.066836105 -0.83668769
#> 84  -0.444333571 -0.715766312 -0.85121894
#> 85  -0.968875193 -1.687613536 -0.79915130
#> 86   0.959440062 -1.088246737  1.55349882
#> 87  -0.345251762 -1.686853560  0.65311349
#> 88   0.572082195 -0.813048594 -1.24514160
#> 89  -1.365521552 -1.373149134 -1.59569125
#> 90   1.176097394  2.119269863  1.42040520
#> 91   1.221033387 -0.577310879 -0.56619463
#> 92   0.108998277  0.648942202 -0.11418245
#> 93   0.101151686  0.111384203  0.73298987
#> 94   0.586259342  1.453725946  1.13935521
#> 95   0.153452555  1.348353614 -0.27134221
#> 96   0.235523130  0.480340620 -0.05456858
#> 97   0.785445485  1.876836130 -0.04612489
#> 98  -0.198978331 -0.525346365 -0.12181418
#> 99   0.629380511 -0.357590811  0.51006913
#> 100 -2.276017019 -0.001639159 -1.01436632

rnorm_multi(100, 3, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A            B           C
#> 1    0.98053619  1.397345080  1.24702308
#> 2    0.28494917 -0.929721452  0.11231993
#> 3   -0.48926432 -0.317158657  0.91387543
#> 4    0.62746163  0.690257228 -0.04676857
#> 5    0.52497027 -0.006770402 -0.54982277
#> 6   -0.08113113 -1.203695112 -0.38591526
#> 7    0.10950230 -0.448304595 -0.72749899
#> 8   -0.98054763 -1.413085775  0.44500300
#> 9   -0.22717869  0.588496717  1.47955055
#> 10  -0.08167748  0.129810359  0.62297867
#> 11  -1.06794579  1.622348860  2.17266535
#> 12   1.23713517  1.254934760  0.21802500
#> 13   2.34268548 -0.127564780 -1.84686336
#> 14  -0.79906111 -0.670100344 -0.78681222
#> 15   1.14762821  0.425896059 -0.19353105
#> 16  -0.68792927  0.746387469  1.17342419
#> 17  -0.42239464 -2.279769415 -1.45252677
#> 18  -0.71983873 -0.782993158 -0.48339888
#> 19  -0.27644209 -0.078365574 -0.15548499
#> 20   0.65702277 -2.167247198 -2.28768427
#> 21   1.36346504  2.807540375  0.61058213
#> 22   2.26453757 -0.107450280 -1.03851429
#> 23   1.07057295  1.036447960  0.77105891
#> 24  -0.75535093  0.189251094  1.72491550
#> 25  -0.68100214 -0.665149580 -0.04357655
#> 26  -0.47017350 -0.314580812 -0.27934568
#> 27   0.73476477  1.147051694  0.19216533
#> 28  -0.09149070  0.876867711 -0.08339957
#> 29  -2.05839553 -1.187120062 -0.84369069
#> 30  -0.62779679  0.680592120  0.13823150
#> 31   0.82896856  0.492073064 -0.10702775
#> 32   1.12230254  0.029101542 -0.69550195
#> 33   1.41109327 -0.547954608 -1.55760838
#> 34  -1.28998785 -0.644308963  0.45818445
#> 35  -0.74524292  0.423205419  2.23457494
#> 36   0.18163205  0.475178614  0.81660170
#> 37  -1.29214617  0.507173875  0.60586872
#> 38  -0.52453660 -0.422831247  0.14219949
#> 39   2.71082498  1.025301778 -0.77900169
#> 40   0.67234061 -0.927892656 -0.94160161
#> 41  -0.22932465 -0.452949466 -0.75393801
#> 42   0.79662820  0.012013809 -0.66456985
#> 43  -1.48598921 -1.485862845 -0.34075299
#> 44  -2.73794234 -0.333565773  1.35801078
#> 45   1.04730267 -0.239862856 -0.94911295
#> 46   0.79570220  0.409234356 -0.02512850
#> 47  -0.29797799 -0.960151155  0.32858669
#> 48   1.32398457  0.086940658 -0.84502893
#> 49   0.70645034  0.763352966  0.66553042
#> 50  -1.19011584 -0.898273386  0.01265418
#> 51   0.17313328  0.375553597 -0.01675357
#> 52   1.04061188 -0.219130525 -0.77337525
#> 53  -0.43883826 -0.042922039  1.27360295
#> 54   0.67987637 -1.058120498 -0.27502560
#> 55  -0.51468288  0.062507127  0.76999444
#> 56  -0.80241295 -0.945210621  0.34889021
#> 57  -0.47060134 -1.268959331  0.70686003
#> 58   2.49959376  0.682354301 -1.46377249
#> 59  -0.76711289 -0.641906992  0.26363487
#> 60   0.84037060 -0.130252667 -0.61136080
#> 61  -0.47759972  0.405378724  0.06637905
#> 62  -1.35795459  0.303024258  1.33414293
#> 63  -0.86120089 -1.445283701  0.98588190
#> 64   0.65405904 -1.262137582 -1.45782179
#> 65  -1.52487511 -1.185720620 -0.29528001
#> 66  -0.60310623  0.543202595  0.16109147
#> 67  -1.34640940  0.111464903  0.03001872
#> 68  -1.20726090 -0.211677096  0.84891723
#> 69  -0.07838011  0.191885581  0.98564804
#> 70   1.13262307  0.785543693 -0.06059885
#> 71   1.36490199 -0.252005899 -2.00968693
#> 72   1.18092139 -0.345183489 -0.80804496
#> 73  -1.50057211 -0.143484529  0.56351319
#> 74  -0.67605220 -0.161655297  0.70612018
#> 75  -0.39630979 -0.711680593 -0.01193046
#> 76  -1.06729240  0.790110596  0.42953674
#> 77   0.25135374  0.004217116 -1.46584325
#> 78  -0.62606452 -0.500831364  0.35233621
#> 79   0.22130814 -1.054521838 -0.13425086
#> 80  -0.82249462 -0.106502250  1.12587401
#> 81   1.13659660  0.694685870 -0.63335312
#> 82   0.63912932 -0.346001998 -0.15632373
#> 83   0.50735668 -0.121686372 -0.35290927
#> 84   0.95787156  0.663581444 -0.85623386
#> 85  -0.91227314 -1.085103286  0.84931769
#> 86  -0.23904032  0.820166860  0.16160248
#> 87  -0.71341295 -0.149361316  0.56590330
#> 88   1.64368977 -1.175172449 -2.12371033
#> 89  -0.35317370  0.697507425 -0.12301334
#> 90  -1.24671366 -0.243868992  1.74687093
#> 91  -0.64148189  0.722090091  1.47996273
#> 92   0.22026898  1.174067582  1.19864762
#> 93  -0.88325604  0.271478223  0.35463827
#> 94   0.07313435  1.405022266  1.00772946
#> 95  -1.47150701  0.560249257  1.03483886
#> 96  -1.10044697 -1.187067944  0.66307881
#> 97  -2.23696083  0.102519789  2.21376158
#> 98  -0.92624102 -1.158465566  0.37674556
#> 99  -1.39758231  0.256387690  1.30028559
#> 100 -0.21576305 -0.512530549 -0.64474625