Multiple Normally Distributed Vectors

rnorm_multi makes multiple normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

dataframe of vars vectors

Examples

rnorm_multi(100, 3, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>               A            B           C
#> 1   -0.22938009  0.012786425 -0.90004352
#> 2    1.56046133 -0.385432656  0.75531705
#> 3    0.12914257 -0.656736867 -1.93726361
#> 4   -0.07215368 -0.395541430  0.49442871
#> 5   -2.24668003 -1.356275428  1.12155272
#> 6    0.26548664 -0.290537897  1.64841161
#> 7   -0.89850890 -0.158541994 -1.32181766
#> 8   -0.43317818 -0.136663326  0.25267470
#> 9   -1.86365445  0.627127133  0.25580039
#> 10   2.84893203  1.378454920  0.21729455
#> 11  -0.95339839  0.191117349 -0.05369938
#> 12  -0.14063788 -0.999410460  1.12898204
#> 13   0.98636258  0.658537288  2.56502490
#> 14  -1.07051311 -0.041799453 -1.13098940
#> 15   0.74214155  0.968599861  1.01242695
#> 16   0.27510590  2.526850459 -1.36301149
#> 17   0.75851442  0.824858441 -0.02201251
#> 18   0.61540640 -1.284281532  0.44739809
#> 19  -1.25229977  0.649401784  1.04299634
#> 20   1.14120612  1.472838252 -0.41407483
#> 21  -0.13893587  2.439369263  0.96815442
#> 22   0.12442575  0.849453778  0.86295126
#> 23   0.24020439  1.791814046 -0.80454355
#> 24   0.32430107  0.826566547  1.23777632
#> 25  -0.79843255  0.507583992  2.38850493
#> 26   1.43339006  0.422294733  0.40255490
#> 27   0.54042449  1.003611011 -0.43584735
#> 28  -0.27270248  0.137349427  0.48404969
#> 29   0.90875306 -0.128264082 -0.32678895
#> 30  -1.32895871  0.592149422 -1.01239965
#> 31   0.43742535 -0.746199669  1.43536319
#> 32  -1.11281425  1.246184665  0.67648861
#> 33   0.94702859  0.481719629  0.74368546
#> 34  -1.36251843 -0.905117010  2.12354888
#> 35   0.27105336 -0.544912060  0.61203808
#> 36  -0.43953477  0.854341594  0.36600299
#> 37  -0.64519574 -0.455202501 -1.41008747
#> 38   0.87524470  0.113104781  0.22076276
#> 39   1.35337579  1.294961626  0.18655402
#> 40  -1.48650474  0.467304440  1.56730788
#> 41  -0.70281495  0.237323663  0.57003485
#> 42   1.51763370 -0.427310169 -0.13912332
#> 43   1.30018974  2.276505621  0.45003510
#> 44   1.40376784  1.166440199  0.24851656
#> 45  -1.23127078  1.379956696  0.94479712
#> 46   1.58291086  1.721780992  3.25541758
#> 47   0.20312594 -0.719710829  0.54653138
#> 48   0.12211318  0.914599819  1.07770907
#> 49   0.64142823 -1.109099836  0.61819487
#> 50   0.32892290  1.932741480 -1.41172049
#> 51  -0.63021426  0.829147371  1.36208648
#> 52  -0.30359291  0.522103414  0.25676326
#> 53  -0.99364118 -0.738012401  0.29391281
#> 54  -0.55172332 -0.158015129  0.51917759
#> 55   1.65584140  1.452538537  0.52956075
#> 56  -0.62860353  1.077683644  1.04982754
#> 57   0.48977446  0.438499547 -1.77411486
#> 58  -0.28005348  0.043618813  3.18255718
#> 59  -0.40482935  1.182844102  0.13877874
#> 60   1.66011018  1.204411616  0.71335575
#> 61   0.34967935 -1.500060823  1.57434588
#> 62  -1.23332110  1.335784286 -0.16508825
#> 63   0.18969668  0.090948497  1.61395242
#> 64  -0.01223603  0.663066677  0.25410359
#> 65  -0.70634018 -1.390591856 -0.67756331
#> 66  -1.90215248 -0.388258845 -0.47585062
#> 67   2.09336046 -0.733021669  1.56505732
#> 68  -0.76812584  0.763652568  0.63167063
#> 69   0.09739696  0.114112086  0.98862881
#> 70   0.43995957  0.917669134 -1.19945057
#> 71   0.26089889  0.297091330 -0.97073631
#> 72  -1.97757603 -0.574069593  0.78415034
#> 73   0.08213986  1.670672301  1.83732041
#> 74   0.31229479  1.360864787  0.73669628
#> 75   0.20788620  1.168721370  1.81829338
#> 76   2.07774387  1.435930771  1.02390979
#> 77   2.35875655 -0.073887074  1.10674805
#> 78   0.90971452 -0.875334875  0.39006433
#> 79   0.96698338  0.094379326  0.67218172
#> 80  -0.10821142  2.611769487  0.40967271
#> 81   1.21200185 -0.641668381  2.42434334
#> 82  -0.71905160 -0.746523850  1.79839276
#> 83   0.76338008 -1.275327303  1.24879127
#> 84   0.52248275  1.925938655  1.05622433
#> 85   0.56667436  0.954185515 -0.04825726
#> 86   1.32983515  2.393110265  1.61053489
#> 87  -0.74149808  0.245879260 -2.11233433
#> 88   0.41783764  2.964408338  0.34430622
#> 89   1.61541229  1.461999145  0.93388979
#> 90  -0.18373305  1.542694878  0.11804889
#> 91   0.02591363  1.523838843  0.92418757
#> 92  -0.02174452  0.002998507  1.56310200
#> 93  -0.80952872 -0.423261151  1.54871262
#> 94   0.68072527 -0.178884625  0.46189711
#> 95   1.80440733  2.163789378  0.98614892
#> 96  -1.31502453  0.532992146  2.17288261
#> 97  -1.21602391  0.776499328  0.14563884
#> 98   1.07677733  1.538707653  1.44634789
#> 99   0.82413241  1.641263075  1.81682636
#> 100  2.31227729  1.012090945  0.20335998

rnorm_multi(100, 3, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#> Error in rnorm_multi(100, 3, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1),     varnames = c("A", "B", "C")): the length of mu must be 1 or vars