Multiple Normally Distributed Vectors

rnorm_multi() makes multiple normally distributed vectors with specified relationships.

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>               A           B           C
#> 1   -0.39525051  1.60204155 -1.23596470
#> 2   -0.83094715 -0.61028059 -0.41526215
#> 3    1.10741695 -0.03410401  1.65078512
#> 4    1.05210280  0.74874664  0.34818001
#> 5    0.08348574 -0.57347797 -2.03954672
#> 6    0.51392142 -0.42364161 -0.55702577
#> 7   -0.16083059  1.41765130  0.34226227
#> 8   -0.18604875 -0.78244004 -0.18806801
#> 9    0.02217013  1.33082736 -0.07445351
#> 10   1.93314424 -0.71190304  0.35216204
#> 11   1.13192897  2.85607016  2.27255396
#> 12   0.25306878  1.87393752  1.53455062
#> 13   0.57349022 -0.56124720  0.28778523
#> 14  -0.76618908 -0.36797718  0.15947318
#> 15  -0.86673846  0.49170388 -0.87400738
#> 16   1.19326250 -0.90314503 -0.34567135
#> 17   0.16587238 -0.36976487 -0.16845418
#> 18   1.18293818  1.06505451  2.01240690
#> 19   1.52078071  1.44946591  1.06521539
#> 20  -0.11153672  0.24446570 -0.28113035
#> 21  -1.79427147 -0.89291789 -1.59678672
#> 22   0.27963076 -0.74334237 -0.99811207
#> 23   0.36067275 -0.96167980 -0.43168608
#> 24   2.98561995  0.13226221  0.75851986
#> 25  -0.86309449 -1.19168073 -0.76918283
#> 26  -0.53510182  0.09482174 -0.45453407
#> 27  -2.16118976  0.22326464 -1.19740940
#> 28  -2.13050458 -0.04733086 -1.15877187
#> 29  -0.95857208  0.86676742 -0.80483657
#> 30   0.90359026 -0.62613780 -0.39722442
#> 31   0.41126505  1.67597975  1.92406936
#> 32   0.12505746  0.12142971 -1.82585892
#> 33   0.41618281 -0.03802799  0.80550964
#> 34   0.72075911 -1.80858071 -0.19806700
#> 35  -1.39746358 -1.22602731  0.15271762
#> 36  -0.38689373  1.75676693  1.32114809
#> 37  -1.09551546  0.49889726  0.25582316
#> 38  -0.42339316  0.56332884  0.49539281
#> 39  -2.19448659 -0.12634163 -0.94499684
#> 40  -0.22014440 -0.37569587  0.18194389
#> 41  -1.37772190  0.65761467  0.76689481
#> 42  -0.10055969  1.72884663 -0.85306991
#> 43  -1.89710893 -0.14856126  0.09617118
#> 44  -1.00564708 -0.16505142 -1.08175138
#> 45   0.79576066 -0.71648334  0.90164361
#> 46  -1.10177595  1.44075288 -0.78370873
#> 47   0.69528255  0.19495379 -0.08656786
#> 48   0.03991584  1.04429548 -0.44228705
#> 49   0.35335412 -0.68943796 -0.04036325
#> 50   0.52609538 -1.10777437 -0.22795153
#> 51   1.59230939  1.36278919  0.28358808
#> 52  -0.19708036  0.51482813  0.27239468
#> 53   0.65962546  2.11191083  1.49041715
#> 54   0.01359814 -0.35840262 -0.54596646
#> 55  -1.17467321  0.43455507  0.17723704
#> 56   0.44446053  0.21501372  0.64837585
#> 57  -0.79525427  0.32998408  0.74805375
#> 58  -1.48700921 -0.24961922  0.52639834
#> 59   0.72980205 -1.22801671  0.73566519
#> 60  -0.30544681  0.16009936 -0.82906136
#> 61   1.65280871 -0.77528616  0.91390042
#> 62   0.85236017  0.03751425  0.40782773
#> 63  -2.61988457 -0.41366722 -1.00136884
#> 64  -1.07065582  0.21999742 -0.37180596
#> 65  -1.34754019 -0.85315809 -0.57809090
#> 66   1.35292712  0.90846281  1.44426742
#> 67  -0.44056970 -0.14443574 -0.27723210
#> 68   1.30041041 -1.25086537  0.80851961
#> 69  -2.02325560 -0.00568078  0.25399138
#> 70  -1.41285534 -1.96469749 -0.43600872
#> 71   0.87184190 -0.72327842 -1.88787837
#> 72  -0.58943778  0.94653519  0.23771436
#> 73  -1.12750863 -0.45313743  1.12760240
#> 74   1.31885257 -1.59831861  0.87225229
#> 75  -2.17970901  1.42255872 -0.76277824
#> 76   0.10960734  0.83787242 -0.05185770
#> 77  -1.52224670  0.09814647  0.29013920
#> 78   0.52639747  1.53003906  2.27465755
#> 79  -1.14982378  0.14386347  0.12302238
#> 80   0.08258408 -0.45208982 -0.76482025
#> 81  -0.99207884  0.65867277 -1.07855054
#> 82   0.98589597  0.01556392  0.02187493
#> 83   0.50332539  0.29071744  1.76348694
#> 84  -0.21626402 -0.32166622  1.01345053
#> 85   0.17581993  0.55340869 -0.54534382
#> 86  -0.05887059  0.77427087 -0.58903907
#> 87   0.16371372  0.19632405 -1.42414321
#> 88   0.25974208 -2.08925242 -2.03063566
#> 89   1.59041059  2.48143158  1.12230141
#> 90  -0.21657097 -1.97550487 -2.27058843
#> 91   0.28798690 -0.36334205  0.35343553
#> 92   1.50495541  1.86526768  2.37377086
#> 93  -1.24136053 -0.55846340  0.52021520
#> 94   1.42688581  0.80863983  0.83316460
#> 95   0.50108335 -0.05058629  0.49215458
#> 96  -0.21799323  0.26254376  0.45034340
#> 97  -0.29374580 -0.38221338 -0.07553967
#> 98  -1.67268172  0.53499312 -0.04372332
#> 99   1.02130201  1.52833314 -0.13247479
#> 100 -1.30467600 -1.58932914 -0.72436986

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A           B            C
#> 1   -0.68161309 -1.20288439  0.024631764
#> 2   -1.91406499 -0.81171684  0.584294213
#> 3   -0.11602244  0.53070678  0.574831668
#> 4   -0.99792397 -1.01772030 -1.299895895
#> 5    0.18512493 -1.93611052 -0.314616038
#> 6    0.29888848  0.78732200  0.007358958
#> 7   -1.28731876 -1.36722082 -0.004458478
#> 8   -1.22786203 -1.27403073 -0.757835286
#> 9   -2.58490115 -2.23273497 -0.686239973
#> 10  -0.34114416  0.29870897  0.235780898
#> 11  -0.34902380 -0.13502155 -0.429484661
#> 12  -0.19222795  0.49283375  0.704182001
#> 13  -0.04085203 -0.57243863  0.032697404
#> 14  -1.01590084 -0.49062775  0.817781958
#> 15  -1.03111281 -1.77993888 -0.716833124
#> 16   0.12903052 -0.50816745  0.237706547
#> 17  -0.22422302 -0.82343328 -0.451196776
#> 18  -0.55824647  0.03021233 -1.373883411
#> 19  -1.17921721 -0.06361306 -0.124782831
#> 20   1.20649883  0.34639931 -0.579932318
#> 21   0.23247874 -1.25412528 -1.676805628
#> 22  -0.83546915  0.52936083  0.491464663
#> 23   2.18364484  0.54783713 -1.942204283
#> 24  -0.84107304  1.59626885  1.701165004
#> 25   0.11162520  1.80656334  0.738606179
#> 26   1.08372783  0.85914282 -0.267416428
#> 27  -1.15764294  0.25806675  0.445184559
#> 28   0.54779410  0.49427098 -0.016485813
#> 29  -1.60590603 -1.22763718  1.630174907
#> 30  -0.16834723  0.94219640 -0.628194393
#> 31  -1.29866848 -0.54453712 -0.210026687
#> 32  -1.32245125  0.67681163  0.647138052
#> 33   2.27499807 -0.47997482 -2.066984585
#> 34  -0.63508085  0.70215302  2.135298971
#> 35   0.58575485 -0.09891610 -0.110227025
#> 36   0.91016908  1.17805408  0.240692335
#> 37  -0.19731272 -0.37943561 -0.513258747
#> 38  -0.72654863 -0.03853057  0.941388173
#> 39   0.01832143  0.76686348 -0.191737732
#> 40   0.96108229 -0.22492339 -0.363827567
#> 41  -0.77756791  0.64791226  1.385480907
#> 42  -1.98491794  0.57495771  2.015623134
#> 43  -1.53040542 -1.32621033 -0.572276632
#> 44   0.46217351 -1.09726705 -2.773676392
#> 45   1.34479316 -0.55681854 -0.999473583
#> 46   0.08069190  1.07978264  0.068916068
#> 47   0.21205978  0.79809626  0.074656478
#> 48   0.34920010  0.32224731 -0.586447212
#> 49  -0.99690821 -1.18992992 -0.098436428
#> 50  -0.57254399  1.24608611  0.730610991
#> 51  -0.43032558 -0.04506700  0.548194989
#> 52  -0.19511949  0.58951505 -0.062111492
#> 53  -3.82981033 -0.92554898  0.598775780
#> 54   0.23513345  0.41073453  0.411390579
#> 55  -1.20170412 -0.24246132  1.829629471
#> 56   0.67922982  1.32898477 -0.285465522
#> 57  -0.18187201 -0.46006436 -0.508355788
#> 58  -0.87501123 -1.12717683  0.238650104
#> 59   0.45549049  1.60604440  1.153891989
#> 60   0.48245429 -0.68277095 -1.522902273
#> 61   0.88900639  1.97390420 -0.249667343
#> 62  -1.17329056  0.48180678  0.413531714
#> 63  -1.99038499  0.43387480  1.106563276
#> 64  -0.14844003  0.44189860  0.059664504
#> 65  -1.74390300 -1.73715739  0.116134452
#> 66  -0.04058235  0.44364181 -0.361572906
#> 67   0.57922843  0.85960869  1.063692806
#> 68  -0.24416071  0.01345569 -0.035080490
#> 69   0.25171564 -0.61316952  0.159202009
#> 70   1.48071233 -0.94372969 -1.730528708
#> 71  -0.42403688 -1.30440451 -0.214612899
#> 72  -1.06824801 -0.92952482 -0.450667092
#> 73   0.06492193  0.12121887 -0.233886798
#> 74   0.54342516 -0.73984720 -0.045437971
#> 75  -0.94013746  0.50991037  0.543237307
#> 76  -0.73964675  0.63939273  1.241963313
#> 77   1.49493395 -0.91650381 -1.751453236
#> 78  -0.96897517 -0.59981329 -1.033744269
#> 79  -1.21374554 -0.63096824 -0.398748366
#> 80  -1.37369164 -1.59838349 -1.006616825
#> 81   1.40632362  0.81658933 -0.284814393
#> 82   1.35470331 -0.39577014 -1.259107221
#> 83  -0.91138353  0.37292308  0.492012723
#> 84   0.56024999 -0.67255808 -0.592477315
#> 85   1.41923457 -0.75362101 -0.238010106
#> 86  -0.06029125  1.94567060  0.505453259
#> 87   0.20947228  0.27092938 -0.381428144
#> 88   0.76004309  1.81701605  1.218908844
#> 89  -1.17376131 -0.51436656  1.174284606
#> 90   1.38331950  1.72288969 -0.494501240
#> 91  -1.23256042  1.35680285  1.983084939
#> 92   1.20461797  1.18496366 -0.238604682
#> 93   1.14901254 -1.55873837 -0.759859907
#> 94   1.12880844  0.47107465 -1.358582031
#> 95   0.74677120  0.23484557  0.189701952
#> 96   0.58469684  0.36023573 -0.470643830
#> 97  -0.07911216 -0.37132240 -0.440357289
#> 98   1.01900220  0.77896839 -0.789277098
#> 99   0.74895827 -0.47350036 -1.768109911
#> 100  2.09714685  0.98942208  0.537047756