Make normally distributed vectors with specified relationships

Make normally distributed vectors with specified relationships

rnorm_multi(n, vars = 3, mu = 0, sd = 1, r = 0, varnames = NULL,
  empirical = FALSE, as.matrix = FALSE, cors = NULL)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

cors
(deprecated; use r)

Value

a tbl of vars vectors

Examples

rnorm_multi(100, 3, 0, 1, c(0.2, 0.4, 0.5), varnames=c("A", "B", "C"))

#>                A           B           C
#> 1    1.455494708  0.70680521  1.83865351
#> 2    1.290022979 -0.73163972 -0.10013057
#> 3   -0.664458112  0.80821701  1.08126776
#> 4   -1.079205563 -0.22818018 -1.06824876
#> 5    0.001749963 -1.09439165  0.34113084
#> 6   -0.128064778 -0.47206951 -0.63278869
#> 7   -0.249500764  0.28748018 -1.07000526
#> 8    0.459678301 -0.46015391 -0.35459423
#> 9    0.486273115 -1.34978235  1.31910316
#> 10   0.963109018 -1.09675093  0.52756108
#> 11   1.150034661  0.99481298  0.84879370
#> 12  -1.032398021  0.16225162 -0.18857251
#> 13   0.649569144 -1.34620994 -0.46883863
#> 14  -0.294242693  0.84100556 -0.50794557
#> 15   0.161760140  0.00426693 -0.34679567
#> 16   0.732430390 -1.10373239  0.17985701
#> 17  -0.998133964  0.46153404  1.26073449
#> 18  -0.521495541  0.77170134  1.58449782
#> 19   0.732504388 -0.02160091  0.40847222
#> 20  -1.245537980 -1.07043330 -0.38380842
#> 21  -0.696733653 -0.87635503 -1.69286763
#> 22   0.800555142 -0.99340112 -0.11595438
#> 23  -0.359002079  2.04640462  1.28852905
#> 24   0.918928033  3.13101959  2.11151768
#> 25   0.570920922  0.71208384 -0.94770593
#> 26   0.400941828 -2.00856358 -0.32322151
#> 27   0.457181632 -1.65963179 -0.86089726
#> 28   1.295734970  0.39574875  0.77515439
#> 29   0.624472140 -1.00784560 -0.46075524
#> 30  -0.274134166 -0.17967564  0.09351818
#> 31  -0.206072959  0.27663729 -0.07679561
#> 32  -0.563932305 -1.31138727 -0.68529554
#> 33  -0.836397638  0.48759158  0.60004123
#> 34   2.781830997  2.52555891  2.51438406
#> 35   1.216573454  1.44375211  0.25274179
#> 36  -1.383841234  0.40297677  1.14231741
#> 37  -0.822303229  0.85799861  1.52176248
#> 38  -0.729657322  0.15863593 -0.53415066
#> 39   0.991776997  0.14213522  0.06588001
#> 40   0.545183586  0.53769758  0.81116494
#> 41   0.511791068  2.10247305  0.99653779
#> 42   0.413004050  0.69099069  0.41617236
#> 43  -0.831900121  1.48367632  1.35601277
#> 44   0.148411838  0.01507207 -1.04530939
#> 45   0.025100620 -1.18666399 -1.59851024
#> 46   0.793220037 -0.55948502 -1.37401189
#> 47  -0.279165213  0.38926127 -0.84921163
#> 48  -0.023175615 -0.32915874 -1.15680443
#> 49  -0.742517128 -0.17898513 -0.04450863
#> 50   1.212695690 -0.63069122  0.71089951
#> 51  -0.565379158  0.02673954 -0.46349787
#> 52   0.526427873  0.26814135  1.37143354
#> 53   0.283522827  0.19737834  0.04793519
#> 54   0.118631375  0.53548813  1.98650878
#> 55  -0.245018090 -2.96210372 -2.35554482
#> 56   0.207030710  1.06542364 -0.98473426
#> 57   0.545422985  0.50247303 -0.92585148
#> 58  -0.603165882  0.40866626 -1.44726242
#> 59  -1.375376691 -1.25431470 -1.12715474
#> 60   2.615952621  1.59059789  0.29518129
#> 61   0.376313790 -0.56754715  0.17399995
#> 62   0.088498736 -0.74351275  0.44939920
#> 63  -1.044480595 -1.56511221 -0.28567838
#> 64  -0.842087815  0.05977675  0.23951556
#> 65   0.616771320  0.45849890  0.71769780
#> 66   0.916910824  2.09922760  0.48415284
#> 67  -0.178477537 -0.01042503  0.43363641
#> 68   1.507953977  1.79485537  1.35590090
#> 69   0.637622780 -0.99274749 -0.21538439
#> 70   0.660560492  1.83188867 -0.08081496
#> 71   0.437983132 -0.37777562 -0.10320093
#> 72  -0.636345138 -0.87998900 -0.96582159
#> 73   0.603285582  0.41325718  0.82042362
#> 74   0.709925705 -1.16531239  0.34816720
#> 75  -0.812479502  0.22781172 -0.05965569
#> 76  -0.567945398 -0.65249288 -0.46178952
#> 77   0.120977588  1.25315189  0.84209595
#> 78  -1.221405213 -0.03384571 -0.70354493
#> 79  -2.244111893 -0.54466094  0.27515704
#> 80  -1.847513959 -1.69224573 -1.04738598
#> 81   1.073375473  0.04580455  0.72117153
#> 82   0.184734069  0.27804166  0.65780636
#> 83   0.085860507 -0.18568367  0.02115438
#> 84  -0.199782580  1.15698772 -0.43305709
#> 85   0.020035570  0.59891962  0.91862219
#> 86  -0.859801491  0.05671391 -1.37049548
#> 87  -1.166308151  0.79442776 -1.29352999
#> 88   1.279851538 -1.02246542  0.19629225
#> 89  -0.056786823 -2.08530389 -0.73122392
#> 90  -0.227897669 -0.27859253  0.67400228
#> 91   1.404879547 -0.81179302 -0.29513205
#> 92  -1.325579512  0.95896719  0.20246814
#> 93  -0.454162671 -1.10493854  0.17663247
#> 94   1.327293537 -0.70163013 -0.30973451
#> 95  -0.621726678  1.04267030  0.27736113
#> 96   0.151704620  1.04065332  0.54616860
#> 97   0.769234352 -0.71746612 -1.01682891
#> 98   0.570374371 -2.16852015 -1.18810921
#> 99   1.038694500  1.14723653  2.20652663
#> 100 -0.968697017 -1.33467006 -1.00500030

rnorm_multi(100, 3, 0, 1, c(1, 0.2, -0.5, 0.2, 1, 0.5, -0.5, 0.5, 1), varnames=c("A", "B", "C"))

#>               A            B           C
#> 1   -2.76661330 -1.226537497 -0.07308821
#> 2   -0.93394099  0.039960253  0.83463025
#> 3   -0.93614444  0.472987580  1.15686571
#> 4   -2.17204176  0.224963355  2.41618501
#> 5   -0.25269429  0.253131770 -0.84206405
#> 6   -0.33877909 -0.410000027 -0.52316101
#> 7    0.96065138  0.335671165 -0.42052879
#> 8    0.12803731 -0.630146440  0.26165368
#> 9    0.47343795  1.243080850  1.46468650
#> 10  -0.89908363  1.036809603  1.38021529
#> 11  -0.46577705 -1.013846029 -0.01311433
#> 12   0.61722865  1.163960773 -0.40171116
#> 13   0.24721779 -0.382408941 -0.40424546
#> 14   1.21547065  0.051647440 -0.73674537
#> 15   0.29628990  0.604113287  0.08546142
#> 16   0.59045815  1.901089585  0.88171638
#> 17   0.45738652 -0.255371951 -0.55085547
#> 18   0.89811622  0.196530011 -0.47467335
#> 19  -1.38459333 -1.170197914  0.84471574
#> 20   0.14936018 -0.817112897 -0.27523779
#> 21  -1.26931503 -1.744872790 -0.61295484
#> 22   0.46241048  0.318020490 -0.11722484
#> 23  -0.49075669 -1.800664361 -0.81602411
#> 24  -1.44456645  0.905820867  1.78031538
#> 25  -0.27259067  0.574231089  0.75377412
#> 26  -0.62593700  0.665468834 -0.18485879
#> 27  -0.48387218 -0.592125827  0.73915920
#> 28  -1.51281624 -0.428518068  0.35462554
#> 29  -0.72786857 -0.467327406  0.81668961
#> 30  -0.02606863  1.316437101  0.41402043
#> 31  -1.83358300  0.459418709  0.94197614
#> 32  -2.12066302  0.596687743  0.71424085
#> 33   0.85658918  2.103329911  0.61188532
#> 34  -0.69730589  0.237213642  0.87912952
#> 35  -0.75473300 -0.438674366 -0.48795684
#> 36   1.51937177  0.510736398 -1.23877209
#> 37  -0.10846757  1.442567525  0.82430990
#> 38  -1.62025740  0.649752795  1.95372057
#> 39   0.57916894 -0.795842678 -0.02080833
#> 40  -0.14479211 -1.213065157 -0.54928606
#> 41  -1.20970237 -0.830338729 -0.11330828
#> 42   0.72034119  1.871730843  0.64506121
#> 43   0.36612959  1.799070846  1.39707225
#> 44  -1.53115211 -0.857566002  0.36251713
#> 45   0.23808933  0.705749375  1.19590154
#> 46  -0.17220031  1.844877652  0.31902249
#> 47  -0.64169494  0.999799498  0.45821655
#> 48   1.44584994  0.752648640 -1.17379273
#> 49  -0.58895639 -0.482447441 -0.25314196
#> 50  -0.82551848 -0.007871281 -0.03891820
#> 51  -1.49448258 -0.579500444  2.22417054
#> 52   0.96068919 -1.189290483 -0.87733227
#> 53  -1.59781123 -1.206229717  0.81778499
#> 54  -1.69641885 -1.297388456  0.90893931
#> 55   1.43945816  0.654478586  0.71532058
#> 56   0.82559682 -0.640593869 -1.04623587
#> 57  -1.14764679 -0.022530967  0.71447671
#> 58  -0.43120756  0.311918754  1.39006106
#> 59  -0.80262254  0.665742529  1.21791230
#> 60  -0.25080518 -0.089439563 -0.24405554
#> 61   0.03891411  1.214762250  0.31417079
#> 62  -0.18884404  1.187389420  1.67163658
#> 63  -0.50465366  0.615891205  0.52950097
#> 64  -0.42448777  0.116184059  0.98601732
#> 65   0.11119361 -0.142030373 -1.15359127
#> 66   0.26124642  1.708920228  0.99973307
#> 67   0.17390369  0.227053507 -0.39696921
#> 68   0.95211941  0.725241414 -1.11924830
#> 69   0.55494085  1.526592879  1.42264312
#> 70  -1.25545789 -0.839215905 -0.07752894
#> 71   0.62167173  0.399152292  0.26311775
#> 72   1.17854779  1.167054231  0.64815602
#> 73   0.61816537  1.592229914  0.46308299
#> 74  -0.13334226  1.206805871 -0.09216836
#> 75   0.46501468 -0.305830138 -0.54708670
#> 76  -0.28928238  1.596484677  1.53799525
#> 77  -1.34047198 -0.360145462 -0.70933658
#> 78   0.21808679  0.720547178  0.70320389
#> 79  -0.36928941  0.648894109  1.56089490
#> 80   0.56662360 -0.353804501  0.07889013
#> 81  -0.71038926 -0.743392232 -0.88947426
#> 82   0.95585222  0.935523427  0.18497840
#> 83   0.84610776  3.041141472  0.53663511
#> 84   0.54643504  0.087251742 -1.22266045
#> 85   0.57137003  1.233439948  0.44350043
#> 86  -1.40820001 -1.511809406 -0.56365802
#> 87   1.56163227  2.076162184 -0.65345781
#> 88  -1.82063566 -0.659841006  0.03169085
#> 89   0.27293075  0.305371208  0.83750777
#> 90   1.09035810  2.673374783 -0.18156038
#> 91  -1.23201496  0.872265052  1.89293326
#> 92  -0.93457496 -0.002236790  0.53061826
#> 93   0.09485478  0.206750365  0.49673672
#> 94   0.16935894 -1.322066436 -1.25325294
#> 95   1.57873634 -1.273265352 -2.10392912
#> 96  -1.35645433 -0.595687467  0.01857906
#> 97  -0.69041545 -1.818226670 -2.19537492
#> 98  -0.89563717  0.304137251  2.04027464
#> 99   1.77507581  0.718474794 -3.09184342
#> 100  0.92204569 -0.352932112 -0.31410476