Multiple correlated normal distributions

Make normally distributed vectors with specified relationships. See vignette("rnorm_multi", package = "faux") for details.

rnorm_multi(
  n = 100,
  vars = NULL,
  mu = 0,
  sd = 1,
  r = 0,
  varnames = NULL,
  empirical = FALSE,
  as.matrix = FALSE,
  seed = NULL
)

Arguments

n
the number of samples required

vars
the number of variables to return

mu
a vector giving the means of the variables (numeric vector of length 1 or vars)

sd
the standard deviations of the variables (numeric vector of length 1 or vars)

r
the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames
optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix
logical. If true, returns a matrix

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl of vars vectors

Examples

# 4 10-item vectors each correlated r = .5
rnorm_multi(10, 4, r = 0.5)

#>             X1         X2         X3          X4
#> 1   0.65088236 -1.5725673 -0.4379606  1.22320755
#> 2   2.06698870  3.1848750  1.4113127  0.75750764
#> 3  -0.16593405  0.1693742 -1.3032044  0.07178465
#> 4   2.60420026  2.0861865  1.2688811  2.40853848
#> 5  -0.63283937  0.3711274 -0.2808316 -0.08139221
#> 6   0.14468638  0.5790868  0.5638104 -1.17799550
#> 7  -0.24150804 -0.9870688 -0.4782898 -0.64592783
#> 8   1.34320986  0.2392003  1.8729049 -0.14845353
#> 9  -0.66713659  0.4574988 -0.8338574  0.16163391
#> 10 -0.08773183 -1.5996034  1.2314022  0.25082339

# set r with the upper right triangle
b <- rnorm_multi(100, 3, c(0, .5, 1), 1, 
                 r = c(0.2, -0.5, 0.5), 
                 varnames=c("A", "B", "C"))
cor(b)

#>             A          B          C
#> A  1.00000000 0.03395991 -0.5182312
#> B  0.03395991 1.00000000  0.6462245
#> C -0.51823120 0.64622455  1.0000000

# set r with a correlation matrix and column names from mu names
c <- rnorm_multi(
  n = 100, 
  mu = c(A = 0, B = 0.5, C = 1),
  r = c( 1,   0.2, -0.5, 
         0.2, 1,    0.5, 
        -0.5, 0.5,  1)
)
cor(c)

#>             A          B          C
#> A  1.00000000 0.08488627 -0.5397680
#> B  0.08488627 1.00000000  0.5918308
#> C -0.53976797 0.59183083  1.0000000