Multiple correlated normal distributions

Make normally distributed vectors with specified relationships. See vignette("rnorm_multi", package = "faux") for details.

rnorm_multi(
  n = 100,
  vars = NULL,
  mu = 0,
  sd = 1,
  r = 0,
  varnames = NULL,
  empirical = FALSE,
  as.matrix = FALSE,
  seed = NULL
)

Arguments

n

the number of samples required

vars

the number of variables to return

mu

a vector giving the means of the variables (numeric vector of length 1 or vars)

sd

the standard deviations of the variables (numeric vector of length 1 or vars)

r

the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames

optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical

logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix

logical. If true, returns a matrix

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl of vars vectors

Examples

# 4 10-item vectors each correlated r = .5
rnorm_multi(10, 4, r = 0.5)
#>            X1         X2            X3         X4
#> 1   0.9803694  1.0976654  0.0001894583  0.8452522
#> 2  -0.8020175 -0.9243474 -1.1450187701 -0.4347333
#> 3  -0.7241138 -1.4937781 -0.5307496604 -1.8831182
#> 4  -1.7623689 -1.6165031 -1.5172104943 -0.1424312
#> 5  -1.6716397 -1.0340988 -1.9206741566  0.5367103
#> 6   1.9368301  0.2323904  0.3411704463  0.8077779
#> 7  -0.9009644 -0.3444944 -1.3911371326  0.2576777
#> 8   0.9588554 -0.6597887 -0.1657576832  0.3789258
#> 9  -1.0027970  0.1128529 -0.7049475972 -1.0118001
#> 10 -1.1289531 -1.1962519 -0.6493908468 -0.8098978

# set r with the upper right triangle
b <- rnorm_multi(100, 3, c(0, .5, 1), 1, 
                 r = c(0.2, -0.5, 0.5), 
                 varnames=c("A", "B", "C"))
cor(b)
#>            A         B          C
#> A  1.0000000 0.2395587 -0.4563639
#> B  0.2395587 1.0000000  0.4425086
#> C -0.4563639 0.4425086  1.0000000

# set r with a correlation matrix and column names from mu names
c <- rnorm_multi(
  n = 100, 
  mu = c(A = 0, B = 0.5, C = 1),
  r = c( 1,   0.2, -0.5, 
         0.2, 1,    0.5, 
        -0.5, 0.5,  1)
)
cor(c)
#>            A         B          C
#> A  1.0000000 0.1607749 -0.5771048
#> B  0.1607749 1.0000000  0.4526899
#> C -0.5771048 0.4526899  1.0000000