Multiple correlated normal distributions

Make normally distributed vectors with specified relationships. See vignette("rnorm_multi", package = "faux") for details.

Usage

rnorm_multi(
  n = 100,
  vars = NULL,
  mu = 0,
  sd = 1,
  r = 0,
  varnames = NULL,
  empirical = FALSE,
  as.matrix = FALSE,
  seed = NULL
)

Arguments

n

the number of samples required

vars

the number of variables to return

mu

a vector giving the means of the variables (numeric vector of length 1 or vars)

sd

the standard deviations of the variables (numeric vector of length 1 or vars)

r

the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)

varnames

optional names for the variables (string vector of length vars) defaults if r is a matrix with column names

empirical

logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

as.matrix

logical. If true, returns a matrix

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl of vars vectors

Examples

# 4 10-item vectors each correlated r = .5
rnorm_multi(10, 4, r = 0.5)
#>              X1         X2          X3         X4
#> 1  -0.815780066 -1.2351185 -0.40726533 -1.6746540
#> 2  -0.963176706  0.5643239 -0.07538908  1.1697940
#> 3  -1.785264990  0.2400259 -1.55855641 -1.2618257
#> 4  -2.680583102 -2.7197955 -2.32027558 -1.8531943
#> 5   0.771659601 -0.3030900  0.73508147 -0.5872710
#> 6  -0.507771241 -1.8094514  0.50592089 -0.4763213
#> 7   1.469498042 -0.1037821  1.22485031  0.8477937
#> 8  -0.165759340  0.8674088  0.64154726  0.5942794
#> 9  -0.007278901 -0.5067206 -0.49613358 -1.3637240
#> 10  0.597292365  0.3570845  0.66837550 -0.9823218

# set r with the upper right triangle
b <- rnorm_multi(100, 3, c(0, .5, 1), 1, 
                 r = c(0.2, -0.5, 0.5), 
                 varnames=c("A", "B", "C"))
cor(b)
#>            A         B          C
#> A  1.0000000 0.2091250 -0.4688501
#> B  0.2091250 1.0000000  0.4498988
#> C -0.4688501 0.4498988  1.0000000

# set r with a correlation matrix and column names from mu names
c <- rnorm_multi(
  n = 100, 
  mu = c(A = 0, B = 0.5, C = 1),
  r = c( 1,   0.2, -0.5, 
         0.2, 1,    0.5, 
        -0.5, 0.5,  1)
)
cor(c)
#>            A         B          C
#> A  1.0000000 0.2142911 -0.5252217
#> B  0.2142911 1.0000000  0.4153675
#> C -0.5252217 0.4153675  1.0000000