Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id         val grand_i       sub_i      item_i         err
#> 1      S01     I01  0.64282330       0  1.20649622  0.44111382 -1.00478673
#> 2      S02     I01  3.14516916       0  1.07408920  0.44111382  1.62996614
#> 3      S03     I01 -1.43663866       0 -1.86244852  0.44111382 -0.01530396
#> 4      S04     I01  1.83870906       0  1.20906999  0.44111382  0.18852526
#> 5      S05     I01  1.55861500       0 -0.08769039  0.44111382  1.20519158
#> 6      S06     I01  0.56515998       0  0.62586368  0.44111382 -0.50181751
#> 7      S07     I01 -2.23288427       0 -0.90233632  0.44111382 -1.77166177
#> 8      S08     I01  0.70966944       0  0.84130873  0.44111382 -0.57275310
#> 9      S09     I01  1.71852643       0  1.03672788  0.44111382  0.24068473
#> 10     S10     I01  3.18016545       0  1.17310971  0.44111382  1.56594192
#> 11     S01     I02 -0.78802776       0  1.20649622 -1.90681772 -0.08770626
#> 12     S02     I02 -0.38525008       0  1.07408920 -1.90681772  0.44747844
#> 13     S03     I02 -3.97343307       0 -1.86244852 -1.90681772 -0.20416683
#> 14     S04     I02  0.77926671       0  1.20906999 -1.90681772  1.47701445
#> 15     S05     I02 -0.74755994       0 -0.08769039 -1.90681772  1.24694817
#> 16     S06     I02 -1.62323448       0  0.62586368 -1.90681772 -0.34228043
#> 17     S07     I02 -3.27958799       0 -0.90233632 -1.90681772 -0.47043395
#> 18     S08     I02 -2.12939491       0  0.84130873 -1.90681772 -1.06388591
#> 19     S09     I02 -1.20262184       0  1.03672788 -1.90681772 -0.33253200
#> 20     S10     I02 -1.36938393       0  1.17310971 -1.90681772 -0.63567592
#> 21     S01     I03  0.44148687       0  1.20649622 -0.29564945 -0.46935990
#> 22     S02     I03  0.92333128       0  1.07408920 -0.29564945  0.14489152
#> 23     S03     I03 -2.98685635       0 -1.86244852 -0.29564945 -0.82875838
#> 24     S04     I03  1.09536932       0  1.20906999 -0.29564945  0.18194878
#> 25     S05     I03 -2.07233533       0 -0.08769039 -0.29564945 -1.68899549
#> 26     S06     I03  0.10122287       0  0.62586368 -0.29564945 -0.22899135
#> 27     S07     I03 -3.07859647       0 -0.90233632 -0.29564945 -1.88061071
#> 28     S08     I03 -0.26391866       0  0.84130873 -0.29564945 -0.80957793
#> 29     S09     I03 -0.79064681       0  1.03672788 -0.29564945 -1.53172524
#> 30     S10     I03  1.15219664       0  1.17310971 -0.29564945  0.27473638
#> 31     S01     I04 -0.60320064       0  1.20649622 -0.67034125 -1.13935562
#> 32     S02     I04  1.10187506       0  1.07408920 -0.67034125  0.69812710
#> 33     S03     I04 -0.94686382       0 -1.86244852 -0.67034125  1.58592594
#> 34     S04     I04  0.56390417       0  1.20906999 -0.67034125  0.02517543
#> 35     S05     I04 -0.40325898       0 -0.08769039 -0.67034125  0.35477266
#> 36     S06     I04  0.86980689       0  0.62586368 -0.67034125  0.91428446
#> 37     S07     I04 -0.39495318       0 -0.90233632 -0.67034125  1.17772438
#> 38     S08     I04  0.75602440       0  0.84130873 -0.67034125  0.58505692
#> 39     S09     I04  0.98887526       0  1.03672788 -0.67034125  0.62248862
#> 40     S10     I04  0.51645424       0  1.17310971 -0.67034125  0.01368577
#> 41     S01     I05 -0.47352064       0  1.20649622 -0.15728455 -1.52273230
#> 42     S02     I05  0.65425032       0  1.07408920 -0.15728455 -0.26255433
#> 43     S03     I05 -1.20576321       0 -1.86244852 -0.15728455  0.81396986
#> 44     S04     I05  1.64764738       0  1.20906999 -0.15728455  0.59586195
#> 45     S05     I05  0.53549527       0 -0.08769039 -0.15728455  0.78047022
#> 46     S06     I05  0.49700336       0  0.62586368 -0.15728455  0.02842424
#> 47     S07     I05 -3.44003863       0 -0.90233632 -0.15728455 -2.38041775
#> 48     S08     I05  0.77972786       0  0.84130873 -0.15728455  0.09570369
#> 49     S09     I05  0.07468133       0  1.03672788 -0.15728455 -0.80476200
#> 50     S10     I05  0.31681054       0  1.17310971 -0.15728455 -0.69901461
#> 51     S01     I06  2.66357694       0  1.20649622  0.76831203  0.68876869
#> 52     S02     I06  2.02302413       0  1.07408920  0.76831203  0.18062289
#> 53     S03     I06 -1.73570909       0 -1.86244852  0.76831203 -0.64157260
#> 54     S04     I06  2.94669824       0  1.20906999  0.76831203  0.96931622
#> 55     S05     I06  1.18490192       0 -0.08769039  0.76831203  0.50428029
#> 56     S06     I06  0.71161640       0  0.62586368  0.76831203 -0.68255931
#> 57     S07     I06  0.37137220       0 -0.90233632  0.76831203  0.50539649
#> 58     S08     I06 -0.19329102       0  0.84130873  0.76831203 -1.80291178
#> 59     S09     I06  1.87752261       0  1.03672788  0.76831203  0.07248270
#> 60     S10     I06  1.50592157       0  1.17310971  0.76831203 -0.43550017
#> 61     S01     I07 -0.53841768       0  1.20649622 -1.06177086 -0.68314305
#> 62     S02     I07 -0.65042713       0  1.07408920 -1.06177086 -0.66274548
#> 63     S03     I07 -2.35268328       0 -1.86244852 -1.06177086  0.57153610
#> 64     S04     I07 -1.47867376       0  1.20906999 -1.06177086 -1.62597289
#> 65     S05     I07 -1.33118709       0 -0.08769039 -1.06177086 -0.18172584
#> 66     S06     I07 -1.07656756       0  0.62586368 -1.06177086 -0.64066038
#> 67     S07     I07 -0.25889054       0 -0.90233632 -1.06177086  1.70521663
#> 68     S08     I07  0.28045567       0  0.84130873 -1.06177086  0.50091780
#> 69     S09     I07  0.69855179       0  1.03672788 -1.06177086  0.72359477
#> 70     S10     I07  2.12081980       0  1.17310971 -1.06177086  2.00948095
#> 71     S01     I08  1.14767649       0  1.20649622 -0.22796548  0.16914575
#> 72     S02     I08 -0.07343508       0  1.07408920 -0.22796548 -0.91955881
#> 73     S03     I08 -2.79844535       0 -1.86244852 -0.22796548 -0.70803136
#> 74     S04     I08  2.29867645       0  1.20906999 -0.22796548  1.31757194
#> 75     S05     I08 -0.17348734       0 -0.08769039 -0.22796548  0.14216853
#> 76     S06     I08  0.09082893       0  0.62586368 -0.22796548 -0.30706927
#> 77     S07     I08  0.22086089       0 -0.90233632 -0.22796548  1.35116269
#> 78     S08     I08  0.26523519       0  0.84130873 -0.22796548 -0.34810806
#> 79     S09     I08 -0.10887464       0  1.03672788 -0.22796548 -0.91763704
#> 80     S10     I08 -0.50384902       0  1.17310971 -0.22796548 -1.44899326
#> 81     S01     I09 -0.88449477       0  1.20649622 -0.73865953 -1.35233146
#> 82     S02     I09 -0.95282156       0  1.07408920 -0.73865953 -1.28825123
#> 83     S03     I09 -2.05355447       0 -1.86244852 -0.73865953  0.54755358
#> 84     S04     I09  0.58019376       0  1.20906999 -0.73865953  0.10978331
#> 85     S05     I09  0.48038373       0 -0.08769039 -0.73865953  1.30673366
#> 86     S06     I09 -0.11495201       0  0.62586368 -0.73865953 -0.00215616
#> 87     S07     I09 -1.59255839       0 -0.90233632 -0.73865953  0.04843746
#> 88     S08     I09 -0.38456174       0  0.84130873 -0.73865953 -0.48721094
#> 89     S09     I09  0.54835367       0  1.03672788 -0.73865953  0.25028531
#> 90     S10     I09  1.30404117       0  1.17310971 -0.73865953  0.86959099
#> 91     S01     I10  1.49413571       0  1.20649622 -0.07112165  0.35876114
#> 92     S02     I10  0.39772236       0  1.07408920 -0.07112165 -0.60524519
#> 93     S03     I10 -1.61841733       0 -1.86244852 -0.07112165  0.31515284
#> 94     S04     I10  0.25840786       0  1.20906999 -0.07112165 -0.87954048
#> 95     S05     I10 -0.49991820       0 -0.08769039 -0.07112165 -0.34110616
#> 96     S06     I10 -0.22686254       0  0.62586368 -0.07112165 -0.78160457
#> 97     S07     I10 -1.39855869       0 -0.90233632 -0.07112165 -0.42510072
#> 98     S08     I10 -0.86421004       0  0.84130873 -0.07112165 -1.63439712
#> 99     S09     I10  1.82776429       0  1.03672788 -0.07112165  0.86215805
#> 100    S10     I10  1.84218645       0  1.17310971 -0.07112165  0.74019839