Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i      sub_i      item_i         err
#> 1      S01     I01 -0.52223836       0 -0.5634951 -0.83959143  0.88084817
#> 2      S02     I01 -1.80665640       0 -0.3393464 -0.83959143 -0.62771855
#> 3      S03     I01  1.47479106       0  1.4571920 -0.83959143  0.85719052
#> 4      S04     I01  0.92776437       0  0.7892063 -0.83959143  0.97814950
#> 5      S05     I01 -0.13496731       0  0.2783617 -0.83959143  0.42626244
#> 6      S06     I01  1.10896782       0 -0.1466364 -0.83959143  2.09519567
#> 7      S07     I01 -1.95864338       0 -1.4367902 -0.83959143  0.31773821
#> 8      S08     I01 -2.45116838       0 -0.6365287 -0.83959143 -0.97504824
#> 9      S09     I01 -2.74241120       0 -0.8846102 -0.83959143 -1.01820956
#> 10     S10     I01 -1.59457718       0 -0.3461640 -0.83959143 -0.40882170
#> 11     S01     I02 -0.27428403       0 -0.5634951  1.34544328 -1.05623222
#> 12     S02     I02  0.05047898       0 -0.3393464  1.34544328 -0.95561788
#> 13     S03     I02  3.03553641       0  1.4571920  1.34544328  0.23290116
#> 14     S04     I02  1.40242916       0  0.7892063  1.34544328 -0.73222042
#> 15     S05     I02  1.86257002       0  0.2783617  1.34544328  0.23876506
#> 16     S06     I02  1.69854502       0 -0.1466364  1.34544328  0.49973816
#> 17     S07     I02 -1.03614133       0 -1.4367902  1.34544328 -0.94479446
#> 18     S08     I02  0.69871546       0 -0.6365287  1.34544328 -0.01019911
#> 19     S09     I02 -1.02914302       0 -0.8846102  1.34544328 -1.48997609
#> 20     S10     I02  1.46055150       0 -0.3461640  1.34544328  0.46127226
#> 21     S01     I03  0.06249064       0 -0.5634951  0.47154658  0.15443916
#> 22     S02     I03  0.74183084       0 -0.3393464  0.47154658  0.60963068
#> 23     S03     I03  1.99853657       0  1.4571920  0.47154658  0.06979803
#> 24     S04     I03  1.44035589       0  0.7892063  0.47154658  0.17960301
#> 25     S05     I03  0.17691158       0  0.2783617  0.47154658 -0.57299667
#> 26     S06     I03  1.03461872       0 -0.1466364  0.47154658  0.70970857
#> 27     S07     I03 -0.26987140       0 -1.4367902  0.47154658  0.69537217
#> 28     S08     I03 -1.34563322       0 -0.6365287  0.47154658 -1.18065108
#> 29     S09     I03  0.76058095       0 -0.8846102  0.47154658  1.17364459
#> 30     S10     I03  0.60353930       0 -0.3461640  0.47154658  0.47815677
#> 31     S01     I04  2.22206572       0 -0.5634951  1.32958213  1.45597868
#> 32     S02     I04  0.50379012       0 -0.3393464  1.32958213 -0.48644560
#> 33     S03     I04  4.08011383       0  1.4571920  1.32958213  1.29333974
#> 34     S04     I04  2.58322863       0  0.7892063  1.32958213  0.46444019
#> 35     S05     I04  0.45515754       0  0.2783617  1.32958213 -1.15278627
#> 36     S06     I04  0.54718393       0 -0.1466364  1.32958213 -0.63576177
#> 37     S07     I04 -0.45745822       0 -1.4367902  1.32958213 -0.35025019
#> 38     S08     I04  0.83923500       0 -0.6365287  1.32958213  0.14618159
#> 39     S09     I04  2.85802291       0 -0.8846102  1.32958213  2.41305099
#> 40     S10     I04  1.01755414       0 -0.3461640  1.32958213  0.03413606
#> 41     S01     I05  0.65308632       0 -0.5634951  1.37396848 -0.15738706
#> 42     S02     I05  1.48793789       0 -0.3393464  1.37396848  0.45331583
#> 43     S03     I05  3.52736856       0  1.4571920  1.37396848  0.69620812
#> 44     S04     I05  2.51721700       0  0.7892063  1.37396848  0.35404223
#> 45     S05     I05  2.61948899       0  0.2783617  1.37396848  0.96715884
#> 46     S06     I05  1.19045055       0 -0.1466364  1.37396848 -0.03688150
#> 47     S07     I05 -1.48400038       0 -1.4367902  1.37396848 -1.42117870
#> 48     S08     I05  1.43801561       0 -0.6365287  1.37396848  0.70057585
#> 49     S09     I05  0.17491388       0 -0.8846102  1.37396848 -0.31444438
#> 50     S10     I05  0.68425132       0 -0.3461640  1.37396848 -0.34355310
#> 51     S01     I06 -1.06827572       0 -0.5634951 -0.27383307 -0.23094756
#> 52     S02     I06 -0.22385408       0 -0.3393464 -0.27383307  0.38932541
#> 53     S03     I06  0.90873960       0  1.4571920 -0.27383307 -0.27461929
#> 54     S04     I06 -1.31573241       0  0.7892063 -0.27383307 -1.83110564
#> 55     S05     I06  0.76780742       0  0.2783617 -0.27383307  0.76327881
#> 56     S06     I06  0.05371287       0 -0.1466364 -0.27383307  0.47418237
#> 57     S07     I06 -1.07619967       0 -1.4367902 -0.27383307  0.63442356
#> 58     S08     I06  1.49667198       0 -0.6365287 -0.27383307  2.40703376
#> 59     S09     I06 -1.79324619       0 -0.8846102 -0.27383307 -0.63480291
#> 60     S10     I06 -1.02082617       0 -0.3461640 -0.27383307 -0.40082906
#> 61     S01     I07  0.87674246       0 -0.5634951  1.37900960  0.06122795
#> 62     S02     I07  1.67872847       0 -0.3393464  1.37900960  0.63906529
#> 63     S03     I07  2.40774536       0  1.4571920  1.37900960 -0.42845620
#> 64     S04     I07  1.67066036       0  0.7892063  1.37900960 -0.49755554
#> 65     S05     I07  1.18350915       0  0.2783617  1.37900960 -0.47386212
#> 66     S06     I07  4.30283406       0 -0.1466364  1.37900960  3.07046088
#> 67     S07     I07 -0.17989864       0 -1.4367902  1.37900960 -0.12211808
#> 68     S08     I07  0.97042981       0 -0.6365287  1.37900960  0.22794892
#> 69     S09     I07  1.05219992       0 -0.8846102  1.37900960  0.55780053
#> 70     S10     I07  2.24657126       0 -0.3461640  1.37900960  1.21372571
#> 71     S01     I08 -0.32619607       0 -0.5634951  0.12755768  0.10974134
#> 72     S02     I08 -0.56587260       0 -0.3393464  0.12755768 -0.35408386
#> 73     S03     I08  2.71829438       0  1.4571920  0.12755768  1.13354473
#> 74     S04     I08  0.28081356       0  0.7892063  0.12755768 -0.63595042
#> 75     S05     I08  2.26387569       0  0.2783617  0.12755768  1.85795634
#> 76     S06     I08  0.69621881       0 -0.1466364  0.12755768  0.71529755
#> 77     S07     I08 -2.34486987       0 -1.4367902  0.12755768 -1.03563740
#> 78     S08     I08  1.16280748       0 -0.6365287  0.12755768  1.67177852
#> 79     S09     I08 -1.24503232       0 -0.8846102  0.12755768 -0.48797979
#> 80     S10     I08  0.30114872       0 -0.3461640  0.12755768  0.51975509
#> 81     S01     I09  1.13784331       0 -0.5634951 -0.07307081  1.77440921
#> 82     S02     I09  0.70805976       0 -0.3393464 -0.07307081  1.12047699
#> 83     S03     I09  2.39359588       0  1.4571920 -0.07307081  1.00947472
#> 84     S04     I09 -0.29945562       0  0.7892063 -0.07307081 -1.01559111
#> 85     S05     I09  0.42263565       0  0.2783617 -0.07307081  0.21734479
#> 86     S06     I09 -0.68282996       0 -0.1466364 -0.07307081 -0.46312272
#> 87     S07     I09 -2.04299088       0 -1.4367902 -0.07307081 -0.53312991
#> 88     S08     I09 -0.58539929       0 -0.6365287 -0.07307081  0.12420023
#> 89     S09     I09 -3.16132742       0 -0.8846102 -0.07307081 -2.20364640
#> 90     S10     I09 -0.42983251       0 -0.3461640 -0.07307081 -0.01059765
#> 91     S01     I10 -1.11213574       0 -0.5634951 -0.06787327 -0.48076738
#> 92     S02     I10 -0.02073002       0 -0.3393464 -0.06787327  0.38648967
#> 93     S03     I10  0.57998557       0  1.4571920 -0.06787327 -0.80933313
#> 94     S04     I10  1.62306885       0  0.7892063 -0.06787327  0.90173582
#> 95     S05     I10  1.70847654       0  0.2783617 -0.06787327  1.49798813
#> 96     S06     I10  0.01606598       0 -0.1466364 -0.06787327  0.23057567
#> 97     S07     I10 -3.12176368       0 -1.4367902 -0.06787327 -1.61710026
#> 98     S08     I10 -1.05312250       0 -0.6365287 -0.06787327 -0.34872052
#> 99     S09     I10 -1.85907721       0 -0.8846102 -0.06787327 -0.90659373
#> 100    S10     I10 -0.43187812       0 -0.3461640 -0.06787327 -0.01784081