Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i       sub_i     item_i         err
#> 1      S01     I01  0.40979961       0  0.50401285 -0.6998284  0.60561513
#> 2      S02     I01  0.54666646       0  0.05960143 -0.6998284  1.18689339
#> 3      S03     I01 -0.80142928       0 -0.40966864 -0.6998284  0.30806772
#> 4      S04     I01 -0.78585066       0  0.72624097 -0.6998284 -0.81226326
#> 5      S05     I01 -0.29943561       0 -0.34722981 -0.6998284  0.74762256
#> 6      S06     I01  0.50515082       0  0.19876863 -0.6998284  1.00621056
#> 7      S07     I01 -0.67285435       0  0.37872362 -0.6998284 -0.35174960
#> 8      S08     I01 -3.42342536       0 -0.96101296 -0.6998284 -1.76258403
#> 9      S09     I01  0.73979186       0  0.15174082 -0.6998284  1.28787940
#> 10     S10     I01 -0.06132409       0  1.52927149 -0.6998284 -0.89076721
#> 11     S01     I02  1.39123818       0  0.50401285  0.6845270  0.20269832
#> 12     S02     I02  0.38603268       0  0.05960143  0.6845270 -0.35809578
#> 13     S03     I02  1.66518330       0 -0.40966864  0.6845270  1.39032492
#> 14     S04     I02  1.64034143       0  0.72624097  0.6845270  0.22957344
#> 15     S05     I02 -1.43047284       0 -0.34722981  0.6845270 -1.76777006
#> 16     S06     I02  2.04077341       0  0.19876863  0.6845270  1.15747776
#> 17     S07     I02  0.94323379       0  0.37872362  0.6845270 -0.12001685
#> 18     S08     I02  0.02393215       0 -0.96101296  0.6845270  0.30041810
#> 19     S09     I02  0.44551515       0  0.15174082  0.6845270 -0.39075270
#> 20     S10     I02  3.44057516       0  1.52927149  0.6845270  1.22677665
#> 21     S01     I03 -1.23477380       0  0.50401285 -0.3073055 -1.43148118
#> 22     S02     I03  0.15326721       0  0.05960143 -0.3073055  0.40097125
#> 23     S03     I03 -0.46510997       0 -0.40966864 -0.3073055  0.25186413
#> 24     S04     I03 -0.68356731       0  0.72624097 -0.3073055 -1.10250282
#> 25     S05     I03 -1.69670314       0 -0.34722981 -0.3073055 -1.04216787
#> 26     S06     I03 -1.74055594       0  0.19876863 -0.3073055 -1.63201911
#> 27     S07     I03 -1.15933687       0  0.37872362 -0.3073055 -1.23075502
#> 28     S08     I03 -0.79352443       0 -0.96101296 -0.3073055  0.47479400
#> 29     S09     I03 -0.42064278       0  0.15174082 -0.3073055 -0.26507814
#> 30     S10     I03  1.18328230       0  1.52927149 -0.3073055 -0.03868372
#> 31     S01     I04 -0.93516730       0  0.50401285 -1.1094537 -0.32972644
#> 32     S02     I04 -1.44680768       0  0.05960143 -1.1094537 -0.39695540
#> 33     S03     I04 -3.18768580       0 -0.40966864 -1.1094537 -1.66856345
#> 34     S04     I04 -0.81229590       0  0.72624097 -1.1094537 -0.42908316
#> 35     S05     I04 -0.32268651       0 -0.34722981 -1.1094537  1.13399701
#> 36     S06     I04  0.17094716       0  0.19876863 -1.1094537  1.08163224
#> 37     S07     I04 -1.29189838       0  0.37872362 -1.1094537 -0.56116829
#> 38     S08     I04 -1.48776068       0 -0.96101296 -1.1094537  0.58270600
#> 39     S09     I04 -1.63863494       0  0.15174082 -1.1094537 -0.68092205
#> 40     S10     I04  0.90222427       0  1.52927149 -1.1094537  0.48240650
#> 41     S01     I05 -0.08320141       0  0.50401285 -0.8719208  0.28470652
#> 42     S02     I05 -0.16087355       0  0.05960143 -0.8719208  0.65144579
#> 43     S03     I05 -2.03102928       0 -0.40966864 -0.8719208 -0.74943987
#> 44     S04     I05 -1.05065885       0  0.72624097 -0.8719208 -0.90497904
#> 45     S05     I05 -0.73205384       0 -0.34722981 -0.8719208  0.48709675
#> 46     S06     I05 -0.19732282       0  0.19876863 -0.8719208  0.47582933
#> 47     S07     I05 -1.36429926       0  0.37872362 -0.8719208 -0.87110210
#> 48     S08     I05 -1.05804429       0 -0.96101296 -0.8719208  0.77488945
#> 49     S09     I05 -0.32992511       0  0.15174082 -0.8719208  0.39025484
#> 50     S10     I05  2.53087943       0  1.52927149 -0.8719208  1.87352872
#> 51     S01     I06  0.21410124       0  0.50401285 -1.1341598  0.84424821
#> 52     S02     I06 -1.71290890       0  0.05960143 -1.1341598 -0.63835051
#> 53     S03     I06 -2.94691740       0 -0.40966864 -1.1341598 -1.40308894
#> 54     S04     I06 -3.04861023       0  0.72624097 -1.1341598 -2.64069138
#> 55     S05     I06 -3.15310855       0 -0.34722981 -1.1341598 -1.67171891
#> 56     S06     I06 -0.86194193       0  0.19876863 -1.1341598  0.07344926
#> 57     S07     I06 -1.12050829       0  0.37872362 -1.1341598 -0.36507208
#> 58     S08     I06 -0.55752103       0 -0.96101296 -1.1341598  1.53765176
#> 59     S09     I06 -0.39294856       0  0.15174082 -1.1341598  0.58947044
#> 60     S10     I06  1.38023963       0  1.52927149 -1.1341598  0.98512796
#> 61     S01     I07 -0.63048925       0  0.50401285 -1.7568253  0.62232318
#> 62     S02     I07 -1.98315328       0  0.05960143 -1.7568253 -0.28592943
#> 63     S03     I07 -0.61878206       0 -0.40966864 -1.7568253  1.54771186
#> 64     S04     I07  0.48772967       0  0.72624097 -1.7568253  1.51831398
#> 65     S05     I07 -2.70388447       0 -0.34722981 -1.7568253 -0.59982938
#> 66     S06     I07 -1.12631444       0  0.19876863 -1.7568253  0.43174221
#> 67     S07     I07 -1.28871465       0  0.37872362 -1.7568253  0.08938701
#> 68     S08     I07 -2.54469683       0 -0.96101296 -1.7568253  0.17314141
#> 69     S09     I07 -3.12551924       0  0.15174082 -1.7568253 -1.52043479
#> 70     S10     I07 -0.25949185       0  1.52927149 -1.7568253 -0.03193806
#> 71     S01     I08 -0.04496929       0  0.50401285 -0.8734816  0.32449943
#> 72     S02     I08 -0.99154561       0  0.05960143 -0.8734816 -0.17766547
#> 73     S03     I08 -2.04013946       0 -0.40966864 -0.8734816 -0.75698925
#> 74     S04     I08 -1.74144927       0  0.72624097 -0.8734816 -1.59420867
#> 75     S05     I08 -1.17797498       0 -0.34722981 -0.8734816  0.04273641
#> 76     S06     I08  0.51599552       0  0.19876863 -0.8734816  1.19070846
#> 77     S07     I08  0.40683762       0  0.37872362 -0.8734816  0.90159557
#> 78     S08     I08 -1.60105507       0 -0.96101296 -0.8734816  0.23343947
#> 79     S09     I08 -0.61417767       0  0.15174082 -0.8734816  0.10756308
#> 80     S10     I08  1.47566933       0  1.52927149 -0.8734816  0.81987942
#> 81     S01     I09  1.77845997       0  0.50401285  0.9097883  0.36465883
#> 82     S02     I09  1.02812843       0  0.05960143  0.9097883  0.05873870
#> 83     S03     I09  1.57064881       0 -0.40966864  0.9097883  1.07052915
#> 84     S04     I09  1.23234759       0  0.72624097  0.9097883 -0.40368168
#> 85     S05     I09 -0.13031039       0 -0.34722981  0.9097883 -0.69286888
#> 86     S06     I09  1.18272442       0  0.19876863  0.9097883  0.07416749
#> 87     S07     I09  0.48427446       0  0.37872362  0.9097883 -0.80423745
#> 88     S08     I09  0.87949909       0 -0.96101296  0.9097883  0.93072375
#> 89     S09     I09  0.95488064       0  0.15174082  0.9097883 -0.10664848
#> 90     S10     I09  1.49313332       0  1.52927149  0.9097883 -0.94592647
#> 91     S01     I10  0.32599687       0  0.50401285 -1.0694739  0.89145795
#> 92     S02     I10  0.26506997       0  0.05960143 -1.0694739  1.27494247
#> 93     S03     I10 -2.68626331       0 -0.40966864 -1.0694739 -1.20712075
#> 94     S04     I10 -1.64394389       0  0.72624097 -1.0694739 -1.30071093
#> 95     S05     I10 -1.38742166       0 -0.34722981 -1.0694739  0.02928208
#> 96     S06     I10 -1.80262641       0  0.19876863 -1.0694739 -0.93192111
#> 97     S07     I10 -0.85695452       0  0.37872362 -1.0694739 -0.16620421
#> 98     S08     I10 -2.76777742       0 -0.96101296 -1.0694739 -0.73729053
#> 99     S09     I10 -0.81014363       0  0.15174082 -1.0694739  0.10758947
#> 100    S10     I10 -0.20305984       0  1.52927149 -1.0694739 -0.66285740