Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, empirical = FALSE, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id            y grand_i       sub_i      item_i          err
#> 1      S01     I01 -0.113439555       0 -0.13940586  0.73761259 -0.711646284
#> 2      S02     I01  2.230829805       0  1.16826379  0.73761259  0.324953430
#> 3      S03     I01  2.824122236       0  0.33700758  0.73761259  1.749502063
#> 4      S04     I01 -0.179892938       0  0.09314852  0.73761259 -1.010654046
#> 5      S05     I01 -0.721732375       0 -1.22332650  0.73761259 -0.236018469
#> 6      S06     I01 -0.975110643       0 -1.74283246  0.73761259  0.030109230
#> 7      S07     I01  0.652475786       0  1.15027657  0.73761259 -1.235413376
#> 8      S08     I01 -0.150330493       0 -0.13930795  0.73761259 -0.748635133
#> 9      S09     I01  0.731980423       0  0.42087195  0.73761259 -0.426504121
#> 10     S10     I01 -1.452925179       0 -0.39166755  0.73761259 -1.798870215
#> 11     S01     I02  0.444429935       0 -0.13940586  1.59081993 -1.006984131
#> 12     S02     I02  1.244568664       0  1.16826379  1.59081993 -1.514515048
#> 13     S03     I02  4.138769237       0  0.33700758  1.59081993  2.210941727
#> 14     S04     I02  0.573334561       0  0.09314852  1.59081993 -1.110633884
#> 15     S05     I02  0.381647276       0 -1.22332650  1.59081993  0.014153846
#> 16     S06     I02  0.464242109       0 -1.74283246  1.59081993  0.616254646
#> 17     S07     I02  3.071365624       0  1.15027657  1.59081993  0.330269125
#> 18     S08     I02  0.157648669       0 -0.13930795  1.59081993 -1.293863308
#> 19     S09     I02  2.568475199       0  0.42087195  1.59081993  0.556783319
#> 20     S10     I02  1.010196290       0 -0.39166755  1.59081993 -0.188956082
#> 21     S01     I03  0.003802968       0 -0.13940586 -0.15767936  0.300888190
#> 22     S02     I03  1.357293685       0  1.16826379 -0.15767936  0.346709260
#> 23     S03     I03  2.037919270       0  0.33700758 -0.15767936  1.858591047
#> 24     S04     I03  0.182804480       0  0.09314852 -0.15767936  0.247335323
#> 25     S05     I03 -1.952725799       0 -1.22332650 -0.15767936 -0.571719941
#> 26     S06     I03 -1.126590575       0 -1.74283246 -0.15767936  0.773921249
#> 27     S07     I03 -0.411589344       0  1.15027657 -0.15767936 -1.404186556
#> 28     S08     I03 -0.025500231       0 -0.13930795 -0.15767936  0.271487080
#> 29     S09     I03 -1.827321174       0  0.42087195 -0.15767936 -2.090513766
#> 30     S10     I03 -2.452761635       0 -0.39166755 -0.15767936 -1.903414721
#> 31     S01     I04 -1.054230384       0 -0.13940586 -0.40660552 -0.508219001
#> 32     S02     I04  0.861032902       0  1.16826379 -0.40660552  0.099374638
#> 33     S03     I04 -0.284071987       0  0.33700758 -0.40660552 -0.214474049
#> 34     S04     I04 -0.963685430       0  0.09314852 -0.40660552 -0.650228427
#> 35     S05     I04  0.380395229       0 -1.22332650 -0.40660552  2.010327246
#> 36     S06     I04 -1.798757840       0 -1.74283246 -0.40660552  0.350680144
#> 37     S07     I04  2.036115393       0  1.15027657 -0.40660552  1.292444342
#> 38     S08     I04  0.267097833       0 -0.13930795 -0.40660552  0.813011304
#> 39     S09     I04 -0.735172899       0  0.42087195 -0.40660552 -0.749439331
#> 40     S10     I04  0.652002183       0 -0.39166755 -0.40660552  1.450275258
#> 41     S01     I05 -0.074045406       0 -0.13940586 -0.65067392  0.716034370
#> 42     S02     I05  0.268659887       0  1.16826379 -0.65067392 -0.248929984
#> 43     S03     I05  0.829427733       0  0.33700758 -0.65067392  1.143094065
#> 44     S04     I05 -0.654479730       0  0.09314852 -0.65067392 -0.096954333
#> 45     S05     I05 -1.720509867       0 -1.22332650 -0.65067392  0.153490545
#> 46     S06     I05 -2.973656978       0 -1.74283246 -0.65067392 -0.580150600
#> 47     S07     I05  1.036000876       0  1.15027657 -0.65067392  0.536398218
#> 48     S08     I05  0.966974732       0 -0.13930795 -0.65067392  1.756956596
#> 49     S09     I05  1.365261511       0  0.42087195 -0.65067392  1.595063472
#> 50     S10     I05 -2.772460235       0 -0.39166755 -0.65067392 -1.730118766
#> 51     S01     I06  0.846943336       0 -0.13940586  2.02042815 -1.034078950
#> 52     S02     I06  3.346673526       0  1.16826379  2.02042815  0.157981593
#> 53     S03     I06  3.767610080       0  0.33700758  2.02042815  1.410174350
#> 54     S04     I06  1.293351256       0  0.09314852  2.02042815 -0.820225409
#> 55     S05     I06  2.502510809       0 -1.22332650  2.02042815  1.705409159
#> 56     S06     I06 -0.032643461       0 -1.74283246  2.02042815 -0.310239144
#> 57     S07     I06  3.382947034       0  1.15027657  2.02042815  0.212242315
#> 58     S08     I06  2.910437885       0 -0.13930795  2.02042815  1.029317687
#> 59     S09     I06  2.544978935       0  0.42087195  2.02042815  0.103678835
#> 60     S10     I06  0.418197223       0 -0.39166755  2.02042815 -1.210563370
#> 61     S01     I07 -2.823281514       0 -0.13940586 -2.12339449 -0.560481163
#> 62     S02     I07 -2.182694296       0  1.16826379 -2.12339449 -1.227563593
#> 63     S03     I07 -1.041101646       0  0.33700758 -2.12339449  0.745285261
#> 64     S04     I07 -2.479678587       0  0.09314852 -2.12339449 -0.449432616
#> 65     S05     I07 -3.531129956       0 -1.22332650 -2.12339449 -0.184408969
#> 66     S06     I07 -5.594485688       0 -1.74283246 -2.12339449 -1.728258736
#> 67     S07     I07  0.022576521       0  1.15027657 -2.12339449  0.995694438
#> 68     S08     I07 -2.227437925       0 -0.13930795 -2.12339449  0.035264515
#> 69     S09     I07  0.119785388       0  0.42087195 -2.12339449  1.822307924
#> 70     S10     I07 -2.519086231       0 -0.39166755 -2.12339449 -0.004024187
#> 71     S01     I08 -0.961286928       0 -0.13940586 -1.34238572  0.520504655
#> 72     S02     I08  0.181027792       0  1.16826379 -1.34238572  0.355149728
#> 73     S03     I08 -2.671397025       0  0.33700758 -1.34238572 -1.666018886
#> 74     S04     I08 -0.745833894       0  0.09314852 -1.34238572  0.503403311
#> 75     S05     I08 -3.035516873       0 -1.22332650 -1.34238572 -0.469804654
#> 76     S06     I08 -1.719728730       0 -1.74283246 -1.34238572  1.365489455
#> 77     S07     I08 -0.114054813       0  1.15027657 -1.34238572  0.078054336
#> 78     S08     I08 -0.163525308       0 -0.13930795 -1.34238572  1.318168364
#> 79     S09     I08 -1.334602127       0  0.42087195 -1.34238572 -0.413088359
#> 80     S10     I08 -1.351703876       0 -0.39166755 -1.34238572  0.382349400
#> 81     S01     I09  1.946242147       0 -0.13940586  1.00919292  1.076455089
#> 82     S02     I09  2.242287453       0  1.16826379  1.00919292  0.064830749
#> 83     S03     I09  1.188715655       0  0.33700758  1.00919292 -0.157484847
#> 84     S04     I09  1.275311709       0  0.09314852  1.00919292  0.172970273
#> 85     S05     I09 -0.884355253       0 -1.22332650  1.00919292 -0.670221675
#> 86     S06     I09 -1.282979497       0 -1.74283246  1.00919292 -0.549339952
#> 87     S07     I09  1.082713146       0  1.15027657  1.00919292 -1.076756345
#> 88     S08     I09  1.035240485       0 -0.13930795  1.00919292  0.165355516
#> 89     S09     I09  3.755942620       0  0.42087195  1.00919292  2.325877748
#> 90     S10     I09  0.079086468       0 -0.39166755  1.00919292 -0.538438897
#> 91     S01     I10  0.178569123       0 -0.13940586 -0.09098883  0.408963818
#> 92     S02     I10  0.192259915       0  1.16826379 -0.09098883 -0.885015037
#> 93     S03     I10  0.077315228       0  0.33700758 -0.09098883 -0.168703521
#> 94     S04     I10 -0.536876922       0  0.09314852 -0.09098883 -0.539036606
#> 95     S05     I10 -3.115161011       0 -1.22332650 -0.09098883 -1.800845680
#> 96     S06     I10 -2.477438976       0 -1.74283246 -0.09098883 -0.643617680
#> 97     S07     I10  2.458609717       0  1.15027657 -0.09098883  1.399321978
#> 98     S08     I10  1.241869474       0 -0.13930795 -0.09098883  1.472166257
#> 99     S09     I10 -0.872428721       0  0.42087195 -0.09098883 -1.202311841
#> 100    S10     I10  0.297415929       0 -0.39166755 -0.09098883  0.780072316