Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n

the number of subjects

item_n

the number of items

grand_i

the grand intercept (overall mean)

sub_sd

the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd

the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd

the SD of the error term

empirical

Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)
#>     sub_id item_id           y grand_i      sub_i      item_i          err
#> 1      S01     I01  0.64028791       0  0.5578782  0.46851960 -0.386109893
#> 15     S01     I02  1.28575527       0  0.5578782  0.06850421  0.659372864
#> 29     S01     I03  3.23804134       0  0.5578782  1.91698874  0.763174403
#> 31     S01     I04 -0.28775036       0  0.5578782 -1.48932282  0.643694255
#> 44     S01     I05  0.01916476       0  0.5578782 -0.03200038 -0.506713054
#> 58     S01     I06  0.44026144       0  0.5578782  1.54188817 -1.659504933
#> 61     S01     I07  2.72653393       0  0.5578782  1.56028336  0.608372375
#> 73     S01     I08  0.97890241       0  0.5578782  0.02755786  0.393466357
#> 87     S01     I09  2.09281240       0  0.5578782 -0.81124192  2.346176114
#> 91     S01     I10 -1.98374209       0  0.5578782 -2.61271975  0.071099465
#> 7      S02     I01  1.68359479       0  0.7835692  0.46851960  0.431505968
#> 17     S02     I02  1.18790694       0  0.7835692  0.06850421  0.335833517
#> 30     S02     I03  3.57483586       0  0.7835692  1.91698874  0.874277899
#> 36     S02     I04 -1.86712241       0  0.7835692 -1.48932282 -1.161368807
#> 46     S02     I05  0.78198191       0  0.7835692 -0.03200038  0.030413079
#> 59     S02     I06  0.90390870       0  0.7835692  1.54188817 -1.421548686
#> 69     S02     I07  4.09389652       0  0.7835692  1.56028336  1.750043951
#> 75     S02     I08  1.01452316       0  0.7835692  0.02755786  0.203396091
#> 88     S02     I09 -0.47198039       0  0.7835692 -0.81124192 -0.444307684
#> 98     S02     I10 -1.22696674       0  0.7835692 -2.61271975  0.602183794
#> 4      S03     I01 -1.48691659       0 -1.3964965  0.46851960 -0.558939662
#> 18     S03     I02 -0.42924736       0 -1.3964965  0.06850421  0.898744965
#> 23     S03     I03 -0.15822912       0 -1.3964965  1.91698874 -0.678721332
#> 33     S03     I04 -3.23558266       0 -1.3964965 -1.48932282 -0.349763313
#> 47     S03     I05 -1.43092070       0 -1.3964965 -0.03200038 -0.002423784
#> 60     S03     I06  0.30418948       0 -1.3964965  1.54188817  0.158797836
#> 62     S03     I07 -0.20810137       0 -1.3964965  1.56028336 -0.371888193
#> 76     S03     I08  0.73578821       0 -1.3964965  0.02755786  2.104726882
#> 89     S03     I09 -2.97574802       0 -1.3964965 -0.81124192 -0.768009570
#> 95     S03     I10 -4.81850701       0 -1.3964965 -2.61271975 -0.809290727
#> 5      S04     I01 -1.13233651       0 -1.3795460  0.46851960 -0.221310093
#> 19     S04     I02 -3.15219574       0 -1.3795460  0.06850421 -1.841153934
#> 21     S04     I03  0.09120813       0 -1.3795460  1.91698874 -0.446234594
#> 34     S04     I04 -2.46448242       0 -1.3795460 -1.48932282  0.404386409
#> 48     S04     I05 -1.71762735       0 -1.3795460 -0.03200038 -0.306080953
#> 51     S04     I06 -0.97143091       0 -1.3795460  1.54188817 -1.133773068
#> 63     S04     I07  0.59806274       0 -1.3795460  1.56028336  0.417325393
#> 77     S04     I08 -0.76394876       0 -1.3795460  0.02755786  0.588039393
#> 90     S04     I09 -1.67370722       0 -1.3795460 -0.81124192  0.517080715
#> 92     S04     I10 -2.62791388       0 -1.3795460 -2.61271975  1.364351882
#> 2      S05     I01  0.88849485       0  1.0028141  0.46851960 -0.582838804
#> 16     S05     I02  1.60457368       0  1.0028141  0.06850421  0.533255423
#> 22     S05     I03  2.69361727       0  1.0028141  1.91698874 -0.226185523
#> 35     S05     I04  0.08706578       0  1.0028141 -1.48932282  0.573574546
#> 45     S05     I05  2.85357882       0  1.0028141 -0.03200038  1.882765148
#> 52     S05     I06  1.56374253       0  1.0028141  1.54188817 -0.980959696
#> 64     S05     I07  1.55818393       0  1.0028141  1.56028336 -1.004913481
#> 74     S05     I08  1.81495618       0  1.0028141  0.02755786  0.784584269
#> 81     S05     I09  1.05231532       0  1.0028141 -0.81124192  0.860743189
#> 93     S05     I10 -1.87718090       0  1.0028141 -2.61271975 -0.267275195
#> 8      S06     I01 -0.17432369       0  0.1420402  0.46851960 -0.784883487
#> 20     S06     I02 -1.07736136       0  0.1420402  0.06850421 -1.287905763
#> 27     S06     I03  3.06377329       0  0.1420402  1.91698874  1.004744360
#> 37     S06     I04 -2.27150151       0  0.1420402 -1.48932282 -0.924218888
#> 49     S06     I05  0.37479678       0  0.1420402 -0.03200038  0.264756966
#> 56     S06     I06  1.47681462       0  0.1420402  1.54188817 -0.207113746
#> 66     S06     I07  1.61624467       0  0.1420402  1.56028336 -0.086078874
#> 79     S06     I08 -0.73896475       0  0.1420402  0.02755786 -0.908562800
#> 85     S06     I09  0.51938339       0  0.1420402 -0.81124192  1.188585115
#> 99     S06     I10 -1.39316652       0  0.1420402 -2.61271975  1.077513038
#> 9      S07     I01  1.48808890       0  0.7611355  0.46851960  0.258433818
#> 11     S07     I02  0.30909695       0  0.7611355  0.06850421 -0.520542735
#> 24     S07     I03  3.29715801       0  0.7611355  1.91698874  0.619033787
#> 38     S07     I04 -0.69046122       0  0.7611355 -1.48932282  0.037726117
#> 50     S07     I05  2.38818099       0  0.7611355 -0.03200038  1.659045889
#> 53     S07     I06 -0.06373369       0  0.7611355  1.54188817 -2.366757346
#> 67     S07     I07  1.14697830       0  0.7611355  1.56028336 -1.174440535
#> 80     S07     I08  1.33159212       0  0.7611355  0.02755786  0.542898788
#> 82     S07     I09  1.35733860       0  0.7611355 -0.81124192  1.407445044
#> 96     S07     I10 -0.19416939       0  0.7611355 -2.61271975  1.657414883
#> 6      S08     I01  1.64105785       0  0.1683146  0.46851960  1.004223683
#> 12     S08     I02  0.92409030       0  0.1683146  0.06850421  0.687271524
#> 25     S08     I03  1.50909887       0  0.1683146  1.91698874 -0.576204444
#> 39     S08     I04 -0.65658498       0  0.1683146 -1.48932282  0.664423273
#> 41     S08     I05  2.46029227       0  0.1683146 -0.03200038  2.323978085
#> 54     S08     I06  1.96088978       0  0.1683146  1.54188817  0.250687040
#> 68     S08     I07  4.15327828       0  0.1683146  1.56028336  2.424680357
#> 78     S08     I08  1.67931314       0  0.1683146  0.02755786  1.483440714
#> 83     S08     I09 -0.39730638       0  0.1683146 -0.81124192  0.245620971
#> 97     S08     I10 -2.06691129       0  0.1683146 -2.61271975  0.377493897
#> 3      S09     I01 -1.13511631       0 -0.4418931  0.46851960 -1.161742793
#> 13     S09     I02 -0.24652543       0 -0.4418931  0.06850421  0.126863476
#> 26     S09     I03 -0.64799278       0 -0.4418931  1.91698874 -2.123088408
#> 32     S09     I04 -0.58450200       0 -0.4418931 -1.48932282  1.346713936
#> 42     S09     I05 -0.71723503       0 -0.4418931 -0.03200038 -0.243341533
#> 55     S09     I06 -0.49600304       0 -0.4418931  1.54188817 -1.595998102
#> 65     S09     I07  1.55566122       0 -0.4418931  1.56028336  0.437270980
#> 71     S09     I08  0.58834273       0 -0.4418931  0.02755786  1.002677989
#> 84     S09     I09 -0.63556222       0 -0.4418931 -0.81124192  0.617572814
#> 94     S09     I10 -2.88804782       0 -0.4418931 -2.61271975  0.166565047
#> 10     S10     I01  2.28769842       0  1.0802580  0.46851960  0.738920855
#> 14     S10     I02  3.32685149       0  1.0802580  0.06850421  2.178089320
#> 28     S10     I03  4.99648653       0  1.0802580  1.91698874  1.999239828
#> 40     S10     I04 -0.51783955       0  1.0802580 -1.48932282 -0.108774687
#> 43     S10     I05  0.76087554       0  1.0802580 -0.03200038 -0.287382039
#> 57     S10     I06  1.64241397       0  1.0802580  1.54188817 -0.979732163
#> 70     S10     I07  2.45435233       0  1.0802580  1.56028336 -0.186188991
#> 72     S10     I08  2.30951178       0  1.0802580  0.02755786  1.201695960
#> 86     S10     I09  1.77560256       0  1.0802580 -0.81124192  1.506586523
#> 100    S10     I10 -2.16893363       0  1.0802580 -2.61271975 -0.636471842