Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, empirical = FALSE, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id            y grand_i        sub_i      item_i         err
#> 1      S01     I01  2.323282198       0  0.425291422  1.00307668  0.89491410
#> 2      S02     I01 -1.075607585       0 -1.619147259  1.00307668 -0.45953700
#> 3      S03     I01 -0.740253822       0 -0.450717347  1.00307668 -1.29261315
#> 4      S04     I01 -0.827831874       0 -0.507179499  1.00307668 -1.32372905
#> 5      S05     I01  0.835179527       0  0.716625163  1.00307668 -0.88452231
#> 6      S06     I01  0.001280315       0 -1.151070325  1.00307668  0.14927396
#> 7      S07     I01  3.163814877       0  0.608975051  1.00307668  1.55176315
#> 8      S08     I01  1.450597365       0 -0.060413162  1.00307668  0.50793385
#> 9      S09     I01  1.028860593       0 -0.002024643  1.00307668  0.02780856
#> 10     S10     I01  2.597403390       0  0.674414634  1.00307668  0.91991208
#> 11     S01     I02  1.867493407       0  0.425291422  0.15040044  1.29180154
#> 12     S02     I02 -2.485230331       0 -1.619147259  0.15040044 -1.01648352
#> 13     S03     I02 -0.393341873       0 -0.450717347  0.15040044 -0.09302497
#> 14     S04     I02 -0.870988211       0 -0.507179499  0.15040044 -0.51420916
#> 15     S05     I02  0.592723071       0  0.716625163  0.15040044 -0.27430254
#> 16     S06     I02 -1.272168640       0 -1.151070325  0.15040044 -0.27149876
#> 17     S07     I02 -0.199255359       0  0.608975051  0.15040044 -0.95863085
#> 18     S08     I02  0.174827441       0 -0.060413162  0.15040044  0.08484016
#> 19     S09     I02  0.985436592       0 -0.002024643  0.15040044  0.83706079
#> 20     S10     I02  1.107774358       0  0.674414634  0.15040044  0.28295928
#> 21     S01     I03  1.174415480       0  0.425291422 -0.17834509  0.92746915
#> 22     S02     I03 -2.023331036       0 -1.619147259 -0.17834509 -0.22583869
#> 23     S03     I03 -0.879531337       0 -0.450717347 -0.17834509 -0.25046890
#> 24     S04     I03  0.479029021       0 -0.507179499 -0.17834509  1.16455361
#> 25     S05     I03  2.159005566       0  0.716625163 -0.17834509  1.62072549
#> 26     S06     I03 -1.652083626       0 -1.151070325 -0.17834509 -0.32266821
#> 27     S07     I03 -0.921268112       0  0.608975051 -0.17834509 -1.35189807
#> 28     S08     I03  1.371223214       0 -0.060413162 -0.17834509  1.60998147
#> 29     S09     I03 -0.122181343       0 -0.002024643 -0.17834509  0.05818839
#> 30     S10     I03  1.358142415       0  0.674414634 -0.17834509  0.86207287
#> 31     S01     I04 -1.172760668       0  0.425291422 -1.35978161 -0.23827048
#> 32     S02     I04 -3.256989446       0 -1.619147259 -1.35978161 -0.27806058
#> 33     S03     I04 -1.933534391       0 -0.450717347 -1.35978161 -0.12303544
#> 34     S04     I04 -2.253373963       0 -0.507179499 -1.35978161 -0.38641286
#> 35     S05     I04 -0.343748187       0  0.716625163 -1.35978161  0.29940826
#> 36     S06     I04 -2.185454657       0 -1.151070325 -1.35978161  0.32539728
#> 37     S07     I04  0.916001368       0  0.608975051 -1.35978161  1.66680793
#> 38     S08     I04 -0.111049822       0 -0.060413162 -1.35978161  1.30914495
#> 39     S09     I04  1.651513643       0 -0.002024643 -1.35978161  3.01331989
#> 40     S10     I04 -1.592731157       0  0.674414634 -1.35978161 -0.90736418
#> 41     S01     I05 -0.099601139       0  0.425291422 -0.06569968 -0.45919288
#> 42     S02     I05 -2.841561099       0 -1.619147259 -0.06569968 -1.15671416
#> 43     S03     I05 -2.981254456       0 -0.450717347 -0.06569968 -2.46483742
#> 44     S04     I05  0.075709679       0 -0.507179499 -0.06569968  0.64858886
#> 45     S05     I05 -0.018889027       0  0.716625163 -0.06569968 -0.66981450
#> 46     S06     I05 -1.481627886       0 -1.151070325 -0.06569968 -0.26485788
#> 47     S07     I05  1.395449458       0  0.608975051 -0.06569968  0.85217409
#> 48     S08     I05  0.511708811       0 -0.060413162 -0.06569968  0.63782166
#> 49     S09     I05  0.915163393       0 -0.002024643 -0.06569968  0.98288772
#> 50     S10     I05  0.500507018       0  0.674414634 -0.06569968 -0.10820793
#> 51     S01     I06  1.514580837       0  0.425291422  1.01649307  0.07279634
#> 52     S02     I06 -0.489198436       0 -1.619147259  1.01649307  0.11345575
#> 53     S03     I06  1.096639819       0 -0.450717347  1.01649307  0.53086409
#> 54     S04     I06  0.810034900       0 -0.507179499  1.01649307  0.30072133
#> 55     S05     I06  1.786021586       0  0.716625163  1.01649307  0.05290335
#> 56     S06     I06 -2.136059534       0 -1.151070325  1.01649307 -2.00148228
#> 57     S07     I06  1.497658519       0  0.608975051  1.01649307 -0.12780960
#> 58     S08     I06  0.070024477       0 -0.060413162  1.01649307 -0.88605543
#> 59     S09     I06 -0.925963778       0 -0.002024643  1.01649307 -1.94043221
#> 60     S10     I06  2.001306431       0  0.674414634  1.01649307  0.31039872
#> 61     S01     I07 -1.669445526       0  0.425291422 -2.11710537  0.02236842
#> 62     S02     I07 -4.017459912       0 -1.619147259 -2.11710537 -0.28120729
#> 63     S03     I07 -1.727770279       0 -0.450717347 -2.11710537  0.84005243
#> 64     S04     I07 -3.066475727       0 -0.507179499 -2.11710537 -0.44219086
#> 65     S05     I07  0.040353558       0  0.716625163 -2.11710537  1.44083376
#> 66     S06     I07 -4.519653716       0 -1.151070325 -2.11710537 -1.25147802
#> 67     S07     I07 -2.465564826       0  0.608975051 -2.11710537 -0.95743451
#> 68     S08     I07 -3.176783802       0 -0.060413162 -2.11710537 -0.99926527
#> 69     S09     I07 -0.828472817       0 -0.002024643 -2.11710537  1.29065719
#> 70     S10     I07 -0.666375049       0  0.674414634 -2.11710537  0.77631568
#> 71     S01     I08 -0.915770606       0  0.425291422 -1.57614669  0.23508466
#> 72     S02     I08 -3.544527512       0 -1.619147259 -1.57614669 -0.34923356
#> 73     S03     I08 -1.859786132       0 -0.450717347 -1.57614669  0.16707791
#> 74     S04     I08 -1.590888342       0 -0.507179499 -1.57614669  0.49243785
#> 75     S05     I08  0.891688939       0  0.716625163 -1.57614669  1.75121047
#> 76     S06     I08 -2.668699259       0 -1.151070325 -1.57614669  0.05851776
#> 77     S07     I08 -0.792319260       0  0.608975051 -1.57614669  0.17485238
#> 78     S08     I08 -1.323125770       0 -0.060413162 -1.57614669  0.31343408
#> 79     S09     I08 -0.684456601       0 -0.002024643 -1.57614669  0.89371473
#> 80     S10     I08 -1.186905866       0  0.674414634 -1.57614669 -0.28517381
#> 81     S01     I09  1.225483238       0  0.425291422  1.07672924 -0.27653743
#> 82     S02     I09  0.230014594       0 -1.619147259  1.07672924  0.77243261
#> 83     S03     I09  0.258590126       0 -0.450717347  1.07672924 -0.36742177
#> 84     S04     I09 -0.795290032       0 -0.507179499  1.07672924 -1.36483978
#> 85     S05     I09  0.429066362       0  0.716625163  1.07672924 -1.36428804
#> 86     S06     I09  0.511623340       0 -1.151070325  1.07672924  0.58596442
#> 87     S07     I09  1.471985452       0  0.608975051  1.07672924 -0.21371884
#> 88     S08     I09  0.377370800       0 -0.060413162  1.07672924 -0.63894528
#> 89     S09     I09 -0.981038879       0 -0.002024643  1.07672924 -2.05574348
#> 90     S10     I09  0.575566751       0  0.674414634  1.07672924 -1.17557713
#> 91     S01     I10  2.312971379       0  0.425291422  1.16992451  0.71775545
#> 92     S02     I10  0.282847476       0 -1.619147259  1.16992451  0.73207023
#> 93     S03     I10  0.855728583       0 -0.450717347  1.16992451  0.13652142
#> 94     S04     I10  0.478154133       0 -0.507179499  1.16992451 -0.18459087
#> 95     S05     I10  2.245950205       0  0.716625163  1.16992451  0.35940054
#> 96     S06     I10  1.069791870       0 -1.151070325  1.16992451  1.05093769
#> 97     S07     I10  0.904354587       0  0.608975051  1.16992451 -0.87454497
#> 98     S08     I10 -0.563118375       0 -0.060413162  1.16992451 -1.67262972
#> 99     S09     I10  0.499149656       0 -0.002024643  1.16992451 -0.66875021
#> 100    S10     I10  0.852085704       0  0.674414634  1.16992451 -0.99225344