Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i        sub_i      item_i          err
#> 1      S01     I01  2.15453148       0  1.246547000  0.41655671  0.491427768
#> 15     S01     I02  1.07322432       0  1.246547000  0.21130684 -0.384629521
#> 29     S01     I03  3.23831212       0  1.246547000  0.01882627  1.972938844
#> 31     S01     I04  1.95461973       0  1.246547000  0.89066536 -0.182592624
#> 44     S01     I05 -2.46693810       0  1.246547000 -1.30699432 -2.406490772
#> 58     S01     I06  0.85751622       0  1.246547000  0.37753180 -0.766562581
#> 61     S01     I07  0.31504883       0  1.246547000  0.11118400 -1.042682165
#> 73     S01     I08  1.09538996       0  1.246547000  0.19025509 -0.341412133
#> 87     S01     I09  2.53226470       0  1.246547000  1.01432471  0.271392995
#> 91     S01     I10  2.36406340       0  1.246547000  0.40981881  0.707697596
#> 7      S02     I01  1.08932140       0  0.006142866  0.41655671  0.666621818
#> 17     S02     I02 -0.45884065       0  0.006142866  0.21130684 -0.676290351
#> 30     S02     I03  0.27891506       0  0.006142866  0.01882627  0.253945923
#> 36     S02     I04  1.07360525       0  0.006142866  0.89066536  0.176797026
#> 46     S02     I05  0.51059464       0  0.006142866 -1.30699432  1.811446100
#> 59     S02     I06 -0.47330043       0  0.006142866  0.37753180 -0.856975100
#> 69     S02     I07 -0.20691770       0  0.006142866  0.11118400 -0.324244567
#> 75     S02     I08 -0.00837479       0  0.006142866  0.19025509 -0.204772748
#> 88     S02     I09 -0.96371059       0  0.006142866  1.01432471 -1.984178161
#> 98     S02     I10  1.15223106       0  0.006142866  0.40981881  0.736269389
#> 4      S03     I01  1.53814237       0 -0.059806157  0.41655671  1.181391813
#> 18     S03     I02  0.70881036       0 -0.059806157  0.21130684  0.557309683
#> 23     S03     I03 -1.52082896       0 -0.059806157  0.01882627 -1.479849077
#> 33     S03     I04  1.63408603       0 -0.059806157  0.89066536  0.803226825
#> 47     S03     I05 -1.53606032       0 -0.059806157 -1.30699432 -0.169259840
#> 60     S03     I06  0.61926942       0 -0.059806157  0.37753180  0.301543775
#> 62     S03     I07 -0.40445477       0 -0.059806157  0.11118400 -0.455832607
#> 76     S03     I08 -1.46482837       0 -0.059806157  0.19025509 -1.595277311
#> 89     S03     I09 -1.46029880       0 -0.059806157  1.01432471 -2.414817352
#> 95     S03     I10 -1.59991020       0 -0.059806157  0.40981881 -1.949922857
#> 5      S04     I01  2.02685787       0  1.501903884  0.41655671  0.108397276
#> 19     S04     I02  2.36020327       0  1.501903884  0.21130684  0.646992551
#> 21     S04     I03 -0.15819334       0  1.501903884  0.01882627 -1.678923499
#> 34     S04     I04  2.92709707       0  1.501903884  0.89066536  0.534527830
#> 48     S04     I05 -0.41831350       0  1.501903884 -1.30699432 -0.613223062
#> 51     S04     I06  2.03549561       0  1.501903884  0.37753180  0.156059929
#> 63     S04     I07  1.73731144       0  1.501903884  0.11118400  0.124223563
#> 77     S04     I08  1.74745716       0  1.501903884  0.19025509  0.055298184
#> 90     S04     I09  3.73978216       0  1.501903884  1.01432471  1.223553567
#> 92     S04     I10  1.81125770       0  1.501903884  0.40981881 -0.100464996
#> 2      S05     I01  2.79497822       0  0.830053002  0.41655671  1.548368507
#> 16     S05     I02  1.87615760       0  0.830053002  0.21130684  0.834797766
#> 22     S05     I03 -0.08348189       0  0.830053002  0.01882627 -0.932361163
#> 35     S05     I04  1.97015294       0  0.830053002  0.89066536  0.249434583
#> 45     S05     I05 -0.95058255       0  0.830053002 -1.30699432 -0.473641226
#> 52     S05     I06  0.45868587       0  0.830053002  0.37753180 -0.748898930
#> 64     S05     I07  1.50589228       0  0.830053002  0.11118400  0.564655280
#> 74     S05     I08  1.93970228       0  0.830053002  0.19025509  0.919394185
#> 81     S05     I09  1.83579435       0  0.830053002  1.01432471 -0.008583359
#> 93     S05     I10  2.75881341       0  0.830053002  0.40981881  1.518941602
#> 8      S06     I01 -0.81561528       0 -0.527490735  0.41655671 -0.704681259
#> 20     S06     I02 -0.29278369       0 -0.527490735  0.21130684  0.023400212
#> 27     S06     I03  0.64319586       0 -0.527490735  0.01882627  1.151860320
#> 37     S06     I04 -0.25918088       0 -0.527490735  0.89066536 -0.622355502
#> 49     S06     I05 -1.87540689       0 -0.527490735 -1.30699432 -0.040921834
#> 56     S06     I06 -0.46611520       0 -0.527490735  0.37753180 -0.316156270
#> 66     S06     I07 -1.61744325       0 -0.527490735  0.11118400 -1.201136514
#> 79     S06     I08 -2.22329083       0 -0.527490735  0.19025509 -1.886055183
#> 85     S06     I09 -0.60318303       0 -0.527490735  1.01432471 -1.090017001
#> 99     S06     I10 -0.20985615       0 -0.527490735  0.40981881 -0.092184224
#> 9      S07     I01  2.76753943       0  1.312766281  0.41655671  1.038216432
#> 11     S07     I02  2.18561454       0  1.312766281  0.21130684  0.661541419
#> 24     S07     I03  2.21481879       0  1.312766281  0.01882627  0.883226238
#> 38     S07     I04  2.02971841       0  1.312766281  0.89066536 -0.173713227
#> 50     S07     I05  0.29941713       0  1.312766281 -1.30699432  0.293645170
#> 53     S07     I06  1.87500260       0  1.312766281  0.37753180  0.184704522
#> 67     S07     I07  1.13178839       0  1.312766281  0.11118400 -0.292161891
#> 80     S07     I08  0.76601443       0  1.312766281  0.19025509 -0.737006942
#> 82     S07     I09  2.09115849       0  1.312766281  1.01432471 -0.235932500
#> 96     S07     I10  2.51646914       0  1.312766281  0.40981881  0.793884053
#> 6      S08     I01  0.27963552       0 -0.179192600  0.41655671  0.042271403
#> 12     S08     I02  1.91242176       0 -0.179192600  0.21130684  1.880307518
#> 25     S08     I03  0.83283757       0 -0.179192600  0.01882627  0.993203894
#> 39     S08     I04  0.21817685       0 -0.179192600  0.89066536 -0.493295910
#> 41     S08     I05 -2.77356615       0 -0.179192600 -1.30699432 -1.287379227
#> 54     S08     I06 -0.46885016       0 -0.179192600  0.37753180 -0.667189357
#> 68     S08     I07 -0.25864666       0 -0.179192600  0.11118400 -0.190638062
#> 78     S08     I08 -0.84402809       0 -0.179192600  0.19025509 -0.855090584
#> 83     S08     I09 -1.23544493       0 -0.179192600  1.01432471 -2.070577042
#> 97     S08     I10 -0.24635884       0 -0.179192600  0.40981881 -0.476985050
#> 3      S09     I01  0.74229382       0  0.088416864  0.41655671  0.237320245
#> 13     S09     I02 -0.64623057       0  0.088416864  0.21130684 -0.945954269
#> 26     S09     I03  0.37836220       0  0.088416864  0.01882627  0.271119065
#> 32     S09     I04  1.13867848       0  0.088416864  0.89066536  0.159596260
#> 42     S09     I05 -0.31914551       0  0.088416864 -1.30699432  0.899431949
#> 55     S09     I06  0.76511857       0  0.088416864  0.37753180  0.299169903
#> 65     S09     I07 -0.23768896       0  0.088416864  0.11118400 -0.437289821
#> 71     S09     I08 -1.44798281       0  0.088416864  0.19025509 -1.726654769
#> 84     S09     I09  0.62135831       0  0.088416864  1.01432471 -0.481383257
#> 94     S09     I10  0.42321053       0  0.088416864  0.40981881 -0.075025139
#> 10     S10     I01 -0.02904971       0 -0.893098054  0.41655671  0.447491634
#> 14     S10     I02  0.51214476       0 -0.893098054  0.21130684  1.193935978
#> 28     S10     I03 -1.03776086       0 -0.893098054  0.01882627 -0.163489074
#> 40     S10     I04 -0.06868898       0 -0.893098054  0.89066536 -0.066256281
#> 43     S10     I05 -2.40544160       0 -0.893098054 -1.30699432 -0.205349219
#> 57     S10     I06  1.24708321       0 -0.893098054  0.37753180  1.762649460
#> 70     S10     I07 -0.61011833       0 -0.893098054  0.11118400  0.171795728
#> 72     S10     I08 -0.57904616       0 -0.893098054  0.19025509  0.123796797
#> 86     S10     I09 -0.73708982       0 -0.893098054  1.01432471 -0.858316477
#> 100    S10     I10 -0.59830934       0 -0.893098054  0.40981881 -0.115030094