Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#> # A tibble: 100 x 7
#>    sub_id item_id     y grand_i sub_i  item_i     err
#>    <chr>  <chr>   <dbl>   <dbl> <dbl>   <dbl>   <dbl>
#>  1 S01    I01     2.00        0  1.30  1.41   -0.709 
#>  2 S01    I02     1.17        0  1.30 -0.924   0.791 
#>  3 S01    I03     2.22        0  1.30  0.986  -0.0654
#>  4 S01    I04     0.776       0  1.30 -0.837   0.309 
#>  5 S01    I05     2.22        0  1.30  1.19   -0.271 
#>  6 S01    I06     0.222       0  1.30 -1.06   -0.0233
#>  7 S01    I07     0.468       0  1.30 -0.0653 -0.770 
#>  8 S01    I08     1.29        0  1.30 -1.23    1.22  
#>  9 S01    I09     0.675       0  1.30 -1.04    0.409 
#> 10 S01    I10     2.48        0  1.30  1.52   -0.343 
#> # … with 90 more rows