Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i        sub_i      item_i          err
#> 1      S01     I01  0.89827035       0  1.048229287 -0.52749074  0.377531800
#> 15     S01     I02  2.47217957       0  1.048229287  1.31276628  0.111183999
#> 29     S01     I03  1.05929178       0  1.048229287 -0.17919260  0.190255093
#> 31     S01     I04  2.15097086       0  1.048229287  0.08841686  1.014324708
#> 44     S01     I05  0.56495004       0  1.048229287 -0.89309805  0.409818809
#> 58     S01     I06  1.95621377       0  1.048229287  0.41655671  0.491427768
#> 61     S01     I07  0.87490660       0  1.048229287  0.21130684 -0.384629521
#> 73     S01     I08  3.03999440       0  1.048229287  0.01882627  1.972938844
#> 87     S01     I09  1.75630202       0  1.048229287  0.89066536 -0.182592624
#> 91     S01     I10 -2.66525581       0  1.048229287 -1.30699432 -2.406490772
#> 7      S02     I01 -0.13923821       0  1.154815105 -0.52749074 -0.766562581
#> 17     S02     I02  1.42489922       0  1.154815105  1.31276628 -1.042682165
#> 30     S02     I03  0.63421037       0  1.154815105 -0.17919260 -0.341412133
#> 36     S02     I04  1.51462496       0  1.154815105  0.08841686  0.271392995
#> 46     S02     I05  0.96941465       0  1.154815105 -0.89309805  0.707697596
#> 59     S02     I06  2.23799364       0  1.154815105  0.41655671  0.666621818
#> 69     S02     I07  0.68983159       0  1.154815105  0.21130684 -0.676290351
#> 75     S02     I08  1.42758730       0  1.154815105  0.01882627  0.253945923
#> 88     S02     I09  2.22227749       0  1.154815105  0.89066536  0.176797026
#> 98     S02     I10  1.65926688       0  1.154815105 -1.30699432  1.811446100
#> 4      S03     I01 -1.41759580       0 -0.033129970 -0.52749074 -0.856975100
#> 18     S03     I02  0.95539174       0 -0.033129970  1.31276628 -0.324244567
#> 23     S03     I03 -0.41709532       0 -0.033129970 -0.17919260 -0.204772748
#> 33     S03     I04 -1.92889127       0 -0.033129970  0.08841686 -1.984178161
#> 47     S03     I05 -0.18995863       0 -0.033129970 -0.89309805  0.736269389
#> 60     S03     I06  1.56481856       0 -0.033129970  0.41655671  1.181391813
#> 62     S03     I07  0.73548655       0 -0.033129970  0.21130684  0.557309683
#> 76     S03     I08 -1.49415277       0 -0.033129970  0.01882627 -1.479849077
#> 89     S03     I09  1.66076221       0 -0.033129970  0.89066536  0.803226825
#> 95     S03     I10 -1.50938413       0 -0.033129970 -1.30699432 -0.169259840
#> 5      S04     I01 -0.14256186       0  0.083385100 -0.52749074  0.301543775
#> 19     S04     I02  0.94031877       0  0.083385100  1.31276628 -0.455832607
#> 21     S04     I03 -1.69108481       0  0.083385100 -0.17919260 -1.595277311
#> 34     S04     I04 -2.24301539       0  0.083385100  0.08841686 -2.414817352
#> 48     S04     I05 -2.75963581       0  0.083385100 -0.89309805 -1.949922857
#> 51     S04     I06  0.60833909       0  0.083385100  0.41655671  0.108397276
#> 63     S04     I07  0.94168449       0  0.083385100  0.21130684  0.646992551
#> 77     S04     I08 -1.57671213       0  0.083385100  0.01882627 -1.678923499
#> 90     S04     I09  1.50857829       0  0.083385100  0.89066536  0.534527830
#> 92     S04     I10 -1.83683229       0  0.083385100 -1.30699432 -0.613223062
#> 2      S05     I01 -2.64653523       0 -2.275104427 -0.52749074  0.156059929
#> 16     S05     I02 -0.83811458       0 -2.275104427  1.31276628  0.124223563
#> 22     S05     I03 -2.39899884       0 -2.275104427 -0.17919260  0.055298184
#> 35     S05     I04 -0.96313400       0 -2.275104427  0.08841686  1.223553567
#> 45     S05     I05 -3.26866748       0 -2.275104427 -0.89309805 -0.100464996
#> 52     S05     I06 -0.31017921       0 -2.275104427  0.41655671  1.548368507
#> 64     S05     I07 -1.22899982       0 -2.275104427  0.21130684  0.834797766
#> 74     S05     I08 -3.18863932       0 -2.275104427  0.01882627 -0.932361163
#> 81     S05     I09 -1.13500449       0 -2.275104427  0.89066536  0.249434583
#> 93     S05     I10 -4.05573998       0 -2.275104427 -1.30699432 -0.473641226
#> 8      S06     I01 -0.02984267       0  1.246547000 -0.52749074 -0.748898930
#> 20     S06     I02  3.12396856       0  1.246547000  1.31276628  0.564655280
#> 27     S06     I03  1.98674858       0  1.246547000 -0.17919260  0.919394185
#> 37     S06     I04  1.32638050       0  1.246547000  0.08841686 -0.008583359
#> 49     S06     I05  1.87239055       0  1.246547000 -0.89309805  1.518941602
#> 56     S06     I06  0.95842246       0  1.246547000  0.41655671 -0.704681259
#> 66     S06     I07  1.48125405       0  1.246547000  0.21130684  0.023400212
#> 79     S06     I08  2.41723359       0  1.246547000  0.01882627  1.151860320
#> 85     S06     I09  1.51485686       0  1.246547000  0.89066536 -0.622355502
#> 99     S06     I10 -0.10136916       0  1.246547000 -1.30699432 -0.040921834
#> 9      S07     I01 -0.83750414       0  0.006142866 -0.52749074 -0.316156270
#> 11     S07     I02  0.11777263       0  0.006142866  1.31276628 -1.201136514
#> 24     S07     I03 -2.05910492       0  0.006142866 -0.17919260 -1.886055183
#> 38     S07     I04 -0.99545727       0  0.006142866  0.08841686 -1.090017001
#> 50     S07     I05 -0.97913941       0  0.006142866 -0.89309805 -0.092184224
#> 53     S07     I06  1.46091601       0  0.006142866  0.41655671  1.038216432
#> 67     S07     I07  0.87899112       0  0.006142866  0.21130684  0.661541419
#> 80     S07     I08  0.90819538       0  0.006142866  0.01882627  0.883226238
#> 82     S07     I09  0.72309500       0  0.006142866  0.89066536 -0.173713227
#> 96     S07     I10 -1.00720629       0  0.006142866 -1.30699432  0.293645170
#> 6      S08     I01 -0.40259237       0 -0.059806157 -0.52749074  0.184704522
#> 12     S08     I02  0.96079823       0 -0.059806157  1.31276628 -0.292161891
#> 25     S08     I03 -0.97600570       0 -0.059806157 -0.17919260 -0.737006942
#> 39     S08     I04 -0.20732179       0 -0.059806157  0.08841686 -0.235932500
#> 41     S08     I05 -0.15902016       0 -0.059806157 -0.89309805  0.793884053
#> 54     S08     I06  0.39902196       0 -0.059806157  0.41655671  0.042271403
#> 68     S08     I07  2.03180820       0 -0.059806157  0.21130684  1.880307518
#> 78     S08     I08  0.95222401       0 -0.059806157  0.01882627  0.993203894
#> 83     S08     I09  0.33756329       0 -0.059806157  0.89066536 -0.493295910
#> 97     S08     I10 -2.65417971       0 -0.059806157 -1.30699432 -1.287379227
#> 3      S09     I01  0.30722379       0  1.501903884 -0.52749074 -0.667189357
#> 13     S09     I02  2.62403210       0  1.501903884  1.31276628 -0.190638062
#> 26     S09     I03  0.46762070       0  1.501903884 -0.17919260 -0.855090584
#> 32     S09     I04 -0.48025629       0  1.501903884  0.08841686 -2.070577042
#> 42     S09     I05  0.13182078       0  1.501903884 -0.89309805 -0.476985050
#> 55     S09     I06  2.15578084       0  1.501903884  0.41655671  0.237320245
#> 65     S09     I07  0.76725645       0  1.501903884  0.21130684 -0.945954269
#> 71     S09     I08  1.79184922       0  1.501903884  0.01882627  0.271119065
#> 84     S09     I09  2.55216550       0  1.501903884  0.89066536  0.159596260
#> 94     S09     I10  1.09434151       0  1.501903884 -1.30699432  0.899431949
#> 10     S10     I01  0.60173217       0  0.830053002 -0.52749074  0.299169903
#> 14     S10     I02  1.70552946       0  0.830053002  1.31276628 -0.437289821
#> 28     S10     I03 -1.07579437       0  0.830053002 -0.17919260 -1.726654769
#> 40     S10     I04  0.43708661       0  0.830053002  0.08841686 -0.481383257
#> 43     S10     I05 -0.13807019       0  0.830053002 -0.89309805 -0.075025139
#> 57     S10     I06  1.69410135       0  0.830053002  0.41655671  0.447491634
#> 70     S10     I07  2.23529582       0  0.830053002  0.21130684  1.193935978
#> 72     S10     I08  0.68539020       0  0.830053002  0.01882627 -0.163489074
#> 86     S10     I09  1.65446208       0  0.830053002  0.89066536 -0.066256281
#> 100    S10     I10 -0.68229054       0  0.830053002 -1.30699432 -0.205349219