Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i        sub_i      item_i         err
#> 1      S01     I01  2.77783583       0  1.075833068  1.54408120  0.15792156
#> 2      S02     I01  1.93554718       0  0.184800537  1.54408120  0.20666544
#> 3      S03     I01  2.95907303       0 -0.003712397  1.54408120  1.41870423
#> 4      S04     I01  1.52624278       0 -0.815588737  1.54408120  0.79775031
#> 5      S05     I01  2.25490971       0  0.363950890  1.54408120  0.34687762
#> 6      S06     I01  2.34753509       0  0.580868397  1.54408120  0.22258549
#> 7      S07     I01 -0.15458212       0 -0.261327459  1.54408120 -1.43733586
#> 8      S08     I01  0.88318135       0  0.174232550  1.54408120 -0.83513240
#> 9      S09     I01  1.69706920       0  0.568555899  1.54408120 -0.41556790
#> 10     S10     I01  0.49991413       0  0.564016859  1.54408120 -1.60818393
#> 11     S01     I02  2.12870027       0  1.075833068 -0.05226319  1.10513039
#> 12     S02     I02 -0.52619014       0  0.184800537 -0.05226319 -0.65872749
#> 13     S03     I02  0.68333242       0 -0.003712397 -0.05226319  0.73930800
#> 14     S04     I02 -0.66844077       0 -0.815588737 -0.05226319  0.19941116
#> 15     S05     I02  3.45170111       0  0.363950890 -0.05226319  3.14001340
#> 16     S06     I02 -0.75650627       0  0.580868397 -0.05226319 -1.28511148
#> 17     S07     I02 -0.70522902       0 -0.261327459 -0.05226319 -0.39163838
#> 18     S08     I02  2.32851468       0  0.174232550 -0.05226319  2.20654531
#> 19     S09     I02  0.80020741       0  0.568555899 -0.05226319  0.28391470
#> 20     S10     I02  1.25040298       0  0.564016859 -0.05226319  0.73864930
#> 21     S01     I03  0.97943972       0  1.075833068 -0.17213751  0.07574416
#> 22     S02     I03 -0.95285542       0  0.184800537 -0.17213751 -0.96551845
#> 23     S03     I03 -0.28129525       0 -0.003712397 -0.17213751 -0.10544534
#> 24     S04     I03 -1.04128484       0 -0.815588737 -0.17213751 -0.05355860
#> 25     S05     I03  0.12446706       0  0.363950890 -0.17213751 -0.06734632
#> 26     S06     I03  0.26411072       0  0.580868397 -0.17213751 -0.14462017
#> 27     S07     I03 -0.20318687       0 -0.261327459 -0.17213751  0.23027810
#> 28     S08     I03  0.05649938       0  0.174232550 -0.17213751  0.05440434
#> 29     S09     I03  1.79224742       0  0.568555899 -0.17213751  1.39582903
#> 30     S10     I03  0.35936540       0  0.564016859 -0.17213751 -0.03251395
#> 31     S01     I04  1.73862350       0  1.075833068  0.84271223 -0.17992180
#> 32     S02     I04  0.58991686       0  0.184800537  0.84271223 -0.43759591
#> 33     S03     I04  2.61252787       0 -0.003712397  0.84271223  1.77352803
#> 34     S04     I04 -0.19658805       0 -0.815588737  0.84271223 -0.22371154
#> 35     S05     I04  1.87546722       0  0.363950890  0.84271223  0.66880410
#> 36     S06     I04  1.64133809       0  0.580868397  0.84271223  0.21775746
#> 37     S07     I04  1.21875542       0 -0.261327459  0.84271223  0.63737065
#> 38     S08     I04 -0.94562715       0  0.174232550  0.84271223 -1.96257193
#> 39     S09     I04  0.17351686       0  0.568555899  0.84271223 -1.23775126
#> 40     S10     I04  0.77348629       0  0.564016859  0.84271223 -0.63324279
#> 41     S01     I05  1.13133054       0  1.075833068  1.05227295 -0.99677548
#> 42     S02     I05  2.40513815       0  0.184800537  1.05227295  1.16806467
#> 43     S03     I05  0.44170385       0 -0.003712397  1.05227295 -0.60685670
#> 44     S04     I05  0.48120832       0 -0.815588737  1.05227295  0.24452411
#> 45     S05     I05  1.55855580       0  0.363950890  1.05227295  0.14233197
#> 46     S06     I05 -0.42467326       0  0.580868397  1.05227295 -2.05781461
#> 47     S07     I05 -0.44377156       0 -0.261327459  1.05227295 -1.23471705
#> 48     S08     I05  1.00084455       0  0.174232550  1.05227295 -0.22566095
#> 49     S09     I05  1.04463106       0  0.568555899  1.05227295 -0.57619779
#> 50     S10     I05  2.80271735       0  0.564016859  1.05227295  1.18642754
#> 51     S01     I06  2.62788984       0  1.075833068  0.51845786  1.03359891
#> 52     S02     I06  1.77408315       0  0.184800537  0.51845786  1.07082475
#> 53     S03     I06 -0.16828987       0 -0.003712397  0.51845786 -0.68303533
#> 54     S04     I06  0.53720571       0 -0.815588737  0.51845786  0.83433658
#> 55     S05     I06  0.60996767       0  0.363950890  0.51845786 -0.27244109
#> 56     S06     I06  0.68643341       0  0.580868397  0.51845786 -0.41289285
#> 57     S07     I06  0.26919942       0 -0.261327459  0.51845786  0.01206901
#> 58     S08     I06  0.87443510       0  0.174232550  0.51845786  0.18174469
#> 59     S09     I06  0.18114051       0  0.568555899  0.51845786 -0.90587326
#> 60     S10     I06  0.65144588       0  0.564016859  0.51845786 -0.43102885
#> 61     S01     I07  2.94463420       0  1.075833068  1.40810731  0.46069382
#> 62     S02     I07  1.49176851       0  0.184800537  1.40810731 -0.10113934
#> 63     S03     I07  0.33250371       0 -0.003712397  1.40810731 -1.07189120
#> 64     S04     I07  0.04241288       0 -0.815588737  1.40810731 -0.55010569
#> 65     S05     I07  0.87480408       0  0.363950890  1.40810731 -0.89725412
#> 66     S06     I07  1.37848605       0  0.580868397  1.40810731 -0.61048966
#> 67     S07     I07  2.39035689       0 -0.261327459  1.40810731  1.24357704
#> 68     S08     I07  1.27620110       0  0.174232550  1.40810731 -0.30613876
#> 69     S09     I07  3.06343913       0  0.568555899  1.40810731  1.08677592
#> 70     S10     I07  2.07215170       0  0.564016859  1.40810731  0.10002753
#> 71     S01     I08  1.74128738       0  1.075833068  1.04951022 -0.38405591
#> 72     S02     I08  0.81472982       0  0.184800537  1.04951022 -0.41958094
#> 73     S03     I08  0.91750832       0 -0.003712397  1.04951022 -0.12828950
#> 74     S04     I08 -0.24723835       0 -0.815588737  1.04951022 -0.48115983
#> 75     S05     I08  1.55436710       0  0.363950890  1.04951022  0.14090599
#> 76     S06     I08  1.28269770       0  0.580868397  1.04951022 -0.34768092
#> 77     S07     I08  0.37626145       0 -0.261327459  1.04951022 -0.41192132
#> 78     S08     I08  0.12856640       0  0.174232550  1.04951022 -1.09517637
#> 79     S09     I08  1.35615526       0  0.568555899  1.04951022 -0.26191086
#> 80     S10     I08  1.71667790       0  0.564016859  1.04951022  0.10315082
#> 81     S01     I09  1.65840264       0  1.075833068  0.26520077  0.31736880
#> 82     S02     I09 -1.29371852       0  0.184800537  0.26520077 -1.74371983
#> 83     S03     I09  1.21343565       0 -0.003712397  0.26520077  0.95194728
#> 84     S04     I09  0.16620400       0 -0.815588737  0.26520077  0.71659196
#> 85     S05     I09  0.84938362       0  0.363950890  0.26520077  0.22023196
#> 86     S06     I09 -0.18944195       0  0.580868397  0.26520077 -1.03551113
#> 87     S07     I09  0.92954165       0 -0.261327459  0.26520077  0.92566834
#> 88     S08     I09  0.73390055       0  0.174232550  0.26520077  0.29446722
#> 89     S09     I09  0.80532671       0  0.568555899  0.26520077 -0.02842996
#> 90     S10     I09  1.11111122       0  0.564016859  0.26520077  0.28189359
#> 91     S01     I10  2.59009141       0  1.075833068  0.92826208  0.58599626
#> 92     S02     I10 -0.49631275       0  0.184800537  0.92826208 -1.60937536
#> 93     S03     I10  1.00278105       0 -0.003712397  0.92826208  0.07823137
#> 94     S04     I10 -0.55196964       0 -0.815588737  0.92826208 -0.66464298
#> 95     S05     I10  0.60110926       0  0.363950890  0.92826208 -0.69110371
#> 96     S06     I10  0.87507103       0  0.580868397  0.92826208 -0.63405945
#> 97     S07     I10  2.33235524       0 -0.261327459  0.92826208  1.66542062
#> 98     S08     I10  0.93781323       0  0.174232550  0.92826208 -0.16468140
#> 99     S09     I10  0.39202963       0  0.568555899  0.92826208 -1.10478835
#> 100    S10     I10 -0.35371935       0  0.564016859  0.92826208 -1.84599829