Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id          val grand_i       sub_i      item_i         err
#> 1      S01     I01 -1.070726917       0 -0.48509448  0.17188601 -0.75751844
#> 2      S02     I01  1.983591235       0  0.96103270  0.17188601  0.85067253
#> 3      S03     I01  2.658315520       0  2.62331596  0.17188601 -0.13688645
#> 4      S04     I01 -0.597116672       0 -0.15973365  0.17188601 -0.60926903
#> 5      S05     I01 -0.665630520       0 -0.32712036  0.17188601 -0.51039617
#> 6      S06     I01 -1.706815580       0 -1.40710108  0.17188601 -0.47160051
#> 7      S07     I01  0.403842300       0 -1.14135570  0.17188601  1.37331200
#> 8      S08     I01 -2.512571738       0  0.08896821  0.17188601 -2.77342596
#> 9      S09     I01  0.207844222       0  0.98910793  0.17188601 -0.95314972
#> 10     S10     I01  2.675761523       0  0.68879865  0.17188601  1.81507686
#> 11     S01     I02 -1.206833698       0 -0.48509448 -0.46730082 -0.25443840
#> 12     S02     I02  0.553678168       0  0.96103270 -0.46730082  0.05994629
#> 13     S03     I02  2.333808984       0  2.62331596 -0.46730082  0.17779384
#> 14     S04     I02 -0.967360281       0 -0.15973365 -0.46730082 -0.34032582
#> 15     S05     I02 -1.571765946       0 -0.32712036 -0.46730082 -0.77734477
#> 16     S06     I02 -1.265615354       0 -1.40710108 -0.46730082  0.60878654
#> 17     S07     I02 -0.002769923       0 -1.14135570 -0.46730082  1.60588660
#> 18     S08     I02  0.935090766       0  0.08896821 -0.46730082  1.31342337
#> 19     S09     I02  1.241126399       0  0.98910793 -0.46730082  0.71931928
#> 20     S10     I02  1.469803153       0  0.68879865 -0.46730082  1.24830532
#> 21     S01     I03  0.610814709       0 -0.48509448  0.05177736  1.04413184
#> 22     S02     I03  1.212405412       0  0.96103270  0.05177736  0.19959536
#> 23     S03     I03  3.539359845       0  2.62331596  0.05177736  0.86426653
#> 24     S04     I03  2.651922603       0 -0.15973365  0.05177736  2.75987890
#> 25     S05     I03  0.354455681       0 -0.32712036  0.05177736  0.62979868
#> 26     S06     I03 -0.406930444       0 -1.40710108  0.05177736  0.94839328
#> 27     S07     I03 -0.838849873       0 -1.14135570  0.05177736  0.25072847
#> 28     S08     I03  1.053263166       0  0.08896821  0.05177736  0.91251760
#> 29     S09     I03  2.349538312       0  0.98910793  0.05177736  1.30865302
#> 30     S10     I03 -1.800395656       0  0.68879865  0.05177736 -2.54097167
#> 31     S01     I04 -0.693293266       0 -0.48509448 -0.14917492 -0.05902387
#> 32     S02     I04  1.640861028       0  0.96103270 -0.14917492  0.82900325
#> 33     S03     I04  2.908364439       0  2.62331596 -0.14917492  0.43422340
#> 34     S04     I04 -0.796708980       0 -0.15973365 -0.14917492 -0.48780041
#> 35     S05     I04 -0.115381493       0 -0.32712036 -0.14917492  0.36091378
#> 36     S06     I04 -1.342243478       0 -1.40710108 -0.14917492  0.21403252
#> 37     S07     I04 -2.172716125       0 -1.14135570 -0.14917492 -0.88218550
#> 38     S08     I04 -1.065731758       0  0.08896821 -0.14917492 -1.00552505
#> 39     S09     I04  2.463792770       0  0.98910793 -0.14917492  1.62385975
#> 40     S10     I04  1.722337838       0  0.68879865 -0.14917492  1.18271410
#> 41     S01     I05  0.102446098       0 -0.48509448  0.99305599 -0.40551541
#> 42     S02     I05  0.178248032       0  0.96103270  0.99305599 -1.77584066
#> 43     S03     I05  4.323572619       0  2.62331596  0.99305599  0.70720067
#> 44     S04     I05  1.124858980       0 -0.15973365  0.99305599  0.29153664
#> 45     S05     I05  2.981871548       0 -0.32712036  0.99305599  2.31593591
#> 46     S06     I05 -0.244271765       0 -1.40710108  0.99305599  0.16977332
#> 47     S07     I05 -0.180206219       0 -1.14135570  0.99305599 -0.03190651
#> 48     S08     I05  2.487752379       0  0.08896821  0.99305599  1.40572818
#> 49     S09     I05  3.319228469       0  0.98910793  0.99305599  1.33706454
#> 50     S10     I05  1.984359293       0  0.68879865  0.99305599  0.30250465
#> 51     S01     I06 -0.777734839       0 -0.48509448 -0.72809615  0.43545579
#> 52     S02     I06  0.113821950       0  0.96103270 -0.72809615 -0.11911460
#> 53     S03     I06  2.502962421       0  2.62331596 -0.72809615  0.60774261
#> 54     S04     I06 -0.268750443       0 -0.15973365 -0.72809615  0.61907935
#> 55     S05     I06 -0.241653658       0 -0.32712036 -0.72809615  0.81356285
#> 56     S06     I06 -0.962284577       0 -1.40710108 -0.72809615  1.17291265
#> 57     S07     I06 -1.274198835       0 -1.14135570 -0.72809615  0.59525302
#> 58     S08     I06 -1.132213776       0  0.08896821 -0.72809615 -0.49308584
#> 59     S09     I06 -0.107189836       0  0.98910793 -0.72809615 -0.36820162
#> 60     S10     I06 -2.013780736       0  0.68879865 -0.72809615 -1.97448324
#> 61     S01     I07  2.431592413       0 -0.48509448  1.77193865  1.14474824
#> 62     S02     I07  3.038617755       0  0.96103270  1.77193865  0.30564640
#> 63     S03     I07  4.046810689       0  2.62331596  1.77193865 -0.34844392
#> 64     S04     I07  0.855787996       0 -0.15973365  1.77193865 -0.75641701
#> 65     S05     I07  1.643360250       0 -0.32712036  1.77193865  0.19854195
#> 66     S06     I07  0.062364989       0 -1.40710108  1.77193865 -0.30247259
#> 67     S07     I07  0.049903091       0 -1.14135570  1.77193865 -0.58067986
#> 68     S08     I07  2.137514759       0  0.08896821  1.77193865  0.27660790
#> 69     S09     I07  3.946039811       0  0.98910793  1.77193865  1.18499322
#> 70     S10     I07  4.389011763       0  0.68879865  1.77193865  1.92827446
#> 71     S01     I08 -1.510399666       0 -0.48509448 -1.54921594  0.52391075
#> 72     S02     I08 -1.286218443       0  0.96103270 -1.54921594 -0.69803521
#> 73     S03     I08  1.349095671       0  2.62331596 -1.54921594  0.27499565
#> 74     S04     I08 -1.805083063       0 -0.15973365 -1.54921594 -0.09613348
#> 75     S05     I08 -2.209783134       0 -0.32712036 -1.54921594 -0.33344684
#> 76     S06     I08 -0.277341074       0 -1.40710108 -1.54921594  2.67897594
#> 77     S07     I08 -2.334880517       0 -1.14135570 -1.54921594  0.35569112
#> 78     S08     I08 -2.398109069       0  0.08896821 -1.54921594 -0.93786134
#> 79     S09     I08 -0.423655767       0  0.98910793 -1.54921594  0.13645224
#> 80     S10     I08 -0.085866706       0  0.68879865 -1.54921594  0.77455058
#> 81     S01     I09 -0.722377294       0 -0.48509448  0.68067583 -0.91795864
#> 82     S02     I09  3.058850470       0  0.96103270  0.68067583  1.41714194
#> 83     S03     I09  0.506304727       0  2.62331596  0.68067583 -2.79768706
#> 84     S04     I09 -0.828998971       0 -0.15973365  0.68067583 -1.34994115
#> 85     S05     I09 -0.438363515       0 -0.32712036  0.68067583 -0.79191899
#> 86     S06     I09 -0.105594723       0 -1.40710108  0.68067583  0.62083052
#> 87     S07     I09 -1.325379022       0 -1.14135570  0.68067583 -0.86469915
#> 88     S08     I09  0.499341133       0  0.08896821  0.68067583 -0.27030291
#> 89     S09     I09  3.351904583       0  0.98910793  0.68067583  1.68212082
#> 90     S10     I09  2.467954143       0  0.68879865  0.68067583  1.09847966
#> 91     S01     I10  0.249644470       0 -0.48509448  0.88586719 -0.15112824
#> 92     S02     I10 -0.085891900       0  0.96103270  0.88586719 -1.93279179
#> 93     S03     I10  4.238070764       0  2.62331596  0.88586719  0.72888761
#> 94     S04     I10  0.634758886       0 -0.15973365  0.88586719 -0.09137466
#> 95     S05     I10  0.704621008       0 -0.32712036  0.88586719  0.14587417
#> 96     S06     I10 -0.218330070       0 -1.40710108  0.88586719  0.30290381
#> 97     S07     I10 -0.944570767       0 -1.14135570  0.88586719 -0.68908226
#> 98     S08     I10  1.794607143       0  0.08896821  0.88586719  0.81977174
#> 99     S09     I10  0.777619637       0  0.98910793  0.88586719 -1.09735549
#> 100    S10     I10  0.320868447       0  0.68879865  0.88586719 -1.25379740