Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i      sub_i     item_i         err
#> 1      S01     I01  1.28306972       0  0.5661166  0.9040914 -0.18713833
#> 15     S01     I02  1.38980946       0  0.5661166  1.5744611 -0.75076824
#> 29     S01     I03  2.34036248       0  0.5661166  1.5699988  0.20424701
#> 31     S01     I04  0.27913179       0  0.5661166 -0.8979437  0.61095882
#> 44     S01     I05  0.10734818       0  0.5661166 -0.3730264 -0.08574201
#> 58     S01     I06  1.29697870       0  0.5661166  1.0319977 -0.30113566
#> 61     S01     I07 -1.79426829       0  0.5661166 -1.0106886 -1.34969631
#> 73     S01     I08  0.88619012       0  0.5661166  0.3755291 -0.05545564
#> 87     S01     I09  1.21718060       0  0.5661166  1.0211462 -0.37008223
#> 91     S01     I10  1.64616493       0  0.5661166  0.2375360  0.84251228
#> 7      S02     I01  1.39681263       0  0.3057251  0.9040914  0.18699608
#> 17     S02     I02  1.33412893       0  0.3057251  1.5744611 -0.54605728
#> 30     S02     I03  0.01446239       0  0.3057251  1.5699988 -1.86126158
#> 36     S02     I04 -0.29416045       0  0.3057251 -0.8979437  0.29805808
#> 46     S02     I05  0.60018200       0  0.3057251 -0.3730264  0.66748331
#> 59     S02     I06  1.85381742       0  0.3057251  1.0319977  0.51609456
#> 69     S02     I07 -0.01278841       0  0.3057251 -1.0106886  0.69217507
#> 75     S02     I08  2.50127039       0  0.3057251  0.3755291  1.82001614
#> 88     S02     I09  2.89807819       0  0.3057251  1.0211462  1.57120686
#> 98     S02     I10  1.63919078       0  0.3057251  0.2375360  1.09592963
#> 4      S03     I01  0.79684289       0  1.1490500  0.9040914 -1.25629847
#> 18     S03     I02  2.05027553       0  1.1490500  1.5744611 -0.67323549
#> 23     S03     I03  3.13289638       0  1.1490500  1.5699988  0.41384760
#> 33     S03     I04 -0.03656242       0  1.1490500 -0.8979437 -0.28766871
#> 47     S03     I05  0.93206113       0  1.1490500 -0.3730264  0.15603762
#> 60     S03     I06  1.13984884       0  1.1490500  1.0319977 -1.04119884
#> 62     S03     I07 -0.87008755       0  1.1490500 -1.0106886 -1.00844889
#> 76     S03     I08  1.41299933       0  1.1490500  0.3755291 -0.11157975
#> 89     S03     I09  3.70143465       0  1.1490500  1.0211462  1.53123850
#> 95     S03     I10  0.81391316       0  1.1490500  0.2375360 -0.57267280
#> 5      S04     I01 -0.13725104       0  0.7972705  0.9040914 -1.83861300
#> 19     S04     I02  1.53732566       0  0.7972705  1.5744611 -0.83440595
#> 21     S04     I03  3.88171393       0  0.7972705  1.5699988  1.51444456
#> 34     S04     I04  0.13582112       0  0.7972705 -0.8979437  0.23649423
#> 48     S04     I05 -0.53002431       0  0.7972705 -0.3730264 -0.95426840
#> 51     S04     I06  2.15998197       0  0.7972705  1.0319977  0.33071370
#> 63     S04     I07 -1.79328716       0  0.7972705 -1.0106886 -1.57986909
#> 77     S04     I08  1.89799227       0  0.7972705  0.3755291  0.72519260
#> 90     S04     I09  2.55883042       0  0.7972705  1.0211462  0.74041368
#> 92     S04     I10  0.42052167       0  0.7972705  0.2375360 -0.61428489
#> 2      S05     I01  2.97106937       0  1.0363797  0.9040914  1.03059825
#> 16     S05     I02  3.23239808       0  1.0363797  1.5744611  0.62155730
#> 22     S05     I03  2.34414855       0  1.0363797  1.5699988 -0.26222999
#> 35     S05     I04  0.82648132       0  1.0363797 -0.8979437  0.68804527
#> 45     S05     I05  0.95438135       0  1.0363797 -0.3730264  0.29102809
#> 52     S05     I06  2.48918738       0  1.0363797  1.0319977  0.42080995
#> 64     S05     I07 -0.23050894       0  1.0363797 -1.0106886 -0.25620003
#> 74     S05     I08  1.03875049       0  1.0363797  0.3755291 -0.37315835
#> 81     S05     I09  2.09819029       0  1.0363797  1.0211462  0.04066438
#> 93     S05     I10 -0.40280704       0  1.0363797  0.2375360 -1.67672276
#> 8      S06     I01  2.07224666       0  1.7278340  0.9040914 -0.55967874
#> 20     S06     I02  2.86034958       0  1.7278340  1.5744611 -0.44194548
#> 27     S06     I03  2.90064158       0  1.7278340  1.5699988 -0.39719124
#> 37     S06     I04  0.65614371       0  1.7278340 -0.8979437 -0.17374661
#> 49     S06     I05  0.85767363       0  1.7278340 -0.3730264 -0.49713391
#> 56     S06     I06  2.83584172       0  1.7278340  1.0319977  0.07601001
#> 66     S06     I07  1.15061665       0  1.7278340 -1.0106886  0.43347128
#> 79     S06     I08  3.40040991       0  1.7278340  0.3755291  1.29704680
#> 85     S06     I09  2.23674475       0  1.7278340  1.0211462 -0.51223543
#> 99     S06     I10 -0.22469451       0  1.7278340  0.2375360 -2.19006451
#> 9      S07     I01  0.51692841       0 -0.9946508  0.9040914  0.60748776
#> 11     S07     I02  0.95977424       0 -0.9946508  1.5744611  0.37996393
#> 24     S07     I03  3.05219635       0 -0.9946508  1.5699988  2.47684829
#> 38     S07     I04 -2.86076016       0 -0.9946508 -0.8979437 -0.96816573
#> 50     S07     I05 -0.21675631       0 -0.9946508 -0.3730264  1.15092091
#> 53     S07     I06 -1.33588420       0 -0.9946508  1.0319977 -1.37323116
#> 67     S07     I07 -2.34461676       0 -0.9946508 -1.0106886 -0.33927738
#> 80     S07     I08 -0.75859299       0 -0.9946508  0.3755291 -0.13947135
#> 82     S07     I09 -1.03625552       0 -0.9946508  1.0211462 -1.06275095
#> 96     S07     I10 -0.33745491       0 -0.9946508  0.2375360  0.41965984
#> 6      S08     I01  1.00235519       0  1.3858512  0.9040914 -1.28758738
#> 12     S08     I02  0.62438969       0  1.3858512  1.5744611 -2.33592254
#> 25     S08     I03  2.64231696       0  1.3858512  1.5699988 -0.31353303
#> 39     S08     I04 -1.21838069       0  1.3858512 -0.8979437 -1.70628819
#> 41     S08     I05  0.32566969       0  1.3858512 -0.3730264 -0.68715502
#> 54     S08     I06  1.10027726       0  1.3858512  1.0319977 -1.31757162
#> 68     S08     I07 -0.80296997       0  1.3858512 -1.0106886 -1.17813251
#> 78     S08     I08  1.57852869       0  1.3858512  0.3755291 -0.18285159
#> 83     S08     I09  1.43285319       0  1.3858512  1.0211462 -0.97414416
#> 97     S08     I10  2.54556864       0  1.3858512  0.2375360  0.92218146
#> 3      S09     I01 -0.13455335       0 -0.0108732  0.9040914 -1.02777157
#> 13     S09     I02  1.67407636       0 -0.0108732  1.5744611  0.11048850
#> 26     S09     I03  2.72320393       0 -0.0108732  1.5699988  1.16407830
#> 32     S09     I04 -2.43604136       0 -0.0108732 -0.8979437 -1.52722450
#> 42     S09     I05  0.13121248       0 -0.0108732 -0.3730264  0.51511213
#> 55     S09     I06  3.68455808       0 -0.0108732  1.0319977  2.66343356
#> 65     S09     I07  0.43712856       0 -0.0108732 -1.0106886  1.45869038
#> 71     S09     I08  1.26947702       0 -0.0108732  0.3755291  0.90482110
#> 84     S09     I09  0.92260097       0 -0.0108732  1.0211462 -0.08767202
#> 94     S09     I10  0.15681399       0 -0.0108732  0.2375360 -0.06984882
#> 10     S10     I01 -1.45225022       0 -0.9582099  0.9040914 -1.39813170
#> 14     S10     I02  0.59426948       0 -0.9582099  1.5744611 -0.02198165
#> 28     S10     I03 -0.00955125       0 -0.9582099  1.5699988 -0.62134015
#> 40     S10     I04 -1.21660809       0 -0.9582099 -0.8979437  0.63954550
#> 43     S10     I05  0.20083426       0 -0.9582099 -0.3730264  1.53207065
#> 57     S10     I06 -1.71927967       0 -0.9582099  1.0319977 -1.79306746
#> 70     S10     I07 -0.95912138       0 -0.9582099 -1.0106886  1.00977717
#> 72     S10     I08 -0.92179528       0 -0.9582099  0.3755291 -0.33911447
#> 86     S10     I09  0.81515038       0 -0.9582099  1.0211462  0.75221412
#> 100    S10     I10 -1.99862965       0 -0.9582099  0.2375360 -1.27795573