Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, empirical = FALSE, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i       sub_i     item_i           err
#> 1      S01     I01  1.97210453       0  0.04217639  1.6027842  0.3271439090
#> 2      S02     I01  2.25161078       0  0.22819161  1.6027842  0.4206349389
#> 3      S03     I01  0.64109018       0 -0.64615330  1.6027842 -0.3155407594
#> 4      S04     I01  2.44142129       0 -0.71346653  1.6027842  1.5521035848
#> 5      S05     I01 -0.29480639       0  0.07323848  1.6027842 -1.9708291081
#> 6      S06     I01  3.47503733       0  0.73331173  1.6027842  1.1389413693
#> 7      S07     I01  0.33642034       0 -1.12456469  1.6027842 -0.1417992078
#> 8      S08     I01  2.96479336       0  0.34519039  1.6027842  1.0168187354
#> 9      S09     I01  0.03940938       0 -1.22974103  1.6027842 -0.3336338169
#> 10     S10     I01  2.76546725       0  0.20058942  1.6027842  0.9620935945
#> 11     S01     I02  1.02391197       0  0.04217639  0.3913837  0.5903519005
#> 12     S02     I02  0.45447079       0  0.22819161  0.3913837 -0.1651044976
#> 13     S03     I02 -1.22912433       0 -0.64615330  0.3913837 -0.9743547166
#> 14     S04     I02  0.04529084       0 -0.71346653  0.3913837  0.3673736928
#> 15     S05     I02 -0.25259085       0  0.07323848  0.3913837 -0.7172130142
#> 16     S06     I02 -0.97398393       0  0.73331173  0.3913837 -2.0986793352
#> 17     S07     I02 -0.49406538       0 -1.12456469  0.3913837  0.2391156257
#> 18     S08     I02  0.64800554       0  0.34519039  0.3913837 -0.0885685321
#> 19     S09     I02 -0.42554240       0 -1.22974103  0.3913837  0.4128149453
#> 20     S10     I02  1.66688911       0  0.20058942  0.3913837  1.0749160043
#> 21     S01     I03 -1.04487249       0  0.04217639 -0.3485825 -0.7384663743
#> 22     S02     I03 -0.42568570       0  0.22819161 -0.3485825 -0.3052948066
#> 23     S03     I03 -1.75580833       0 -0.64615330 -0.3485825 -0.7610725258
#> 24     S04     I03 -2.54997993       0 -0.71346653 -0.3485825 -1.4879308973
#> 25     S05     I03 -1.58831114       0  0.07323848 -0.3485825 -1.3129671206
#> 26     S06     I03  0.03572491       0  0.73331173 -0.3485825 -0.3490043135
#> 27     S07     I03 -2.08015346       0 -1.12456469 -0.3485825 -0.6070062718
#> 28     S08     I03  0.35292964       0  0.34519039 -0.3485825  0.3563217566
#> 29     S09     I03 -0.45492981       0 -1.22974103 -0.3485825  1.1233937263
#> 30     S10     I03  1.76409854       0  0.20058942 -0.3485825  1.9120916196
#> 31     S01     I04  0.74701390       0  0.04217639  1.0589849 -0.3541474021
#> 32     S02     I04  0.61564067       0  0.22819161  1.0589849 -0.6715358484
#> 33     S03     I04  1.35397618       0 -0.64615330  1.0589849  0.9411445714
#> 34     S04     I04  0.63274038       0 -0.71346653  1.0589849  0.2872220058
#> 35     S05     I04  1.76264870       0  0.07323848  1.0589849  0.6304253059
#> 36     S06     I04  1.93057502       0  0.73331173  1.0589849  0.1382783815
#> 37     S07     I04 -0.35706627       0 -1.12456469  1.0589849 -0.2914864933
#> 38     S08     I04  1.25253126       0  0.34519039  1.0589849 -0.1516440374
#> 39     S09     I04 -0.04565918       0 -1.22974103  1.0589849  0.1250969411
#> 40     S10     I04  0.83138289       0  0.20058942  1.0589849 -0.4281914450
#> 41     S01     I05  0.09680222       0  0.04217639  0.2251011 -0.1704752782
#> 42     S02     I05 -0.31225848       0  0.22819161  0.2251011 -0.7655511900
#> 43     S03     I05 -1.46608752       0 -0.64615330  0.2251011 -1.0450353220
#> 44     S04     I05  0.46467691       0 -0.71346653  0.2251011  0.9530423358
#> 45     S05     I05  1.15546637       0  0.07323848  0.2251011  0.8571267894
#> 46     S06     I05  1.61298893       0  0.73331173  0.2251011  0.6545760970
#> 47     S07     I05 -1.43251780       0 -1.12456469  0.2251011 -0.5330542115
#> 48     S08     I05  0.67058553       0  0.34519039  0.2251011  0.1002940404
#> 49     S09     I05 -0.17050573       0 -1.22974103  0.2251011  0.8341342019
#> 50     S10     I05  0.05976629       0  0.20058942  0.2251011 -0.3659242365
#> 51     S01     I06 -2.24023924       0  0.04217639 -1.2135727 -1.0688428849
#> 52     S02     I06  0.56735960       0  0.22819161 -1.2135727  1.5527407402
#> 53     S03     I06 -2.09274062       0 -0.64615330 -1.2135727 -0.2330145754
#> 54     S04     I06 -2.09882813       0 -0.71346653 -1.2135727 -0.1717888529
#> 55     S05     I06 -1.91787952       0  0.07323848 -1.2135727 -0.7775452530
#> 56     S06     I06  0.01249567       0  0.73331173 -1.2135727  0.4927566908
#> 57     S07     I06 -3.14676600       0 -1.12456469 -1.2135727 -0.8086285592
#> 58     S08     I06 -0.78143177       0  0.34519039 -1.2135727  0.0869505907
#> 59     S09     I06 -3.38243385       0 -1.22974103 -1.2135727 -0.9391200735
#> 60     S10     I06 -1.23339043       0  0.20058942 -1.2135727 -0.2204071062
#> 61     S01     I07 -1.17217659       0  0.04217639 -0.6813815 -0.5329715190
#> 62     S02     I07 -1.75089099       0  0.22819161 -0.6813815 -1.2977011313
#> 63     S03     I07 -2.21321755       0 -0.64615330 -0.6813815 -0.8856827851
#> 64     S04     I07  0.09323872       0 -0.71346653 -0.6813815  1.4880867126
#> 65     S05     I07 -1.90553199       0  0.07323848 -0.6813815 -1.2973890063
#> 66     S06     I07 -0.56791573       0  0.73331173 -0.6813815 -0.6198459917
#> 67     S07     I07 -1.22686718       0 -1.12456469 -0.6813815  0.5790789738
#> 68     S08     I07 -0.94339606       0  0.34519039 -0.6813815 -0.6072049817
#> 69     S09     I07 -1.76336728       0 -1.22974103 -0.6813815  0.1477552196
#> 70     S10     I07  0.60664819       0  0.20058942 -0.6813815  1.0874402282
#> 71     S01     I08 -0.15005756       0  0.04217639 -0.8144630  0.6222290916
#> 72     S02     I08 -0.26641670       0  0.22819161 -0.8144630  0.3198547319
#> 73     S03     I08 -1.69830619       0 -0.64615330 -0.8144630 -0.2376898524
#> 74     S04     I08 -0.75656788       0 -0.71346653 -0.8144630  0.7713616974
#> 75     S05     I08 -2.07084442       0  0.07323848 -0.8144630 -1.3296198554
#> 76     S06     I08 -0.39424454       0  0.73331173 -0.8144630 -0.3130932271
#> 77     S07     I08 -1.50885206       0 -1.12456469 -0.8144630  0.4301756720
#> 78     S08     I08  0.04922681       0  0.34519039 -0.8144630  0.5184994654
#> 79     S09     I08 -2.04387174       0 -1.22974103 -0.8144630  0.0003323363
#> 80     S10     I08 -0.39872021       0  0.20058942 -0.8144630  0.2151534150
#> 81     S01     I09  0.52359416       0  0.04217639  1.1266349 -0.6452171406
#> 82     S02     I09  0.93219551       0  0.22819161  1.1266349 -0.4226310080
#> 83     S03     I09 -0.21045712       0 -0.64615330  1.1266349 -0.6909387353
#> 84     S04     I09  0.19760934       0 -0.71346653  1.1266349 -0.2155590435
#> 85     S05     I09  1.04414969       0  0.07323848  1.1266349 -0.1557236999
#> 86     S06     I09  2.45203559       0  0.73331173  1.1266349  0.5920889569
#> 87     S07     I09  0.54144023       0 -1.12456469  1.1266349  0.5393700080
#> 88     S08     I09 -1.43369402       0  0.34519039  1.1266349 -2.9055193209
#> 89     S09     I09 -0.54056217       0 -1.22974103  1.1266349 -0.4374560487
#> 90     S10     I09  1.63176383       0  0.20058942  1.1266349  0.3045394996
#> 91     S01     I10 -1.68561297       0  0.04217639 -1.3873328 -0.3404565311
#> 92     S02     I10  0.13413271       0  0.22819161 -1.3873328  1.2932739283
#> 93     S03     I10 -1.15305321       0 -0.64615330 -1.3873328  0.8804329184
#> 94     S04     I10 -2.20701483       0 -0.71346653 -1.3873328 -0.1062154662
#> 95     S05     I10 -1.59960846       0  0.07323848 -1.3873328 -0.2855141130
#> 96     S06     I10 -0.68814339       0  0.73331173 -1.3873328 -0.0341222852
#> 97     S07     I10 -2.71316224       0 -1.12456469 -1.3873328 -0.2012647182
#> 98     S08     I10 -2.94156864       0  0.34519039 -1.3873328 -1.8994261995
#> 99     S09     I10 -4.58354500       0 -1.22974103 -1.3873328 -1.9664711391
#> 100    S10     I10 -1.28249953       0  0.20058942 -1.3873328 -0.0957561170