Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i       sub_i     item_i         err
#> 1      S01     I01 -0.57149255       0  1.27662687 -0.4788730 -1.36924644
#> 15     S01     I02  0.75470727       0  1.27662687 -0.4279532 -0.09396641
#> 29     S01     I03  0.56921386       0  1.27662687 -0.9791689  0.27175593
#> 31     S01     I04  0.86021936       0  1.27662687 -1.7457328  1.32932529
#> 44     S01     I05  0.03770291       0  1.27662687 -1.5333346  0.29441061
#> 58     S01     I06  0.19660132       0  1.27662687 -0.6034774 -0.47654819
#> 61     S01     I07  1.71675319       0  1.27662687 -1.2814310  1.72155733
#> 73     S01     I08 -1.20865637       0  1.27662687 -1.2186536 -1.26662960
#> 87     S01     I09  1.32465785       0  1.27662687  1.0071978 -0.95916684
#> 91     S01     I10  1.06166400       0  1.27662687 -0.2361820  0.02121911
#> 7      S02     I01  1.52401242       0  1.31772174 -0.4788730  0.68516365
#> 17     S02     I02  0.78540959       0  1.31772174 -0.4279532 -0.10435896
#> 30     S02     I03 -0.34563077       0  1.31772174 -0.9791689 -0.68418357
#> 36     S02     I04 -0.45139451       0  1.31772174 -1.7457328 -0.02338344
#> 46     S02     I05 -0.19029434       0  1.31772174 -1.5333346  0.02531850
#> 59     S02     I06  0.30735848       0  1.31772174 -0.6034774 -0.40688590
#> 69     S02     I07  0.21568408       0  1.31772174 -1.2814310  0.17939335
#> 75     S02     I08 -0.88048981       0  1.31772174 -1.2186536 -0.97955791
#> 88     S02     I09  2.58007996       0  1.31772174  1.0071978  0.25516040
#> 98     S02     I10  1.32799947       0  1.31772174 -0.2361820  0.24645970
#> 4      S03     I01 -1.44514135       0 -0.50308233 -0.4788730 -0.46318605
#> 18     S03     I02 -0.76312175       0 -0.50308233 -0.4279532  0.16791377
#> 23     S03     I03 -1.27376013       0 -0.50308233 -0.9791689  0.20849114
#> 33     S03     I04 -2.99965894       0 -0.50308233 -1.7457328 -0.75084381
#> 47     S03     I05 -2.42928596       0 -0.50308233 -1.5333346 -0.39286906
#> 60     S03     I06  0.87891595       0 -0.50308233 -0.6034774  1.98547563
#> 62     S03     I07 -1.63613358       0 -0.50308233 -1.2814310  0.14837976
#> 76     S03     I08 -1.81384844       0 -0.50308233 -1.2186536 -0.09211247
#> 89     S03     I09  0.26194348       0 -0.50308233  1.0071978 -0.24217202
#> 95     S03     I10 -0.68351826       0 -0.50308233 -0.2361820  0.05574604
#> 5      S04     I01  2.17241681       0  0.22596692 -0.4788730  2.42532286
#> 19     S04     I02  0.08038692       0  0.22596692 -0.4279532  0.28237319
#> 21     S04     I03  0.64459726       0  0.22596692 -0.9791689  1.39779928
#> 34     S04     I04 -2.47567302       0  0.22596692 -1.7457328 -0.95590714
#> 48     S04     I05 -2.28490710       0  0.22596692 -1.5333346 -0.97753945
#> 51     S04     I06 -0.20627578       0  0.22596692 -0.6034774  0.17123465
#> 63     S04     I07 -0.85777488       0  0.22596692 -1.2814310  0.19768920
#> 77     S04     I08 -0.51370385       0  0.22596692 -1.2186536  0.47898287
#> 90     S04     I09  0.60731237       0  0.22596692  1.0071978 -0.62585237
#> 92     S04     I10 -0.32075334       0  0.22596692 -0.2361820 -0.31053829
#> 2      S05     I01 -0.50902090       0  1.35618977 -0.4788730 -1.38633770
#> 16     S05     I02 -0.74943255       0  1.35618977 -0.4279532 -1.67766914
#> 22     S05     I03  2.66459235       0  1.35618977 -0.9791689  2.28757152
#> 35     S05     I04 -0.02397857       0  1.35618977 -1.7457328  0.36556447
#> 45     S05     I05 -1.28569787       0  1.35618977 -1.5333346 -1.10855306
#> 52     S05     I06 -0.58451148       0  1.35618977 -0.6034774 -1.33722389
#> 64     S05     I07  1.90725267       0  1.35618977 -1.2814310  1.83249391
#> 74     S05     I08  0.95141767       0  1.35618977 -1.2186536  0.81388155
#> 81     S05     I09  3.10305217       0  1.35618977  1.0071978  0.73966458
#> 93     S05     I10  2.26934531       0  1.35618977 -0.2361820  1.14933752
#> 8      S06     I01 -1.11851209       0  0.16154024 -0.4788730 -0.80117935
#> 20     S06     I02 -0.20341757       0  0.16154024 -0.4279532  0.06299538
#> 27     S06     I03 -0.02042033       0  0.16154024 -0.9791689  0.79720838
#> 37     S06     I04 -0.79469002       0  0.16154024 -1.7457328  0.78950255
#> 49     S06     I05 -0.27632761       0  0.16154024 -1.5333346  1.09546672
#> 56     S06     I06  0.59606256       0  0.16154024 -0.6034774  1.03799968
#> 66     S06     I07 -0.03739999       0  0.16154024 -1.2814310  1.08249079
#> 79     S06     I08 -0.97679008       0  0.16154024 -1.2186536  0.08032333
#> 85     S06     I09 -0.00106152       0  0.16154024  1.0071978 -1.16979958
#> 99     S06     I10  1.41488428       0  0.16154024 -0.2361820  1.48952602
#> 9      S07     I01 -0.83902262       0 -1.02959421 -0.4788730  0.66944458
#> 11     S07     I02 -1.80424940       0 -1.02959421 -0.4279532 -0.34670200
#> 24     S07     I03 -2.22307020       0 -1.02959421 -0.9791689 -0.21430705
#> 38     S07     I04 -4.84904482       0 -1.02959421 -1.7457328 -2.07371780
#> 50     S07     I05 -3.62027654       0 -1.02959421 -1.5333346 -1.05734775
#> 53     S07     I06 -1.40386842       0 -1.02959421 -0.6034774  0.22920314
#> 67     S07     I07 -1.55739674       0 -1.02959421 -1.2814310  0.75362849
#> 80     S07     I08 -2.89668798       0 -1.02959421 -1.2186536 -0.64844012
#> 82     S07     I09  1.01589213       0 -1.02959421  1.0071978  1.03828853
#> 96     S07     I10  0.12908425       0 -1.02959421 -0.2361820  1.39486045
#> 6      S08     I01 -0.72915479       0 -0.88644058 -0.4788730  0.63615876
#> 12     S08     I02 -1.90713476       0 -0.88644058 -0.4279532 -0.59274099
#> 25     S08     I03 -2.44045255       0 -0.88644058 -0.9791689 -0.57484303
#> 39     S08     I04 -1.74102940       0 -0.88644058 -1.7457328  0.89114398
#> 41     S08     I05 -2.70828244       0 -0.88644058 -1.5333346 -0.28850729
#> 54     S08     I06  1.57572301       0 -0.88644058 -0.6034774  3.06564095
#> 68     S08     I07 -0.42232405       0 -0.88644058 -1.2814310  1.74554754
#> 78     S08     I08 -2.23236120       0 -0.88644058 -1.2186536 -0.12726698
#> 83     S08     I09 -0.31508727       0 -0.88644058  1.0071978 -0.43584451
#> 97     S08     I10 -0.04531878       0 -0.88644058 -0.2361820  1.07730378
#> 3      S09     I01 -0.94013179       0  0.02681869 -0.4788730 -0.48807751
#> 13     S09     I02 -0.74552651       0  0.02681869 -0.4279532 -0.34439201
#> 26     S09     I03 -1.63053328       0  0.02681869 -0.9791689 -0.67818303
#> 32     S09     I04 -3.15663227       0  0.02681869 -1.7457328 -1.43771815
#> 42     S09     I05 -1.59599813       0  0.02681869 -1.5333346 -0.08948225
#> 55     S09     I06  0.65302390       0  0.02681869 -0.6034774  1.22968257
#> 65     S09     I07  0.05511114       0  0.02681869 -1.2814310  1.30972345
#> 71     S09     I08 -1.88026951       0  0.02681869 -1.2186536 -0.68843455
#> 84     S09     I09  0.39625947       0  0.02681869  1.0071978 -0.63775704
#> 94     S09     I10  0.95572504       0  0.02681869 -0.2361820  1.16508833
#> 10     S10     I01 -0.36634797       0  0.39093855 -0.4788730 -0.27841354
#> 14     S10     I02 -0.25371924       0  0.39093855 -0.4279532 -0.21670461
#> 28     S10     I03  0.24185820       0  0.39093855 -0.9791689  0.83008858
#> 40     S10     I04 -1.70280346       0  0.39093855 -1.7457328 -0.34800921
#> 43     S10     I05 -0.40166490       0  0.39093855 -1.5333346  0.74073112
#> 57     S10     I06 -0.51502208       0  0.39093855 -0.6034774 -0.30248328
#> 70     S10     I07 -0.70303698       0  0.39093855 -1.2814310  0.18745548
#> 72     S10     I08 -0.75975198       0  0.39093855 -1.2186536  0.06796312
#> 86     S10     I09  0.80072562       0  0.39093855  1.0071978 -0.59741075
#> 100    S10     I10  1.57543893       0  0.39093855 -0.2361820  1.42068236