Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id             y grand_i      sub_i      item_i         err
#> 1      S01     I01  2.0004192704       0  1.3038108  1.40562619 -0.70901773
#> 15     S01     I02  1.1713293550       0  1.3038108 -0.92361708  0.79113563
#> 29     S01     I03  2.2246354050       0  1.3038108  0.98624885 -0.06542426
#> 31     S01     I04  0.7756846367       0  1.3038108 -0.83710697  0.30898080
#> 44     S01     I05  2.2240424485       0  1.3038108  1.19158137 -0.27134973
#> 58     S01     I06  0.2222517633       0  1.3038108 -1.05821892 -0.02334013
#> 61     S01     I07  0.4680648191       0  1.3038108 -0.06533498 -0.77041102
#> 73     S01     I08  1.2877222965       0  1.3038108 -1.23481702  1.21872850
#> 87     S01     I09  0.6749877595       0  1.3038108 -1.03765448  0.40883142
#> 91     S01     I10  2.4782862031       0  1.3038108  1.51763371 -0.34315832
#> 7      S02     I01  0.6275421017       0 -0.4741852  1.40562619 -0.30389888
#> 17     S02     I02 -2.2218857338       0 -0.4741852 -0.92361708 -0.82408344
#> 30     S02     I03  0.1050223311       0 -0.4741852  0.98624885 -0.40704131
#> 36     S02     I04 -1.7065566543       0 -0.4741852 -0.83710697 -0.39526447
#> 46     S02     I05 -0.6462740706       0 -0.4741852  1.19158137 -1.36367022
#> 59     S02     I06 -1.9357491987       0 -0.4741852 -1.05821892 -0.40334507
#> 69     S02     I07 -1.9867247095       0 -0.4741852 -0.06533498 -1.44720452
#> 75     S02     I08 -0.8648522638       0 -0.4741852 -1.23481702  0.84414997
#> 88     S02     I09 -2.0515671479       0 -0.4741852 -1.03765448 -0.53972746
#> 98     S02     I10  0.5986747320       0 -0.4741852  1.51763371 -0.44477377
#> 4      S03     I01  4.7647700669       0  2.5564164  1.40562619  0.80272745
#> 18     S03     I02  1.3811783905       0  2.5564164 -0.92361708 -0.25162095
#> 23     S03     I03  3.7774632450       0  2.5564164  0.98624885  0.23479797
#> 33     S03     I04 -0.3708535117       0  2.5564164 -0.83710697 -2.09016296
#> 47     S03     I05  2.4215702243       0  2.5564164  1.19158137 -1.32642756
#> 60     S03     I06  0.1318672257       0  2.5564164 -1.05821892 -1.36633028
#> 62     S03     I07  2.3890005906       0  2.5564164 -0.06533498 -0.10208086
#> 76     S03     I08  1.1076501575       0  2.5564164 -1.23481702 -0.21394925
#> 89     S03     I09  0.2363029469       0  2.5564164 -1.03765448 -1.28245900
#> 95     S03     I10  3.8221935337       0  2.5564164  1.51763371 -0.25185660
#> 5      S04     I01 -1.1967094977       0 -0.6964105  1.40562619 -1.90592521
#> 19     S04     I02  0.4139191645       0 -0.6964105 -0.92361708  2.03394672
#> 21     S04     I03  1.5809068180       0 -0.6964105  0.98624885  1.29106845
#> 34     S04     I04 -2.8393110036       0 -0.6964105 -0.83710697 -1.30579355
#> 48     S04     I05  0.8196696047       0 -0.6964105  1.19158137  0.32449872
#> 51     S04     I06 -2.6758793279       0 -0.6964105 -1.05821892 -0.92124993
#> 63     S04     I07 -1.7108677977       0 -0.6964105 -0.06533498 -0.94912234
#> 77     S04     I08 -1.0812624550       0 -0.6964105 -1.23481702  0.84996504
#> 90     S04     I09 -0.7495894137       0 -0.6964105 -1.03765448  0.98447554
#> 92     S04     I10  0.6219933692       0 -0.6964105  1.51763371 -0.19922987
#> 2      S05     I01  1.4514028019       0 -0.5288203  1.40562619  0.57459688
#> 16     S05     I02 -2.9402833664       0 -0.5288203 -0.92361708 -1.48784601
#> 22     S05     I03  0.1618011193       0 -0.5288203  0.98624885 -0.29562746
#> 35     S05     I04 -2.8429112569       0 -0.5288203 -0.83710697 -1.47698401
#> 45     S05     I05  0.5270435380       0 -0.5288203  1.19158137 -0.13571755
#> 52     S05     I06 -0.9239336745       0 -0.5288203 -1.05821892  0.66310552
#> 64     S05     I07  1.2779964400       0 -0.5288203 -0.06533498  1.87215169
#> 74     S05     I08 -2.0521918816       0 -0.5288203 -1.23481702 -0.28855459
#> 81     S05     I09 -0.3618586237       0 -0.5288203 -1.03765448  1.20461613
#> 93     S05     I10  1.2378346438       0 -0.5288203  1.51763371  0.24902121
#> 8      S06     I01  0.7606374818       0  0.7704003  1.40562619 -1.41538905
#> 20     S06     I02  0.1104094357       0  0.7704003 -0.92361708  0.26362618
#> 27     S06     I03  2.4021160888       0  0.7704003  0.98624885  0.64546690
#> 37     S06     I04  0.8740931483       0  0.7704003 -0.83710697  0.94079978
#> 49     S06     I05  1.8297413231       0  0.7704003  1.19158137 -0.13224038
#> 56     S06     I06  1.1017429161       0  0.7704003 -1.05821892  1.38956150
#> 66     S06     I07  0.2067194110       0  0.7704003 -0.06533498 -0.49834595
#> 79     S06     I08 -1.9692157499       0  0.7704003 -1.23481702 -1.50479907
#> 85     S06     I09 -1.5192483906       0  0.7704003 -1.03765448 -1.25199425
#> 99     S06     I10  3.7064258970       0  0.7704003  1.51763371  1.41839185
#> 9      S07     I01 -0.4608676076       0 -0.2093763  1.40562619 -1.65711748
#> 11     S07     I02  0.2107250962       0 -0.2093763 -0.92361708  1.34371850
#> 24     S07     I03  0.5727876551       0 -0.2093763  0.98624885 -0.20408487
#> 38     S07     I04 -2.4390585932       0 -0.2093763 -0.83710697 -1.39257530
#> 50     S07     I05  2.1465887439       0 -0.2093763  1.19158137  1.16438370
#> 53     S07     I06 -0.7559042672       0 -0.2093763 -1.05821892  0.51169097
#> 67     S07     I07 -0.1547317483       0 -0.2093763 -0.06533498  0.11997955
#> 80     S07     I08 -2.2911624679       0 -0.2093763 -1.23481702 -0.84696913
#> 82     S07     I09 -1.9431221030       0 -0.2093763 -1.03765448 -0.69609131
#> 96     S07     I10 -0.7163302996       0 -0.2093763  1.51763371 -2.02458769
#> 6      S08     I01 -0.1664058679       0 -1.2123358  1.40562619 -0.35969624
#> 12     S08     I02 -2.8394137510       0 -1.2123358 -0.92361708 -0.70346085
#> 25     S08     I03  0.4366254976       0 -1.2123358  0.98624885  0.66271247
#> 39     S08     I04 -2.7376119491       0 -1.2123358 -0.83710697 -0.68816916
#> 41     S08     I05  0.5625193949       0 -1.2123358  1.19158137  0.58327385
#> 54     S08     I06 -1.4777349395       0 -1.2123358 -1.05821892  0.79281980
#> 68     S08     I07 -1.8198370231       0 -1.2123358 -0.06533498 -0.54216623
#> 78     S08     I08 -2.0599580373       0 -1.2123358 -1.23481702  0.38719480
#> 83     S08     I09 -2.8193331860       0 -1.2123358 -1.03765448 -0.56934289
#> 97     S08     I10  0.1421580397       0 -1.2123358  1.51763371 -0.16313985
#> 3      S09     I01  1.8513709889       0 -0.2780400  1.40562619  0.72378480
#> 13     S09     I02 -1.7921206752       0 -0.2780400 -0.92361708 -0.59046359
#> 26     S09     I03  2.0086345929       0 -0.2780400  0.98624885  1.30042575
#> 32     S09     I04 -0.7019323863       0 -0.2780400 -0.83710697  0.41321459
#> 42     S09     I05 -0.0404953193       0 -0.2780400  1.19158137 -0.95403668
#> 55     S09     I06 -0.1949078703       0 -0.2780400 -1.05821892  1.14135105
#> 65     S09     I07  0.1956432889       0 -0.2780400 -0.06533498  0.53901827
#> 71     S09     I08 -1.3373997362       0 -0.2780400 -1.23481702  0.17545728
#> 84     S09     I09 -2.1770315279       0 -0.2780400 -1.03765448 -0.86133705
#> 94     S09     I10  2.0045214368       0 -0.2780400  1.51763371  0.76492773
#> 10     S10     I01 -0.9079536543       0 -1.2690850  1.40562619 -1.04449487
#> 14     S10     I02 -3.4067982210       0 -1.2690850 -0.92361708 -1.21409616
#> 28     S10     I03 -0.3692341161       0 -1.2690850  0.98624885 -0.08639799
#> 40     S10     I04 -3.1889142801       0 -1.2690850 -0.83710697 -1.08272233
#> 43     S10     I05  0.9685118098       0 -1.2690850  1.19158137  1.04601542
#> 57     S10     I06 -2.7333323066       0 -1.2690850 -1.05821892 -0.40602841
#> 70     S10     I07  0.1861507922       0 -1.2690850 -0.06533498  1.52057075
#> 72     S10     I08 -2.0177451512       0 -1.2690850 -1.23481702  0.48615685
#> 86     S10     I09 -2.5041761949       0 -1.2690850 -1.03765448 -0.19743674
#> 100    S10     I10  0.0008962212       0 -1.2690850  1.51763371 -0.24765251