Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id            y grand_i      sub_i     item_i         err
#> 1      S01     I01 -3.251555640       0 -1.0933128 -2.2805280  0.12228514
#> 15     S01     I02 -1.851005204       0 -1.0933128 -1.0255685  0.26787615
#> 29     S01     I03 -1.390906708       0 -1.0933128 -0.5102049  0.21261102
#> 31     S01     I04 -1.617827203       0 -1.0933128 -0.3906218 -0.13389262
#> 44     S01     I05 -0.306424397       0 -1.0933128  0.2110268  0.57586158
#> 58     S01     I06 -1.192460606       0 -1.0933128 -0.1854311  0.08628335
#> 61     S01     I07 -4.025276824       0 -1.0933128 -2.1994373 -0.73252676
#> 73     S01     I08 -0.343640393       0 -1.0933128 -0.1822684  0.93194080
#> 87     S01     I09 -2.429076596       0 -1.0933128  0.1812328 -1.51699658
#> 91     S01     I10  0.944345854       0 -1.0933128  1.2116043  0.82605439
#> 7      S02     I01 -0.773433000       0  1.2269399 -2.2805280  0.28015505
#> 17     S02     I02  1.532269630       0  1.2269399 -1.0255685  1.33089825
#> 30     S02     I03  1.631832238       0  1.2269399 -0.5102049  0.91509724
#> 36     S02     I04  1.944916820       0  1.2269399 -0.3906218  1.10859867
#> 46     S02     I05  3.134732788       0  1.2269399  0.2110268  1.69676604
#> 59     S02     I06 -1.486422043       0  1.2269399 -0.1854311 -2.52793082
#> 69     S02     I07  0.163382637       0  1.2269399 -2.1994373  1.13587997
#> 75     S02     I08  2.060254032       0  1.2269399 -0.1822684  1.01558250
#> 88     S02     I09  1.215851593       0  1.2269399  0.1812328 -0.19232112
#> 98     S02     I10  2.324614560       0  1.2269399  1.2116043 -0.11392964
#> 4      S03     I01 -0.721495163       0 -0.1957095 -2.2805280  1.75474230
#> 18     S03     I02 -2.151125422       0 -0.1957095 -1.0255685 -0.92984739
#> 23     S03     I03 -1.336808367       0 -0.1957095 -0.5102049 -0.63089395
#> 33     S03     I04 -0.003187314       0 -0.1957095 -0.3906218  0.58314395
#> 47     S03     I05  0.859461960       0 -0.1957095  0.2110268  0.84414462
#> 60     S03     I06 -0.662850779       0 -0.1957095 -0.1854311 -0.28171014
#> 62     S03     I07 -0.642091135       0 -0.1957095 -2.1994373  1.75305561
#> 76     S03     I08 -0.771932591       0 -0.1957095 -0.1822684 -0.39395471
#> 89     S03     I09 -1.053161622       0 -0.1957095  0.1812328 -1.03868492
#> 95     S03     I10  2.074628263       0 -0.1957095  1.2116043  1.05873348
#> 5      S04     I01 -0.230073329       0  1.4196822 -2.2805280  0.63077248
#> 19     S04     I02  1.271973590       0  1.4196822 -1.0255685  0.87785997
#> 21     S04     I03  0.893478213       0  1.4196822 -0.5102049 -0.01599903
#> 34     S04     I04  0.521050191       0  1.4196822 -0.3906218 -0.50801020
#> 48     S04     I05  2.728176842       0  1.4196822  0.2110268  1.09746785
#> 51     S04     I06  0.252287855       0  1.4196822 -0.1854311 -0.98196316
#> 63     S04     I07 -1.702166936       0  1.4196822 -2.1994373 -0.92241184
#> 77     S04     I08  1.210144474       0  1.4196822 -0.1822684 -0.02726930
#> 90     S04     I09  1.863358807       0  1.4196822  0.1812328  0.26244385
#> 92     S04     I10  1.398374433       0  1.4196822  1.2116043 -1.23291200
#> 2      S05     I01 -3.068392143       0 -0.9949279 -2.2805280  0.20706376
#> 16     S05     I02 -2.736549670       0 -0.9949279 -1.0255685 -0.71605320
#> 22     S05     I03 -2.069097882       0 -0.9949279 -0.5102049 -0.56396503
#> 35     S05     I04  0.015326650       0 -0.9949279 -0.3906218  1.40087635
#> 45     S05     I05 -1.735127692       0 -0.9949279  0.2110268 -0.95122659
#> 52     S05     I06 -2.490137013       0 -0.9949279 -0.1854311 -1.30977793
#> 64     S05     I07 -2.763086707       0 -0.9949279 -2.1994373  0.43127848
#> 74     S05     I08 -1.353850058       0 -0.9949279 -0.1822684 -0.17665374
#> 81     S05     I09 -0.171143537       0 -0.9949279  0.1812328  0.64255160
#> 93     S05     I10  0.381921817       0 -0.9949279  1.2116043  0.16524548
#> 8      S06     I01 -2.661141564       0 -0.5320612 -2.2805280  0.15144762
#> 20     S06     I02 -0.910032228       0 -0.5320612 -1.0255685  0.64759753
#> 27     S06     I03 -1.762424505       0 -0.5320612 -0.5102049 -0.72015837
#> 37     S06     I04  0.003285302       0 -0.5320612 -0.3906218  0.92596829
#> 49     S06     I05 -0.164212061       0 -0.5320612  0.2110268  0.15682232
#> 56     S06     I06  0.335041751       0 -0.5320612 -0.1854311  1.05253411
#> 66     S06     I07 -2.510777609       0 -0.5320612 -2.1994373  0.22072086
#> 79     S06     I08 -3.403794678       0 -0.5320612 -0.1822684 -2.68946508
#> 85     S06     I09 -0.754394318       0 -0.5320612  0.1812328 -0.40356590
#> 99     S06     I10  1.029510980       0 -0.5320612  1.2116043  0.34996792
#> 9      S07     I01 -2.593790416       0 -0.4006895 -2.2805280  0.08742706
#> 11     S07     I02 -0.304845036       0 -0.4006895 -1.0255685  1.12141301
#> 24     S07     I03  0.619818589       0 -0.4006895 -0.5102049  1.53071301
#> 38     S07     I04 -2.480214929       0 -0.4006895 -0.3906218 -1.68890365
#> 50     S07     I05 -1.479370173       0 -0.4006895  0.2110268 -1.28970750
#> 53     S07     I06  0.515777993       0 -0.4006895 -0.1854311  1.10189864
#> 67     S07     I07 -3.143974999       0 -0.4006895 -2.1994373 -0.54384824
#> 80     S07     I08 -0.407739052       0 -0.4006895 -0.1822684  0.17521884
#> 82     S07     I09 -0.155472015       0 -0.4006895  0.1812328  0.06398470
#> 96     S07     I10  0.485831518       0 -0.4006895  1.2116043 -0.32508325
#> 6      S08     I01 -0.935827145       0  0.4811886 -2.2805280  0.86351220
#> 12     S08     I02 -0.649548158       0  0.4811886 -1.0255685 -0.10516825
#> 25     S08     I03 -0.086549245       0  0.4811886 -0.5102049 -0.05753295
#> 39     S08     I04  1.453692551       0  0.4811886 -0.3906218  1.36312569
#> 41     S08     I05  2.225776944       0  0.4811886  0.2110268  1.53356148
#> 54     S08     I06  0.019908610       0  0.4811886 -0.1854311 -0.27584887
#> 68     S08     I07 -1.520430713       0  0.4811886 -2.1994373  0.19781791
#> 78     S08     I08  0.855809028       0  0.4811886 -0.1822684  0.55688879
#> 83     S08     I09  0.841050376       0  0.4811886  0.1812328  0.17862895
#> 97     S08     I10  0.853546501       0  0.4811886  1.2116043 -0.83924640
#> 3      S09     I01 -1.036473841       0  1.3491860 -2.2805280 -0.10513183
#> 13     S09     I02  0.552036698       0  1.3491860 -1.0255685  0.22841928
#> 26     S09     I03  0.790027261       0  1.3491860 -0.5102049 -0.04895378
#> 32     S09     I04  1.264072873       0  1.3491860 -0.3906218  0.30550868
#> 42     S09     I05  0.547848354       0  1.3491860  0.2110268 -1.01236444
#> 55     S09     I06  1.399430774       0  1.3491860 -0.1854311  0.23567596
#> 65     S09     I07  0.305820358       0  1.3491860 -2.1994373  1.15607166
#> 71     S09     I08  0.619861481       0  1.3491860 -0.1822684 -0.54705609
#> 84     S09     I09  1.983678167       0  1.3491860  0.1812328  0.45325941
#> 94     S09     I10  4.128354187       0  1.3491860  1.2116043  1.56756395
#> 10     S10     I01 -2.163925260       0  0.3304327 -2.2805280 -0.21382994
#> 14     S10     I02 -0.521211412       0  0.3304327 -1.0255685  0.17392447
#> 28     S10     I03 -0.675178314       0  0.3304327 -0.5102049 -0.49540605
#> 40     S10     I04 -0.180856589       0  0.3304327 -0.3906218 -0.12066747
#> 43     S10     I05  0.097784937       0  0.3304327  0.2110268 -0.44367455
#> 57     S10     I06  1.341901847       0  0.3304327 -0.1854311  1.19690034
#> 70     S10     I07 -1.388569829       0  0.3304327 -2.1994373  0.48043477
#> 72     S10     I08 -2.191840031       0  0.3304327 -0.1822684 -2.34000430
#> 86     S10     I09  0.913213917       0  0.3304327  0.1812328  0.40154847
#> 100    S10     I10  2.065417563       0  0.3304327  1.2116043  0.52338063