Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, empirical = FALSE, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i       sub_i      item_i         err
#> 1      S01     I01 -3.51781818       0 -0.26485788 -2.00148228 -1.25147802
#> 2      S02     I01 -2.10674270       0  0.85217409 -2.00148228 -0.95743451
#> 3      S03     I01 -2.36292590       0  0.63782166 -2.00148228 -0.99926527
#> 4      S04     I01  0.27206263       0  0.98288772 -2.00148228  1.29065719
#> 5      S05     I01 -1.33337453       0 -0.10820793 -2.00148228  0.77631568
#> 6      S06     I01 -1.69360127       0  0.07279634 -2.00148228  0.23508466
#> 7      S07     I01 -2.23726009       0  0.11345575 -2.00148228 -0.34923356
#> 8      S08     I01 -1.30354028       0  0.53086409 -2.00148228  0.16707791
#> 9      S09     I01 -1.20832310       0  0.30072133 -2.00148228  0.49243785
#> 10     S10     I01 -0.19736846       0  0.05290335 -2.00148228  1.75121047
#> 11     S01     I02 -0.33414972       0 -0.26485788 -0.12780960  0.05851776
#> 12     S02     I02  0.89921687       0  0.85217409 -0.12780960  0.17485238
#> 13     S03     I02  0.82344614       0  0.63782166 -0.12780960  0.31343408
#> 14     S04     I02  1.74879285       0  0.98288772 -0.12780960  0.89371473
#> 15     S05     I02 -0.52119134       0 -0.10820793 -0.12780960 -0.28517381
#> 16     S06     I02 -0.33155069       0  0.07279634 -0.12780960 -0.27653743
#> 17     S07     I02  0.75807876       0  0.11345575 -0.12780960  0.77243261
#> 18     S08     I02  0.03563272       0  0.53086409 -0.12780960 -0.36742177
#> 19     S09     I02 -1.19192805       0  0.30072133 -0.12780960 -1.36483978
#> 20     S10     I02 -1.43919430       0  0.05290335 -0.12780960 -1.36428804
#> 21     S01     I03 -0.56494889       0 -0.26485788 -0.88605543  0.58596442
#> 22     S02     I03 -0.24760018       0  0.85217409 -0.88605543 -0.21371884
#> 23     S03     I03 -0.88717906       0  0.63782166 -0.88605543 -0.63894528
#> 24     S04     I03 -1.95891119       0  0.98288772 -0.88605543 -2.05574348
#> 25     S05     I03 -2.16984049       0 -0.10820793 -0.88605543 -1.17557713
#> 26     S06     I03 -0.09550364       0  0.07279634 -0.88605543  0.71775545
#> 27     S07     I03 -0.04052945       0  0.11345575 -0.88605543  0.73207023
#> 28     S08     I03 -0.21866991       0  0.53086409 -0.88605543  0.13652142
#> 29     S09     I03 -0.76992498       0  0.30072133 -0.88605543 -0.18459087
#> 30     S10     I03 -0.47375154       0  0.05290335 -0.88605543  0.35940054
#> 31     S01     I04 -1.15435239       0 -0.26485788 -1.94043221  1.05093769
#> 32     S02     I04 -1.96280308       0  0.85217409 -1.94043221 -0.87454497
#> 33     S03     I04 -2.97524027       0  0.63782166 -1.94043221 -1.67262972
#> 34     S04     I04 -1.62629469       0  0.98288772 -1.94043221 -0.66875021
#> 35     S05     I04 -3.04089358       0 -0.10820793 -1.94043221 -0.99225344
#> 36     S06     I04 -1.71674406       0  0.07279634 -1.94043221  0.15089180
#> 37     S07     I04 -0.32484064       0  0.11345575 -1.94043221  1.50213582
#> 38     S08     I04 -0.87167850       0  0.53086409 -1.94043221  0.53788962
#> 39     S09     I04 -1.65878659       0  0.30072133 -1.94043221 -0.01907571
#> 40     S10     I04 -2.85625200       0  0.05290335 -1.94043221 -0.96872315
#> 41     S01     I05  0.07751757       0 -0.26485788  0.31039872  0.03197672
#> 42     S02     I05  0.80284189       0  0.85217409  0.31039872 -0.35973093
#> 43     S03     I05  0.77594992       0  0.63782166  0.31039872 -0.17227046
#> 44     S04     I05  2.78554651       0  0.98288772  0.31039872  1.49226007
#> 45     S05     I05 -0.73512182       0 -0.10820793  0.31039872 -0.93731261
#> 46     S06     I05 -0.98857162       0  0.07279634  0.31039872 -1.37176668
#> 47     S07     I05  1.32749140       0  0.11345575  0.31039872  0.90363692
#> 48     S08     I05 -0.02029090       0  0.53086409  0.31039872 -0.86155372
#> 49     S09     I05 -0.78266947       0  0.30072133  0.31039872 -1.39378952
#> 50     S10     I05  0.06866119       0  0.05290335  0.31039872 -0.29464089
#> 51     S01     I06 -0.22736305       0 -0.26485788  0.02236842  0.01512641
#> 52     S02     I06  1.79895749       0  0.85217409  0.02236842  0.92441498
#> 53     S03     I06  1.77163364       0  0.63782166  0.02236842  1.11144356
#> 54     S04     I06  0.61796737       0  0.98288772  0.02236842 -0.38728876
#> 55     S05     I06  0.67280931       0 -0.10820793  0.02236842  0.75864882
#> 56     S06     I06  1.80159527       0  0.07279634  0.02236842  1.70643051
#> 57     S07     I06  0.83308172       0  0.11345575  0.02236842  0.69725756
#> 58     S08     I06  1.73760601       0  0.53086409  0.02236842  1.18437350
#> 59     S09     I06 -0.82560122       0  0.30072133  0.02236842 -1.14869097
#> 60     S10     I06 -0.42129321       0  0.05290335  0.02236842 -0.49656498
#> 61     S01     I07 -1.77975307       0 -0.26485788 -0.28120729 -1.23368791
#> 62     S02     I07  2.37469030       0  0.85217409 -0.28120729  1.80372349
#> 63     S03     I07  0.31882933       0  0.63782166 -0.28120729 -0.03778504
#> 64     S04     I07  2.19987828       0  0.98288772 -0.28120729  1.49819784
#> 65     S05     I07  0.07153688       0 -0.10820793 -0.28120729  0.46095210
#> 66     S06     I07 -0.04136671       0  0.07279634 -0.28120729  0.16704423
#> 67     S07     I07 -1.16113904       0  0.11345575 -0.28120729 -0.99338750
#> 68     S08     I07 -0.89573365       0  0.53086409 -0.28120729 -1.14539045
#> 69     S09     I07  0.85139065       0  0.30072133 -0.28120729  0.83187661
#> 70     S10     I07  1.24431803       0  0.05290335 -0.28120729  1.47262196
#> 71     S01     I08  0.74771478       0 -0.26485788  0.84005243  0.17252022
#> 72     S02     I08  3.15687522       0  0.85217409  0.84005243  1.46464870
#> 73     S03     I08  3.35064742       0  0.63782166  0.84005243  1.87277333
#> 74     S04     I08  2.94743364       0  0.98288772  0.84005243  1.12449348
#> 75     S05     I08  1.01801119       0 -0.10820793  0.84005243  0.28616669
#> 76     S06     I08  0.72884193       0  0.07279634  0.84005243 -0.18400685
#> 77     S07     I08  2.46816660       0  0.11345575  0.84005243  1.51465841
#> 78     S08     I08  1.14977631       0  0.53086409  0.84005243 -0.22114022
#> 79     S09     I08  4.13784233       0  0.30072133  0.84005243  2.99706857
#> 80     S10     I08  0.84017765       0  0.05290335  0.84005243 -0.05277814
#> 81     S01     I09 -0.31018182       0 -0.26485788 -0.44219086  0.39686692
#> 82     S02     I09 -0.74499044       0  0.85217409 -0.44219086 -1.15497367
#> 83     S03     I09 -0.02270472       0  0.63782166 -0.44219086 -0.21833552
#> 84     S04     I09 -1.67644100       0  0.98288772 -0.44219086 -2.21713786
#> 85     S05     I09  1.15437413       0 -0.10820793 -0.44219086  1.70477292
#> 86     S06     I09  1.56759097       0  0.07279634 -0.44219086  1.93698549
#> 87     S07     I09 -1.23222960       0  0.11345575 -0.44219086 -0.90349449
#> 88     S08     I09 -2.12706105       0  0.53086409 -0.44219086 -2.21573428
#> 89     S09     I09  1.38634433       0  0.30072133 -0.44219086  1.52781386
#> 90     S10     I09 -0.17809258       0  0.05290335 -0.44219086  0.21119493
#> 91     S01     I10  0.68664777       0 -0.26485788  1.44083376 -0.48932812
#> 92     S02     I10  1.94872968       0  0.85217409  1.44083376 -0.34427817
#> 93     S03     I10  2.27325873       0  0.63782166  1.44083376  0.19460331
#> 94     S04     I10  2.17519026       0  0.98288772  1.44083376 -0.24853122
#> 95     S05     I10  0.61312137       0 -0.10820793  1.44083376 -0.71950446
#> 96     S06     I10  0.08458673       0  0.07279634  1.44083376 -1.42904338
#> 97     S07     I10  3.52490711       0  0.11345575  1.44083376  1.97061760
#> 98     S08     I10  2.02810815       0  0.53086409  1.44083376  0.05641029
#> 99     S09     I10  0.50964444       0  0.30072133  1.44083376 -1.23191065
#> 100    S10     I10  1.28585071       0  0.05290335  1.44083376 -0.20788640