Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id         val grand_i      sub_i     item_i         err
#> 1      S01     I01 -1.78387199       0 -0.6827736 -0.2909904 -0.81010800
#> 2      S02     I01  0.14875682       0 -0.6445324 -0.2909904  1.08427964
#> 3      S03     I01 -0.22433187       0  0.9972362 -0.2909904 -0.93057767
#> 4      S04     I01  0.31646067       0  0.7755156 -0.2909904 -0.16806450
#> 5      S05     I01  0.37728298       0  0.1623586 -0.2909904  0.50591478
#> 6      S06     I01  2.34714872       0  1.5194632 -0.2909904  1.11867598
#> 7      S07     I01 -2.18814639       0 -0.5331349 -0.2909904 -1.36402105
#> 8      S08     I01 -0.38294716       0 -1.0273115 -0.2909904  0.93535477
#> 9      S09     I01 -2.54923935       0 -0.3219598 -0.2909904 -1.93628909
#> 10     S10     I01 -0.35170116       0 -0.3599592 -0.2909904  0.29924845
#> 11     S01     I02 -0.29113570       0 -0.6827736 -0.7136342  1.10527208
#> 12     S02     I02 -0.19930299       0 -0.6445324 -0.7136342  1.15886361
#> 13     S03     I02 -1.18725684       0  0.9972362 -0.7136342 -1.47085885
#> 14     S04     I02  0.62050914       0  0.7755156 -0.7136342  0.55862776
#> 15     S05     I02 -1.67801345       0  0.1623586 -0.7136342 -1.12673785
#> 16     S06     I02  2.48448587       0  1.5194632 -0.7136342  1.67865693
#> 17     S07     I02 -0.86345033       0 -0.5331349 -0.7136342  0.38331880
#> 18     S08     I02 -1.64093807       0 -1.0273115 -0.7136342  0.10000766
#> 19     S09     I02  0.75019854       0 -0.3219598 -0.7136342  1.78579259
#> 20     S10     I02 -1.42721699       0 -0.3599592 -0.7136342 -0.35362358
#> 21     S01     I03 -0.65925804       0 -0.6827736 -0.7989801  0.82249562
#> 22     S02     I03 -1.18750921       0 -0.6445324 -0.7989801  0.25600328
#> 23     S03     I03  0.86564643       0  0.9972362 -0.7989801  0.66739030
#> 24     S04     I03  0.13284868       0  0.7755156 -0.7989801  0.15631319
#> 25     S05     I03 -1.05549070       0  0.1623586 -0.7989801 -0.41886922
#> 26     S06     I03  2.46466407       0  1.5194632 -0.7989801  1.74418101
#> 27     S07     I03 -2.25499315       0 -0.5331349 -0.7989801 -0.92287813
#> 28     S08     I03 -0.76929718       0 -1.0273115 -0.7989801  1.05699443
#> 29     S09     I03 -0.81892962       0 -0.3219598 -0.7989801  0.30201032
#> 30     S10     I03 -3.83532664       0 -0.3599592 -0.7989801 -2.67638735
#> 31     S01     I04 -2.06777416       0 -0.6827736  0.4371359 -1.82213653
#> 32     S02     I04  0.30872641       0 -0.6445324  0.4371359  0.51612286
#> 33     S03     I04  1.63328949       0  0.9972362  0.4371359  0.19891733
#> 34     S04     I04  1.06780009       0  0.7755156  0.4371359 -0.14485144
#> 35     S05     I04  1.48222912       0  0.1623586  0.4371359  0.88273457
#> 36     S06     I04  0.99219222       0  1.5194632  0.4371359 -0.96440687
#> 37     S07     I04  0.79543538       0 -0.5331349  0.4371359  0.89143436
#> 38     S08     I04 -2.13003814       0 -1.0273115  0.4371359 -1.53986257
#> 39     S09     I04  0.64055601       0 -0.3219598  0.4371359  0.52537991
#> 40     S10     I04 -0.44798948       0 -0.3599592  0.4371359 -0.52516623
#> 41     S01     I05  0.09355736       0 -0.6827736  0.9583824 -0.18205147
#> 42     S02     I05  0.67990682       0 -0.6445324  0.9583824  0.36605682
#> 43     S03     I05  0.86466381       0  0.9972362  0.9583824 -1.09095481
#> 44     S04     I05 -0.04642096       0  0.7755156  0.9583824 -1.78031894
#> 45     S05     I05  1.81282375       0  0.1623586  0.9583824  0.69208274
#> 46     S06     I05  3.44990224       0  1.5194632  0.9583824  0.97205670
#> 47     S07     I05 -0.28983311       0 -0.5331349  0.9583824 -0.71508058
#> 48     S08     I05 -1.45634884       0 -1.0273115  0.9583824 -1.38741972
#> 49     S09     I05 -0.27392104       0 -0.3219598  0.9583824 -0.91034359
#> 50     S10     I05  2.32377025       0 -0.3599592  0.9583824  1.72534705
#> 51     S01     I06 -1.82395420       0 -0.6827736 -0.3945358 -0.74664480
#> 52     S02     I06 -1.48243869       0 -0.6445324 -0.3945358 -0.44337047
#> 53     S03     I06  0.06309513       0  0.9972362 -0.3945358 -0.53960526
#> 54     S04     I06  1.98359151       0  0.7755156 -0.3945358  1.60261176
#> 55     S05     I06  1.06783094       0  0.1623586 -0.3945358  1.30000816
#> 56     S06     I06  0.66370080       0  1.5194632 -0.3945358 -0.46122652
#> 57     S07     I06  0.62183784       0 -0.5331349 -0.3945358  1.54950859
#> 58     S08     I06 -2.27233780       0 -1.0273115 -0.3945358 -0.85049046
#> 59     S09     I06 -2.00829887       0 -0.3219598 -0.3945358 -1.29180320
#> 60     S10     I06  0.60802870       0 -0.3599592 -0.3945358  1.36252372
#> 61     S01     I07 -0.51400054       0 -0.6827736  0.7519394 -0.58316641
#> 62     S02     I07 -0.41504025       0 -0.6445324  0.7519394 -0.52244729
#> 63     S03     I07  1.53236461       0  0.9972362  0.7519394 -0.21681105
#> 64     S04     I07  2.13810541       0  0.7755156  0.7519394  0.61065039
#> 65     S05     I07 -0.01274531       0  0.1623586  0.7519394 -0.92704336
#> 66     S06     I07  1.88753250       0  1.5194632  0.7519394 -0.38387008
#> 67     S07     I07 -0.60291521       0 -0.5331349  0.7519394 -0.82171973
#> 68     S08     I07  0.70557399       0 -1.0273115  0.7519394  0.98094607
#> 69     S09     I07  1.59088295       0 -0.3219598  0.7519394  1.16090335
#> 70     S10     I07  3.78781115       0 -0.3599592  0.7519394  3.39583090
#> 71     S01     I08 -0.48891821       0 -0.6827736  0.4468574 -0.25300201
#> 72     S02     I08 -0.21832301       0 -0.6445324  0.4468574 -0.02064799
#> 73     S03     I08  1.02428486       0  0.9972362  0.4468574 -0.41980873
#> 74     S04     I08  1.87205360       0  0.7755156  0.4468574  0.64968064
#> 75     S05     I08  1.26248679       0  0.1623586  0.4468574  0.65327081
#> 76     S06     I08  1.46696946       0  1.5194632  0.4468574 -0.49935106
#> 77     S07     I08  0.84191886       0 -0.5331349  0.4468574  0.92819641
#> 78     S08     I08 -0.22370914       0 -1.0273115  0.4468574  0.35674500
#> 79     S09     I08  2.73761579       0 -0.3219598  0.4468574  2.61271826
#> 80     S10     I08  1.41308294       0 -0.3599592  0.4468574  1.32618476
#> 81     S01     I09 -0.19849395       0 -0.6827736  0.3519087  0.13237093
#> 82     S02     I09 -0.21249457       0 -0.6445324  0.3519087  0.08012914
#> 83     S03     I09  1.60370577       0  0.9972362  0.3519087  0.25456087
#> 84     S04     I09  0.68927049       0  0.7755156  0.3519087 -0.43815378
#> 85     S05     I09  0.64782422       0  0.1623586  0.3519087  0.13355692
#> 86     S06     I09  1.36064667       0  1.5194632  0.3519087 -0.51072517
#> 87     S07     I09 -1.50325996       0 -0.5331349  0.3519087 -1.32203372
#> 88     S08     I09 -0.21573033       0 -1.0273115  0.3519087  0.45967250
#> 89     S09     I09  0.68492930       0 -0.3219598  0.3519087  0.65498046
#> 90     S10     I09  1.06461300       0 -0.3599592  0.3519087  1.07266351
#> 91     S01     I10 -1.09483203       0 -0.6827736 -0.1982794 -0.21377910
#> 92     S02     I10  0.59158042       0 -0.6445324 -0.1982794  1.43439217
#> 93     S03     I10  1.62811169       0  0.9972362 -0.1982794  0.82915483
#> 94     S04     I10  0.85223011       0  0.7755156 -0.1982794  0.27499388
#> 95     S05     I10  1.14224777       0  0.1623586 -0.1982794  1.17816852
#> 96     S06     I10  2.11594976       0  1.5194632 -0.1982794  0.79476597
#> 97     S07     I10 -0.27927018       0 -0.5331349 -0.1982794  0.45214411
#> 98     S08     I10 -1.76724142       0 -1.0273115 -0.1982794 -0.54165055
#> 99     S09     I10  0.55813539       0 -0.3219598 -0.1982794  1.07837459
#> 100    S10     I10 -0.49699895       0 -0.3599592 -0.1982794  0.06123960