Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i        sub_i     item_i         err
#> 1      S01     I01  2.00849569       0  0.538287644  0.5661166  0.90409141
#> 15     S01     I02  2.41847385       0  0.538287644  0.3057251  1.57446107
#> 29     S01     I03  3.25733642       0  0.538287644  1.1490500  1.56999883
#> 31     S01     I04  0.43761453       0  0.538287644  0.7972705 -0.89794366
#> 44     S01     I05  1.20164090       0  0.538287644  1.0363797 -0.37302645
#> 58     S01     I06  3.29811935       0  0.538287644  1.7278340  1.03199772
#> 61     S01     I07 -1.46705174       0  0.538287644 -0.9946508 -1.01068862
#> 73     S01     I08  2.29966793       0  0.538287644  1.3858512  0.37552912
#> 87     S01     I09  1.54856064       0  0.538287644 -0.0108732  1.02114619
#> 91     S01     I10 -0.18238628       0  0.538287644 -0.9582099  0.23753601
#> 7      S02     I01 -0.17764942       0 -0.556627723  0.5661166 -0.18713833
#> 17     S02     I02 -1.00167083       0 -0.556627723  0.3057251 -0.75076824
#> 30     S02     I03  0.79666924       0 -0.556627723  1.1490500  0.20424701
#> 36     S02     I04  0.85160164       0 -0.556627723  0.7972705  0.61095882
#> 46     S02     I05  0.39400998       0 -0.556627723  1.0363797 -0.08574201
#> 59     S02     I06  0.87007060       0 -0.556627723  1.7278340 -0.30113566
#> 69     S02     I07 -2.90097480       0 -0.556627723 -0.9946508 -1.34969631
#> 75     S02     I08  0.77376779       0 -0.556627723  1.3858512 -0.05545564
#> 88     S02     I09 -0.93758316       0 -0.556627723 -0.0108732 -0.37008223
#> 98     S02     I10 -0.67232538       0 -0.556627723 -0.9582099  0.84251228
#> 4      S03     I01  0.64581771       0 -0.107295007  0.5661166  0.18699608
#> 18     S03     I02 -0.34762715       0 -0.107295007  0.3057251 -0.54605728
#> 23     S03     I03 -0.81950663       0 -0.107295007  1.1490500 -1.86126158
#> 33     S03     I04  0.98803362       0 -0.107295007  0.7972705  0.29805808
#> 47     S03     I05  1.59656802       0 -0.107295007  1.0363797  0.66748331
#> 60     S03     I06  2.13663354       0 -0.107295007  1.7278340  0.51609456
#> 62     S03     I07 -0.40977070       0 -0.107295007 -0.9946508  0.69217507
#> 76     S03     I08  3.09857229       0 -0.107295007  1.3858512  1.82001614
#> 89     S03     I09  1.45303865       0 -0.107295007 -0.0108732  1.57120686
#> 95     S03     I10  0.03042469       0 -0.107295007 -0.9582099  1.09592963
#> 5      S04     I01 -0.43193427       0  0.258247569  0.5661166 -1.25629847
#> 19     S04     I02 -0.10926279       0  0.258247569  0.3057251 -0.67323549
#> 21     S04     I03  1.82114512       0  0.258247569  1.1490500  0.41384760
#> 34     S04     I04  0.76784940       0  0.258247569  0.7972705 -0.28766871
#> 48     S04     I05  1.45066490       0  0.258247569  1.0363797  0.15603762
#> 51     S04     I06  0.94488272       0  0.258247569  1.7278340 -1.04119884
#> 63     S04     I07 -1.74485208       0  0.258247569 -0.9946508 -1.00844889
#> 77     S04     I08  1.53251898       0  0.258247569  1.3858512 -0.11157975
#> 90     S04     I09  1.77861287       0  0.258247569 -0.0108732  1.53123850
#> 92     S04     I10 -1.27263517       0  0.258247569 -0.9582099 -0.57267280
#> 2      S05     I01 -1.26873632       0  0.003760049  0.5661166 -1.83861300
#> 16     S05     I02 -0.52492077       0  0.003760049  0.3057251 -0.83440595
#> 22     S05     I03  2.66725456       0  0.003760049  1.1490500  1.51444456
#> 35     S05     I04  1.03752483       0  0.003760049  0.7972705  0.23649423
#> 45     S05     I05  0.08587136       0  0.003760049  1.0363797 -0.95426840
#> 52     S05     I06  2.06230773       0  0.003760049  1.7278340  0.33071370
#> 64     S05     I07 -2.57075980       0  0.003760049 -0.9946508 -1.57986909
#> 74     S05     I08  2.11480381       0  0.003760049  1.3858512  0.72519260
#> 81     S05     I09  0.73330053       0  0.003760049 -0.0108732  0.74041368
#> 93     S05     I10 -1.56873477       0  0.003760049 -0.9582099 -0.61428489
#> 8      S06     I01  0.83930100       0 -0.757413884  0.5661166  1.03059825
#> 20     S06     I02  0.16986855       0 -0.757413884  0.3057251  0.62155730
#> 27     S06     I03  0.12940608       0 -0.757413884  1.1490500 -0.26222999
#> 37     S06     I04  0.72790193       0 -0.757413884  0.7972705  0.68804527
#> 49     S06     I05  0.56999392       0 -0.757413884  1.0363797  0.29102809
#> 56     S06     I06  1.39123005       0 -0.757413884  1.7278340  0.42080995
#> 66     S06     I07 -2.00826468       0 -0.757413884 -0.9946508 -0.25620003
#> 79     S06     I08  0.25527893       0 -0.757413884  1.3858512 -0.37315835
#> 85     S06     I09 -0.72762270       0 -0.757413884 -0.0108732  0.04066438
#> 99     S06     I10 -3.39234658       0 -0.757413884 -0.9582099 -1.67672276
#> 9      S07     I01  0.95185049       0  0.945412596  0.5661166 -0.55967874
#> 11     S07     I02  0.80919225       0  0.945412596  0.3057251 -0.44194548
#> 24     S07     I03  1.69727131       0  0.945412596  1.1490500 -0.39719124
#> 38     S07     I04  1.56893653       0  0.945412596  0.7972705 -0.17374661
#> 50     S07     I05  1.48465840       0  0.945412596  1.0363797 -0.49713391
#> 53     S07     I06  2.74925659       0  0.945412596  1.7278340  0.07601001
#> 67     S07     I07  0.38423311       0  0.945412596 -0.9946508  0.43347128
#> 80     S07     I08  3.62831055       0  0.945412596  1.3858512  1.29704680
#> 82     S07     I09  0.42230397       0  0.945412596 -0.0108732 -0.51223543
#> 96     S07     I10 -2.20286185       0  0.945412596 -0.9582099 -2.19006451
#> 6      S08     I01  2.12828558       0  0.954681184  0.5661166  0.60748776
#> 12     S08     I02  1.64037025       0  0.954681184  0.3057251  0.37996393
#> 25     S08     I03  4.58057942       0  0.954681184  1.1490500  2.47684829
#> 39     S08     I04  0.78378599       0  0.954681184  0.7972705 -0.96816573
#> 41     S08     I05  3.14198180       0  0.954681184  1.0363797  1.15092091
#> 54     S08     I06  1.30928401       0  0.954681184  1.7278340 -1.37323116
#> 68     S08     I07 -0.37924696       0  0.954681184 -0.9946508 -0.33927738
#> 78     S08     I08  2.20106100       0  0.954681184  1.3858512 -0.13947135
#> 83     S08     I09 -0.11894296       0  0.954681184 -0.0108732 -1.06275095
#> 97     S08     I10  0.41613109       0  0.954681184 -0.9582099  0.41965984
#> 3      S09     I01 -0.58564705       0  0.135823699  0.5661166 -1.28758738
#> 13     S09     I02 -1.89437371       0  0.135823699  0.3057251 -2.33592254
#> 26     S09     I03  0.97134062       0  0.135823699  1.1490500 -0.31353303
#> 32     S09     I04 -0.77319395       0  0.135823699  0.7972705 -1.70628819
#> 42     S09     I05  0.48504839       0  0.135823699  1.0363797 -0.68715502
#> 55     S09     I06  0.54608607       0  0.135823699  1.7278340 -1.31757162
#> 65     S09     I07 -2.03695958       0  0.135823699 -0.9946508 -1.17813251
#> 71     S09     I08  1.33882327       0  0.135823699  1.3858512 -0.18285159
#> 84     S09     I09 -0.84919366       0  0.135823699 -0.0108732 -0.97414416
#> 94     S09     I10  0.09979523       0  0.135823699 -0.9582099  0.92218146
#> 10     S10     I01 -0.97499080       0 -0.513335863  0.5661166 -1.02777157
#> 14     S10     I02 -0.09712223       0 -0.513335863  0.3057251  0.11048850
#> 28     S10     I03  1.79979239       0 -0.513335863  1.1490500  1.16407830
#> 40     S10     I04 -1.24328982       0 -0.513335863  0.7972705 -1.52722450
#> 43     S10     I05  1.03815597       0 -0.513335863  1.0363797  0.51511213
#> 57     S10     I06  3.87793168       0 -0.513335863  1.7278340  2.66343356
#> 70     S10     I07 -0.04929625       0 -0.513335863 -0.9946508  1.45869038
#> 72     S10     I08  1.77733639       0 -0.513335863  1.3858512  0.90482110
#> 86     S10     I09 -0.61188109       0 -0.513335863 -0.0108732 -0.08767202
#> 100    S10     I10 -1.54139462       0 -0.513335863 -0.9582099 -0.06984882