Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items

sim_mixed_cc(sub_n = 100, item_n = 20, grand_i = 0, sub_sd = 1,
  item_sd = 1, error_sd = 1, empirical = FALSE, seed = NULL)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts

item_sd
the SD of item random intercepts

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id            y grand_i       sub_i     item_i         err
#> 1      S01     I01  0.478964211       0  1.30167314 -0.2654794 -0.55722951
#> 2      S02     I01 -1.135522197       0 -0.10310205 -0.2654794 -0.76694073
#> 3      S03     I01  0.708353847       0  1.73990989 -0.2654794 -0.76607662
#> 4      S04     I01 -0.955031919       0 -0.72224945 -0.2654794  0.03269694
#> 5      S05     I01  2.603779830       0  1.79007992 -0.2654794  1.07917933
#> 6      S06     I01  0.826596635       0 -0.08721526 -0.2654794  1.17929131
#> 7      S07     I01  0.232782751       0  0.25052609 -0.2654794  0.24773608
#> 8      S08     I01 -0.755984048       0 -1.36445746 -0.2654794  0.87395283
#> 9      S09     I01  0.520546262       0  0.82052309 -0.2654794 -0.03449742
#> 10     S10     I01 -3.018058130       0 -1.73864789 -0.2654794 -1.01393082
#> 11     S01     I02  0.676063311       0  1.30167314 -0.5834017 -0.04220810
#> 12     S02     I02 -0.982679976       0 -0.10310205 -0.5834017 -0.29617620
#> 13     S03     I02  2.905648735       0  1.73990989 -0.5834017  1.74914058
#> 14     S04     I02 -1.805213215       0 -0.72224945 -0.5834017 -0.49956204
#> 15     S05     I02  0.674813603       0  1.79007992 -0.5834017 -0.53186458
#> 16     S06     I02  0.857082262       0 -0.08721526 -0.5834017  1.52769925
#> 17     S07     I02 -0.643272283       0  0.25052609 -0.5834017 -0.31039664
#> 18     S08     I02 -1.817410619       0 -1.36445746 -0.5834017  0.13044857
#> 19     S09     I02 -0.671830444       0  0.82052309 -0.5834017 -0.90895181
#> 20     S10     I02  0.125584712       0 -1.73864789 -0.5834017  2.44763433
#> 21     S01     I03  0.441384001       0  1.30167314 -0.2269004 -0.63338878
#> 22     S02     I03  0.315746321       0 -0.10310205 -0.2269004  0.64574872
#> 23     S03     I03  0.190231802       0  1.73990989 -0.2269004 -1.32277773
#> 24     S04     I03 -0.345197136       0 -0.72224945 -0.2269004  0.60395267
#> 25     S05     I03  1.856416487       0  1.79007992 -0.2269004  0.29323692
#> 26     S06     I03 -2.180759944       0 -0.08721526 -0.2269004 -1.86664433
#> 27     S07     I03  1.697725633       0  0.25052609 -0.2269004  1.67409990
#> 28     S08     I03 -2.985944329       0 -1.36445746 -0.2269004 -1.39458652
#> 29     S09     I03  1.519043963       0  0.82052309 -0.2269004  0.92542122
#> 30     S10     I03 -3.187603168       0 -1.73864789 -0.2269004 -1.22205492
#> 31     S01     I04  0.534577223       0  1.30167314  0.7651259 -1.53222181
#> 32     S02     I04  0.678828944       0 -0.10310205  0.7651259  0.01680510
#> 33     S03     I04  4.030331996       0  1.73990989  0.7651259  1.52529621
#> 34     S04     I04  0.637645262       0 -0.72224945  0.7651259  0.59476881
#> 35     S05     I04  1.373145866       0  1.79007992  0.7651259 -1.18205995
#> 36     S06     I04  0.771504824       0 -0.08721526  0.7651259  0.09359419
#> 37     S07     I04 -0.501880040       0  0.25052609  0.7651259 -1.51753202
#> 38     S08     I04 -2.353282246       0 -1.36445746  0.7651259 -1.75395068
#> 39     S09     I04  1.980729373       0  0.82052309  0.7651259  0.39508038
#> 40     S10     I04 -0.802171646       0 -1.73864789  0.7651259  0.17135035
#> 41     S01     I05  1.694776756       0  1.30167314  0.2373136  0.15579006
#> 42     S02     I05 -0.242535435       0 -0.10310205  0.2373136 -0.37674694
#> 43     S03     I05  2.265938393       0  1.73990989  0.2373136  0.28871495
#> 44     S04     I05 -0.463397508       0 -0.72224945  0.2373136  0.02153838
#> 45     S05     I05  2.542412421       0  1.79007992  0.2373136  0.51501895
#> 46     S06     I05  1.011571057       0 -0.08721526  0.2373136  0.86147276
#> 47     S07     I05 -0.307004732       0  0.25052609  0.2373136 -0.79484437
#> 48     S08     I05 -0.334930122       0 -1.36445746  0.2373136  0.79221378
#> 49     S09     I05  0.359039978       0  0.82052309  0.2373136 -0.69879667
#> 50     S10     I05 -2.241448753       0 -1.73864789  0.2373136 -0.74011442
#> 51     S01     I06  0.713914529       0  1.30167314  0.4523127 -1.04007133
#> 52     S02     I06  0.063093454       0 -0.10310205  0.4523127 -0.28611722
#> 53     S03     I06  3.635444110       0  1.73990989  0.4523127  1.44322150
#> 54     S04     I06 -0.885506209       0 -0.72224945  0.4523127 -0.61556949
#> 55     S05     I06  1.821660004       0  1.79007992  0.4523127 -0.42073264
#> 56     S06     I06  1.371488541       0 -0.08721526  0.4523127  1.00639107
#> 57     S07     I06  0.140318336       0  0.25052609  0.4523127 -0.56252047
#> 58     S08     I06 -1.422210657       0 -1.36445746  0.4523127 -0.51006592
#> 59     S09     I06  0.939526501       0  0.82052309  0.4523127 -0.33330932
#> 60     S10     I06 -1.449041066       0 -1.73864789  0.4523127 -0.16270590
#> 61     S01     I07  0.489629921       0  1.30167314 -1.2409111  0.42886787
#> 62     S02     I07 -0.570866374       0 -0.10310205 -1.2409111  0.77314676
#> 63     S03     I07 -1.564583243       0  1.73990989 -1.2409111 -2.06358204
#> 64     S04     I07 -1.519329748       0 -0.72224945 -1.2409111  0.44383079
#> 65     S05     I07  0.160164208       0  1.79007992 -1.2409111 -0.38900462
#> 66     S06     I07 -0.512746243       0 -0.08721526 -1.2409111  0.81538010
#> 67     S07     I07 -2.162615409       0  0.25052609 -1.2409111 -1.17223041
#> 68     S08     I07 -3.395309438       0 -1.36445746 -1.2409111 -0.78994089
#> 69     S09     I07 -0.844123966       0  0.82052309 -1.2409111 -0.42373597
#> 70     S10     I07 -2.781169046       0 -1.73864789 -1.2409111  0.19838994
#> 71     S01     I08  2.705090360       0  1.30167314  1.5030017 -0.09958452
#> 72     S02     I08 -0.015186750       0 -0.10310205  1.5030017 -1.41508644
#> 73     S03     I08  2.390246665       0  1.73990989  1.5030017 -0.85266496
#> 74     S04     I08 -0.445652805       0 -0.72224945  1.5030017 -1.22640510
#> 75     S05     I08  2.809023559       0  1.79007992  1.5030017 -0.48405810
#> 76     S06     I08  2.351900072       0 -0.08721526  1.5030017  0.93611359
#> 77     S07     I08  1.190310366       0  0.25052609  1.5030017 -0.56321746
#> 78     S08     I08 -0.034623686       0 -1.36445746  1.5030017 -0.17316797
#> 79     S09     I08  2.397444081       0  0.82052309  1.5030017  0.07391925
#> 80     S10     I08 -0.503974160       0 -1.73864789  1.5030017 -0.26832801
#> 81     S01     I09 -0.135828903       0  1.30167314  0.1555071 -1.59300914
#> 82     S02     I09  0.648942573       0 -0.10310205  0.1555071  0.59653752
#> 83     S03     I09  1.797562919       0  1.73990989  0.1555071 -0.09785407
#> 84     S04     I09  0.513296325       0 -0.72224945  0.1555071  1.08003867
#> 85     S05     I09  0.492075429       0  1.79007992  0.1555071 -1.45351159
#> 86     S06     I09  1.149049339       0 -0.08721526  0.1555071  1.08075750
#> 87     S07     I09  0.364342554       0  0.25052609  0.1555071 -0.04169063
#> 88     S08     I09 -0.418674908       0 -1.36445746  0.1555071  0.79027545
#> 89     S09     I09  1.762971036       0  0.82052309  0.1555071  0.78694084
#> 90     S10     I09 -0.839953964       0 -1.73864789  0.1555071  0.74318683
#> 91     S01     I10 -0.527112217       0  1.30167314 -1.1919374 -0.63684796
#> 92     S02     I10 -2.612319833       0 -0.10310205 -1.1919374 -1.31728039
#> 93     S03     I10  0.012758637       0  1.73990989 -1.1919374 -0.53521386
#> 94     S04     I10 -0.540893869       0 -0.72224945 -1.1919374  1.37329297
#> 95     S05     I10  0.004583627       0  1.79007992 -1.1919374 -0.59355890
#> 96     S06     I10  0.478979529       0 -0.08721526 -1.1919374  1.75813218
#> 97     S07     I10 -1.324937912       0  0.25052609 -1.1919374 -0.38352661
#> 98     S08     I10 -3.646305182       0 -1.36445746 -1.1919374 -1.08991033
#> 99     S09     I10  0.813531580       0  0.82052309 -1.1919374  1.18494587
#> 100    S10     I10 -3.721908912       0 -1.73864789 -1.1919374 -0.79132363