Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

Usage

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n

the number of subjects

item_n

the number of items

grand_i

the grand intercept (overall mean)

sub_sd

the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd

the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd

the SD of the error term

empirical

Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)
#>     sub_id item_id           y grand_i      sub_i      item_i         err
#> 1      S01     I01  0.72312865       0 -1.0691224  1.76467545  0.02757556
#> 15     S01     I02  0.03643503       0 -1.0691224 -0.40483200  1.51038940
#> 29     S01     I03 -0.16533889       0 -1.0691224  1.03760287 -0.13381940
#> 31     S01     I04  0.34207882       0 -1.0691224  0.77432266  0.63687852
#> 44     S01     I05 -3.53651960       0 -1.0691224 -1.35739022 -1.11000701
#> 58     S01     I06 -1.91699695       0 -1.0691224 -1.06589893  0.21802434
#> 61     S01     I07 -2.81809970       0 -1.0691224 -1.10327258 -0.64570475
#> 73     S01     I08 -1.06578978       0 -1.0691224 -0.05367636  0.05700895
#> 87     S01     I09 -1.34003986       0 -1.0691224  0.70935652 -0.98027402
#> 91     S01     I10  2.86815569       0 -1.0691224  1.24100151  2.69627654
#> 7      S02     I01  1.31316845       0  0.1469474  1.76467545 -0.59845438
#> 17     S02     I02 -0.11737393       0  0.1469474 -0.40483200  0.14051069
#> 30     S02     I03  0.41995112       0  0.1469474  1.03760287 -0.76459914
#> 36     S02     I04 -0.94897519       0  0.1469474  0.77432266 -1.87024523
#> 46     S02     I05  0.02482756       0  0.1469474 -1.35739022  1.23527041
#> 59     S02     I06 -1.13768053       0  0.1469474 -1.06589893 -0.21872899
#> 69     S02     I07 -0.19702652       0  0.1469474 -1.10327258  0.75929868
#> 75     S02     I08  0.45598473       0  0.1469474 -0.05367636  0.36271371
#> 88     S02     I09  0.23190857       0  0.1469474  0.70935652 -0.62439533
#> 98     S02     I10 -0.44361044       0  0.1469474  1.24100151 -1.83155934
#> 4      S03     I01  1.17682815       0 -0.7217386  1.76467545  0.13389125
#> 18     S03     I02  0.29457382       0 -0.7217386 -0.40483200  1.42114437
#> 23     S03     I03 -0.21813660       0 -0.7217386  1.03760287 -0.53400092
#> 33     S03     I04  0.48047456       0 -0.7217386  0.77432266  0.42789045
#> 47     S03     I05 -0.95254601       0 -0.7217386 -1.35739022  1.12658277
#> 60     S03     I06 -1.03395667       0 -0.7217386 -1.06589893  0.75368081
#> 62     S03     I07 -1.56996505       0 -0.7217386 -1.10327258  0.25504609
#> 76     S03     I08 -0.03491919       0 -0.7217386 -0.05367636  0.74049572
#> 89     S03     I09  0.05660278       0 -0.7217386  0.70935652  0.06898481
#> 95     S03     I10  1.12708541       0 -0.7217386  1.24100151  0.60782245
#> 5      S04     I01  2.80574053       0 -0.7257852  1.76467545  1.76685024
#> 19     S04     I02 -1.91386564       0 -0.7257852 -0.40483200 -0.78324848
#> 21     S04     I03 -0.20409699       0 -0.7257852  1.03760287 -0.51591470
#> 34     S04     I04  1.18026609       0 -0.7257852  0.77432266  1.13172859
#> 48     S04     I05 -1.81288704       0 -0.7257852 -1.35739022  0.27028835
#> 51     S04     I06 -4.08700957       0 -0.7257852 -1.06589893 -2.29532548
#> 63     S04     I07  0.93518834       0 -0.7257852 -1.10327258  2.76424608
#> 77     S04     I08 -0.52849502       0 -0.7257852 -0.05367636  0.25096649
#> 90     S04     I09 -0.91106641       0 -0.7257852  0.70935652 -0.89463777
#> 92     S04     I10  1.44740298       0 -0.7257852  1.24100151  0.93218662
#> 2      S05     I01  0.47842112       0 -0.0229003  1.76467545 -1.26335404
#> 16     S05     I02 -0.63175510       0 -0.0229003 -0.40483200 -0.20402280
#> 22     S05     I03  1.44141065       0 -0.0229003  1.03760287  0.42670807
#> 35     S05     I04  0.28742071       0 -0.0229003  0.77432266 -0.46400166
#> 45     S05     I05 -3.00838806       0 -0.0229003 -1.35739022 -1.62809754
#> 52     S05     I06 -1.56695912       0 -0.0229003 -1.06589893 -0.47815990
#> 64     S05     I07 -1.70427172       0 -0.0229003 -1.10327258 -0.57809884
#> 74     S05     I08 -0.93791945       0 -0.0229003 -0.05367636 -0.86134280
#> 81     S05     I09  2.70403233       0 -0.0229003  0.70935652  2.01757610
#> 93     S05     I10 -0.95810543       0 -0.0229003  1.24100151 -2.17620665
#> 8      S06     I01  0.56360492       0 -0.6740328  1.76467545 -0.52703768
#> 20     S06     I02  0.55387231       0 -0.6740328 -0.40483200  1.63273716
#> 27     S06     I03 -0.75921888       0 -0.6740328  1.03760287 -1.12278891
#> 37     S06     I04 -2.04126076       0 -0.6740328  0.77432266 -2.14155058
#> 49     S06     I05 -0.94522876       0 -0.6740328 -1.35739022  1.08619431
#> 56     S06     I06 -2.38794826       0 -0.6740328 -1.06589893 -0.64801649
#> 66     S06     I07 -1.97038759       0 -0.6740328 -1.10327258 -0.19308216
#> 79     S06     I08 -1.28934959       0 -0.6740328 -0.05367636 -0.56164038
#> 85     S06     I09  0.75127802       0 -0.6740328  0.70935652  0.71595434
#> 99     S06     I10  0.46551376       0 -0.6740328  1.24100151 -0.10145491
#> 9      S07     I01  3.55206184       0  0.9324205  1.76467545  0.85496585
#> 11     S07     I02 -0.41387170       0  0.9324205 -0.40483200 -0.94146023
#> 24     S07     I03  2.03609461       0  0.9324205  1.03760287  0.06607120
#> 38     S07     I04  1.75476690       0  0.9324205  0.77432266  0.04802370
#> 50     S07     I05 -0.48154167       0  0.9324205 -1.35739022 -0.05657198
#> 53     S07     I06 -0.03052695       0  0.9324205 -1.06589893  0.10295145
#> 67     S07     I07  1.11828422       0  0.9324205 -1.10327258  1.28913627
#> 80     S07     I08  1.61749568       0  0.9324205 -0.05367636  0.73875150
#> 82     S07     I09  1.33658953       0  0.9324205  0.70935652 -0.30518753
#> 96     S07     I10  0.94966362       0  0.9324205  1.24100151 -1.22375843
#> 6      S08     I01  2.94132517       0 -0.4575641  1.76467545  1.63421383
#> 12     S08     I02 -1.66360154       0 -0.4575641 -0.40483200 -0.80120542
#> 25     S08     I03  0.79307880       0 -0.4575641  1.03760287  0.21304005
#> 39     S08     I04  0.13074485       0 -0.4575641  0.77432266 -0.18601370
#> 41     S08     I05 -2.19622788       0 -0.4575641 -1.35739022 -0.38127354
#> 54     S08     I06 -2.12474587       0 -0.4575641 -1.06589893 -0.60128282
#> 68     S08     I07 -1.19787675       0 -0.4575641 -1.10327258  0.36295995
#> 78     S08     I08 -1.87221167       0 -0.4575641 -0.05367636 -1.36097119
#> 83     S08     I09  0.74191488       0 -0.4575641  0.70935652  0.49012247
#> 97     S08     I10 -1.77728778       0 -0.4575641  1.24100151 -2.56072518
#> 3      S09     I01  2.59743264       0  1.3696082  1.76467545 -0.53685097
#> 13     S09     I02  1.85074456       0  1.3696082 -0.40483200  0.88596841
#> 26     S09     I03  4.28852091       0  1.3696082  1.03760287  1.88130988
#> 32     S09     I04  1.24756734       0  1.3696082  0.77432266 -0.89636348
#> 42     S09     I05  0.32625774       0  1.3696082 -1.35739022  0.31403981
#> 55     S09     I06  1.77401885       0  1.3696082 -1.06589893  1.47030962
#> 65     S09     I07 -2.11304524       0  1.3696082 -1.10327258 -2.37938081
#> 71     S09     I08  0.89194438       0  1.3696082 -0.05367636 -0.42398741
#> 84     S09     I09  1.06488083       0  1.3696082  0.70935652 -1.01408385
#> 94     S09     I10  2.53429414       0  1.3696082  1.24100151 -0.07631553
#> 10     S10     I01  2.63011527       0 -1.3210337  1.76467545  2.18647349
#> 14     S10     I02 -3.94578435       0 -1.3210337 -0.40483200 -2.21991868
#> 28     S10     I03 -1.66676104       0 -1.3210337  1.03760287 -1.38333024
#> 40     S10     I04 -0.97772268       0 -1.3210337  0.77432266 -0.43101168
#> 43     S10     I05 -1.71842393       0 -1.3210337 -1.35739022  0.95999996
#> 57     S10     I06 -1.03649765       0 -1.3210337 -1.06589893  1.35043495
#> 70     S10     I07 -1.92579137       0 -1.3210337 -1.10327258  0.49851489
#> 72     S10     I08 -0.81073662       0 -1.3210337 -0.05367636  0.56397341
#> 86     S10     I09 -1.64115479       0 -1.3210337  0.70935652 -1.02947765
#> 100    S10     I10 -0.42432897       0 -1.3210337  1.24100151 -0.34429682