Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

Usage

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n

the number of subjects

item_n

the number of items

grand_i

the grand intercept (overall mean)

sub_sd

the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd

the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd

the SD of the error term

empirical

Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)
#>     sub_id item_id            y grand_i       sub_i     item_i         err
#> 1      S01     I01 -1.409399619       0 -0.18902605 -1.0922544 -0.12811919
#> 15     S01     I02 -0.833687787       0 -0.18902605 -0.9874887  0.34282694
#> 29     S01     I03 -0.538930011       0 -0.18902605 -0.5440981  0.19419413
#> 31     S01     I04  1.523286569       0 -0.18902605  0.1692168  1.54309582
#> 44     S01     I05  0.055032990       0 -0.18902605 -0.5968080  0.84086707
#> 58     S01     I06 -0.277471494       0 -0.18902605  0.6353539 -0.72379930
#> 61     S01     I07  0.628653115       0 -0.18902605 -0.2653338  1.08301300
#> 73     S01     I08  1.020638461       0 -0.18902605  1.3963789 -0.18671441
#> 87     S01     I09 -1.331830999       0 -0.18902605 -0.7947550 -0.34804996
#> 91     S01     I10  0.430967192       0 -0.18902605 -1.1095781  1.72957133
#> 7      S02     I01 -3.237938821       0 -0.58729648 -1.0922544 -1.55838796
#> 17     S02     I02 -2.678845798       0 -0.58729648 -0.9874887 -1.10406064
#> 30     S02     I03 -0.954243407       0 -0.58729648 -0.5440981  0.17715116
#> 36     S02     I04 -0.473881065       0 -0.58729648  0.1692168 -0.05580138
#> 46     S02     I05 -0.460827678       0 -0.58729648 -0.5968080  0.72327683
#> 59     S02     I06  0.287956955       0 -0.58729648  0.6353539  0.23989958
#> 69     S02     I07 -1.793350935       0 -0.58729648 -0.2653338 -0.94072062
#> 75     S02     I08  1.108525741       0 -0.58729648  1.3963789  0.29944330
#> 88     S02     I09 -2.287170523       0 -0.58729648 -0.7947550 -0.90511905
#> 98     S02     I10 -2.029212917       0 -0.58729648 -1.1095781 -0.33233836
#> 4      S03     I01 -2.439553199       0 -1.65674757 -1.0922544  0.30944875
#> 18     S03     I02 -3.648070467       0 -1.65674757 -0.9874887 -1.00383422
#> 23     S03     I03 -2.446642235       0 -1.65674757 -0.5440981 -0.24579658
#> 33     S03     I04 -1.616151407       0 -1.65674757  0.1692168 -0.12862063
#> 47     S03     I05 -2.159483337       0 -1.65674757 -0.5968080  0.09407226
#> 60     S03     I06 -1.569371123       0 -1.65674757  0.6353539 -0.54797741
#> 62     S03     I07 -1.459497024       0 -1.65674757 -0.2653338  0.46258438
#> 76     S03     I08 -0.407908704       0 -1.65674757  1.3963789 -0.14754006
#> 89     S03     I09 -2.619350588       0 -1.65674757 -0.7947550 -0.16784803
#> 95     S03     I10 -1.424492344       0 -1.65674757 -1.1095781  1.34183331
#> 5      S04     I01 -1.358569192       0 -1.37707395 -1.0922544  1.11075913
#> 19     S04     I02 -4.278890333       0 -1.37707395 -0.9874887 -1.91432771
#> 21     S04     I03 -2.275216690       0 -1.37707395 -0.5440981 -0.35404466
#> 34     S04     I04 -0.344123960       0 -1.37707395  0.1692168  0.86373319
#> 48     S04     I05 -2.096498177       0 -1.37707395 -0.5968080 -0.12261620
#> 51     S04     I06 -0.371372474       0 -1.37707395  0.6353539  0.37034762
#> 63     S04     I07 -0.952250516       0 -1.37707395 -0.2653338  0.69015727
#> 77     S04     I08 -0.667643465       0 -1.37707395  1.3963789 -0.68694844
#> 90     S04     I09 -2.876732805       0 -1.37707395 -0.7947550 -0.70490387
#> 92     S04     I10 -3.823782415       0 -1.37707395 -1.1095781 -1.33713039
#> 2      S05     I01 -4.535894113       0 -1.80023932 -1.0922544 -1.64340041
#> 16     S05     I02 -3.532914165       0 -1.80023932 -0.9874887 -0.74518617
#> 22     S05     I03 -1.444935680       0 -1.80023932 -0.5440981  0.89940173
#> 35     S05     I04 -0.841633415       0 -1.80023932  0.1692168  0.78938911
#> 45     S05     I05 -3.515759916       0 -1.80023932 -0.5968080 -1.11871256
#> 52     S05     I06 -0.004341175       0 -1.80023932  0.6353539  1.16054429
#> 64     S05     I07 -1.549091327       0 -1.80023932 -0.2653338  0.51648183
#> 74     S05     I08 -1.237744642       0 -1.80023932  1.3963789 -0.83388424
#> 81     S05     I09 -4.442790371       0 -1.80023932 -0.7947550 -1.84779606
#> 93     S05     I10 -4.302704060       0 -1.80023932 -1.1095781 -1.39288666
#> 8      S06     I01 -1.792809705       0  0.06492634 -1.0922544 -0.76548167
#> 20     S06     I02  0.134685563       0  0.06492634 -0.9874887  1.05724790
#> 27     S06     I03  0.113354290       0  0.06492634 -0.5440981  0.59252604
#> 37     S06     I04  0.323720144       0  0.06492634  0.1692168  0.08957701
#> 49     S06     I05  0.193792222       0  0.06492634 -0.5968080  0.72567391
#> 56     S06     I06  1.931335697       0  0.06492634  0.6353539  1.23105550
#> 66     S06     I07 -0.647614603       0  0.06492634 -0.2653338 -0.44720711
#> 79     S06     I08  0.549137960       0  0.06492634  1.3963789 -0.91216731
#> 85     S06     I09 -0.180213641       0  0.06492634 -0.7947550  0.54961501
#> 99     S06     I10 -0.007457697       0  0.06492634 -1.1095781  1.03719404
#> 9      S07     I01  0.484023785       0  0.31641974 -1.0922544  1.25985842
#> 11     S07     I02 -0.480258136       0  0.31641974 -0.9874887  0.19081080
#> 24     S07     I03 -0.365887380       0  0.31641974 -0.5440981 -0.13820903
#> 38     S07     I04 -0.572303391       0  0.31641974  0.1692168 -1.05793993
#> 50     S07     I05 -1.536531748       0  0.31641974 -0.5968080 -1.25614346
#> 53     S07     I06  0.969073776       0  0.31641974  0.6353539  0.01730018
#> 67     S07     I07 -0.536733247       0  0.31641974 -0.2653338 -0.58781915
#> 80     S07     I08  3.786816584       0  0.31641974  1.3963789  2.07401792
#> 82     S07     I09 -1.042831257       0  0.31641974 -0.7947550 -0.56449600
#> 96     S07     I10 -0.257338995       0  0.31641974 -1.1095781  0.53581935
#> 6      S08     I01 -2.748295533       0 -1.40405396 -1.0922544 -0.25198719
#> 12     S08     I02 -1.608419907       0 -1.40405396 -0.9874887  0.78312273
#> 25     S08     I03 -3.544310556       0 -1.40405396 -0.5440981 -1.59615850
#> 39     S08     I04 -3.313283239       0 -1.40405396  0.1692168 -2.07844607
#> 41     S08     I05 -1.938397863       0 -1.40405396 -0.5968080  0.06246413
#> 54     S08     I06 -1.574526957       0 -1.40405396  0.6353539 -0.80582685
#> 68     S08     I07 -2.178673381       0 -1.40405396 -0.2653338 -0.50928558
#> 78     S08     I08  0.419085911       0 -1.40405396  1.3963789  0.42676095
#> 83     S08     I09 -1.939856886       0 -1.40405396 -0.7947550  0.25895207
#> 97     S08     I10 -2.977594851       0 -1.40405396 -1.1095781 -0.46396280
#> 3      S09     I01 -1.579951138       0  0.25868823 -1.0922544 -0.74638499
#> 13     S09     I02 -0.642168067       0  0.25868823 -0.9874887  0.08663238
#> 26     S09     I03  1.098891920       0  0.25868823 -0.5440981  1.38430178
#> 32     S09     I04  2.115082232       0  0.25868823  0.1692168  1.68717720
#> 42     S09     I05 -1.425013908       0  0.25868823 -0.5968080 -1.08689411
#> 55     S09     I06  1.931512871       0  0.25868823  0.6353539  1.03747079
#> 65     S09     I07 -0.286848373       0  0.25868823 -0.2653338 -0.28020277
#> 71     S09     I08  3.060329809       0  0.25868823  1.3963789  1.40526265
#> 84     S09     I09 -1.077320605       0  0.25868823 -0.7947550 -0.54125384
#> 94     S09     I10 -1.603093871       0  0.25868823 -1.1095781 -0.75220402
#> 10     S10     I01 -4.778540415       0 -1.55147462 -1.0922544 -2.13481142
#> 14     S10     I02 -3.744991698       0 -1.55147462 -0.9874887 -1.20602840
#> 28     S10     I03 -2.418470187       0 -1.55147462 -0.5440981 -0.32289748
#> 40     S10     I04 -0.274662420       0 -1.55147462  0.1692168  1.10759540
#> 43     S10     I05 -2.105791881       0 -1.55147462 -0.5968080  0.04249077
#> 57     S10     I06  0.654077499       0 -1.55147462  0.6353539  1.57019827
#> 70     S10     I07 -2.535221606       0 -1.55147462 -0.2653338 -0.71841315
#> 72     S10     I08 -0.713296008       0 -1.55147462  1.3963789 -0.55820031
#> 86     S10     I09 -3.271739260       0 -1.55147462 -0.7947550 -0.92550965
#> 100    S10     I10 -2.473085779       0 -1.55147462 -1.1095781  0.18796692