Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i      sub_i     item_i         err
#> 1      S01     I01 -1.79728652       0 -0.9267769 -1.2508904  0.38038075
#> 15     S01     I02 -0.44400510       0 -0.9267769 -0.1383032  0.62107502
#> 29     S01     I03 -0.54401179       0 -0.9267769 -0.4284905  0.81125564
#> 31     S01     I04 -2.26391450       0 -0.9267769 -0.1582815 -1.17885605
#> 44     S01     I05 -0.37116518       0 -0.9267769  0.3987451  0.15686668
#> 58     S01     I06  0.12399281       0 -0.9267769  0.1638725  0.88689721
#> 61     S01     I07 -1.96794770       0 -0.9267769 -0.4220881 -0.61908265
#> 73     S01     I08 -2.58208632       0 -0.9267769 -1.5426317 -0.11267773
#> 87     S01     I09 -3.27000406       0 -0.9267769 -0.8490586 -1.49416851
#> 91     S01     I10 -0.94594327       0 -0.9267769 -0.6288139  0.60964759
#> 7      S02     I01  0.38843085       0  0.6804446 -1.2508904  0.95887657
#> 17     S02     I02  1.09271645       0  0.6804446 -0.1383032  0.55057502
#> 30     S02     I03  0.36132041       0  0.6804446 -0.4284905  0.10936629
#> 36     S02     I04 -0.63365419       0  0.6804446 -0.1582815 -1.15581729
#> 46     S02     I05  2.31893043       0  0.6804446  0.3987451  1.23974074
#> 59     S02     I06  1.35478022       0  0.6804446  0.1638725  0.51046306
#> 69     S02     I07  1.31047647       0  0.6804446 -0.4220881  1.05211996
#> 75     S02     I08 -0.36961221       0  0.6804446 -1.5426317  0.49257482
#> 88     S02     I09  1.40336688       0  0.6804446 -0.8490586  1.57198088
#> 98     S02     I10 -0.50683352       0  0.6804446 -0.6288139 -0.55846423
#> 4      S03     I01 -1.77440627       0 -0.4602496 -1.2508904 -0.06326634
#> 18     S03     I02 -0.05815690       0 -0.4602496 -0.1383032  0.54039587
#> 23     S03     I03 -0.35931214       0 -0.4602496 -0.4284905  0.52942795
#> 33     S03     I04 -0.34687654       0 -0.4602496 -0.1582815  0.27165456
#> 47     S03     I05 -1.56333736       0 -0.4602496  0.3987451 -1.50183285
#> 60     S03     I06  0.68021702       0 -0.4602496  0.1638725  0.97659407
#> 62     S03     I07 -0.69327550       0 -0.4602496 -0.4220881  0.18906219
#> 76     S03     I08 -1.91533055       0 -0.4602496 -1.5426317  0.08755069
#> 89     S03     I09  1.25227650       0 -0.4602496 -0.8490586  2.56158471
#> 95     S03     I10 -1.62857645       0 -0.4602496 -0.6288139 -0.53951294
#> 5      S04     I01 -0.95602512       0  0.4623437 -1.2508904 -0.16747843
#> 19     S04     I02  0.74470777       0  0.4623437 -0.1383032  0.42066730
#> 21     S04     I03 -0.89564826       0  0.4623437 -0.4284905 -0.92950142
#> 34     S04     I04  0.38379939       0  0.4623437 -0.1582815  0.07973725
#> 48     S04     I05  2.47008843       0  0.4623437  0.3987451  1.60899970
#> 51     S04     I06  0.77137479       0  0.4623437  0.1638725  0.14515859
#> 63     S04     I07  0.58176745       0  0.4623437 -0.4220881  0.54151190
#> 77     S04     I08 -1.10696293       0  0.4623437 -1.5426317 -0.02667494
#> 90     S04     I09 -0.26314481       0  0.4623437 -0.8490586  0.12357015
#> 92     S04     I10 -0.49563547       0  0.4623437 -0.6288139 -0.32916521
#> 2      S05     I01 -0.39846326       0  0.3918912 -1.2508904  0.46053589
#> 16     S05     I02 -1.02263216       0  0.3918912 -0.1383032 -1.27622017
#> 22     S05     I03 -0.40113610       0  0.3918912 -0.4284905 -0.36453679
#> 35     S05     I04  0.65033739       0  0.3918912 -0.1582815  0.41672771
#> 45     S05     I05 -0.09507895       0  0.3918912  0.3987451 -0.88571521
#> 52     S05     I06  1.16185245       0  0.3918912  0.1638725  0.60608872
#> 64     S05     I07 -0.83220518       0  0.3918912 -0.4220881 -0.80200827
#> 74     S05     I08 -2.64908440       0  0.3918912 -1.5426317 -1.49834394
#> 81     S05     I09  1.36915507       0  0.3918912 -0.8490586  1.82632249
#> 93     S05     I10 -0.16937315       0  0.3918912 -0.6288139  0.06754958
#> 8      S06     I01  1.00668516       0  1.3054596 -1.2508904  0.95211590
#> 20     S06     I02 -0.23519344       0  1.3054596 -0.1383032 -1.40234985
#> 27     S06     I03  0.14199010       0  1.3054596 -0.4284905 -0.73497899
#> 37     S06     I04 -0.36044251       0  1.3054596 -0.1582815 -1.50762058
#> 49     S06     I05  2.25818476       0  1.3054596  0.3987451  0.55398009
#> 56     S06     I06  3.71483396       0  1.3054596  0.1638725  2.24550183
#> 66     S06     I07 -0.14031628       0  1.3054596 -0.4220881 -1.02368777
#> 79     S06     I08 -1.81776233       0  1.3054596 -1.5426317 -1.58059027
#> 85     S06     I09  0.69299360       0  1.3054596 -0.8490586  0.23659263
#> 99     S06     I10  1.37568816       0  1.3054596 -0.6288139  0.69904248
#> 9      S07     I01 -1.95400153       0  0.1847753 -1.2508904 -0.88788650
#> 11     S07     I02  0.25260098       0  0.1847753 -0.1383032  0.20612884
#> 24     S07     I03  1.13405956       0  0.1847753 -0.4284905  1.37777474
#> 38     S07     I04 -2.10867315       0  0.1847753 -0.1582815 -2.13516695
#> 50     S07     I05  1.35193691       0  0.1847753  0.3987451  0.76841651
#> 53     S07     I06 -0.37904817       0  0.1847753  0.1638725 -0.72769603
#> 67     S07     I07 -1.35277127       0  0.1847753 -0.4220881 -1.11545848
#> 80     S07     I08 -1.42753270       0  0.1847753 -1.5426317 -0.06967636
#> 82     S07     I09 -1.27595766       0  0.1847753 -0.8490586 -0.61167436
#> 96     S07     I10 -0.46554200       0  0.1847753 -0.6288139 -0.02150340
#> 6      S08     I01 -2.99543140       0 -2.3116912 -1.2508904  0.56715016
#> 12     S08     I02 -3.74289632       0 -2.3116912 -0.1383032 -1.29290192
#> 25     S08     I03 -3.25591229       0 -2.3116912 -0.4284905 -0.51573057
#> 39     S08     I04 -2.78403597       0 -2.3116912 -0.1582815 -0.31406324
#> 41     S08     I05 -2.93866739       0 -2.3116912  0.3987451 -1.02572125
#> 54     S08     I06 -0.19607330       0 -2.3116912  0.1638725  1.95174538
#> 68     S08     I07 -3.91854235       0 -2.3116912 -0.4220881 -1.18476303
#> 78     S08     I08 -5.19546246       0 -2.3116912 -1.5426317 -1.34113959
#> 83     S08     I09 -1.25025102       0 -2.3116912 -0.8490586  1.91049881
#> 97     S08     I10 -3.58340843       0 -2.3116912 -0.6288139 -0.64290330
#> 3      S09     I01 -0.32811200       0  0.1429659 -1.2508904  0.77981244
#> 13     S09     I02 -1.39981447       0  0.1429659 -0.1383032 -1.40447718
#> 26     S09     I03  0.50787996       0  0.1429659 -0.4284905  0.79340457
#> 32     S09     I04  0.76562137       0  0.1429659 -0.1582815  0.78093699
#> 42     S09     I05  1.01195832       0  0.1429659  0.3987451  0.47024735
#> 55     S09     I06  0.02731895       0  0.1429659  0.1638725 -0.27951949
#> 65     S09     I07 -1.36765178       0  0.1429659 -0.4220881 -1.08852957
#> 71     S09     I08 -1.18305746       0  0.1429659 -1.5426317  0.21660830
#> 84     S09     I09 -0.51097492       0  0.1429659 -0.8490586  0.19511780
#> 94     S09     I10  0.29120018       0  0.1429659 -0.6288139  0.77704820
#> 10     S10     I01 -2.73062488       0 -0.5516974 -1.2508904 -0.92803716
#> 14     S10     I02 -0.70403212       0 -0.5516974 -0.1383032 -0.01403155
#> 28     S10     I03 -0.15939519       0 -0.5516974 -0.4284905  0.82079269
#> 40     S10     I04 -0.91274452       0 -0.5516974 -0.1582815 -0.20276562
#> 43     S10     I05 -0.05536677       0 -0.5516974  0.3987451  0.09758554
#> 57     S10     I06 -0.67305470       0 -0.5516974  0.1638725 -0.28522986
#> 70     S10     I07  1.08321012       0 -0.5516974 -0.4220881  2.05699561
#> 72     S10     I08 -1.27805653       0 -0.5516974 -1.5426317  0.81627250
#> 86     S10     I09 -1.00037534       0 -0.5516974 -0.8490586  0.40038066
#> 100    S10     I10 -0.80974696       0 -0.5516974 -0.6288139  0.37076434