Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

Usage

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n

the number of subjects

item_n

the number of items

grand_i

the grand intercept (overall mean)

sub_sd

the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd

the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd

the SD of the error term

empirical

Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)
#>     sub_id item_id           y grand_i       sub_i      item_i          err
#> 1      S01     I01 -0.13968429       0 -0.72658585  0.16470704  0.422194515
#> 15     S01     I02  0.86368172       0 -0.72658585  0.79567069  0.794596879
#> 29     S01     I03 -3.13640413       0 -0.72658585 -1.20802674 -1.201791539
#> 31     S01     I04  3.45406584       0 -0.72658585  1.65928157  2.521370113
#> 44     S01     I05 -0.71358959       0 -0.72658585  0.07577959 -0.062783339
#> 58     S01     I06 -0.41545380       0 -0.72658585 -0.92417838  1.235310423
#> 61     S01     I07  0.30482345       0 -0.72658585  1.49075715 -0.459347861
#> 73     S01     I08 -1.67657043       0 -0.72658585 -1.26620660  0.316222016
#> 87     S01     I09  0.74404972       0 -0.72658585  1.34325276  0.127382806
#> 91     S01     I10 -1.73778943       0 -0.72658585 -0.62949073 -0.381712858
#> 7      S02     I01 -0.73560089       0 -1.18872045  0.16470704  0.288412525
#> 17     S02     I02 -0.03343747       0 -1.18872045  0.79567069  0.359612294
#> 30     S02     I03 -2.76772689       0 -1.18872045 -1.20802674 -0.370979695
#> 36     S02     I04  1.49366389       0 -1.18872045  1.65928157  1.023102772
#> 46     S02     I05 -1.23973339       0 -1.18872045  0.07577959 -0.126792538
#> 59     S02     I06 -2.86416544       0 -1.18872045 -0.92417838 -0.751266620
#> 69     S02     I07 -1.47285364       0 -1.18872045  1.49075715 -1.774890342
#> 75     S02     I08 -1.80425376       0 -1.18872045 -1.26620660  0.650673287
#> 88     S02     I09 -0.43973169       0 -1.18872045  1.34325276 -0.594264005
#> 98     S02     I10  0.04728649       0 -1.18872045 -0.62949073  1.865497665
#> 4      S03     I01  2.59950245       0  0.66869032  0.16470704  1.766105101
#> 18     S03     I02  0.74224404       0  0.66869032  0.79567069 -0.722116967
#> 23     S03     I03 -0.27130276       0  0.66869032 -1.20802674  0.268033668
#> 33     S03     I04  2.18729291       0  0.66869032  1.65928157 -0.140678976
#> 47     S03     I05  1.29314833       0  0.66869032  0.07577959  0.548678415
#> 60     S03     I06 -0.57434811       0  0.66869032 -0.92417838 -0.318860052
#> 62     S03     I07  1.44924477       0  0.66869032  1.49075715 -0.710202697
#> 76     S03     I08  1.19782141       0  0.66869032 -1.26620660  1.795337692
#> 89     S03     I09  2.07334414       0  0.66869032  1.34325276  0.061401057
#> 95     S03     I10 -0.97034281       0  0.66869032 -0.62949073 -1.009542404
#> 5      S04     I01  1.06786952       0  1.46320659  0.16470704 -0.560044112
#> 19     S04     I02  1.69095645       0  1.46320659  0.79567069 -0.567920829
#> 21     S04     I03  0.96601565       0  1.46320659 -1.20802674  0.710835801
#> 34     S04     I04  3.10996250       0  1.46320659  1.65928157 -0.012525661
#> 48     S04     I05  2.40882301       0  1.46320659  0.07577959  0.869836829
#> 51     S04     I06 -0.32724845       0  1.46320659 -0.92417838 -0.866276663
#> 63     S04     I07  3.73870920       0  1.46320659  1.49075715  0.784745460
#> 77     S04     I08  0.93152970       0  1.46320659 -1.26620660  0.734529710
#> 90     S04     I09  1.97102776       0  1.46320659  1.34325276 -0.835431598
#> 92     S04     I10  0.22188394       0  1.46320659 -0.62949073 -0.611831926
#> 2      S05     I01 -0.10137687       0  0.02284036  0.16470704 -0.288924265
#> 16     S05     I02  1.30435477       0  0.02284036  0.79567069  0.485843724
#> 22     S05     I03 -0.54461081       0  0.02284036 -1.20802674  0.640575578
#> 35     S05     I04  2.47792883       0  0.02284036  1.65928157  0.795806903
#> 45     S05     I05  0.16919845       0  0.02284036  0.07577959  0.070578497
#> 52     S05     I06 -1.82299895       0  0.02284036 -0.92417838 -0.921660933
#> 64     S05     I07  2.04979880       0  0.02284036  1.49075715  0.536201287
#> 74     S05     I08 -1.21750634       0  0.02284036 -1.26620660  0.025859899
#> 81     S05     I09 -1.36905369       0  0.02284036  1.34325276 -2.735146813
#> 93     S05     I10 -1.53412566       0  0.02284036 -0.62949073 -0.927475292
#> 8      S06     I01  1.33393551       0  0.08795165  0.16470704  1.081276818
#> 20     S06     I02  2.26823655       0  0.08795165  0.79567069  1.384614216
#> 27     S06     I03 -0.03204851       0  0.08795165 -1.20802674  1.088026584
#> 37     S06     I04  3.04076440       0  0.08795165  1.65928157  1.293531178
#> 49     S06     I05  0.29528274       0  0.08795165  0.07577959  0.131551493
#> 56     S06     I06  0.36993278       0  0.08795165 -0.92417838  1.206159504
#> 66     S06     I07  2.47808550       0  0.08795165  1.49075715  0.899376701
#> 79     S06     I08  0.55574783       0  0.08795165 -1.26620660  1.734002776
#> 85     S06     I09  0.22186883       0  0.08795165  1.34325276 -1.209335585
#> 99     S06     I10 -0.52515528       0  0.08795165 -0.62949073  0.016383795
#> 9      S07     I01 -1.04090889       0 -0.24641902  0.16470704 -0.959196903
#> 11     S07     I02  0.66636683       0 -0.24641902  0.79567069  0.117115163
#> 24     S07     I03 -1.74785349       0 -0.24641902 -1.20802674 -0.293407729
#> 38     S07     I04  2.34853281       0 -0.24641902  1.65928157  0.935670259
#> 50     S07     I05  0.22684453       0 -0.24641902  0.07577959  0.397483952
#> 53     S07     I06 -2.29285030       0 -0.24641902 -0.92417838 -1.122252909
#> 67     S07     I07  1.97793161       0 -0.24641902  1.49075715  0.733593471
#> 80     S07     I08 -0.15133855       0 -0.24641902 -1.26620660  1.361287067
#> 82     S07     I09 -0.59715040       0 -0.24641902  1.34325276 -1.693984149
#> 96     S07     I10  0.94149217       0 -0.24641902 -0.62949073  1.817401912
#> 6      S08     I01  2.58575180       0  1.48494859  0.16470704  0.936096181
#> 12     S08     I02  2.25945508       0  1.48494859  0.79567069 -0.021164194
#> 25     S08     I03  2.82989187       0  1.48494859 -1.20802674  2.552970031
#> 39     S08     I04  3.36918176       0  1.48494859  1.65928157  0.224951608
#> 41     S08     I05  0.85566375       0  1.48494859  0.07577959 -0.705064426
#> 54     S08     I06 -0.18278540       0  1.48494859 -0.92417838 -0.743555605
#> 68     S08     I07  1.72133907       0  1.48494859  1.49075715 -1.254366670
#> 78     S08     I08  0.39280434       0  1.48494859 -1.26620660  0.174062355
#> 83     S08     I09  4.03516877       0  1.48494859  1.34325276  1.206967424
#> 97     S08     I10  0.37015478       0  1.48494859 -0.62949073 -0.485303082
#> 3      S09     I01  0.83197308       0  0.19652868  0.16470704  0.470737361
#> 13     S09     I02 -0.79519899       0  0.19652868  0.79567069 -1.787398363
#> 26     S09     I03 -1.01631645       0  0.19652868 -1.20802674 -0.004818383
#> 32     S09     I04  2.37436122       0  0.19652868  1.65928157  0.518550969
#> 42     S09     I05  0.66450953       0  0.19652868  0.07577959  0.392201254
#> 55     S09     I06 -1.29750640       0  0.19652868 -0.92417838 -0.569856701
#> 65     S09     I07  1.94709671       0  0.19652868  1.49075715  0.259810881
#> 71     S09     I08 -1.19854039       0  0.19652868 -1.26620660 -0.128862475
#> 84     S09     I09 -1.01661600       0  0.19652868  1.34325276 -2.556397450
#> 94     S09     I10 -0.34503023       0  0.19652868 -0.62949073  0.087931813
#> 10     S10     I01 -2.94181700       0 -1.28319631  0.16470704 -1.823327731
#> 14     S10     I02  0.38919865       0 -1.28319631  0.79567069  0.876724272
#> 28     S10     I03 -2.11670968       0 -1.28319631 -1.20802674  0.374513371
#> 40     S10     I04 -0.35193517       0 -1.28319631  1.65928157 -0.728020429
#> 43     S10     I05 -1.53188355       0 -1.28319631  0.07577959 -0.324466836
#> 57     S10     I06 -3.65145293       0 -1.28319631 -0.92417838 -1.444078246
#> 70     S10     I07  0.05683838       0 -1.28319631  1.49075715 -0.150722467
#> 72     S10     I08 -2.45174726       0 -1.28319631 -1.26620660  0.097655648
#> 86     S10     I09  0.53171151       0 -1.28319631  1.34325276  0.471655050
#> 100    S10     I10 -1.81204078       0 -1.28319631 -0.62949073  0.100646251