Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id            y grand_i       sub_i      item_i         err
#> 1      S01     I01 -1.793893352       0 -0.59090568 -1.01660298 -0.18638469
#> 15     S01     I02 -0.763910394       0 -0.59090568  0.88886409 -1.06186880
#> 29     S01     I03  0.226566306       0 -0.59090568  0.42594280  0.39152919
#> 31     S01     I04 -1.848858823       0 -0.59090568  0.08014553 -1.33809867
#> 44     S01     I05 -3.037397085       0 -0.59090568 -1.35340447 -1.09308693
#> 58     S01     I06 -0.534833687       0 -0.59090568  1.00501576 -0.94894376
#> 61     S01     I07 -1.570240413       0 -0.59090568 -0.49763308 -0.48170165
#> 73     S01     I08 -0.923906680       0 -0.59090568  0.19475232 -0.52775332
#> 87     S01     I09 -1.807811574       0 -0.59090568 -0.04031917 -1.17658672
#> 91     S01     I10 -1.224638292       0 -0.59090568  0.28022273 -0.91395534
#> 7      S02     I01 -2.414269368       0 -0.70815518 -1.01660298 -0.68951120
#> 17     S02     I02  1.416409250       0 -0.70815518  0.88886409  1.23570034
#> 30     S02     I03  0.404171409       0 -0.70815518  0.42594280  0.68638379
#> 36     S02     I04 -0.989007669       0 -0.70815518  0.08014553 -0.36099802
#> 46     S02     I05 -1.126924466       0 -0.70815518 -1.35340447  0.93463519
#> 59     S02     I06  0.090902381       0 -0.70815518  1.00501576 -0.20595820
#> 69     S02     I07 -0.001254428       0 -0.70815518 -0.49763308  1.20453383
#> 75     S02     I08 -2.319936911       0 -0.70815518  0.19475232 -1.80653405
#> 88     S02     I09 -0.962313229       0 -0.70815518 -0.04031917 -0.21383888
#> 98     S02     I10 -0.894686937       0 -0.70815518  0.28022273 -0.46675448
#> 4      S03     I01 -2.576653634       0 -0.99177457 -1.01660298 -0.56827608
#> 18     S03     I02 -1.396105449       0 -0.99177457  0.88886409 -1.29319497
#> 23     S03     I03 -1.769733369       0 -0.99177457  0.42594280 -1.20390160
#> 33     S03     I04 -0.249819690       0 -0.99177457  0.08014553  0.66180935
#> 47     S03     I05 -4.339868182       0 -0.99177457 -1.35340447 -1.99468914
#> 60     S03     I06 -0.745254065       0 -0.99177457  1.00501576 -0.75849525
#> 62     S03     I07 -0.213739238       0 -0.99177457 -0.49763308  1.27566841
#> 76     S03     I08 -1.142421234       0 -0.99177457  0.19475232 -0.34539898
#> 89     S03     I09 -1.692656607       0 -0.99177457 -0.04031917 -0.66056287
#> 95     S03     I10  0.088075499       0 -0.99177457  0.28022273  0.79962734
#> 5      S04     I01  2.861861202       0  1.90849295 -1.01660298  1.96997123
#> 19     S04     I02  3.651425131       0  1.90849295  0.88886409  0.85406809
#> 21     S04     I03  4.657050023       0  1.90849295  0.42594280  2.32261427
#> 34     S04     I04  1.695813447       0  1.90849295  0.08014553 -0.29282503
#> 48     S04     I05  1.603691100       0  1.90849295 -1.35340447  1.04860262
#> 51     S04     I06  5.721062820       0  1.90849295  1.00501576  2.80755411
#> 63     S04     I07  1.005021236       0  1.90849295 -0.49763308 -0.40583864
#> 77     S04     I08  2.460387155       0  1.90849295  0.19475232  0.35714188
#> 90     S04     I09  1.680249176       0  1.90849295 -0.04031917 -0.18792460
#> 92     S04     I10  5.067278459       0  1.90849295  0.28022273  2.87856278
#> 2      S05     I01  1.332516478       0  0.03670272 -1.01660298  2.31241674
#> 16     S05     I02  1.655310572       0  0.03670272  0.88886409  0.72974377
#> 22     S05     I03 -0.339512282       0  0.03670272  0.42594280 -0.80215780
#> 35     S05     I04  1.349210025       0  0.03670272  0.08014553  1.23236178
#> 45     S05     I05 -2.878496380       0  0.03670272 -1.35340447 -1.56179462
#> 52     S05     I06  0.402314034       0  0.03670272  1.00501576 -0.63940444
#> 64     S05     I07 -3.590134498       0  0.03670272 -0.49763308 -3.12920414
#> 74     S05     I08  1.157198467       0  0.03670272  0.19475232  0.92574343
#> 81     S05     I09  0.440231172       0  0.03670272 -0.04031917  0.44384763
#> 93     S05     I10  0.476719246       0  0.03670272  0.28022273  0.15979380
#> 8      S06     I01 -1.733452643       0 -0.51436372 -1.01660298 -0.20248594
#> 20     S06     I02 -0.829008271       0 -0.51436372  0.88886409 -1.20350864
#> 27     S06     I03  0.161322778       0 -0.51436372  0.42594280  0.24974370
#> 37     S06     I04  0.200868998       0 -0.51436372  0.08014553  0.63508719
#> 49     S06     I05 -2.000901057       0 -0.51436372 -1.35340447 -0.13313286
#> 56     S06     I06 -1.126687935       0 -0.51436372  1.00501576 -1.61733997
#> 66     S06     I07  1.629651817       0 -0.51436372 -0.49763308  2.64164862
#> 79     S06     I08 -2.298852424       0 -0.51436372  0.19475232 -1.97924102
#> 85     S06     I09 -0.257018146       0 -0.51436372 -0.04031917  0.29766474
#> 99     S06     I10 -0.095326646       0 -0.51436372  0.28022273  0.13881434
#> 9      S07     I01 -1.190056752       0 -0.38407400 -1.01660298  0.21062023
#> 11     S07     I02  1.809983733       0 -0.38407400  0.88886409  1.30519365
#> 24     S07     I03  0.845171677       0 -0.38407400  0.42594280  0.80330288
#> 38     S07     I04 -1.045551865       0 -0.38407400  0.08014553 -0.74162339
#> 50     S07     I05 -1.805973478       0 -0.38407400 -1.35340447 -0.06849500
#> 53     S07     I06  1.226065806       0 -0.38407400  1.00501576  0.60512405
#> 67     S07     I07 -2.141669650       0 -0.38407400 -0.49763308 -1.25996257
#> 80     S07     I08  0.628557004       0 -0.38407400  0.19475232  0.81787869
#> 82     S07     I09 -0.579978040       0 -0.38407400 -0.04031917 -0.15558487
#> 96     S07     I10  0.167197487       0 -0.38407400  0.28022273  0.27104876
#> 6      S08     I01  0.497472501       0  1.21703584 -1.01660298  0.29703965
#> 12     S08     I02  2.445849375       0  1.21703584  0.88886409  0.33994945
#> 25     S08     I03  2.568517819       0  1.21703584  0.42594280  0.92553918
#> 39     S08     I04  1.977838284       0  1.21703584  0.08014553  0.68065692
#> 41     S08     I05 -0.392648098       0  1.21703584 -1.35340447 -0.25627946
#> 54     S08     I06  3.582122634       0  1.21703584  1.00501576  1.36007104
#> 68     S08     I07  1.950458261       0  1.21703584 -0.49763308  1.23105551
#> 78     S08     I08  1.934374768       0  1.21703584  0.19475232  0.52258662
#> 83     S08     I09  0.350963871       0  1.21703584 -0.04031917 -0.82575279
#> 97     S08     I10  1.736483393       0  1.21703584  0.28022273  0.23922483
#> 3      S09     I01  2.553588160       0  1.56285588 -1.01660298  2.00733526
#> 13     S09     I02  3.894598782       0  1.56285588  0.88886409  1.44287881
#> 26     S09     I03  1.421513479       0  1.56285588  0.42594280 -0.56728520
#> 32     S09     I04  1.662002972       0  1.56285588  0.08014553  0.01900156
#> 42     S09     I05  0.971027187       0  1.56285588 -1.35340447  0.76157578
#> 55     S09     I06  2.220621234       0  1.56285588  1.00501576 -0.34725041
#> 65     S09     I07  1.802481233       0  1.56285588 -0.49763308  0.73725843
#> 71     S09     I08  1.058650656       0  1.56285588  0.19475232 -0.69895754
#> 84     S09     I09  2.152885289       0  1.56285588 -0.04031917  0.63034858
#> 94     S09     I10  0.889223072       0  1.56285588  0.28022273 -0.95385554
#> 10     S10     I01  2.499136451       0  2.69604356 -1.01660298  0.81969587
#> 14     S10     I02  3.441591623       0  2.69604356  0.88886409 -0.14331602
#> 28     S10     I03  4.051150950       0  2.69604356  0.42594280  0.92916459
#> 40     S10     I04  2.307202467       0  2.69604356  0.08014553 -0.46898662
#> 43     S10     I05  0.275316924       0  2.69604356 -1.35340447 -1.06732216
#> 57     S10     I06  3.016925982       0  2.69604356  1.00501576 -0.68413334
#> 70     S10     I07  1.082463319       0  2.69604356 -0.49763308 -1.11594716
#> 72     S10     I08  1.875591884       0  2.69604356  0.19475232 -1.01520399
#> 86     S10     I09  2.826066272       0  2.69604356 -0.04031917  0.17034188
#> 100    S10     I10  3.434347636       0  2.69604356  0.28022273  0.45808135