Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

Usage

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n

the number of subjects

item_n

the number of items

grand_i

the grand intercept (overall mean)

sub_sd

the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd

the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd

the SD of the error term

empirical

Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)
#>     sub_id item_id           y grand_i        sub_i     item_i         err
#> 1      S01     I01  0.79873488       0 -0.230518876  0.1384996  0.89075414
#> 15     S01     I02  0.45266767       0 -0.230518876  0.8944780 -0.21129144
#> 29     S01     I03  1.97096422       0 -0.230518876  0.6006202  1.60086292
#> 31     S01     I04 -0.73246451       0 -0.230518876  0.9553465 -1.45729217
#> 44     S01     I05  0.94470794       0 -0.230518876  1.6750840 -0.49985722
#> 58     S01     I06 -1.22869160       0 -0.230518876  0.2620355 -1.26020820
#> 61     S01     I07  1.06813600       0 -0.230518876  0.9446057  0.35404914
#> 73     S01     I08 -0.40708537       0 -0.230518876  0.4493581 -0.62592457
#> 87     S01     I09 -1.03918150       0 -0.230518876 -0.6399552 -0.16870740
#> 91     S01     I10  1.73995267       0 -0.230518876  1.1530081  0.81746340
#> 7      S02     I01  0.74103460       0 -1.049536453  0.1384996  1.65207143
#> 17     S02     I02 -0.56164670       0 -1.049536453  0.8944780 -0.40658823
#> 30     S02     I03 -0.90601120       0 -1.049536453  0.6006202 -0.45709493
#> 36     S02     I04 -0.74240986       0 -1.049536453  0.9553465 -0.64821994
#> 46     S02     I05  1.06446516       0 -1.049536453  1.6750840  0.43891757
#> 59     S02     I06 -0.87959999       0 -1.049536453  0.2620355 -0.09209901
#> 69     S02     I07  0.66015900       0 -1.049536453  0.9446057  0.76508972
#> 75     S02     I08 -0.41628504       0 -1.049536453  0.4493581  0.18389333
#> 88     S02     I09 -1.78610060       0 -1.049536453 -0.6399552 -0.09660893
#> 98     S02     I10 -0.34758325       0 -1.049536453  1.1530081 -0.45105495
#> 4      S03     I01  0.60385832       0  1.223107440  0.1384996 -0.75774874
#> 18     S03     I02  1.13216725       0  1.223107440  0.8944780 -0.98541818
#> 23     S03     I03  2.10866827       0  1.223107440  0.6006202  0.28494066
#> 33     S03     I04  2.56525136       0  1.223107440  0.9553465  0.38679738
#> 47     S03     I05  2.96321979       0  1.223107440  1.6750840  0.06502830
#> 60     S03     I06  3.87667160       0  1.223107440  0.2620355  2.39152869
#> 62     S03     I07  2.42533380       0  1.223107440  0.9446057  0.25762063
#> 76     S03     I08  2.15718602       0  1.223107440  0.4493581  0.48472050
#> 89     S03     I09  2.33108419       0  1.223107440 -0.6399552  1.74793197
#> 95     S03     I10  2.92207567       0  1.223107440  1.1530081  0.54596009
#> 5      S04     I01  0.04783186       0  0.008421251  0.1384996 -0.09908902
#> 19     S04     I02  0.71843318       0  0.008421251  0.8944780 -0.18446605
#> 21     S04     I03 -0.67265854       0  0.008421251  0.6006202 -1.28169997
#> 34     S04     I04  0.35753535       0  0.008421251  0.9553465 -0.60623244
#> 48     S04     I05  3.08771532       0  0.008421251  1.6750840  1.40421003
#> 51     S04     I06 -0.39531417       0  0.008421251  0.2620355 -0.66577090
#> 63     S04     I07  1.86102349       0  0.008421251  0.9446057  0.90799651
#> 77     S04     I08  1.80185009       0  0.008421251  0.4493581  1.34407075
#> 90     S04     I09  0.04106562       0  0.008421251 -0.6399552  0.67259959
#> 92     S04     I10  1.67028424       0  0.008421251  1.1530081  0.50885484
#> 2      S05     I01  1.86747666       0  1.235342028  0.1384996  0.49363501
#> 16     S05     I02  3.17000249       0  1.235342028  0.8944780  1.04018248
#> 22     S05     I03  2.14003733       0  1.235342028  0.6006202  0.30407512
#> 35     S05     I04  2.69744234       0  1.235342028  0.9553465  0.50675378
#> 45     S05     I05  3.73101173       0  1.235342028  1.6750840  0.82058566
#> 52     S05     I06  2.99958355       0  1.235342028  0.2620355  1.50220604
#> 64     S05     I07  1.27956881       0  1.235342028  0.9446057 -0.90037895
#> 74     S05     I08  2.32101378       0  1.235342028  0.4493581  0.63631367
#> 81     S05     I09  1.48303395       0  1.235342028 -0.6399552  0.88764714
#> 93     S05     I10  1.37823829       0  1.235342028  1.1530081 -1.01011188
#> 8      S06     I01  1.31687229       0  1.092753326  0.1384996  0.08561934
#> 20     S06     I02  2.48104896       0  1.092753326  0.8944780  0.49381765
#> 27     S06     I03  4.59266796       0  1.092753326  0.6006202  2.89929446
#> 37     S06     I04  2.79688900       0  1.092753326  0.9553465  0.74878914
#> 49     S06     I05  1.55576116       0  1.092753326  1.6750840 -1.21207621
#> 56     S06     I06  1.71699049       0  1.092753326  0.2620355  0.36220168
#> 66     S06     I07  3.29807581       0  1.092753326  0.9446057  1.26071676
#> 79     S06     I08  1.61053642       0  1.092753326  0.4493581  0.06842502
#> 85     S06     I09  1.99838162       0  1.092753326 -0.6399552  1.54558352
#> 99     S06     I10  2.01892931       0  1.092753326  1.1530081 -0.22683216
#> 9      S07     I01  1.26472569       0  0.188142962  0.1384996  0.93808311
#> 11     S07     I02  1.01759501       0  0.188142962  0.8944780 -0.06502594
#> 24     S07     I03 -0.16358740       0  0.188142962  0.6006202 -0.95235054
#> 38     S07     I04  1.41782314       0  0.188142962  0.9553465  0.27433364
#> 50     S07     I05  0.88014068       0  0.188142962  1.6750840 -0.98308632
#> 53     S07     I06  1.67877325       0  0.188142962  0.2620355  1.22859481
#> 67     S07     I07  1.31229217       0  0.188142962  0.9446057  0.17954348
#> 80     S07     I08 -0.47397357       0  0.188142962  0.4493581 -1.11147461
#> 82     S07     I09  0.95629737       0  0.188142962 -0.6399552  1.40810963
#> 96     S07     I10  2.09140336       0  0.188142962  1.1530081  0.75025225
#> 6      S08     I01  0.04939942       0 -0.959012838  0.1384996  0.86991264
#> 12     S08     I02  0.70930301       0 -0.959012838  0.8944780  0.77383786
#> 25     S08     I03  1.49975623       0 -0.959012838  0.6006202  1.85814889
#> 39     S08     I04  1.01277929       0 -0.959012838  0.9553465  1.01644559
#> 41     S08     I05 -0.57946843       0 -0.959012838  1.6750840 -1.29553963
#> 54     S08     I06 -1.00439896       0 -0.959012838  0.2620355 -0.30742160
#> 68     S08     I07  1.06313310       0 -0.959012838  0.9446057  1.07754021
#> 78     S08     I08 -0.48166731       0 -0.959012838  0.4493581  0.02798745
#> 83     S08     I09 -0.27996832       0 -0.959012838 -0.6399552  1.31899974
#> 97     S08     I10  0.34851500       0 -0.959012838  1.1530081  0.15451969
#> 3      S09     I01  2.53734842       0  1.235159450  0.1384996  1.16368935
#> 13     S09     I02  2.28582323       0  1.235159450  0.8944780  0.15618580
#> 26     S09     I03  0.87667853       0  1.235159450  0.6006202 -0.95910110
#> 32     S09     I04  1.10019597       0  1.235159450  0.9553465 -1.09031002
#> 42     S09     I05  2.70652816       0  1.235159450  1.6750840 -0.20371533
#> 55     S09     I06  2.72938143       0  1.235159450  0.2620355  1.23218651
#> 65     S09     I07  3.05612619       0  1.235159450  0.9446057  0.87636101
#> 71     S09     I08  3.39205760       0  1.235159450  0.4493581  1.70754007
#> 84     S09     I09  0.82599844       0  1.235159450 -0.6399552  0.23079422
#> 94     S09     I10  3.08030975       0  1.235159450  1.1530081  0.69214215
#> 10     S10     I01 -1.78373615       0 -0.258017769  0.1384996 -1.66421800
#> 14     S10     I02  0.69710173       0 -0.258017769  0.8944780  0.06064151
#> 28     S10     I03 -1.45447884       0 -0.258017769  0.6006202 -1.79708124
#> 40     S10     I04 -0.26078002       0 -0.258017769  0.9553465 -0.95810879
#> 43     S10     I05  0.13774732       0 -0.258017769  1.6750840 -1.27931896
#> 57     S10     I06  0.28574627       0 -0.258017769  0.2620355  0.28172856
#> 70     S10     I07  0.09600028       0 -0.258017769  0.9446057 -0.59058768
#> 72     S10     I08 -0.12071988       0 -0.258017769  0.4493581 -0.31206019
#> 86     S10     I09 -1.31601553       0 -0.258017769 -0.6399552 -0.41804254
#> 100    S10     I10  0.75208663       0 -0.258017769  1.1530081 -0.14290374