Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i       sub_i      item_i          err
#> 1      S01     I01  2.68240516       0  1.25849769  1.41843783  0.005469635
#> 15     S01     I02  0.35008882       0  1.25849769  0.09166167 -1.000070539
#> 29     S01     I03  3.11508711       0  1.25849769  0.15001932  1.706570096
#> 31     S01     I04  2.42709504       0  1.25849769  0.86091975  0.307677593
#> 44     S01     I05  0.46491647       0  1.25849769  0.40760612 -1.201187338
#> 58     S01     I06  0.78317159       0  1.25849769 -0.52474846  0.049422350
#> 61     S01     I07 -0.29472824       0  1.25849769 -0.60867558 -0.944550354
#> 73     S01     I08 -3.07655560       0  1.25849769 -1.70809505 -2.626958246
#> 87     S01     I09  2.04745997       0  1.25849769 -0.15835100  0.947313273
#> 91     S01     I10  1.19630828       0  1.25849769 -1.67857591  1.616386498
#> 7      S02     I01  0.11279540       0 -0.38877796  1.41843783 -0.916864472
#> 17     S02     I02  1.56050431       0 -0.38877796  0.09166167  1.857620601
#> 30     S02     I03  0.28497982       0 -0.38877796  0.15001932  0.523738459
#> 36     S02     I04 -0.13714955       0 -0.38877796  0.86091975 -0.609291349
#> 46     S02     I05 -0.79198463       0 -0.38877796  0.40760612 -0.810812790
#> 59     S02     I06 -0.79435646       0 -0.38877796 -0.52474846  0.119169958
#> 69     S02     I07 -0.04958474       0 -0.38877796 -0.60867558  0.947868796
#> 75     S02     I08 -2.76760478       0 -0.38877796 -1.70809505 -0.670731766
#> 88     S02     I09 -1.29015230       0 -0.38877796 -0.15835100 -0.743023334
#> 98     S02     I10 -1.97880586       0 -0.38877796 -1.67857591  0.088548012
#> 4      S03     I01  3.15071378       0  2.31093328  1.41843783 -0.578657330
#> 18     S03     I02  0.87566799       0  2.31093328  0.09166167 -1.526926959
#> 23     S03     I03  3.14032189       0  2.31093328  0.15001932  0.679369291
#> 33     S03     I04  3.12936218       0  2.31093328  0.86091975 -0.042490860
#> 47     S03     I05  3.21887880       0  2.31093328  0.40760612  0.500339403
#> 60     S03     I06  3.43270367       0  2.31093328 -0.52474846  1.646518846
#> 62     S03     I07  1.16518939       0  2.31093328 -0.60867558 -0.537068316
#> 76     S03     I08  1.39081171       0  2.31093328 -1.70809505  0.787973478
#> 89     S03     I09  3.47654678       0  2.31093328 -0.15835100  1.323964497
#> 95     S03     I10 -0.25584223       0  2.31093328 -1.67857591 -0.888199600
#> 5      S04     I01  1.86066528       0 -1.26736217  1.41843783  1.709589619
#> 19     S04     I02 -3.29516740       0 -1.26736217  0.09166167 -2.119466896
#> 21     S04     I03 -3.37422007       0 -1.26736217  0.15001932 -2.256877218
#> 34     S04     I04  0.32530965       0 -1.26736217  0.86091975  0.731752060
#> 48     S04     I05 -0.55102733       0 -1.26736217  0.40760612  0.308728720
#> 51     S04     I06 -0.64044290       0 -1.26736217 -0.52474846  1.151667720
#> 63     S04     I07 -0.15827097       0 -1.26736217 -0.60867558  1.717766773
#> 77     S04     I08 -2.29553226       0 -1.26736217 -1.70809505  0.679924952
#> 90     S04     I09 -1.86658057       0 -1.26736217 -0.15835100 -0.440867399
#> 92     S04     I10 -3.16474783       0 -1.26736217 -1.67857591 -0.218809754
#> 2      S05     I01  1.34908479       0 -0.06053322  1.41843783 -0.008819819
#> 16     S05     I02 -2.97759477       0 -0.06053322  0.09166167 -3.008723217
#> 22     S05     I03  0.37688315       0 -0.06053322  0.15001932  0.287397053
#> 35     S05     I04 -1.87333735       0 -0.06053322  0.86091975 -2.673723886
#> 45     S05     I05  0.25971494       0 -0.06053322  0.40760612 -0.087357953
#> 52     S05     I06 -3.12665921       0 -0.06053322 -0.52474846 -2.541377539
#> 64     S05     I07 -0.62914110       0 -0.06053322 -0.60867558  0.040067695
#> 74     S05     I08 -1.95600990       0 -0.06053322 -1.70809505 -0.187381628
#> 81     S05     I09 -0.38151390       0 -0.06053322 -0.15835100 -0.162629680
#> 93     S05     I10 -0.42973268       0 -0.06053322 -1.67857591  1.309376453
#> 8      S06     I01  1.78975123       0 -0.11397900  1.41843783  0.485292406
#> 20     S06     I02 -1.43162398       0 -0.11397900  0.09166167 -1.409306645
#> 27     S06     I03 -1.72284544       0 -0.11397900  0.15001932 -1.758885758
#> 37     S06     I04 -0.09811033       0 -0.11397900  0.86091975 -0.845051082
#> 49     S06     I05 -0.51368630       0 -0.11397900  0.40760612 -0.807313410
#> 56     S06     I06 -0.80098259       0 -0.11397900 -0.52474846 -0.162255134
#> 66     S06     I07 -0.87933885       0 -0.11397900 -0.60867558 -0.156684270
#> 79     S06     I08 -2.85958026       0 -0.11397900 -1.70809505 -1.037506206
#> 85     S06     I09 -0.26980269       0 -0.11397900 -0.15835100  0.002527318
#> 99     S06     I10 -2.76718699       0 -0.11397900 -1.67857591 -0.974632076
#> 9      S07     I01  3.75304054       0  0.92153643  1.41843783  1.413066277
#> 11     S07     I02  0.78778623       0  0.92153643  0.09166167 -0.225411866
#> 24     S07     I03  0.37885193       0  0.92153643  0.15001932 -0.692703826
#> 38     S07     I04  1.63407273       0  0.92153643  0.86091975 -0.148383453
#> 50     S07     I05  2.31895639       0  0.92153643  0.40760612  0.989813835
#> 53     S07     I06  1.53582271       0  0.92153643 -0.52474846  1.139034731
#> 67     S07     I07  1.12649844       0  0.92153643 -0.60867558  0.813637585
#> 80     S07     I08 -1.33120408       0  0.92153643 -1.70809505 -0.544645468
#> 82     S07     I09  3.13715721       0  0.92153643 -0.15835100  2.373971778
#> 96     S07     I10 -0.26579748       0  0.92153643 -1.67857591  0.491242000
#> 6      S08     I01  2.59702611       0  0.97253168  1.41843783  0.206056608
#> 12     S08     I02  1.87397794       0  0.97253168  0.09166167  0.809784599
#> 25     S08     I03  2.04830948       0  0.97253168  0.15001932  0.925758489
#> 39     S08     I04  3.42576017       0  0.97253168  0.86091975  1.592308736
#> 41     S08     I05  1.94495542       0  0.97253168  0.40760612  0.564817631
#> 54     S08     I06 -0.01103292       0  0.97253168 -0.52474846 -0.458816139
#> 68     S08     I07  0.49892862       0  0.97253168 -0.60867558  0.135072527
#> 78     S08     I08 -0.77803276       0  0.97253168 -1.70809505 -0.042469384
#> 83     S08     I09  1.30017294       0  0.97253168 -0.15835100  0.485992267
#> 97     S08     I10 -1.82632225       0  0.97253168 -1.67857591 -1.120278010
#> 3      S09     I01  2.55435426       0 -0.36728700  1.41843783  1.503203425
#> 13     S09     I02  0.47170321       0 -0.36728700  0.09166167  0.747328539
#> 26     S09     I03 -0.87341171       0 -0.36728700  0.15001932 -0.656144027
#> 32     S09     I04  0.22986107       0 -0.36728700  0.86091975 -0.263771684
#> 42     S09     I05  0.42045869       0 -0.36728700  0.40760612  0.380139571
#> 55     S09     I06 -0.21773070       0 -0.36728700 -0.52474846  0.674304750
#> 65     S09     I07 -2.46697618       0 -0.36728700 -0.60867558 -1.491013606
#> 71     S09     I08 -1.71851419       0 -0.36728700 -1.70809505  0.356867854
#> 84     S09     I09 -0.53295713       0 -0.36728700 -0.15835100 -0.007319136
#> 94     S09     I10 -2.56583619       0 -0.36728700 -1.67857591 -0.519973283
#> 10     S10     I01  0.97647849       0  0.08964848  1.41843783 -0.531607818
#> 14     S10     I02 -0.16229397       0  0.08964848  0.09166167 -0.343604117
#> 28     S10     I03  1.47500116       0  0.08964848  0.15001932  1.235333359
#> 40     S10     I04  0.15229758       0  0.08964848  0.86091975 -0.798270660
#> 43     S10     I05  0.38897259       0  0.08964848  0.40760612 -0.108282007
#> 57     S10     I06  1.27019422       0  0.08964848 -0.52474846  1.705294195
#> 70     S10     I07 -1.90511464       0  0.08964848 -0.60867558 -1.386087547
#> 72     S10     I08 -1.25275736       0  0.08964848 -1.70809505  0.365689203
#> 86     S10     I09 -1.35947074       0  0.08964848 -0.15835100 -1.290768225
#> 100    S10     I10 -0.90612682       0  0.08964848 -1.67857591  0.682800605