Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n

the number of subjects

item_n

the number of items

grand_i

the grand intercept (overall mean)

sub_sd

the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd

the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd

the SD of the error term

empirical

Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)
#>     sub_id item_id           y grand_i       sub_i      item_i         err
#> 1      S01     I01 -4.02381730       0 -0.45008419 -2.04579568 -1.52793743
#> 15     S01     I02  0.99088263       0 -0.45008419  0.74670049  0.69426633
#> 29     S01     I03 -0.18325820       0 -0.45008419  1.08351350 -0.81668751
#> 31     S01     I04  0.99219056       0 -0.45008419  0.77276395  0.66951080
#> 44     S01     I05 -1.41440971       0 -0.45008419  0.44743538 -1.41176090
#> 58     S01     I06  1.20219004       0 -0.45008419  1.30340797  0.34886625
#> 61     S01     I07  0.71408847       0 -0.45008419  0.03907525  1.12509741
#> 73     S01     I08  1.42246674       0 -0.45008419  0.54139545  1.33115548
#> 87     S01     I09 -0.05528848       0 -0.45008419 -0.72526772  1.12006343
#> 91     S01     I10 -1.73494262       0 -0.45008419 -1.31340714  0.02854871
#> 7      S02     I01 -3.90149932       0 -1.17290856 -2.04579568 -0.68279508
#> 17     S02     I02 -0.88861522       0 -1.17290856  0.74670049 -0.46240714
#> 30     S02     I03 -1.08353827       0 -1.17290856  1.08351350 -0.99414321
#> 36     S02     I04 -0.57255430       0 -1.17290856  0.77276395 -0.17240968
#> 46     S02     I05 -0.86086452       0 -1.17290856  0.44743538 -0.13539133
#> 59     S02     I06 -0.05187827       0 -1.17290856  1.30340797 -0.18237767
#> 69     S02     I07 -1.71828021       0 -1.17290856  0.03907525 -0.58444689
#> 75     S02     I08 -1.19819241       0 -1.17290856  0.54139545 -0.56667930
#> 88     S02     I09 -2.86939913       0 -1.17290856 -0.72526772 -0.97122285
#> 98     S02     I10 -2.96669693       0 -1.17290856 -1.31340714 -0.48038122
#> 4      S03     I01 -1.48191928       0 -0.47346018 -2.04579568  1.03733658
#> 18     S03     I02  1.62267151       0 -0.47346018  0.74670049  1.34943120
#> 23     S03     I03  0.85188953       0 -0.47346018  1.08351350  0.24183621
#> 33     S03     I04 -0.44728921       0 -0.47346018  0.77276395 -0.74659298
#> 47     S03     I05  1.02094068       0 -0.47346018  0.44743538  1.04696548
#> 60     S03     I06  0.58880568       0 -0.47346018  1.30340797 -0.24114211
#> 62     S03     I07 -1.00568020       0 -0.47346018  0.03907525 -0.57129526
#> 76     S03     I08  1.05803572       0 -0.47346018  0.54139545  0.99010045
#> 89     S03     I09 -1.56442697       0 -0.47346018 -0.72526772 -0.36569908
#> 95     S03     I10 -1.25637920       0 -0.47346018 -1.31340714  0.53048812
#> 5      S04     I01 -0.97350808       0  0.19853334 -2.04579568  0.87375426
#> 19     S04     I02 -1.47794437       0  0.19853334  0.74670049 -2.42317820
#> 21     S04     I03  1.67182821       0  0.19853334  1.08351350  0.38978137
#> 34     S04     I04 -0.41570075       0  0.19853334  0.77276395 -1.38699804
#> 48     S04     I05  0.70144065       0  0.19853334  0.44743538  0.05547193
#> 51     S04     I06  1.71168928       0  0.19853334  1.30340797  0.20974797
#> 63     S04     I07  0.26104824       0  0.19853334  0.03907525  0.02343965
#> 77     S04     I08 -0.21938080       0  0.19853334  0.54139545 -0.95930959
#> 90     S04     I09 -1.76173544       0  0.19853334 -0.72526772 -1.23500107
#> 92     S04     I10  0.31555868       0  0.19853334 -1.31340714  1.43043248
#> 2      S05     I01 -2.43388245       0 -0.01562385 -2.04579568 -0.37246292
#> 16     S05     I02 -0.11126050       0 -0.01562385  0.74670049 -0.84233714
#> 22     S05     I03 -0.22702385       0 -0.01562385  1.08351350 -1.29491350
#> 35     S05     I04  0.99396380       0 -0.01562385  0.77276395  0.23682369
#> 45     S05     I05 -0.47270710       0 -0.01562385  0.44743538 -0.90451863
#> 52     S05     I06  0.73984326       0 -0.01562385  1.30340797 -0.54794087
#> 64     S05     I07 -1.12650006       0 -0.01562385  0.03907525 -1.14995146
#> 74     S05     I08  2.01916724       0 -0.01562385  0.54139545  1.49339564
#> 81     S05     I09 -2.00584455       0 -0.01562385 -0.72526772 -1.26495298
#> 93     S05     I10 -1.02712281       0 -0.01562385 -1.31340714  0.30190818
#> 8      S06     I01 -0.44127814       0  1.22975074 -2.04579568  0.37476680
#> 20     S06     I02  2.03697203       0  1.22975074  0.74670049  0.06052080
#> 27     S06     I03  0.90124283       0  1.22975074  1.08351350 -1.41202141
#> 37     S06     I04  4.33572770       0  1.22975074  0.77276395  2.33321301
#> 49     S06     I05 -0.01778273       0  1.22975074  0.44743538 -1.69496884
#> 56     S06     I06  0.66288606       0  1.22975074  1.30340797 -1.87027265
#> 66     S06     I07  0.60284749       0  1.22975074  0.03907525 -0.66597849
#> 79     S06     I08  0.58230529       0  1.22975074  0.54139545 -1.18884089
#> 85     S06     I09 -0.58009013       0  1.22975074 -0.72526772 -1.08457314
#> 99     S06     I10 -1.55275893       0  1.22975074 -1.31340714 -1.46910252
#> 9      S07     I01 -2.95431322       0 -0.82503247 -2.04579568 -0.08348508
#> 11     S07     I02 -1.14278855       0 -0.82503247  0.74670049 -1.06445657
#> 24     S07     I03  0.71073642       0 -0.82503247  1.08351350  0.45225539
#> 38     S07     I04 -1.21703392       0 -0.82503247  0.77276395 -1.16476540
#> 50     S07     I05 -0.45711794       0 -0.82503247  0.44743538 -0.07952085
#> 53     S07     I06  2.88205569       0 -0.82503247  1.30340797  2.40368018
#> 67     S07     I07 -1.36730832       0 -0.82503247  0.03907525 -0.58135110
#> 80     S07     I08 -0.86360874       0 -0.82503247  0.54139545 -0.57997172
#> 82     S07     I09 -0.85842135       0 -0.82503247 -0.72526772  0.69187884
#> 96     S07     I10 -1.82299631       0 -0.82503247 -1.31340714  0.31544329
#> 6      S08     I01 -4.30759580       0 -2.23232604 -2.04579568 -0.02947408
#> 12     S08     I02  0.09448143       0 -2.23232604  0.74670049  1.58010699
#> 25     S08     I03 -2.57917614       0 -2.23232604  1.08351350 -1.43036360
#> 39     S08     I04 -2.24800090       0 -2.23232604  0.77276395 -0.78843881
#> 41     S08     I05 -1.32896408       0 -2.23232604  0.44743538  0.45592659
#> 54     S08     I06 -2.54776056       0 -2.23232604  1.30340797 -1.61884249
#> 68     S08     I07 -3.25544714       0 -2.23232604  0.03907525 -1.06219634
#> 78     S08     I08  0.68160041       0 -2.23232604  0.54139545  2.37253101
#> 83     S08     I09 -2.66929824       0 -2.23232604 -0.72526772  0.28829552
#> 97     S08     I10 -1.40651748       0 -2.23232604 -1.31340714  2.13921571
#> 3      S09     I01  0.53316465       0  0.67394128 -2.04579568  1.90501904
#> 13     S09     I02  1.51094038       0  0.67394128  0.74670049  0.09029860
#> 26     S09     I03  2.13033449       0  0.67394128  1.08351350  0.37287970
#> 32     S09     I04  1.89024240       0  0.67394128  0.77276395  0.44353716
#> 42     S09     I05  0.44328618       0  0.67394128  0.44743538 -0.67809048
#> 55     S09     I06  0.53471656       0  0.67394128  1.30340797 -1.44263270
#> 65     S09     I07  1.21352410       0  0.67394128  0.03907525  0.50050757
#> 71     S09     I08 -0.22348964       0  0.67394128  0.54139545 -1.43882638
#> 84     S09     I09 -0.46193574       0  0.67394128 -0.72526772 -0.41060931
#> 94     S09     I10 -0.46310586       0  0.67394128 -1.31340714  0.17635999
#> 10     S10     I01 -1.36598789       0 -0.09672313 -2.04579568  0.77653091
#> 14     S10     I02 -0.34714086       0 -0.09672313  0.74670049 -0.99711822
#> 28     S10     I03 -0.81527854       0 -0.09672313  1.08351350 -1.80206891
#> 40     S10     I04  1.46919578       0 -0.09672313  0.77276395  0.79315495
#> 43     S10     I05  3.27739849       0 -0.09672313  0.44743538  2.92668624
#> 57     S10     I06  1.76779585       0 -0.09672313  1.30340797  0.56111100
#> 70     S10     I07 -0.93069088       0 -0.09672313  0.03907525 -0.87304300
#> 72     S10     I08 -0.59891701       0 -0.09672313  0.54139545 -1.04358934
#> 86     S10     I09 -0.23983780       0 -0.09672313 -0.72526772  0.58215304
#> 100    S10     I10 -1.16801978       0 -0.09672313 -1.31340714  0.24211048