Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

Usage

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n

the number of subjects

item_n

the number of items

grand_i

the grand intercept (overall mean)

sub_sd

the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd

the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd

the SD of the error term

empirical

Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed

DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)
#>     sub_id item_id           y grand_i        sub_i      item_i          err
#> 1      S01     I01  0.88353805       0 -0.001028001  0.65517753  0.229388527
#> 15     S01     I02 -1.50091956       0 -0.001028001 -1.95338755  0.453495991
#> 29     S01     I03  0.07328855       0 -0.001028001  0.08885549 -0.014538933
#> 31     S01     I04 -0.28578937       0 -0.001028001 -0.46806097  0.183299594
#> 44     S01     I05  1.07905635       0 -0.001028001 -0.50391394  1.583998291
#> 58     S01     I06 -1.29492761       0 -0.001028001 -0.07328558 -1.220614022
#> 61     S01     I07  0.14184694       0 -0.001028001  0.10540185  0.037473085
#> 73     S01     I08  1.38143269       0 -0.001028001  0.82431425  0.558146449
#> 87     S01     I09 -2.04433153       0 -0.001028001 -2.02684623 -0.016457297
#> 91     S01     I10 -0.87620190       0 -0.001028001  0.49818530 -1.373359200
#> 7      S02     I01  0.31929714       0 -0.172580518  0.65517753 -0.163299868
#> 17     S02     I02 -2.13731687       0 -0.172580518 -1.95338755 -0.011348803
#> 30     S02     I03 -0.78022771       0 -0.172580518  0.08885549 -0.696502681
#> 36     S02     I04  0.53041438       0 -0.172580518 -0.46806097  1.171055862
#> 46     S02     I05 -0.54832557       0 -0.172580518 -0.50391394  0.128168889
#> 59     S02     I06 -0.57740114       0 -0.172580518 -0.07328558 -0.331535039
#> 69     S02     I07 -0.66184222       0 -0.172580518  0.10540185 -0.594663554
#> 75     S02     I08 -0.75428022       0 -0.172580518  0.82431425 -1.406013947
#> 88     S02     I09 -1.04050908       0 -0.172580518 -2.02684623  1.158917666
#> 98     S02     I10  2.61376321       0 -0.172580518  0.49818530  2.288158425
#> 4      S03     I01 -0.12613280       0 -0.554840401  0.65517753 -0.226469920
#> 18     S03     I02 -1.32910691       0 -0.554840401 -1.95338755  1.179121044
#> 23     S03     I03 -0.33873543       0 -0.554840401  0.08885549  0.127249487
#> 33     S03     I04 -0.50610656       0 -0.554840401 -0.46806097  0.516794808
#> 47     S03     I05 -0.12702779       0 -0.554840401 -0.50391394  0.931726552
#> 60     S03     I06 -1.38401754       0 -0.554840401 -0.07328558 -0.755891559
#> 62     S03     I07  1.00149840       0 -0.554840401  0.10540185  1.450936941
#> 76     S03     I08  0.67310566       0 -0.554840401  0.82431425  0.403631810
#> 89     S03     I09 -2.57772463       0 -0.554840401 -2.02684623  0.003962001
#> 95     S03     I10  0.97550952       0 -0.554840401  0.49818530  1.032164621
#> 5      S04     I01  0.07195537       0 -1.484229034  0.65517753  0.901006876
#> 19     S04     I02 -4.35642930       0 -1.484229034 -1.95338755 -0.918812713
#> 21     S04     I03 -1.05309796       0 -1.484229034  0.08885549  0.342275588
#> 34     S04     I04 -3.07250377       0 -1.484229034 -0.46806097 -1.120213769
#> 48     S04     I05 -1.61766351       0 -1.484229034 -0.50391394  0.370479469
#> 51     S04     I06 -1.35177953       0 -1.484229034 -0.07328558  0.205735085
#> 63     S04     I07 -2.72774972       0 -1.484229034  0.10540185 -1.348922537
#> 77     S04     I08 -0.96194164       0 -1.484229034  0.82431425 -0.302026851
#> 90     S04     I09 -3.56093131       0 -1.484229034 -2.02684623 -0.049856039
#> 92     S04     I10  0.93934154       0 -1.484229034  0.49818530  1.925385281
#> 2      S05     I01 -0.71750197       0 -1.327679785  0.65517753 -0.044999711
#> 16     S05     I02 -3.87760407       0 -1.327679785 -1.95338755 -0.596536733
#> 22     S05     I03 -0.79725061       0 -1.327679785  0.08885549  0.441573690
#> 35     S05     I04 -1.57456729       0 -1.327679785 -0.46806097  0.221173463
#> 45     S05     I05 -0.23230526       0 -1.327679785 -0.50391394  1.599288472
#> 52     S05     I06 -0.61753051       0 -1.327679785 -0.07328558  0.783434856
#> 64     S05     I07 -0.36839544       0 -1.327679785  0.10540185  0.853882493
#> 74     S05     I08  0.52476489       0 -1.327679785  0.82431425  1.028130429
#> 81     S05     I09 -3.82698731       0 -1.327679785 -2.02684623 -0.472461288
#> 93     S05     I10 -0.55219837       0 -1.327679785  0.49818530  0.277296115
#> 8      S06     I01  0.70537850       0 -1.296166059  0.65517753  1.346367031
#> 20     S06     I02 -3.77462957       0 -1.296166059 -1.95338755 -0.525075954
#> 27     S06     I03 -0.14652433       0 -1.296166059  0.08885549  1.060786246
#> 37     S06     I04 -2.09882185       0 -1.296166059 -0.46806097 -0.334594826
#> 49     S06     I05 -2.59947240       0 -1.296166059 -0.50391394 -0.799392400
#> 56     S06     I06 -1.92056777       0 -1.296166059 -0.07328558 -0.551116133
#> 66     S06     I07 -1.69883008       0 -1.296166059  0.10540185 -0.508065879
#> 79     S06     I08  1.12042295       0 -1.296166059  0.82431425  1.592274760
#> 85     S06     I09 -3.89081991       0 -1.296166059 -2.02684623 -0.567807623
#> 99     S06     I10 -0.90442030       0 -1.296166059  0.49818530 -0.106439543
#> 9      S07     I01 -1.59556254       0 -0.585691129  0.65517753 -1.665048937
#> 11     S07     I02 -3.00819942       0 -0.585691129 -1.95338755 -0.469120734
#> 24     S07     I03 -0.27395399       0 -0.585691129  0.08885549  0.222881651
#> 38     S07     I04 -2.89610635       0 -0.585691129 -0.46806097 -1.842354252
#> 50     S07     I05 -1.46233488       0 -0.585691129 -0.50391394 -0.372729812
#> 53     S07     I06  0.43161950       0 -0.585691129 -0.07328558  1.090596217
#> 67     S07     I07 -0.93763678       0 -0.585691129  0.10540185 -0.457347505
#> 80     S07     I08 -0.51342146       0 -0.585691129  0.82431425 -0.752044575
#> 82     S07     I09 -1.83212533       0 -0.585691129 -2.02684623  0.780412029
#> 96     S07     I10 -0.64759537       0 -0.585691129  0.49818530 -0.560089534
#> 6      S08     I01  1.64807969       0 -0.458404314  0.65517753  1.451306482
#> 12     S08     I02 -2.25991772       0 -0.458404314 -1.95338755  0.151874152
#> 25     S08     I03  0.26516677       0 -0.458404314  0.08885549  0.634715601
#> 39     S08     I04  0.80311172       0 -0.458404314 -0.46806097  1.729576997
#> 41     S08     I05 -0.84269743       0 -0.458404314 -0.50391394  0.119620832
#> 54     S08     I06 -0.54833963       0 -0.458404314 -0.07328558 -0.016649736
#> 68     S08     I07  0.51177805       0 -0.458404314  0.10540185  0.864780513
#> 78     S08     I08  1.34662081       0 -0.458404314  0.82431425  0.980710881
#> 83     S08     I09 -0.75356840       0 -0.458404314 -2.02684623  1.731682149
#> 97     S08     I10 -0.04416547       0 -0.458404314  0.49818530 -0.083946453
#> 3      S09     I01  1.50473981       0  1.118133108  0.65517753 -0.268570820
#> 13     S09     I02 -1.04481642       0  1.118133108 -1.95338755 -0.209561976
#> 26     S09     I03  0.41453586       0  1.118133108  0.08885549 -0.792452733
#> 32     S09     I04  1.49208379       0  1.118133108 -0.46806097  0.842011645
#> 42     S09     I05  1.45983205       0  1.118133108 -0.50391394  0.845612884
#> 55     S09     I06 -0.05541059       0  1.118133108 -0.07328558 -1.100258111
#> 65     S09     I07  2.37763911       0  1.118133108  0.10540185  1.154104148
#> 71     S09     I08  2.51773568       0  1.118133108  0.82431425  0.575288323
#> 84     S09     I09 -0.34023225       0  1.118133108 -2.02684623  0.568480874
#> 94     S09     I10  0.57916822       0  1.118133108  0.49818530 -1.037150188
#> 10     S10     I01  0.80040353       0 -0.796871985  0.65517753  0.942097986
#> 14     S10     I02 -1.83677658       0 -0.796871985 -1.95338755  0.913482955
#> 28     S10     I03 -0.77282406       0 -0.796871985  0.08885549 -0.064807564
#> 40     S10     I04 -0.22711059       0 -0.796871985 -0.46806097  1.037822357
#> 43     S10     I05 -0.61085599       0 -0.796871985 -0.50391394  0.689929942
#> 57     S10     I06 -0.36950985       0 -0.796871985 -0.07328558  0.500647713
#> 70     S10     I07 -0.99872782       0 -0.796871985  0.10540185 -0.307257686
#> 72     S10     I08  1.06613220       0 -0.796871985  0.82431425  1.038689939
#> 86     S10     I09 -3.34260052       0 -0.796871985 -2.02684623 -0.518882301
#> 100    S10     I10 -1.02408652       0 -0.796871985  0.49818530 -0.725399829