Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id            y grand_i      sub_i     item_i         err
#> 1      S01     I01  0.968901255       0 -0.9170673  0.4930427  1.39292585
#> 15     S01     I02 -1.167369123       0 -0.9170673  0.5508002 -0.80110199
#> 29     S01     I03 -2.015530060       0 -0.9170673 -1.3856475  0.28718479
#> 31     S01     I04  1.263606654       0 -0.9170673  1.5466122  0.63406181
#> 44     S01     I05 -2.453368451       0 -0.9170673 -1.2356505 -0.30065060
#> 58     S01     I06 -0.264373062       0 -0.9170673 -0.5312514  1.18394567
#> 61     S01     I07 -1.403422591       0 -0.9170673 -0.7195197  0.23316446
#> 73     S01     I08 -1.650858581       0 -0.9170673  0.4758174 -1.20960864
#> 87     S01     I09 -1.058520181       0 -0.9170673 -0.3148856  0.17343273
#> 91     S01     I10 -3.059220807       0 -0.9170673 -1.6898239 -0.45232963
#> 7      S02     I01  0.895625856       0 -0.3029440  0.4930427  0.70552714
#> 17     S02     I02 -0.593971365       0 -0.3029440  0.5508002 -0.84182754
#> 30     S02     I03 -0.655150572       0 -0.3029440 -1.3856475  1.03344097
#> 36     S02     I04  1.714470163       0 -0.3029440  1.5466122  0.47080201
#> 46     S02     I05 -2.887768835       0 -0.3029440 -1.2356505 -1.34917429
#> 59     S02     I06 -0.493189014       0 -0.3029440 -0.5312514  0.34100641
#> 69     S02     I07  0.257576489       0 -0.3029440 -0.7195197  1.28004023
#> 75     S02     I08  0.427823051       0 -0.3029440  0.4758174  0.25494968
#> 88     S02     I09 -1.478294925       0 -0.3029440 -0.3148856 -0.86046532
#> 98     S02     I10 -1.248313867       0 -0.3029440 -1.6898239  0.74445400
#> 4      S03     I01 -1.181435030       0 -0.1016664  0.4930427 -1.57281132
#> 18     S03     I02  2.418436293       0 -0.1016664  0.5508002  1.96930255
#> 23     S03     I03 -0.381716102       0 -0.1016664 -1.3856475  1.10559787
#> 33     S03     I04 -0.279979297       0 -0.1016664  1.5466122 -1.72492502
#> 47     S03     I05 -2.638653996       0 -0.1016664 -1.2356505 -1.30133702
#> 60     S03     I06 -0.578501549       0 -0.1016664 -0.5312514  0.05441630
#> 62     S03     I07  0.484056128       0 -0.1016664 -0.7195197  1.30524230
#> 76     S03     I08  0.684164202       0 -0.1016664  0.4758174  0.31001326
#> 89     S03     I09  0.445213585       0 -0.1016664 -0.3148856  0.86176562
#> 95     S03     I10 -2.034011950       0 -0.1016664 -1.6898239 -0.24252165
#> 5      S04     I01 -1.615332674       0 -1.5914114  0.4930427 -0.51696404
#> 19     S04     I02 -2.288961861       0 -1.5914114  0.5508002 -1.24835068
#> 21     S04     I03 -3.223013298       0 -1.5914114 -1.3856475 -0.24595441
#> 34     S04     I04  1.541292133       0 -1.5914114  1.5466122  1.58609133
#> 48     S04     I05 -3.110884349       0 -1.5914114 -1.2356505 -0.28382245
#> 51     S04     I06 -2.415371302       0 -1.5914114 -0.5312514 -0.29270853
#> 63     S04     I07 -2.953754046       0 -1.5914114 -0.7195197 -0.64282295
#> 77     S04     I08 -0.493160195       0 -1.5914114  0.4758174  0.62243379
#> 90     S04     I09 -2.474717517       0 -1.5914114 -0.3148856 -0.56842056
#> 92     S04     I10 -3.877869526       0 -1.5914114 -1.6898239 -0.59663430
#> 2      S05     I01  1.330200372       0 -0.1409557  0.4930427  0.97811339
#> 16     S05     I02  1.855865098       0 -0.1409557  0.5508002  1.44602066
#> 22     S05     I03 -2.308034489       0 -0.1409557 -1.3856475 -0.78143122
#> 35     S05     I04  2.640192980       0 -0.1409557  1.5466122  1.23453656
#> 45     S05     I05  0.867682089       0 -0.1409557 -1.2356505  2.24428837
#> 52     S05     I06  0.087531738       0 -0.1409557 -0.5312514  0.75973889
#> 64     S05     I07 -0.269570058       0 -0.1409557 -0.7195197  0.59090542
#> 74     S05     I08  0.200298047       0 -0.1409557  0.4758174 -0.13456359
#> 81     S05     I09 -1.243160667       0 -0.1409557 -0.3148856 -0.78731933
#> 93     S05     I10 -3.082281771       0 -0.1409557 -1.6898239 -1.25150217
#> 8      S06     I01  0.403580831       0  0.5337733  0.4930427 -0.62323524
#> 20     S06     I02  1.418885617       0  0.5337733  0.5508002  0.33431210
#> 27     S06     I03 -0.705119240       0  0.5337733 -1.3856475  0.14675495
#> 37     S06     I04  1.388928242       0  0.5337733  1.5466122 -0.69145726
#> 49     S06     I05  1.363235806       0  0.5337733 -1.2356505  2.06511300
#> 56     S06     I06 -0.077538552       0  0.5337733 -0.5312514 -0.08006048
#> 66     S06     I07  2.081538584       0  0.5337733 -0.7195197  2.26728498
#> 79     S06     I08  2.591947852       0  0.5337733  0.4758174  1.58235713
#> 85     S06     I09  0.991267142       0  0.5337733 -0.3148856  0.77237940
#> 99     S06     I10 -1.839372685       0  0.5337733 -1.6898239 -0.68332217
#> 9      S07     I01  0.670884965       0  1.6274208  0.4930427 -1.44957854
#> 11     S07     I02  1.914937365       0  1.6274208  0.5508002 -0.26328360
#> 24     S07     I03  1.028666997       0  1.6274208 -1.3856475  0.78689374
#> 38     S07     I04  3.484910259       0  1.6274208  1.5466122  0.31087731
#> 50     S07     I05  0.575875265       0  1.6274208 -1.2356505  0.18410502
#> 53     S07     I06 -0.030639060       0  1.6274208 -0.5312514 -1.12680843
#> 67     S07     I07  1.456385252       0  1.6274208 -0.7195197  0.54848420
#> 80     S07     I08  3.083856067       0  1.6274208  0.4758174  0.98061790
#> 82     S07     I09 -0.446860452       0  1.6274208 -0.3148856 -1.75939564
#> 96     S07     I10  1.552465551       0  1.6274208 -1.6898239  1.61486863
#> 6      S08     I01  0.684102263       0  0.4065431  0.4930427 -0.21548354
#> 12     S08     I02  0.058221041       0  0.4065431  0.5508002 -0.89912221
#> 25     S08     I03 -0.347159646       0  0.4065431 -1.3856475  0.63194481
#> 39     S08     I04  2.231503292       0  0.4065431  1.5466122  0.27834806
#> 41     S08     I05 -1.299992509       0  0.4065431 -1.2356505 -0.47088505
#> 54     S08     I06 -1.024036636       0  0.4065431 -0.5312514 -0.89932830
#> 68     S08     I07 -1.167181185       0  0.4065431 -0.7195197 -0.85420452
#> 78     S08     I08  0.053370325       0  0.4065431  0.4758174 -0.82899013
#> 83     S08     I09  0.861994695       0  0.4065431 -0.3148856  0.77033722
#> 97     S08     I10 -3.663281622       0  0.4065431 -1.6898239 -2.38000084
#> 3      S09     I01 -0.307170556       0  0.7672365  0.4930427 -1.56744978
#> 13     S09     I02  0.580973976       0  0.7672365  0.5508002 -0.73706271
#> 26     S09     I03 -0.883077265       0  0.7672365 -1.3856475 -0.26466624
#> 32     S09     I04  2.707004719       0  0.7672365  1.5466122  0.39315606
#> 42     S09     I05 -0.610957113       0  0.7672365 -1.2356505 -0.14254308
#> 55     S09     I06  1.068942669       0  0.7672365 -0.5312514  0.83295758
#> 65     S09     I07 -0.167031801       0  0.7672365 -0.7195197 -0.21474857
#> 71     S09     I08  1.434419149       0  0.7672365  0.4758174  0.19136527
#> 84     S09     I09  0.045614314       0  0.7672365 -0.3148856 -0.40673659
#> 94     S09     I10  0.006192883       0  0.7672365 -1.6898239  0.92878024
#> 10     S10     I01 -1.524995866       0 -0.9441889  0.4930427 -1.07384968
#> 14     S10     I02 -1.424147676       0 -0.9441889  0.5508002 -1.03075894
#> 28     S10     I03 -2.051219206       0 -0.9441889 -1.3856475  0.27861724
#> 40     S10     I04  0.455136263       0 -0.9441889  1.5466122 -0.14728699
#> 43     S10     I05 -2.528173157       0 -0.9441889 -1.2356505 -0.34833371
#> 57     S10     I06 -1.735610802       0 -0.9441889 -0.5312514 -0.26017048
#> 70     S10     I07  1.394614017       0 -0.9441889 -0.7195197  3.05832267
#> 72     S10     I08  0.955766878       0 -0.9441889  0.4758174  1.42413841
#> 86     S10     I09 -1.594197104       0 -0.9441889 -0.3148856 -0.33512260
#> 100    S10     I10 -1.559135162       0 -0.9441889 -1.6898239  1.07487761