Generate a cross-classified sample

Makes a basic cross-classified design with random intercepts for subjects and items. See vignette("sim_mixed", package = "faux") for examples and details.

sim_mixed_cc(
  sub_n = 100,
  item_n = 20,
  grand_i = 0,
  sub_sd = 1,
  item_sd = 1,
  error_sd = 1,
  empirical = FALSE,
  seed = NULL
)

Arguments

sub_n
the number of subjects

item_n
the number of items

grand_i
the grand intercept (overall mean)

sub_sd
the SD of subject random intercepts (or a sub_n-length named vector of random intercepts for each subject)

item_sd
the SD of item random intercepts (or an item_n-length named vector of random intercepts for each item)

error_sd
the SD of the error term

empirical
Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)

seed
DEPRECATED use set.seed() instead before running this function

Value

a tbl

Examples

sim_mixed_cc(10, 10)

#>     sub_id item_id           y grand_i      sub_i     item_i          err
#> 1      S01     I01  1.50975044       0  0.9459906  1.2267632 -0.663003340
#> 15     S01     I02  3.34865203       0  0.9459906  2.2991934  0.103468033
#> 29     S01     I03  1.06185632       0  0.9459906 -0.4075220  0.523387707
#> 31     S01     I04  3.27131968       0  0.9459906 -0.0444485  2.369777586
#> 44     S01     I05  0.24534356       0  0.9459906 -0.3770660 -0.323581056
#> 58     S01     I06  1.05528061       0  0.9459906 -0.8526325  0.961922473
#> 61     S01     I07  1.07252911       0  0.9459906  0.4260213 -0.299482773
#> 73     S01     I08  0.94860648       0  0.9459906 -0.2758645  0.278480353
#> 87     S01     I09 -0.06503802       0  0.9459906 -1.9029931  0.891964448
#> 91     S01     I10 -0.65723195       0  0.9459906 -0.4911469 -1.112075674
#> 7      S02     I01  3.10668813       0 -0.7477811  1.2267632  2.627706060
#> 17     S02     I02  3.01729350       0 -0.7477811  2.2991934  1.465881221
#> 30     S02     I03 -0.76488585       0 -0.7477811 -0.4075220  0.390417244
#> 36     S02     I04  1.07875575       0 -0.7477811 -0.0444485  1.870985369
#> 46     S02     I05 -0.65609405       0 -0.7477811 -0.3770660  0.468753050
#> 59     S02     I06 -1.32574262       0 -0.7477811 -0.8526325  0.274670956
#> 69     S02     I07 -0.20146025       0 -0.7477811  0.4260213  0.120299581
#> 75     S02     I08  0.14681469       0 -0.7477811 -0.2758645  1.170460279
#> 88     S02     I09 -4.33632876       0 -0.7477811 -1.9029931 -1.685554576
#> 98     S02     I10 -1.46905891       0 -0.7477811 -0.4911469 -0.230130910
#> 4      S03     I01  1.51441989       0 -0.4379911  1.2267632  0.725647762
#> 18     S03     I02  1.93609829       0 -0.4379911  2.2991934  0.074895952
#> 23     S03     I03 -1.14604714       0 -0.4379911 -0.4075220 -0.300534098
#> 33     S03     I04 -1.22278673       0 -0.4379911 -0.0444485 -0.740347166
#> 47     S03     I05 -0.98522094       0 -0.4379911 -0.3770660 -0.170163900
#> 60     S03     I06 -0.59829971       0 -0.4379911 -0.8526325  0.692323808
#> 62     S03     I07 -0.60290123       0 -0.4379911  0.4260213 -0.590931454
#> 76     S03     I08 -0.69032885       0 -0.4379911 -0.2758645  0.023526678
#> 89     S03     I09 -2.09782122       0 -0.4379911 -1.9029931  0.243162902
#> 95     S03     I10 -0.29661490       0 -0.4379911 -0.4911469  0.632523038
#> 5      S04     I01  2.22246017       0  0.6438773  1.2267632  0.351819725
#> 19     S04     I02  0.04705236       0  0.6438773  2.2991934 -2.896018301
#> 21     S04     I03 -0.51611000       0  0.6438773 -0.4075220 -0.752465271
#> 34     S04     I04  0.31705391       0  0.6438773 -0.0444485 -0.282374841
#> 48     S04     I05  1.07317884       0  0.6438773 -0.3770660  0.806367574
#> 51     S04     I06  2.84413450       0  0.6438773 -0.8526325  3.052889709
#> 63     S04     I07  0.35751519       0  0.6438773  0.4260213 -0.712383347
#> 77     S04     I08  0.11803291       0  0.6438773 -0.2758645 -0.249979876
#> 90     S04     I09 -3.24011064       0  0.6438773 -1.9029931 -1.980994828
#> 92     S04     I10 -0.49351843       0  0.6438773 -0.4911469 -0.646248810
#> 2      S05     I01  2.34418076       0  1.2248338  1.2267632 -0.107416267
#> 16     S05     I02  3.82046186       0  1.2248338  2.2991934  0.296434615
#> 22     S05     I03 -0.49028249       0  1.2248338 -0.4075220 -1.307594356
#> 35     S05     I04  2.44813011       0  1.2248338 -0.0444485  1.267744775
#> 45     S05     I05  0.11777551       0  1.2248338 -0.3770660 -0.729992344
#> 52     S05     I06 -1.82246476       0  1.2248338 -0.8526325 -2.194666139
#> 64     S05     I07  2.47952279       0  1.2248338  0.4260213  0.828667662
#> 74     S05     I08  0.57132487       0  1.2248338 -0.2758645 -0.377644504
#> 81     S05     I09  0.66253071       0  1.2248338 -1.9029931  1.340689934
#> 93     S05     I10  2.11167290       0  1.2248338 -0.4911469  1.377985935
#> 8      S06     I01  1.17639447       0  1.0738344  1.2267632 -1.124203116
#> 20     S06     I02  2.50047658       0  1.0738344  2.2991934 -0.872551230
#> 27     S06     I03  1.44826704       0  1.0738344 -0.4075220  0.781954612
#> 37     S06     I04  1.35506310       0  1.0738344 -0.0444485  0.325677202
#> 49     S06     I05 -0.95870156       0  1.0738344 -0.3770660 -1.655469978
#> 56     S06     I06 -1.00862495       0  1.0738344 -0.8526325 -1.229826895
#> 66     S06     I07  3.08050255       0  1.0738344  0.4260213  1.580646864
#> 79     S06     I08  1.20135174       0  1.0738344 -0.2758645  0.403381802
#> 85     S06     I09 -0.55572258       0  1.0738344 -1.9029931  0.273436078
#> 99     S06     I10  1.14887107       0  1.0738344 -0.4911469  0.566183546
#> 9      S07     I01  1.14928314       0 -1.0922379  1.2267632  1.014757900
#> 11     S07     I02  0.21351758       0 -1.0922379  2.2991934 -0.993437874
#> 24     S07     I03 -1.42405653       0 -1.0922379 -0.4075220  0.075703396
#> 38     S07     I04 -0.39716095       0 -1.0922379 -0.0444485  0.739525496
#> 50     S07     I05 -1.90897076       0 -1.0922379 -0.3770660 -0.439666830
#> 53     S07     I06 -1.59931309       0 -1.0922379 -0.8526325  0.345557310
#> 67     S07     I07 -0.93206963       0 -1.0922379  0.4260213 -0.265852974
#> 80     S07     I08 -2.30089466       0 -1.0922379 -0.2758645 -0.932792249
#> 82     S07     I09 -1.86789638       0 -1.0922379 -1.9029931  1.127334635
#> 96     S07     I10 -2.51787910       0 -1.0922379 -0.4911469 -0.934494277
#> 6      S08     I01  2.11712777       0  0.2620193  1.2267632  0.628345305
#> 12     S08     I02  1.96305868       0  0.2620193  2.2991934 -0.598154004
#> 25     S08     I03  0.26684146       0  0.2620193 -0.4075220  0.412344156
#> 39     S08     I04  1.28811883       0  0.2620193 -0.0444485  1.070548056
#> 41     S08     I05  0.35968392       0  0.2620193 -0.3770660  0.474730623
#> 54     S08     I06 -0.40792322       0  0.2620193 -0.8526325  0.182689963
#> 68     S08     I07  0.92033021       0  0.2620193  0.4260213  0.232289640
#> 78     S08     I08 -1.56896500       0  0.2620193 -0.2758645 -1.555119806
#> 83     S08     I09 -0.91665284       0  0.2620193 -1.9029931  0.724320946
#> 97     S08     I10 -0.69362383       0  0.2620193 -0.4911469 -0.464496234
#> 3      S09     I01  1.43533160       0  1.1540355  1.2267632 -0.945467086
#> 13     S09     I02  4.38986387       0  1.1540355  2.2991934  0.936634969
#> 26     S09     I03 -0.94223008       0  1.1540355 -0.4075220 -1.688743607
#> 32     S09     I04 -0.12632769       0  1.1540355 -0.0444485 -1.235914691
#> 42     S09     I05 -1.83056565       0  1.1540355 -0.3770660 -2.607535166
#> 55     S09     I06  0.47936041       0  1.1540355 -0.8526325  0.177957371
#> 65     S09     I07  1.58483755       0  1.1540355  0.4260213  0.004780766
#> 71     S09     I08 -0.04791947       0  1.1540355 -0.2758645 -0.926090507
#> 84     S09     I09 -1.48507083       0  1.1540355 -1.9029931 -0.736113265
#> 94     S09     I10  1.26884677       0  1.1540355 -0.4911469  0.605958143
#> 10     S10     I01  2.20714989       0  0.3603791  1.2267632  0.620007589
#> 14     S10     I02  1.93975801       0  0.3603791  2.2991934 -0.719814509
#> 28     S10     I03 -0.16225694       0  0.3603791 -0.4075220 -0.115114079
#> 40     S10     I04  1.52370681       0  0.3603791 -0.0444485  1.207776205
#> 43     S10     I05  0.06851119       0  0.3603791 -0.3770660  0.085198066
#> 57     S10     I06 -1.80244988       0  0.3603791 -0.8526325 -1.310196532
#> 70     S10     I07  0.70031259       0  0.3603791  0.4260213 -0.086087812
#> 72     S10     I08  1.04140579       0  0.3603791 -0.2758645  0.956891146
#> 86     S10     I09 -1.06530802       0  0.3603791 -1.9029931  0.477305936
#> 100    S10     I10 -0.83847795       0  0.3603791 -0.4911469 -0.707710183