Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id           y  grand_i       sub_i      item_i          err
#> 1      S01     I01  3.94305992 3.018906  0.12679876 -1.05893894  1.856294482
#> 2      S02     I01  3.98074785 3.018906  1.26228832 -1.05893894  0.758492857
#> 3      S03     I01  3.70035974 3.018906  0.45200425 -1.05893894  1.288388816
#> 4      S04     I01  0.69436427 3.018906 -0.01602987 -1.05893894 -1.249572537
#> 5      S05     I01  0.60574566 3.018906 -0.81404617 -1.05893894 -0.540174838
#> 6      S06     I01  0.64480347 3.018906  0.02687097 -1.05893894 -1.342034174
#> 7      S07     I01  3.61980637 3.018906 -0.30229234 -1.05893894  1.962132033
#> 8      S08     I01  1.77409937 3.018906 -0.14476387 -1.05893894 -0.041103441
#> 9      S09     I01  4.84373029 3.018906  1.25398944 -1.05893894  1.629774179
#> 10     S10     I01  1.67375002 3.018906 -0.78765099 -1.05893894  0.501434330
#> 11     S01     I02  4.02498387 3.018906  0.12679876  0.69756493  0.181714568
#> 12     S02     I02  3.89812901 3.018906  1.26228832  0.69756493 -1.080629846
#> 13     S03     I02  2.92249263 3.018906  0.45200425  0.69756493 -1.245982162
#> 14     S04     I02  4.60537523 3.018906 -0.01602987  0.69756493  0.904934567
#> 15     S05     I02  4.50437651 3.018906 -0.81404617  0.69756493  1.601952148
#> 16     S06     I02  3.93101298 3.018906  0.02687097  0.69756493  0.187671471
#> 17     S07     I02  5.00745684 3.018906 -0.30229234  0.69756493  1.593278643
#> 18     S08     I02  5.60895271 3.018906 -0.14476387  0.69756493  2.037246039
#> 19     S09     I02  6.19370993 3.018906  1.25398944  0.69756493  1.223249957
#> 20     S10     I02  3.24011828 3.018906 -0.78765099  0.69756493  0.311298725
#> 21     S01     I03  2.28045879 3.018906  0.12679876 -0.66507868 -0.200166897
#> 22     S02     I03  5.26379561 3.018906  1.26228832 -0.66507868  1.647680367
#> 23     S03     I03  2.56526974 3.018906  0.45200425 -0.66507868 -0.240561438
#> 24     S04     I03  5.59807739 3.018906 -0.01602987 -0.66507868  3.260280335
#> 25     S05     I03  1.48236748 3.018906 -0.81404617 -0.66507868 -0.057413274
#> 26     S06     I03  2.81241889 3.018906  0.02687097 -0.66507868  0.431720989
#> 27     S07     I03  0.79512758 3.018906 -0.30229234 -0.66507868 -1.256407009
#> 28     S08     I03  1.97155265 3.018906 -0.14476387 -0.66507868 -0.237510416
#> 29     S09     I03  1.19596265 3.018906  1.25398944 -0.66507868 -2.411853715
#> 30     S10     I03  3.42066726 3.018906 -0.78765099 -0.66507868  1.854491315
#> 31     S01     I04  4.17686235 3.018906  0.12679876 -1.07593953  2.107097506
#> 32     S02     I04  2.22241226 3.018906  1.26228832 -1.07593953 -0.982842148
#> 33     S03     I04 -0.01535653 3.018906  0.45200425 -1.07593953 -2.410326871
#> 34     S04     I04  3.58892739 3.018906 -0.01602987 -1.07593953  1.661991170
#> 35     S05     I04  1.35866263 3.018906 -0.81404617 -1.07593953  0.229742719
#> 36     S06     I04  1.43753464 3.018906  0.02687097 -1.07593953 -0.532302419
#> 37     S07     I04  1.26616001 3.018906 -0.30229234 -1.07593953 -0.374513741
#> 38     S08     I04  2.00989620 3.018906 -0.14476387 -1.07593953  0.211693975
#> 39     S09     I04  2.92659753 3.018906  1.25398944 -1.07593953 -0.270357997
#> 40     S10     I04  0.37262155 3.018906 -0.78765099 -1.07593953 -0.782693554
#> 41     S01     I05  1.36370919 3.018906  0.12679876 -0.22744882 -1.554546359
#> 42     S02     I05  6.19742840 3.018906  1.26228832 -0.22744882  2.143683288
#> 43     S03     I05  3.30482546 3.018906  0.45200425 -0.22744882  0.061364417
#> 44     S04     I05  1.43532610 3.018906 -0.01602987 -0.22744882 -1.340100823
#> 45     S05     I05  1.75126700 3.018906 -0.81404617 -0.22744882 -0.226143622
#> 46     S06     I05  1.34339219 3.018906  0.02687097 -0.22744882 -1.474935574
#> 47     S07     I05  4.76292340 3.018906 -0.30229234 -0.22744882  2.273758942
#> 48     S08     I05  2.54033821 3.018906 -0.14476387 -0.22744882 -0.106354715
#> 49     S09     I05  4.84260966 3.018906  1.25398944 -0.22744882  0.797163432
#> 50     S10     I05  0.70790753 3.018906 -0.78765099 -0.22744882 -1.295898275
#> 51     S01     I06  4.06986983 3.018906  0.12679876  0.01167687  0.912488583
#> 52     S02     I06  3.19129522 3.018906  1.26228832  0.01167687 -1.101575585
#> 53     S03     I06  3.69003984 3.018906  0.45200425  0.01167687  0.207453101
#> 54     S04     I06  3.82513296 3.018906 -0.01602987  0.01167687  0.810580347
#> 55     S05     I06  2.60349474 3.018906 -0.81404617  0.01167687  0.386958424
#> 56     S06     I06  2.95181026 3.018906  0.02687097  0.01167687 -0.105643196
#> 57     S07     I06  3.41850535 3.018906 -0.30229234  0.01167687  0.690215201
#> 58     S08     I06  3.58577995 3.018906 -0.14476387  0.01167687  0.699961329
#> 59     S09     I06  3.51865758 3.018906  1.25398944  0.01167687 -0.765914342
#> 60     S10     I06  2.20376235 3.018906 -0.78765099  0.01167687 -0.039169150
#> 61     S01     I07  4.07726296 3.018906  0.12679876  0.71360460  0.217953984
#> 62     S02     I07  6.52212135 3.018906  1.26228832  0.71360460  1.527322816
#> 63     S03     I07  4.19169638 3.018906  0.45200425  0.71360460  0.007181909
#> 64     S04     I07  3.59628644 3.018906 -0.01602987  0.71360460 -0.120193901
#> 65     S05     I07  2.86079107 3.018906 -0.81404617  0.71360460 -0.057672968
#> 66     S06     I07  5.62190058 3.018906  0.02687097  0.71360460  1.862519401
#> 67     S07     I07  4.99681430 3.018906 -0.30229234  0.71360460  1.566596429
#> 68     S08     I07  3.15416825 3.018906 -0.14476387  0.71360460 -0.433578100
#> 69     S09     I07  5.70899721 3.018906  1.25398944  0.71360460  0.722497562
#> 70     S10     I07  2.54858745 3.018906 -0.78765099  0.71360460 -0.396271780
#> 71     S01     I08  5.93887220 3.018906  0.12679876  0.85798181  1.935186018
#> 72     S02     I08  5.42296318 3.018906  1.26228832  0.85798181  0.283787434
#> 73     S03     I08  3.86096247 3.018906  0.45200425  0.85798181 -0.467929204
#> 74     S04     I08  3.59732397 3.018906 -0.01602987  0.85798181 -0.263533580
#> 75     S05     I08  2.35702283 3.018906 -0.81404617  0.85798181 -0.705818415
#> 76     S06     I08  4.55124916 3.018906  0.02687097  0.85798181  0.647490770
#> 77     S07     I08  1.76008115 3.018906 -0.30229234  0.85798181 -1.814513929
#> 78     S08     I08  6.13970132 3.018906 -0.14476387  0.85798181  2.407577757
#> 79     S09     I08  4.57677369 3.018906  1.25398944  0.85798181 -0.554103169
#> 80     S10     I08  2.12572668 3.018906 -0.78765099  0.85798181 -0.963509760
#> 81     S01     I09  1.12467805 3.018906  0.12679876 -0.29896859 -1.722057728
#> 82     S02     I09  4.12806545 3.018906  1.26228832 -0.29896859  0.145840113
#> 83     S03     I09  2.20958135 3.018906  0.45200425 -0.29896859 -0.962359921
#> 84     S04     I09  3.27452023 3.018906 -0.01602987 -0.29896859  0.570613084
#> 85     S05     I09  1.57905267 3.018906 -0.81404617 -0.29896859 -0.326838175
#> 86     S06     I09  2.68350804 3.018906  0.02687097 -0.29896859 -0.063299945
#> 87     S07     I09  3.79540793 3.018906 -0.30229234 -0.29896859  1.377763247
#> 88     S08     I09  4.23096100 3.018906 -0.14476387 -0.29896859  1.655787847
#> 89     S09     I09  4.21115934 3.018906  1.25398944 -0.29896859  0.237232882
#> 90     S10     I09  2.43138070 3.018906 -0.78765099 -0.29896859  0.499094659
#> 91     S01     I10  4.36425644 3.018906  0.12679876  0.58564097  0.632911092
#> 92     S02     I10  3.45920320 3.018906  1.26228832  0.58564097 -1.407631702
#> 93     S03     I10  5.33899388 3.018906  0.45200425  0.58564097  1.282443044
#> 94     S04     I10  4.75749417 3.018906 -0.01602987  0.58564097  1.168977453
#> 95     S05     I10  3.20168327 3.018906 -0.81404617  0.58564097  0.411182854
#> 96     S06     I10  3.44772946 3.018906  0.02687097  0.58564097 -0.183688095
#> 97     S07     I10  3.59396648 3.018906 -0.30229234  0.58564097  0.291712229
#> 98     S08     I10  2.25286624 3.018906 -0.14476387  0.58564097 -1.206916477
#> 99     S09     I10  6.73485964 3.018906  1.25398944  0.58564097  1.876323616
#> 100    S10     I10  4.23594840 3.018906 -0.78765099  0.58564097  1.419052799