Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id         y  grand_i       sub_i      item_i          err
#> 1      S01     I01 5.5783983 3.018906 -0.13605091  0.83705227  1.858491327
#> 15     S01     I02 3.2125868 3.018906 -0.13605091 -0.75212982  1.081861907
#> 29     S01     I03 3.5990958 3.018906 -0.13605091  0.28463261  0.431608519
#> 31     S01     I04 2.2717703 3.018906 -0.13605091  0.65041613 -1.261500535
#> 44     S01     I05 3.4765330 3.018906 -0.13605091  0.13645406  0.457224204
#> 58     S01     I06 1.9131079 3.018906 -0.13605091 -0.07543029 -0.894316475
#> 61     S01     I07 3.8989441 3.018906 -0.13605091 -0.14689739  1.162986743
#> 73     S01     I08 1.5162663 3.018906 -0.13605091 -1.12003412 -0.246554258
#> 87     S01     I09 2.8657935 3.018906 -0.13605091  0.76356729 -0.780628445
#> 91     S01     I10 2.7618638 3.018906 -0.13605091 -0.55371548  0.432724599
#> 7      S02     I01 6.3578740 3.018906  0.66187772  0.83705227  1.840038401
#> 17     S02     I02 2.7105283 3.018906  0.66187772 -0.75212982 -0.218125265
#> 30     S02     I03 4.2007583 3.018906  0.66187772  0.28463261  0.235342382
#> 36     S02     I04 4.9076222 3.018906  0.66187772  0.65041613  0.576422730
#> 46     S02     I05 3.4343825 3.018906  0.66187772  0.13645406 -0.382854863
#> 59     S02     I06 0.7623946 3.018906  0.66187772 -0.07543029 -2.842958483
#> 69     S02     I07 1.6422306 3.018906  0.66187772 -0.14689739 -1.891655296
#> 75     S02     I08 3.5460805 3.018906  0.66187772 -1.12003412  0.985331290
#> 88     S02     I09 3.4076611 3.018906  0.66187772  0.76356729 -1.036689478
#> 98     S02     I10 2.8673507 3.018906  0.66187772 -0.55371548 -0.259717122
#> 4      S03     I01 5.0402842 3.018906  0.02215492  0.83705227  1.162171402
#> 18     S03     I02 3.7375203 3.018906  0.02215492 -0.75212982  1.448589625
#> 23     S03     I03 3.3163340 3.018906  0.02215492  0.28463261 -0.009359188
#> 33     S03     I04 3.3135786 3.018906  0.02215492  0.65041613 -0.377898052
#> 47     S03     I05 3.7190453 3.018906  0.02215492  0.13645406  0.541530720
#> 60     S03     I06 1.4802505 3.018906  0.02215492 -0.07543029 -1.485379769
#> 62     S03     I07 4.0277408 3.018906  0.02215492 -0.14689739  1.133577642
#> 76     S03     I08 2.5399304 3.018906  0.02215492 -1.12003412  0.618903997
#> 89     S03     I09 4.2776503 3.018906  0.02215492  0.76356729  0.473022471
#> 95     S03     I10 1.4111871 3.018906  0.02215492 -0.55371548 -1.076157968
#> 5      S04     I01 4.6934530 3.018906  0.05300989  0.83705227  0.784485181
#> 19     S04     I02 2.8066360 3.018906  0.05300989 -0.75212982  0.486850304
#> 21     S04     I03 5.3389873 3.018906  0.05300989  0.28463261  1.982439188
#> 34     S04     I04 4.1022663 3.018906  0.05300989  0.65041613  0.379934717
#> 48     S04     I05 1.6565386 3.018906  0.05300989  0.13645406 -1.551830916
#> 51     S04     I06 3.7216447 3.018906  0.05300989 -0.07543029  0.725159477
#> 63     S04     I07 3.3558339 3.018906  0.05300989 -0.14689739  0.430815837
#> 77     S04     I08 2.6924614 3.018906  0.05300989 -1.12003412  0.740579989
#> 90     S04     I09 3.7237941 3.018906  0.05300989  0.76356729 -0.111688681
#> 92     S04     I10 2.9878398 3.018906  0.05300989 -0.55371548  0.469639778
#> 2      S05     I01 5.1304681 3.018906  0.76161944  0.83705227  0.512890807
#> 16     S05     I02 4.0050122 3.018906  0.76161944 -0.75212982  0.976616993
#> 22     S05     I03 3.4994735 3.018906  0.76161944  0.28463261 -0.565684163
#> 35     S05     I04 5.5759749 3.018906  0.76161944  0.65041613  1.145033723
#> 45     S05     I05 3.3741627 3.018906  0.76161944  0.13645406 -0.542816453
#> 52     S05     I06 4.2059437 3.018906  0.76161944 -0.07543029  0.500848967
#> 64     S05     I07 1.0569694 3.018906  0.76161944 -0.14689739 -2.576658256
#> 74     S05     I08 3.2480176 3.018906  0.76161944 -1.12003412  0.587526654
#> 81     S05     I09 4.1517986 3.018906  0.76161944  0.76356729 -0.392293749
#> 93     S05     I10 2.0168638 3.018906  0.76161944 -0.55371548 -1.209945803
#> 8      S06     I01 5.1687377 3.018906  1.25404463  0.83705227  0.058735212
#> 20     S06     I02 4.2334525 3.018906  1.25404463 -0.75212982  0.712632047
#> 27     S06     I03 4.5926846 3.018906  1.25404463  0.28463261  0.035101698
#> 37     S06     I04 3.4264084 3.018906  1.25404463  0.65041613 -1.496957973
#> 49     S06     I05 4.0157728 3.018906  1.25404463  0.13645406 -0.393631458
#> 56     S06     I06 5.1358623 3.018906  1.25404463 -0.07543029  0.938342302
#> 66     S06     I07 3.7201498 3.018906  1.25404463 -0.14689739 -0.405903043
#> 79     S06     I08 2.6308341 3.018906  1.25404463 -1.12003412 -0.522082018
#> 85     S06     I09 5.1634861 3.018906  1.25404463  0.76356729  0.126968571
#> 99     S06     I10 2.9304679 3.018906  1.25404463 -0.55371548 -0.788766836
#> 9      S07     I01 5.9338370 3.018906  0.93814525  0.83705227  1.139733891
#> 11     S07     I02 4.3359763 3.018906  0.93814525 -0.75212982  1.131055259
#> 24     S07     I03 6.2483589 3.018906  0.93814525  0.28463261  2.006675465
#> 38     S07     I04 4.9026907 3.018906  0.93814525  0.65041613  0.295223750
#> 50     S07     I05 5.5463688 3.018906  0.93814525  0.13645406  1.452863866
#> 53     S07     I06 4.5476532 3.018906  0.93814525 -0.07543029  0.666032623
#> 67     S07     I07 2.2943716 3.018906  0.93814525 -0.14689739 -1.515781873
#> 80     S07     I08 0.3605633 3.018906  0.93814525 -1.12003412 -2.476453458
#> 82     S07     I09 2.8967961 3.018906  0.93814525  0.76356729 -1.823822017
#> 96     S07     I10 4.1245185 3.018906  0.93814525 -0.55371548  0.721183078
#> 6      S08     I01 4.0195760 3.018906 -0.33155043  0.83705227  0.495168502
#> 12     S08     I02 1.3465400 3.018906 -0.33155043 -0.75212982 -0.588685333
#> 25     S08     I03 2.6314249 3.018906 -0.33155043  0.28463261 -0.340562903
#> 39     S08     I04 5.6867139 3.018906 -0.33155043  0.65041613  2.348942620
#> 41     S08     I05 2.2036669 3.018906 -0.33155043  0.13645406 -0.620142355
#> 54     S08     I06 2.4106152 3.018906 -0.33155043 -0.07543029 -0.201309738
#> 68     S08     I07 2.5878772 3.018906 -0.33155043 -0.14689739  0.047419430
#> 78     S08     I08 1.7615471 3.018906 -0.33155043 -1.12003412  0.194226081
#> 83     S08     I09 2.5436619 3.018906 -0.33155043  0.76356729 -0.907260599
#> 97     S08     I10 1.2061660 3.018906 -0.33155043 -0.55371548 -0.927473652
#> 3      S09     I01 5.1377342 3.018906  0.37086373  0.83705227  0.910912611
#> 13     S09     I02 2.8748637 3.018906  0.37086373 -0.75212982  0.237224209
#> 26     S09     I03 3.5222314 3.018906  0.37086373  0.28463261 -0.152170563
#> 32     S09     I04 5.7838047 3.018906  0.37086373  0.65041613  1.743619239
#> 42     S09     I05 5.1650516 3.018906  0.37086373  0.13645406  1.638828189
#> 55     S09     I06 3.7001798 3.018906  0.37086373 -0.07543029  0.385840717
#> 65     S09     I07 2.8041221 3.018906  0.37086373 -0.14689739 -0.438749831
#> 71     S09     I08 3.1534733 3.018906  0.37086373 -1.12003412  0.883738042
#> 84     S09     I09 4.4765889 3.018906  0.37086373  0.76356729  0.323252241
#> 94     S09     I10 2.2734220 3.018906  0.37086373 -0.55371548 -0.562631878
#> 10     S10     I01 4.1494860 3.018906  1.91483725  0.83705227 -1.621309123
#> 14     S10     I02 3.8089716 3.018906  1.91483725 -0.75212982 -0.372641485
#> 28     S10     I03 5.4823412 3.018906  1.91483725  0.28463261  0.263965689
#> 40     S10     I04 7.0397013 3.018906  1.91483725  0.65041613  1.455542273
#> 43     S10     I05 4.9260855 3.018906  1.91483725  0.13645406 -0.144111409
#> 57     S10     I06 5.1902714 3.018906  1.91483725 -0.07543029  0.331958862
#> 70     S10     I07 6.2104586 3.018906  1.91483725 -0.14689739  1.423613156
#> 72     S10     I08 4.5256935 3.018906  1.91483725 -1.12003412  0.711984727
#> 86     S10     I09 4.6054083 3.018906  1.91483725  0.76356729 -1.091901894
#> 100    S10     I10 4.5496066 3.018906  1.91483725 -0.55371548  0.169579216