Generate a sample with random intercepts for subjects and items

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(.data, sub_n = NULL, item_n = NULL, dv = 1,
  sub_id = 2, item_id = 3, seed = NULL)

Arguments

.data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id       val  grand_i       sub_i      item_i          err
#> 1      S01     I01 2.7036152 3.018906  0.02805978  0.27344359 -0.616793775
#> 2      S02     I01 2.5862347 3.018906 -0.03908504  0.27344359 -0.667029511
#> 3      S03     I01 4.2063565 3.018906  0.42565642  0.27344359  0.488350908
#> 4      S04     I01 3.6627835 3.018906  0.02611968  0.27344359  0.344314657
#> 5      S05     I01 5.6149842 3.018906  0.15323300  0.27344359  2.169401966
#> 6      S06     I01 3.6144259 3.018906  0.12860568  0.27344359  0.193471044
#> 7      S07     I01 1.6368591 3.018906 -0.72484206  0.27344359 -0.930648018
#> 8      S08     I01 2.7220111 3.018906 -0.01945261  0.27344359 -0.550885477
#> 9      S09     I01 2.9053181 3.018906  0.75503061  0.27344359 -1.142061709
#> 10     S10     I01 4.6585777 3.018906  0.73323587  0.27344359  0.632992648
#> 11     S01     I02 4.0842736 3.018906  0.02805978 -0.06099718  1.098305364
#> 12     S02     I02 3.0659766 3.018906 -0.03908504 -0.06099718  0.147153202
#> 13     S03     I02 2.2008334 3.018906  0.42565642 -0.06099718 -1.182731493
#> 14     S04     I02 2.5931382 3.018906  0.02611968 -0.06099718 -0.390889867
#> 15     S05     I02 4.1553344 3.018906  0.15323300 -0.06099718  1.044193013
#> 16     S06     I02 2.3187857 3.018906  0.12860568 -0.06099718 -0.767728408
#> 17     S07     I02 3.9856749 3.018906 -0.72484206 -0.06099718  1.752608527
#> 18     S08     I02 0.3812322 3.018906 -0.01945261 -0.06099718 -2.557223584
#> 19     S09     I02 5.2537160 3.018906  0.75503061 -0.06099718  1.540776922
#> 20     S10     I02 1.2813724 3.018906  0.73323587 -0.06099718 -2.409771916
#> 21     S01     I03 0.8487761 3.018906  0.02805978  0.42032225 -2.618511556
#> 22     S02     I03 3.6267403 3.018906 -0.03908504  0.42032225  0.226597523
#> 23     S03     I03 3.4892128 3.018906  0.42565642  0.42032225 -0.375671448
#> 24     S04     I03 3.0262226 3.018906  0.02611968  0.42032225 -0.439124917
#> 25     S05     I03 4.3486330 3.018906  0.15323300  0.42032225  0.756172148
#> 26     S06     I03 3.0738084 3.018906  0.12860568  0.42032225 -0.494025153
#> 27     S07     I03 3.9531837 3.018906 -0.72484206  0.42032225  1.238797911
#> 28     S08     I03 3.5478802 3.018906 -0.01945261  0.42032225  0.128104974
#> 29     S09     I03 4.1658232 3.018906  0.75503061  0.42032225 -0.028435234
#> 30     S10     I03 2.4455711 3.018906  0.73323587  0.42032225 -1.726892586
#> 31     S01     I04 1.0395447 3.018906  0.02805978 -1.00181535 -1.005605371
#> 32     S02     I04 1.5107225 3.018906 -0.03908504 -1.00181535 -0.467282731
#> 33     S03     I04 3.3432786 3.018906  0.42565642 -1.00181535  0.900531879
#> 34     S04     I04 1.4320289 3.018906  0.02611968 -1.00181535 -0.611180996
#> 35     S05     I04 1.9145837 3.018906  0.15323300 -1.00181535 -0.255739608
#> 36     S06     I04 1.4980747 3.018906  0.12860568 -1.00181535 -0.647621279
#> 37     S07     I04 0.6689404 3.018906 -0.72484206 -1.00181535 -0.623307851
#> 38     S08     I04 1.6261217 3.018906 -0.01945261 -1.00181535 -0.371515935
#> 39     S09     I04 2.8615743 3.018906  0.75503061 -1.00181535  0.089453453
#> 40     S10     I04 2.4251568 3.018906  0.73323587 -1.00181535 -0.325169325
#> 41     S01     I05 1.5789940 3.018906  0.02805978  0.07673155 -1.544702983
#> 42     S02     I05 3.2765270 3.018906 -0.03908504  0.07673155  0.219974831
#> 43     S03     I05 4.2559120 3.018906  0.42565642  0.07673155  0.734618397
#> 44     S04     I05 3.7076244 3.018906  0.02611968  0.07673155  0.585867522
#> 45     S05     I05 4.7192574 3.018906  0.15323300  0.07673155  1.470387248
#> 46     S06     I05 3.3505867 3.018906  0.12860568  0.07673155  0.126343843
#> 47     S07     I05 1.1575612 3.018906 -0.72484206  0.07673155 -1.213233947
#> 48     S08     I05 3.3281779 3.018906 -0.01945261  0.07673155  0.251993370
#> 49     S09     I05 2.9812084 3.018906  0.75503061  0.07673155 -0.869459348
#> 50     S10     I05 3.3756190 3.018906  0.73323587  0.07673155 -0.453254013
#> 51     S01     I06 2.6061066 3.018906  0.02805978 -0.73411275  0.293253920
#> 52     S02     I06 3.0028335 3.018906 -0.03908504 -0.73411275  0.757125690
#> 53     S03     I06 1.9293074 3.018906  0.42565642 -0.73411275 -0.781141855
#> 54     S04     I06 4.5324660 3.018906  0.02611968 -0.73411275  2.221553456
#> 55     S05     I06 1.7903403 3.018906  0.15323300 -0.73411275 -0.647685545
#> 56     S06     I06 3.4711908 3.018906  0.12860568 -0.73411275  1.057792231
#> 57     S07     I06 2.3875263 3.018906 -0.72484206 -0.73411275  0.827575521
#> 58     S08     I06 4.4369801 3.018906 -0.01945261 -0.73411275  2.171639813
#> 59     S09     I06 2.8530437 3.018906  0.75503061 -0.73411275 -0.186779770
#> 60     S10     I06 2.3898964 3.018906  0.73323587 -0.73411275 -0.628132327
#> 61     S01     I07 1.0811957 3.018906  0.02805978 -0.78132320 -1.184446523
#> 62     S02     I07 1.6276274 3.018906 -0.03908504 -0.78132320 -0.570869986
#> 63     S03     I07 3.2010300 3.018906  0.42565642 -0.78132320  0.537791155
#> 64     S04     I07 2.4223862 3.018906  0.02611968 -0.78132320  0.158684139
#> 65     S05     I07 2.1196958 3.018906  0.15323300 -0.78132320 -0.271119639
#> 66     S06     I07 1.3938027 3.018906  0.12860568 -0.78132320 -0.972385394
#> 67     S07     I07 0.7002856 3.018906 -0.72484206 -0.78132320 -0.812454791
#> 68     S08     I07 2.2212733 3.018906 -0.01945261 -0.78132320  0.003143549
#> 69     S09     I07 4.0575198 3.018906  0.75503061 -0.78132320  1.064906796
#> 70     S10     I07 2.4817761 3.018906  0.73323587 -0.78132320 -0.489042128
#> 71     S01     I08 1.4630860 3.018906  0.02805978  0.36427551 -1.948154923
#> 72     S02     I08 5.5545989 3.018906 -0.03908504  0.36427551  2.210502855
#> 73     S03     I08 2.9518140 3.018906  0.42565642  0.36427551 -0.857023498
#> 74     S04     I08 3.1830514 3.018906  0.02611968  0.36427551 -0.226249414
#> 75     S05     I08 2.9533531 3.018906  0.15323300  0.36427551 -0.583060976
#> 76     S06     I08 3.2022729 3.018906  0.12860568  0.36427551 -0.309513951
#> 77     S07     I08 2.2420840 3.018906 -0.72484206  0.36427551 -0.416255111
#> 78     S08     I08 3.1012345 3.018906 -0.01945261  0.36427551 -0.262494019
#> 79     S09     I08 3.0974267 3.018906  0.75503061  0.36427551 -1.040785056
#> 80     S10     I08 4.7925155 3.018906  0.73323587  0.36427551  0.676098552
#> 81     S01     I09 3.6855343 3.018906  0.02805978  1.00457845 -0.366009519
#> 82     S02     I09 5.2725231 3.018906 -0.03908504  1.00457845  1.288124079
#> 83     S03     I09 3.7950833 3.018906  0.42565642  1.00457845 -0.654057192
#> 84     S04     I09 3.0083248 3.018906  0.02611968  1.00457845 -1.041278951
#> 85     S05     I09 5.2393249 3.018906  0.15323300  1.00457845  1.062607861
#> 86     S06     I09 3.1523666 3.018906  0.12860568  1.00457845 -0.999723114
#> 87     S07     I09 4.0035449 3.018906 -0.72484206  1.00457845  0.704902905
#> 88     S08     I09 4.3329246 3.018906 -0.01945261  1.00457845  0.328893138
#> 89     S09     I09 4.9650986 3.018906  0.75503061  1.00457845  0.186583931
#> 90     S10     I09 5.4323812 3.018906  0.73323587  1.00457845  0.675661288
#> 91     S01     I10 2.7751682 3.018906  0.02805978 -0.01304277 -0.258754436
#> 92     S02     I10 3.6078174 3.018906 -0.03908504 -0.01304277  0.641039557
#> 93     S03     I10 2.0041595 3.018906  0.42565642 -0.01304277 -1.427359737
#> 94     S04     I10 4.7886985 3.018906  0.02611968 -0.01304277  1.756715985
#> 95     S05     I10 2.0099390 3.018906  0.15323300 -0.01304277 -1.149156853
#> 96     S06     I10 5.0740262 3.018906  0.12860568 -0.01304277  1.939557631
#> 97     S07     I10 3.1744809 3.018906 -0.72484206 -0.01304277  0.893460091
#> 98     S08     I10 4.5173552 3.018906 -0.01945261 -0.01304277  1.530944922
#> 99     S09     I10 3.6525313 3.018906  0.75503061 -0.01304277 -0.108362120
#> 100    S10     I10 5.4244101 3.018906  0.73323587 -0.01304277  1.685311400