Generate a sample with random intercepts for subjects and items

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(.data, sub_n = NULL, item_n = NULL, dv = 1,
  sub_id = 2, item_id = 3, seed = NULL)

Arguments

.data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id        val  grand_i       sub_i     item_i          err
#> 1      S01     I01  3.3666387 3.018906 -0.27889026  0.6180103  0.008613058
#> 2      S02     I01  4.3964387 3.018906  0.63494106  0.6180103  0.124581694
#> 3      S03     I01  4.9577488 3.018906  0.37907607  0.6180103  0.941756776
#> 4      S04     I01  4.9455330 3.018906  1.05888211  0.6180103  0.249734958
#> 5      S05     I01  5.5934950 3.018906  1.97063523  0.6180103 -0.014056205
#> 6      S06     I01  4.1629792 3.018906  0.07139596  0.6180103  0.454667293
#> 7      S07     I01  3.9066024 3.018906 -0.44496567  0.6180103  0.714652137
#> 8      S08     I01  4.4443187 3.018906  0.39543822  0.6180103  0.411964598
#> 9      S09     I01  4.1709971 3.018906  0.02036073  0.6180103  0.513720459
#> 10     S10     I01  1.3820943 3.018906 -1.15747743  0.6180103 -1.097344152
#> 11     S01     I02  1.4656752 3.018906 -0.27889026  0.6726218 -1.946962030
#> 12     S02     I02  6.1305851 3.018906  0.63494106  0.6726218  1.804116566
#> 13     S03     I02  2.3328159 3.018906  0.37907607  0.6726218 -1.737787617
#> 14     S04     I02  5.7306988 3.018906  1.05888211  0.6726218  0.980289196
#> 15     S05     I02  6.6263257 3.018906  1.97063523  0.6726218  0.964163009
#> 16     S06     I02  2.5644986 3.018906  0.07139596  0.6726218 -1.198424766
#> 17     S07     I02  2.2237114 3.018906 -0.44496567  0.6726218 -1.022850376
#> 18     S08     I02  1.0571193 3.018906  0.39543822  0.6726218 -3.029846400
#> 19     S09     I02  2.3550054 3.018906  0.02036073  0.6726218 -1.356882730
#> 20     S10     I02  2.0305393 3.018906 -1.15747743  0.6726218 -0.503510701
#> 21     S01     I03  4.1608587 3.018906 -0.27889026  0.9149951  0.505848264
#> 22     S02     I03  5.4898856 3.018906  0.63494106  0.9149951  0.921043810
#> 23     S03     I03  3.0811621 3.018906  0.37907607  0.9149951 -1.231814646
#> 24     S04     I03  5.2295906 3.018906  1.05888211  0.9149951  0.236807766
#> 25     S05     I03  7.6792044 3.018906  1.97063523  0.9149951  1.774668501
#> 26     S06     I03  3.0957687 3.018906  0.07139596  0.9149951 -0.909527965
#> 27     S07     I03  3.7805407 3.018906 -0.44496567  0.9149951  0.291605670
#> 28     S08     I03  4.3203333 3.018906  0.39543822  0.9149951 -0.009005562
#> 29     S09     I03  4.4217167 3.018906  0.02036073  0.9149951  0.467455277
#> 30     S10     I03  4.5039791 3.018906 -1.15747743  0.9149951  1.727555829
#> 31     S01     I04 -0.6333022 3.018906 -0.27889026 -1.0459298 -2.327387723
#> 32     S02     I04  3.2252960 3.018906  0.63494106 -1.0459298  0.617379085
#> 33     S03     I04  2.5588912 3.018906  0.37907607 -1.0459298  0.206839327
#> 34     S04     I04  1.9404564 3.018906  1.05888211 -1.0459298 -1.091401524
#> 35     S05     I04  5.1359559 3.018906  1.97063523 -1.0459298  1.192344825
#> 36     S06     I04  1.0443115 3.018906  0.07139596 -1.0459298 -1.000060322
#> 37     S07     I04  0.7363709 3.018906 -0.44496567 -1.0459298 -0.791639229
#> 38     S08     I04  5.2480680 3.018906  0.39543822 -1.0459298  2.879653924
#> 39     S09     I04  2.8780537 3.018906  0.02036073 -1.0459298  0.884717107
#> 40     S10     I04 -0.5680213 3.018906 -1.15747743 -1.0459298 -1.383519658
#> 41     S01     I05  3.9328541 3.018906 -0.27889026  0.9186049  0.274233783
#> 42     S02     I05  4.0070825 3.018906  0.63494106  0.9186049 -0.565369120
#> 43     S03     I05  5.7397823 3.018906  0.37907607  0.9186049  1.423195657
#> 44     S04     I05  6.1032771 3.018906  1.05888211  0.9186049  1.106884419
#> 45     S05     I05  5.3981563 3.018906  1.97063523  0.9186049 -0.509989451
#> 46     S06     I05  3.2772267 3.018906  0.07139596  0.9186049 -0.731679799
#> 47     S07     I05  2.2041534 3.018906 -0.44496567  0.9186049 -1.288391438
#> 48     S08     I05  5.8942878 3.018906  0.39543822  0.9186049  1.561339073
#> 49     S09     I05  5.1399931 3.018906  0.02036073  0.9186049  1.182121843
#> 50     S10     I05  2.7510765 3.018906 -1.15747743  0.9186049 -0.028956660
#> 51     S01     I06  3.0851708 3.018906 -0.27889026  0.7965244 -0.451368980
#> 52     S02     I06  4.9516055 3.018906  0.63494106  0.7965244  0.501234395
#> 53     S03     I06  4.4291337 3.018906  0.37907607  0.7965244  0.234627658
#> 54     S04     I06  4.5471602 3.018906  1.05888211  0.7965244 -0.327151902
#> 55     S05     I06  4.4209913 3.018906  1.97063523  0.7965244 -1.365073959
#> 56     S06     I06  3.7149547 3.018906  0.07139596  0.7965244 -0.171871307
#> 57     S07     I06  3.2515910 3.018906 -0.44496567  0.7965244 -0.118873337
#> 58     S08     I06  4.7122473 3.018906  0.39543822  0.7965244  0.501379079
#> 59     S09     I06  3.7579993 3.018906  0.02036073  0.7965244 -0.077791472
#> 60     S10     I06  2.5030097 3.018906 -1.15747743  0.7965244 -0.154942857
#> 61     S01     I07  3.5283724 3.018906 -0.27889026  0.6018640  0.186493072
#> 62     S02     I07  7.6462923 3.018906  0.63494106  0.6018640  3.390581665
#> 63     S03     I07  3.8958322 3.018906  0.37907607  0.6018640 -0.104013472
#> 64     S04     I07  3.9224884 3.018906  1.05888211  0.6018640 -0.757163342
#> 65     S05     I07  5.4111356 3.018906  1.97063523  0.6018640 -0.180269202
#> 66     S06     I07  3.8916600 3.018906  0.07139596  0.6018640  0.199494406
#> 67     S07     I07  4.4254371 3.018906 -0.44496567  0.6018640  1.249633128
#> 68     S08     I07  3.1254841 3.018906  0.39543822  0.6018640 -0.890723707
#> 69     S09     I07  0.8033408 3.018906  0.02036073  0.6018640 -2.837789496
#> 70     S10     I07  3.5062259 3.018906 -1.15747743  0.6018640  1.042933717
#> 71     S01     I08  3.4067189 3.018906 -0.27889026  0.5459112  0.120792343
#> 72     S02     I08  5.3660360 3.018906  0.63494106  0.5459112  1.166278159
#> 73     S03     I08  2.8892379 3.018906  0.37907607  0.5459112 -1.054654985
#> 74     S04     I08  7.9598935 3.018906  1.05888211  0.5459112  3.336194592
#> 75     S05     I08  5.9223507 3.018906  1.97063523  0.5459112  0.386898682
#> 76     S06     I08  5.0022541 3.018906  0.07139596  0.5459112  1.366041307
#> 77     S07     I08  1.8758306 3.018906 -0.44496567  0.5459112 -1.244020566
#> 78     S08     I08  3.2868142 3.018906  0.39543822  0.5459112 -0.673440830
#> 79     S09     I08  3.4829318 3.018906  0.02036073  0.5459112 -0.102245729
#> 80     S10     I08  3.0899977 3.018906 -1.15747743  0.5459112  0.682658290
#> 81     S01     I09  2.4521244 3.018906 -0.27889026 -0.3272687  0.039377706
#> 82     S02     I09  5.3307141 3.018906  0.63494106 -0.3272687  2.004136133
#> 83     S03     I09  2.8476623 3.018906  0.37907607 -0.3272687 -0.223050682
#> 84     S04     I09  3.8408135 3.018906  1.05888211 -0.3272687  0.090294434
#> 85     S05     I09  3.1162972 3.018906  1.97063523 -0.3272687 -1.545974974
#> 86     S06     I09  2.4086016 3.018906  0.07139596 -0.3272687 -0.354431253
#> 87     S07     I09  3.4618533 3.018906 -0.44496567 -0.3272687  1.215182065
#> 88     S08     I09  2.9610176 3.018906  0.39543822 -0.3272687 -0.126057521
#> 89     S09     I09  2.9045382 3.018906  0.02036073 -0.3272687  0.192540534
#> 90     S10     I09  2.3260486 3.018906 -1.15747743 -0.3272687  0.791889092
#> 91     S01     I10  2.1808606 3.018906 -0.27889026  0.2035203 -0.762675088
#> 92     S02     I10  3.6061579 3.018906  0.63494106  0.2035203 -0.251209103
#> 93     S03     I10  4.6855673 3.018906  0.37907607  0.2035203  1.084065284
#> 94     S04     I10  2.3126365 3.018906  1.05888211  0.2035203 -1.968671548
#> 95     S05     I10  4.0567527 3.018906  1.97063523  0.2035203 -1.136308496
#> 96     S06     I10  3.4131912 3.018906  0.07139596  0.2035203  0.119369277
#> 97     S07     I10  3.5433295 3.018906 -0.44496567  0.2035203  0.765869252
#> 98     S08     I10  5.5439522 3.018906  0.39543822  0.2035203  1.926088074
#> 99     S09     I10  3.9274550 3.018906  0.02036073  0.2035203  0.684668319
#> 100    S10     I10  3.4064235 3.018906 -1.15747743  0.2035203  1.341475016