Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id           y  grand_i      sub_i     item_i         err
#> 1      S01     I01  4.40559817 3.018906  0.3871324  0.7661444  0.23341575
#> 2      S02     I01  3.44268128 3.018906 -1.1303745  0.7661444  0.78800576
#> 3      S03     I01  4.12590203 3.018906 -1.0993374  0.7661444  1.44018936
#> 4      S04     I01  0.68663220 3.018906 -1.3254376  0.7661444 -1.77298025
#> 5      S05     I01  2.76401536 3.018906 -0.4260909  0.7661444 -0.59494376
#> 6      S06     I01  4.58727357 3.018906  1.3676314  0.7661444 -0.56540786
#> 7      S07     I01  5.03730112 3.018906  0.4924711  0.7661444  0.75977999
#> 8      S08     I01  2.85834201 3.018906  0.4894632  0.7661444 -1.41617129
#> 9      S09     I01  3.28685263 3.018906  0.5409288  0.7661444 -1.03912618
#> 10     S10     I01  4.30969980 3.018906 -0.9899364  0.7661444  1.51458615
#> 11     S01     I02  1.80771745 3.018906  0.3871324 -1.6633537  0.06503313
#> 12     S02     I02  0.66294910 3.018906 -1.1303745 -1.6633537  0.43777167
#> 13     S03     I02  0.02375537 3.018906 -1.0993374 -1.6633537 -0.23245921
#> 14     S04     I02 -0.71292515 3.018906 -1.3254376 -1.6633537 -0.74303952
#> 15     S05     I02  1.24008421 3.018906 -0.4260909 -1.6633537  0.31062318
#> 16     S06     I02  4.03595962 3.018906  1.3676314 -1.6633537  1.31277628
#> 17     S07     I02  2.60704590 3.018906  0.4924711 -1.6633537  0.75902286
#> 18     S08     I02  2.47808129 3.018906  0.4894632 -1.6633537  0.63306608
#> 19     S09     I02  2.32463750 3.018906  0.5409288 -1.6633537  0.42815678
#> 20     S10     I02  1.08664177 3.018906 -0.9899364 -1.6633537  0.72102622
#> 21     S01     I03  3.91319896 3.018906  0.3871324  0.7361624 -0.22900144
#> 22     S02     I03  1.99245454 3.018906 -1.1303745  0.7361624 -0.63223896
#> 23     S03     I03  1.03736661 3.018906 -1.0993374  0.7361624 -1.61836404
#> 24     S04     I03  2.09815453 3.018906 -1.3254376  0.7361624 -0.33147591
#> 25     S05     I03  5.35493113 3.018906 -0.4260909  0.7361624  2.02595402
#> 26     S06     I03  7.17404392 3.018906  1.3676314  0.7361624  2.05134451
#> 27     S07     I03  2.92991285 3.018906  0.4924711  0.7361624 -1.31762627
#> 28     S08     I03  3.54382182 3.018906  0.4894632  0.7361624 -0.70070946
#> 29     S09     I03  4.45880499 3.018906  0.5409288  0.7361624  0.16280819
#> 30     S10     I03  4.06443531 3.018906 -0.9899364  0.7361624  1.29930368
#> 31     S01     I04  4.91946952 3.018906  0.3871324  0.5074318  1.00599979
#> 32     S02     I04  2.99536973 3.018906 -1.1303745  0.5074318  0.59940689
#> 33     S03     I04  1.68806019 3.018906 -1.0993374  0.5074318 -0.73893979
#> 34     S04     I04  2.87303553 3.018906 -1.3254376  0.5074318  0.67213575
#> 35     S05     I04  3.40834124 3.018906 -0.4260909  0.5074318  0.30809480
#> 36     S06     I04  4.29772340 3.018906  1.3676314  0.5074318 -0.59624535
#> 37     S07     I04  4.00488355 3.018906  0.4924711  0.5074318 -0.01392490
#> 38     S08     I04  4.36196460 3.018906  0.4894632  0.5074318  0.34616398
#> 39     S09     I04  4.70530596 3.018906  0.5409288  0.5074318  0.63803983
#> 40     S10     I04  4.32480885 3.018906 -0.9899364  0.5074318  1.78840789
#> 41     S01     I05  3.77971154 3.018906  0.3871324  0.1875071  0.18616652
#> 42     S02     I05  5.19514097 3.018906 -1.1303745  0.1875071  3.11910284
#> 43     S03     I05  3.60191658 3.018906 -1.0993374  0.1875071  1.49484130
#> 44     S04     I05  0.59386295 3.018906 -1.3254376  0.1875071 -1.28711212
#> 45     S05     I05  1.73526145 3.018906 -0.4260909  0.1875071 -1.04506028
#> 46     S06     I05  2.19436175 3.018906  1.3676314  0.1875071 -2.37968230
#> 47     S07     I05  3.47534858 3.018906  0.4924711  0.1875071 -0.22353516
#> 48     S08     I05  3.14293961 3.018906  0.4894632  0.1875071 -0.55293630
#> 49     S09     I05  3.52339263 3.018906  0.5409288  0.1875071 -0.22394880
#> 50     S10     I05  2.08767756 3.018906 -0.9899364  0.1875071 -0.12879870
#> 51     S01     I06  4.03127127 3.018906  0.3871324 -0.3264574  0.95169070
#> 52     S02     I06 -0.46773095 3.018906 -1.1303745 -0.3264574 -2.02980463
#> 53     S03     I06  2.57210396 3.018906 -1.0993374 -0.3264574  0.97899314
#> 54     S04     I06  2.59022345 3.018906 -1.3254376 -0.3264574  1.22321284
#> 55     S05     I06  1.32803661 3.018906 -0.4260909 -0.3264574 -0.93832067
#> 56     S06     I06  4.85366419 3.018906  1.3676314 -0.3264574  0.79358460
#> 57     S07     I06  2.27995877 3.018906  0.4924711 -0.3264574 -0.90496052
#> 58     S08     I06  3.56121237 3.018906  0.4894632 -0.3264574  0.37930091
#> 59     S09     I06  5.59322888 3.018906  0.5409288 -0.3264574  2.35985191
#> 60     S10     I06 -0.68595349 3.018906 -0.9899364 -0.3264574 -2.38846530
#> 61     S01     I07  1.33815838 3.018906  0.3871324 -0.2109181 -1.85696148
#> 62     S02     I07  3.24580984 3.018906 -1.1303745 -0.2109181  1.56819687
#> 63     S03     I07  1.60178435 3.018906 -1.0993374 -0.2109181 -0.10686577
#> 64     S04     I07  2.33546543 3.018906 -1.3254376 -0.2109181  0.85291553
#> 65     S05     I07  3.51510036 3.018906 -0.4260909 -0.2109181  1.13320380
#> 66     S06     I07  4.19134419 3.018906  1.3676314 -0.2109181  0.01572531
#> 67     S07     I07  1.39771315 3.018906  0.4924711 -0.2109181 -1.90274543
#> 68     S08     I07  1.63267037 3.018906  0.4894632 -0.2109181 -1.66478037
#> 69     S09     I07  4.97066649 3.018906  0.5409288 -0.2109181  1.62175023
#> 70     S10     I07  0.17146484 3.018906 -0.9899364 -0.2109181 -1.64658625
#> 71     S01     I08  4.06549907 3.018906  0.3871324 -0.4496278  1.10908887
#> 72     S02     I08  4.09959305 3.018906 -1.1303745 -0.4496278  2.66068974
#> 73     S03     I08 -0.80777102 3.018906 -1.0993374 -0.4496278 -2.27771148
#> 74     S04     I08  2.00399466 3.018906 -1.3254376 -0.4496278  0.76015442
#> 75     S05     I08  1.33279498 3.018906 -0.4260909 -0.4496278 -0.81039193
#> 76     S06     I08  2.91782135 3.018906  1.3676314 -0.4496278 -1.01908788
#> 77     S07     I08  0.77303528 3.018906  0.4924711 -0.4496278 -2.28871364
#> 78     S08     I08  5.36272628 3.018906  0.4894632 -0.4496278  2.30398519
#> 79     S09     I08  3.51782701 3.018906  0.5409288 -0.4496278  0.40762041
#> 80     S10     I08  1.96255175 3.018906 -0.9899364 -0.4496278  0.38321032
#> 81     S01     I09  5.65424427 3.018906  0.3871324  0.4009762  1.84723006
#> 82     S02     I09  3.78767505 3.018906 -1.1303745  0.4009762  1.49816773
#> 83     S03     I09  1.67911188 3.018906 -1.0993374  0.4009762 -0.64143258
#> 84     S04     I09  3.35564977 3.018906 -1.3254376  0.4009762  1.26120553
#> 85     S05     I09  2.80534947 3.018906 -0.4260909  0.4009762 -0.18844144
#> 86     S06     I09  4.65473783 3.018906  1.3676314  0.4009762 -0.13277540
#> 87     S07     I09  2.91889538 3.018906  0.4924711  0.4009762 -0.99345755
#> 88     S08     I09  4.69310461 3.018906  0.4894632  0.4009762  0.78375952
#> 89     S09     I09  1.60881539 3.018906  0.5409288  0.4009762 -2.35199521
#> 90     S10     I09  1.54500245 3.018906 -0.9899364  0.4009762 -0.88494299
#> 91     S01     I10  1.50382632 3.018906  0.3871324 -1.5412093 -0.36100240
#> 92     S02     I10  1.53479422 3.018906 -1.1303745 -1.5412093  1.18747240
#> 93     S03     I10 -0.67475647 3.018906 -1.0993374 -1.5412093 -1.05311544
#> 94     S04     I10  0.43366088 3.018906 -1.3254376 -1.5412093  0.28140213
#> 95     S05     I10  3.32355462 3.018906 -0.4260909 -1.5412093  2.27194920
#> 96     S06     I10  3.35941949 3.018906  1.3676314 -1.5412093  0.51409175
#> 97     S07     I10  0.81545555 3.018906  0.4924711 -1.5412093 -1.15471188
#> 98     S08     I10  0.61314315 3.018906  0.4894632 -1.5412093 -1.35401645
#> 99     S09     I10  2.36144299 3.018906  0.5409288 -1.5412093  0.34281788
#> 100    S10     I10 -0.42598296 3.018906 -0.9899364 -1.5412093 -0.91374291