Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(.data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

.data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id           y  grand_i       sub_i      item_i          err
#> 1      S01     I01  3.25861122 3.018906  0.25216409  0.27277566 -0.285234140
#> 2      S02     I01  2.67114072 3.018906 -0.87021237  0.27277566  0.249671820
#> 3      S03     I01  3.45328341 3.018906  0.05209540  0.27277566  0.109506749
#> 4      S04     I01  1.68911860 3.018906 -0.04143293  0.27277566 -1.561129744
#> 5      S05     I01  2.43003863 3.018906 -1.55533922  0.27277566  0.693696583
#> 6      S06     I01  3.44423323 3.018906  0.22932883  0.27277566 -0.076776871
#> 7      S07     I01  2.03957201 3.018906 -0.55821757  0.27277566 -0.693891697
#> 8      S08     I01  1.77327102 3.018906 -0.70821063  0.27277566 -0.810199624
#> 9      S09     I01  2.34280818 3.018906 -0.81919449  0.27277566 -0.129678605
#> 10     S10     I01 -0.43679832 3.018906 -1.22390850  0.27277566 -2.504571083
#> 11     S01     I02  4.66784540 3.018906  0.25216409  0.47240504  0.924370653
#> 12     S02     I02  2.94198213 3.018906 -0.87021237  0.47240504  0.320883851
#> 13     S03     I02  3.07072627 3.018906  0.05209540  0.47240504 -0.472679770
#> 14     S04     I02  2.97318885 3.018906 -0.04143293  0.47240504 -0.476688869
#> 15     S05     I02  2.89398346 3.018906 -1.55533922  0.47240504  0.958012028
#> 16     S06     I02  4.36641126 3.018906  0.22932883  0.47240504  0.645771773
#> 17     S07     I02  0.84119375 3.018906 -0.55821757  0.47240504 -2.091899334
#> 18     S08     I02  1.23554597 3.018906 -0.70821063  0.47240504 -1.547554057
#> 19     S09     I02  3.68846194 3.018906 -0.81919449  0.47240504  1.016345783
#> 20     S10     I02  1.97801323 3.018906 -1.22390850  0.47240504 -0.289388921
#> 21     S01     I03  0.53472363 3.018906  0.25216409 -1.56925114 -1.167094934
#> 22     S02     I03  0.80255458 3.018906 -0.87021237 -1.56925114  0.223112470
#> 23     S03     I03  1.25340589 3.018906  0.05209540 -1.56925114 -0.248343984
#> 24     S04     I03 -0.75134785 3.018906 -0.04143293 -1.56925114 -2.159569395
#> 25     S05     I03  0.44813128 3.018906 -1.55533922 -1.56925114  0.553816022
#> 26     S06     I03  4.07355262 3.018906  0.22932883 -1.56925114  2.394569314
#> 27     S07     I03  1.44107431 3.018906 -0.55821757 -1.56925114  0.549637401
#> 28     S08     I03  0.74466454 3.018906 -0.70821063 -1.56925114  0.003220689
#> 29     S09     I03  2.18776822 3.018906 -0.81919449 -1.56925114  1.557308231
#> 30     S10     I03  0.38118500 3.018906 -1.22390850 -1.56925114  0.155439029
#> 31     S01     I04  2.36772881 3.018906  0.25216409 -0.46403237 -0.439308523
#> 32     S02     I04  1.42239131 3.018906 -0.87021237 -0.46403237 -0.262269558
#> 33     S03     I04  2.02679243 3.018906  0.05209540 -0.46403237 -0.580176205
#> 34     S04     I04  2.87813807 3.018906 -0.04143293 -0.46403237  0.364697755
#> 35     S05     I04  1.43358195 3.018906 -1.55533922 -0.46403237  0.434047929
#> 36     S06     I04  3.36882646 3.018906  0.22932883 -0.46403237  0.584624383
#> 37     S07     I04  2.38088099 3.018906 -0.55821757 -0.46403237  0.384225319
#> 38     S08     I04  0.77218555 3.018906 -0.70821063 -0.46403237 -1.074477063
#> 39     S09     I04 -0.13251035 3.018906 -0.81919449 -0.46403237 -1.868189097
#> 40     S10     I04  1.85958033 3.018906 -1.22390850 -0.46403237  0.528615597
#> 41     S01     I05  2.92029094 3.018906  0.25216409 -0.08565360 -0.265125170
#> 42     S02     I05  2.87253576 3.018906 -0.87021237 -0.08565360  0.809496112
#> 43     S03     I05  2.12780552 3.018906  0.05209540 -0.08565360 -0.857541895
#> 44     S04     I05  2.71953143 3.018906 -0.04143293 -0.08565360 -0.172287660
#> 45     S05     I05  0.03203695 3.018906 -1.55533922 -0.08565360 -1.345875846
#> 46     S06     I05  2.45746674 3.018906  0.22932883 -0.08565360 -0.705114107
#> 47     S07     I05  1.98232782 3.018906 -0.55821757 -0.08565360 -0.392706624
#> 48     S08     I05  3.08455305 3.018906 -0.70821063 -0.08565360  0.859511659
#> 49     S09     I05  2.31433229 3.018906 -0.81919449 -0.08565360  0.200274760
#> 50     S10     I05 -0.61633875 3.018906 -1.22390850 -0.08565360 -2.325682262
#> 51     S01     I06  2.79682520 3.018906  0.25216409  0.49091258 -0.965157081
#> 52     S02     I06  1.86684874 3.018906 -0.87021237  0.49091258 -0.772757083
#> 53     S03     I06  2.92459026 3.018906  0.05209540  0.49091258 -0.637323329
#> 54     S04     I06  4.15868545 3.018906 -0.04143293  0.49091258  0.690300192
#> 55     S05     I06  2.04339653 3.018906 -1.55533922  0.49091258  0.088917561
#> 56     S06     I06  5.46383722 3.018906  0.22932883  0.49091258  1.724690195
#> 57     S07     I06  4.23335377 3.018906 -0.55821757  0.49091258  1.281753142
#> 58     S08     I06  1.61617304 3.018906 -0.70821063  0.49091258 -1.185434521
#> 59     S09     I06  2.83445668 3.018906 -0.81919449  0.49091258  0.143832983
#> 60     S10     I06  2.47977450 3.018906 -1.22390850  0.49091258  0.193864810
#> 61     S01     I07  0.46502907 3.018906  0.25216409 -0.95497667 -1.851063971
#> 62     S02     I07  0.04721845 3.018906 -0.87021237 -0.95497667 -1.146498124
#> 63     S03     I07  3.78678184 3.018906  0.05209540 -0.95497667  1.670757499
#> 64     S04     I07  2.41077450 3.018906 -0.04143293 -0.95497667  0.388278485
#> 65     S05     I07 -0.55261301 3.018906 -1.55533922 -0.95497667 -1.061202736
#> 66     S06     I07  3.68140785 3.018906  0.22932883 -0.95497667  1.388150073
#> 67     S07     I07  1.18213165 3.018906 -0.55821757 -0.95497667 -0.323579729
#> 68     S08     I07 -0.12397733 3.018906 -0.70821063 -0.95497667 -1.479695646
#> 69     S09     I07  1.67992799 3.018906 -0.81919449 -0.95497667  0.435193538
#> 70     S10     I07  0.81132211 3.018906 -1.22390850 -0.95497667 -0.028698332
#> 71     S01     I08  3.08563022 3.018906  0.25216409  0.33253792 -0.517977410
#> 72     S02     I08  2.34290785 3.018906 -0.87021237  0.33253792 -0.138323317
#> 73     S03     I08  4.37774778 3.018906  0.05209540  0.33253792  0.974208845
#> 74     S04     I08  1.02428183 3.018906 -0.04143293  0.33253792 -2.285728780
#> 75     S05     I08  3.39853866 3.018906 -1.55533922  0.33253792  1.602434340
#> 76     S06     I08  1.89243671 3.018906  0.22932883  0.33253792 -1.688335661
#> 77     S07     I08  4.98787447 3.018906 -0.55821757  0.33253792  2.194648496
#> 78     S08     I08  2.91987993 3.018906 -0.70821063  0.33253792  0.276647018
#> 79     S09     I08  4.45454528 3.018906 -0.81919449  0.33253792  1.922296233
#> 80     S10     I08  0.47787165 3.018906 -1.22390850  0.33253792 -1.649663380
#> 81     S01     I09  0.39289304 3.018906  0.25216409 -0.06814383 -2.810032835
#> 82     S02     I09  1.48315104 3.018906 -0.87021237 -0.06814383 -0.597398378
#> 83     S03     I09  4.14491111 3.018906  0.05209540 -0.06814383  1.142053935
#> 84     S04     I09  2.56626255 3.018906 -0.04143293 -0.06814383 -0.343066304
#> 85     S05     I09  1.37987904 3.018906 -1.55533922 -0.06814383 -0.015543525
#> 86     S06     I09  3.36010233 3.018906  0.22932883 -0.06814383  0.180011717
#> 87     S07     I09  2.34245213 3.018906 -0.55821757 -0.06814383 -0.050092086
#> 88     S08     I09  2.52477622 3.018906 -0.70821063 -0.06814383  0.282225060
#> 89     S09     I09  1.38113033 3.018906 -0.81919449 -0.06814383 -0.750436963
#> 90     S10     I09  3.40278673 3.018906 -1.22390850 -0.06814383  1.675933455
#> 91     S01     I10  2.67299426 3.018906  0.25216409 -0.04679977 -0.551275676
#> 92     S02     I10  3.59750414 3.018906 -0.87021237 -0.04679977  1.495610673
#> 93     S03     I10  2.79036203 3.018906  0.05209540 -0.04679977 -0.233839204
#> 94     S04     I10  3.54277496 3.018906 -0.04143293 -0.04679977  0.612102050
#> 95     S05     I10  1.16721548 3.018906 -1.55533922 -0.04679977 -0.249551142
#> 96     S06     I10  2.88640149 3.018906  0.22932883 -0.04679977 -0.315033184
#> 97     S07     I10  2.22236880 3.018906 -0.55821757 -0.04679977 -0.191519469
#> 98     S08     I10  0.27848174 3.018906 -0.70821063 -0.04679977 -1.985413467
#> 99     S09     I10  1.92744906 3.018906 -0.81919449 -0.04679977 -0.225462292
#> 100    S10     I10  0.61689366 3.018906 -1.22390850 -0.04679977 -1.131303679