Generate a sample with random intercepts for subjects and items

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(.data, sub_n = 100, item_n = 25, dv = 1, sub_id = 2,
  item_id = 3, seed = NULL)

Arguments

.data
the existing tbl

sub_n
the number of subjects to simulate

item_n
the number of items to simulate

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id        val  grand_i       sub_i      item_i         err
#> 1      S01     I01  4.5508203 3.018906  0.19136983  1.12673449  0.21381037
#> 2      S02     I01  4.5969762 3.018906 -0.48247088  1.12673449  0.93380703
#> 3      S03     I01  3.4946267 3.018906  0.18529730  1.12673449 -0.83631072
#> 4      S04     I01  4.0266176 3.018906 -0.28053080  1.12673449  0.16150831
#> 5      S05     I01  4.4643170 3.018906  0.12187036  1.12673449  0.19680652
#> 6      S06     I01  5.5457024 3.018906  1.21614257  1.12673449  0.18391970
#> 7      S07     I01  3.8671385 3.018906  0.95736712  1.12673449 -1.23586871
#> 8      S08     I01  2.9003153 3.018906 -0.09651158  1.12673449 -1.14881319
#> 9      S09     I01  4.4797018 3.018906  0.52064570  1.12673449 -0.18658401
#> 10     S10     I01  4.2684961 3.018906  0.66323604  1.12673449 -0.54038008
#> 11     S01     I02  3.0523151 3.018906  0.19136983 -0.16098453  0.00302416
#> 12     S02     I02  3.3835070 3.018906 -0.48247088 -0.16098453  1.00805679
#> 13     S03     I02  1.6426491 3.018906  0.18529730 -0.16098453 -1.40056932
#> 14     S04     I02  3.4357912 3.018906 -0.28053080 -0.16098453  0.85840089
#> 15     S05     I02  3.9930330 3.018906  0.12187036 -0.16098453  1.01324159
#> 16     S06     I02  5.0597052 3.018906  1.21614257 -0.16098453  0.98564157
#> 17     S07     I02  3.9371732 3.018906  0.95736712 -0.16098453  0.12188496
#> 18     S08     I02  4.0964161 3.018906 -0.09651158 -0.16098453  1.33500656
#> 19     S09     I02  4.4218897 3.018906  0.52064570 -0.16098453  1.04332290
#> 20     S10     I02  4.2864161 3.018906  0.66323604 -0.16098453  0.76525902
#> 21     S01     I03  2.0445254 3.018906  0.19136983  0.20005204 -1.36580210
#> 22     S02     I03  2.5256102 3.018906 -0.48247088  0.20005204 -0.21087655
#> 23     S03     I03  4.4220577 3.018906  0.18529730  0.20005204  1.01780278
#> 24     S04     I03  3.4942463 3.018906 -0.28053080  0.20005204  0.55581946
#> 25     S05     I03  3.2289920 3.018906  0.12187036  0.20005204 -0.11183602
#> 26     S06     I03  4.6052000 3.018906  1.21614257  0.20005204  0.17009981
#> 27     S07     I03  4.0806994 3.018906  0.95736712  0.20005204 -0.09562539
#> 28     S08     I03  2.9051509 3.018906 -0.09651158  0.20005204 -0.21729521
#> 29     S09     I03  3.0821370 3.018906  0.52064570  0.20005204 -0.65746630
#> 30     S10     I03  4.9263817 3.018906  0.66323604  0.20005204  1.04418797
#> 31     S01     I04  3.0061144 3.018906  0.19136983  0.93334270 -1.13750372
#> 32     S02     I04  3.1600681 3.018906 -0.48247088  0.93334270 -0.30970936
#> 33     S03     I04  5.1693220 3.018906  0.18529730  0.93334270  1.03177636
#> 34     S04     I04  4.2878482 3.018906 -0.28053080  0.93334270  0.61613065
#> 35     S05     I04  2.8594473 3.018906  0.12187036  0.93334270 -1.21467140
#> 36     S06     I04  5.0884221 3.018906  1.21614257  0.93334270 -0.07996874
#> 37     S07     I04  6.5944650 3.018906  0.95736712  0.93334270  1.68484959
#> 38     S08     I04  2.8479354 3.018906 -0.09651158  0.93334270 -1.00780136
#> 39     S09     I04  4.2996206 3.018906  0.52064570  0.93334270 -0.17327340
#> 40     S10     I04  3.4489218 3.018906  0.66323604  0.93334270 -1.16656260
#> 41     S01     I05  2.5540625 3.018906  0.19136983 -1.34048584  0.68427288
#> 42     S02     I05  1.2571714 3.018906 -0.48247088 -1.34048584  0.06122247
#> 43     S03     I05  1.4300948 3.018906  0.18529730 -1.34048584 -0.43362226
#> 44     S04     I05  2.2419624 3.018906 -0.28053080 -1.34048584  0.84407343
#> 45     S05     I05  3.4016131 3.018906  0.12187036 -1.34048584  1.60132293
#> 46     S06     I05  1.7720808 3.018906  1.21614257 -1.34048584 -1.12248153
#> 47     S07     I05  2.6715301 3.018906  0.95736712 -1.34048584  0.03574322
#> 48     S08     I05  1.3875379 3.018906 -0.09651158 -1.34048584 -0.19437024
#> 49     S09     I05  3.8287658 3.018906  0.52064570 -1.34048584  1.62970036
#> 50     S10     I05  3.7302372 3.018906  0.66323604 -1.34048584  1.38858137
#> 51     S01     I06  4.1455720 3.018906  0.19136983  0.88213608  0.05316050
#> 52     S02     I06  4.4241646 3.018906 -0.48247088  0.88213608  1.00559379
#> 53     S03     I06  2.4300508 3.018906  0.18529730  0.88213608 -1.65628820
#> 54     S04     I06  4.3531638 3.018906 -0.28053080  0.88213608  0.73265289
#> 55     S05     I06  3.5679337 3.018906  0.12187036  0.88213608 -0.45497835
#> 56     S06     I06  4.7711772 3.018906  1.21614257  0.88213608 -0.34600710
#> 57     S07     I06  6.8223365 3.018906  0.95736712  0.88213608  1.96392772
#> 58     S08     I06  2.6260217 3.018906 -0.09651158  0.88213608 -1.17850840
#> 59     S09     I06  5.7402282 3.018906  0.52064570  0.88213608  1.31854079
#> 60     S10     I06  5.9262515 3.018906  0.66323604  0.88213608  1.36197376
#> 61     S01     I07  4.2858716 3.018906  0.19136983 -0.29022705  1.36582319
#> 62     S02     I07  3.2683454 3.018906 -0.48247088 -0.29022705  1.02213772
#> 63     S03     I07  1.8989021 3.018906  0.18529730 -0.29022705 -1.01507380
#> 64     S04     I07  3.6403711 3.018906 -0.28053080 -0.29022705  1.19222336
#> 65     S05     I07  1.6711299 3.018906  0.12187036 -0.29022705 -1.17941902
#> 66     S06     I07  1.9196423 3.018906  1.21614257 -0.29022705 -2.02517878
#> 67     S07     I07  2.8781300 3.018906  0.95736712 -0.29022705 -0.80791567
#> 68     S08     I07  3.4208595 3.018906 -0.09651158 -0.29022705  0.78869250
#> 69     S09     I07  2.4090473 3.018906  0.52064570 -0.29022705 -0.84027700
#> 70     S10     I07  2.8981367 3.018906  0.66323604 -0.29022705 -0.49377789
#> 71     S01     I08  5.2220700 3.018906  0.19136983  0.40735785  1.60443672
#> 72     S02     I08  2.1285653 3.018906 -0.48247088  0.40735785 -0.81522726
#> 73     S03     I08  3.2315612 3.018906  0.18529730  0.40735785 -0.37999952
#> 74     S04     I08  4.4735135 3.018906 -0.28053080  0.40735785  1.32778082
#> 75     S05     I08  2.2762440 3.018906  0.12187036  0.40735785 -1.27188980
#> 76     S06     I08  3.1937150 3.018906  1.21614257  0.40735785 -1.44869102
#> 77     S07     I08  2.4556127 3.018906  0.95736712  0.40735785 -1.92801788
#> 78     S08     I08  3.3025190 3.018906 -0.09651158  0.40735785 -0.02723289
#> 79     S09     I08  5.6334002 3.018906  0.52064570  0.40735785  1.68649099
#> 80     S10     I08  4.7835169 3.018906  0.66323604  0.40735785  0.69401739
#> 81     S01     I09  4.5485653 3.018906  0.19136983 -0.06498277  1.40327263
#> 82     S02     I09  0.3981457 3.018906 -0.48247088 -0.06498277 -2.07330624
#> 83     S03     I09  3.3902371 3.018906  0.18529730 -0.06498277  0.25101696
#> 84     S04     I09  2.4268503 3.018906 -0.28053080 -0.06498277 -0.24654177
#> 85     S05     I09  3.7163240 3.018906  0.12187036 -0.06498277  0.64053083
#> 86     S06     I09  4.8771743 3.018906  1.21614257 -0.06498277  0.70710888
#> 87     S07     I09  3.3048450 3.018906  0.95736712 -0.06498277 -0.60644493
#> 88     S08     I09 -0.1139544 3.018906 -0.09651158 -0.06498277 -2.97136566
#> 89     S09     I09  3.1692541 3.018906  0.52064570 -0.06498277 -0.30531445
#> 90     S10     I09  3.0002262 3.018906  0.66323604 -0.06498277 -0.61693267
#> 91     S01     I10  2.9830655 3.018906  0.19136983  0.77601487 -1.00322482
#> 92     S02     I10  2.9897216 3.018906 -0.48247088  0.77601487 -0.32272803
#> 93     S03     I10  1.5549279 3.018906  0.18529730  0.77601487 -2.42528985
#> 94     S04     I10  4.7552680 3.018906 -0.28053080  0.77601487  1.24087831
#> 95     S05     I10  3.1221253 3.018906  0.12187036  0.77601487 -0.79466551
#> 96     S06     I10  3.3438481 3.018906  1.21614257  0.77601487 -1.66721495
#> 97     S07     I10  4.3811539 3.018906  0.95736712  0.77601487 -0.37113375
#> 98     S08     I10  3.8004794 3.018906 -0.09651158  0.77601487  0.10207048
#> 99     S09     I10  4.0969316 3.018906  0.52064570  0.77601487 -0.21863459
#> 100    S10     I10  4.7447760 3.018906  0.66323604  0.77601487  0.28661943