Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id           y  grand_i      sub_i        item_i          err
#> 1      S01     I01  3.17992601 3.018906 -1.3590229  0.4519197386  1.068123528
#> 2      S02     I01  5.88891354 3.018906  0.5145774  0.4519197386  1.903510743
#> 3      S03     I01  3.29577993 3.018906 -0.5137715  0.4519197386  0.338726030
#> 4      S04     I01  4.95854533 3.018906  0.5142260  0.4519197386  0.973494022
#> 5      S05     I01  3.13158889 3.018906 -1.0735144  0.4519197386  0.734277981
#> 6      S06     I01  2.82286755 3.018906 -1.2838183  0.4519197386  0.635860542
#> 7      S07     I01  4.29087699 3.018906  1.1182171  0.4519197386 -0.298165480
#> 8      S08     I01  3.15535363 3.018906 -0.3213612  0.4519197386  0.005889481
#> 9      S09     I01  4.44021289 3.018906  1.0881448  0.4519197386 -0.118757251
#> 10     S10     I01  4.42537667 3.018906  0.7042877  0.4519197386  0.250263630
#> 11     S01     I02  2.36975247 3.018906 -1.3590229  0.7209477895 -0.011078070
#> 12     S02     I02  2.16678007 3.018906  0.5145774  0.7209477895 -2.087650774
#> 13     S03     I02  3.95954261 3.018906 -0.5137715  0.7209477895  0.733460662
#> 14     S04     I02  3.97414069 3.018906  0.5142260  0.7209477895 -0.279938669
#> 15     S05     I02  3.16992819 3.018906 -1.0735144  0.7209477895  0.503589226
#> 16     S06     I02  1.15699418 3.018906 -1.2838183  0.7209477895 -1.299040872
#> 17     S07     I02  4.35233367 3.018906  1.1182171  0.7209477895 -0.505736851
#> 18     S08     I02  4.48490140 3.018906 -0.3213612  0.7209477895  1.066409199
#> 19     S09     I02  5.45624623 3.018906  1.0881448  0.7209477895  0.628248034
#> 20     S10     I02  3.98969259 3.018906  0.7042877  0.7209477895 -0.454448497
#> 21     S01     I03  2.46656509 3.018906 -1.3590229 -0.0783522885  0.885034630
#> 22     S02     I03  2.15232365 3.018906  0.5145774 -0.0783522885 -1.302807122
#> 23     S03     I03  0.74951075 3.018906 -0.5137715 -0.0783522885 -1.677271118
#> 24     S04     I03  2.81825392 3.018906  0.5142260 -0.0783522885 -0.636525355
#> 25     S05     I03  2.07491652 3.018906 -1.0735144 -0.0783522885  0.207877639
#> 26     S06     I03  1.39510660 3.018906 -1.2838183 -0.0783522885 -0.261628380
#> 27     S07     I03  3.33877985 3.018906  1.1182171 -0.0783522885 -0.719990591
#> 28     S08     I03  4.13498534 3.018906 -0.3213612 -0.0783522885  1.515793226
#> 29     S09     I03  3.66173961 3.018906  1.0881448 -0.0783522885 -0.366958511
#> 30     S10     I03  3.30674901 3.018906  0.7042877 -0.0783522885 -0.338092000
#> 31     S01     I04  1.18915046 3.018906 -1.3590229 -0.5362115914  0.065479309
#> 32     S02     I04  2.42252000 3.018906  0.5145774 -0.5362115914 -0.574751462
#> 33     S03     I04  2.39414398 3.018906 -0.5137715 -0.5362115914  0.425221409
#> 34     S04     I04  1.57246516 3.018906  0.5142260 -0.5362115914 -1.424454812
#> 35     S05     I04  1.05027274 3.018906 -1.0735144 -0.5362115914 -0.358906836
#> 36     S06     I04  1.27174393 3.018906 -1.2838183 -0.5362115914  0.072868259
#> 37     S07     I04  3.07174629 3.018906  1.1182171 -0.5362115914 -0.529164844
#> 38     S08     I04 -0.90163451 3.018906 -0.3213612 -0.5362115914 -3.062967322
#> 39     S09     I04  4.33000991 3.018906  1.0881448 -0.5362115914  0.759171097
#> 40     S10     I04  4.80513637 3.018906  0.7042877 -0.5362115914  1.618154657
#> 41     S01     I05  2.19720372 3.018906 -1.3590229  0.5403205086 -0.002999534
#> 42     S02     I05  4.31733833 3.018906  0.5145774  0.5403205086  0.243534764
#> 43     S03     I05  2.57379937 3.018906 -0.5137715  0.5403205086 -0.471655296
#> 44     S04     I05  5.35179226 3.018906  0.5142260  0.5403205086  1.278340185
#> 45     S05     I05  2.47127870 3.018906 -1.0735144  0.5403205086 -0.014432980
#> 46     S06     I05  1.76350928 3.018906 -1.2838183  0.5403205086 -0.511898497
#> 47     S07     I05  5.73966362 3.018906  1.1182171  0.5403205086  1.062220385
#> 48     S08     I05  3.66252371 3.018906 -0.3213612  0.5403205086  0.424658791
#> 49     S09     I05  3.97860951 3.018906  1.0881448  0.5403205086 -0.668761406
#> 50     S10     I05  4.31752163 3.018906  0.7042877  0.5403205086  0.054007821
#> 51     S01     I06  1.65270575 3.018906 -1.3590229  0.4686248718 -0.475801871
#> 52     S02     I06  3.37854658 3.018906  0.5145774  0.4686248718 -0.623561344
#> 53     S03     I06  2.65193564 3.018906 -0.5137715  0.4686248718 -0.321823391
#> 54     S04     I06  3.02676176 3.018906  0.5142260  0.4686248718 -0.974994678
#> 55     S05     I06  1.78607000 3.018906 -1.0735144  0.4686248718 -0.627946047
#> 56     S06     I06  3.36111484 3.018906 -1.2838183  0.4686248718  1.157402697
#> 57     S07     I06  2.24562501 3.018906  1.1182171  0.4686248718 -2.360122589
#> 58     S08     I06  2.99017459 3.018906 -0.3213612  0.4686248718 -0.175994685
#> 59     S09     I06  5.11597470 3.018906  1.0881448  0.4686248718  0.540299421
#> 60     S10     I06  5.29025023 3.018906  0.7042877  0.4686248718  1.098432055
#> 61     S01     I07 -0.37886927 3.018906 -1.3590229 -0.3240288266 -1.714723188
#> 62     S02     I07  5.36403558 3.018906  0.5145774 -0.3240288266  2.154581348
#> 63     S03     I07  1.11597722 3.018906 -0.5137715 -0.3240288266 -1.065128110
#> 64     S04     I07  4.82416062 3.018906  0.5142260 -0.3240288266  1.615057879
#> 65     S05     I07  1.17260517 3.018906 -1.0735144 -0.3240288266 -0.448757176
#> 66     S06     I07  0.52404537 3.018906 -1.2838183 -0.3240288266 -0.887013071
#> 67     S07     I07  3.41241311 3.018906  1.1182171 -0.3240288266 -0.400680793
#> 68     S08     I07  4.51295781 3.018906 -0.3213612 -0.3240288266  2.139442232
#> 69     S09     I07  3.82302551 3.018906  1.0881448 -0.3240288266  0.040003927
#> 70     S10     I07  4.05997196 3.018906  0.7042877 -0.3240288266  0.660807483
#> 71     S01     I08 -1.49610431 3.018906 -1.3590229 -0.2437706311 -2.912216423
#> 72     S02     I08  4.75403279 3.018906  0.5145774 -0.2437706311  1.464320366
#> 73     S03     I08  3.15625146 3.018906 -0.5137715 -0.2437706311  0.894887934
#> 74     S04     I08  1.75065892 3.018906  0.5142260 -0.2437706311 -1.538702011
#> 75     S05     I08  1.23939314 3.018906 -1.0735144 -0.2437706311 -0.462227403
#> 76     S06     I08  1.81867315 3.018906 -1.2838183 -0.2437706311  0.327356513
#> 77     S07     I08  4.56905434 3.018906  1.1182171 -0.2437706311  0.675702246
#> 78     S08     I08  3.29100995 3.018906 -0.3213612 -0.2437706311  0.837236173
#> 79     S09     I08  3.42294649 3.018906  1.0881448 -0.2437706311 -0.440333280
#> 80     S10     I08  1.42601558 3.018906  0.7042877 -0.2437706311 -2.053407088
#> 81     S01     I09 -2.18936539 3.018906 -1.3590229 -0.3441143841 -3.505133756
#> 82     S02     I09  4.01677987 3.018906  0.5145774 -0.3441143841  0.827411202
#> 83     S03     I09  2.09880924 3.018906 -0.5137715 -0.3441143841 -0.062210540
#> 84     S04     I09  3.10334694 3.018906  0.5142260 -0.3441143841 -0.085670241
#> 85     S05     I09 -0.99957184 3.018906 -1.0735144 -0.3441143841 -2.600848623
#> 86     S06     I09 -1.15460900 3.018906 -1.2838183 -0.3441143841 -2.545581886
#> 87     S07     I09  5.08012367 3.018906  1.1182171 -0.3441143841  1.287115327
#> 88     S08     I09  3.47116590 3.018906 -0.3213612 -0.3441143841  1.117735882
#> 89     S09     I09  3.32934924 3.018906  1.0881448 -0.3441143841 -0.433586781
#> 90     S10     I09  4.27124275 3.018906  0.7042877 -0.3441143841  0.892163831
#> 91     S01     I10  1.48817230 3.018906 -1.3590229 -0.0005711141 -0.171139328
#> 92     S02     I10  4.10497579 3.018906  0.5145774 -0.0005711141  0.572063850
#> 93     S03     I10  0.43724168 3.018906 -0.5137715 -0.0005711141 -2.067321369
#> 94     S04     I10  2.98869818 3.018906  0.5142260 -0.0005711141 -0.543862266
#> 95     S05     I10  2.66104629 3.018906 -1.0735144 -0.0005711141  0.716226232
#> 96     S06     I10 -0.04455255 3.018906 -1.2838183 -0.0005711141 -1.779068708
#> 97     S07     I10  3.67659118 3.018906  1.1182171 -0.0005711141 -0.459960434
#> 98     S08     I10  3.11847675 3.018906 -0.3213612 -0.0005711141  0.421503461
#> 99     S09     I10  3.26510912 3.018906  1.0881448 -0.0005711141 -0.841370166
#> 100    S10     I10  2.06176845 3.018906  0.7042877 -0.0005711141 -1.660853742