Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id          y  grand_i      sub_i      item_i          err
#> 1      S01     I01  1.3404557 3.018906 -1.0395381 -0.70561485  0.066703019
#> 15     S01     I02  2.0524603 3.018906 -1.0395381  0.90761293 -0.834520179
#> 29     S01     I03  0.7735606 3.018906 -1.0395381 -0.65856072 -0.547246213
#> 31     S01     I04  2.7480409 3.018906 -1.0395381 -0.22679409  0.995467452
#> 44     S01     I05  0.7215181 3.018906 -1.0395381 -0.18290229 -1.074947155
#> 58     S01     I06  1.5683075 3.018906 -1.0395381  0.54499305 -0.956053104
#> 61     S01     I07  3.9603648 3.018906 -1.0395381  0.98270080  0.998296419
#> 73     S01     I08  0.3960146 3.018906 -1.0395381 -0.94891467 -0.634438274
#> 87     S01     I09  0.9220197 3.018906 -1.0395381 -0.15256392 -0.904783877
#> 91     S01     I10  3.2178820 3.018906 -1.0395381  0.09822643  1.140288003
#> 7      S02     I01  4.2051827 3.018906  0.6356574 -0.70561485  1.256234515
#> 17     S02     I02  4.5261364 3.018906  0.6356574  0.90761293 -0.036039546
#> 30     S02     I03  3.0867105 3.018906  0.6356574 -0.65856072  0.090708236
#> 36     S02     I04  0.9528578 3.018906  0.6356574 -0.22679409 -2.474911087
#> 46     S02     I05  4.8276833 3.018906  0.6356574 -0.18290229  1.356022586
#> 59     S02     I06  4.2062384 3.018906  0.6356574  0.54499305  0.006682351
#> 69     S02     I07  4.5722053 3.018906  0.6356574  0.98270080 -0.065058517
#> 75     S02     I08  4.3394540 3.018906  0.6356574 -0.94891467  1.633805697
#> 88     S02     I09  4.4049499 3.018906  0.6356574 -0.15256392  0.902950806
#> 98     S02     I10  3.1789728 3.018906  0.6356574  0.09822643 -0.573816592
#> 4      S03     I01  4.2317379 3.018906  0.4903897 -0.70561485  1.428057448
#> 18     S03     I02  4.2219784 3.018906  0.4903897  0.90761293 -0.194929844
#> 23     S03     I03  2.9469165 3.018906  0.4903897 -0.65856072  0.096181904
#> 33     S03     I04  2.3109686 3.018906  0.4903897 -0.22679409 -0.971532668
#> 47     S03     I05  3.7795330 3.018906  0.4903897 -0.18290229  0.453140005
#> 60     S03     I06  4.2841528 3.018906  0.4903897  0.54499305  0.229864453
#> 62     S03     I07  4.5124758 3.018906  0.4903897  0.98270080  0.020479653
#> 76     S03     I08  3.5292670 3.018906  0.4903897 -0.94891467  0.968886326
#> 89     S03     I09  1.9349528 3.018906  0.4903897 -0.15256392 -1.421778581
#> 95     S03     I10  4.0182095 3.018906  0.4903897  0.09822643  0.410687803
#> 5      S04     I01  1.7407086 3.018906 -0.6935307 -0.70561485  0.120948519
#> 19     S04     I02  3.4399517 3.018906 -0.6935307  0.90761293  0.206963879
#> 21     S04     I03  2.7702200 3.018906 -0.6935307 -0.65856072  1.103405820
#> 34     S04     I04  2.5443912 3.018906 -0.6935307 -0.22679409  0.445810355
#> 48     S04     I05  2.6770591 3.018906 -0.6935307 -0.18290229  0.534586464
#> 51     S04     I06  2.4519591 3.018906 -0.6935307  0.54499305 -0.418408867
#> 63     S04     I07  5.4542841 3.018906 -0.6935307  0.98270080  2.146208394
#> 77     S04     I08  1.1778318 3.018906 -0.6935307 -0.94891467 -0.198628469
#> 90     S04     I09 -0.4450252 3.018906 -0.6935307 -0.15256392 -2.617836184
#> 92     S04     I10  1.5897169 3.018906 -0.6935307  0.09822643 -0.833884462
#> 2      S05     I01  0.9028685 3.018906 -0.2761686 -0.70561485 -1.134253716
#> 16     S05     I02  3.2789540 3.018906 -0.2761686  0.90761293 -0.371395996
#> 22     S05     I03  2.3794039 3.018906 -0.2761686 -0.65856072  0.295227562
#> 35     S05     I04  3.2857927 3.018906 -0.2761686 -0.22679409  0.769849773
#> 45     S05     I05  3.2850013 3.018906 -0.2761686 -0.18290229  0.725166594
#> 52     S05     I06  2.5520458 3.018906 -0.2761686  0.54499305 -0.735684247
#> 64     S05     I07  4.0016861 3.018906 -0.2761686  0.98270080  0.276248234
#> 74     S05     I08  1.9861463 3.018906 -0.2761686 -0.94891467  0.192323884
#> 81     S05     I09  4.5607060 3.018906 -0.2761686 -0.15256392  1.970532861
#> 93     S05     I10  1.9087262 3.018906 -0.2761686  0.09822643 -0.932237285
#> 8      S06     I01  2.5917969 3.018906  0.6312269 -0.70561485 -0.352720721
#> 20     S06     I02  4.3349889 3.018906  0.6312269  0.90761293 -0.222756520
#> 27     S06     I03  0.8331370 3.018906  0.6312269 -0.65856072 -2.158434782
#> 37     S06     I04  4.2242692 3.018906  0.6312269 -0.22679409  0.800930829
#> 49     S06     I05  4.7523755 3.018906  0.6312269 -0.18290229  1.285145271
#> 56     S06     I06  4.8013799 3.018906  0.6312269  0.54499305  0.606254332
#> 66     S06     I07  3.0230193 3.018906  0.6312269  0.98270080 -1.609813969
#> 79     S06     I08  3.5749865 3.018906  0.6312269 -0.94891467  0.873768672
#> 85     S06     I09  3.3134438 3.018906  0.6312269 -0.15256392 -0.184124760
#> 99     S06     I10  4.0763852 3.018906  0.6312269  0.09822643  0.328026275
#> 9      S07     I01  1.3629624 3.018906 -0.4544631 -0.70561485 -0.495865227
#> 11     S07     I02  1.7366760 3.018906 -0.4544631  0.90761293 -1.735379465
#> 24     S07     I03 -0.7210122 3.018906 -0.4544631 -0.65856072 -2.626894030
#> 38     S07     I04  0.2164774 3.018906 -0.4544631 -0.22679409 -2.121171072
#> 50     S07     I05  2.4994573 3.018906 -0.4544631 -0.18290229  0.117917058
#> 53     S07     I06  3.8132490 3.018906 -0.4544631  0.54499305  0.703813424
#> 67     S07     I07  1.7207716 3.018906 -0.4544631  0.98270080 -1.826371718
#> 80     S07     I08  2.1969996 3.018906 -0.4544631 -0.94891467  0.581471706
#> 82     S07     I09  1.7448004 3.018906 -0.4544631 -0.15256392 -0.667078195
#> 96     S07     I10  2.8324345 3.018906 -0.4544631  0.09822643  0.169765592
#> 6      S08     I01  3.6924997 3.018906  1.2440757 -0.70561485  0.135133257
#> 12     S08     I02  5.2307488 3.018906  1.2440757  0.90761293  0.060154640
#> 25     S08     I03  4.9354303 3.018906  1.2440757 -0.65856072  1.331009799
#> 39     S08     I04  3.9268990 3.018906  1.2440757 -0.22679409 -0.109288141
#> 41     S08     I05  5.7644300 3.018906  1.2440757 -0.18290229  1.684350986
#> 54     S08     I06  5.7160866 3.018906  1.2440757  0.54499305  0.908112270
#> 68     S08     I07  4.2314381 3.018906  1.2440757  0.98270080 -1.014243985
#> 78     S08     I08  3.5854073 3.018906  1.2440757 -0.94891467  0.271340694
#> 83     S08     I09  3.5951795 3.018906  1.2440757 -0.15256392 -0.515237854
#> 97     S08     I10  3.5465382 3.018906  1.2440757  0.09822643 -0.814669534
#> 3      S09     I01  3.3429430 3.018906  0.4154072 -0.70561485  0.614245042
#> 13     S09     I02  5.3420639 3.018906  0.4154072  0.90761293  1.000138206
#> 26     S09     I03  2.7664149 3.018906  0.4154072 -0.65856072 -0.009337176
#> 32     S09     I04  4.8598584 3.018906  0.4154072 -0.22679409  1.652339720
#> 42     S09     I05  2.4848419 3.018906  0.4154072 -0.18290229 -0.766568532
#> 55     S09     I06  4.0047611 3.018906  0.4154072  0.54499305  0.025455291
#> 65     S09     I07  5.6700338 3.018906  0.4154072  0.98270080  1.253020232
#> 71     S09     I08  1.8083854 3.018906  0.4154072 -0.94891467 -0.677012674
#> 84     S09     I09  3.2372331 3.018906  0.4154072 -0.15256392 -0.044515716
#> 94     S09     I10  3.1886171 3.018906  0.4154072  0.09822643 -0.343922085
#> 10     S10     I01  1.8493198 3.018906  0.8426530 -0.70561485 -1.306624013
#> 14     S10     I02  2.7174772 3.018906  0.8426530  0.90761293 -2.051694343
#> 28     S10     I03  2.0172523 3.018906  0.8426530 -0.65856072 -1.185745642
#> 40     S10     I04  3.5344844 3.018906  0.8426530 -0.22679409 -0.100280126
#> 43     S10     I05  4.8080521 3.018906  0.8426530 -0.18290229  1.129395788
#> 57     S10     I06  5.1261917 3.018906  0.8426530  0.54499305  0.719640019
#> 70     S10     I07  5.8050533 3.018906  0.8426530  0.98270080  0.960793880
#> 72     S10     I08  2.7236747 3.018906  0.8426530 -0.94891467 -0.188969256
#> 86     S10     I09  4.0284287 3.018906  0.8426530 -0.15256392  0.319433991
#> 100    S10     I10  4.1607109 3.018906  0.8426530  0.09822643  0.200925841