Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id          y  grand_i       sub_i     item_i          err
#> 1      S01     I01  4.4357804 3.018906 -0.38094376  0.5245776  1.273240858
#> 2      S02     I01  1.8896110 3.018906 -0.99358181  0.5245776 -0.660290431
#> 3      S03     I01  3.5967389 3.018906  0.32064565  0.5245776 -0.267390000
#> 4      S04     I01  1.7010592 3.018906 -0.68977882  0.5245776 -1.152645206
#> 5      S05     I01  1.7588845 3.018906  0.74452130  0.5245776 -2.529120010
#> 6      S06     I01  3.2784155 3.018906 -0.06154169  0.5245776 -0.203526019
#> 7      S07     I01  3.1037644 3.018906  0.35626195  0.5245776 -0.795980840
#> 8      S08     I01  3.1864040 3.018906 -1.03650785  0.5245776  0.679428640
#> 9      S09     I01  3.8573163 3.018906  1.63841229  0.5245776 -1.324579231
#> 10     S10     I01  3.9630693 3.018906 -0.51805479  0.5245776  0.937640793
#> 11     S01     I02  2.2537371 3.018906 -0.38094376  0.5480155 -0.932240172
#> 12     S02     I02  2.3642438 3.018906 -0.99358181  0.5480155 -0.209095451
#> 13     S03     I02  3.3449652 3.018906  0.32064565  0.5480155 -0.542601541
#> 14     S04     I02  0.6006869 3.018906 -0.68977882  0.5480155 -2.276455331
#> 15     S05     I02  6.0766964 3.018906  0.74452130  0.5480155  1.765254050
#> 16     S06     I02  2.9725057 3.018906 -0.06154169  0.5480155 -0.532873638
#> 17     S07     I02  6.1436971 3.018906  0.35626195  0.5480155  2.220514098
#> 18     S08     I02  1.5468834 3.018906 -1.03650785  0.5480155 -0.983529781
#> 19     S09     I02  5.2041408 3.018906  1.63841229  0.5480155 -0.001192608
#> 20     S10     I02  1.7984473 3.018906 -0.51805479  0.5480155 -1.250419024
#> 21     S01     I03  3.6734055 3.018906 -0.38094376  1.0090702  0.026373469
#> 22     S02     I03  2.3829707 3.018906 -0.99358181  1.0090702 -0.651423260
#> 23     S03     I03  4.4061595 3.018906  0.32064565  1.0090702  0.057538071
#> 24     S04     I03  4.1622715 3.018906 -0.68977882  1.0090702  0.824074516
#> 25     S05     I03  3.5909604 3.018906  0.74452130  1.0090702 -1.181536732
#> 26     S06     I03  4.1882763 3.018906 -0.06154169  1.0090702  0.221842189
#> 27     S07     I03  2.6435682 3.018906  0.35626195  1.0090702 -1.740669538
#> 28     S08     I03  3.5987484 3.018906 -1.03650785  1.0090702  0.607280468
#> 29     S09     I03  6.2732246 3.018906  1.63841229  1.0090702  0.606836510
#> 30     S10     I03  4.3366840 3.018906 -0.51805479  1.0090702  0.826762993
#> 31     S01     I04  3.3167553 3.018906 -0.38094376  0.5620832  0.116710268
#> 32     S02     I04  2.5005197 3.018906 -0.99358181  0.5620832 -0.086887237
#> 33     S03     I04  3.9136358 3.018906  0.32064565  0.5620832  0.012001363
#> 34     S04     I04  2.4486712 3.018906 -0.68977882  0.5620832 -0.442538777
#> 35     S05     I04  5.4873860 3.018906  0.74452130  0.5620832  1.161875915
#> 36     S06     I04  3.1061402 3.018906 -0.06154169  0.5620832 -0.413306921
#> 37     S07     I04  2.3926067 3.018906  0.35626195  0.5620832 -1.544643976
#> 38     S08     I04  1.8296960 3.018906 -1.03650785  0.5620832 -0.714784945
#> 39     S09     I04  3.7507564 3.018906  1.63841229  0.5620832 -1.468644662
#> 40     S10     I04  3.2740925 3.018906 -0.51805479  0.5620832  0.211158541
#> 41     S01     I05  3.7346005 3.018906 -0.38094376  0.4261726  0.670466065
#> 42     S02     I05  1.3335994 3.018906 -0.99358181  0.4261726 -1.117896969
#> 43     S03     I05  4.9209945 3.018906  0.32064565  0.4261726  1.155270642
#> 44     S04     I05  2.7384245 3.018906 -0.68977882  0.4261726 -0.016874898
#> 45     S05     I05  5.1574196 3.018906  0.74452130  0.4261726  0.967820055
#> 46     S06     I05  2.2661675 3.018906 -0.06154169  0.4261726 -1.117368962
#> 47     S07     I05  3.9239921 3.018906  0.35626195  0.4261726  0.122651967
#> 48     S08     I05  1.4523968 3.018906 -1.03650785  0.4261726 -0.956173561
#> 49     S09     I05  4.7317604 3.018906  1.63841229  0.4261726 -0.351730067
#> 50     S10     I05  3.5943190 3.018906 -0.51805479  0.4261726  0.667295553
#> 51     S01     I06  2.7622621 3.018906 -0.38094376 -1.2529558  1.377256020
#> 52     S02     I06 -0.1170381 3.018906 -0.99358181 -1.2529558 -0.889406136
#> 53     S03     I06  1.7882653 3.018906  0.32064565 -1.2529558 -0.298330215
#> 54     S04     I06  1.0178041 3.018906 -0.68977882 -1.2529558 -0.058366901
#> 55     S05     I06  1.4757211 3.018906  0.74452130 -1.2529558 -1.034750052
#> 56     S06     I06  1.5936588 3.018906 -0.06154169 -1.2529558 -0.110749314
#> 57     S07     I06  0.3528574 3.018906  0.35626195 -1.2529558 -1.769354408
#> 58     S08     I06  0.7150324 3.018906 -1.03650785 -1.2529558 -0.014409558
#> 59     S09     I06  3.4775235 3.018906  1.63841229 -1.2529558  0.073161362
#> 60     S10     I06  2.4788619 3.018906 -0.51805479 -1.2529558  1.230966904
#> 61     S01     I07  4.1433240 3.018906 -0.38094376  0.5696848  0.935677338
#> 62     S02     I07  4.0113552 3.018906 -0.99358181  0.5696848  1.416346503
#> 63     S03     I07  4.0969145 3.018906  0.32064565  0.5696848  0.187678396
#> 64     S04     I07  4.9130670 3.018906 -0.68977882  0.5696848  2.014255346
#> 65     S05     I07  4.6774699 3.018906  0.74452130  0.5696848  0.344358192
#> 66     S06     I07  3.6059192 3.018906 -0.06154169  0.5696848  0.078870483
#> 67     S07     I07  2.5852189 3.018906  0.35626195  0.5696848 -1.359633457
#> 68     S08     I07  2.8385337 3.018906 -1.03650785  0.5696848  0.286451093
#> 69     S09     I07  2.6496893 3.018906  1.63841229  0.5696848 -2.577313498
#> 70     S10     I07  2.2622948 3.018906 -0.51805479  0.5696848 -0.808240813
#> 71     S01     I08  1.2859572 3.018906 -0.38094376 -1.4240384  0.072033764
#> 72     S02     I08 -0.9559706 3.018906 -0.99358181 -1.4240384 -1.557256008
#> 73     S03     I08  0.8145723 3.018906  0.32064565 -1.4240384 -1.100940522
#> 74     S04     I08  0.9190765 3.018906 -0.68977882 -1.4240384  0.013988153
#> 75     S05     I08  4.7170565 3.018906  0.74452130 -1.4240384  2.377668030
#> 76     S06     I08 -1.1186983 3.018906 -0.06154169 -1.4240384 -2.652023749
#> 77     S07     I08  1.3312520 3.018906  0.35626195 -1.4240384 -0.619877101
#> 78     S08     I08  0.8058533 3.018906 -1.03650785 -1.4240384  0.247493992
#> 79     S09     I08  3.2537028 3.018906  1.63841229 -1.4240384  0.020423317
#> 80     S10     I08 -0.1495579 3.018906 -0.51805479 -1.4240384 -1.226370315
#> 81     S01     I09  4.4509257 3.018906 -0.38094376  0.6761977  1.136766067
#> 82     S02     I09  4.2815446 3.018906 -0.99358181  0.6761977  1.580023050
#> 83     S03     I09  1.1672965 3.018906  0.32064565  0.6761977 -2.848452478
#> 84     S04     I09  2.2650994 3.018906 -0.68977882  0.6761977 -0.740225074
#> 85     S05     I09  3.2621379 3.018906  0.74452130  0.6761977 -1.177486747
#> 86     S06     I09  3.8041966 3.018906 -0.06154169  0.6761977  0.170634994
#> 87     S07     I09  3.9222501 3.018906  0.35626195  0.6761977 -0.129115221
#> 88     S08     I09  2.0104918 3.018906 -1.03650785  0.6761977 -0.648103709
#> 89     S09     I09  8.3487476 3.018906  1.63841229  0.6761977  3.015231996
#> 90     S10     I09  2.9517408 3.018906 -0.51805479  0.6761977 -0.225307752
#> 91     S01     I10  3.4145677 3.018906 -0.38094376  0.1689911  0.607614704
#> 92     S02     I10  2.4254406 3.018906 -0.99358181  0.1689911  0.231125695
#> 93     S03     I10  3.4276616 3.018906  0.32064565  0.1689911 -0.080880840
#> 94     S04     I10  2.6694079 3.018906 -0.68977882  0.1689911  0.171289967
#> 95     S05     I10  3.7180579 3.018906  0.74452130  0.1689911 -0.214360180
#> 96     S06     I10  5.4651227 3.018906 -0.06154169  0.1689911  2.338767646
#> 97     S07     I10  3.8316102 3.018906  0.35626195  0.1689911  0.287451525
#> 98     S08     I10  2.3210072 3.018906 -1.03650785  0.1689911  0.169618269
#> 99     S09     I10  5.7081539 3.018906  1.63841229  0.1689911  0.881844895
#> 100    S10     I10  1.4999409 3.018906 -0.51805479  0.1689911 -1.169901036