Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id          y  grand_i       sub_i       item_i          err
#> 1      S01     I01 3.66708902 3.018906 -1.13936834 -0.074967657  1.862519401
#> 2      S02     I01 6.26698327 3.018906  1.75644888 -0.074967657  1.566596429
#> 3      S03     I01 2.42820224 3.018906 -0.08215762 -0.074967657 -0.433578100
#> 4      S04     I01 4.28223379 3.018906  0.61579827 -0.074967657  0.722497562
#> 5      S05     I01 1.54660180 3.018906 -1.00106438 -0.074967657 -0.396271780
#> 6      S06     I01 5.58400941 3.018906  0.70488543 -0.074967657  1.935186018
#> 7      S07     I01 2.37677274 3.018906 -0.85095265 -0.074967657  0.283787434
#> 8      S08     I01 2.63626354 3.018906  0.16025479 -0.074967657 -0.467929204
#> 9      S09     I01 3.30656704 3.018906  0.62616266 -0.074967657 -0.263533580
#> 10     S10     I01 2.53703984 3.018906  0.29892029 -0.074967657 -0.705818415
#> 11     S01     I02 3.01682595 3.018906 -1.13936834  0.489797903  0.647490770
#> 12     S02     I02 3.45063847 3.018906  1.75644888  0.489797903 -1.814513929
#> 13     S03     I02 5.83412365 3.018906 -0.08215762  0.489797903  2.407577757
#> 14     S04     I02 3.57039862 3.018906  0.61579827  0.489797903 -0.554103169
#> 15     S05     I02 1.54412938 3.018906 -1.00106438  0.489797903 -0.963509760
#> 16     S06     I02 2.49153122 3.018906  0.70488543  0.489797903 -1.722057728
#> 17     S07     I02 2.80359097 3.018906 -0.85095265  0.489797903  0.145840113
#> 18     S08     I02 2.70659839 3.018906  0.16025479  0.489797903 -0.962359921
#> 19     S09     I02 4.70547926 3.018906  0.62616266  0.489797903  0.570613084
#> 20     S10     I02 3.48078564 3.018906  0.29892029  0.489797903 -0.326838175
#> 21     S01     I03 2.31295139 3.018906 -1.13936834  0.496714055 -0.063299945
#> 22     S02     I03 6.64983180 3.018906  1.75644888  0.496714055  1.377763247
#> 23     S03     I03 5.08924990 3.018906 -0.08215762  0.496714055  1.655787847
#> 24     S04     I03 4.36865082 3.018906  0.61579827  0.496714055  0.237232882
#> 25     S05     I03 3.01364995 3.018906 -1.00106438  0.496714055  0.499094659
#> 26     S06     I03 4.85341619 3.018906  0.70488543  0.496714055  0.632911092
#> 27     S07     I03 1.25703531 3.018906 -0.85095265  0.496714055 -1.407631702
#> 28     S08     I03 4.95831750 3.018906  0.16025479  0.496714055  1.282443044
#> 29     S09     I03 5.31075978 3.018906  0.62616266  0.496714055  1.168977453
#> 30     S10     I03 4.22572282 3.018906  0.29892029  0.496714055  0.411182854
#> 31     S01     I04 1.15233284 3.018906 -1.13936834 -0.543516339 -0.183688095
#> 32     S02     I04 4.52355039 3.018906  1.75644888 -0.543516339  0.291712229
#> 33     S03     I04 1.18631518 3.018906 -0.08215762 -0.543516339 -1.206916477
#> 34     S04     I04 4.96751116 3.018906  0.61579827 -0.543516339  1.876323616
#> 35     S05     I04 2.89337770 3.018906 -1.00106438 -0.543516339  1.419052799
#> 36     S06     I04 4.13509712 3.018906  0.70488543 -0.543516339  0.954822413
#> 37     S07     I04 1.28155091 3.018906 -0.85095265 -0.543516339 -0.342885708
#> 38     S08     I04 2.26950055 3.018906  0.16025479 -0.543516339 -0.366143514
#> 39     S09     I04 3.99324056 3.018906  0.62616266 -0.543516339  0.891688623
#> 40     S10     I04 0.01991926 3.018906  0.29892029 -0.543516339 -2.754390306
#> 41     S01     I05 4.36694840 3.018906 -1.13936834 -0.027795632  2.515206759
#> 42     S02     I05 4.74442370 3.018906  1.75644888 -0.027795632 -0.003135162
#> 43     S03     I05 2.87854687 3.018906 -0.08215762 -0.027795632 -0.030405492
#> 44     S04     I05 4.21006243 3.018906  0.61579827 -0.027795632  0.603154185
#> 45     S05     I05 2.94698300 3.018906 -1.00106438 -0.027795632  0.956937395
#> 46     S06     I05 1.44025711 3.018906  0.70488543 -0.027795632 -2.255738303
#> 47     S07     I05 0.77026099 3.018906 -0.85095265 -0.027795632 -1.369896338
#> 48     S08     I05 3.60480714 3.018906  0.16025479 -0.027795632  0.453442367
#> 49     S09     I05 5.12085178 3.018906  0.62616266 -0.027795632  1.503579144
#> 50     S10     I05 1.81898015 3.018906  0.29892029 -0.027795632 -1.471050124
#> 51     S01     I06 3.19816633 3.018906 -1.13936834  0.154666841  1.163962217
#> 52     S02     I06 3.26544606 3.018906  1.75644888  0.154666841 -1.664575281
#> 53     S03     I06 2.94287789 3.018906 -0.08215762  0.154666841 -0.148536940
#> 54     S04     I06 4.34391589 3.018906  0.61579827  0.154666841  0.554545170
#> 55     S05     I06 2.97764740 3.018906 -1.00106438  0.154666841  0.805139323
#> 56     S06     I06 3.60063655 3.018906  0.70488543  0.154666841 -0.277821336
#> 57     S07     I06 2.24298655 3.018906 -0.85095265  0.154666841 -0.079633247
#> 58     S08     I06 4.15242422 3.018906  0.16025479  0.154666841  0.818596970
#> 59     S09     I06 4.19964396 3.018906  0.62616266  0.154666841  0.399908850
#> 60     S10     I06 2.59651298 3.018906  0.29892029  0.154666841 -0.875979765
#> 61     S01     I07 2.63332110 3.018906 -1.13936834  1.083835175 -0.330051352
#> 62     S02     I07 5.13938428 3.018906  1.75644888  1.083835175 -0.719805394
#> 63     S03     I07 3.30780970 3.018906 -0.08215762  1.083835175 -0.712773463
#> 64     S04     I07 3.98060584 3.018906  0.61579827  1.083835175 -0.737933218
#> 65     S05     I07 3.58558074 3.018906 -1.00106438  1.083835175  0.483904328
#> 66     S06     I07 3.75453415 3.018906  0.70488543  1.083835175 -1.053092071
#> 67     S07     I07 3.58614761 3.018906 -0.85095265  1.083835175  0.334359479
#> 68     S08     I07 4.55577747 3.018906  0.16025479  1.083835175  0.292781888
#> 69     S09     I07 4.60335244 3.018906  0.62616266  1.083835175 -0.125551009
#> 70     S10     I07 4.76641263 3.018906  0.29892029  1.083835175  0.364751545
#> 71     S01     I08 0.36146832 3.018906 -1.13936834  0.005096503 -1.523165455
#> 72     S02     I08 7.13298042 3.018906  1.75644888  0.005096503  2.352529424
#> 73     S03     I08 4.21625071 3.018906 -0.08215762  0.005096503  1.274406212
#> 74     S04     I08 1.47239231 3.018906  0.61579827  0.005096503 -2.167408075
#> 75     S05     I08 0.66159040 3.018906 -1.00106438  0.005096503 -1.361347343
#> 76     S06     I08 4.49762696 3.018906  0.70488543  0.005096503  0.768739408
#> 77     S07     I08 4.40264425 3.018906 -0.85095265  0.005096503  2.229594792
#> 78     S08     I08 3.52019231 3.018906  0.16025479  0.005096503  0.335935399
#> 79     S09     I08 3.78550578 3.018906  0.62616266  0.005096503  0.135341004
#> 80     S10     I08 4.14805323 3.018906  0.29892029  0.005096503  0.825130822
#> 81     S01     I09 0.72863886 3.018906 -1.13936834 -0.085293284 -1.065605133
#> 82     S02     I09 5.02068175 3.018906  1.75644888 -0.085293284  0.330620536
#> 83     S03     I09 3.76239145 3.018906 -0.08215762 -0.085293284  0.910936742
#> 84     S04     I09 3.77410319 3.018906  0.61579827 -0.085293284  0.224692591
#> 85     S05     I09 1.80776565 3.018906 -1.00106438 -0.085293284 -0.124782301
#> 86     S06     I09 4.58013009 3.018906  0.70488543 -0.085293284  0.941632328
#> 87     S07     I09 4.02331241 3.018906 -0.85095265 -0.085293284  1.940652739
#> 88     S08     I09 2.15827674 3.018906  0.16025479 -0.085293284 -0.935590381
#> 89     S09     I09 4.38293165 3.018906  0.62616266 -0.085293284  0.823156662
#> 90     S10     I09 1.98983147 3.018906  0.29892029 -0.085293284 -1.242701156
#> 91     S01     I10 2.22476832 3.018906 -1.13936834 -0.040926509  0.386157553
#> 92     S02     I10 6.58674117 3.018906  1.75644888 -0.040926509  1.852313183
#> 93     S03     I10 1.42561782 3.018906 -0.08215762 -0.040926509 -1.470203664
#> 94     S04     I10 2.89550831 3.018906  0.61579827 -0.040926509 -0.698269066
#> 95     S05     I10 1.81538779 3.018906 -1.00106438 -0.040926509 -0.161526932
#> 96     S06     I10 5.54688464 3.018906  0.70488543 -0.040926509  1.864020107
#> 97     S07     I10 3.41424889 3.018906 -0.85095265 -0.040926509  1.287222441
#> 98     S08     I10 3.17301576 3.018906  0.16025479 -0.040926509  0.034781867
#> 99     S09     I10 3.51645046 3.018906  0.62616266 -0.040926509 -0.087691298
#> 100    S10     I10 4.05821754 3.018906  0.29892029 -0.040926509  0.781318140