Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id            y  grand_i       sub_i      item_i         err
#> 1      S01     I01  4.424421612 3.018906  0.27574486  1.10087374  0.02889740
#> 15     S01     I02  3.848263802 3.018906  0.27574486 -0.29432480  0.84793813
#> 29     S01     I03  3.785327270 3.018906  0.27574486  0.66207708 -0.17140028
#> 31     S01     I04  3.943650655 3.018906  0.27574486 -0.06026755  0.70926774
#> 44     S01     I05  3.811372999 3.018906  0.27574486  0.35539173  0.16133080
#> 58     S01     I06  2.262307572 3.018906  0.27574486  0.47775678 -1.51009968
#> 61     S01     I07  3.025427855 3.018906  0.27574486  0.16918745 -0.43841007
#> 73     S01     I08  0.000458143 3.018906  0.27574486 -1.19288132 -2.10131101
#> 87     S01     I09  2.865387300 3.018906  0.27574486 -1.39150630  0.96224313
#> 91     S01     I10  4.537730307 3.018906  0.27574486  0.82765762  0.41542222
#> 7      S02     I01  6.280560237 3.018906  1.66441057  1.10087374  0.49637031
#> 17     S02     I02  4.689140856 3.018906  1.66441057 -0.29432480  0.30014947
#> 30     S02     I03  6.588109489 3.018906  1.66441057  0.66207708  1.24271622
#> 36     S02     I04  4.728864675 3.018906  1.66441057 -0.06026755  0.10581605
#> 46     S02     I05  5.385213547 3.018906  1.66441057  0.35539173  0.34650563
#> 59     S02     I06  3.796701545 3.018906  1.66441057  0.47775678 -1.36437142
#> 69     S02     I07  5.260397660 3.018906  1.66441057  0.16918745  0.40789402
#> 75     S02     I08  5.570571780 3.018906  1.66441057 -1.19288132  2.08013692
#> 88     S02     I09  3.085937140 3.018906  1.66441057 -1.39150630 -0.20587274
#> 98     S02     I10  5.270496368 3.018906  1.66441057  0.82765762 -0.24047743
#> 4      S03     I01  5.031856389 3.018906 -0.64871111  1.10087374  1.56078814
#> 18     S03     I02  3.535569428 3.018906 -0.64871111 -0.29432480  1.45969972
#> 23     S03     I03  1.611172105 3.018906 -0.64871111  0.66207708 -1.42109948
#> 33     S03     I04  1.686115558 3.018906 -0.64871111 -0.06026755 -0.62381139
#> 47     S03     I05  2.818068358 3.018906 -0.64871111  0.35539173  0.09248213
#> 60     S03     I06  2.671771114 3.018906 -0.64871111  0.47775678 -0.17618017
#> 62     S03     I07  2.207099356 3.018906 -0.64871111  0.16918745 -0.33228260
#> 76     S03     I08  1.456642777 3.018906 -0.64871111 -1.19288132  0.27932960
#> 89     S03     I09 -0.295417588 3.018906 -0.64871111 -1.39150630 -1.27410579
#> 95     S03     I10  2.031117195 3.018906 -0.64871111  0.82765762 -1.16673493
#> 5      S04     I01  6.259332482 3.018906  0.76380851  1.10087374  1.37574462
#> 19     S04     I02  3.212498501 3.018906  0.76380851 -0.29432480 -0.27589082
#> 21     S04     I03  4.516015018 3.018906  0.76380851  0.66207708  0.07122382
#> 34     S04     I04  3.194254925 3.018906  0.76380851 -0.06026755 -0.52819164
#> 48     S04     I05  2.551205503 3.018906  0.76380851  0.35539173 -1.58690035
#> 51     S04     I06  4.877556330 3.018906  0.76380851  0.47775678  0.61708542
#> 63     S04     I07  2.697940782 3.018906  0.76380851  0.16918745 -1.25396080
#> 77     S04     I08  1.947861475 3.018906  0.76380851 -1.19288132 -0.64197132
#> 90     S04     I09  2.289380845 3.018906  0.76380851 -1.39150630 -0.10182698
#> 92     S04     I10  5.118078235 3.018906  0.76380851  0.82765762  0.50770650
#> 2      S05     I01  4.457456023 3.018906  0.39888947  1.10087374 -0.06121280
#> 16     S05     I02  3.513413913 3.018906  0.39888947 -0.29432480  0.38994363
#> 22     S05     I03  5.424294759 3.018906  0.39888947  0.66207708  1.34442260
#> 35     S05     I04  2.530743258 3.018906  0.39888947 -0.06026755 -0.82678427
#> 45     S05     I05  4.567672071 3.018906  0.39888947  0.35539173  0.79448526
#> 52     S05     I06  7.928758083 3.018906  0.39888947  0.47775678  4.03320622
#> 64     S05     I07  4.693960808 3.018906  0.39888947  0.16918745  1.10697827
#> 74     S05     I08  4.258985161 3.018906  0.39888947 -1.19288132  2.03407141
#> 81     S05     I09  1.757036357 3.018906  0.39888947 -1.39150630 -0.26925242
#> 93     S05     I10  5.990775983 3.018906  0.39888947  0.82765762  1.74532329
#> 8      S06     I01  3.960750402 3.018906  1.29685525  1.10087374 -1.45588420
#> 20     S06     I02  4.596527163 3.018906  1.29685525 -0.29432480  0.57509110
#> 27     S06     I03  7.013250157 3.018906  1.29685525  0.66207708  2.03541222
#> 37     S06     I04  4.442822051 3.018906  1.29685525 -0.06026755  0.18732875
#> 49     S06     I05  4.723494160 3.018906  1.29685525  0.35539173  0.05234157
#> 56     S06     I06  3.728508472 3.018906  1.29685525  0.47775678 -1.06500917
#> 66     S06     I07  4.665463248 3.018906  1.29685525  0.16918745  0.18051494
#> 79     S06     I08  3.628043346 3.018906  1.29685525 -1.19288132  0.50516381
#> 85     S06     I09  4.508358129 3.018906  1.29685525 -1.39150630  1.58410357
#> 99     S06     I10  5.635571960 3.018906  1.29685525  0.82765762  0.49215349
#> 9      S07     I01  2.826082584 3.018906  0.15339455  1.10087374 -1.44709132
#> 11     S07     I02  3.843341231 3.018906  0.15339455 -0.29432480  0.96536587
#> 24     S07     I03  5.220462235 3.018906  0.15339455  0.66207708  1.38608500
#> 38     S07     I04  2.683770548 3.018906  0.15339455 -0.06026755 -0.42826206
#> 50     S07     I05  2.526762314 3.018906  0.15339455  0.35539173 -1.00092957
#> 53     S07     I06  3.287777605 3.018906  0.15339455  0.47775678 -0.36227934
#> 67     S07     I07  3.184504882 3.018906  0.15339455  0.16918745 -0.15698273
#> 80     S07     I08  3.118143707 3.018906  0.15339455 -1.19288132  1.13872487
#> 82     S07     I09  2.308340330 3.018906  0.15339455 -1.39150630  0.52754647
#> 96     S07     I10  4.377227313 3.018906  0.15339455  0.82765762  0.37726954
#> 6      S08     I01  3.969224023 3.018906  0.07060137  1.10087374 -0.22115670
#> 12     S08     I02  2.493360383 3.018906  0.07060137 -0.29432480 -0.30182180
#> 25     S08     I03  5.275960411 3.018906  0.07060137  0.66207708  1.52437635
#> 39     S08     I04  1.773591175 3.018906  0.07060137 -0.06026755 -1.25564825
#> 41     S08     I05  3.967957794 3.018906  0.07060137  0.35539173  0.52305908
#> 54     S08     I06  5.616046665 3.018906  0.07060137  0.47775678  2.04878290
#> 68     S08     I07  3.287270260 3.018906  0.07060137  0.16918745  0.02857582
#> 78     S08     I08 -0.205105787 3.018906  0.07060137 -1.19288132 -2.10173145
#> 83     S08     I09  3.096950233 3.018906  0.07060137 -1.39150630  1.39894955
#> 97     S08     I10  5.691103943 3.018906  0.07060137  0.82765762  1.77393934
#> 3      S09     I01  5.075011603 3.018906 -0.46411811  1.10087374  1.41935035
#> 13     S09     I02  1.772426955 3.018906 -0.46411811 -0.29432480 -0.48803575
#> 26     S09     I03  4.293142153 3.018906 -0.46411811  0.66207708  1.07627757
#> 32     S09     I04  2.088943044 3.018906 -0.46411811 -0.06026755 -0.40557691
#> 42     S09     I05  2.692697114 3.018906 -0.46411811  0.35539173 -0.21748212
#> 55     S09     I06  3.803878732 3.018906 -0.46411811  0.47775678  0.77133444
#> 65     S09     I07  1.726367973 3.018906 -0.46411811  0.16918745 -0.99760699
#> 71     S09     I08  1.032356700 3.018906 -0.46411811 -1.19288132 -0.32954948
#> 84     S09     I09  1.052686107 3.018906 -0.46411811 -1.39150630 -0.11059510
#> 94     S09     I10  1.651271125 3.018906 -0.46411811  0.82765762 -1.73117400
#> 10     S10     I01  5.193669245 3.018906  1.22722480  1.10087374 -0.15333491
#> 14     S10     I02  4.532456567 3.018906  1.22722480 -0.29432480  0.58065095
#> 28     S10     I03  6.678553366 3.018906  1.22722480  0.66207708  1.77034588
#> 40     S10     I04  4.628109792 3.018906  1.22722480 -0.06026755  0.44224694
#> 43     S10     I05  5.436139705 3.018906  1.22722480  0.35539173  0.83461757
#> 57     S10     I06  3.696776711 3.018906  1.22722480  0.47775678 -1.02711048
#> 70     S10     I07  4.951661119 3.018906  1.22722480  0.16918745  0.53634325
#> 72     S10     I08  3.652422232 3.018906  1.22722480 -1.19288132  0.59917315
#> 86     S10     I09  1.347284756 3.018906  1.22722480 -1.39150630 -1.50733935
#> 100    S10     I10  6.756228420 3.018906  1.22722480  0.82765762  1.68244039