Generate a sample with random intercepts for subjects and items

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(.data, sub_n = 100, item_n = 25, dv = 1, sub_id = 2,
  item_id = 3)

Arguments

.data
the existing tbl

sub_n
the number of subjects to simulate

item_n
the number of items to simulate

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id      sub_i        item_i         dv
#> 1        1       1 -0.2396229 -0.6240398122  2.8518221
#> 2        2       1 -0.5442409 -0.6240398122  1.9668874
#> 3        3       1 -1.0690789 -0.6240398122 -0.2629545
#> 4        4       1 -0.1472031 -0.6240398122  2.7428086
#> 5        5       1 -1.8498327 -0.6240398122  1.9288361
#> 6        6       1  0.4872866 -0.6240398122  0.7987672
#> 7        7       1  0.0527697 -0.6240398122  2.7039147
#> 8        8       1  1.4916886 -0.6240398122  5.0352024
#> 9        9       1  1.1627588 -0.6240398122  4.0272131
#> 10      10       1  1.3705433 -0.6240398122  1.8661317
#> 11       1       2 -0.2396229  0.0737720044  5.7190067
#> 12       2       2 -0.5442409  0.0737720044  2.5149279
#> 13       3       2 -1.0690789  0.0737720044 -0.4307355
#> 14       4       2 -0.1472031  0.0737720044  4.9436124
#> 15       5       2 -1.8498327  0.0737720044  1.7851634
#> 16       6       2  0.4872866  0.0737720044  3.7499332
#> 17       7       2  0.0527697  0.0737720044  3.7508227
#> 18       8       2  1.4916886  0.0737720044  5.3318171
#> 19       9       2  1.1627588  0.0737720044  0.6610708
#> 20      10       2  1.3705433  0.0737720044  5.4186885
#> 21       1       3 -0.2396229  0.7077325203  5.0765159
#> 22       2       3 -0.5442409  0.7077325203  3.0633821
#> 23       3       3 -1.0690789  0.7077325203  4.0995844
#> 24       4       3 -0.1472031  0.7077325203  4.8202937
#> 25       5       3 -1.8498327  0.7077325203  0.8647338
#> 26       6       3  0.4872866  0.7077325203  4.0022090
#> 27       7       3  0.0527697  0.7077325203  4.7845643
#> 28       8       3  1.4916886  0.7077325203  4.9096337
#> 29       9       3  1.1627588  0.7077325203  5.5897579
#> 30      10       3  1.3705433  0.7077325203  6.4079320
#> 31       1       4 -0.2396229  0.3657847830  2.0720024
#> 32       2       4 -0.5442409  0.3657847830  4.3231517
#> 33       3       4 -1.0690789  0.3657847830  1.6225195
#> 34       4       4 -0.1472031  0.3657847830  1.8747657
#> 35       5       4 -1.8498327  0.3657847830  0.9092096
#> 36       6       4  0.4872866  0.3657847830  2.0484583
#> 37       7       4  0.0527697  0.3657847830  3.7233531
#> 38       8       4  1.4916886  0.3657847830  3.3307094
#> 39       9       4  1.1627588  0.3657847830  5.3892575
#> 40      10       4  1.3705433  0.3657847830  4.0696969
#> 41       1       5 -0.2396229 -0.3051109814 -0.2706484
#> 42       2       5 -0.5442409 -0.3051109814  3.3070643
#> 43       3       5 -1.0690789 -0.3051109814  2.4620409
#> 44       4       5 -0.1472031 -0.3051109814  3.4348493
#> 45       5       5 -1.8498327 -0.3051109814 -0.8413550
#> 46       6       5  0.4872866 -0.3051109814  1.9528294
#> 47       7       5  0.0527697 -0.3051109814  3.2404056
#> 48       8       5  1.4916886 -0.3051109814  5.3511790
#> 49       9       5  1.1627588 -0.3051109814  4.9246782
#> 50      10       5  1.3705433 -0.3051109814  4.8392120
#> 51       1       6 -0.2396229  0.3314753477  2.8604816
#> 52       2       6 -0.5442409  0.3314753477  1.7884178
#> 53       3       6 -1.0690789  0.3314753477  1.2994108
#> 54       4       6 -0.1472031  0.3314753477  3.1696075
#> 55       5       6 -1.8498327  0.3314753477  0.5555800
#> 56       6       6  0.4872866  0.3314753477  6.2175933
#> 57       7       6  0.0527697  0.3314753477  5.0767993
#> 58       8       6  1.4916886  0.3314753477  2.7947075
#> 59       9       6  1.1627588  0.3314753477  5.5094497
#> 60      10       6  1.3705433  0.3314753477  4.2839022
#> 61       1       7 -0.2396229  0.6568782867  2.0689045
#> 62       2       7 -0.5442409  0.6568782867  4.4199572
#> 63       3       7 -1.0690789  0.6568782867  3.7038534
#> 64       4       7 -0.1472031  0.6568782867  2.3687959
#> 65       5       7 -1.8498327  0.6568782867  0.6508695
#> 66       6       7  0.4872866  0.6568782867  2.5024103
#> 67       7       7  0.0527697  0.6568782867  2.2378004
#> 68       8       7  1.4916886  0.6568782867  3.7602979
#> 69       9       7  1.1627588  0.6568782867  1.6174663
#> 70      10       7  1.3705433  0.6568782867  6.2707761
#> 71       1       8 -0.2396229 -0.1315365784  2.6521274
#> 72       2       8 -0.5442409 -0.1315365784  2.5200856
#> 73       3       8 -1.0690789 -0.1315365784  0.2040451
#> 74       4       8 -0.1472031 -0.1315365784  4.0073279
#> 75       5       8 -1.8498327 -0.1315365784 -0.3466308
#> 76       6       8  0.4872866 -0.1315365784  2.1630615
#> 77       7       8  0.0527697 -0.1315365784  1.5184225
#> 78       8       8  1.4916886 -0.1315365784  4.1895499
#> 79       9       8  1.1627588 -0.1315365784  3.3068711
#> 80      10       8  1.3705433 -0.1315365784  3.5817560
#> 81       1       9 -0.2396229 -0.0006109263  2.6098816
#> 82       2       9 -0.5442409 -0.0006109263  1.2921410
#> 83       3       9 -1.0690789 -0.0006109263  2.5477599
#> 84       4       9 -0.1472031 -0.0006109263  4.3883662
#> 85       5       9 -1.8498327 -0.0006109263 -0.5593484
#> 86       6       9  0.4872866 -0.0006109263  5.7547464
#> 87       7       9  0.0527697 -0.0006109263  1.2249057
#> 88       8       9  1.4916886 -0.0006109263  5.4460634
#> 89       9       9  1.1627588 -0.0006109263  3.7680685
#> 90      10       9  1.3705433 -0.0006109263  6.3987326
#> 91       1      10 -0.2396229  0.6643744909  4.5724435
#> 92       2      10 -0.5442409  0.6643744909  1.5404019
#> 93       3      10 -1.0690789  0.6643744909 -0.4066272
#> 94       4      10 -0.1472031  0.6643744909  3.4523366
#> 95       5      10 -1.8498327  0.6643744909  0.7951289
#> 96       6      10  0.4872866  0.6643744909  4.9414845
#> 97       7      10  0.0527697  0.6643744909  5.6692058
#> 98       8      10  1.4916886  0.6643744909  6.1173654
#> 99       9      10  1.1627588  0.6643744909  5.0124822
#> 100     10      10  1.3705433  0.6643744909  5.0033603