Generate a sample with random intercepts for subjects and items

sim_mixed_df Produces a dataframe with the same distributions of by-subject and by-item random intercepts as an existing dataframe

sim_mixed_df(dat, sub_n = 100, item_n = 25, dv = 1, sub_id = 2,
  item_id = 3)

Arguments

dat
the existing dataframe

sub_n
the number of subjects to simulate

item_n
the number of items to simulate

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

tibble

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id       sub_i      item_i         dv
#> 1        1       1 -1.31204794 -0.91527468 -0.5821852
#> 2        2       1  0.58686412 -0.91527468  3.6621355
#> 3        3       1  0.03209700 -0.91527468  1.1597510
#> 4        4       1  0.19719216 -0.91527468  2.9041436
#> 5        5       1  0.85825196 -0.91527468  1.6956532
#> 6        6       1  1.23958914 -0.91527468  3.3635152
#> 7        7       1 -0.04067057 -0.91527468  2.4206597
#> 8        8       1 -0.49972036 -0.91527468  1.2830612
#> 9        9       1 -0.39832107 -0.91527468  2.4032570
#> 10      10       1 -0.27425086 -0.91527468  2.8374416
#> 11       1       2 -1.31204794 -0.75442016  2.8801949
#> 12       2       2  0.58686412 -0.75442016  4.6594123
#> 13       3       2  0.03209700 -0.75442016  2.3269642
#> 14       4       2  0.19719216 -0.75442016  4.2190505
#> 15       5       2  0.85825196 -0.75442016  1.6568714
#> 16       6       2  1.23958914 -0.75442016  3.7147063
#> 17       7       2 -0.04067057 -0.75442016  2.6823381
#> 18       8       2 -0.49972036 -0.75442016  1.2207087
#> 19       9       2 -0.39832107 -0.75442016  0.1317110
#> 20      10       2 -0.27425086 -0.75442016  3.5317627
#> 21       1       3 -1.31204794 -0.52699466  2.3226909
#> 22       2       3  0.58686412 -0.52699466  2.4430166
#> 23       3       3  0.03209700 -0.52699466  1.9510354
#> 24       4       3  0.19719216 -0.52699466  2.5561268
#> 25       5       3  0.85825196 -0.52699466  2.7123415
#> 26       6       3  1.23958914 -0.52699466  3.4618648
#> 27       7       3 -0.04067057 -0.52699466  3.3228664
#> 28       8       3 -0.49972036 -0.52699466  4.7390622
#> 29       9       3 -0.39832107 -0.52699466  2.3873381
#> 30      10       3 -0.27425086 -0.52699466  1.2534605
#> 31       1       4 -1.31204794 -0.02510984  0.8446420
#> 32       2       4  0.58686412 -0.02510984  4.0745936
#> 33       3       4  0.03209700 -0.02510984  2.2571747
#> 34       4       4  0.19719216 -0.02510984  2.1582137
#> 35       5       4  0.85825196 -0.02510984  2.6919926
#> 36       6       4  1.23958914 -0.02510984  3.2690829
#> 37       7       4 -0.04067057 -0.02510984  3.6804353
#> 38       8       4 -0.49972036 -0.02510984  0.2449312
#> 39       9       4 -0.39832107 -0.02510984  4.4384063
#> 40      10       4 -0.27425086 -0.02510984  2.5131083
#> 41       1       5 -1.31204794  1.14363329  3.3097340
#> 42       2       5  0.58686412  1.14363329  4.0280504
#> 43       3       5  0.03209700  1.14363329  3.4736644
#> 44       4       5  0.19719216  1.14363329  4.4829729
#> 45       5       5  0.85825196  1.14363329  6.0465548
#> 46       6       5  1.23958914  1.14363329  4.0794730
#> 47       7       5 -0.04067057  1.14363329  4.0704548
#> 48       8       5 -0.49972036  1.14363329  3.4539076
#> 49       9       5 -0.39832107  1.14363329  4.4518836
#> 50      10       5 -0.27425086  1.14363329  3.7230423
#> 51       1       6 -1.31204794  0.02181121  1.7452042
#> 52       2       6  0.58686412  0.02181121  4.5943491
#> 53       3       6  0.03209700  0.02181121  3.0840564
#> 54       4       6  0.19719216  0.02181121  3.6131343
#> 55       5       6  0.85825196  0.02181121  2.9055922
#> 56       6       6  1.23958914  0.02181121  3.2747244
#> 57       7       6 -0.04067057  0.02181121  2.5952921
#> 58       8       6 -0.49972036  0.02181121  2.2170451
#> 59       9       6 -0.39832107  0.02181121  4.3551644
#> 60      10       6 -0.27425086  0.02181121  0.6118565
#> 61       1       7 -1.31204794 -0.53113887  0.9948724
#> 62       2       7  0.58686412 -0.53113887  3.2054074
#> 63       3       7  0.03209700 -0.53113887  1.8000139
#> 64       4       7  0.19719216 -0.53113887  2.1077354
#> 65       5       7  0.85825196 -0.53113887  3.0181285
#> 66       6       7  1.23958914 -0.53113887  3.0721583
#> 67       7       7 -0.04067057 -0.53113887  2.1009421
#> 68       8       7 -0.49972036 -0.53113887  2.3233091
#> 69       9       7 -0.39832107 -0.53113887  2.9582357
#> 70      10       7 -0.27425086 -0.53113887  4.1178312
#> 71       1       8 -1.31204794 -0.69069006  1.3048000
#> 72       2       8  0.58686412 -0.69069006  3.2600878
#> 73       3       8  0.03209700 -0.69069006  2.1985137
#> 74       4       8  0.19719216 -0.69069006  2.8759249
#> 75       5       8  0.85825196 -0.69069006  2.8337853
#> 76       6       8  1.23958914 -0.69069006  2.7156484
#> 77       7       8 -0.04067057 -0.69069006  2.2678694
#> 78       8       8 -0.49972036 -0.69069006  1.6444369
#> 79       9       8 -0.39832107 -0.69069006  1.7506832
#> 80      10       8 -0.27425086 -0.69069006  2.8302596
#> 81       1       9 -1.31204794 -0.41158246  1.5533167
#> 82       2       9  0.58686412 -0.41158246  0.6675154
#> 83       3       9  0.03209700 -0.41158246  2.7254063
#> 84       4       9  0.19719216 -0.41158246  2.3380499
#> 85       5       9  0.85825196 -0.41158246  3.0601278
#> 86       6       9  1.23958914 -0.41158246  3.7980488
#> 87       7       9 -0.04067057 -0.41158246  4.0910659
#> 88       8       9 -0.49972036 -0.41158246  1.7803072
#> 89       9       9 -0.39832107 -0.41158246  1.1584322
#> 90      10       9 -0.27425086 -0.41158246  4.1641162
#> 91       1      10 -1.31204794  0.45071102  1.3128191
#> 92       2      10  0.58686412  0.45071102  4.3255009
#> 93       3      10  0.03209700  0.45071102  2.4496289
#> 94       4      10  0.19719216  0.45071102  2.5378068
#> 95       5      10  0.85825196  0.45071102  3.1817382
#> 96       6      10  1.23958914  0.45071102  4.4856500
#> 97       7      10 -0.04067057  0.45071102  4.8521086
#> 98       8      10 -0.49972036  0.45071102  3.5287348
#> 99       9      10 -0.39832107  0.45071102  1.5642742
#> 100     10      10 -0.27425086  0.45071102  2.1061951