Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id          y  grand_i       sub_i      item_i         err
#> 1      S01     I01  3.1278258 3.018906 -1.67684152  0.08138621  1.70437549
#> 15     S01     I02  2.5178021 3.018906 -1.67684152  0.60558819  0.57014980
#> 29     S01     I03  1.7924090 3.018906 -1.67684152 -0.94560820  1.39595314
#> 31     S01     I04  1.3811393 3.018906 -1.67684152 -0.10053204  0.13960723
#> 44     S01     I05  1.4350208 3.018906 -1.67684152  0.85725988 -0.76430316
#> 58     S01     I06  1.8878330 3.018906 -1.67684152 -0.89406578  1.43983472
#> 61     S01     I07  1.7265937 3.018906 -1.67684152  0.55564089 -0.17111125
#> 73     S01     I08  3.4946301 3.018906 -1.67684152  1.04326387  1.10930214
#> 87     S01     I09  1.3662224 3.018906 -1.67684152 -0.02055614  0.04471447
#> 91     S01     I10  1.5530898 3.018906 -1.67684152 -0.07791713  0.28894282
#> 7      S02     I01  4.1879489 3.018906  0.60850587  0.08138621  0.47915121
#> 17     S02     I02  5.0775457 3.018906  0.60850587  0.60558819  0.84454599
#> 30     S02     I03  2.1395769 3.018906  0.60850587 -0.94560820 -0.54222637
#> 36     S02     I04  4.8443791 3.018906  0.60850587 -0.10053204  1.31749967
#> 46     S02     I05  2.9145828 3.018906  0.60850587  0.85725988 -1.57008858
#> 59     S02     I06  2.6521829 3.018906  0.60850587 -0.89406578 -0.08116277
#> 69     S02     I07  3.6300258 3.018906  0.60850587  0.55564089 -0.55302653
#> 75     S02     I08  4.7799447 3.018906  0.60850587  1.04326387  0.10926931
#> 88     S02     I09  2.7762793 3.018906  0.60850587 -0.02055614 -0.83057599
#> 98     S02     I10  1.9059712 3.018906  0.60850587 -0.07791713 -1.64352314
#> 4      S03     I01  3.0338978 3.018906  0.04150974  0.08138621 -0.10790379
#> 18     S03     I02  4.8014058 3.018906  0.04150974  0.60558819  1.13540230
#> 23     S03     I03  3.2591527 3.018906  0.04150974 -0.94560820  1.14434553
#> 33     S03     I04  5.3792716 3.018906  0.04150974 -0.10053204  2.41938831
#> 47     S03     I05  4.0445539 3.018906  0.04150974  0.85725988  0.12687869
#> 60     S03     I06  2.9155580 3.018906  0.04150974 -0.89406578  0.74920838
#> 62     S03     I07  6.8361553 3.018906  0.04150974  0.55564089  3.22009907
#> 76     S03     I08  4.3673534 3.018906  0.04150974  1.04326387  0.26367420
#> 89     S03     I09  2.9644835 3.018906  0.04150974 -0.02055614 -0.07537573
#> 95     S03     I10  2.0292929 3.018906  0.04150974 -0.07791713 -0.95320530
#> 5      S04     I01  2.5587788 3.018906  0.19403865  0.08138621 -0.73555167
#> 19     S04     I02  4.6982347 3.018906  0.19403865  0.60558819  0.87970225
#> 21     S04     I03  1.2811414 3.018906  0.19403865 -0.94560820 -0.98619464
#> 34     S04     I04  3.8024529 3.018906  0.19403865 -0.10053204  0.69004066
#> 48     S04     I05  4.3468109 3.018906  0.19403865  0.85725988  0.27660674
#> 51     S04     I06  0.8934603 3.018906  0.19403865 -0.89406578 -1.42541820
#> 63     S04     I07  3.8568434 3.018906  0.19403865  0.55564089  0.08825825
#> 77     S04     I08  4.5564114 3.018906  0.19403865  1.04326387  0.30020324
#> 90     S04     I09  3.0499571 3.018906  0.19403865 -0.02055614 -0.14243104
#> 92     S04     I10  2.2026983 3.018906  0.19403865 -0.07791713 -0.93232884
#> 2      S05     I01  1.8270154 3.018906 -1.85428145  0.08138621  0.58100506
#> 16     S05     I02  0.9234527 3.018906 -1.85428145  0.60558819 -0.84675970
#> 22     S05     I03 -0.9951084 3.018906 -1.85428145 -0.94560820 -1.21412438
#> 35     S05     I04  0.8023149 3.018906 -1.85428145 -0.10053204 -0.26177717
#> 45     S05     I05  0.7741331 3.018906 -1.85428145  0.85725988 -1.24775094
#> 52     S05     I06  2.2533490 3.018906 -1.85428145 -0.89406578  1.98279063
#> 64     S05     I07  1.7990978 3.018906 -1.85428145  0.55564089  0.07883275
#> 74     S05     I08  0.5285580 3.018906 -1.85428145  1.04326387 -1.67933005
#> 81     S05     I09  2.4034940 3.018906 -1.85428145 -0.02055614  1.25942600
#> 93     S05     I10  1.2385019 3.018906 -1.85428145 -0.07791713  0.15179485
#> 8      S06     I01  3.3077444 3.018906  1.08456280  0.08138621 -0.87711022
#> 20     S06     I02  4.7364315 3.018906  1.08456280  0.60558819  0.02737486
#> 27     S06     I03  3.3861675 3.018906  1.08456280 -0.94560820  0.22830727
#> 37     S06     I04  2.7381016 3.018906  1.08456280 -0.10053204 -1.26483474
#> 49     S06     I05  6.4029147 3.018906  1.08456280  0.85725988  1.44218638
#> 56     S06     I06  3.4944629 3.018906  1.08456280 -0.89406578  0.28506027
#> 66     S06     I07  4.9555189 3.018906  1.08456280  0.55564089  0.29640961
#> 79     S06     I08  4.8829175 3.018906  1.08456280  1.04326387 -0.26381473
#> 85     S06     I09  3.8950730 3.018906  1.08456280 -0.02055614 -0.18783924
#> 99     S06     I10  4.3134767 3.018906  1.08456280 -0.07791713  0.28792545
#> 9      S07     I01  3.1461509 3.018906  1.05382827  0.08138621 -1.00796921
#> 11     S07     I02  4.2392443 3.018906  1.05382827  0.60558819 -0.43907777
#> 24     S07     I03  3.5294281 3.018906  1.05382827 -0.94560820  0.40230242
#> 38     S07     I04  4.8054483 3.018906  1.05382827 -0.10053204  0.83324650
#> 50     S07     I05  4.8409439 3.018906  1.05382827  0.85725988 -0.08904986
#> 53     S07     I06  3.9910787 3.018906  1.05382827 -0.89406578  0.81241055
#> 67     S07     I07  5.9918638 3.018906  1.05382827  0.55564089  1.36348899
#> 80     S07     I08  5.3183275 3.018906  1.05382827  1.04326387  0.20232971
#> 82     S07     I09  3.3975185 3.018906  1.05382827 -0.02055614 -0.65465925
#> 96     S07     I10  3.4153331 3.018906  1.05382827 -0.07791713 -0.57948362
#> 6      S08     I01  4.0449056 3.018906  0.40617835  0.08138621  0.53843543
#> 12     S08     I02  3.2784501 3.018906  0.40617835  0.60558819 -0.75222209
#> 25     S08     I03  3.4008367 3.018906  0.40617835 -0.94560820  0.92136093
#> 39     S08     I04  4.3439805 3.018906  0.40617835 -0.10053204  1.01942856
#> 41     S08     I05  5.2727405 3.018906  0.40617835  0.85725988  0.99039667
#> 54     S08     I06  3.4175055 3.018906  0.40617835 -0.89406578  0.88648728
#> 68     S08     I07  2.8882369 3.018906  0.40617835  0.55564089 -1.09248799
#> 78     S08     I08  3.8699794 3.018906  0.40617835  1.04326387 -0.59836845
#> 83     S08     I09  4.0985897 3.018906  0.40617835 -0.02055614  0.69406190
#> 97     S08     I10  2.6611451 3.018906  0.40617835 -0.07791713 -0.68602176
#> 3      S09     I01  4.9525361 3.018906  1.51706950  0.08138621  0.33517481
#> 13     S09     I02  5.3832387 3.018906  1.51706950  0.60558819  0.24167544
#> 26     S09     I03  2.0654256 3.018906  1.51706950 -0.94560820 -1.52494133
#> 32     S09     I04  3.7798412 3.018906  1.51706950 -0.10053204 -0.65560187
#> 42     S09     I05  5.9999839 3.018906  1.51706950  0.85725988  0.60674886
#> 55     S09     I06  2.2191263 3.018906  1.51706950 -0.89406578 -1.42278303
#> 65     S09     I07  4.3330141 3.018906  1.51706950  0.55564089 -0.75860192
#> 71     S09     I08  4.9287667 3.018906  1.51706950  1.04326387 -0.65047225
#> 84     S09     I09  4.3813831 3.018906  1.51706950 -0.02055614 -0.13403589
#> 94     S09     I10  4.8996444 3.018906  1.51706950 -0.07791713  0.44158639
#> 10     S10     I01  3.3000988 3.018906 -0.71568371  0.08138621  0.91549064
#> 14     S10     I02  4.7651133 3.018906 -0.71568371  0.60558819  1.85630316
#> 28     S10     I03  1.7167237 3.018906 -0.71568371 -0.94560820  0.35911000
#> 40     S10     I04  3.1018159 3.018906 -0.71568371 -0.10053204  0.89912603
#> 43     S10     I05  3.3829766 3.018906 -0.71568371  0.85725988  0.22249483
#> 57     S10     I06  0.8819678 3.018906 -0.71568371 -0.89406578 -0.52718838
#> 70     S10     I07  2.5323462 3.018906 -0.71568371  0.55564089 -0.32651664
#> 72     S10     I08  4.9847200 3.018906 -0.71568371  1.04326387  1.63823421
#> 86     S10     I09  2.7658243 3.018906 -0.71568371 -0.02055614  0.48315856
#> 100    S10     I10  1.1152143 3.018906 -0.71568371 -0.07791713 -1.11009048