Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id           y  grand_i       sub_i     item_i          err
#> 1      S01     I01 -0.10527831 3.018906 -1.17489064 -0.9549524 -0.994340919
#> 15     S01     I02  3.70694314 3.018906 -1.17489064  0.5839349  1.278993302
#> 29     S01     I03  1.36646937 3.018906 -1.17489064  0.4504874 -0.928033007
#> 31     S01     I04  0.88732135 3.018906 -1.17489064 -0.6370991 -0.319594531
#> 44     S01     I05  1.33257492 3.018906 -1.17489064 -0.2536971 -0.257742925
#> 58     S01     I06  3.19187512 3.018906 -1.17489064  0.5798649  0.767995301
#> 61     S01     I07  2.81133680 3.018906 -1.17489064 -0.4174841  1.384805909
#> 73     S01     I08  1.64966694 3.018906 -1.17489064  1.1428470 -1.337195050
#> 87     S01     I09  2.01063047 3.018906 -1.17489064  0.3816061 -0.214990576
#> 91     S01     I10  2.75652162 3.018906 -1.17489064  0.7740876  0.138419074
#> 7      S02     I01  2.11218444 3.018906 -0.01847182 -0.9549524  0.066703019
#> 17     S02     I02  2.74984848 3.018906 -0.01847182  0.5839349 -0.834520179
#> 30     S02     I03  2.90367498 3.018906 -0.01847182  0.4504874 -0.547246213
#> 36     S02     I04  3.35880215 3.018906 -0.01847182 -0.6370991  0.995467452
#> 46     S02     I05  1.67178950 3.018906 -0.01847182 -0.2536971 -1.074947155
#> 59     S02     I06  2.62424554 3.018906 -0.01847182  0.5798649 -0.956053104
#> 69     S02     I07  3.58124613 3.018906 -0.01847182 -0.4174841  0.998296419
#> 75     S02     I08  3.50884253 3.018906 -0.01847182  1.1428470 -0.634438274
#> 88     S02     I09  2.47725598 3.018906 -0.01847182  0.3816061 -0.904783877
#> 98     S02     I10  4.91480937 3.018906 -0.01847182  0.7740876  1.140288003
#> 4      S03     I01  2.79805761 3.018906 -0.52213016 -0.9549524  1.256234515
#> 18     S03     I02  3.04467078 3.018906 -0.52213016  0.5839349 -0.036039546
#> 23     S03     I03  3.03797109 3.018906 -0.52213016  0.4504874  0.090708236
#> 33     S03     I04 -0.61523472 3.018906 -0.52213016 -0.6370991 -2.474911087
#> 47     S03     I05  3.59910091 3.018906 -0.52213016 -0.2536971  1.356022586
#> 60     S03     I06  3.08332266 3.018906 -0.52213016  0.5798649  0.006682351
#> 62     S03     I07  2.01423286 3.018906 -0.52213016 -0.4174841 -0.065058517
#> 76     S03     I08  5.27342817 3.018906 -0.52213016  1.1428470  1.633805697
#> 89     S03     I09  3.78133233 3.018906 -0.52213016  0.3816061  0.902950806
#> 95     S03     I10  2.69704644 3.018906 -0.52213016  0.7740876 -0.573816592
#> 5      S04     I01  4.02943934 3.018906  0.53742865 -0.9549524  1.428057448
#> 19     S04     I02  3.94533928 3.018906  0.53742865  0.5839349 -0.194929844
#> 21     S04     I03  4.10300356 3.018906  0.53742865  0.4504874  0.096181904
#> 34     S04     I04  1.94770250 3.018906  0.53742865 -0.6370991 -0.971532668
#> 48     S04     I05  3.75577713 3.018906  0.53742865 -0.2536971  0.453140005
#> 51     S04     I06  4.36606356 3.018906  0.53742865  0.5798649  0.229864453
#> 63     S04     I07  3.15932983 3.018906  0.53742865 -0.4174841  0.020479653
#> 77     S04     I08  5.66806760 3.018906  0.53742865  1.1428470  0.968886326
#> 90     S04     I09  2.51616175 3.018906  0.53742865  0.3816061 -1.421778581
#> 92     S04     I10  4.74110964 3.018906  0.53742865  0.7740876  0.410687803
#> 2      S05     I01  3.47234519 3.018906  1.28744342 -0.9549524  0.120948519
#> 16     S05     I02  5.09724778 3.018906  1.28744342  0.5839349  0.206963879
#> 22     S05     I03  5.86024226 3.018906  1.28744342  0.4504874  1.103405820
#> 35     S05     I04  4.11506030 3.018906  1.28744342 -0.6370991  0.445810355
#> 45     S05     I05  4.58723837 3.018906  1.28744342 -0.2536971  0.534586464
#> 52     S05     I06  4.46780502 3.018906  1.28744342  0.5798649 -0.418408867
#> 64     S05     I07  6.03507335 3.018906  1.28744342 -0.4174841  2.146208394
#> 74     S05     I08  5.25056758 3.018906  1.28744342  1.1428470 -0.198628469
#> 81     S05     I09  2.07011892 3.018906  1.28744342  0.3816061 -2.617836184
#> 93     S05     I10  4.24655215 3.018906  1.28744342  0.7740876 -0.833884462
#> 8      S06     I01 -0.57706709 3.018906 -1.50676662 -0.9549524 -1.134253716
#> 20     S06     I02  1.72467786 3.018906 -1.50676662  0.5839349 -0.371395996
#> 27     S06     I03  2.25785395 3.018906 -1.50676662  0.4504874  0.295227562
#> 37     S06     I04  1.64488967 3.018906 -1.50676662 -0.6370991  0.769849773
#> 49     S06     I05  1.98360845 3.018906 -1.50676662 -0.2536971  0.725166594
#> 56     S06     I06  1.35631959 3.018906 -1.50676662  0.5798649 -0.735684247
#> 66     S06     I07  1.37090315 3.018906 -1.50676662 -0.4174841  0.276248234
#> 79     S06     I08  2.84730989 3.018906 -1.50676662  1.1428470  0.192323884
#> 85     S06     I09  3.86427792 3.018906 -1.50676662  0.3816061  1.970532861
#> 99     S06     I10  1.35398928 3.018906 -1.50676662  0.7740876 -0.932237285
#> 9      S07     I01  2.55977758 3.018906  0.84854506 -0.9549524 -0.352720721
#> 11     S07     I02  4.22862901 3.018906  0.84854506  0.5839349 -0.222756520
#> 24     S07     I03  2.15950329 3.018906  0.84854506  0.4504874 -2.158434782
#> 38     S07     I04  4.03128241 3.018906  0.84854506 -0.6370991  0.800930829
#> 50     S07     I05  4.89889881 3.018906  0.84854506 -0.2536971  1.285145271
#> 53     S07     I06  5.05356985 3.018906  0.84854506  0.5798649  0.606254332
#> 67     S07     I07  1.84015262 3.018906  0.84854506 -0.4174841 -1.609813969
#> 80     S07     I08  5.88406635 3.018906  0.84854506  1.1428470  0.873768672
#> 82     S07     I09  4.06493198 3.018906  0.84854506  0.3816061 -0.184124760
#> 96     S07     I10  4.96956452 3.018906  0.84854506  0.7740876  0.328026275
#> 6      S08     I01  1.28312029 3.018906 -0.28496773 -0.9549524 -0.495865227
#> 12     S08     I02  1.58249328 3.018906 -0.28496773  0.5839349 -1.735379465
#> 25     S08     I03  0.55753125 3.018906 -0.28496773  0.4504874 -2.626894030
#> 39     S08     I04 -0.02433228 3.018906 -0.28496773 -0.6370991 -2.121171072
#> 41     S08     I05  2.59815781 3.018906 -0.28496773 -0.2536971  0.117917058
#> 54     S08     I06  4.01761615 3.018906 -0.28496773  0.5798649  0.703813424
#> 68     S08     I07  0.49008208 3.018906 -0.28496773 -0.4174841 -1.826371718
#> 78     S08     I08  4.45825660 3.018906 -0.28496773  1.1428470  0.581471706
#> 83     S08     I09  2.44846576 3.018906 -0.28496773  0.3816061 -0.667078195
#> 97     S08     I10  3.67779105 3.018906 -0.28496773  0.7740876  0.169765592
#> 3      S09     I01  2.83119386 3.018906  0.63210735 -0.9549524  0.135133257
#> 13     S09     I02  4.29510247 3.018906  0.63210735  0.5839349  0.060154640
#> 26     S09     I03  5.43251016 3.018906  0.63210735  0.4504874  1.331009799
#> 32     S09     I04  2.90462573 3.018906  0.63210735 -0.6370991 -0.109288141
#> 42     S09     I05  5.08166682 3.018906  0.63210735 -0.2536971  1.684350986
#> 55     S09     I06  5.13899008 3.018906  0.63210735  0.5798649  0.908112270
#> 65     S09     I07  2.21928490 3.018906  0.63210735 -0.4174841 -1.014243985
#> 71     S09     I08  5.06520067 3.018906  0.63210735  1.1428470  0.271340694
#> 84     S09     I09  3.51738118 3.018906  0.63210735  0.3816061 -0.515237854
#> 94     S09     I10  3.61043100 3.018906  0.63210735  0.7740876 -0.814669534
#> 10     S10     I01  1.60429501 3.018906 -1.07390328 -0.9549524  0.614245042
#> 14     S10     I02  3.52907540 3.018906 -1.07390328  0.5839349  1.000138206
#> 28     S10     I03  2.38615255 3.018906 -1.07390328  0.4504874 -0.009337176
#> 40     S10     I04  2.96024296 3.018906 -1.07390328 -0.6370991  1.652339720
#> 43     S10     I05  0.92473667 3.018906 -1.07390328 -0.2536971 -0.766568532
#> 57     S10     I06  2.55032247 3.018906 -1.07390328  0.5798649  0.025455291
#> 70     S10     I07  2.78053848 3.018906 -1.07390328 -0.4174841  1.253020232
#> 72     S10     I08  2.41083667 3.018906 -1.07390328  1.1428470 -0.677012674
#> 86     S10     I09  2.28209269 3.018906 -1.07390328  0.3816061 -0.044515716
#> 100    S10     I10  2.37516782 3.018906 -1.07390328  0.7740876 -0.343922085