Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id",
  seed = NULL
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#> # A tibble: 100 x 7
#>    sub_id item_id     y grand_i  sub_i  item_i    err
#>    <chr>  <chr>   <dbl>   <dbl>  <dbl>   <dbl>  <dbl>
#>  1 S01    I01     2.85     3.02 -0.242  0.878  -0.805
#>  2 S01    I02     2.27     3.02 -0.242 -0.0954 -0.409
#>  3 S01    I03     3.00     3.02 -0.242  1.02   -0.797
#>  4 S01    I04     0.736    3.02 -0.242 -0.767  -1.27 
#>  5 S01    I05     2.37     3.02 -0.242 -0.225  -0.178
#>  6 S01    I06     2.30     3.02 -0.242  0.0701 -0.550
#>  7 S01    I07     5.20     3.02 -0.242  0.832   1.59 
#>  8 S01    I08     3.63     3.02 -0.242 -1.28    2.13 
#>  9 S01    I09     2.25     3.02 -0.242  0.630  -1.16 
#> 10 S01    I10     1.47     3.02 -0.242  0.191  -1.50 
#> # … with 90 more rows