Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id         y  grand_i      sub_i      item_i          err
#> 1      S01     I01 3.9617779 3.018906  1.0691659  0.18672851 -0.313022117
#> 2      S02     I01 4.2117022 3.018906 -0.8073357  0.18672851  1.813403784
#> 3      S03     I01 7.2278486 3.018906  0.9505762  0.18672851  3.071638230
#> 4      S04     I01 4.4284256 3.018906 -0.5114046  0.18672851  1.734196100
#> 5      S05     I01 3.0211767 3.018906 -0.3496485  0.18672851  0.165191099
#> 6      S06     I01 7.0296913 3.018906  1.2081781  0.18672851  2.615879093
#> 7      S07     I01 4.5564778 3.018906  0.1503317  0.18672851  1.200511984
#> 8      S08     I01 1.8544600 3.018906 -0.7782791  0.18672851 -0.572895043
#> 9      S09     I01 4.6824839 3.018906  0.9660476  0.18672851  0.510802083
#> 10     S10     I01 2.8910486 3.018906 -0.7009329  0.18672851  0.386347382
#> 11     S01     I02 3.1169191 3.018906  1.0691659 -0.79960952 -0.171542864
#> 12     S02     I02 2.1273418 3.018906 -0.8073357 -0.79960952  0.715381448
#> 13     S03     I02 3.9632819 3.018906  0.9505762 -0.79960952  0.793409555
#> 14     S04     I02 0.6582600 3.018906 -0.5114046 -0.79960952 -1.049631440
#> 15     S05     I02 0.5809438 3.018906 -0.3496485 -0.79960952 -1.288703732
#> 16     S06     I02 5.1821513 3.018906  1.2081781 -0.79960952  1.754677142
#> 17     S07     I02 3.4538382 3.018906  0.1503317 -0.79960952  1.084210436
#> 18     S08     I02 2.4996841 3.018906 -0.7782791 -0.79960952  1.058667161
#> 19     S09     I02 3.4269626 3.018906  0.9660476 -0.79960952  0.241618909
#> 20     S10     I02 0.5417475 3.018906 -0.7009329 -0.79960952 -0.976615704
#> 21     S01     I03 5.6444714 3.018906  1.0691659  1.21278097  0.343619008
#> 22     S02     I03 4.0533695 3.018906 -0.8073357  1.21278097  0.629018624
#> 23     S03     I03 4.7895441 3.018906  0.9505762  1.21278097 -0.392718685
#> 24     S04     I03 4.2338357 3.018906 -0.5114046  1.21278097  0.513553732
#> 25     S05     I03 3.5720688 3.018906 -0.3496485  1.21278097 -0.309969230
#> 26     S06     I03 5.1704766 3.018906  1.2081781  1.21278097 -0.269388092
#> 27     S07     I03 4.4804086 3.018906  0.1503317  1.21278097  0.098390369
#> 28     S08     I03 1.9020148 3.018906 -0.7782791  1.21278097 -1.551392679
#> 29     S09     I03 5.3382018 3.018906  0.9660476  1.21278097  0.140467544
#> 30     S10     I03 3.0085075 3.018906 -0.7009329  1.21278097 -0.522246241
#> 31     S01     I04 3.7582121 3.018906  1.0691659  0.01160521 -0.341464535
#> 32     S02     I04 0.4279178 3.018906 -0.8073357  0.01160521 -1.795257342
#> 33     S03     I04 4.8908687 3.018906  0.9505762  0.01160521  0.909781670
#> 34     S04     I04 3.2618329 3.018906 -0.5114046  0.01160521  0.742726753
#> 35     S05     I04 3.6571109 3.018906 -0.3496485  0.01160521  0.976248623
#> 36     S06     I04 3.8820373 3.018906  1.2081781  0.01160521 -0.356651552
#> 37     S07     I04 2.7388271 3.018906  0.1503317  0.01160521 -0.442015393
#> 38     S08     I04 0.2048896 3.018906 -0.7782791  0.01160521 -2.047342086
#> 39     S09     I04 4.0410701 3.018906  0.9660476  0.01160521  0.044511676
#> 40     S10     I04 1.9923658 3.018906 -0.7009329  0.01160521 -0.337212125
#> 41     S01     I05 5.7582763 3.018906  1.0691659  0.46567601  1.204528788
#> 42     S02     I05 1.2861569 3.018906 -0.8073357  0.46567601 -1.391088989
#> 43     S03     I05 5.3646417 3.018906  0.9505762  0.46567601  0.929483793
#> 44     S04     I05 3.3717344 3.018906 -0.5114046  0.46567601  0.398557383
#> 45     S05     I05 3.5073176 3.018906 -0.3496485  0.46567601  0.372384456
#> 46     S06     I05 5.9578221 3.018906  1.2081781  0.46567601  1.265062407
#> 47     S07     I05 4.1169263 3.018906  0.1503317  0.46567601  0.482012942
#> 48     S08     I05 1.3570715 3.018906 -0.7782791  0.46567601 -1.349231049
#> 49     S09     I05 4.0654179 3.018906  0.9660476  0.46567601 -0.385211368
#> 50     S10     I05 2.7218771 3.018906 -0.7009329  0.46567601 -0.061771658
#> 51     S01     I06 4.1043035 3.018906  1.0691659  0.45643207 -0.440200074
#> 52     S02     I06 2.7753206 3.018906 -0.8073357  0.45643207  0.107318587
#> 53     S03     I06 4.8633477 3.018906  0.9505762  0.45643207  0.437433830
#> 54     S04     I06 2.8196739 3.018906 -0.5114046  0.45643207 -0.144259144
#> 55     S05     I06 1.8217365 3.018906 -0.3496485  0.45643207 -1.303952624
#> 56     S06     I06 6.1440061 3.018906  1.2081781  0.45643207  1.460490316
#> 57     S07     I06 1.3360029 3.018906  0.1503317  0.45643207 -2.289666499
#> 58     S08     I06 3.1692035 3.018906 -0.7782791  0.45643207  0.472144936
#> 59     S09     I06 5.0579239 3.018906  0.9660476  0.45643207  0.616538578
#> 60     S10     I06 0.7890577 3.018906 -0.7009329  0.45643207 -1.985347149
#> 61     S01     I07 2.8410112 3.018906  1.0691659 -0.95416846 -0.292891840
#> 62     S02     I07 0.1205241 3.018906 -0.8073357 -0.95416846 -1.136877306
#> 63     S03     I07 2.8571338 3.018906  0.9505762 -0.95416846 -0.158179606
#> 64     S04     I07 1.2254879 3.018906 -0.5114046 -0.95416846 -0.327844608
#> 65     S05     I07 4.1704771 3.018906 -0.3496485 -0.95416846  2.455388431
#> 66     S06     I07 3.4248307 3.018906  1.2081781 -0.95416846  0.151915493
#> 67     S07     I07 3.5863946 3.018906  0.1503317 -0.95416846  1.371325757
#> 68     S08     I07 2.8786513 3.018906 -0.7782791 -0.95416846  1.592193268
#> 69     S09     I07 1.5679376 3.018906  0.9660476 -0.95416846 -1.462847230
#> 70     S10     I07 1.1683348 3.018906 -0.7009329 -0.95416846 -0.195469464
#> 71     S01     I08 6.4857209 3.018906  1.0691659  1.18732830  1.210321165
#> 72     S02     I08 3.0749077 3.018906 -0.8073357  1.18732830 -0.323990461
#> 73     S03     I08 5.7183510 3.018906  0.9505762  1.18732830  0.561540844
#> 74     S04     I08 5.1322774 3.018906 -0.5114046  1.18732830  1.437448078
#> 75     S05     I08 3.8817765 3.018906 -0.3496485  1.18732830  0.025191110
#> 76     S06     I08 4.8680590 3.018906  1.2081781  1.18732830 -0.546352960
#> 77     S07     I08 3.7300336 3.018906  0.1503317  1.18732830 -0.626532028
#> 78     S08     I08 3.1713619 3.018906 -0.7782791  1.18732830 -0.256592903
#> 79     S09     I08 6.1623785 3.018906  0.9660476  1.18732830  0.990096917
#> 80     S10     I08 2.8299266 3.018906 -0.7009329  1.18732830 -0.675374408
#> 81     S01     I09 3.3444813 3.018906  1.0691659 -0.20821949 -0.535370714
#> 82     S02     I09 1.5092707 3.018906 -0.8073357 -0.20821949 -0.494079716
#> 83     S03     I09 2.7339333 3.018906  0.9505762 -0.20821949 -1.027329079
#> 84     S04     I09 1.4230769 3.018906 -0.5114046 -0.20821949 -0.876204608
#> 85     S05     I09 3.5780502 3.018906 -0.3496485 -0.20821949  1.117012561
#> 86     S06     I09 2.9278240 3.018906  1.2081781 -0.20821949 -1.091040166
#> 87     S07     I09 1.6416381 3.018906  0.1503317 -0.20821949 -1.319379750
#> 88     S08     I09 1.8256129 3.018906 -0.7782791 -0.20821949 -0.206794071
#> 89     S09     I09 3.0383730 3.018906  0.9660476 -0.20821949 -0.738360761
#> 90     S10     I09 0.6306693 3.018906 -0.7009329 -0.20821949 -1.479083933
#> 91     S01     I10 4.0780425 3.018906  1.0691659  0.64987807 -0.659907028
#> 92     S02     I10 0.7889467 3.018906 -0.8073357  0.64987807 -2.072501243
#> 93     S03     I10 3.7429505 3.018906  0.9505762  0.64987807 -0.876409416
#> 94     S04     I10 4.5652617 3.018906 -0.5114046  0.64987807  1.407882644
#> 95     S05     I10 2.9694788 3.018906 -0.3496485  0.64987807 -0.349656375
#> 96     S06     I10 5.4627010 3.018906  1.2081781  0.64987807  0.585739236
#> 97     S07     I10 3.8156273 3.018906  0.1503317  0.64987807 -0.003488031
#> 98     S08     I10 2.6187219 3.018906 -0.7782791  0.64987807 -0.271782632
#> 99     S09     I10 5.4893499 3.018906  0.9660476  0.64987807  0.854518549
#> 100    S10     I10 0.6614829 3.018906 -0.7009329  0.64987807 -2.306367886