Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(.data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

.data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id          y  grand_i       sub_i      item_i         err
#> 1      S01     I01  2.2340426 3.018906 -0.06169490 -0.20764517 -0.51552290
#> 2      S02     I01  1.7900382 3.018906 -1.52599932 -0.20764517  0.50477710
#> 3      S03     I01  4.1545903 3.018906  0.20737160 -0.20764517  1.13595822
#> 4      S04     I01  2.4590501 3.018906  0.45370435 -0.20764517 -0.80591472
#> 5      S05     I01  3.7585866 3.018906 -1.16142361 -0.20764517  2.10874980
#> 6      S06     I01  2.7271015 3.018906  0.01357193 -0.20764517 -0.09773092
#> 7      S07     I01  2.8879969 3.018906 -0.12216573 -0.20764517  0.19890215
#> 8      S08     I01  3.0787750 3.018906  1.95758393 -0.20764517 -1.69006941
#> 9      S09     I01  4.7575764 3.018906  0.87327392 -0.20764517  1.07304206
#> 10     S10     I01  2.4961403 3.018906  0.53171488 -0.20764517 -0.84683504
#> 11     S01     I02  5.3228693 3.018906 -0.06169490  1.16081981  1.20483876
#> 12     S02     I02  2.5475579 3.018906 -1.52599932  1.16081981 -0.10616817
#> 13     S03     I02  5.6692365 3.018906  0.20737160  1.16081981  1.28213944
#> 14     S04     I02  5.7317362 3.018906  0.45370435  1.16081981  1.09830638
#> 15     S05     I02  1.7958188 3.018906 -1.16142361  1.16081981 -1.22248301
#> 16     S06     I02  5.5674832 3.018906  0.01357193  1.16081981  1.37418587
#> 17     S07     I02  3.7564919 3.018906 -0.12216573  1.16081981 -0.30106779
#> 18     S08     I02  5.0520417 3.018906  1.95758393  1.16081981 -1.08526761
#> 19     S09     I02  6.2774774 3.018906  0.87327392  1.16081981  1.22447801
#> 20     S10     I02  6.2817419 3.018906  0.53171488  1.16081981  1.57030156
#> 21     S01     I03  2.3427792 3.018906 -0.06169490  0.88946627 -1.50389783
#> 22     S02     I03  3.3850499 3.018906 -1.52599932  0.88946627  1.00267736
#> 23     S03     I03  6.0734973 3.018906  0.20737160  0.88946627  1.95775382
#> 24     S04     I03  5.9060348 3.018906  0.45370435  0.88946627  1.54395859
#> 25     S05     I03  2.9695023 3.018906 -1.16142361  0.88946627  0.22255407
#> 26     S06     I03  2.9550669 3.018906  0.01357193  0.88946627 -0.96687690
#> 27     S07     I03  5.1475332 3.018906 -0.12216573  0.88946627  1.36132701
#> 28     S08     I03  4.2232089 3.018906  1.95758393  0.88946627 -1.64274692
#> 29     S09     I03  4.1906264 3.018906  0.87327392  0.88946627 -0.59101939
#> 30     S10     I03  5.3981861 3.018906  0.53171488  0.88946627  0.95809936
#> 31     S01     I04  0.5933262 3.018906 -0.06169490 -1.97122214 -0.39266234
#> 32     S02     I04  0.1300622 3.018906 -1.52599932 -1.97122214  0.60837802
#> 33     S03     I04  1.0933996 3.018906  0.20737160 -1.97122214 -0.16165550
#> 34     S04     I04  0.1862121 3.018906  0.45370435 -1.97122214 -1.31517573
#> 35     S05     I04  0.8427791 3.018906 -1.16142361 -1.97122214  0.95651919
#> 36     S06     I04  2.4482834 3.018906  0.01357193 -1.97122214  1.38702799
#> 37     S07     I04  0.6827606 3.018906 -0.12216573 -1.97122214 -0.24275712
#> 38     S08     I04  4.6847649 3.018906  1.95758393 -1.97122214  1.67949751
#> 39     S09     I04  1.7320520 3.018906  0.87327392 -1.97122214 -0.18890542
#> 40     S10     I04  3.0407544 3.018906  0.53171488 -1.97122214  1.46135607
#> 41     S01     I05  4.7030742 3.018906 -0.06169490 -0.12158259  1.86744611
#> 42     S02     I05  0.8629521 3.018906 -1.52599932 -0.12158259 -0.50837159
#> 43     S03     I05  6.4013013 3.018906  0.20737160 -0.12158259  3.29660669
#> 44     S04     I05  1.8898069 3.018906  0.45370435 -0.12158259 -1.46122050
#> 45     S05     I05  1.8313415 3.018906 -1.16142361 -0.12158259  0.09544205
#> 46     S06     I05  5.5515057 3.018906  0.01357193 -0.12158259  2.64061073
#> 47     S07     I05  2.6989957 3.018906 -0.12216573 -0.12158259 -0.07616159
#> 48     S08     I05  4.1551140 3.018906  1.95758393 -0.12158259 -0.69979296
#> 49     S09     I05  5.1606492 3.018906  0.87327392 -0.12158259  1.39005221
#> 50     S10     I05  1.9713647 3.018906  0.53171488 -0.12158259 -1.45767317
#> 51     S01     I06  3.3595187 3.018906 -0.06169490  0.01981099  0.38249696
#> 52     S02     I06  1.6761077 3.018906 -1.52599932  0.01981099  0.16339046
#> 53     S03     I06  4.7533269 3.018906  0.20737160  0.01981099  1.50723866
#> 54     S04     I06  4.7193318 3.018906  0.45370435  0.01981099  1.22691086
#> 55     S05     I06  2.3541800 3.018906 -1.16142361  0.01981099  0.47688705
#> 56     S06     I06  4.0921993 3.018906  0.01357193  0.01981099  1.03991076
#> 57     S07     I06  1.4391585 3.018906 -0.12216573  0.01981099 -1.47739242
#> 58     S08     I06  6.5933271 3.018906  1.95758393  0.01981099  1.59702653
#> 59     S09     I06  3.3783697 3.018906  0.87327392  0.01981099 -0.53362083
#> 60     S10     I06  4.3919369 3.018906  0.53171488  0.01981099  0.82150538
#> 61     S01     I07  3.1507904 3.018906 -0.06169490 -0.17198854  0.36556827
#> 62     S02     I07  0.3190953 3.018906 -1.52599932 -0.17198854 -1.00182244
#> 63     S03     I07  3.2745188 3.018906  0.20737160 -0.17198854  0.22023014
#> 64     S04     I07  1.5620807 3.018906  0.45370435 -0.17198854 -1.73854076
#> 65     S05     I07  3.0764385 3.018906 -1.16142361 -0.17198854  1.39094505
#> 66     S06     I07  2.1535840 3.018906  0.01357193 -0.17198854 -0.70690498
#> 67     S07     I07  2.6368287 3.018906 -0.12216573 -0.17198854 -0.08792260
#> 68     S08     I07  5.4220805 3.018906  1.95758393 -0.17198854  0.61757947
#> 69     S09     I07  3.3157635 3.018906  0.87327392 -0.17198854 -0.40442745
#> 70     S10     I07  3.8256002 3.018906  0.53171488 -0.17198854  0.44696830
#> 71     S01     I08  4.1054677 3.018906 -0.06169490 -0.32884988  1.47710683
#> 72     S02     I08  2.3600037 3.018906 -1.52599932 -0.32884988  1.19594732
#> 73     S03     I08  2.7620722 3.018906  0.20737160 -0.32884988 -0.13535514
#> 74     S04     I08  1.1265578 3.018906  0.45370435 -0.32884988 -2.01720231
#> 75     S05     I08  1.0721878 3.018906 -1.16142361 -0.32884988 -0.45644430
#> 76     S06     I08  4.0813522 3.018906  0.01357193 -0.32884988  1.37772451
#> 77     S07     I08 -0.6955771 3.018906 -0.12216573 -0.32884988 -3.26346712
#> 78     S08     I08  4.0010660 3.018906  1.95758393 -0.32884988 -0.64657371
#> 79     S09     I08  4.8764820 3.018906  0.87327392 -0.32884988  1.31315231
#> 80     S10     I08  3.2497822 3.018906  0.53171488 -0.32884988  0.02801162
#> 81     S01     I09  1.6100043 3.018906 -0.06169490  0.01291578 -1.36012215
#> 82     S02     I09  1.9101661 3.018906 -1.52599932  0.01291578  0.40434399
#> 83     S03     I09  2.2513573 3.018906  0.20737160  0.01291578 -0.98783572
#> 84     S04     I09  2.9546311 3.018906  0.45370435  0.01291578 -0.53089469
#> 85     S05     I09  2.9531147 3.018906 -1.16142361  0.01291578  1.08271687
#> 86     S06     I09  2.5835337 3.018906  0.01357193  0.01291578 -0.46185958
#> 87     S07     I09  4.8518804 3.018906 -0.12216573  0.01291578  1.94222472
#> 88     S08     I09  3.9327980 3.018906  1.95758393  0.01291578 -1.05660734
#> 89     S09     I09  3.0632995 3.018906  0.87327392  0.01291578 -0.84179583
#> 90     S10     I09  4.8249479 3.018906  0.53171488  0.01291578  1.26141165
#> 91     S01     I10  1.8598863 3.018906 -0.06169490 -0.96097942 -0.13634495
#> 92     S02     I10  0.4144696 3.018906 -1.52599932 -0.96097942 -0.11745731
#> 93     S03     I10  2.5773876 3.018906  0.20737160 -0.96097942  0.31208985
#> 94     S04     I10  3.6237131 3.018906  0.45370435 -0.96097942  1.11208252
#> 95     S05     I10  1.1022706 3.018906 -1.16142361 -0.96097942  0.20576797
#> 96     S06     I10  2.2487871 3.018906  0.01357193 -0.96097942  0.17728901
#> 97     S07     I10  2.7517175 3.018906 -0.12216573 -0.96097942  0.81595703
#> 98     S08     I10  4.8047013 3.018906  1.95758393 -0.96097942  0.78919120
#> 99     S09     I10  2.3343889 3.018906  0.87327392 -0.96097942 -0.59681126
#> 100    S10     I10  3.8306755 3.018906  0.53171488 -0.96097942  1.24103438