Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id         y  grand_i       sub_i      item_i         err
#> 1      S01     I01 6.3646902 3.018906  1.48120549  1.19133200  0.67324705
#> 15     S01     I02 4.8794404 3.018906  1.48120549  0.14091305  0.23841620
#> 29     S01     I03 2.8839789 3.018906  1.48120549  0.06485664 -1.68098887
#> 31     S01     I04 2.1128697 3.018906  1.48120549 -0.42635348 -1.96088794
#> 44     S01     I05 6.7938000 3.018906  1.48120549  1.12736729  1.16632159
#> 58     S01     I06 5.6298822 3.018906  1.48120549  1.10087374  0.02889740
#> 61     S01     I07 5.0537244 3.018906  1.48120549 -0.29432480  0.84793813
#> 73     S01     I08 4.9907879 3.018906  1.48120549  0.66207708 -0.17140028
#> 87     S01     I09 5.1491113 3.018906  1.48120549 -0.06026755  0.70926774
#> 91     S01     I10 5.0168336 3.018906  1.48120549  0.35539173  0.16133080
#> 7      S02     I01 2.8445029 3.018906  0.14436494  1.19133200 -1.51009968
#> 17     S02     I02 2.8657735 3.018906  0.14436494  0.14091305 -0.43841007
#> 30     S02     I03 1.1268162 3.018906  0.14436494  0.06485664 -2.10131101
#> 36     S02     I04 3.6991602 3.018906  0.14436494 -0.42635348  0.96224313
#> 46     S02     I05 4.7060601 3.018906  0.14436494  1.12736729  0.41542222
#> 59     S02     I06 4.7605146 3.018906  0.14436494  1.10087374  0.49637031
#> 69     S02     I07 3.1690952 3.018906  0.14436494 -0.29432480  0.30014947
#> 75     S02     I08 5.0680639 3.018906  0.14436494  0.66207708  1.24271622
#> 88     S02     I09 3.2088190 3.018906  0.14436494 -0.06026755  0.10581605
#> 98     S02     I10 3.8651679 3.018906  0.14436494  0.35539173  0.34650563
#> 4      S03     I01 2.9498962 3.018906  0.10403004  1.19133200 -1.36437142
#> 18     S03     I02 3.6717427 3.018906  0.10403004  0.14091305  0.40789402
#> 23     S03     I03 5.2679292 3.018906  0.10403004  0.06485664  2.08013692
#> 33     S03     I04 2.4907094 3.018906  0.10403004 -0.42635348 -0.20587274
#> 47     S03     I05 4.0098255 3.018906  0.10403004  1.12736729 -0.24047743
#> 60     S03     I06 5.7845975 3.018906  0.10403004  1.10087374  1.56078814
#> 62     S03     I07 4.2883106 3.018906  0.10403004 -0.29432480  1.45969972
#> 76     S03     I08 2.3639133 3.018906  0.10403004  0.66207708 -1.42109948
#> 89     S03     I09 2.4388567 3.018906  0.10403004 -0.06026755 -0.62381139
#> 95     S03     I10 3.5708095 3.018906  0.10403004  0.35539173  0.09248213
#> 5      S04     I01 3.3127892 3.018906 -0.72126824  1.19133200 -0.17618017
#> 19     S04     I02 2.1062678 3.018906 -0.72126824  0.14091305 -0.33228260
#> 21     S04     I03 2.6418236 3.018906 -0.72126824  0.06485664  0.27932960
#> 34     S04     I04 0.5971781 3.018906 -0.72126824 -0.42635348 -1.27410579
#> 48     S04     I05 2.2582697 3.018906 -0.72126824  1.12736729 -1.16673493
#> 51     S04     I06 4.7742557 3.018906 -0.72126824  1.10087374  1.37574462
#> 63     S04     I07 1.7274217 3.018906 -0.72126824 -0.29432480 -0.27589082
#> 77     S04     I08 3.0309383 3.018906 -0.72126824  0.66207708  0.07122382
#> 90     S04     I09 1.7091782 3.018906 -0.72126824 -0.06026755 -0.52819164
#> 92     S04     I10 1.0661287 3.018906 -0.72126824  0.35539173 -1.58690035
#> 2      S05     I01 4.7306599 3.018906 -0.09666313  1.19133200  0.61708542
#> 16     S05     I02 1.8091947 3.018906 -0.09666313  0.14091305 -1.25396080
#> 22     S05     I03 2.3451278 3.018906 -0.09666313  0.06485664 -0.64197132
#> 35     S05     I04 2.3940620 3.018906 -0.09666313 -0.42635348 -0.10182698
#> 45     S05     I05 4.5573163 3.018906 -0.09666313  1.12736729  0.50770650
#> 52     S05     I06 3.9619034 3.018906 -0.09666313  1.10087374 -0.06121280
#> 64     S05     I07 3.0178613 3.018906 -0.09666313 -0.29432480  0.38994363
#> 74     S05     I08 4.9287422 3.018906 -0.09666313  0.66207708  1.34442260
#> 81     S05     I09 2.0351907 3.018906 -0.09666313 -0.06026755 -0.82678427
#> 93     S05     I10 4.0721195 3.018906 -0.09666313  0.35539173  0.79448526
#> 8      S06     I01 8.5191887 3.018906  0.27574486  1.19133200  4.03320622
#> 20     S06     I02 4.5425418 3.018906  0.27574486  0.14091305  1.10697827
#> 27     S06     I03 5.3935785 3.018906  0.27574486  0.06485664  2.03407141
#> 37     S06     I04 2.5990446 3.018906  0.27574486 -0.42635348 -0.26925242
#> 49     S06     I05 6.1673410 3.018906  0.27574486  1.12736729  1.74532329
#> 56     S06     I06 2.9396400 3.018906  0.27574486  1.10087374 -1.45588420
#> 66     S06     I07 3.5754168 3.018906  0.27574486 -0.29432480  0.57509110
#> 79     S06     I08 5.9921398 3.018906  0.27574486  0.66207708  2.03541222
#> 85     S06     I09 3.4217117 3.018906  0.27574486 -0.06026755  0.18732875
#> 99     S06     I10 3.7023838 3.018906  0.27574486  0.35539173  0.05234157
#> 9      S07     I01 4.8096390 3.018906  1.66441057  1.19133200 -1.06500917
#> 11     S07     I02 5.0047442 3.018906  1.66441057  0.14091305  0.18051494
#> 24     S07     I03 5.2533366 3.018906  1.66441057  0.06485664  0.50516381
#> 38     S07     I04 5.8410663 3.018906  1.66441057 -0.42635348  1.58410357
#> 50     S07     I05 6.3028370 3.018906  1.66441057  1.12736729  0.49215349
#> 53     S07     I06 4.3370986 3.018906  1.66441057  1.10087374 -1.44709132
#> 67     S07     I07 5.3543573 3.018906  1.66441057 -0.29432480  0.96536587
#> 80     S07     I08 6.7314783 3.018906  1.66441057  0.66207708  1.38608500
#> 82     S07     I09 4.1947866 3.018906  1.66441057 -0.06026755 -0.42826206
#> 96     S07     I10 4.0377783 3.018906  1.66441057  0.35539173 -1.00092957
#> 6      S08     I01 3.1992472 3.018906 -0.64871111  1.19133200 -0.36227934
#> 12     S08     I02 2.3541248 3.018906 -0.64871111  0.14091305 -0.15698273
#> 25     S08     I03 3.5737760 3.018906 -0.64871111  0.06485664  1.13872487
#> 39     S08     I04 2.4713875 3.018906 -0.64871111 -0.42635348  0.52754647
#> 41     S08     I05 3.8748313 3.018906 -0.64871111  1.12736729  0.37726954
#> 54     S08     I06 3.2499115 3.018906 -0.64871111  1.10087374 -0.22115670
#> 68     S08     I07 1.7740479 3.018906 -0.64871111 -0.29432480 -0.30182180
#> 78     S08     I08 4.5566479 3.018906 -0.64871111  0.66207708  1.52437635
#> 83     S08     I09 1.0542787 3.018906 -0.64871111 -0.06026755 -1.25564825
#> 97     S08     I10 3.2486453 3.018906 -0.64871111  0.35539173  0.52305908
#> 3      S09     I01 7.0228290 3.018906  0.76380851  1.19133200  2.04878290
#> 13     S09     I02 3.9522030 3.018906  0.76380851  0.14091305  0.02857582
#> 26     S09     I03 1.7458393 3.018906  0.76380851  0.06485664 -2.10173145
#> 32     S09     I04 4.7553102 3.018906  0.76380851 -0.42635348  1.39894955
#> 42     S09     I05 6.6840208 3.018906  0.76380851  1.12736729  1.77393934
#> 55     S09     I06 6.3029382 3.018906  0.76380851  1.10087374  1.41935035
#> 65     S09     I07 3.0003536 3.018906  0.76380851 -0.29432480 -0.48803575
#> 71     S09     I08 5.5210688 3.018906  0.76380851  0.66207708  1.07627757
#> 84     S09     I09 3.3168697 3.018906  0.76380851 -0.06026755 -0.40557691
#> 94     S09     I10 3.9206237 3.018906  0.76380851  0.35539173 -0.21748212
#> 10     S10     I01 5.3804615 3.018906  0.39888947  1.19133200  0.77133444
#> 14     S10     I02 2.5611011 3.018906  0.39888947  0.14091305 -0.99760699
#> 28     S10     I03 3.1531022 3.018906  0.39888947  0.06485664 -0.32954948
#> 40     S10     I04 2.8808465 3.018906  0.39888947 -0.42635348 -0.11059510
#> 43     S10     I05 2.8139884 3.018906  0.39888947  1.12736729 -1.73117400
#> 57     S10     I06 4.3653339 3.018906  0.39888947  1.10087374 -0.15333491
#> 70     S10     I07 3.7041212 3.018906  0.39888947 -0.29432480  0.58065095
#> 72     S10     I08 5.8502180 3.018906  0.39888947  0.66207708  1.77034588
#> 86     S10     I09 3.7997745 3.018906  0.39888947 -0.06026755  0.44224694
#> 100    S10     I10 4.6078044 3.018906  0.39888947  0.35539173  0.83461757