Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(data, sub_n = NULL, item_n = NULL, dv = "y",
  sub_id = "sub_id", item_id = "item_id", seed = NULL)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

seed
a single value, interpreted as an integer, or NULL (see set.seed)

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id          y  grand_i      sub_i      item_i         err
#> 1      S01     I01  2.4008697 3.018906 -1.4858292  1.10996057 -0.24216729
#> 2      S02     I01  4.9717877 3.018906 -0.3736548  1.10996057  1.21657631
#> 3      S03     I01  5.4189031 3.018906 -0.2300894  1.10996057  1.52012623
#> 4      S04     I01  5.3620990 3.018906  0.4321905  1.10996057  0.80104224
#> 5      S05     I01  2.9027353 3.018906 -0.3381691  1.10996057 -0.88796187
#> 6      S06     I01  4.2411257 3.018906 -0.3734755  1.10996057  0.48573496
#> 7      S07     I01  3.2022008 3.018906 -1.5441701  1.10996057  0.61750473
#> 8      S08     I01  4.9224786 3.018906  0.8966993  1.10996057 -0.10308689
#> 9      S09     I01  2.6882234 3.018906 -1.1034909  1.10996057 -0.33715185
#> 10     S10     I01  3.6903697 3.018906 -1.2635262  1.10996057  0.82502977
#> 11     S01     I02  1.8991966 3.018906 -1.4858292  0.44193802 -0.07581786
#> 12     S02     I02  4.5901387 3.018906 -0.3736548  0.44193802  1.50294990
#> 13     S03     I02  2.3495705 3.018906 -0.2300894  0.44193802 -0.88118381
#> 14     S04     I02  4.6641174 3.018906  0.4321905  0.44193802  0.77108325
#> 15     S05     I02  1.4325322 3.018906 -0.3381691  0.44193802 -1.69014233
#> 16     S06     I02  3.4973336 3.018906 -0.3734755  0.44193802  0.40996540
#> 17     S07     I02  2.1575913 3.018906 -1.5441701  0.44193802  0.24091774
#> 18     S08     I02  4.0323027 3.018906  0.8966993  0.44193802 -0.32524018
#> 19     S09     I02  2.4516536 3.018906 -1.1034909  0.44193802  0.09430086
#> 20     S10     I02  0.1344304 3.018906 -1.2635262  0.44193802 -2.06288703
#> 21     S01     I03 -0.1002820 3.018906 -1.4858292 -0.62891672 -1.00444171
#> 22     S02     I03  2.1162079 3.018906 -0.3736548 -0.62891672  0.09987377
#> 23     S03     I03  1.8309002 3.018906 -0.2300894 -0.62891672 -0.32899930
#> 24     S04     I03  4.3704758 3.018906  0.4321905 -0.62891672  1.54829637
#> 25     S05     I03  1.7214935 3.018906 -0.3381691 -0.62891672 -0.33032635
#> 26     S06     I03  0.8899895 3.018906 -0.3734755 -0.62891672 -1.12652394
#> 27     S07     I03  0.1619067 3.018906 -1.5441701 -0.62891672 -0.68391208
#> 28     S08     I03  2.3906376 3.018906  0.8966993 -0.62891672 -0.89605061
#> 29     S09     I03  1.4635686 3.018906 -1.1034909 -0.62891672  0.17707062
#> 30     S10     I03  1.6449717 3.018906 -1.2635262 -0.62891672  0.51850906
#> 31     S01     I04  2.0914241 3.018906 -1.4858292  0.23567328  0.32267445
#> 32     S02     I04  2.7992890 3.018906 -0.3736548  0.23567328 -0.08163504
#> 33     S03     I04  3.3126252 3.018906 -0.2300894  0.23567328  0.28813566
#> 34     S04     I04  1.7338304 3.018906  0.4321905  0.23567328 -1.95293899
#> 35     S05     I04  2.5813247 3.018906 -0.3381691  0.23567328 -0.33508515
#> 36     S06     I04  3.5256130 3.018906 -0.3734755  0.23567328  0.64450953
#> 37     S07     I04  1.6335678 3.018906 -1.5441701  0.23567328 -0.07684097
#> 38     S08     I04  4.8555461 3.018906  0.8966993  0.23567328  0.70426789
#> 39     S09     I04  3.7811426 3.018906 -1.1034909  0.23567328  1.63005464
#> 40     S10     I04  2.5131255 3.018906 -1.2635262  0.23567328  0.52207280
#> 41     S01     I05  1.0154786 3.018906 -1.4858292  0.44627473 -0.96387251
#> 42     S02     I05  2.8578026 3.018906 -0.3736548  0.44627473 -0.23372297
#> 43     S03     I05  3.0433703 3.018906 -0.2300894  0.44627473 -0.19172065
#> 44     S04     I05  4.3426054 3.018906  0.4321905  0.44627473  0.44523456
#> 45     S05     I05  2.2109829 3.018906 -0.3381691  0.44627473 -0.91602836
#> 46     S06     I05  3.8526137 3.018906 -0.3734755  0.44627473  0.76090879
#> 47     S07     I05  0.3902985 3.018906 -1.5441701  0.44627473 -1.53071173
#> 48     S08     I05  3.8573251 3.018906  0.8966993  0.44627473 -0.50455451
#> 49     S09     I05  1.3611386 3.018906 -1.1034909  0.44627473 -1.00055084
#> 50     S10     I05  2.2509881 3.018906 -1.2635262  0.44627473  0.04933399
#> 51     S01     I06  1.8427136 3.018906 -1.4858292  0.29684130  0.01279592
#> 52     S02     I06  4.1802134 3.018906 -0.3736548  0.29684130  1.23812134
#> 53     S03     I06  1.8498935 3.018906 -0.2300894  0.29684130 -1.23576401
#> 54     S04     I06  1.7601252 3.018906  0.4321905  0.29684130 -1.98781220
#> 55     S05     I06  3.9704656 3.018906 -0.3381691  0.29684130  0.99288778
#> 56     S06     I06  4.0702273 3.018906 -0.3734755  0.29684130  1.12795585
#> 57     S07     I06  2.3501160 3.018906 -1.5441701  0.29684130  0.57853915
#> 58     S08     I06  4.4776663 3.018906  0.8966993  0.29684130  0.26522010
#> 59     S09     I06  1.7178921 3.018906 -1.1034909  0.29684130 -0.49436383
#> 60     S10     I06  2.5762166 3.018906 -1.2635262  0.29684130  0.52399594
#> 61     S01     I07  1.9880027 3.018906 -1.4858292  1.82020722 -1.36528090
#> 62     S02     I07  4.8421560 3.018906 -0.3736548  1.82020722  0.37669794
#> 63     S03     I07  3.7636113 3.018906 -0.2300894  1.82020722 -0.84541214
#> 64     S04     I07  6.2954059 3.018906  0.4321905  1.82020722  1.02410255
#> 65     S05     I07  4.1332852 3.018906 -0.3381691  1.82020722 -0.36765854
#> 66     S06     I07  5.5859448 3.018906 -0.3734755  1.82020722  1.12030746
#> 67     S07     I07  2.2586655 3.018906 -1.5441701  1.82020722 -1.03627726
#> 68     S08     I07  7.5108759 3.018906  0.8966993  1.82020722  1.77506382
#> 69     S09     I07  5.7193069 3.018906 -1.1034909  1.82020722  1.98368495
#> 70     S10     I07  3.4306896 3.018906 -1.2635262  1.82020722 -0.14489703
#> 71     S01     I08  4.5394902 3.018906 -1.4858292  1.59042372  1.41599007
#> 72     S02     I08  4.1235177 3.018906 -0.3736548  1.59042372 -0.11215679
#> 73     S03     I08  6.2719541 3.018906 -0.2300894  1.59042372  1.89271412
#> 74     S04     I08  4.2558402 3.018906  0.4321905  1.59042372 -0.78567961
#> 75     S05     I08  6.2184505 3.018906 -0.3381691  1.59042372  1.94729023
#> 76     S06     I08  4.1409791 3.018906 -0.3734755  1.59042372 -0.09487477
#> 77     S07     I08  3.3376873 3.018906 -1.5441701  1.59042372  0.27252806
#> 78     S08     I08  4.0217403 3.018906  0.8966993  1.59042372 -1.48428831
#> 79     S09     I08  4.3984224 3.018906 -1.1034909  1.59042372  0.89258395
#> 80     S10     I08  1.4544618 3.018906 -1.2635262  1.59042372 -1.89134129
#> 81     S01     I09  1.2566182 3.018906 -1.4858292  0.01233648 -0.28879463
#> 82     S02     I09  2.0229494 3.018906 -0.3736548  0.01233648 -0.63463786
#> 83     S03     I09  2.5543253 3.018906 -0.2300894  0.01233648 -0.24682744
#> 84     S04     I09  4.2957542 3.018906  0.4321905  0.01233648  0.83232160
#> 85     S05     I09  2.9512282 3.018906 -0.3381691  0.01233648  0.25815516
#> 86     S06     I09  3.1498029 3.018906 -0.3734755  0.01233648  0.49203622
#> 87     S07     I09  0.1371803 3.018906 -1.5441701  0.01233648 -1.34989171
#> 88     S08     I09  5.5629414 3.018906  0.8966993  0.01233648  1.63499997
#> 89     S09     I09  2.0969154 3.018906 -1.1034909  0.01233648  0.16916421
#> 90     S10     I09  0.4710989 3.018906 -1.2635262  0.01233648 -1.29661699
#> 91     S01     I10  0.7106320 3.018906 -1.4858292 -0.21627724 -0.60616712
#> 92     S02     I10  1.5946777 3.018906 -0.3736548 -0.21627724 -0.83429582
#> 93     S03     I10  1.7391832 3.018906 -0.2300894 -0.21627724 -0.83335583
#> 94     S04     I10  3.2703874 3.018906  0.4321905 -0.21627724  0.03556849
#> 95     S05     I10  3.6384155 3.018906 -0.3381691 -0.21627724  1.17395617
#> 96     S06     I10  3.7120132 3.018906 -0.3734755 -0.21627724  1.28286029
#> 97     S07     I10  1.5279513 3.018906 -1.5441701 -0.21627724  0.26949302
#> 98     S08     I10  4.6500337 3.018906  0.8966993 -0.21627724  0.95070605
#> 99     S09     I10  1.6616104 3.018906 -1.1034909 -0.21627724 -0.03752709
#> 100    S10     I10  0.4361248 3.018906 -1.2635262 -0.21627724 -1.10297734