Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

Usage

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id          y  grand_i       sub_i     item_i          err
#> 1      S01     I01  5.1445385 3.018906  1.29000219  0.1384851  0.697145637
#> 15     S01     I02  4.5831518 3.018906  1.29000219  0.7862365 -0.511992534
#> 29     S01     I03  6.2903638 3.018906  1.29000219  1.1522226  0.829233394
#> 31     S01     I04  1.9659806 3.018906  1.29000219 -1.9761691 -0.366758028
#> 44     S01     I05  2.2797487 3.018906  1.29000219 -1.3879907 -0.641168432
#> 58     S01     I06  4.1391827 3.018906  1.29000219  0.7830445 -0.952769613
#> 61     S01     I07  4.5435632 3.018906  1.29000219 -0.2193914  0.454046713
#> 73     S01     I08  5.3625272 3.018906  1.29000219  0.1705658  0.883053635
#> 87     S01     I09  3.7728960 3.018906  1.29000219  0.5902264 -1.126238260
#> 91     S01     I10  5.2397040 3.018906  1.29000219  0.1595821  0.771214147
#> 7      S02     I01  4.7460033 3.018906  0.93650865  0.1384851  0.652103921
#> 17     S02     I02  6.1721877 3.018906  0.93650865  0.7862365  1.430536960
#> 30     S02     I03  4.6238804 3.018906  0.93650865  1.1522226 -0.483756460
#> 36     S02     I04  0.8693031 3.018906  0.93650865 -1.9761691 -1.109941990
#> 46     S02     I05  3.2792254 3.018906  0.93650865 -1.3879907  0.711801779
#> 59     S02     I06  2.8901540 3.018906  0.93650865  0.7830445 -1.848304707
#> 69     S02     I07  3.2680378 3.018906  0.93650865 -0.2193914 -0.467985130
#> 75     S02     I08  3.1824453 3.018906  0.93650865  0.1705658 -0.943534683
#> 88     S02     I09  4.6343094 3.018906  0.93650865  0.5902264  0.088668662
#> 98     S02     I10  4.2472706 3.018906  0.93650865  0.1595821  0.132274295
#> 4      S03     I01  5.8755051 3.018906  0.17581093  0.1384851  2.542303498
#> 18     S03     I02  4.9975938 3.018906  0.17581093  0.7862365  1.016640741
#> 23     S03     I03  2.5076014 3.018906  0.17581093  1.1522226 -1.839337784
#> 33     S03     I04  1.7281696 3.018906  0.17581093 -1.9761691  0.509622185
#> 47     S03     I05  3.1947292 3.018906  0.17581093 -1.3879907  1.388003358
#> 60     S03     I06  4.0770324 3.018906  0.17581093  0.7830445  0.099271395
#> 62     S03     I07  3.1270347 3.018906  0.17581093 -0.2193914  0.151709486
#> 76     S03     I08  4.9891856 3.018906  0.17581093  0.1705658  1.623903306
#> 89     S03     I09  3.9156067 3.018906  0.17581093  0.5902264  0.130663698
#> 95     S03     I10  3.3753272 3.018906  0.17581093  0.1595821  0.021028606
#> 5      S04     I01  3.4884043 3.018906  1.44840672  0.1384851 -1.117393093
#> 19     S04     I02  2.8538712 3.018906  1.44840672  0.7862365 -2.399677611
#> 21     S04     I03  2.9233846 3.018906  1.44840672  1.1522226 -2.696150366
#> 34     S04     I04  3.4378336 3.018906  1.44840672 -1.9761691  0.946690442
#> 48     S04     I05  1.9327416 3.018906  1.44840672 -1.3879907 -1.146580082
#> 51     S04     I06  5.0508056 3.018906  1.44840672  0.7830445 -0.199551219
#> 63     S04     I07  4.5124689 3.018906  1.44840672 -0.2193914  0.264547882
#> 77     S04     I08  5.8413190 3.018906  1.44840672  0.1705658  1.203440887
#> 90     S04     I09  4.8970500 3.018906  1.44840672  0.5902264 -0.160488741
#> 92     S04     I10  3.6451019 3.018906  1.44840672  0.1595821 -0.981792477
#> 2      S05     I01  3.9742561 3.018906 -0.35702302  0.1384851  1.173888449
#> 16     S05     I02  5.8325461 3.018906 -0.35702302  0.7862365  2.384427062
#> 22     S05     I03  3.7567127 3.018906 -0.35702302  1.1522226 -0.057392487
#> 35     S05     I04 -0.7987169 3.018906 -0.35702302 -1.9761691 -1.484430384
#> 45     S05     I05  1.1239567 3.018906 -0.35702302 -1.3879907 -0.149935190
#> 52     S05     I06  5.4956829 3.018906 -0.35702302  0.7830445  2.050755816
#> 64     S05     I07  3.2291273 3.018906 -0.35702302 -0.2193914  0.786636022
#> 74     S05     I08  3.9854702 3.018906 -0.35702302  0.1705658  1.153021887
#> 81     S05     I09  2.7436375 3.018906 -0.35702302  0.5902264 -0.508471510
#> 93     S05     I10  2.1653187 3.018906 -0.35702302  0.1595821 -0.656146006
#> 8      S06     I01  2.6866056 3.018906 -0.67872620  0.1384851  0.207941066
#> 20     S06     I02  3.3394005 3.018906 -0.67872620  0.7862365  0.212984577
#> 27     S06     I03  0.8448473 3.018906 -0.67872620  1.1522226 -2.647554671
#> 37     S06     I04 -0.2589242 3.018906 -0.67872620 -1.9761691 -0.622934462
#> 49     S06     I05  0.6714229 3.018906 -0.67872620 -1.3879907 -0.280765807
#> 56     S06     I06  5.5312650 3.018906 -0.67872620  0.7830445  2.408041091
#> 66     S06     I07  3.5175394 3.018906 -0.67872620 -0.2193914  1.396751377
#> 79     S06     I08  2.5753009 3.018906 -0.67872620  0.1705658  0.064555670
#> 85     S06     I09  3.3175521 3.018906 -0.67872620  0.5902264  0.387146295
#> 99     S06     I10  1.7700478 3.018906 -0.67872620  0.1595821 -0.729713720
#> 9      S07     I01  3.4676971 3.018906 -0.07213936  0.1384851  0.382445780
#> 11     S07     I02  4.0748293 3.018906 -0.07213936  0.7862365  0.341826527
#> 24     S07     I03  4.1038070 3.018906 -0.07213936  1.1522226  0.004818110
#> 38     S07     I04  2.9059294 3.018906 -0.07213936 -1.9761691  1.935332287
#> 50     S07     I05  1.6839386 3.018906 -0.07213936 -1.3879907  0.125163028
#> 53     S07     I06  2.9000453 3.018906 -0.07213936  0.7830445 -0.829765450
#> 67     S07     I07  3.0235152 3.018906 -0.07213936 -0.2193914  0.296140340
#> 80     S07     I08  1.2309482 3.018906 -0.07213936  0.1705658 -1.886383820
#> 82     S07     I09  1.8800771 3.018906 -0.07213936  0.5902264 -1.656915555
#> 96     S07     I10  3.2645299 3.018906 -0.07213936  0.1595821  0.158181606
#> 6      S08     I01  2.5744348 3.018906 -0.42460035  0.1384851 -0.158355575
#> 12     S08     I02  2.9587862 3.018906 -0.42460035  0.7862365 -0.421755509
#> 25     S08     I03  1.8185478 3.018906 -0.42460035  1.1522226 -1.927980050
#> 39     S08     I04  1.9413767 3.018906 -0.42460035 -1.9761691  1.323240566
#> 41     S08     I05 -0.3944326 3.018906 -0.42460035 -1.3879907 -1.600747147
#> 54     S08     I06  2.3034704 3.018906 -0.42460035  0.7830445 -1.073879336
#> 68     S08     I07  2.7146272 3.018906 -0.42460035 -0.2193914  0.339713298
#> 78     S08     I08  3.5806118 3.018906 -0.42460035  0.1705658  0.815740737
#> 83     S08     I09  3.4077312 3.018906 -0.42460035  0.5902264  0.223199497
#> 97     S08     I10  2.9202905 3.018906 -0.42460035  0.1595821  0.166403151
#> 3      S09     I01  5.1023352 3.018906  1.50752855  0.1384851  0.437415918
#> 13     S09     I02  6.7325976 3.018906  1.50752855  0.7862365  1.419926919
#> 26     S09     I03  6.3379190 3.018906  1.50752855  1.1522226  0.659262241
#> 32     S09     I04  4.0736702 3.018906  1.50752855 -1.9761691  1.523405126
#> 42     S09     I05  2.8985229 3.018906  1.50752855 -1.3879907 -0.239920564
#> 55     S09     I06  4.9331600 3.018906  1.50752855  0.7830445 -0.376318593
#> 65     S09     I07  3.1782824 3.018906  1.50752855 -0.2193914 -1.128760446
#> 71     S09     I08  6.0464353 3.018906  1.50752855  0.1705658  1.349435402
#> 84     S09     I09  5.7644610 3.018906  1.50752855  0.5902264  0.647800436
#> 94     S09     I10  5.1842353 3.018906  1.50752855  0.1595821  0.498219026
#> 10     S10     I01  5.5891458 3.018906  0.74288624  0.1384851  1.688868852
#> 14     S10     I02  2.4439146 3.018906  0.74288624  0.7862365 -2.104113704
#> 28     S10     I03  6.5402573 3.018906  0.74288624  1.1522226  1.626242804
#> 40     S10     I04  2.5593023 3.018906  0.74288624 -1.9761691  0.773679576
#> 43     S10     I05  4.2730525 3.018906  0.74288624 -1.3879907  1.899251312
#> 57     S10     I06  5.4240240 3.018906  0.74288624  0.7830445  0.879187724
#> 70     S10     I07  2.5600420 3.018906  0.74288624 -0.2193914 -0.982358528
#> 72     S10     I08  4.8877662 3.018906  0.74288624  0.1705658  0.955408583
#> 86     S10     I09  4.3399112 3.018906  0.74288624  0.5902264 -0.012107065
#> 100    S10     I10  3.9264272 3.018906  0.74288624  0.1595821  0.005053226