Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

Usage

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id          y  grand_i        sub_i     item_i         err
#> 1      S01     I01  1.4243580 3.018906 -0.290252297 -1.0797522 -0.22454315
#> 15     S01     I02  1.4399799 3.018906 -0.290252297 -0.3773387 -0.91133470
#> 29     S01     I03 -0.6339352 3.018906 -0.290252297 -2.0195229 -1.34306559
#> 31     S01     I04  2.1534965 3.018906 -0.290252297 -1.0681476  0.49299075
#> 44     S01     I05  5.0850665 3.018906 -0.290252297  1.5873368  0.76907633
#> 58     S01     I06  3.6168314 3.018906 -0.290252297  0.1047402  0.78343786
#> 61     S01     I07  4.7270912 3.018906 -0.290252297  1.1127491  0.88568881
#> 73     S01     I08  0.8670285 3.018906 -0.290252297  0.1039260 -1.96555080
#> 87     S01     I09  3.7173207 3.018906 -0.290252297  0.1455229  0.84314448
#> 91     S01     I10  4.7781721 3.018906 -0.290252297  1.0389862  1.01053252
#> 7      S02     I01  0.4560245 3.018906 -0.849168495 -1.0797522 -0.63396042
#> 17     S02     I02  3.5330370 3.018906 -0.849168495 -0.3773387  1.74063857
#> 30     S02     I03  0.7693366 3.018906 -0.849168495 -2.0195229  0.61912236
#> 36     S02     I04  0.3812361 3.018906 -0.849168495 -1.0681476 -0.72035345
#> 46     S02     I05  2.4040521 3.018906 -0.849168495  1.5873368 -1.35302180
#> 59     S02     I06  1.2551642 3.018906 -0.849168495  0.1047402 -1.01931313
#> 69     S02     I07  3.6395446 3.018906 -0.849168495  1.1127491  0.35705846
#> 75     S02     I08  0.9832117 3.018906 -0.849168495  0.1039260 -1.29045149
#> 88     S02     I09  2.5287186 3.018906 -0.849168495  0.1455229  0.21345858
#> 98     S02     I10  2.9117986 3.018906 -0.849168495  1.0389862 -0.29692479
#> 4      S03     I01  2.5625935 3.018906  2.505858738 -1.0797522 -1.88241867
#> 18     S03     I02  1.7988836 3.018906  2.505858738 -0.3773387 -3.34854199
#> 23     S03     I03  4.1514437 3.018906  2.505858738 -2.0195229  0.64620218
#> 33     S03     I04  4.9832177 3.018906  2.505858738 -1.0681476  0.52660091
#> 47     S03     I05  5.9898986 3.018906  2.505858738  1.5873368 -1.12220258
#> 60     S03     I06  5.8509157 3.018906  2.505858738  0.1047402  0.22141119
#> 62     S03     I07  6.8808179 3.018906  2.505858738  1.1127491  0.24330452
#> 76     S03     I08  6.3580458 3.018906  2.505858738  0.1039260  0.72935543
#> 89     S03     I09  6.3843388 3.018906  2.505858738  0.1455229  0.71405162
#> 95     S03     I10  4.8099155 3.018906  2.505858738  1.0389862 -1.75383506
#> 5      S04     I01  3.3865495 3.018906  0.076078144 -1.0797522  1.37131794
#> 19     S04     I02  5.5966711 3.018906  0.076078144 -0.3773387  2.87902610
#> 21     S04     I03  0.3690713 3.018906  0.076078144 -2.0195229 -0.70638953
#> 34     S04     I04  2.4708917 3.018906  0.076078144 -1.0681476  0.44405553
#> 48     S04     I05  4.0641114 3.018906  0.076078144  1.5873368 -0.61820915
#> 51     S04     I06  1.6258899 3.018906  0.076078144  0.1047402 -1.57383403
#> 63     S04     I07  5.1381946 3.018906  0.076078144  1.1127491  0.93046177
#> 77     S04     I08  1.8656648 3.018906  0.076078144  0.1039260 -1.33324497
#> 90     S04     I09  3.8945743 3.018906  0.076078144  0.1455229  0.65406770
#> 92     S04     I10  4.7882769 3.018906  0.076078144  1.0389862  0.65430690
#> 2      S05     I01  1.8372530 3.018906  0.126464694 -1.0797522 -0.22836515
#> 16     S05     I02  3.5303375 3.018906  0.126464694 -0.3773387  0.76230592
#> 22     S05     I03 -0.8996093 3.018906  0.126464694 -2.0195229 -2.02545669
#> 35     S05     I04  2.7431538 3.018906  0.126464694 -1.0681476  0.66593110
#> 45     S05     I05  4.3769949 3.018906  0.126464694  1.5873368 -0.35571226
#> 52     S05     I06  3.9118975 3.018906  0.126464694  0.1047402  0.66178703
#> 64     S05     I07  3.0578203 3.018906  0.126464694  1.1127491 -1.20029912
#> 74     S05     I08  1.4060333 3.018906  0.126464694  0.1039260 -1.84326304
#> 81     S05     I09  1.6894256 3.018906  0.126464694  0.1455229 -1.60146754
#> 93     S05     I10  3.7148103 3.018906  0.126464694  1.0389862 -0.46954622
#> 8      S06     I01  2.4783745 3.018906 -0.217127712 -1.0797522  0.75634877
#> 20     S06     I02  1.3084870 3.018906 -0.217127712 -0.3773387 -1.11595218
#> 27     S06     I03  0.3588164 3.018906 -0.217127712 -2.0195229 -0.42343865
#> 37     S06     I04  1.8086161 3.018906 -0.217127712 -1.0681476  0.07498583
#> 49     S06     I05  3.9140666 3.018906 -0.217127712  1.5873368 -0.47504807
#> 56     S06     I06  2.8517099 3.018906 -0.217127712  0.1047402 -0.05480824
#> 66     S06     I07  2.1140058 3.018906 -0.217127712  1.1127491 -1.80052114
#> 79     S06     I08  3.0135384 3.018906 -0.217127712  0.1039260  0.10783448
#> 85     S06     I09  3.5463178 3.018906 -0.217127712  0.1455229  0.59901703
#> 99     S06     I10  5.1589974 3.018906 -0.217127712  1.0389862  1.31823329
#> 9      S07     I01  2.4081956 3.018906  0.868188070 -1.0797522 -0.39914597
#> 11     S07     I02  3.0048827 3.018906  0.868188070 -0.3773387 -0.50487227
#> 24     S07     I03  1.3462491 3.018906  0.868188070 -2.0195229 -0.52132171
#> 38     S07     I04  2.6799047 3.018906  0.868188070 -1.0681476 -0.13904137
#> 50     S07     I05  4.7946689 3.018906  0.868188070  1.5873368 -0.67976163
#> 53     S07     I06  5.0149360 3.018906  0.868188070  0.1047402  1.02310214
#> 67     S07     I07  4.5989739 3.018906  0.868188070  1.1127491 -0.40086881
#> 80     S07     I08  4.4645966 3.018906  0.868188070  0.1039260  0.47357685
#> 82     S07     I09  3.5783589 3.018906  0.868188070  0.1455229 -0.45425765
#> 96     S07     I10  4.4238533 3.018906  0.868188070  1.0389862 -0.50222660
#> 6      S08     I01  2.8963291 3.018906 -0.006752759 -1.0797522  0.96392842
#> 12     S08     I02  3.3085028 3.018906 -0.006752759 -0.3773387  0.67368873
#> 25     S08     I03  0.3296455 3.018906 -0.006752759 -2.0195229 -0.66298451
#> 39     S08     I04  1.8630845 3.018906 -0.006752759 -1.0681476 -0.08092075
#> 41     S08     I05  4.1572777 3.018906 -0.006752759  1.5873368 -0.44221197
#> 54     S08     I06  1.9766239 3.018906 -0.006752759  0.1047402 -1.14026914
#> 68     S08     I07  5.0674361 3.018906 -0.006752759  1.1127491  0.94253422
#> 78     S08     I08  3.9549169 3.018906 -0.006752759  0.1039260  0.83883800
#> 83     S08     I09  4.4615487 3.018906 -0.006752759  0.1455229  1.30387302
#> 97     S08     I10  4.5070772 3.018906 -0.006752759  1.0389862  0.45593808
#> 3      S09     I01  0.8378567 3.018906  0.321940218 -1.0797522 -1.42323696
#> 13     S09     I02  3.0762202 3.018906  0.321940218 -0.3773387  0.11271314
#> 26     S09     I03  2.5816914 3.018906  0.321940218 -2.0195229  1.26036840
#> 32     S09     I04  0.9466719 3.018906  0.321940218 -1.0681476 -1.32602632
#> 42     S09     I05  5.0083062 3.018906  0.321940218  1.5873368  0.08012351
#> 55     S09     I06  3.7357459 3.018906  0.321940218  0.1047402  0.29015989
#> 65     S09     I07  5.5658355 3.018906  0.321940218  1.1127491  1.11224064
#> 71     S09     I08  4.0171788 3.018906  0.321940218  0.1039260  0.57240693
#> 84     S09     I09  4.4104957 3.018906  0.321940218  0.1455229  0.92412701
#> 94     S09     I10  4.4486647 3.018906  0.321940218  1.0389862  0.06883267
#> 10     S10     I01  3.1930621 3.018906 -0.494652343 -1.0797522  1.74856094
#> 14     S10     I02  3.6512037 3.018906 -0.494652343 -0.3773387  1.50428921
#> 28     S10     I03  0.3701096 3.018906 -0.494652343 -2.0195229 -0.13462077
#> 40     S10     I04  2.3479909 3.018906 -0.494652343 -1.0681476  0.89188519
#> 43     S10     I05  4.7885049 3.018906 -0.494652343  1.5873368  0.67691479
#> 57     S10     I06  1.6677044 3.018906 -0.494652343  0.1047402 -0.96128913
#> 70     S10     I07  4.2990399 3.018906 -0.494652343  1.1127491  0.66203758
#> 72     S10     I08  2.3259090 3.018906 -0.494652343  0.1039260 -0.30227026
#> 86     S10     I09  3.3051108 3.018906 -0.494652343  0.1455229  0.63533464
#> 100    S10     I10  4.9826269 3.018906 -0.494652343  1.0389862  1.41938745