Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id          y  grand_i      sub_i     item_i         err
#> 1      S01     I01  4.0511208 3.018906  0.8574016 -0.6613285  0.83614203
#> 15     S01     I02  4.2521524 3.018906  0.8574016  0.1236549  0.25219027
#> 29     S01     I03  5.1697720 3.018906  0.8574016 -0.2752233  1.56868800
#> 31     S01     I04  4.2529267 3.018906  0.8574016  0.1362527  0.24036680
#> 44     S01     I05  4.2874696 3.018906  0.8574016 -0.3474690  0.75863140
#> 58     S01     I06  3.7948070 3.018906  0.8574016 -0.7926878  0.71118756
#> 61     S01     I07  5.7229490 3.018906  0.8574016  1.2756102  0.57103153
#> 73     S01     I08  2.9613371 3.018906  0.8574016 -0.2874909 -0.62747926
#> 87     S01     I09  3.5617686 3.018906  0.8574016 -0.3980562  0.08351753
#> 91     S01     I10  3.4361355 3.018906  0.8574016  0.5185442 -0.95871595
#> 7      S02     I01  0.6259855 3.018906 -0.1179668 -0.6613285 -1.61362486
#> 17     S02     I02  2.3980448 3.018906 -0.1179668  0.1236549 -0.62654901
#> 30     S02     I03  1.5502546 3.018906 -0.1179668 -0.2752233 -1.07546096
#> 36     S02     I04  4.4551603 3.018906 -0.1179668  0.1362527  1.41796880
#> 46     S02     I05  2.9412913 3.018906 -0.1179668 -0.3474690  0.38782146
#> 59     S02     I06  2.4972964 3.018906 -0.1179668 -0.7926878  0.38904531
#> 69     S02     I07  4.6306322 3.018906 -0.1179668  1.2756102  0.45408316
#> 75     S02     I08  3.7843953 3.018906 -0.1179668 -0.2874909  1.17094734
#> 88     S02     I09  2.9880721 3.018906 -0.1179668 -0.3980562  0.48518949
#> 98     S02     I10  3.3429221 3.018906 -0.1179668  0.5185442 -0.07656097
#> 4      S03     I01 -0.4884503 3.018906 -0.9017917 -0.6613285 -1.94423582
#> 18     S03     I02  3.1878200 3.018906 -0.9017917  0.1236549  0.94705114
#> 23     S03     I03  3.1587626 3.018906 -0.9017917 -0.2752233  1.31687188
#> 33     S03     I04  1.6318079 3.018906 -0.9017917  0.1362527 -0.62155867
#> 47     S03     I05  3.9318612 3.018906 -0.9017917 -0.3474690  2.16221625
#> 60     S03     I06  2.8722779 3.018906 -0.9017917 -0.7926878  1.54785175
#> 62     S03     I07  1.2558431 3.018906 -0.9017917  1.2756102 -2.13688107
#> 76     S03     I08  1.2579804 3.018906 -0.9017917 -0.2874909 -0.57164265
#> 89     S03     I09  1.6059111 3.018906 -0.9017917 -0.3980562 -0.11314669
#> 95     S03     I10  2.0934125 3.018906 -0.9017917  0.5185442 -0.54224565
#> 5      S04     I01  2.8606742 3.018906 -0.6999993 -0.6613285  1.20309632
#> 19     S04     I02  3.5829446 3.018906 -0.6999993  0.1236549  1.14038335
#> 21     S04     I03  1.5919308 3.018906 -0.6999993 -0.2752233 -0.45175222
#> 34     S04     I04  2.4147208 3.018906 -0.6999993  0.1362527 -0.04043820
#> 48     S04     I05  2.8489902 3.018906 -0.6999993 -0.3474690  0.87755288
#> 51     S04     I06  2.3103315 3.018906 -0.6999993 -0.7926878  0.78411296
#> 63     S04     I07  4.5758598 3.018906 -0.6999993  1.2756102  0.98134328
#> 77     S04     I08  0.8115642 3.018906 -0.6999993 -0.2874909 -1.21985122
#> 90     S04     I09  2.5972656 3.018906 -0.6999993 -0.3980562  0.67641548
#> 92     S04     I10  4.2299522 3.018906 -0.6999993  0.5185442  1.39250172
#> 2      S05     I01  3.9874376 3.018906  1.1274375 -0.6613285  0.50242296
#> 16     S05     I02  3.2457267 3.018906  1.1274375  0.1236549 -1.02427137
#> 22     S05     I03  4.1850530 3.018906  1.1274375 -0.2752233  0.31393317
#> 35     S05     I04  5.0310730 3.018906  1.1274375  0.1362527  0.74847724
#> 45     S05     I05  3.8430188 3.018906  1.1274375 -0.3474690  0.04414471
#> 52     S05     I06  3.6293249 3.018906  1.1274375 -0.7926878  0.27566950
#> 64     S05     I07  6.3315220 3.018906  1.1274375  1.2756102  0.90956873
#> 74     S05     I08  4.5529385 3.018906  1.1274375 -0.2874909  0.69408630
#> 81     S05     I09  2.4977296 3.018906  1.1274375 -0.3980562 -1.25055732
#> 93     S05     I10  5.9565441 3.018906  1.1274375  0.5185442  1.29165675
#> 8      S06     I01  1.6190567 3.018906  0.5704626 -0.6613285 -1.30898300
#> 20     S06     I02  4.3008230 3.018906  0.5704626  0.1236549  0.58779984
#> 27     S06     I03  2.8685319 3.018906  0.5704626 -0.2752233 -0.44561307
#> 37     S06     I04  2.6783321 3.018906  0.5704626  0.1362527 -1.04728875
#> 49     S06     I05  2.1749881 3.018906  0.5704626 -0.3474690 -1.06691112
#> 56     S06     I06  2.4391513 3.018906  0.5704626 -0.7926878 -0.35752916
#> 66     S06     I07  4.0401858 3.018906  0.5704626  1.2756102 -0.82479256
#> 79     S06     I08  2.6868892 3.018906  0.5704626 -0.2874909 -0.61498816
#> 85     S06     I09  3.4714632 3.018906  0.5704626 -0.3980562  0.28015116
#> 99     S06     I10  2.6290792 3.018906  0.5704626  0.5185442 -1.47883324
#> 9      S07     I01  1.6372808 3.018906 -1.1516260 -0.6613285  0.43132967
#> 11     S07     I02  5.5925333 3.018906 -1.1516260  0.1236549  3.60159874
#> 24     S07     I03  2.3883181 3.018906 -1.1516260 -0.2752233  0.79626177
#> 38     S07     I04  2.0266735 3.018906 -1.1516260  0.1362527  0.02314120
#> 50     S07     I05  0.7156571 3.018906 -1.1516260 -0.3474690 -0.80415349
#> 53     S07     I06  1.5005414 3.018906 -1.1516260 -0.7926878  0.42594957
#> 67     S07     I07  4.2092142 3.018906 -1.1516260  1.2756102  1.06632445
#> 80     S07     I08  2.3820883 3.018906 -1.1516260 -0.2874909  0.80229956
#> 82     S07     I09  1.7572786 3.018906 -1.1516260 -0.3980562  0.28805521
#> 96     S07     I10  1.5924957 3.018906 -1.1516260  0.5185442 -0.79332807
#> 6      S08     I01  1.4640703 3.018906 -0.9554480 -0.6613285  0.06194119
#> 12     S08     I02  2.7890756 3.018906 -0.9554480  0.1236549  0.60196311
#> 25     S08     I03  1.8350899 3.018906 -0.9554480 -0.2752233  0.04685558
#> 39     S08     I04  1.8239862 3.018906 -0.9554480  0.1362527 -0.37572400
#> 41     S08     I05  1.3028169 3.018906 -0.9554480 -0.3474690 -0.41317174
#> 54     S08     I06  0.7050262 3.018906 -0.9554480 -0.7926878 -0.56574367
#> 68     S08     I07  0.8266966 3.018906 -0.9554480  1.2756102 -2.51237116
#> 78     S08     I08  2.3593087 3.018906 -0.9554480 -0.2874909  0.58334197
#> 83     S08     I09  0.8431222 3.018906 -0.9554480 -0.3980562 -0.82227915
#> 97     S08     I10  3.2837753 3.018906 -0.9554480  0.5185442  0.70177347
#> 3      S09     I01  2.6328120 3.018906  0.3472619 -0.6613285 -0.07202703
#> 13     S09     I02  3.4717832 3.018906  0.3472619  0.1236549 -0.01803927
#> 26     S09     I03  2.3591694 3.018906  0.3472619 -0.2752233 -0.73177479
#> 32     S09     I04  4.1340621 3.018906  0.3472619  0.1362527  0.63164191
#> 42     S09     I05  4.1905888 3.018906  0.3472619 -0.3474690  1.17189034
#> 55     S09     I06  1.0909280 3.018906  0.3472619 -0.7926878 -1.48255176
#> 65     S09     I07  3.4855579 3.018906  0.3472619  1.2756102 -1.15621976
#> 71     S09     I08  1.7218769 3.018906  0.3472619 -0.2874909 -1.35679967
#> 84     S09     I09  3.8409553 3.018906  0.3472619 -0.3980562  0.87284403
#> 94     S09     I10  4.2981792 3.018906  0.3472619  0.5185442  0.41346747
#> 10     S10     I01  0.8124803 3.018906  0.5572061 -0.6613285 -2.10230292
#> 14     S10     I02  3.1847850 3.018906  0.5572061  0.1236549 -0.51498167
#> 28     S10     I03  3.7295094 3.018906  0.5572061 -0.2752233  0.42862098
#> 40     S10     I04  6.3655131 3.018906  0.5572061  0.1362527  2.65314870
#> 43     S10     I05  4.0707565 3.018906  0.5572061 -0.3474690  0.84211379
#> 57     S10     I06  1.4996674 3.018906  0.5572061 -0.7926878 -1.28375651
#> 70     S10     I07  5.5649492 3.018906  0.5572061  1.2756102  0.71322729
#> 72     S10     I08  2.8446188 3.018906  0.5572061 -0.2874909 -0.44400199
#> 86     S10     I09  2.6293299 3.018906  0.5572061 -0.3980562 -0.54872560
#> 100    S10     I10  5.6679392 3.018906  0.5572061  0.5185442  1.57328326