Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id             y  grand_i      sub_i     item_i         err
#> 1      S01     I01  2.1716950631 3.018906 -0.7798356  1.0263826 -1.09375755
#> 15     S01     I02  3.3012387446 3.018906 -0.7798356  0.2941346  0.76803412
#> 29     S01     I03  0.9840653184 3.018906 -0.7798356 -0.7345863 -0.52041839
#> 31     S01     I04  3.4734724632 3.018906 -0.7798356  0.6676717  0.56673072
#> 44     S01     I05  3.2031341301 3.018906 -0.7798356  0.3318599  0.63220420
#> 58     S01     I06  1.9065026912 3.018906 -0.7798356 -0.3062988 -0.02626852
#> 61     S01     I07  4.3534968942 3.018906 -0.7798356  0.3069992  1.80742760
#> 73     S01     I08  2.3057216078 3.018906 -0.7798356 -0.8295020  0.89615353
#> 87     S01     I09 -0.0004433304 3.018906 -0.7798356 -0.7193702 -1.52014316
#> 91     S01     I10  3.4333762050 3.018906 -0.7798356  0.2905044  0.90380179
#> 7      S02     I01  5.4431080524 3.018906  0.3715555  1.0263826  1.02626443
#> 17     S02     I02  4.8759783125 3.018906  0.3715555  0.2941346  1.19138268
#> 30     S02     I03  2.3042707868 3.018906  0.3715555 -0.7345863 -0.35160393
#> 36     S02     I04  4.7886419092 3.018906  0.3715555  0.6676717  0.73050915
#> 46     S02     I05  1.5127625194 3.018906  0.3715555  0.3318599 -2.20955842
#> 59     S02     I06  3.2705068207 3.018906  0.3715555 -0.3062988  0.18634460
#> 69     S02     I07  3.9124547844 3.018906  0.3715555  0.3069992  0.21499448
#> 75     S02     I08  4.9947804252 3.018906  0.3715555 -0.8295020  2.43382134
#> 88     S02     I09  3.3994926373 3.018906  0.3715555 -0.7193702  0.72840180
#> 98     S02     I10  2.2138385228 3.018906  0.3715555  0.2905044 -1.46712690
#> 4      S03     I01  5.6263460246 3.018906  0.1477006  1.0263826  1.43335727
#> 18     S03     I02  4.9082804482 3.018906  0.1477006  0.2941346  1.44753968
#> 23     S03     I03  1.9908304792 3.018906  0.1477006 -0.7345863 -0.44118938
#> 33     S03     I04  2.6185611126 3.018906  0.1477006  0.6676717 -1.21571678
#> 47     S03     I05  3.1786364469 3.018906  0.1477006  0.3318599 -0.31982963
#> 60     S03     I06  1.9640404448 3.018906  0.1477006 -0.3062988 -0.89626690
#> 62     S03     I07  3.6575944906 3.018906  0.1477006  0.3069992  0.18398905
#> 76     S03     I08  2.2030959547 3.018906  0.1477006 -0.8295020 -0.13400826
#> 89     S03     I09  2.5530806953 3.018906  0.1477006 -0.7193702  0.10584472
#> 95     S03     I10  3.8376110090 3.018906  0.1477006  0.2905044  0.38050045
#> 5      S04     I01  6.1555226823 3.018906  1.1208782  1.0263826  0.98935627
#> 19     S04     I02  5.2128689267 3.018906  1.1208782  0.2941346  0.77895050
#> 21     S04     I03  3.5459683375 3.018906  1.1208782 -0.7345863  0.14077083
#> 34     S04     I04  4.0685989208 3.018906  1.1208782  0.6676717 -0.73885662
#> 48     S04     I05  4.1272194761 3.018906  1.1208782  0.3318599 -0.34442425
#> 51     S04     I06  5.1312236369 3.018906  1.1208782 -0.3062988  1.29773863
#> 63     S04     I07  5.0508494287 3.018906  1.1208782  0.3069992  0.60406634
#> 77     S04     I08  5.0514461121 3.018906  1.1208782 -0.8295020  1.74116424
#> 90     S04     I09  3.3324463443 3.018906  1.1208782 -0.7193702 -0.08796728
#> 92     S04     I10  4.8382028480 3.018906  1.1208782  0.2905044  0.40791463
#> 2      S05     I01  5.0725435326 3.018906  0.8670641  1.0263826  0.16019124
#> 16     S05     I02  3.2183873381 3.018906  0.8670641  0.2941346 -0.96171697
#> 22     S05     I03  5.2763168828 3.018906  0.8670641 -0.7345863  2.12493349
#> 35     S05     I04  5.6507650260 3.018906  0.8670641  0.6676717  1.09712360
#> 45     S05     I05  5.2640229699 3.018906  0.8670641  0.3318599  1.04619336
#> 52     S05     I06  2.8472948949 3.018906  0.8670641 -0.3062988 -0.73237599
#> 64     S05     I07  5.3308969048 3.018906  0.8670641  0.3069992  1.13792793
#> 74     S05     I08  1.6187290108 3.018906  0.8670641 -0.8295020 -1.43773875
#> 81     S05     I09  4.1203064717 3.018906  0.8670641 -0.7193702  0.95370696
#> 93     S05     I10  3.7832061580 3.018906  0.8670641  0.2905044 -0.39326794
#> 8      S06     I01  3.1187444733 3.018906  0.5086388  1.0263826 -1.43518252
#> 20     S06     I02  5.0985762639 3.018906  0.5086388  0.2941346  1.27689726
#> 27     S06     I03  0.3395967847 3.018906  0.5086388 -0.7345863 -2.45336130
#> 37     S06     I04  5.2379965086 3.018906  0.5086388  0.6676717  1.04278039
#> 49     S06     I05  3.2045497622 3.018906  0.5086388  0.3318599 -0.65485454
#> 56     S06     I06  3.4337299932 3.018906  0.5086388 -0.3062988  0.21248441
#> 66     S06     I07  5.2900719915 3.018906  0.5086388  0.3069992  1.45552832
#> 79     S06     I08  3.9496169820 3.018906  0.5086388 -0.8295020  1.25157453
#> 85     S06     I09  2.1176594188 3.018906  0.5086388 -0.7193702 -0.69051479
#> 99     S06     I10  1.8055450309 3.018906  0.5086388  0.2905044 -2.01250376
#> 9      S07     I01  4.7565442755 3.018906  0.2733068  1.0263826  0.43794934
#> 11     S07     I02  2.3969917375 3.018906  0.2733068  0.2941346 -1.18935521
#> 24     S07     I03  2.3642225017 3.018906  0.2733068 -0.7345863 -0.19340353
#> 38     S07     I04  3.1900049746 3.018906  0.2733068  0.6676717 -0.76987909
#> 50     S07     I05  0.9247226494 3.018906  0.2733068  0.3318599 -2.69934960
#> 53     S07     I06  3.5179391160 3.018906  0.2733068 -0.3062988  0.53202559
#> 67     S07     I07  3.5148507549 3.018906  0.2733068  0.3069992 -0.08436086
#> 80     S07     I08  1.0561780365 3.018906  0.2733068 -0.8295020 -1.40653236
#> 82     S07     I09  2.0904104531 3.018906  0.2733068 -0.7193702 -0.48243170
#> 96     S07     I10  3.4084586735 3.018906  0.2733068  0.2905044 -0.17425806
#> 6      S08     I01  4.6769184993 3.018906 -0.9731389  1.0263826  1.60476923
#> 12     S08     I02  1.9308331899 3.018906 -0.9731389  0.2941346 -0.40906809
#> 25     S08     I03  0.7871535667 3.018906 -0.9731389 -0.7345863 -0.52402680
#> 39     S08     I04  3.6132230721 3.018906 -0.9731389  0.6676717  0.89978467
#> 41     S08     I05  0.6527052147 3.018906 -0.9731389  0.3318599 -1.72492137
#> 54     S08     I06  2.8229103376 3.018906 -0.9731389 -0.3062988  1.08344248
#> 68     S08     I07  2.4757979911 3.018906 -0.9731389  0.3069992  0.12303204
#> 78     S08     I08  3.2135689751 3.018906 -0.9731389 -0.8295020  1.99730424
#> 83     S08     I09  0.9363220462 3.018906 -0.9731389 -0.7193702 -0.39007444
#> 97     S08     I10  3.5409692270 3.018906 -0.9731389  0.2905044  1.20469815
#> 3      S09     I01  5.2447785809 3.018906  0.9019945  1.0263826  0.29749593
#> 13     S09     I02  3.2722166632 3.018906  0.9019945  0.2941346 -0.94281800
#> 26     S09     I03  3.9432165186 3.018906  0.9019945 -0.7345863  0.75690277
#> 32     S09     I04  5.2398726527 3.018906  0.9019945  0.6676717  0.65130087
#> 42     S09     I05  3.9004162710 3.018906  0.9019945  0.3318599 -0.35234369
#> 55     S09     I06  3.1024238732 3.018906  0.9019945 -0.3062988 -0.51217737
#> 65     S09     I07  3.9047345483 3.018906  0.9019945  0.3069992 -0.32316478
#> 71     S09     I08  2.5707832704 3.018906  0.9019945 -0.8295020 -0.52061484
#> 84     S09     I09  4.8035410620 3.018906  0.9019945 -0.7193702  1.60201120
#> 94     S09     I10  5.1638374633 3.018906  0.9019945  0.2905044  0.95243301
#> 10     S10     I01  2.8469030990 3.018906  0.3028606  1.0263826 -1.50124571
#> 14     S10     I02  3.6806264138 3.018906  0.3028606  0.2941346  0.06472559
#> 28     S10     I03  3.0845557980 3.018906  0.3028606 -0.7345863  0.49737589
#> 40     S10     I04  3.1459167525 3.018906  0.3028606  0.6676717 -0.84352119
#> 43     S10     I05  4.4655991405 3.018906  0.3028606  0.3318599  0.81197301
#> 57     S10     I06  3.4284093961 3.018906  0.3028606 -0.3062988  0.41294199
#> 70     S10     I07  4.3023891250 3.018906  0.3028606  0.3069992  0.67362363
#> 72     S10     I08  3.6784962246 3.018906  0.3028606 -0.8295020  1.18623195
#> 86     S10     I09  1.5195247655 3.018906  0.3028606 -0.7193702 -1.08287126
#> 100    S10     I10  3.8358417281 3.018906  0.3028606  0.2905044  0.22357111