Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id          y  grand_i       sub_i      item_i          err
#> 1      S01     I01  5.4160880 3.018906  1.67195078 -0.34445779  1.069689347
#> 15     S01     I02  1.3265202 3.018906  1.67195078 -0.57623686 -2.788099298
#> 29     S01     I03  4.0723932 3.018906  1.67195078 -1.26414572  0.645682548
#> 31     S01     I04  2.6066250 3.018906  1.67195078 -0.03848531 -2.045746076
#> 44     S01     I05  5.0128546 3.018906  1.67195078 -1.24200121  1.563999379
#> 58     S01     I06  5.8120946 3.018906  1.67195078  0.26557544  0.855662807
#> 61     S01     I07  5.3076237 3.018906  1.67195078 -0.07185764  0.688624934
#> 73     S01     I08  3.3132011 3.018906  1.67195078 -1.12640626 -0.251249025
#> 87     S01     I09  6.1060449 3.018906  1.67195078  0.95735857  0.457829950
#> 91     S01     I10  6.0976367 3.018906  1.67195078 -0.61412187  2.020902129
#> 7      S02     I01  1.9745403 3.018906 -0.12307504 -0.34445779 -0.576832448
#> 17     S02     I02  2.7216505 3.018906 -0.12307504 -0.57623686  0.402056802
#> 30     S02     I03  2.5743716 3.018906 -0.12307504 -1.26414572  0.942686753
#> 36     S02     I04  2.2477749 3.018906 -0.12307504 -0.03848531 -0.609570308
#> 46     S02     I05  1.6467903 3.018906 -0.12307504 -1.24200121 -0.007039054
#> 59     S02     I06  4.0626058 3.018906 -0.12307504  0.26557544  0.901199819
#> 69     S02     I07  5.6518440 3.018906 -0.12307504 -0.07185764  2.827871094
#> 75     S02     I08  3.7286530 3.018906 -0.12307504 -1.12640626  1.959228726
#> 88     S02     I09  5.3903269 3.018906 -0.12307504  0.95735857  1.537137799
#> 98     S02     I10  1.2457852 3.018906 -0.12307504 -0.61412187 -1.035923495
#> 4      S03     I01  3.1451082 3.018906  0.02619833 -0.34445779  0.444462057
#> 18     S03     I02  1.5862689 3.018906  0.02619833 -0.57623686 -0.882598170
#> 23     S03     I03  2.1999238 3.018906  0.02619833 -1.26414572  0.418965586
#> 33     S03     I04  3.4930515 3.018906  0.02619833 -0.03848531  0.486432828
#> 47     S03     I05  2.5420155 3.018906  0.02619833 -1.24200121  0.738912747
#> 60     S03     I06  3.4628352 3.018906  0.02619833  0.26557544  0.152155781
#> 62     S03     I07  2.9833241 3.018906  0.02619833 -0.07185764  0.010077825
#> 76     S03     I08  2.7737547 3.018906  0.02619833 -1.12640626  0.855057058
#> 89     S03     I09  5.1807967 3.018906  0.02619833  0.95735857  1.178334193
#> 95     S03     I10  1.1539116 3.018906  0.02619833 -0.61412187 -1.277070481
#> 5      S04     I01  1.5057186 3.018906 -0.72189888 -0.34445779 -0.446830336
#> 19     S04     I02  1.9131367 3.018906 -0.72189888 -0.57623686  0.192366858
#> 21     S04     I03  1.0171734 3.018906 -0.72189888 -1.26414572 -0.015687589
#> 34     S04     I04  2.1294179 3.018906 -0.72189888 -0.03848531 -0.129103540
#> 48     S04     I05 -0.8306559 3.018906 -0.72189888 -1.24200121 -1.885661452
#> 51     S04     I06  2.7284624 3.018906 -0.72189888  0.26557544  0.165880208
#> 63     S04     I07  1.2189978 3.018906 -0.72189888 -0.07185764 -1.006151268
#> 77     S04     I08  3.5132829 3.018906 -0.72189888 -1.12640626  2.342682441
#> 90     S04     I09  3.1824514 3.018906 -0.72189888  0.95735857 -0.071913930
#> 92     S04     I10  0.8092712 3.018906 -0.72189888 -0.61412187 -0.873613679
#> 2      S05     I01  1.1276839 3.018906  0.21726244 -0.34445779 -1.764026333
#> 16     S05     I02  3.3305305 3.018906  0.21726244 -0.57623686  0.670599319
#> 22     S05     I03  2.3502125 3.018906  0.21726244 -1.26414572  0.378190187
#> 35     S05     I04  3.6567581 3.018906  0.21726244 -0.03848531  0.459075383
#> 45     S05     I05  2.9853304 3.018906  0.21726244 -1.24200121  0.991163581
#> 52     S05     I06  4.6690300 3.018906  0.21726244  0.26557544  1.167286481
#> 64     S05     I07  3.6142290 3.018906  0.21726244 -0.07185764  0.449918570
#> 74     S05     I08  1.4038982 3.018906  0.21726244 -1.12640626 -0.705863640
#> 81     S05     I09  1.6823701 3.018906  0.21726244  0.95735857 -2.511156551
#> 93     S05     I10  4.1144597 3.018906  0.21726244 -0.61412187  1.492413564
#> 8      S06     I01  4.4244990 3.018906  0.35888763 -0.34445779  1.391163505
#> 20     S06     I02  4.8321552 3.018906  0.35888763 -0.57623686  2.030598811
#> 27     S06     I03  2.6464247 3.018906  0.35888763 -1.26414572  0.532777182
#> 37     S06     I04  2.1309189 3.018906  0.35888763 -0.03848531 -1.208389036
#> 49     S06     I05  2.1971096 3.018906  0.35888763 -1.24200121  0.061317609
#> 56     S06     I06  2.1938310 3.018906  0.35888763  0.26557544 -1.449537682
#> 66     S06     I07  4.9678878 3.018906  0.35888763 -0.07185764  1.661952182
#> 79     S06     I08  2.0170179 3.018906  0.35888763 -1.12640626 -0.234369069
#> 85     S06     I09  5.5506782 3.018906  0.35888763  0.95735857  1.215526409
#> 99     S06     I10  2.4831103 3.018906  0.35888763 -0.61412187 -0.280561095
#> 9      S07     I01  3.2133371 3.018906  0.35263346 -0.34445779  0.186255828
#> 11     S07     I02  2.8629979 3.018906  0.35263346 -0.57623686  0.067695680
#> 24     S07     I03  1.8953601 3.018906  0.35263346 -1.26414572 -0.212033218
#> 38     S07     I04  3.4765820 3.018906  0.35263346 -0.03848531  0.143528276
#> 50     S07     I05  0.3587327 3.018906  0.35263346 -1.24200121 -1.770805195
#> 53     S07     I06  3.7732270 3.018906  0.35263346  0.26557544  0.136112501
#> 67     S07     I07  5.0819469 3.018906  0.35263346 -0.07185764  1.782265458
#> 80     S07     I08  1.3159275 3.018906  0.35263346 -1.12640626 -0.929205354
#> 82     S07     I09  4.4261611 3.018906  0.35263346  0.95735857  0.097263470
#> 96     S07     I10  2.9098584 3.018906  0.35263346 -0.61412187  0.152441215
#> 6      S08     I01  3.8323192 3.018906 -0.70633316 -0.34445779  1.864204562
#> 12     S08     I02  2.4879523 3.018906 -0.70633316 -0.57623686  0.751616719
#> 25     S08     I03  3.0949539 3.018906 -0.70633316 -1.26414572  2.046527169
#> 39     S08     I04  2.9010604 3.018906 -0.70633316 -0.03848531  0.626973215
#> 41     S08     I05  0.1981297 3.018906 -0.70633316 -1.24200121 -0.872441542
#> 54     S08     I06  3.3828574 3.018906 -0.70633316  0.26557544  0.804709468
#> 68     S08     I07  0.4680403 3.018906 -0.70633316 -0.07185764 -1.772674533
#> 78     S08     I08  1.2912209 3.018906 -0.70633316 -1.12640626  0.105054659
#> 83     S08     I09  2.9415138 3.018906 -0.70633316  0.95735857 -0.328417180
#> 97     S08     I10 -0.4235129 3.018906 -0.70633316 -0.61412187 -2.121963532
#> 3      S09     I01  3.2696906 3.018906  0.97458062 -0.34445779 -0.379337839
#> 13     S09     I02  4.7497953 3.018906  0.97458062 -0.57623686  1.332545910
#> 26     S09     I03  3.9435035 3.018906  0.97458062 -1.26414572  1.214162953
#> 32     S09     I04  5.6807924 3.018906  0.97458062 -0.03848531  1.725791443
#> 42     S09     I05  3.2885406 3.018906  0.97458062 -1.24200121  0.537055588
#> 55     S09     I06  3.8099870 3.018906  0.97458062  0.26557544 -0.449074628
#> 65     S09     I07  5.0864202 3.018906  0.97458062 -0.07185764  1.164791599
#> 71     S09     I08  3.3097543 3.018906  0.97458062 -1.12640626  0.442674361
#> 84     S09     I09  4.1845325 3.018906  0.97458062  0.95735857 -0.766312272
#> 94     S09     I10  3.2282923 3.018906  0.97458062 -0.61412187 -0.151072070
#> 10     S10     I01  2.9304900 3.018906 -1.42538574 -0.34445779  1.681427893
#> 14     S10     I02  2.4549980 3.018906 -1.42538574 -0.57623686  1.437715030
#> 28     S10     I03 -0.1976979 3.018906 -1.42538574 -1.26414572 -0.527072071
#> 40     S10     I04  0.8733709 3.018906 -1.42538574 -0.03848531 -0.681663694
#> 43     S10     I05  1.4753131 3.018906 -1.42538574 -1.24200121  1.123794415
#> 57     S10     I06  3.0357359 3.018906 -1.42538574  0.26557544  1.176640547
#> 70     S10     I07  3.0052639 3.018906 -1.42538574 -0.07185764  1.483601631
#> 72     S10     I08 -2.1055708 3.018906 -1.42538574 -1.12640626 -2.572684455
#> 86     S10     I09  2.5336864 3.018906 -1.42538574  0.95735857 -0.017192075
#> 100    S10     I10  2.1809097 3.018906 -1.42538574 -0.61412187  1.201511675