Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

Usage

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id          y  grand_i        sub_i     item_i          err
#> 1      S01     I01  5.8705090 3.018906  1.157556991  0.5524068  1.141639591
#> 15     S01     I02  2.4799882 3.018906  1.157556991 -1.1857130 -0.510761341
#> 29     S01     I03  2.9520550 3.018906  1.157556991 -1.1521728 -0.072234831
#> 31     S01     I04  6.1595280 3.018906  1.157556991  0.3174069  1.665658576
#> 44     S01     I05  4.1188480 3.018906  1.157556991 -0.8475896  0.789975038
#> 58     S01     I06  3.3842406 3.018906  1.157556991 -0.8684548  0.076232789
#> 61     S01     I07  4.2988416 3.018906  1.157556991  0.1363663 -0.013987295
#> 73     S01     I08  3.4893811 3.018906  1.157556991 -0.4399390 -0.247142467
#> 87     S01     I09  4.2856218 3.018906  1.157556991  1.2586441 -1.149484924
#> 91     S01     I10  5.1901401 3.018906  1.157556991  0.9073945  0.106283014
#> 7      S02     I01  1.8283435 3.018906 -0.516397190  0.5524068 -1.226571754
#> 17     S02     I02 -0.3661918 3.018906 -0.516397190 -1.1857130 -1.682987209
#> 30     S02     I03  1.6230033 3.018906 -0.516397190 -1.1521728  0.272667677
#> 36     S02     I04  2.8384243 3.018906 -0.516397190  0.3174069  0.018508977
#> 46     S02     I05  2.7482028 3.018906 -0.516397190 -0.8475896  1.093283968
#> 59     S02     I06  1.5971279 3.018906 -0.516397190 -0.8684548 -0.036925720
#> 69     S02     I07  2.2287113 3.018906 -0.516397190  0.1363663 -0.410163449
#> 75     S02     I08  2.3134564 3.018906 -0.516397190 -0.4399390  0.250887001
#> 88     S02     I09  4.6864241 3.018906 -0.516397190  1.2586441  0.925271591
#> 98     S02     I10  3.2060479 3.018906 -0.516397190  0.9073945 -0.203854996
#> 4      S03     I01  3.0366232 3.018906 -1.455322491  0.5524068  0.920633187
#> 18     S03     I02 -0.3674664 3.018906 -1.455322491 -1.1857130 -0.745336471
#> 23     S03     I03 -1.4741634 3.018906 -1.455322491 -1.1521728 -1.885573727
#> 33     S03     I04  2.7128198 3.018906 -1.455322491  0.3174069  0.831829838
#> 47     S03     I05  1.1579550 3.018906 -1.455322491 -0.8475896  0.441961542
#> 60     S03     I06  0.4493406 3.018906 -1.455322491 -0.8684548 -0.245787740
#> 62     S03     I07  0.1331160 3.018906 -1.455322491  0.1363663 -1.566833428
#> 76     S03     I08  4.0009389 3.018906 -1.455322491 -0.4399390  2.877294841
#> 89     S03     I09  3.4567185 3.018906 -1.455322491  1.2586441  0.634491297
#> 95     S03     I10  2.7995042 3.018906 -1.455322491  0.9073945  0.328526593
#> 5      S04     I01  7.4426675 3.018906  0.009725293  0.5524068  3.861629777
#> 19     S04     I02  2.6825545 3.018906  0.009725293 -1.1857130  0.839636597
#> 21     S04     I03  0.8719674 3.018906  0.009725293 -1.1521728 -1.004490744
#> 34     S04     I04  4.6273472 3.018906  0.009725293  0.3174069  1.281309396
#> 48     S04     I05  1.9672711 3.018906  0.009725293 -0.8475896 -0.213770153
#> 51     S04     I06  1.7172197 3.018906  0.009725293 -0.8684548 -0.442956376
#> 63     S04     I07  2.0395297 3.018906  0.009725293  0.1363663 -1.125467568
#> 77     S04     I08  0.9059727 3.018906  0.009725293 -0.4399390 -1.682719112
#> 90     S04     I09  4.5352373 3.018906  0.009725293  1.2586441  0.247962237
#> 92     S04     I10  3.8473116 3.018906  0.009725293  0.9073945 -0.088713797
#> 2      S05     I01  5.5299908 3.018906  0.367091477  0.5524068  1.591586876
#> 16     S05     I02  3.1098332 3.018906  0.367091477 -1.1857130  0.909549193
#> 22     S05     I03  1.4206604 3.018906  0.367091477 -1.1521728 -0.813163906
#> 35     S05     I04  3.6235920 3.018906  0.367091477  0.3174069 -0.079811963
#> 45     S05     I05  4.5703297 3.018906  0.367091477 -0.8475896  2.031922253
#> 52     S05     I06  3.7356314 3.018906  0.367091477 -0.8684548  1.218089148
#> 64     S05     I07  1.7493992 3.018906  0.367091477  0.1363663 -1.772964222
#> 74     S05     I08  2.9130510 3.018906  0.367091477 -0.4399390 -0.033007074
#> 81     S05     I09  3.1273784 3.018906  0.367091477  1.2586441 -1.517262823
#> 93     S05     I10  3.4706670 3.018906  0.367091477  0.9073945 -0.822724526
#> 8      S06     I01  5.4228159 3.018906  0.952160743  0.5524068  0.899342678
#> 20     S06     I02  3.4949617 3.018906  0.952160743 -1.1857130  0.709608361
#> 27     S06     I03  4.8500239 3.018906  0.952160743 -1.1521728  2.031130361
#> 37     S06     I04  4.9805727 3.018906  0.952160743  0.3174069  0.692099523
#> 49     S06     I05  2.0579839 3.018906  0.952160743 -0.8475896 -1.065492854
#> 56     S06     I06  4.1442431 3.018906  0.952160743 -0.8684548  1.041631581
#> 66     S06     I07  2.1899729 3.018906  0.952160743  0.1363663 -1.917459828
#> 79     S06     I08  3.5401710 3.018906  0.952160743 -0.4399390  0.009043665
#> 85     S06     I09  5.6463757 3.018906  0.952160743  1.2586441  0.416665219
#> 99     S06     I10  5.3461299 3.018906  0.952160743  0.9073945  0.467669123
#> 9      S07     I01  3.9097574 3.018906 -0.401338042  0.5524068  0.739782965
#> 11     S07     I02  2.9453618 3.018906 -0.401338042 -1.1857130  1.513507285
#> 24     S07     I03  1.2077257 3.018906 -0.401338042 -1.1521728 -0.257669041
#> 38     S07     I04  3.8404993 3.018906 -0.401338042  0.3174069  0.905524908
#> 50     S07     I05  1.4873728 3.018906 -0.401338042 -0.8475896 -0.282605170
#> 53     S07     I06  2.2279268 3.018906 -0.401338042 -0.8684548  0.478814035
#> 67     S07     I07  2.1279559 3.018906 -0.401338042  0.1363663 -0.625978034
#> 80     S07     I08  1.6054858 3.018906 -0.401338042 -0.4399390 -0.572142706
#> 82     S07     I09  5.9752414 3.018906 -0.401338042  1.2586441  2.099029743
#> 96     S07     I10  2.2311517 3.018906 -0.401338042  0.9073945 -1.293810387
#> 6      S08     I01  3.2515372 3.018906  0.702682746  0.5524068 -1.022458044
#> 12     S08     I02  3.7402885 3.018906  0.702682746 -1.1857130  1.204413216
#> 25     S08     I03  3.2385663 3.018906  0.702682746 -1.1521728  0.669150778
#> 39     S08     I04  5.7522285 3.018906  0.702682746  0.3174069  1.713233308
#> 41     S08     I05  3.9797102 3.018906  0.702682746 -0.8475896  1.105711474
#> 54     S08     I06  3.2946077 3.018906  0.702682746 -0.8684548  0.441474190
#> 68     S08     I07  1.1532217 3.018906  0.702682746  0.1363663 -2.704733037
#> 78     S08     I08  2.3104917 3.018906  0.702682746 -0.4399390 -0.971157598
#> 83     S08     I09  7.1425032 3.018906  0.702682746  1.2586441  2.162270768
#> 97     S08     I10  5.6841290 3.018906  0.702682746  0.9073945  1.055146212
#> 3      S09     I01  2.3513821 3.018906  0.911857315  0.5524068 -2.131787631
#> 13     S09     I02  2.7197759 3.018906  0.911857315 -1.1857130 -0.025273958
#> 26     S09     I03  3.0126423 3.018906  0.911857315 -1.1521728  0.234052203
#> 32     S09     I04  2.9122027 3.018906  0.911857315  0.3174069 -1.335967126
#> 42     S09     I05  3.2620266 3.018906  0.911857315 -0.8475896  0.178853291
#> 55     S09     I06  4.1561734 3.018906  0.911857315 -0.8684548  1.093865335
#> 65     S09     I07  3.8120971 3.018906  0.911857315  0.1363663 -0.255032155
#> 71     S09     I08  2.6876547 3.018906  0.911857315 -0.4399390 -0.803169139
#> 84     S09     I09  4.1678197 3.018906  0.911857315  1.2586441 -1.021587305
#> 94     S09     I10  4.0816714 3.018906  0.911857315  0.9073945 -0.756485997
#> 10     S10     I01  3.4761689 3.018906 -0.009847584  0.5524068 -0.085295956
#> 14     S10     I02  2.6857394 3.018906 -0.009847584 -1.1857130  0.862394370
#> 28     S10     I03  2.2583047 3.018906 -0.009847584 -1.1521728  0.401419517
#> 40     S10     I04  3.8639121 3.018906 -0.009847584  0.3174069  0.537447185
#> 43     S10     I05  2.9547812 3.018906 -0.009847584 -0.8475896  0.793312753
#> 57     S10     I06  2.9005605 3.018906 -0.009847584 -0.8684548  0.759957311
#> 70     S10     I07  4.4972966 3.018906 -0.009847584  0.1363663  1.351872241
#> 72     S10     I08  2.5608134 3.018906 -0.009847584 -0.4399390 -0.008305563
#> 86     S10     I09  3.6957436 3.018906 -0.009847584  1.2586441 -0.571958506
#> 100    S10     I10  4.6951100 3.018906 -0.009847584  0.9073945  0.778657527