Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id           y  grand_i       sub_i     item_i          err
#> 1      S01     I01  3.50492086 3.018906  0.06334694 -0.8073076  1.229975891
#> 15     S01     I02  4.67338245 3.018906  0.06334694  0.9577043  0.633425558
#> 29     S01     I03  2.06609153 3.018906  0.06334694 -0.7794740 -0.236687044
#> 31     S01     I04  1.89221997 3.018906  0.06334694 -0.7740645 -0.415968101
#> 44     S01     I05  2.29630700 3.018906  0.06334694 -0.9147330  0.128787496
#> 58     S01     I06  3.36037336 3.018906  0.06334694 -1.2683194  1.546440201
#> 61     S01     I07  1.66827629 3.018906  0.06334694 -0.8139721 -0.600004184
#> 73     S01     I08  3.26792062 3.018906  0.06334694 -0.2172159  0.402883925
#> 87     S01     I09  3.50613049 3.018906  0.06334694 -0.8640981  1.287976059
#> 91     S01     I10  3.53230494 3.018906  0.06334694  0.5679891 -0.117936751
#> 7      S02     I01  0.55395250 3.018906 -0.47728978 -0.8073076 -1.180355751
#> 17     S02     I02  4.20625062 3.018906 -0.47728978  0.9577043  0.706930444
#> 30     S02     I03  1.20275612 3.018906 -0.47728978 -0.7794740 -0.559385739
#> 36     S02     I04 -0.82951271 3.018906 -0.47728978 -0.7740645 -2.597064063
#> 46     S02     I05  2.36004505 3.018906 -0.47728978 -0.9147330  0.733162255
#> 59     S02     I06 -1.28702574 3.018906 -0.47728978 -1.2683194 -2.560322180
#> 69     S02     I07  1.38526302 3.018906 -0.47728978 -0.8139721 -0.342380738
#> 75     S02     I08  3.02926513 3.018906 -0.47728978 -0.2172159  0.704865153
#> 88     S02     I09  1.54488775 3.018906 -0.47728978 -0.8640981 -0.132629959
#> 98     S02     I10  3.07227683 3.018906 -0.47728978  0.5679891 -0.037328140
#> 4      S03     I01  4.08986262 3.018906  0.71390697 -0.8073076  1.164357618
#> 18     S03     I02  4.06394514 3.018906  0.71390697  0.9577043 -0.626571789
#> 23     S03     I03  3.01874332 3.018906  0.71390697 -0.7794740  0.065404709
#> 33     S03     I04  3.21815738 3.018906  0.71390697 -0.7740645  0.259409282
#> 47     S03     I05  2.34625264 3.018906  0.71390697 -0.9147330 -0.471826908
#> 60     S03     I06  2.69151484 3.018906  0.71390697 -1.2683194  0.227021653
#> 62     S03     I07  3.73336607 3.018906  0.71390697 -0.8139721  0.814525565
#> 76     S03     I08  4.04284263 3.018906  0.71390697 -0.2172159  0.527245897
#> 89     S03     I09  3.74472603 3.018906  0.71390697 -0.8640981  0.876011561
#> 95     S03     I10  3.77639174 3.018906  0.71390697  0.5679891 -0.524409985
#> 5      S04     I01  4.22192144 3.018906  0.19360452 -0.8073076  1.816718893
#> 19     S04     I02  4.55624064 3.018906  0.19360452  0.9577043  0.386026170
#> 21     S04     I03  3.18641436 3.018906  0.19360452 -0.7794740  0.753378207
#> 34     S04     I04  3.91026089 3.018906  0.19360452 -0.7740645  1.471815247
#> 48     S04     I05  1.98581166 3.018906  0.19360452 -0.9147330 -0.311965430
#> 51     S04     I06  2.29990318 3.018906  0.19360452 -1.2683194  0.355712440
#> 63     S04     I07  0.62377976 3.018906  0.19360452 -0.8139721 -1.774758294
#> 77     S04     I08  3.69108753 3.018906  0.19360452 -0.2172159  0.695793263
#> 90     S04     I09  1.74474205 3.018906  0.19360452 -0.8640981 -0.603669955
#> 92     S04     I10  2.92831779 3.018906  0.19360452  0.5679891 -0.852181483
#> 2      S05     I01  2.05706735 3.018906 -0.20250459 -0.8073076  0.047973911
#> 16     S05     I02  5.74098539 3.018906 -0.20250459  0.9577043  1.966880018
#> 22     S05     I03  2.82213484 3.018906 -0.20250459 -0.7794740  0.785207784
#> 35     S05     I04  2.33591113 3.018906 -0.20250459 -0.7740645  0.293574583
#> 45     S05     I05  0.21749716 3.018906 -0.20250459 -0.9147330 -1.684170823
#> 52     S05     I06  1.94537523 3.018906 -0.20250459 -1.2683194  0.397293595
#> 64     S05     I07  0.08592144 3.018906 -0.20250459 -0.8139721 -1.916507512
#> 74     S05     I08  2.42585649 3.018906 -0.20250459 -0.2172159 -0.173328675
#> 81     S05     I09  1.15895531 3.018906 -0.20250459 -0.8640981 -0.793347601
#> 93     S05     I10  4.00930333 3.018906 -0.20250459  0.5679891  0.624913165
#> 8      S06     I01  1.38429069 3.018906 -0.18676344 -0.8073076 -0.640543889
#> 20     S06     I02  3.77948608 3.018906 -0.18676344  0.9577043 -0.010360434
#> 27     S06     I03  2.60096489 3.018906 -0.18676344 -0.7794740  0.548296694
#> 37     S06     I04  2.10293267 3.018906 -0.18676344 -0.7740645  0.044854984
#> 49     S06     I05  1.70411743 3.018906 -0.18676344 -0.9147330 -0.213291696
#> 56     S06     I06  2.13587191 3.018906 -0.18676344 -1.2683194  0.572049132
#> 66     S06     I07  1.89974832 3.018906 -0.18676344 -0.8139721 -0.118421773
#> 79     S06     I08  3.76018506 3.018906 -0.18676344 -0.2172159  1.145258750
#> 85     S06     I09  4.16138257 3.018906 -0.18676344 -0.8640981  2.193338526
#> 99     S06     I10  3.90094113 3.018906 -0.18676344  0.5679891  0.500809822
#> 9      S07     I01  0.83504432 3.018906 -0.69613448 -0.8073076 -0.680419225
#> 11     S07     I02  4.90569031 3.018906 -0.69613448  0.9577043  1.625214836
#> 24     S07     I03  1.80759204 3.018906 -0.69613448 -0.7794740  0.264294876
#> 38     S07     I04  1.15839420 3.018906 -0.69613448 -0.7740645 -0.390312446
#> 50     S07     I05  0.67798106 3.018906 -0.69613448 -0.9147330 -0.730057028
#> 53     S07     I06  2.22630397 3.018906 -0.69613448 -1.2683194  1.171852227
#> 67     S07     I07  0.66184235 3.018906 -0.69613448 -0.8139721 -0.846956708
#> 80     S07     I08  0.04976664 3.018906 -0.69613448 -0.2172159 -2.055788632
#> 82     S07     I09  2.17427164 3.018906 -0.69613448 -0.8640981  0.715598623
#> 96     S07     I10  2.72515402 3.018906 -0.69613448  0.5679891 -0.165606256
#> 6      S08     I01  1.33785227 3.018906  0.28551278 -0.8073076 -1.159258541
#> 12     S08     I02  3.64896904 3.018906  0.28551278  0.9577043 -0.613153700
#> 25     S08     I03  4.45352014 3.018906  0.28551278 -0.7794740  1.928575715
#> 39     S08     I04  3.72304983 3.018906  0.28551278 -0.7740645  1.192695918
#> 41     S08     I05  5.04870375 3.018906  0.28551278 -0.9147330  2.659018395
#> 54     S08     I06  2.11577483 3.018906  0.28551278 -1.2683194  0.079675825
#> 68     S08     I07  4.52953481 3.018906  0.28551278 -0.8139721  2.039088493
#> 78     S08     I08  3.07851348 3.018906  0.28551278 -0.2172159 -0.008689057
#> 83     S08     I09  2.90681383 3.018906  0.28551278 -0.8640981  0.466493556
#> 97     S08     I10  6.47722027 3.018906  0.28551278  0.5679891  2.604812729
#> 3      S09     I01  3.38318782 3.018906 -0.06244629 -0.8073076  1.234036079
#> 13     S09     I02  2.60104483 3.018906 -0.06244629  0.9577043 -1.313118833
#> 26     S09     I03  3.74022576 3.018906 -0.06244629 -0.7794740  1.563240408
#> 32     S09     I04  1.74553195 3.018906 -0.06244629 -0.7740645 -0.436862893
#> 42     S09     I05  1.25159442 3.018906 -0.06244629 -0.9147330 -0.790131860
#> 55     S09     I06  2.96542148 3.018906 -0.06244629 -1.2683194  1.277281550
#> 65     S09     I07  0.91819962 3.018906 -0.06244629 -0.8139721 -1.224287625
#> 71     S09     I08  0.77187161 3.018906 -0.06244629 -0.2172159 -1.967371855
#> 84     S09     I09  1.37235106 3.018906 -0.06244629 -0.8640981 -0.720010142
#> 94     S09     I10  3.42659763 3.018906 -0.06244629  0.5679891 -0.097850839
#> 10     S10     I01  4.50754643 3.018906  1.78220512 -0.8073076  0.513743287
#> 14     S10     I02  5.85793863 3.018906  1.78220512  0.9577043  0.099123562
#> 28     S10     I03  6.64551333 3.018906  1.78220512 -0.7794740  2.623876570
#> 40     S10     I04  4.63033655 3.018906  1.78220512 -0.7740645  0.603290301
#> 43     S10     I05  3.42694259 3.018906  1.78220512 -0.9147330 -0.459435102
#> 57     S10     I06  3.30383750 3.018906  1.78220512 -1.2683194 -0.228953841
#> 70     S10     I07  5.84489894 3.018906  1.78220512 -0.8139721  1.857760290
#> 72     S10     I08  5.20518733 3.018906  1.78220512 -0.2172159  0.621292455
#> 86     S10     I09  3.63613157 3.018906  1.78220512 -0.8640981 -0.300881036
#> 100    S10     I10  6.76756021 3.018906  1.78220512  0.5679891  1.398460340