Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

Usage

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id           y  grand_i      sub_i      item_i         err
#> 1      S01     I01  1.08706391 3.018906 -0.5984481 -1.35431210  0.02091849
#> 15     S01     I02  1.36565175 3.018906 -0.5984481 -0.10926927 -0.94553651
#> 29     S01     I03  0.37082497 3.018906 -0.5984481 -2.32224648  0.27261392
#> 31     S01     I04  2.48973221 3.018906 -0.5984481 -0.30730849  0.37658318
#> 44     S01     I05  2.41732990 3.018906 -0.5984481  0.01253761 -0.01566523
#> 58     S01     I06  3.18923840 3.018906 -0.5984481  0.08815511  0.68062578
#> 61     S01     I07  3.18712544 3.018906 -0.5984481 -0.47142095  1.23808888
#> 73     S01     I08  3.44894457 3.018906 -0.5984481  0.78598369  0.24250336
#> 87     S01     I09  4.15039637 3.018906 -0.5984481 -0.21917677  1.94911563
#> 91     S01     I10  2.30206424 3.018906 -0.5984481  0.10429483 -0.22268811
#> 7      S02     I01  0.37782846 3.018906 -0.8359018 -1.35431210 -0.45086320
#> 17     S02     I02  2.51272109 3.018906 -0.8359018 -0.10926927  0.43898659
#> 30     S02     I03 -0.23621530 3.018906 -0.8359018 -2.32224648 -0.09697260
#> 36     S02     I04  1.24021882 3.018906 -0.8359018 -0.30730849 -0.63547646
#> 46     S02     I05  1.16140843 3.018906 -0.8359018  0.01253761 -1.03413294
#> 59     S02     I06  2.48113548 3.018906 -0.8359018  0.08815511  0.20997661
#> 69     S02     I07  1.58384153 3.018906 -0.8359018 -0.47142095 -0.12774129
#> 75     S02     I08  2.57910050 3.018906 -0.8359018  0.78598369 -0.38988696
#> 88     S02     I09  2.81736755 3.018906 -0.8359018 -0.21917677  0.85354056
#> 98     S02     I10  1.21170660 3.018906 -0.8359018  0.10429483 -1.07559200
#> 4      S03     I01  0.73174842 3.018906 -0.4271029 -1.35431210 -0.50574222
#> 18     S03     I02  3.16937939 3.018906 -0.4271029 -0.10926927  0.68684591
#> 23     S03     I03  0.11344576 3.018906 -0.4271029 -2.32224648 -0.15611051
#> 33     S03     I04  1.98496720 3.018906 -0.4271029 -0.30730849 -0.29952706
#> 47     S03     I05  5.35072243 3.018906 -0.4271029  0.01253761  2.74638208
#> 60     S03     I06  1.42792438 3.018906 -0.4271029  0.08815511 -1.25203348
#> 62     S03     I07  2.37999723 3.018906 -0.4271029 -0.47142095  0.25961544
#> 76     S03     I08  4.33635173 3.018906 -0.4271029  0.78598369  0.95856529
#> 89     S03     I09  3.15123457 3.018906 -0.4271029 -0.21917677  0.77860860
#> 95     S03     I10  2.63995688 3.018906 -0.4271029  0.10429483 -0.05614070
#> 5      S04     I01  2.31280378 3.018906 -0.5344649 -1.35431210  1.18267517
#> 19     S04     I02  2.41721191 3.018906 -0.5344649 -0.10926927  0.04204046
#> 21     S04     I03  0.02338423 3.018906 -0.5344649 -2.32224648 -0.13881001
#> 34     S04     I04  1.82913686 3.018906 -0.5344649 -0.30730849 -0.34799537
#> 48     S04     I05  1.06102032 3.018906 -0.5344649  0.01253761 -1.43595800
#> 51     S04     I06  2.54407569 3.018906 -0.5344649  0.08815511 -0.02852013
#> 63     S04     I07  1.83389992 3.018906 -0.5344649 -0.47142095 -0.17911985
#> 77     S04     I08  3.05262386 3.018906 -0.5344649  0.78598369 -0.21780055
#> 90     S04     I09  1.82941587 3.018906 -0.5344649 -0.21917677 -0.43584807
#> 92     S04     I10  3.12494382 3.018906 -0.5344649  0.10429483  0.53620828
#> 2      S05     I01  3.31302423 3.018906  0.8972961 -1.35431210  0.75113462
#> 16     S05     I02  4.26518807 3.018906  0.8972961 -0.10926927  0.45825563
#> 22     S05     I03  1.53821351 3.018906  0.8972961 -2.32224648 -0.05574172
#> 35     S05     I04  2.74779897 3.018906  0.8972961 -0.30730849 -0.86109425
#> 45     S05     I05  2.76795380 3.018906  0.8972961  0.01253761 -1.16078552
#> 52     S05     I06  2.70961607 3.018906  0.8972961  0.08815511 -1.29474074
#> 64     S05     I07  3.78376397 3.018906  0.8972961 -0.47142095  0.33898321
#> 74     S05     I08  4.68436297 3.018906  0.8972961  0.78598369 -0.01782244
#> 81     S05     I09  5.98591518 3.018906  0.8972961 -0.21917677  2.28889024
#> 93     S05     I10  6.65532688 3.018906  0.8972961  0.10429483  2.63483034
#> 8      S06     I01  3.67777298 3.018906 -0.2831109 -1.35431210  2.29629042
#> 20     S06     I02  2.52728665 3.018906 -0.2831109 -0.10926927 -0.09923875
#> 27     S06     I03  1.25998643 3.018906 -0.2831109 -2.32224648  0.84643824
#> 37     S06     I04  2.52930872 3.018906 -0.2831109 -0.30730849  0.10082254
#> 49     S06     I05  2.81290514 3.018906 -0.2831109  0.01253761  0.06457286
#> 56     S06     I06  2.63093950 3.018906 -0.2831109  0.08815511 -0.19301027
#> 66     S06     I07  0.92929480 3.018906 -0.2831109 -0.47142095 -1.33507892
#> 79     S06     I08  4.63083299 3.018906 -0.2831109  0.78598369  1.10905463
#> 85     S06     I09  1.27182367 3.018906 -0.2831109 -0.21917677 -1.24479423
#> 99     S06     I10  4.08739148 3.018906 -0.2831109  0.10429483  1.24730198
#> 9      S07     I01  0.62769279 3.018906  0.8490571 -1.35431210 -1.88595780
#> 11     S07     I02  4.24420650 3.018906  0.8490571 -0.10926927  0.48551307
#> 24     S07     I03  1.93089570 3.018906  0.8490571 -2.32224648  0.38517948
#> 38     S07     I04  4.90441668 3.018906  0.8490571 -0.30730849  1.34376248
#> 50     S07     I05  2.09754137 3.018906  0.8490571  0.01253761 -1.78295894
#> 53     S07     I06  4.47806002 3.018906  0.8490571  0.08815511  0.52194222
#> 67     S07     I07  2.96675065 3.018906  0.8490571 -0.47142095 -0.42979109
#> 80     S07     I08  3.37348971 3.018906  0.8490571  0.78598369 -1.28045668
#> 82     S07     I09  2.71008106 3.018906  0.8490571 -0.21917677 -0.93870486
#> 96     S07     I10  4.04737350 3.018906  0.8490571  0.10429483  0.07511597
#> 6      S08     I01  0.42074315 3.018906 -1.1006616 -1.35431210 -0.14318873
#> 12     S08     I02  1.46106976 3.018906 -1.1006616 -0.10926927 -0.34790496
#> 25     S08     I03  1.01379504 3.018906 -1.1006616 -2.32224648  1.41779754
#> 39     S08     I04  1.82761494 3.018906 -1.1006616 -0.30730849  0.21667945
#> 41     S08     I05  1.67318995 3.018906 -1.1006616  0.01253761 -0.25759164
#> 54     S08     I06  0.90590868 3.018906 -1.1006616  0.08815511 -1.10049041
#> 68     S08     I07  3.68852932 3.018906 -1.1006616 -0.47142095  2.24170629
#> 78     S08     I08  1.95743731 3.018906 -1.1006616  0.78598369 -0.74679037
#> 83     S08     I09  2.42695256 3.018906 -1.1006616 -0.21917677  0.72788535
#> 97     S08     I10  2.89113158 3.018906 -1.1006616  0.10429483  0.86859278
#> 3      S09     I01  4.11515269 3.018906  1.4597740 -1.35431210  0.99078517
#> 13     S09     I02  3.38758513 3.018906  1.4597740 -0.10926927 -0.98182522
#> 26     S09     I03  1.52423283 3.018906  1.4597740 -2.32224648 -0.63220032
#> 32     S09     I04  5.32100842 3.018906  1.4597740 -0.30730849  1.14963728
#> 42     S09     I05  3.45553597 3.018906  1.4597740  0.01253761 -1.03568126
#> 55     S09     I06  5.24981522 3.018906  1.4597740  0.08815511  0.68298049
#> 65     S09     I07  3.60737888 3.018906  1.4597740 -0.47142095 -0.39987979
#> 71     S09     I08  4.08299587 3.018906  1.4597740  0.78598369 -1.18166745
#> 84     S09     I09  4.72866659 3.018906  1.4597740 -0.21917677  0.46916373
#> 94     S09     I10  3.42402180 3.018906  1.4597740  0.10429483 -1.15895266
#> 10     S10     I01  0.41335198 3.018906 -1.0980540 -1.35431210 -0.15318755
#> 14     S10     I02  0.95005984 3.018906 -1.0980540 -0.10926927 -0.86152252
#> 28     S10     I03  0.43856723 3.018906 -1.0980540 -2.32224648  0.83996207
#> 40     S10     I04  1.68439394 3.018906 -1.0980540 -0.30730849  0.07085080
#> 43     S10     I05  1.35933368 3.018906 -1.0980540  0.01253761 -0.57405556
#> 57     S10     I06  5.55528604 3.018906 -1.0980540  0.08815511  3.54627930
#> 70     S10     I07  1.36939249 3.018906 -1.0980540 -0.47142095 -0.08003819
#> 72     S10     I08  4.23124615 3.018906 -1.0980540  0.78598369  1.52441082
#> 86     S10     I09  1.15654102 3.018906 -1.0980540 -0.21917677 -0.54513384
#> 100    S10     I10  2.23099134 3.018906 -1.0980540  0.10429483  0.20584488