Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id           y  grand_i       sub_i      item_i         err
#> 1      S01     I01  2.18961940 3.018906 -0.79691418 -0.74639314  0.71402110
#> 15     S01     I02  4.85830367 3.018906 -0.79691418  1.67348237  0.96282986
#> 29     S01     I03  3.29075871 3.018906 -0.79691418 -0.08902361  1.15779089
#> 31     S01     I04 -0.13306713 3.018906 -0.79691418 -0.55597224 -1.79908632
#> 44     S01     I05  1.87688557 3.018906 -0.79691418 -0.97973891  0.63463305
#> 58     S01     I06  2.12180885 3.018906 -0.79691418 -0.07292973 -0.02725286
#> 61     S01     I07  2.92573310 3.018906 -0.79691418 -0.29586461  0.99960628
#> 73     S01     I08  1.32103287 3.018906 -0.79691418  0.60198801 -1.50294658
#> 87     S01     I09  3.24226330 3.018906 -0.79691418  1.54574768 -0.52547581
#> 91     S01     I10  0.01109906 3.018906 -0.79691418 -0.41128960 -1.79960277
#> 7      S02     I01  3.64900832 3.018906  1.91154309 -0.74639314 -0.53504724
#> 17     S02     I02  7.87920015 3.018906  1.91154309  1.67348237  1.27526907
#> 30     S02     I03  4.86805437 3.018906  1.91154309 -0.08902361  0.02662927
#> 36     S02     I04  3.93929493 3.018906  1.91154309 -0.55597224 -0.43518153
#> 46     S02     I05  4.01049451 3.018906  1.91154309 -0.97973891  0.05978473
#> 59     S02     I06  4.32304356 3.018906  1.91154309 -0.07292973 -0.53447542
#> 69     S02     I07  6.31482257 3.018906  1.91154309 -0.29586461  1.68023848
#> 75     S02     I08  6.35978328 3.018906  1.91154309  0.60198801  0.82734656
#> 88     S02     I09  6.35597698 3.018906  1.91154309  1.54574768 -0.12021939
#> 98     S02     I10  3.83756669 3.018906  1.91154309 -0.41128960 -0.68159241
#> 4      S03     I01  3.34403342 3.018906  0.09192124 -0.74639314  0.97959970
#> 18     S03     I02  4.99906985 3.018906  0.09192124  1.67348237  0.21476062
#> 23     S03     I03  0.16551955 3.018906  0.09192124 -0.08902361 -2.85628370
#> 33     S03     I04  0.83732009 3.018906  0.09192124 -0.55597224 -1.71753452
#> 47     S03     I05  2.62946492 3.018906  0.09192124 -0.97973891  0.49837698
#> 60     S03     I06  3.16351661 3.018906  0.09192124 -0.07292973  0.12561949
#> 62     S03     I07  4.19946643 3.018906  0.09192124 -0.29586461  1.38450418
#> 76     S03     I08  4.11719410 3.018906  0.09192124  0.60198801  0.40437923
#> 89     S03     I09  6.31605796 3.018906  0.09192124  1.54574768  1.65948343
#> 95     S03     I10  3.91604845 3.018906  0.09192124 -0.41128960  1.21651119
#> 5      S04     I01  2.23696595 3.018906  0.86934943 -0.74639314 -0.90489595
#> 19     S04     I02  3.75413218 3.018906  0.86934943  1.67348237 -1.80760524
#> 21     S04     I03  0.73748191 3.018906  0.86934943 -0.08902361 -3.06174953
#> 34     S04     I04  2.49413662 3.018906  0.86934943 -0.55597224 -0.83814618
#> 48     S04     I05  1.52173889 3.018906  0.86934943 -0.97973891 -1.38677724
#> 51     S04     I06  3.50896102 3.018906  0.86934943 -0.07292973 -0.30636429
#> 63     S04     I07  2.52829010 3.018906  0.86934943 -0.29586461 -1.06410033
#> 77     S04     I08  6.41509368 3.018906  0.86934943  0.60198801  1.92485062
#> 90     S04     I09  5.24925895 3.018906  0.86934943  1.54574768 -0.18474377
#> 92     S04     I10  4.59109402 3.018906  0.86934943 -0.41128960  1.11412857
#> 2      S05     I01  1.57069609 3.018906 -0.59889839 -0.74639314 -0.10291800
#> 16     S05     I02  2.46653542 3.018906 -0.59889839  1.67348237 -1.62695417
#> 22     S05     I03  2.22123773 3.018906 -0.59889839 -0.08902361 -0.10974589
#> 35     S05     I04  2.56010657 3.018906 -0.59889839 -0.55597224  0.69607159
#> 45     S05     I05  1.06796672 3.018906 -0.59889839 -0.97973891 -0.37230159
#> 52     S05     I06  1.87563941 3.018906 -0.59889839 -0.07292973 -0.47143808
#> 64     S05     I07  0.93891934 3.018906 -0.59889839 -0.29586461 -1.18522328
#> 74     S05     I08  3.07255128 3.018906 -0.59889839  0.60198801  0.05055604
#> 81     S05     I09  4.47040507 3.018906 -0.59889839  1.54574768  0.50465017
#> 93     S05     I10 -0.07854354 3.018906 -0.59889839 -0.41128960 -2.08726116
#> 8      S06     I01  2.82574772 3.018906 -0.42172141 -0.74639314  0.97495666
#> 20     S06     I02  3.82212677 3.018906 -0.42172141  1.67348237 -0.44853980
#> 27     S06     I03  2.55177439 3.018906 -0.42172141 -0.08902361  0.04361380
#> 37     S06     I04  1.93337663 3.018906 -0.42172141 -0.55597224 -0.10783532
#> 49     S06     I05  3.42189090 3.018906 -0.42172141 -0.97973891  1.80444561
#> 56     S06     I06  5.18969466 3.018906 -0.42172141 -0.07292973  2.66544019
#> 66     S06     I07  2.71367988 3.018906 -0.42172141 -0.29586461  0.41236029
#> 79     S06     I08  4.49365529 3.018906 -0.42172141  0.60198801  1.29448308
#> 85     S06     I09  4.98551224 3.018906 -0.42172141  1.54574768  0.84258036
#> 99     S06     I10  2.79925557 3.018906 -0.42172141 -0.41128960  0.61336097
#> 9      S07     I01  3.24774922 3.018906 -0.42392211 -0.74639314  1.39915886
#> 11     S07     I02  2.11531717 3.018906 -0.42392211  1.67348237 -2.15314871
#> 24     S07     I03  1.68015916 3.018906 -0.42392211 -0.08902361 -0.82580073
#> 38     S07     I04  2.33443768 3.018906 -0.42392211 -0.55597224  0.29542642
#> 50     S07     I05  1.05110350 3.018906 -0.42392211 -0.97973891 -0.56414108
#> 53     S07     I06  2.68775281 3.018906 -0.42392211 -0.07292973  0.16569903
#> 67     S07     I07  2.45586293 3.018906 -0.42392211 -0.29586461  0.15674404
#> 80     S07     I08  2.43352802 3.018906 -0.42392211  0.60198801 -0.76344349
#> 82     S07     I09  4.59239607 3.018906 -0.42392211  1.54574768  0.45166490
#> 96     S07     I10  0.57395427 3.018906 -0.42392211 -0.41128960 -1.60973963
#> 6      S08     I01  3.64704056 3.018906  0.60710990 -0.74639314  0.76741819
#> 12     S08     I02  4.84433894 3.018906  0.60710990  1.67348237 -0.45515895
#> 25     S08     I03  5.11030366 3.018906  0.60710990 -0.08902361  1.57331175
#> 39     S08     I04  1.16086312 3.018906  0.60710990 -0.55597224 -1.90918016
#> 41     S08     I05  2.87600548 3.018906  0.60710990 -0.97973891  0.22972888
#> 54     S08     I06  3.84131444 3.018906  0.60710990 -0.07292973  0.28822865
#> 68     S08     I07  3.57417019 3.018906  0.60710990 -0.29586461  0.24401929
#> 78     S08     I08  3.82212213 3.018906  0.60710990  0.60198801 -0.40588139
#> 83     S08     I09  5.81805891 3.018906  0.60710990  1.54574768  0.64629572
#> 97     S08     I10  0.34883442 3.018906  0.60710990 -0.41128960 -2.86589150
#> 3      S09     I01  0.49694151 3.018906  0.42369458 -0.74639314 -2.19926555
#> 13     S09     I02  6.14265526 3.018906  0.42369458  1.67348237  1.02657269
#> 26     S09     I03  3.77423729 3.018906  0.42369458 -0.08902361  0.42066071
#> 32     S09     I04  2.46312640 3.018906  0.42369458 -0.55597224 -0.42350155
#> 42     S09     I05  3.02132564 3.018906  0.42369458 -0.97973891  0.55846436
#> 55     S09     I06  2.16592876 3.018906  0.42369458 -0.07292973 -1.20374171
#> 65     S09     I07  2.69944946 3.018906  0.42369458 -0.29586461 -0.44728613
#> 71     S09     I08  4.20624568 3.018906  0.42369458  0.60198801  0.16165747
#> 84     S09     I09  5.71721134 3.018906  0.42369458  1.54574768  0.72886347
#> 94     S09     I10  1.01114241 3.018906  0.42369458 -0.41128960 -2.02016818
#> 10     S10     I01  2.43996213 3.018906 -0.15960485 -0.74639314  0.32705450
#> 14     S10     I02  3.66136472 3.018906 -0.15960485  1.67348237 -0.87141842
#> 28     S10     I03  1.77896685 3.018906 -0.15960485 -0.08902361 -0.99131031
#> 40     S10     I04  1.74967094 3.018906 -0.15960485 -0.55597224 -0.55365758
#> 43     S10     I05  1.34489985 3.018906 -0.15960485 -0.97973891 -0.53466201
#> 57     S10     I06  3.33889277 3.018906 -0.15960485 -0.07292973  0.55252173
#> 70     S10     I07  1.79332081 3.018906 -0.15960485 -0.29586461 -0.77011534
#> 72     S10     I08  3.26789916 3.018906 -0.15960485  0.60198801 -0.19338962
#> 86     S10     I09  5.01827784 3.018906 -0.15960485  1.54574768  0.61322940
#> 100    S10     I10  2.66479161 3.018906 -0.15960485 -0.41128960  0.21678044