Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

Usage

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id            y  grand_i       sub_i       item_i          err
#> 1      S01     I01  4.320472154 3.018906  0.29587997  1.195712059 -0.190025483
#> 15     S01     I02  4.786391204 3.018906  0.29587997  0.006075405  1.465530220
#> 29     S01     I03  2.466024676 3.018906  0.29587997 -0.849962510  0.001201608
#> 31     S01     I04  4.978864510 3.018906  0.29587997  0.792180548  0.871898384
#> 44     S01     I05  2.508917502 3.018906  0.29587997 -0.729131847 -0.076736229
#> 58     S01     I06  1.828510192 3.018906  0.29587997 -1.498689084  0.012413698
#> 61     S01     I07  2.515319333 3.018906  0.29587997 -0.002152109 -0.797314136
#> 73     S01     I08  1.278015373 3.018906  0.29587997 -1.406754998 -0.630015207
#> 87     S01     I09  4.041724911 3.018906  0.29587997 -0.676883534  1.403822867
#> 91     S01     I10  4.390553562 3.018906  0.29587997 -0.074833065  1.150601049
#> 7      S02     I01  3.344195242 3.018906 -0.47515066  1.195712059 -0.395271774
#> 17     S02     I02  2.080591647 3.018906 -0.47515066  0.006075405 -0.469238716
#> 30     S02     I03  1.458233969 3.018906 -0.47515066 -0.849962510 -0.235558479
#> 36     S02     I04  4.104469124 3.018906 -0.47515066  0.792180548  0.768533618
#> 46     S02     I05  2.226081401 3.018906 -0.47515066 -0.729131847  0.411458291
#> 59     S02     I06 -0.511410891 3.018906 -0.47515066 -1.498689084 -1.556476765
#> 69     S02     I07  3.258503910 3.018906 -0.47515066 -0.002152109  0.716901062
#> 75     S02     I08  1.359268058 3.018906 -0.47515066 -1.406754998  0.222268099
#> 88     S02     I09  2.373341237 3.018906 -0.47515066 -0.676883534  0.506469814
#> 98     S02     I10  1.464174778 3.018906 -0.47515066 -0.074833065 -1.004747114
#> 4      S03     I01  2.112172288 3.018906 -0.56341458  1.195712059 -1.539030803
#> 18     S03     I02  1.946848597 3.018906 -0.56341458  0.006075405 -0.514717840
#> 23     S03     I03  2.349296408 3.018906 -0.56341458 -0.849962510  0.743767886
#> 33     S03     I04  3.429913242 3.018906 -0.56341458  0.792180548  0.182241662
#> 47     S03     I05  1.329122851 3.018906 -0.56341458 -0.729131847 -0.397236334
#> 60     S03     I06  2.308757280 3.018906 -0.56341458 -1.498689084  1.351955332
#> 62     S03     I07  2.476882919 3.018906 -0.56341458 -0.002152109  0.023543995
#> 76     S03     I08  1.127752886 3.018906 -0.56341458 -1.406754998  0.079016852
#> 89     S03     I09  1.649523590 3.018906 -0.56341458 -0.676883534 -0.129083909
#> 95     S03     I10  1.133324289 3.018906 -0.56341458 -0.074833065 -1.247333679
#> 5      S04     I01  5.532828046 3.018906 -0.37288499  1.195712059  1.691095369
#> 19     S04     I02  1.919484689 3.018906 -0.37288499  0.006075405 -0.732611334
#> 21     S04     I03  0.952201553 3.018906 -0.37288499 -0.849962510 -0.843856555
#> 34     S04     I04  1.956923824 3.018906 -0.37288499  0.792180548 -1.481277342
#> 48     S04     I05  4.516080508 3.018906 -0.37288499 -0.729131847  2.599191737
#> 51     S04     I06  1.106407793 3.018906 -0.37288499 -1.498689084 -0.040923741
#> 63     S04     I07  2.376313811 3.018906 -0.37288499 -0.002152109 -0.267554698
#> 77     S04     I08  3.759703416 3.018906 -0.37288499 -1.406754998  2.520437797
#> 90     S04     I09  2.191698571 3.018906 -0.37288499 -0.676883534  0.222561487
#> 92     S04     I10  2.031085908 3.018906 -0.37288499 -0.074833065 -0.540101645
#> 2      S05     I01  5.744495874 3.018906  0.52472456  1.195712059  1.005153647
#> 16     S05     I02  4.403843643 3.018906  0.52472456  0.006075405  0.854138070
#> 22     S05     I03  3.535494979 3.018906  0.52472456 -0.849962510  0.841827321
#> 35     S05     I04  5.257868283 3.018906  0.52472456  0.792180548  0.922057567
#> 45     S05     I05  2.812009545 3.018906  0.52472456 -0.729131847 -0.002488777
#> 52     S05     I06  3.110593913 3.018906  0.52472456 -1.498689084  1.065652829
#> 64     S05     I07  3.191038717 3.018906  0.52472456 -0.002152109 -0.350439342
#> 74     S05     I08  3.017546964 3.018906  0.52472456 -1.406754998  0.880671794
#> 81     S05     I09  2.030981203 3.018906  0.52472456 -0.676883534 -0.835765431
#> 93     S05     I10  2.033522150 3.018906  0.52472456 -0.074833065 -1.435274953
#> 8      S06     I01  4.946223800 3.018906 -0.35659636  1.195712059  1.088202489
#> 20     S06     I02 -0.418632810 3.018906 -0.35659636  0.006075405 -3.087017467
#> 27     S06     I03  2.875915052 3.018906 -0.35659636 -0.849962510  1.063568310
#> 37     S06     I04  2.330256863 3.018906 -0.35659636  0.792180548 -1.124232936
#> 49     S06     I05 -0.210998402 3.018906 -0.35659636 -0.729131847 -2.144175806
#> 56     S06     I06  1.045112651 3.018906 -0.35659636 -1.498689084 -0.118507516
#> 66     S06     I07  3.471333977 3.018906 -0.35659636 -0.002152109  0.811176834
#> 79     S06     I08  0.055651431 3.018906 -0.35659636 -1.406754998 -1.199902822
#> 85     S06     I09  0.026751698 3.018906 -0.35659636 -0.676883534 -1.958674019
#> 99     S06     I10  4.022883525 3.018906 -0.35659636 -0.074833065  1.435407339
#> 9      S07     I01  3.916520009 3.018906 -0.16910721  1.195712059 -0.128990448
#> 11     S07     I02  0.874657454 3.018906 -0.16910721  0.006075405 -1.981216350
#> 24     S07     I03  3.212121805 3.018906 -0.16910721 -0.849962510  1.212285917
#> 38     S07     I04  3.975861487 3.018906 -0.16910721  0.792180548  0.333882541
#> 50     S07     I05  1.774805309 3.018906 -0.16910721 -0.729131847 -0.345861242
#> 53     S07     I06  3.021809906 3.018906 -0.16910721 -1.498689084  1.670700593
#> 67     S07     I07  2.546155308 3.018906 -0.16910721 -0.002152109 -0.301490982
#> 80     S07     I08  0.857776819 3.018906 -0.16910721 -1.406754998 -0.585266581
#> 82     S07     I09  2.462478229 3.018906 -0.16910721 -0.676883534  0.289563364
#> 96     S07     I10  1.639575452 3.018906 -0.16910721 -0.074833065 -1.135389881
#> 6      S08     I01  3.112204495 3.018906 -0.02392489  1.195712059 -1.078488286
#> 12     S08     I02  4.400069858 3.018906 -0.02392489  0.006075405  1.399013731
#> 25     S08     I03  2.910136120 3.018906 -0.02392489 -0.849962510  0.765117908
#> 39     S08     I04  4.140267959 3.018906 -0.02392489  0.792180548  0.353106689
#> 41     S08     I05  2.800661260 3.018906 -0.02392489 -0.729131847  0.534812385
#> 54     S08     I06  3.314393980 3.018906 -0.02392489 -1.498689084  1.818102343
#> 68     S08     I07  2.371576068 3.018906 -0.02392489 -0.002152109 -0.621252545
#> 78     S08     I08  4.037200184 3.018906 -0.02392489 -1.406754998  2.448974460
#> 83     S08     I09  4.761232263 3.018906 -0.02392489 -0.676883534  2.443135075
#> 97     S08     I10  3.033050172 3.018906 -0.02392489 -0.074833065  0.112902515
#> 3      S09     I01  3.058538755 3.018906 -0.48394703  1.195712059 -0.672131882
#> 13     S09     I02  2.413533702 3.018906 -0.48394703  0.006075405 -0.127500282
#> 26     S09     I03 -0.097493988 3.018906 -0.48394703 -0.849962510 -1.782490056
#> 32     S09     I04  3.954625943 3.018906 -0.48394703  0.792180548  0.627486817
#> 42     S09     I05  0.599599087 3.018906 -0.48394703 -0.729131847 -1.206227645
#> 55     S09     I06  2.334109873 3.018906 -0.48394703 -1.498689084  1.297840379
#> 65     S09     I07  3.210850072 3.018906 -0.48394703 -0.002152109  0.678043603
#> 71     S09     I08  2.905929155 3.018906 -0.48394703 -1.406754998  1.777725576
#> 84     S09     I09 -0.002418746 3.018906 -0.48394703 -0.676883534 -1.860493790
#> 94     S09     I10  1.913294652 3.018906 -0.48394703 -0.074833065 -0.546830862
#> 10     S10     I01  4.687522438 3.018906  1.01523871  1.195712059 -0.542333947
#> 14     S10     I02  4.351904159 3.018906  1.01523871  0.006075405  0.311684428
#> 28     S10     I03  1.710793218 3.018906  1.01523871 -0.849962510 -1.473388598
#> 40     S10     I04  4.122970288 3.018906  1.01523871  0.792180548 -0.703354586
#> 43     S10     I05  4.196209970 3.018906  1.01523871 -0.729131847  0.891197491
#> 57     S10     I06  2.582200474 3.018906  1.01523871 -1.498689084  0.046745231
#> 70     S10     I07  4.705416597 3.018906  1.01523871 -0.002152109  0.673424379
#> 72     S10     I08  0.717186382 3.018906  1.01523871 -1.406754998 -1.910202946
#> 86     S10     I09  2.440676561 3.018906  1.01523871 -0.676883534 -0.916584232
#> 100    S10     I10  3.520483374 3.018906  1.01523871 -0.074833065 -0.438827887