Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id         y  grand_i       sub_i      item_i           err
#> 1      S01     I01 4.1152324 3.018906  0.56998287  0.06372689  0.4626170223
#> 15     S01     I02 4.4224176 3.018906  0.56998287 -0.61913239  1.4526614811
#> 29     S01     I03 3.8339794 3.018906  0.56998287 -0.47084975  0.7159406247
#> 31     S01     I04 3.6693181 3.018906  0.56998287 -0.79426537  0.8746950317
#> 44     S01     I05 4.1682274 3.018906  0.56998287  0.04751609  0.5318228254
#> 58     S01     I06 4.5589567 3.018906  0.56998287 -0.03988620  1.0099544077
#> 61     S01     I07 3.3139561 3.018906  0.56998287 -0.15429743 -0.1206349769
#> 73     S01     I08 3.4466844 3.018906  0.56998287 -0.10140898 -0.0407951098
#> 87     S01     I09 4.1150472 3.018906  0.56998287 -0.44323124  0.9693899904
#> 91     S01     I10 3.7546416 3.018906  0.56998287 -0.41773369  0.5834868363
#> 7      S02     I01 4.4331212 3.018906  0.15562574  0.06372689  1.1948629436
#> 17     S02     I02 2.6376456 3.018906  0.15562574 -0.61913239  0.0822465999
#> 30     S02     I03 1.7765036 3.018906  0.15562574 -0.47084975 -0.9271780136
#> 36     S02     I04 2.0848855 3.018906  0.15562574 -0.79426537 -0.2953805179
#> 46     S02     I05 2.7319999 3.018906  0.15562574  0.04751609 -0.4900475211
#> 59     S02     I06 2.6284272 3.018906  0.15562574 -0.03988620 -0.5062179545
#> 69     S02     I07 2.3668419 3.018906  0.15562574 -0.15429743 -0.6533920233
#> 75     S02     I08 3.4096195 3.018906  0.15562574 -0.10140898  0.3364971592
#> 88     S02     I09 2.0566689 3.018906  0.15562574 -0.44323124 -0.6746311797
#> 98     S02     I10 2.3720810 3.018906  0.15562574 -0.41773369 -0.3847166733
#> 4      S03     I01 4.7030774 3.018906  0.03692411  0.06372689  1.5835208395
#> 18     S03     I02 0.8793870 3.018906  0.03692411 -0.61913239 -1.5573102959
#> 23     S03     I03 2.0037907 3.018906  0.03692411 -0.47084975 -0.5811892504
#> 33     S03     I04 1.0014127 3.018906  0.03692411 -0.79426537 -1.2601516722
#> 47     S03     I05 3.2211549 3.018906  0.03692411  0.04751609  0.1178091041
#> 60     S03     I06 1.3304699 3.018906  0.03692411 -0.03988620 -1.6854735746
#> 62     S03     I07 3.1215507 3.018906  0.03692411 -0.15429743  0.2200184260
#> 76     S03     I08 5.3608647 3.018906  0.03692411 -0.10140898  2.4064439566
#> 89     S03     I09 2.5733421 3.018906  0.03692411 -0.44323124 -0.0392563757
#> 95     S03     I10 3.3314952 3.018906  0.03692411 -0.41773369  0.6933991949
#> 5      S04     I01 2.6792280 3.018906  0.19476739  0.06372689 -0.5981719043
#> 19     S04     I02 1.8084207 3.018906  0.19476739 -0.61913239 -0.7861199053
#> 21     S04     I03 3.9052572 3.018906  0.19476739 -0.47084975  1.1624339153
#> 34     S04     I04 0.2667572 3.018906  0.19476739 -0.79426537 -2.1526504501
#> 48     S04     I05 2.0248756 3.018906  0.19476739  0.04751609 -1.2363135229
#> 51     S04     I06 1.6265464 3.018906  0.19476739 -0.03988620 -1.5472403653
#> 63     S04     I07 1.7877199 3.018906  0.19476739 -0.15429743 -1.2716556202
#> 77     S04     I08 2.5907081 3.018906  0.19476739 -0.10140898 -0.5215558932
#> 90     S04     I09 2.0019422 3.018906  0.19476739 -0.44323124 -0.7684995082
#> 92     S04     I10 3.9799185 3.018906  0.19476739 -0.41773369  1.1839792246
#> 2      S05     I01 3.5292700 3.018906 -0.75807487  0.06372689  1.2047123566
#> 16     S05     I02 1.3561640 3.018906 -0.75807487 -0.61913239 -0.2855343293
#> 22     S05     I03 3.2562826 3.018906 -0.75807487 -0.47084975  1.4663015969
#> 35     S05     I04 2.2140772 3.018906 -0.75807487 -0.79426537  0.7475118666
#> 45     S05     I05 3.8999286 3.018906 -0.75807487  0.04751609  1.5915817696
#> 52     S05     I06 4.4560178 3.018906 -0.75807487 -0.03988620  2.2350733124
#> 64     S05     I07 3.5746937 3.018906 -0.75807487 -0.15429743  1.4681603850
#> 74     S05     I08 0.3091243 3.018906 -0.75807487 -0.10140898 -1.8502974398
#> 81     S05     I09 2.2126284 3.018906 -0.75807487 -0.44323124  0.3950288774
#> 93     S05     I10 0.5045372 3.018906 -0.75807487 -0.41773369 -1.3385598100
#> 8      S06     I01 3.0560640 3.018906  0.48363404  0.06372689 -0.5102025526
#> 20     S06     I02 3.8891499 3.018906  0.48363404 -0.61913239  1.0057426556
#> 27     S06     I03 4.8701173 3.018906  0.48363404 -0.47084975  1.8384274413
#> 37     S06     I04 1.4001795 3.018906  0.48363404 -0.79426537 -1.3080948287
#> 49     S06     I05 2.0216018 3.018906  0.48363404  0.04751609 -1.5284539771
#> 56     S06     I06 3.7873565 3.018906  0.48363404 -0.03988620  0.3247030803
#> 66     S06     I07 3.2805782 3.018906  0.48363404 -0.15429743 -0.0676640293
#> 79     S06     I08 4.7155581 3.018906  0.48363404 -0.10140898  1.3144273905
#> 85     S06     I09 3.2717270 3.018906  0.48363404 -0.44323124  0.2124186328
#> 99     S06     I10 3.2642692 3.018906  0.48363404 -0.41773369  0.1794631937
#> 9      S07     I01 3.6939536 3.018906  0.54553255  0.06372689  0.0657885048
#> 11     S07     I02 1.2170700 3.018906  0.54553255 -0.61913239 -1.7282357430
#> 24     S07     I03 4.2110089 3.018906  0.54553255 -0.47084975  1.1174204576
#> 38     S07     I04 2.1033565 3.018906  0.54553255 -0.79426537 -0.6668162783
#> 50     S07     I05 3.7363105 3.018906  0.54553255  0.04751609  0.1243562894
#> 53     S07     I06 3.6785335 3.018906  0.54553255 -0.03988620  0.1539815503
#> 67     S07     I07 4.7664235 3.018906  0.54553255 -0.15429743  1.3562827924
#> 80     S07     I08 3.0045458 3.018906  0.54553255 -0.10140898 -0.4584834292
#> 82     S07     I09 3.8416183 3.018906  0.54553255 -0.44323124  0.7204114222
#> 96     S07     I10 2.8716905 3.018906  0.54553255 -0.41773369 -0.2750140285
#> 6      S08     I01 2.2301931 3.018906 -0.40144258  0.06372689 -0.4509967811
#> 12     S08     I02 2.3133628 3.018906 -0.40144258 -0.61913239  0.3150321147
#> 25     S08     I03 2.0169064 3.018906 -0.40144258 -0.47084975 -0.1297068522
#> 39     S08     I04 2.0436948 3.018906 -0.40144258 -0.79426537  0.2204971325
#> 41     S08     I05 0.9715504 3.018906 -0.40144258  0.04751609 -1.6934287350
#> 54     S08     I06 1.8520727 3.018906 -0.40144258 -0.03988620 -0.7255041591
#> 68     S08     I07 2.4390171 3.018906 -0.40144258 -0.15429743 -0.0241485454
#> 78     S08     I08 1.0305279 3.018906 -0.40144258 -0.10140898 -1.4855261434
#> 83     S08     I09 1.9171300 3.018906 -0.40144258 -0.44323124 -0.2571018196
#> 97     S08     I10 1.6690544 3.018906 -0.40144258 -0.41773369 -0.5306749508
#> 3      S09     I01 2.9555480 3.018906  0.57304625  0.06372689 -0.7001307111
#> 13     S09     I02 2.2941068 3.018906  0.57304625 -0.61913239 -0.6787127129
#> 26     S09     I03 0.7553089 3.018906  0.57304625 -0.47084975 -2.3657931602
#> 32     S09     I04 3.7013706 3.018906  0.57304625 -0.79426537  0.9036840992
#> 42     S09     I05 4.5290509 3.018906  0.57304625  0.04751609  0.8895829209
#> 55     S09     I06 3.7888207 3.018906  0.57304625 -0.03988620  0.2367550325
#> 65     S09     I07 3.4651418 3.018906  0.57304625 -0.15429743  0.0274874126
#> 71     S09     I08 4.5560139 3.018906  0.57304625 -0.10140898  1.0654710594
#> 84     S09     I09 2.9545448 3.018906  0.57304625 -0.44323124 -0.1941758110
#> 94     S09     I10 2.7663826 3.018906  0.57304625 -0.41773369 -0.4078355519
#> 10     S10     I01 4.9049963 3.018906  0.42651506  0.06372689  1.3958487131
#> 14     S10     I02 2.2826745 3.018906  0.42651506 -0.61913239 -0.5436137684
#> 28     S10     I03 3.0489810 3.018906  0.42651506 -0.47084975  0.0744100611
#> 40     S10     I04 2.2858762 3.018906  0.42651506 -0.79426537 -0.3652791241
#> 43     S10     I05 2.1215325 3.018906  0.42651506  0.04751609 -1.3714042494
#> 57     S10     I06 4.4669262 3.018906  0.42651506 -0.03988620  1.0613917740
#> 70     S10     I07 3.0090947 3.018906  0.42651506 -0.15429743 -0.2820285529
#> 72     S10     I08 4.0185082 3.018906  0.42651506 -0.10140898  0.6744965044
#> 86     S10     I09 1.0367855 3.018906  0.42651506 -0.44323124 -1.9654039006
#> 100    S10     I10 3.0267011 3.018906  0.42651506 -0.41773369 -0.0009858907