Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data
the existing tbl

sub_n
the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n
the number of items to simulate (if NULL, returns data for the same items)

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id          y  grand_i       sub_i      item_i          err
#> 1      S01     I01  3.6598733 3.018906  0.35054166 -0.42312159  0.713547579
#> 15     S01     I02  4.5871715 3.018906  0.35054166  0.76588428  0.451839974
#> 29     S01     I03  2.1024761 3.018906  0.35054166 -0.11211864 -1.154852555
#> 31     S01     I04  4.6039422 3.018906  0.35054166  0.08501020  1.149484683
#> 44     S01     I05  3.5936695 3.018906  0.35054166  0.10363931  0.120582914
#> 58     S01     I06  4.8268173 3.018906  0.35054166  0.64079992  0.816570078
#> 61     S01     I07  3.3850485 3.018906  0.35054166  0.54672988 -0.531128640
#> 73     S01     I08  5.2660592 3.018906  0.35054166  0.12294393  1.773668015
#> 87     S01     I09  2.8228659 3.018906  0.35054166 -0.58899663  0.042415204
#> 91     S01     I10  2.6925674 3.018906  0.35054166 -0.05474597 -0.622133930
#> 7      S02     I01  1.1057872 3.018906 -0.29850683 -0.42312159 -1.191489975
#> 17     S02     I02  5.9067406 3.018906 -0.29850683  0.76588428  2.420457526
#> 30     S02     I03  3.5591342 3.018906 -0.29850683 -0.11211864  0.950854089
#> 36     S02     I04  1.1803405 3.018906 -0.29850683  0.08501020 -1.625068470
#> 46     S02     I05  4.8201477 3.018906 -0.29850683  0.10363931  1.996109629
#> 59     S02     I06  2.6075151 3.018906 -0.29850683  0.64079992 -0.753683641
#> 69     S02     I07  2.4954212 3.018906 -0.29850683  0.54672988 -0.771707454
#> 75     S02     I08  2.8421034 3.018906 -0.29850683  0.12294393 -0.001239313
#> 88     S02     I09  2.4284483 3.018906 -0.29850683 -0.58899663  0.297046149
#> 98     S02     I10  4.8443424 3.018906 -0.29850683 -0.05474597  2.178689620
#> 4      S03     I01  3.7495406 3.018906 -0.01801732 -0.42312159  1.171773852
#> 18     S03     I02  3.9209782 3.018906 -0.01801732  0.76588428  0.154205608
#> 23     S03     I03  3.3971310 3.018906 -0.01801732 -0.11211864  0.508361324
#> 33     S03     I04  1.4553026 3.018906 -0.01801732  0.08501020 -1.630595850
#> 47     S03     I05  2.3612690 3.018906 -0.01801732  0.10363931 -0.743258589
#> 60     S03     I06  3.8741605 3.018906 -0.01801732  0.64079992  0.232472282
#> 62     S03     I07  3.1980033 3.018906 -0.01801732  0.54672988 -0.349614854
#> 76     S03     I08  2.8881243 3.018906 -0.01801732  0.12294393 -0.235707898
#> 89     S03     I09  3.4986986 3.018906 -0.01801732 -0.58899663  1.086806929
#> 95     S03     I10  4.4176194 3.018906 -0.01801732 -0.05474597  1.471477069
#> 5      S04     I01  3.9800018 3.018906  1.99784682 -0.42312159 -0.613629062
#> 19     S04     I02  4.5347835 3.018906  1.99784682  0.76588428 -1.247853263
#> 21     S04     I03  3.9446989 3.018906  1.99784682 -0.11211864 -0.959934931
#> 34     S04     I04  4.2552398 3.018906  1.99784682  0.08501020 -0.846522871
#> 48     S04     I05  6.1791485 3.018906  1.99784682  0.10363931  1.058756744
#> 51     S04     I06  6.4722440 3.018906  1.99784682  0.64079992  0.814691695
#> 63     S04     I07  6.3671994 3.018906  1.99784682  0.54672988  0.803717109
#> 77     S04     I08  4.7020295 3.018906  1.99784682  0.12294393 -0.437666894
#> 90     S04     I09  3.6158815 3.018906  1.99784682 -0.58899663 -0.811874322
#> 92     S04     I10  4.4355517 3.018906  1.99784682 -0.05474597 -0.526454808
#> 2      S05     I01  1.7731304 3.018906 -0.39614798 -0.42312159 -0.426505617
#> 16     S05     I02  6.8783842 3.018906 -0.39614798  0.76588428  3.489742245
#> 22     S05     I03  2.0238223 3.018906 -0.39614798 -0.11211864 -0.486816699
#> 35     S05     I04  1.5657032 3.018906 -0.39614798  0.08501020 -1.142064588
#> 45     S05     I05  3.8763320 3.018906 -0.39614798  0.10363931  1.149935067
#> 52     S05     I06  3.3048877 3.018906 -0.39614798  0.64079992  0.041330123
#> 64     S05     I07  2.5533200 3.018906 -0.39614798  0.54672988 -0.616167482
#> 74     S05     I08  1.9596771 3.018906 -0.39614798  0.12294393 -0.786024501
#> 81     S05     I09  2.4103206 3.018906 -0.39614798 -0.58899663  0.376559597
#> 93     S05     I10  2.3685636 3.018906 -0.39614798 -0.05474597 -0.199448090
#> 8      S06     I01  2.4153677 3.018906  0.15451961 -0.42312159 -0.334935890
#> 20     S06     I02  3.8111750 3.018906  0.15451961  0.76588428 -0.128134529
#> 27     S06     I03  3.9077231 3.018906  0.15451961 -0.11211864  0.846416479
#> 37     S06     I04  2.1471394 3.018906  0.15451961  0.08501020 -1.111296036
#> 49     S06     I05  3.4089956 3.018906  0.15451961  0.10363931  0.131931118
#> 56     S06     I06  3.7628887 3.018906  0.15451961  0.64079992 -0.051336433
#> 66     S06     I07  4.0466916 3.018906  0.15451961  0.54672988  0.326536486
#> 79     S06     I08  2.6017252 3.018906  0.15451961  0.12294393 -0.694643908
#> 85     S06     I09  2.7559422 3.018906  0.15451961 -0.58899663  0.171513598
#> 99     S06     I10  1.1666824 3.018906  0.15451961 -0.05474597 -1.951996863
#> 9      S07     I01 -0.7937663 3.018906 -1.62346398 -0.42312159 -1.766086341
#> 11     S07     I02  0.6214938 3.018906 -1.62346398  0.76588428 -1.539832088
#> 24     S07     I03  0.5252600 3.018906 -1.62346398 -0.11211864 -0.758062957
#> 38     S07     I04  2.8431762 3.018906 -1.62346398  0.08501020  1.362724337
#> 50     S07     I05  1.2803736 3.018906 -1.62346398  0.10363931 -0.218707313
#> 53     S07     I06  3.2328369 3.018906 -1.62346398  0.64079992  1.196595365
#> 67     S07     I07  3.4709560 3.018906 -1.62346398  0.54672988  1.528784540
#> 80     S07     I08  1.7355326 3.018906 -1.62346398  0.12294393  0.217147063
#> 82     S07     I09  1.8378476 3.018906 -1.62346398 -0.58899663  1.031402608
#> 96     S07     I10  1.6288733 3.018906 -1.62346398 -0.05474597  0.288177591
#> 6      S08     I01  2.2970676 3.018906  0.90746292 -0.42312159 -1.206179292
#> 12     S08     I02  5.0593517 3.018906  0.90746292  0.76588428  0.367098902
#> 25     S08     I03  4.8035193 3.018906  0.90746292 -0.11211864  0.989269461
#> 39     S08     I04  4.0162783 3.018906  0.90746292  0.08501020  0.004899548
#> 41     S08     I05  4.9231985 3.018906  0.90746292  0.10363931  0.893190675
#> 54     S08     I06  4.2439289 3.018906  0.90746292  0.64079992 -0.323239534
#> 68     S08     I07  5.0362879 3.018906  0.90746292  0.54672988  0.563189532
#> 78     S08     I08  3.6099113 3.018906  0.90746292  0.12294393 -0.439401154
#> 83     S08     I09  3.1179068 3.018906  0.90746292 -0.58899663 -0.219465141
#> 97     S08     I10  4.8818045 3.018906  0.90746292 -0.05474597  1.010181950
#> 3      S09     I01  4.8792240 3.018906  0.89298523 -0.42312159  1.390454751
#> 13     S09     I02  4.9945040 3.018906  0.89298523  0.76588428  0.316728923
#> 26     S09     I03  4.6723348 3.018906  0.89298523 -0.11211864  0.872562619
#> 32     S09     I04  2.2009519 3.018906  0.89298523  0.08501020 -1.795949139
#> 42     S09     I05  4.3429545 3.018906  0.89298523  0.10363931  0.327424383
#> 55     S09     I06  4.2477225 3.018906  0.89298523  0.64079992 -0.304968281
#> 65     S09     I07  5.9259951 3.018906  0.89298523  0.54672988  1.467374361
#> 71     S09     I08  3.2720309 3.018906  0.89298523  0.12294393 -0.762803823
#> 84     S09     I09  2.8645253 3.018906  0.89298523 -0.58899663 -0.458368950
#> 94     S09     I10  3.5151255 3.018906  0.89298523 -0.05474597 -0.342019412
#> 10     S10     I01  4.1928784 3.018906  1.27052246 -0.42312159  0.326571954
#> 14     S10     I02  5.5588878 3.018906  1.27052246  0.76588428  0.503575420
#> 28     S10     I03  3.7562076 3.018906  1.27052246 -0.11211864 -0.421101846
#> 40     S10     I04  3.0417523 3.018906  1.27052246  0.08501020 -1.332685950
#> 43     S10     I05  3.9010089 3.018906  1.27052246  0.10363931 -0.492058444
#> 57     S10     I06  5.1383014 3.018906  1.27052246  0.64079992  0.208073424
#> 70     S10     I07  4.7974349 3.018906  1.27052246  0.54672988 -0.038723086
#> 72     S10     I08  3.7451785 3.018906  1.27052246  0.12294393 -0.667193545
#> 86     S10     I09  3.2907038 3.018906  1.27052246 -0.58899663 -0.409727632
#> 100    S10     I10  3.0542615 3.018906  1.27052246 -0.05474597 -1.180420555