Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

Usage

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id          y  grand_i       sub_i      item_i          err
#> 1      S01     I01  1.4111718 3.018906  0.48261992 -1.21967536 -0.870678367
#> 15     S01     I02  1.8847926 3.018906  0.48261992 -0.38246272 -1.234270231
#> 29     S01     I03  1.1626072 3.018906  0.48261992 -1.29494118 -1.043977107
#> 31     S01     I04  3.2212267 3.018906  0.48261992 -0.48503121  0.204732384
#> 44     S01     I05  3.1272017 3.018906  0.48261992 -0.28134948 -0.092974343
#> 58     S01     I06  3.8966252 3.018906  0.48261992  0.90264916 -0.507549501
#> 61     S01     I07  3.2543928 3.018906  0.48261992  0.09490194 -0.342034615
#> 73     S01     I08  3.1580881 3.018906  0.48261992 -0.47690499  0.133467611
#> 87     S01     I09  1.6518644 3.018906  0.48261992 -0.83001773 -1.019643366
#> 91     S01     I10  3.7874857 3.018906  0.48261992 -0.58787995  0.873840164
#> 7      S02     I01  2.5149268 3.018906  0.67521972 -1.21967536  0.040476841
#> 17     S02     I02  3.8591276 3.018906  0.67521972 -0.38246272  0.547464975
#> 30     S02     I03  2.0257125 3.018906  0.67521972 -1.29494118 -0.373471703
#> 36     S02     I04  2.5783007 3.018906  0.67521972 -0.48503121 -0.630793375
#> 46     S02     I05  2.5386140 3.018906  0.67521972 -0.28134948 -0.874161828
#> 59     S02     I06  5.9025511 3.018906  0.67521972  0.90264916  1.305776603
#> 69     S02     I07  3.3907877 3.018906  0.67521972  0.09490194 -0.398239590
#> 75     S02     I08  1.8302869 3.018906  0.67521972 -0.47690499 -1.386933475
#> 88     S02     I09  3.0062540 3.018906  0.67521972 -0.83001773  0.142146416
#> 98     S02     I10  2.6931470 3.018906  0.67521972 -0.58787995 -0.413098361
#> 4      S03     I01  1.9764829 3.018906  0.64406637 -1.21967536 -0.466813768
#> 18     S03     I02  3.1715589 3.018906  0.64406637 -0.38246272 -0.108950372
#> 23     S03     I03  1.1677294 3.018906  0.64406637 -1.29494118 -1.200301450
#> 33     S03     I04  4.3248227 3.018906  0.64406637 -0.48503121  1.146881957
#> 47     S03     I05  4.2481028 3.018906  0.64406637 -0.28134948  0.866480336
#> 60     S03     I06  3.3632403 3.018906  0.64406637  0.90264916 -1.202380880
#> 62     S03     I07  4.3790412 3.018906  0.64406637  0.09490194  0.621167275
#> 76     S03     I08  4.8520707 3.018906  0.64406637 -0.47690499  1.666003734
#> 89     S03     I09  1.2900668 3.018906  0.64406637 -0.83001773 -1.542887499
#> 95     S03     I10  2.3192194 3.018906  0.64406637 -0.58787995 -0.755872610
#> 5      S04     I01  1.5825889 3.018906  0.79055178 -1.21967536 -1.007193099
#> 19     S04     I02  2.4443221 3.018906  0.79055178 -0.38246272 -0.982672609
#> 21     S04     I03  5.0232108 3.018906  0.79055178 -1.29494118  2.508694612
#> 34     S04     I04  3.0704525 3.018906  0.79055178 -0.48503121 -0.253973664
#> 48     S04     I05  2.3975350 3.018906  0.79055178 -0.28134948 -1.130572954
#> 51     S04     I06  4.7524339 3.018906  0.79055178  0.90264916  0.040327356
#> 63     S04     I07  3.7075309 3.018906  0.79055178  0.09490194 -0.196828445
#> 77     S04     I08  4.9550084 3.018906  0.79055178 -0.47690499  1.622456038
#> 90     S04     I09  2.3867455 3.018906  0.79055178 -0.83001773 -0.592694159
#> 92     S04     I10  2.2743951 3.018906  0.79055178 -0.58787995 -0.947182376
#> 2      S05     I01  1.7927658 3.018906 -0.77680283 -1.21967536  0.770338411
#> 16     S05     I02  0.4242923 3.018906 -0.77680283 -0.38246272 -1.435347781
#> 22     S05     I03  0.6968630 3.018906 -0.77680283 -1.29494118 -0.250298599
#> 35     S05     I04  3.9907966 3.018906 -0.77680283 -0.48503121  2.233725060
#> 45     S05     I05  0.8699793 3.018906 -0.77680283 -0.28134948 -1.090773974
#> 52     S05     I06  3.0662796 3.018906 -0.77680283  0.90264916 -0.078472388
#> 64     S05     I07  2.4541584 3.018906 -0.77680283  0.09490194  0.117153701
#> 74     S05     I08  0.4298469 3.018906 -0.77680283 -0.47690499 -1.335350888
#> 81     S05     I09  1.7265737 3.018906 -0.77680283 -0.83001773  0.314488688
#> 93     S05     I10  1.5889772 3.018906 -0.77680283 -0.58787995 -0.065245679
#> 8      S06     I01 -0.2484586 3.018906 -0.70048995 -1.21967536 -1.347198953
#> 20     S06     I02  1.9589297 3.018906 -0.70048995 -0.38246272  0.022976766
#> 27     S06     I03  1.3580806 3.018906 -0.70048995 -1.29494118  0.334606088
#> 37     S06     I04  2.0874092 3.018906 -0.70048995 -0.48503121  0.254024736
#> 49     S06     I05  1.1549877 3.018906 -0.70048995 -0.28134948 -0.882078478
#> 56     S06     I06  3.6712528 3.018906 -0.70048995  0.90264916  0.450187938
#> 66     S06     I07  2.3303001 3.018906 -0.70048995  0.09490194 -0.083017454
#> 79     S06     I08  1.1409390 3.018906 -0.70048995 -0.47690499 -0.700571681
#> 85     S06     I09  0.7042130 3.018906 -0.70048995 -0.83001773 -0.784184966
#> 99     S06     I10  1.2336682 3.018906 -0.70048995 -0.58787995 -0.496867533
#> 9      S07     I01  2.9574803 3.018906  1.67411697 -1.21967536 -0.515866904
#> 11     S07     I02  4.6244990 3.018906  1.67411697 -0.38246272  0.313939183
#> 24     S07     I03  3.2574777 3.018906  1.67411697 -1.29494118 -0.140603725
#> 38     S07     I04  2.5115776 3.018906  1.67411697 -0.48503121 -1.696413823
#> 50     S07     I05  5.1166457 3.018906  1.67411697 -0.28134948  0.704972580
#> 53     S07     I06  4.6588424 3.018906  1.67411697  0.90264916 -0.936829372
#> 67     S07     I07  4.2697964 3.018906  1.67411697  0.09490194 -0.518128165
#> 80     S07     I08  3.4482387 3.018906  1.67411697 -0.47690499 -0.767878848
#> 82     S07     I09  3.5352294 3.018906  1.67411697 -0.83001773 -0.327775483
#> 96     S07     I10  3.0749247 3.018906  1.67411697 -0.58787995 -1.030217910
#> 6      S08     I01  2.3700222 3.018906  0.77409127 -1.21967536 -0.203299372
#> 12     S08     I02  3.4696960 3.018906  0.77409127 -0.38246272  0.059161856
#> 25     S08     I03  1.7139378 3.018906  0.77409127 -1.29494118 -0.784117939
#> 39     S08     I04  2.4024263 3.018906  0.77409127 -0.48503121 -0.905539399
#> 41     S08     I05  3.7703035 3.018906  0.77409127 -0.28134948  0.258656077
#> 54     S08     I06  5.3963779 3.018906  0.77409127  0.90264916  0.700731843
#> 68     S08     I07  3.0303924 3.018906  0.77409127  0.09490194 -0.857506409
#> 78     S08     I08  3.1234286 3.018906  0.77409127 -0.47690499 -0.192663325
#> 83     S08     I09  1.5777418 3.018906  0.77409127 -0.83001773 -1.385237387
#> 97     S08     I10  2.5212751 3.018906  0.77409127 -0.58787995 -0.683841836
#> 3      S09     I01  2.7146530 3.018906  0.06184517 -1.21967536  0.853577531
#> 13     S09     I02  3.8269547 3.018906  0.06184517 -0.38246272  1.128666668
#> 26     S09     I03  2.4820110 3.018906  0.06184517 -1.29494118  0.696201438
#> 32     S09     I04  1.7155503 3.018906  0.06184517 -0.48503121 -0.880169300
#> 42     S09     I05  2.9736371 3.018906  0.06184517 -0.28134948  0.174235781
#> 55     S09     I06  4.7984257 3.018906  0.06184517  0.90264916  0.815025797
#> 65     S09     I07  3.8727596 3.018906  0.06184517  0.09490194  0.697106910
#> 71     S09     I08  4.4005973 3.018906  0.06184517 -0.47690499  1.796751505
#> 84     S09     I09  2.4519822 3.018906  0.06184517 -0.83001773  0.201249175
#> 94     S09     I10  2.2404821 3.018906  0.06184517 -0.58787995 -0.252388722
#> 10     S10     I01  4.7014895 3.018906  0.60409975 -1.21967536  2.298159463
#> 14     S10     I02  3.8896310 3.018906  0.60409975 -0.38246272  0.649088363
#> 28     S10     I03  3.2515243 3.018906  0.60409975 -1.29494118  0.923460129
#> 40     S10     I04  2.0571903 3.018906  0.60409975 -0.48503121 -1.080783848
#> 43     S10     I05  5.0324876 3.018906  0.60409975 -0.28134948  1.690831725
#> 57     S10     I06  3.3904591 3.018906  0.60409975  0.90264916 -1.135195439
#> 70     S10     I07  3.7129863 3.018906  0.60409975  0.09490194 -0.004920999
#> 72     S10     I08  4.2578487 3.018906  0.60409975 -0.47690499  1.111748311
#> 86     S10     I09  1.7970834 3.018906  0.60409975 -0.83001773 -0.995904271
#> 100    S10     I10  2.4720447 3.018906  0.60409975 -0.58787995 -0.563080739