Generate a mixed design from existing data

sim_mixed_df() produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.

Usage

sim_mixed_df(
  data,
  sub_n = NULL,
  item_n = NULL,
  dv = "y",
  sub_id = "sub_id",
  item_id = "item_id"
)

Arguments

data

the existing tbl

sub_n

the number of subjects to simulate (if NULL, returns data for the same subjects)

item_n

the number of items to simulate (if NULL, returns data for the same items)

dv

the column name or index containing the DV

sub_id

the column name or index for the subject IDs

item_id

the column name or index for the item IDs

Value

a tbl

Examples

sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")
#>     sub_id item_id            y  grand_i       sub_i     item_i         err
#> 1      S01     I01  1.795239135 3.018906 -1.88866619  0.4766657  0.18833405
#> 15     S01     I02  3.385124427 3.018906 -1.88866619  0.8280414  1.42684358
#> 29     S01     I03 -0.289793063 3.018906 -1.88866619 -0.5095749 -0.91045756
#> 31     S01     I04  1.764204819 3.018906 -1.88866619 -0.6299344  1.26389979
#> 44     S01     I05  1.231894607 3.018906 -1.88866619 -0.2453519  0.34700706
#> 58     S01     I06  2.310078407 3.018906 -1.88866619 -0.1193556  1.29919462
#> 61     S01     I07  0.300352480 3.018906 -1.88866619 -1.1400340  0.31014710
#> 73     S01     I08  0.408384240 3.018906 -1.88866619  0.2595173 -0.98137249
#> 87     S01     I09 -1.789524779 3.018906 -1.88866619 -0.9965762 -1.92318798
#> 91     S01     I10  1.116021054 3.018906 -1.88866619  0.3965225 -0.41074085
#> 7      S02     I01  3.126378763 3.018906 -0.01999370  0.4766657 -0.34919881
#> 17     S02     I02  2.325976210 3.018906 -0.01999370  0.8280414 -1.50097713
#> 30     S02     I03  1.604919934 3.018906 -0.01999370 -0.5095749 -0.88441705
#> 36     S02     I04  4.985919644 3.018906 -0.01999370 -0.6299344  2.61694212
#> 46     S02     I05  2.588221156 3.018906 -0.01999370 -0.2453519 -0.16533889
#> 59     S02     I06  2.462063727 3.018906 -0.01999370 -0.1193556 -0.41749255
#> 69     S02     I07  1.346386258 3.018906 -0.01999370 -1.1400340 -0.51249161
#> 75     S02     I08  3.586171246 3.018906 -0.01999370  0.2595173  0.32774203
#> 88     S02     I09  4.326847496 3.018906 -0.01999370 -0.9965762  2.32451180
#> 98     S02     I10  3.895943848 3.018906 -0.01999370  0.3965225  0.50050945
#> 4      S03     I01  2.256613987 3.018906 -1.12490273  0.4766657 -0.11405456
#> 18     S03     I02  2.271752990 3.018906 -1.12490273  0.8280414 -0.45029132
#> 23     S03     I03  0.411675456 3.018906 -1.12490273 -0.5095749 -0.97275250
#> 33     S03     I04  1.588805583 3.018906 -1.12490273 -0.6299344  0.32473709
#> 47     S03     I05  1.084992421 3.018906 -1.12490273 -0.2453519 -0.56365859
#> 60     S03     I06  2.249038794 3.018906 -1.12490273 -0.1193556  0.47439155
#> 62     S03     I07 -1.738168502 3.018906 -1.12490273 -1.1400340 -2.49213735
#> 76     S03     I08  2.482240717 3.018906 -1.12490273  0.2595173  0.32872053
#> 89     S03     I09  1.277717914 3.018906 -1.12490273 -0.9965762  0.38029125
#> 95     S03     I10  2.415696564 3.018906 -1.12490273  0.3965225  0.12517119
#> 5      S04     I01  5.835511094 3.018906 -0.06805038  0.4766657  2.40799019
#> 19     S04     I02  3.633256510 3.018906 -0.06805038  0.8280414 -0.14564015
#> 21     S04     I03  5.175198779 3.018906 -0.06805038 -0.5095749  2.73391847
#> 34     S04     I04  3.557086113 3.018906 -0.06805038 -0.6299344  1.23616527
#> 48     S04     I05  3.516585778 3.018906 -0.06805038 -0.2453519  0.81108241
#> 51     S04     I06  2.752546889 3.018906 -0.06805038 -0.1193556 -0.07895271
#> 63     S04     I07  1.095534606 3.018906 -0.06805038 -1.1400340 -0.71528659
#> 77     S04     I08  4.168658357 3.018906 -0.06805038  0.2595173  0.95828582
#> 90     S04     I09  1.424271924 3.018906 -0.06805038 -0.9965762 -0.53000709
#> 92     S04     I10  6.841363966 3.018906 -0.06805038  0.3965225  3.49398624
#> 2      S05     I01  4.660223902 3.018906  1.45271316  0.4766657 -0.28806053
#> 16     S05     I02  5.669695065 3.018906  1.45271316  0.8280414  0.37003487
#> 22     S05     I03  4.995457790 3.018906  1.45271316 -0.5095749  1.03341395
#> 35     S05     I04  7.328064691 3.018906  1.45271316 -0.6299344  3.48638031
#> 45     S05     I05  2.320683615 3.018906  1.45271316 -0.2453519 -1.90558329
#> 52     S05     I06  4.194267100 3.018906  1.45271316 -0.1193556 -0.15799603
#> 64     S05     I07  3.567232330 3.018906  1.45271316 -1.1400340  0.23564760
#> 74     S05     I08  6.213466846 3.018906  1.45271316  0.2595173  1.48233077
#> 81     S05     I09  3.774613442 3.018906  1.45271316 -0.9965762  0.29957089
#> 93     S05     I10  4.739126529 3.018906  1.45271316  0.3965225 -0.12901473
#> 8      S06     I01  5.583757342 3.018906  1.17598315  0.4766657  0.91220291
#> 20     S06     I02  2.993617949 3.018906  1.17598315  0.8280414 -2.02931224
#> 27     S06     I03  3.621804582 3.018906  1.17598315 -0.5095749 -0.06350925
#> 37     S06     I04  4.160574616 3.018906  1.17598315 -0.6299344  0.59562024
#> 49     S06     I05  4.965155650 3.018906  1.17598315 -0.2453519  1.01561876
#> 56     S06     I06  3.439335981 3.018906  1.17598315 -0.1193556 -0.63619715
#> 66     S06     I07  4.078942442 3.018906  1.17598315 -1.1400340  1.02408772
#> 79     S06     I08  5.508741295 3.018906  1.17598315  0.2595173  1.05433523
#> 85     S06     I09  4.337096790 3.018906  1.17598315 -0.9965762  1.13878424
#> 99     S06     I10  5.403573855 3.018906  1.17598315  0.3965225  0.81216261
#> 9      S07     I01  3.862369611 3.018906  0.44967163  0.4766657 -0.08287330
#> 11     S07     I02  4.239752965 3.018906  0.44967163  0.8280414 -0.05686570
#> 24     S07     I03  3.803692583 3.018906  0.44967163 -0.5095749  0.84469027
#> 38     S07     I04  2.950187897 3.018906  0.44967163 -0.6299344  0.11154505
#> 50     S07     I05  2.535038025 3.018906  0.44967163 -0.2453519 -0.68818735
#> 53     S07     I06  0.129758020 3.018906  0.44967163 -0.1193556 -3.21946359
#> 67     S07     I07  3.207201002 3.018906  0.44967163 -1.1400340  0.87865780
#> 80     S07     I08  2.389105442 3.018906  0.44967163  0.2595173 -1.33898911
#> 82     S07     I09  2.024036700 3.018906  0.44967163 -0.9965762 -0.44796433
#> 96     S07     I10  3.275401614 3.018906  0.44967163  0.3965225 -0.58969812
#> 6      S08     I01  3.626639613 3.018906 -0.91100928  0.4766657  1.04207762
#> 12     S08     I02  3.359845546 3.018906 -0.91100928  0.8280414  0.42390779
#> 25     S08     I03  0.031271577 3.018906 -0.91100928 -0.5095749 -1.56704983
#> 39     S08     I04  1.750341318 3.018906 -0.91100928 -0.6299344  0.27237938
#> 41     S08     I05 -0.003263509 3.018906 -0.91100928 -0.2453519 -1.86580797
#> 54     S08     I06  0.643213891 3.018906 -0.91100928 -0.1193556 -1.34532680
#> 68     S08     I07  1.528201892 3.018906 -0.91100928 -1.1400340  0.56033960
#> 78     S08     I08  2.790771830 3.018906 -0.91100928  0.2595173  0.42335820
#> 83     S08     I09  0.267607628 3.018906 -0.91100928 -0.9965762 -0.84371248
#> 97     S08     I10  1.059291958 3.018906 -0.91100928  0.3965225 -1.44512686
#> 3      S09     I01  2.539404029 3.018906 -1.14838769  0.4766657  0.19222044
#> 13     S09     I02  2.004751767 3.018906 -1.14838769  0.8280414 -0.69380758
#> 26     S09     I03  0.613069285 3.018906 -1.14838769 -0.5095749 -0.74787371
#> 32     S09     I04  0.307863710 3.018906 -1.14838769 -0.6299344 -0.93271982
#> 42     S09     I05  2.768805059 3.018906 -1.14838769 -0.2453519  1.14363901
#> 55     S09     I06  0.535342137 3.018906 -1.14838769 -0.1193556 -1.21582015
#> 65     S09     I07 -0.714388274 3.018906 -1.14838769 -1.1400340 -1.44487216
#> 71     S09     I08  2.040056437 3.018906 -1.14838769  0.2595173 -0.08997879
#> 84     S09     I09  0.736768844 3.018906 -1.14838769 -0.9965762 -0.13717286
#> 94     S09     I10  0.818846698 3.018906 -1.14838769  0.3965225 -1.44819371
#> 10     S10     I01  5.306068299 3.018906  1.04341572  0.4766657  0.76708130
#> 14     S10     I02  3.972546911 3.018906  1.04341572  0.8280414 -0.91781585
#> 28     S10     I03  5.077634202 3.018906  1.04341572 -0.5095749  1.52488779
#> 40     S10     I04  2.787574127 3.018906  1.04341572 -0.6299344 -0.64481282
#> 43     S10     I05  6.528302349 3.018906  1.04341572 -0.2453519  2.71133288
#> 57     S10     I06  4.415610034 3.018906  1.04341572 -0.1193556  0.47264433
#> 70     S10     I07  1.965674999 3.018906  1.04341572 -1.1400340 -0.95661230
#> 72     S10     I08  5.486316125 3.018906  1.04341572  0.2595173  1.16447748
#> 86     S10     I09  2.859077358 3.018906  1.04341572 -0.9965762 -0.20666776
#> 100    S10     I10  4.366894266 3.018906  1.04341572  0.3965225 -0.09194956