Generate a sample with random intercepts for subjects and items

simdf_mixed Produces a dataframe with the same distributions of by-subject and by-item random intercepts as an existing dataframe

simdf_mixed(dat, sub_n = 100, item_n = 25, dv = 1, sub_id = 2,
  item_id = 3)

Arguments

dat
the existing dataframe

sub_n
the number of subjects to simulate

item_n
the number of items to simulate

dv
the column name or index containing the DV

sub_id
the column name or index for the subject IDs

item_id
the column name or index for the item IDs

Value

tibble

Examples

simdf_mixed(faceratings, 10, 10, "rating", "rater_id", "face_id")

#>     sub_id item_id      sub_i      item_i         dv
#> 1        1       1  0.7171917 -0.16440417  2.7076926
#> 2        2       1  1.5872296 -0.16440417  4.0497225
#> 3        3       1 -0.7289360 -0.16440417  2.8124829
#> 4        4       1  0.6003483 -0.16440417  1.6454961
#> 5        5       1  0.5163209 -0.16440417  2.8399589
#> 6        6       1  0.2563833 -0.16440417  3.0883100
#> 7        7       1 -0.7781445 -0.16440417  3.0922434
#> 8        8       1  1.1424562 -0.16440417  3.7438529
#> 9        9       1  0.1711736 -0.16440417  1.9164980
#> 10      10       1  1.0828005 -0.16440417  2.1094020
#> 11       1       2  0.7171917  0.37636537  6.5942167
#> 12       2       2  1.5872296  0.37636537  3.8486785
#> 13       3       2 -0.7289360  0.37636537  3.4411992
#> 14       4       2  0.6003483  0.37636537  5.3235771
#> 15       5       2  0.5163209  0.37636537  4.2771207
#> 16       6       2  0.2563833  0.37636537  3.3479907
#> 17       7       2 -0.7781445  0.37636537  2.7002602
#> 18       8       2  1.1424562  0.37636537  4.1520841
#> 19       9       2  0.1711736  0.37636537  5.1249236
#> 20      10       2  1.0828005  0.37636537  4.4809776
#> 21       1       3  0.7171917 -0.82422093  3.2861012
#> 22       2       3  1.5872296 -0.82422093  2.8786693
#> 23       3       3 -0.7289360 -0.82422093  1.9902052
#> 24       4       3  0.6003483 -0.82422093  2.0458923
#> 25       5       3  0.5163209 -0.82422093  1.5216236
#> 26       6       3  0.2563833 -0.82422093  3.3659935
#> 27       7       3 -0.7781445 -0.82422093  2.9983831
#> 28       8       3  1.1424562 -0.82422093  3.4864340
#> 29       9       3  0.1711736 -0.82422093  2.7199936
#> 30      10       3  1.0828005 -0.82422093  4.6086848
#> 31       1       4  0.7171917 -0.91128136  1.6789966
#> 32       2       4  1.5872296 -0.91128136  5.6968101
#> 33       3       4 -0.7289360 -0.91128136 -1.0487718
#> 34       4       4  0.6003483 -0.91128136  2.0597989
#> 35       5       4  0.5163209 -0.91128136  2.3732315
#> 36       6       4  0.2563833 -0.91128136  4.2745678
#> 37       7       4 -0.7781445 -0.91128136  1.8469632
#> 38       8       4  1.1424562 -0.91128136  4.1014591
#> 39       9       4  0.1711736 -0.91128136  1.0318854
#> 40      10       4  1.0828005 -0.91128136  4.2240532
#> 41       1       5  0.7171917  1.66678147  5.0273318
#> 42       2       5  1.5872296  1.66678147  3.5821303
#> 43       3       5 -0.7289360  1.66678147  5.4538830
#> 44       4       5  0.6003483  1.66678147  7.9807309
#> 45       5       5  0.5163209  1.66678147  3.7473337
#> 46       6       5  0.2563833  1.66678147  7.1439931
#> 47       7       5 -0.7781445  1.66678147  4.8426279
#> 48       8       5  1.1424562  1.66678147  6.6205912
#> 49       9       5  0.1711736  1.66678147  6.7093920
#> 50      10       5  1.0828005  1.66678147  4.3002833
#> 51       1       6  0.7171917 -0.32253158  3.5070761
#> 52       2       6  1.5872296 -0.32253158  4.5057448
#> 53       3       6 -0.7289360 -0.32253158  0.9024044
#> 54       4       6  0.6003483 -0.32253158  3.9435972
#> 55       5       6  0.5163209 -0.32253158  4.3432880
#> 56       6       6  0.2563833 -0.32253158  1.5616119
#> 57       7       6 -0.7781445 -0.32253158  1.5701492
#> 58       8       6  1.1424562 -0.32253158  1.8015241
#> 59       9       6  0.1711736 -0.32253158  2.9668905
#> 60      10       6  1.0828005 -0.32253158  4.5541003
#> 61       1       7  0.7171917  0.06766548  3.6839723
#> 62       2       7  1.5872296  0.06766548  3.5298722
#> 63       3       7 -0.7289360  0.06766548  1.3879482
#> 64       4       7  0.6003483  0.06766548  3.0134067
#> 65       5       7  0.5163209  0.06766548  3.8805034
#> 66       6       7  0.2563833  0.06766548  2.8650564
#> 67       7       7 -0.7781445  0.06766548  3.9055132
#> 68       8       7  1.1424562  0.06766548  5.1346467
#> 69       9       7  0.1711736  0.06766548  2.2637057
#> 70      10       7  1.0828005  0.06766548  3.8728360
#> 71       1       8  0.7171917 -0.41971766  4.1334290
#> 72       2       8  1.5872296 -0.41971766  4.1028115
#> 73       3       8 -0.7289360 -0.41971766  2.0302724
#> 74       4       8  0.6003483 -0.41971766  2.8500859
#> 75       5       8  0.5163209 -0.41971766  1.2627937
#> 76       6       8  0.2563833 -0.41971766  2.0276961
#> 77       7       8 -0.7781445 -0.41971766  0.9953581
#> 78       8       8  1.1424562 -0.41971766  4.0599506
#> 79       9       8  0.1711736 -0.41971766  3.1317760
#> 80      10       8  1.0828005 -0.41971766  3.4296845
#> 81       1       9  0.7171917 -0.09253167  2.7561118
#> 82       2       9  1.5872296 -0.09253167  4.7486546
#> 83       3       9 -0.7289360 -0.09253167  0.9585535
#> 84       4       9  0.6003483 -0.09253167  3.6791317
#> 85       5       9  0.5163209 -0.09253167  2.7224103
#> 86       6       9  0.2563833 -0.09253167  2.6014336
#> 87       7       9 -0.7781445 -0.09253167  0.6113194
#> 88       8       9  1.1424562 -0.09253167  5.3929322
#> 89       9       9  0.1711736 -0.09253167  3.5604510
#> 90      10       9  1.0828005 -0.09253167  3.8776255
#> 91       1      10  0.7171917  1.38999715  5.8347359
#> 92       2      10  1.5872296  1.38999715  6.5293892
#> 93       3      10 -0.7289360  1.38999715  4.4470914
#> 94       4      10  0.6003483  1.38999715  5.8856564
#> 95       5      10  0.5163209  1.38999715  4.7294257
#> 96       6      10  0.2563833  1.38999715  5.8323700
#> 97       7      10 -0.7781445  1.38999715  3.5409873
#> 98       8      10  1.1424562  1.38999715  6.7181566
#> 99       9      10  0.1711736  1.38999715  4.5778140
#> 100     10      10  1.0828005  1.38999715  4.5562784