Convert data from wide to long format

Convert data from wide to long format

wide2long(
  data,
  within_factors = c(),
  within_cols = c(),
  dv = "y",
  id = "id",
  sep = faux_options("sep")
)

Arguments

data
the tbl in wide format

within_factors
the names of the within factors

within_cols
the names (or indices) of the within-subject (value) columns

dv
the name of the dv column (defaults to "y")

id
the name of the ID column(s) if they don't exist, a new column will be made (defaults to ("id")

sep
separator for within-columns (to be used in strsplit, so can be regex), defaults to "_"

Value

a tbl in long format

Examples

wide2long(iris, c("Feature", "Measure"), 1:4, sep = "\\.")

#>          id    Species Feature Measure   y
#> S001.1 S001     setosa   Sepal  Length 5.1
#> S002.1 S002     setosa   Sepal  Length 4.9
#> S003.1 S003     setosa   Sepal  Length 4.7
#> S004.1 S004     setosa   Sepal  Length 4.6
#> S005.1 S005     setosa   Sepal  Length 5.0
#> S006.1 S006     setosa   Sepal  Length 5.4
#> S007.1 S007     setosa   Sepal  Length 4.6
#> S008.1 S008     setosa   Sepal  Length 5.0
#> S009.1 S009     setosa   Sepal  Length 4.4
#> S010.1 S010     setosa   Sepal  Length 4.9
#> S011.1 S011     setosa   Sepal  Length 5.4
#> S012.1 S012     setosa   Sepal  Length 4.8
#> S013.1 S013     setosa   Sepal  Length 4.8
#> S014.1 S014     setosa   Sepal  Length 4.3
#> S015.1 S015     setosa   Sepal  Length 5.8
#> S016.1 S016     setosa   Sepal  Length 5.7
#> S017.1 S017     setosa   Sepal  Length 5.4
#> S018.1 S018     setosa   Sepal  Length 5.1
#> S019.1 S019     setosa   Sepal  Length 5.7
#> S020.1 S020     setosa   Sepal  Length 5.1
#> S021.1 S021     setosa   Sepal  Length 5.4
#> S022.1 S022     setosa   Sepal  Length 5.1
#> S023.1 S023     setosa   Sepal  Length 4.6
#> S024.1 S024     setosa   Sepal  Length 5.1
#> S025.1 S025     setosa   Sepal  Length 4.8
#> S026.1 S026     setosa   Sepal  Length 5.0
#> S027.1 S027     setosa   Sepal  Length 5.0
#> S028.1 S028     setosa   Sepal  Length 5.2
#> S029.1 S029     setosa   Sepal  Length 5.2
#> S030.1 S030     setosa   Sepal  Length 4.7
#> S031.1 S031     setosa   Sepal  Length 4.8
#> S032.1 S032     setosa   Sepal  Length 5.4
#> S033.1 S033     setosa   Sepal  Length 5.2
#> S034.1 S034     setosa   Sepal  Length 5.5
#> S035.1 S035     setosa   Sepal  Length 4.9
#> S036.1 S036     setosa   Sepal  Length 5.0
#> S037.1 S037     setosa   Sepal  Length 5.5
#> S038.1 S038     setosa   Sepal  Length 4.9
#> S039.1 S039     setosa   Sepal  Length 4.4
#> S040.1 S040     setosa   Sepal  Length 5.1
#> S041.1 S041     setosa   Sepal  Length 5.0
#> S042.1 S042     setosa   Sepal  Length 4.5
#> S043.1 S043     setosa   Sepal  Length 4.4
#> S044.1 S044     setosa   Sepal  Length 5.0
#> S045.1 S045     setosa   Sepal  Length 5.1
#> S046.1 S046     setosa   Sepal  Length 4.8
#> S047.1 S047     setosa   Sepal  Length 5.1
#> S048.1 S048     setosa   Sepal  Length 4.6
#> S049.1 S049     setosa   Sepal  Length 5.3
#> S050.1 S050     setosa   Sepal  Length 5.0
#> S051.1 S051 versicolor   Sepal  Length 7.0
#> S052.1 S052 versicolor   Sepal  Length 6.4
#> S053.1 S053 versicolor   Sepal  Length 6.9
#> S054.1 S054 versicolor   Sepal  Length 5.5
#> S055.1 S055 versicolor   Sepal  Length 6.5
#> S056.1 S056 versicolor   Sepal  Length 5.7
#> S057.1 S057 versicolor   Sepal  Length 6.3
#> S058.1 S058 versicolor   Sepal  Length 4.9
#> S059.1 S059 versicolor   Sepal  Length 6.6
#> S060.1 S060 versicolor   Sepal  Length 5.2
#> S061.1 S061 versicolor   Sepal  Length 5.0
#> S062.1 S062 versicolor   Sepal  Length 5.9
#> S063.1 S063 versicolor   Sepal  Length 6.0
#> S064.1 S064 versicolor   Sepal  Length 6.1
#> S065.1 S065 versicolor   Sepal  Length 5.6
#> S066.1 S066 versicolor   Sepal  Length 6.7
#> S067.1 S067 versicolor   Sepal  Length 5.6
#> S068.1 S068 versicolor   Sepal  Length 5.8
#> S069.1 S069 versicolor   Sepal  Length 6.2
#> S070.1 S070 versicolor   Sepal  Length 5.6
#> S071.1 S071 versicolor   Sepal  Length 5.9
#> S072.1 S072 versicolor   Sepal  Length 6.1
#> S073.1 S073 versicolor   Sepal  Length 6.3
#> S074.1 S074 versicolor   Sepal  Length 6.1
#> S075.1 S075 versicolor   Sepal  Length 6.4
#> S076.1 S076 versicolor   Sepal  Length 6.6
#> S077.1 S077 versicolor   Sepal  Length 6.8
#> S078.1 S078 versicolor   Sepal  Length 6.7
#> S079.1 S079 versicolor   Sepal  Length 6.0
#> S080.1 S080 versicolor   Sepal  Length 5.7
#> S081.1 S081 versicolor   Sepal  Length 5.5
#> S082.1 S082 versicolor   Sepal  Length 5.5
#> S083.1 S083 versicolor   Sepal  Length 5.8
#> S084.1 S084 versicolor   Sepal  Length 6.0
#> S085.1 S085 versicolor   Sepal  Length 5.4
#> S086.1 S086 versicolor   Sepal  Length 6.0
#> S087.1 S087 versicolor   Sepal  Length 6.7
#> S088.1 S088 versicolor   Sepal  Length 6.3
#> S089.1 S089 versicolor   Sepal  Length 5.6
#> S090.1 S090 versicolor   Sepal  Length 5.5
#> S091.1 S091 versicolor   Sepal  Length 5.5
#> S092.1 S092 versicolor   Sepal  Length 6.1
#> S093.1 S093 versicolor   Sepal  Length 5.8
#> S094.1 S094 versicolor   Sepal  Length 5.0
#> S095.1 S095 versicolor   Sepal  Length 5.6
#> S096.1 S096 versicolor   Sepal  Length 5.7
#> S097.1 S097 versicolor   Sepal  Length 5.7
#> S098.1 S098 versicolor   Sepal  Length 6.2
#> S099.1 S099 versicolor   Sepal  Length 5.1
#> S100.1 S100 versicolor   Sepal  Length 5.7
#> S101.1 S101  virginica   Sepal  Length 6.3
#> S102.1 S102  virginica   Sepal  Length 5.8
#> S103.1 S103  virginica   Sepal  Length 7.1
#> S104.1 S104  virginica   Sepal  Length 6.3
#> S105.1 S105  virginica   Sepal  Length 6.5
#> S106.1 S106  virginica   Sepal  Length 7.6
#> S107.1 S107  virginica   Sepal  Length 4.9
#> S108.1 S108  virginica   Sepal  Length 7.3
#> S109.1 S109  virginica   Sepal  Length 6.7
#> S110.1 S110  virginica   Sepal  Length 7.2
#> S111.1 S111  virginica   Sepal  Length 6.5
#> S112.1 S112  virginica   Sepal  Length 6.4
#> S113.1 S113  virginica   Sepal  Length 6.8
#> S114.1 S114  virginica   Sepal  Length 5.7
#> S115.1 S115  virginica   Sepal  Length 5.8
#> S116.1 S116  virginica   Sepal  Length 6.4
#> S117.1 S117  virginica   Sepal  Length 6.5
#> S118.1 S118  virginica   Sepal  Length 7.7
#> S119.1 S119  virginica   Sepal  Length 7.7
#> S120.1 S120  virginica   Sepal  Length 6.0
#> S121.1 S121  virginica   Sepal  Length 6.9
#> S122.1 S122  virginica   Sepal  Length 5.6
#> S123.1 S123  virginica   Sepal  Length 7.7
#> S124.1 S124  virginica   Sepal  Length 6.3
#> S125.1 S125  virginica   Sepal  Length 6.7
#> S126.1 S126  virginica   Sepal  Length 7.2
#> S127.1 S127  virginica   Sepal  Length 6.2
#> S128.1 S128  virginica   Sepal  Length 6.1
#> S129.1 S129  virginica   Sepal  Length 6.4
#> S130.1 S130  virginica   Sepal  Length 7.2
#> S131.1 S131  virginica   Sepal  Length 7.4
#> S132.1 S132  virginica   Sepal  Length 7.9
#> S133.1 S133  virginica   Sepal  Length 6.4
#> S134.1 S134  virginica   Sepal  Length 6.3
#> S135.1 S135  virginica   Sepal  Length 6.1
#> S136.1 S136  virginica   Sepal  Length 7.7
#> S137.1 S137  virginica   Sepal  Length 6.3
#> S138.1 S138  virginica   Sepal  Length 6.4
#> S139.1 S139  virginica   Sepal  Length 6.0
#> S140.1 S140  virginica   Sepal  Length 6.9
#> S141.1 S141  virginica   Sepal  Length 6.7
#> S142.1 S142  virginica   Sepal  Length 6.9
#> S143.1 S143  virginica   Sepal  Length 5.8
#> S144.1 S144  virginica   Sepal  Length 6.8
#> S145.1 S145  virginica   Sepal  Length 6.7
#> S146.1 S146  virginica   Sepal  Length 6.7
#> S147.1 S147  virginica   Sepal  Length 6.3
#> S148.1 S148  virginica   Sepal  Length 6.5
#> S149.1 S149  virginica   Sepal  Length 6.2
#> S150.1 S150  virginica   Sepal  Length 5.9
#> S001.2 S001     setosa   Sepal   Width 3.5
#> S002.2 S002     setosa   Sepal   Width 3.0
#> S003.2 S003     setosa   Sepal   Width 3.2
#> S004.2 S004     setosa   Sepal   Width 3.1
#> S005.2 S005     setosa   Sepal   Width 3.6
#> S006.2 S006     setosa   Sepal   Width 3.9
#> S007.2 S007     setosa   Sepal   Width 3.4
#> S008.2 S008     setosa   Sepal   Width 3.4
#> S009.2 S009     setosa   Sepal   Width 2.9
#> S010.2 S010     setosa   Sepal   Width 3.1
#> S011.2 S011     setosa   Sepal   Width 3.7
#> S012.2 S012     setosa   Sepal   Width 3.4
#> S013.2 S013     setosa   Sepal   Width 3.0
#> S014.2 S014     setosa   Sepal   Width 3.0
#> S015.2 S015     setosa   Sepal   Width 4.0
#> S016.2 S016     setosa   Sepal   Width 4.4
#> S017.2 S017     setosa   Sepal   Width 3.9
#> S018.2 S018     setosa   Sepal   Width 3.5
#> S019.2 S019     setosa   Sepal   Width 3.8
#> S020.2 S020     setosa   Sepal   Width 3.8
#> S021.2 S021     setosa   Sepal   Width 3.4
#> S022.2 S022     setosa   Sepal   Width 3.7
#> S023.2 S023     setosa   Sepal   Width 3.6
#> S024.2 S024     setosa   Sepal   Width 3.3
#> S025.2 S025     setosa   Sepal   Width 3.4
#> S026.2 S026     setosa   Sepal   Width 3.0
#> S027.2 S027     setosa   Sepal   Width 3.4
#> S028.2 S028     setosa   Sepal   Width 3.5
#> S029.2 S029     setosa   Sepal   Width 3.4
#> S030.2 S030     setosa   Sepal   Width 3.2
#> S031.2 S031     setosa   Sepal   Width 3.1
#> S032.2 S032     setosa   Sepal   Width 3.4
#> S033.2 S033     setosa   Sepal   Width 4.1
#> S034.2 S034     setosa   Sepal   Width 4.2
#> S035.2 S035     setosa   Sepal   Width 3.1
#> S036.2 S036     setosa   Sepal   Width 3.2
#> S037.2 S037     setosa   Sepal   Width 3.5
#> S038.2 S038     setosa   Sepal   Width 3.6
#> S039.2 S039     setosa   Sepal   Width 3.0
#> S040.2 S040     setosa   Sepal   Width 3.4
#> S041.2 S041     setosa   Sepal   Width 3.5
#> S042.2 S042     setosa   Sepal   Width 2.3
#> S043.2 S043     setosa   Sepal   Width 3.2
#> S044.2 S044     setosa   Sepal   Width 3.5
#> S045.2 S045     setosa   Sepal   Width 3.8
#> S046.2 S046     setosa   Sepal   Width 3.0
#> S047.2 S047     setosa   Sepal   Width 3.8
#> S048.2 S048     setosa   Sepal   Width 3.2
#> S049.2 S049     setosa   Sepal   Width 3.7
#> S050.2 S050     setosa   Sepal   Width 3.3
#> S051.2 S051 versicolor   Sepal   Width 3.2
#> S052.2 S052 versicolor   Sepal   Width 3.2
#> S053.2 S053 versicolor   Sepal   Width 3.1
#> S054.2 S054 versicolor   Sepal   Width 2.3
#> S055.2 S055 versicolor   Sepal   Width 2.8
#> S056.2 S056 versicolor   Sepal   Width 2.8
#> S057.2 S057 versicolor   Sepal   Width 3.3
#> S058.2 S058 versicolor   Sepal   Width 2.4
#> S059.2 S059 versicolor   Sepal   Width 2.9
#> S060.2 S060 versicolor   Sepal   Width 2.7
#> S061.2 S061 versicolor   Sepal   Width 2.0
#> S062.2 S062 versicolor   Sepal   Width 3.0
#> S063.2 S063 versicolor   Sepal   Width 2.2
#> S064.2 S064 versicolor   Sepal   Width 2.9
#> S065.2 S065 versicolor   Sepal   Width 2.9
#> S066.2 S066 versicolor   Sepal   Width 3.1
#> S067.2 S067 versicolor   Sepal   Width 3.0
#> S068.2 S068 versicolor   Sepal   Width 2.7
#> S069.2 S069 versicolor   Sepal   Width 2.2
#> S070.2 S070 versicolor   Sepal   Width 2.5
#> S071.2 S071 versicolor   Sepal   Width 3.2
#> S072.2 S072 versicolor   Sepal   Width 2.8
#> S073.2 S073 versicolor   Sepal   Width 2.5
#> S074.2 S074 versicolor   Sepal   Width 2.8
#> S075.2 S075 versicolor   Sepal   Width 2.9
#> S076.2 S076 versicolor   Sepal   Width 3.0
#> S077.2 S077 versicolor   Sepal   Width 2.8
#> S078.2 S078 versicolor   Sepal   Width 3.0
#> S079.2 S079 versicolor   Sepal   Width 2.9
#> S080.2 S080 versicolor   Sepal   Width 2.6
#> S081.2 S081 versicolor   Sepal   Width 2.4
#> S082.2 S082 versicolor   Sepal   Width 2.4
#> S083.2 S083 versicolor   Sepal   Width 2.7
#> S084.2 S084 versicolor   Sepal   Width 2.7
#> S085.2 S085 versicolor   Sepal   Width 3.0
#> S086.2 S086 versicolor   Sepal   Width 3.4
#> S087.2 S087 versicolor   Sepal   Width 3.1
#> S088.2 S088 versicolor   Sepal   Width 2.3
#> S089.2 S089 versicolor   Sepal   Width 3.0
#> S090.2 S090 versicolor   Sepal   Width 2.5
#> S091.2 S091 versicolor   Sepal   Width 2.6
#> S092.2 S092 versicolor   Sepal   Width 3.0
#> S093.2 S093 versicolor   Sepal   Width 2.6
#> S094.2 S094 versicolor   Sepal   Width 2.3
#> S095.2 S095 versicolor   Sepal   Width 2.7
#> S096.2 S096 versicolor   Sepal   Width 3.0
#> S097.2 S097 versicolor   Sepal   Width 2.9
#> S098.2 S098 versicolor   Sepal   Width 2.9
#> S099.2 S099 versicolor   Sepal   Width 2.5
#> S100.2 S100 versicolor   Sepal   Width 2.8
#> S101.2 S101  virginica   Sepal   Width 3.3
#> S102.2 S102  virginica   Sepal   Width 2.7
#> S103.2 S103  virginica   Sepal   Width 3.0
#> S104.2 S104  virginica   Sepal   Width 2.9
#> S105.2 S105  virginica   Sepal   Width 3.0
#> S106.2 S106  virginica   Sepal   Width 3.0
#> S107.2 S107  virginica   Sepal   Width 2.5
#> S108.2 S108  virginica   Sepal   Width 2.9
#> S109.2 S109  virginica   Sepal   Width 2.5
#> S110.2 S110  virginica   Sepal   Width 3.6
#> S111.2 S111  virginica   Sepal   Width 3.2
#> S112.2 S112  virginica   Sepal   Width 2.7
#> S113.2 S113  virginica   Sepal   Width 3.0
#> S114.2 S114  virginica   Sepal   Width 2.5
#> S115.2 S115  virginica   Sepal   Width 2.8
#> S116.2 S116  virginica   Sepal   Width 3.2
#> S117.2 S117  virginica   Sepal   Width 3.0
#> S118.2 S118  virginica   Sepal   Width 3.8
#> S119.2 S119  virginica   Sepal   Width 2.6
#> S120.2 S120  virginica   Sepal   Width 2.2
#> S121.2 S121  virginica   Sepal   Width 3.2
#> S122.2 S122  virginica   Sepal   Width 2.8
#> S123.2 S123  virginica   Sepal   Width 2.8
#> S124.2 S124  virginica   Sepal   Width 2.7
#> S125.2 S125  virginica   Sepal   Width 3.3
#> S126.2 S126  virginica   Sepal   Width 3.2
#> S127.2 S127  virginica   Sepal   Width 2.8
#> S128.2 S128  virginica   Sepal   Width 3.0
#> S129.2 S129  virginica   Sepal   Width 2.8
#> S130.2 S130  virginica   Sepal   Width 3.0
#> S131.2 S131  virginica   Sepal   Width 2.8
#> S132.2 S132  virginica   Sepal   Width 3.8
#> S133.2 S133  virginica   Sepal   Width 2.8
#> S134.2 S134  virginica   Sepal   Width 2.8
#> S135.2 S135  virginica   Sepal   Width 2.6
#> S136.2 S136  virginica   Sepal   Width 3.0
#> S137.2 S137  virginica   Sepal   Width 3.4
#> S138.2 S138  virginica   Sepal   Width 3.1
#> S139.2 S139  virginica   Sepal   Width 3.0
#> S140.2 S140  virginica   Sepal   Width 3.1
#> S141.2 S141  virginica   Sepal   Width 3.1
#> S142.2 S142  virginica   Sepal   Width 3.1
#> S143.2 S143  virginica   Sepal   Width 2.7
#> S144.2 S144  virginica   Sepal   Width 3.2
#> S145.2 S145  virginica   Sepal   Width 3.3
#> S146.2 S146  virginica   Sepal   Width 3.0
#> S147.2 S147  virginica   Sepal   Width 2.5
#> S148.2 S148  virginica   Sepal   Width 3.0
#> S149.2 S149  virginica   Sepal   Width 3.4
#> S150.2 S150  virginica   Sepal   Width 3.0
#> S001.3 S001     setosa   Petal  Length 1.4
#> S002.3 S002     setosa   Petal  Length 1.4
#> S003.3 S003     setosa   Petal  Length 1.3
#> S004.3 S004     setosa   Petal  Length 1.5
#> S005.3 S005     setosa   Petal  Length 1.4
#> S006.3 S006     setosa   Petal  Length 1.7
#> S007.3 S007     setosa   Petal  Length 1.4
#> S008.3 S008     setosa   Petal  Length 1.5
#> S009.3 S009     setosa   Petal  Length 1.4
#> S010.3 S010     setosa   Petal  Length 1.5
#> S011.3 S011     setosa   Petal  Length 1.5
#> S012.3 S012     setosa   Petal  Length 1.6
#> S013.3 S013     setosa   Petal  Length 1.4
#> S014.3 S014     setosa   Petal  Length 1.1
#> S015.3 S015     setosa   Petal  Length 1.2
#> S016.3 S016     setosa   Petal  Length 1.5
#> S017.3 S017     setosa   Petal  Length 1.3
#> S018.3 S018     setosa   Petal  Length 1.4
#> S019.3 S019     setosa   Petal  Length 1.7
#> S020.3 S020     setosa   Petal  Length 1.5
#> S021.3 S021     setosa   Petal  Length 1.7
#> S022.3 S022     setosa   Petal  Length 1.5
#> S023.3 S023     setosa   Petal  Length 1.0
#> S024.3 S024     setosa   Petal  Length 1.7
#> S025.3 S025     setosa   Petal  Length 1.9
#> S026.3 S026     setosa   Petal  Length 1.6
#> S027.3 S027     setosa   Petal  Length 1.6
#> S028.3 S028     setosa   Petal  Length 1.5
#> S029.3 S029     setosa   Petal  Length 1.4
#> S030.3 S030     setosa   Petal  Length 1.6
#> S031.3 S031     setosa   Petal  Length 1.6
#> S032.3 S032     setosa   Petal  Length 1.5
#> S033.3 S033     setosa   Petal  Length 1.5
#> S034.3 S034     setosa   Petal  Length 1.4
#> S035.3 S035     setosa   Petal  Length 1.5
#> S036.3 S036     setosa   Petal  Length 1.2
#> S037.3 S037     setosa   Petal  Length 1.3
#> S038.3 S038     setosa   Petal  Length 1.4
#> S039.3 S039     setosa   Petal  Length 1.3
#> S040.3 S040     setosa   Petal  Length 1.5
#> S041.3 S041     setosa   Petal  Length 1.3
#> S042.3 S042     setosa   Petal  Length 1.3
#> S043.3 S043     setosa   Petal  Length 1.3
#> S044.3 S044     setosa   Petal  Length 1.6
#> S045.3 S045     setosa   Petal  Length 1.9
#> S046.3 S046     setosa   Petal  Length 1.4
#> S047.3 S047     setosa   Petal  Length 1.6
#> S048.3 S048     setosa   Petal  Length 1.4
#> S049.3 S049     setosa   Petal  Length 1.5
#> S050.3 S050     setosa   Petal  Length 1.4
#> S051.3 S051 versicolor   Petal  Length 4.7
#> S052.3 S052 versicolor   Petal  Length 4.5
#> S053.3 S053 versicolor   Petal  Length 4.9
#> S054.3 S054 versicolor   Petal  Length 4.0
#> S055.3 S055 versicolor   Petal  Length 4.6
#> S056.3 S056 versicolor   Petal  Length 4.5
#> S057.3 S057 versicolor   Petal  Length 4.7
#> S058.3 S058 versicolor   Petal  Length 3.3
#> S059.3 S059 versicolor   Petal  Length 4.6
#> S060.3 S060 versicolor   Petal  Length 3.9
#> S061.3 S061 versicolor   Petal  Length 3.5
#> S062.3 S062 versicolor   Petal  Length 4.2
#> S063.3 S063 versicolor   Petal  Length 4.0
#> S064.3 S064 versicolor   Petal  Length 4.7
#> S065.3 S065 versicolor   Petal  Length 3.6
#> S066.3 S066 versicolor   Petal  Length 4.4
#> S067.3 S067 versicolor   Petal  Length 4.5
#> S068.3 S068 versicolor   Petal  Length 4.1
#> S069.3 S069 versicolor   Petal  Length 4.5
#> S070.3 S070 versicolor   Petal  Length 3.9
#> S071.3 S071 versicolor   Petal  Length 4.8
#> S072.3 S072 versicolor   Petal  Length 4.0
#> S073.3 S073 versicolor   Petal  Length 4.9
#> S074.3 S074 versicolor   Petal  Length 4.7
#> S075.3 S075 versicolor   Petal  Length 4.3
#> S076.3 S076 versicolor   Petal  Length 4.4
#> S077.3 S077 versicolor   Petal  Length 4.8
#> S078.3 S078 versicolor   Petal  Length 5.0
#> S079.3 S079 versicolor   Petal  Length 4.5
#> S080.3 S080 versicolor   Petal  Length 3.5
#> S081.3 S081 versicolor   Petal  Length 3.8
#> S082.3 S082 versicolor   Petal  Length 3.7
#> S083.3 S083 versicolor   Petal  Length 3.9
#> S084.3 S084 versicolor   Petal  Length 5.1
#> S085.3 S085 versicolor   Petal  Length 4.5
#> S086.3 S086 versicolor   Petal  Length 4.5
#> S087.3 S087 versicolor   Petal  Length 4.7
#> S088.3 S088 versicolor   Petal  Length 4.4
#> S089.3 S089 versicolor   Petal  Length 4.1
#> S090.3 S090 versicolor   Petal  Length 4.0
#> S091.3 S091 versicolor   Petal  Length 4.4
#> S092.3 S092 versicolor   Petal  Length 4.6
#> S093.3 S093 versicolor   Petal  Length 4.0
#> S094.3 S094 versicolor   Petal  Length 3.3
#> S095.3 S095 versicolor   Petal  Length 4.2
#> S096.3 S096 versicolor   Petal  Length 4.2
#> S097.3 S097 versicolor   Petal  Length 4.2
#> S098.3 S098 versicolor   Petal  Length 4.3
#> S099.3 S099 versicolor   Petal  Length 3.0
#> S100.3 S100 versicolor   Petal  Length 4.1
#> S101.3 S101  virginica   Petal  Length 6.0
#> S102.3 S102  virginica   Petal  Length 5.1
#> S103.3 S103  virginica   Petal  Length 5.9
#> S104.3 S104  virginica   Petal  Length 5.6
#> S105.3 S105  virginica   Petal  Length 5.8
#> S106.3 S106  virginica   Petal  Length 6.6
#> S107.3 S107  virginica   Petal  Length 4.5
#> S108.3 S108  virginica   Petal  Length 6.3
#> S109.3 S109  virginica   Petal  Length 5.8
#> S110.3 S110  virginica   Petal  Length 6.1
#> S111.3 S111  virginica   Petal  Length 5.1
#> S112.3 S112  virginica   Petal  Length 5.3
#> S113.3 S113  virginica   Petal  Length 5.5
#> S114.3 S114  virginica   Petal  Length 5.0
#> S115.3 S115  virginica   Petal  Length 5.1
#> S116.3 S116  virginica   Petal  Length 5.3
#> S117.3 S117  virginica   Petal  Length 5.5
#> S118.3 S118  virginica   Petal  Length 6.7
#> S119.3 S119  virginica   Petal  Length 6.9
#> S120.3 S120  virginica   Petal  Length 5.0
#> S121.3 S121  virginica   Petal  Length 5.7
#> S122.3 S122  virginica   Petal  Length 4.9
#> S123.3 S123  virginica   Petal  Length 6.7
#> S124.3 S124  virginica   Petal  Length 4.9
#> S125.3 S125  virginica   Petal  Length 5.7
#> S126.3 S126  virginica   Petal  Length 6.0
#> S127.3 S127  virginica   Petal  Length 4.8
#> S128.3 S128  virginica   Petal  Length 4.9
#> S129.3 S129  virginica   Petal  Length 5.6
#> S130.3 S130  virginica   Petal  Length 5.8
#> S131.3 S131  virginica   Petal  Length 6.1
#> S132.3 S132  virginica   Petal  Length 6.4
#> S133.3 S133  virginica   Petal  Length 5.6
#> S134.3 S134  virginica   Petal  Length 5.1
#> S135.3 S135  virginica   Petal  Length 5.6
#> S136.3 S136  virginica   Petal  Length 6.1
#> S137.3 S137  virginica   Petal  Length 5.6
#> S138.3 S138  virginica   Petal  Length 5.5
#> S139.3 S139  virginica   Petal  Length 4.8
#> S140.3 S140  virginica   Petal  Length 5.4
#> S141.3 S141  virginica   Petal  Length 5.6
#> S142.3 S142  virginica   Petal  Length 5.1
#> S143.3 S143  virginica   Petal  Length 5.1
#> S144.3 S144  virginica   Petal  Length 5.9
#> S145.3 S145  virginica   Petal  Length 5.7
#> S146.3 S146  virginica   Petal  Length 5.2
#> S147.3 S147  virginica   Petal  Length 5.0
#> S148.3 S148  virginica   Petal  Length 5.2
#> S149.3 S149  virginica   Petal  Length 5.4
#> S150.3 S150  virginica   Petal  Length 5.1
#> S001.4 S001     setosa   Petal   Width 0.2
#> S002.4 S002     setosa   Petal   Width 0.2
#> S003.4 S003     setosa   Petal   Width 0.2
#> S004.4 S004     setosa   Petal   Width 0.2
#> S005.4 S005     setosa   Petal   Width 0.2
#> S006.4 S006     setosa   Petal   Width 0.4
#> S007.4 S007     setosa   Petal   Width 0.3
#> S008.4 S008     setosa   Petal   Width 0.2
#> S009.4 S009     setosa   Petal   Width 0.2
#> S010.4 S010     setosa   Petal   Width 0.1
#> S011.4 S011     setosa   Petal   Width 0.2
#> S012.4 S012     setosa   Petal   Width 0.2
#> S013.4 S013     setosa   Petal   Width 0.1
#> S014.4 S014     setosa   Petal   Width 0.1
#> S015.4 S015     setosa   Petal   Width 0.2
#> S016.4 S016     setosa   Petal   Width 0.4
#> S017.4 S017     setosa   Petal   Width 0.4
#> S018.4 S018     setosa   Petal   Width 0.3
#> S019.4 S019     setosa   Petal   Width 0.3
#> S020.4 S020     setosa   Petal   Width 0.3
#> S021.4 S021     setosa   Petal   Width 0.2
#> S022.4 S022     setosa   Petal   Width 0.4
#> S023.4 S023     setosa   Petal   Width 0.2
#> S024.4 S024     setosa   Petal   Width 0.5
#> S025.4 S025     setosa   Petal   Width 0.2
#> S026.4 S026     setosa   Petal   Width 0.2
#> S027.4 S027     setosa   Petal   Width 0.4
#> S028.4 S028     setosa   Petal   Width 0.2
#> S029.4 S029     setosa   Petal   Width 0.2
#> S030.4 S030     setosa   Petal   Width 0.2
#> S031.4 S031     setosa   Petal   Width 0.2
#> S032.4 S032     setosa   Petal   Width 0.4
#> S033.4 S033     setosa   Petal   Width 0.1
#> S034.4 S034     setosa   Petal   Width 0.2
#> S035.4 S035     setosa   Petal   Width 0.2
#> S036.4 S036     setosa   Petal   Width 0.2
#> S037.4 S037     setosa   Petal   Width 0.2
#> S038.4 S038     setosa   Petal   Width 0.1
#> S039.4 S039     setosa   Petal   Width 0.2
#> S040.4 S040     setosa   Petal   Width 0.2
#> S041.4 S041     setosa   Petal   Width 0.3
#> S042.4 S042     setosa   Petal   Width 0.3
#> S043.4 S043     setosa   Petal   Width 0.2
#> S044.4 S044     setosa   Petal   Width 0.6
#> S045.4 S045     setosa   Petal   Width 0.4
#> S046.4 S046     setosa   Petal   Width 0.3
#> S047.4 S047     setosa   Petal   Width 0.2
#> S048.4 S048     setosa   Petal   Width 0.2
#> S049.4 S049     setosa   Petal   Width 0.2
#> S050.4 S050     setosa   Petal   Width 0.2
#> S051.4 S051 versicolor   Petal   Width 1.4
#> S052.4 S052 versicolor   Petal   Width 1.5
#> S053.4 S053 versicolor   Petal   Width 1.5
#> S054.4 S054 versicolor   Petal   Width 1.3
#> S055.4 S055 versicolor   Petal   Width 1.5
#> S056.4 S056 versicolor   Petal   Width 1.3
#> S057.4 S057 versicolor   Petal   Width 1.6
#> S058.4 S058 versicolor   Petal   Width 1.0
#> S059.4 S059 versicolor   Petal   Width 1.3
#> S060.4 S060 versicolor   Petal   Width 1.4
#> S061.4 S061 versicolor   Petal   Width 1.0
#> S062.4 S062 versicolor   Petal   Width 1.5
#> S063.4 S063 versicolor   Petal   Width 1.0
#> S064.4 S064 versicolor   Petal   Width 1.4
#> S065.4 S065 versicolor   Petal   Width 1.3
#> S066.4 S066 versicolor   Petal   Width 1.4
#> S067.4 S067 versicolor   Petal   Width 1.5
#> S068.4 S068 versicolor   Petal   Width 1.0
#> S069.4 S069 versicolor   Petal   Width 1.5
#> S070.4 S070 versicolor   Petal   Width 1.1
#> S071.4 S071 versicolor   Petal   Width 1.8
#> S072.4 S072 versicolor   Petal   Width 1.3
#> S073.4 S073 versicolor   Petal   Width 1.5
#> S074.4 S074 versicolor   Petal   Width 1.2
#> S075.4 S075 versicolor   Petal   Width 1.3
#> S076.4 S076 versicolor   Petal   Width 1.4
#> S077.4 S077 versicolor   Petal   Width 1.4
#> S078.4 S078 versicolor   Petal   Width 1.7
#> S079.4 S079 versicolor   Petal   Width 1.5
#> S080.4 S080 versicolor   Petal   Width 1.0
#> S081.4 S081 versicolor   Petal   Width 1.1
#> S082.4 S082 versicolor   Petal   Width 1.0
#> S083.4 S083 versicolor   Petal   Width 1.2
#> S084.4 S084 versicolor   Petal   Width 1.6
#> S085.4 S085 versicolor   Petal   Width 1.5
#> S086.4 S086 versicolor   Petal   Width 1.6
#> S087.4 S087 versicolor   Petal   Width 1.5
#> S088.4 S088 versicolor   Petal   Width 1.3
#> S089.4 S089 versicolor   Petal   Width 1.3
#> S090.4 S090 versicolor   Petal   Width 1.3
#> S091.4 S091 versicolor   Petal   Width 1.2
#> S092.4 S092 versicolor   Petal   Width 1.4
#> S093.4 S093 versicolor   Petal   Width 1.2
#> S094.4 S094 versicolor   Petal   Width 1.0
#> S095.4 S095 versicolor   Petal   Width 1.3
#> S096.4 S096 versicolor   Petal   Width 1.2
#> S097.4 S097 versicolor   Petal   Width 1.3
#> S098.4 S098 versicolor   Petal   Width 1.3
#> S099.4 S099 versicolor   Petal   Width 1.1
#> S100.4 S100 versicolor   Petal   Width 1.3
#> S101.4 S101  virginica   Petal   Width 2.5
#> S102.4 S102  virginica   Petal   Width 1.9
#> S103.4 S103  virginica   Petal   Width 2.1
#> S104.4 S104  virginica   Petal   Width 1.8
#> S105.4 S105  virginica   Petal   Width 2.2
#> S106.4 S106  virginica   Petal   Width 2.1
#> S107.4 S107  virginica   Petal   Width 1.7
#> S108.4 S108  virginica   Petal   Width 1.8
#> S109.4 S109  virginica   Petal   Width 1.8
#> S110.4 S110  virginica   Petal   Width 2.5
#> S111.4 S111  virginica   Petal   Width 2.0
#> S112.4 S112  virginica   Petal   Width 1.9
#> S113.4 S113  virginica   Petal   Width 2.1
#> S114.4 S114  virginica   Petal   Width 2.0
#> S115.4 S115  virginica   Petal   Width 2.4
#> S116.4 S116  virginica   Petal   Width 2.3
#> S117.4 S117  virginica   Petal   Width 1.8
#> S118.4 S118  virginica   Petal   Width 2.2
#> S119.4 S119  virginica   Petal   Width 2.3
#> S120.4 S120  virginica   Petal   Width 1.5
#> S121.4 S121  virginica   Petal   Width 2.3
#> S122.4 S122  virginica   Petal   Width 2.0
#> S123.4 S123  virginica   Petal   Width 2.0
#> S124.4 S124  virginica   Petal   Width 1.8
#> S125.4 S125  virginica   Petal   Width 2.1
#> S126.4 S126  virginica   Petal   Width 1.8
#> S127.4 S127  virginica   Petal   Width 1.8
#> S128.4 S128  virginica   Petal   Width 1.8
#> S129.4 S129  virginica   Petal   Width 2.1
#> S130.4 S130  virginica   Petal   Width 1.6
#> S131.4 S131  virginica   Petal   Width 1.9
#> S132.4 S132  virginica   Petal   Width 2.0
#> S133.4 S133  virginica   Petal   Width 2.2
#> S134.4 S134  virginica   Petal   Width 1.5
#> S135.4 S135  virginica   Petal   Width 1.4
#> S136.4 S136  virginica   Petal   Width 2.3
#> S137.4 S137  virginica   Petal   Width 2.4
#> S138.4 S138  virginica   Petal   Width 1.8
#> S139.4 S139  virginica   Petal   Width 1.8
#> S140.4 S140  virginica   Petal   Width 2.1
#> S141.4 S141  virginica   Petal   Width 2.4
#> S142.4 S142  virginica   Petal   Width 2.3
#> S143.4 S143  virginica   Petal   Width 1.9
#> S144.4 S144  virginica   Petal   Width 2.3
#> S145.4 S145  virginica   Petal   Width 2.5
#> S146.4 S146  virginica   Petal   Width 2.3
#> S147.4 S147  virginica   Petal   Width 1.9
#> S148.4 S148  virginica   Petal   Width 2.0
#> S149.4 S149  virginica   Petal   Width 2.3
#> S150.4 S150  virginica   Petal   Width 1.8